% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.r
\name{a.}
\alias{a.}
\alias{a}
\alias{av}
\title{Atomize to a vector of atomic values}
\usage{
a.()

a(...)

av(...)
}
\arguments{
\item{...}{Arguments to be atomized.}
}
\value{
A single atomic vector containing all constituent atomic values in
the full set of arguments supplied in \code{...}.
}
\description{
Reduce all arguments in \code{...} to a single atomic vector of
atomic elements without any additional attributes.
}
\examples{
Fruits <- c("apples" , "bananas", "oranges")
Digits <- c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9)
Both   <- list(Fruits, Digits)
av(Fruits)
av(Digits)
av(Both)
av("celery", "beets", "lettuce")
av("celery", "beets", "lettuce")
}
\seealso{
Other extensions: 
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
