% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\encoding{UTF-8}
\name{alert_help}
\alias{alert_help}
\alias{alert}
\alias{acknowledge}
\alias{choose1}
\alias{chooseN}
\alias{NO}
\alias{YES}
\alias{OK}
\alias{CANCEL}
\alias{ask}
\alias{ask_new}
\alias{choose_dir}
\alias{choose_doc}
\title{Console-Based User Dialog}
\usage{
alert_help()

alert(
  ...,
  title = "alert",
  sub = "",
  ps = "",
  def = "",
  ft = "r|w|b",
  fs = "k|y|p",
  fm = "",
  fp = "k|y|i",
  d = " ",
  clear = FALSE
)

acknowledge(
  ...,
  sub = "",
  ps = "",
  ft = "r|w|b",
  fs = "k|y|p",
  fp = "k|y|i",
  d = " ",
  clear = FALSE
)

choose1(opts, ..., ft = "r|w|b", fs = "k|y|p", d = " ", clear = FALSE)

chooseN(
  opts,
  ...,
  n = NULL,
  min = NULL,
  max = NULL,
  all = base::is.null(c(n, max)),
  none = FALSE,
  ft = "r|w|b",
  fs = "k|y|p",
  d = " ",
  clear = FALSE
)

NO(..., ft = "r|w|b", fs = "k|y|p", d = " ", clear = FALSE)

YES(..., ft = "r|w|b", fs = "k|y|p", d = " ", clear = FALSE)

OK(..., ft = "r|w|b", fs = "k|y|p", d = " ", clear = FALSE)

CANCEL(..., ft = "r|w|b", fs = "k|y|p", d = " ", clear = FALSE)

ask(
  ...,
  Default = "",
  sub = "",
  ft = "r|w|b",
  fs = "k|y|p",
  fm = "",
  fp = "k|y|i",
  d = " ",
  clear = FALSE
)

ask_new(
  old,
  type = "replacement values",
  u = TRUE,
  sub = "",
  ft = "r|w|b",
  fs = "k|y|p",
  fm = "",
  fp = "k|y|i",
  d = "|",
  clear = FALSE
)

choose_dir(
  dir.type = "directory",
  sub = "",
  ft = "r|w|b",
  fs = "k|y|p",
  fm = "",
  fp = "k|y|i",
  d = " ",
  clear = FALSE
)

choose_doc(
  doc.type = "document",
  sub = "",
  ft = "r|w|b",
  fs = "",
  fm = "",
  fp = "k|y|i",
  d = " ",
  clear = FALSE
)
}
\arguments{
\item{...}{An arbitrary, optional number of arguments which are \link[=av]{atomized} into a character scalar message to be posted to the console.}

\item{title}{A title for the alert (\code{type = 'error'} results in stopping execution after posting the alert). Should be short to avoid formatting problems.}

\item{sub}{A character scalar alert subtitle to post to the console on the line following the title. Should be short to avoid formatting problems.}

\item{ps}{A character scalar suffix to post to the console on the line following the alert message contained in \code{...}.}

\item{def}{A character scalar containing a default message if \link[=av]{atomizing} and collapsing \code{...} to a character scalar results in a blank string (\code{""}).}

\item{ft, fs, fm, fp}{Formatting values consistent with the description in the \emph{specifying formats} section giving formatting instructions for, respectively, the title (in \code{title}), subtitle (in \code{sub}), message (in \code{...}), and postscript (in \code{ps}).}

\item{d}{A character scalar delimiter for collapsing \code{...} args into a character scalar.}

\item{clear}{A non-\code{NA} logical scalar indicating whether to clear the console before each interaction with the user.}

\item{opts}{An atomic vector listing opts to choose from.}

\item{n}{An optional \link[=cmp_psw_scl]{complete positive numeric whole-number scalar} (?cmp_psw_scl) indicating the number of opts that must be selected. Must be contained in \code{1:length(opts)}.}

\item{min}{An optional \link[=CMP]{complete} \link[=PSW]{positive numeric whole-number} \link[=SCL]{scalar} indicating the minimum number of opts that may be selected. Must be \code{NULL} when \code{n} is non-\code{NULL}.}

\item{max}{An optional complete positive numeric whole-number scalar indicating the maximum number of opts that may be selected. Must be \code{NULL} when \code{n} is non-\code{NULL}.}

\item{all}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add an \code{{ ALL }} value to \code{opts}.}

\item{none}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add a \code{{ NONE }} value to \code{opts} (implying that it is valid to select none).}

\item{old}{A \link[=chr_vec]{character vec} of unique values to be replaced.}

\item{type}{A character scalar describing the type of replacement values to be entered.}

\item{u}{Scalar \code{TRUE} or \code{FALSE} indicating whether replacement values must be unique.}
}
\value{
\strong{The} \code{NULL} \strong{object} \cr\cr \verb{acknowledge, alert}
\cr\cr  \strong{A character scalar}    \cr\cr \code{choose_dir} (a directory path) \cr \code{choose_doc} (a document path) \cr \code{ask}
\cr\cr  \strong{A character vector}    \cr\cr \code{ask_new}
\cr\cr  \strong{An atomic vector}      \cr\cr \code{chooseN}
\cr\cr  \strong{An atomic scalar}      \cr\cr \code{choose1}
\cr\cr  \strong{A logical scalar}      \cr\cr \verb{CANCEL, YES, NO, OK}
}
\description{
All functions collapse \code{...} args into a \code{d}-delimited prompt. Each posts an alert to the console, posts the prompt (if any), and follows up with a specific action.
}
\details{
The following give templates for what the user sees, where any value derived from arguments will be absent if it is a blank string or resolves to a blank string.
\cr\cr
\strong{\code{alert}}
\tabular{lrll}{   \tab \strong{CONDITION}       \tab   \tab \strong{VALUE}                      \cr
   \tab \code{g(d, title) != ""} \tab   \tab formatted title from \code{title}.  \cr
   \tab \code{g(d, sub) != ""  } \tab   \tab formatted subtitle from \code{sub}. \cr
   \tab \code{g(d, ...) != ""  } \tab   \tab formatted message from \code{...}.  \cr
   \tab                     \tab   \tab \verb{< blank line >}               \cr
   \tab \code{g(d, ps) != ""   } \tab   \tab formatted postscript from \code{ps}.  }
all other templates incorporate a call to \code{alert(.)} with components as shown below.
\cr\cr\strong{\code{acknowledge}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                                \cr
   \tab Title           \tab   \tab \code{'ACKNOWLEDGMENT REQUIRED'}              \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                        \cr
   \tab Message         \tab   \tab \code{'< g(d, ...) >'}                        \cr
   \tab                 \tab   \tab \verb{< blank line >}                         \cr
   \tab Postscript      \tab   \tab \code{'press [return] or [enter] to continue:'} }
\strong{\code{choose1}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                         \cr
   \tab title           \tab   \tab \code{'RESPONSE REQUIRED'}             \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                 \cr
   \tab Message         \tab   \tab \code{'< g(d, ...) >'}                 \cr
   \tab                 \tab   \tab \verb{< blank line >}                  \cr
   \tab Postscript      \tab   \tab \code{'choose an option - Select one'} \cr
   \tab                 \tab   \tab \verb{< blank line >}                  \cr
   \tab CANCEL.Option   \tab   \tab \verb{'X: \{ CANCEL \}}'                 \cr
   \tab Option.1        \tab   \tab \code{'1: < opts[1] >'}                \cr
   \tab Option.2        \tab   \tab \code{'2: < opts[2] >'}                \cr
   \tab ...             \tab   \tab ...                               \cr
   \tab Option.N        \tab   \tab \code{'< N >: < opts[n] >'}            \cr
   \tab Prompt          \tab   \tab \code{'Selection:'}                      }
\strong{\code{chooseN}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                                                                      \cr
   \tab Title           \tab   \tab \code{'RESPONSE REQUIRED'}                                                          \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                                                              \cr
   \tab Message         \tab   \tab \code{'< g(d, ...) >'}                                                              \cr
   \tab                 \tab   \tab \verb{< blank line >}                                                               \cr
   \tab Postscript      \tab   \tab \code{'choose < n | between n1 and n2 > opts - Select one or more'}                 \cr
   \tab                 \tab   \tab \verb{< blank line >}                                                               \cr
   \tab CANCEL.Option   \tab   \tab \code{'X: { CANCEL }'}                                                              \cr
   \tab Option.1        \tab   \tab \code{'1: < opts[1] >'}                                                             \cr
   \tab Option.2        \tab   \tab \code{'2: < opts[2] >'}                                                             \cr
   \tab ...             \tab   \tab ...                                                                            \cr
   \tab Option.N        \tab   \tab \code{'< N >: < opts[n] >'}                                                         \cr
   \tab ALL.option      \tab   \tab \code{'A: { ALL }'}                                                                 \cr
   \tab NONE.option     \tab   \tab \code{'N: { NONE }'}                                                                \cr
   \tab Prompt          \tab   \tab \code{'Enter one or more numbers separated by spaces and then ENTER, or 0 to cancel'} }
\strong{\code{NO}} and \strong{\code{YES}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                         \cr
   \tab Title           \tab   \tab \code{'RESPONSE REQUIRED'}             \cr
   \tab Subtitle        \tab   \tab '\verb{< g(d, sub) >'}                 \cr
   \tab Message         \tab   \tab \code{'< g(d, ...) >'}                 \cr
   \tab                 \tab   \tab \verb{< blank line >}                  \cr
   \tab Postscript      \tab   \tab \code{'choose an option - Select one'} \cr
   \tab                 \tab   \tab \verb{< blank line >}                  \cr
   \tab CANCEL.Option   \tab   \tab \code{'1: { CANCEL }'}                 \cr
   \tab YES.option      \tab   \tab \code{'2: { YES }'}                    \cr
   \tab NO.option       \tab   \tab \code{'3: { NO }'}                     \cr
   \tab Prompt          \tab   \tab \code{'Selection:'}                      }
\strong{\code{OK}} and \strong{\code{CANCEL}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                         \cr
   \tab Title           \tab   \tab \code{'RESPONSE REQUIRED'}             \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                 \cr
   \tab Message         \tab   \tab \code{'< g(d, ...) >'}                 \cr
   \tab                 \tab   \tab \verb{< blank line >}                  \cr
   \tab Postscript      \tab   \tab \code{'choose an option - Select one'} \cr
   \tab                 \tab   \tab \verb{< blank line >}                  \cr
   \tab CANCEL.Option   \tab   \tab \code{'1: { CANCEL }'}                 \cr
   \tab OK.option       \tab   \tab \code{'2: { OK }'}                     \cr
   \tab Prompt          \tab   \tab \code{'Selection: '}                     }
\strong{\code{ask}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}               \cr
   \tab Title           \tab   \tab \code{'RESPONSE REQUIRED'}   \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}       \cr
   \tab Message         \tab   \tab \code{'< g(d, ...) >'}       \cr
   \tab                 \tab   \tab \verb{< blank line >}        \cr
   \tab Postscript      \tab   \tab \code{'enter your response: '} }
\strong{\code{ask_new}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                                                       \cr
   \tab Title           \tab   \tab \code{'RESPONSE REQUIRED'}                                           \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                                               \cr
   \tab Message         \tab   \tab \verb{'Enter a pipe-separated list of < N(opts) > replacement values}\cr
   \tab                 \tab   \tab \verb{for the following pipe-separated original values: '}           \cr
   \tab                 \tab   \tab \verb{< blank line >}                                                \cr
   \tab Old.Values      \tab   \tab \code{'< paste0(old, collapse = " | ") >'}                           \cr
   \tab                 \tab   \tab \verb{< blank line >}                                                \cr
   \tab Postscript      \tab   \tab \code{'enter your response: '}                                       }
\strong{\code{choose_dir}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                                        \cr
   \tab Title           \tab   \tab \code{'ACKNOWLEDGMENT REQUIRED'}                      \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                                \cr
   \tab Message         \tab   \tab \code{'In the next dialog box, select a < dir.type >.'} }
\strong{\code{choose_doc}}
\tabular{lrll}{   \tab \strong{COMPONENT}   \tab   \tab \strong{VALUE}                                        \cr
   \tab Title           \tab   \tab \code{'ACKNOWLEDGMENT REQUIRED'}                      \cr
   \tab Subtitle        \tab   \tab \code{'< g(d, sub) >'}                                \cr
   \tab Message         \tab   \tab \code{'In the next dialog box, select a < doc.type >.'} }
}
\section{Functions}{
\itemize{
\item \code{alert()}: Collapses \code{...} into an alert message using \code{d} as a delimiter and issues the alert to the console with the specified title, subtitle, postscript and style/color formatting. Clears the console first if \code{clear = TRUE}.

\item \code{acknowledge()}: Posts an alert message with the title \verb{ACKNOWLEDGMENT REQUIRED} and waits for the user to press the \verb{[return]} or \verb{[enter]} key.

\item \code{choose1()}: Posts an alert message and prompts the user to select one of the options in \code{opts}.

\item \code{chooseN()}: Posts an alert message and prompts the user to select from \code{0} to \code{N} of the \code{N} options in \code{opts}. The minimum and maximum number of options to be selected are controlled by the arguments \code{n}, \code{min}, \code{max}, \code{all}, and \code{none}.

\item \code{NO()}: Posts an alert message and prompts the user to choose either \code{yes} or \code{no} as a response. Returns \code{TRUE} when the user chooses \code{yes}. Returns \code{FALSE} when the user chooses \code{no}.

\item \code{YES()}: Posts an alert message and prompts the user to choose either \code{yes} or \code{no} as a response. Returns \code{TRUE} when the user chooses \code{no}. Returns \code{FALSE} when the user chooses \code{yes}.

\item \code{OK()}: Posts an alert message and prompts the user to choose either \code{ok} or \code{cancel} as a response. Returns \code{TRUE} when the user chooses \code{ok}. Returns \code{FALSE} when the user chooses \code{cancel}.

\item \code{CANCEL()}: Posts an alert message and prompts the user to choose either \code{ok} or \code{cancel} as a response. Returns \code{TRUE} when the user chooses \code{cancel}. Returns \code{FALSE} when the user chooses \code{ok}.

\item \code{ask()}: Posts an alert question and prompts the user to answer the question via entering text and hitting \link{return} or \link{enter}. Returns the entered text.

\item \code{ask_new()}: Posts an alert message describing multiple values and prompts the user to enter new/replacement values. Returns a vector of new/replacement values.

\item \code{choose_dir()}: Posts an alert message notifying the user that a dialog box will open and prompting the user to choose a directory/folder. Returns the path to the selected directory/folder.

\item \code{choose_doc()}: Posts an alert message notifying the user that a dialog box will open and prompting the user to choose a file/document. Returns the path to the selected file/document.

}}
\section{Specifying formats}{
 When formatting arguments (\code{ft}, \code{fs}, \code{fm}, and \code{fp}) take the special value \code{""}, the corresponding alert elements (\code{title}, \code{sub}, \code{...}, and \code{ps}, respectively) are posted to the console without special formatting.
\cr\cr Otherwise, formatting arguments must be \link[uj::cmp_str_vec]{complete string vecs} that when \link[=av]{atomized} and \link[=ssplit]{split along pipes} results in a three-element character vector, the first element of which is used to specify \link[=bg]{text background color} and must be a value from \code{\link{bg_vals}()}, the second element of which is used to specify \link[=fg]{text foreground color} and must be a value from \code{\link{fg_vals}()}, and the last of which specifies \link[=st]{text style} and must be a value from \code{\link{st_vals}()}.
}

\examples{
\dontrun{
  egAlert <- function() {
    partA <- "two-part"
    partB <- "message"
    fmtT  <- base::c("yellow", "red", "plain")
    fmtS  <- base::c("blk|wht", "und")
    fmtM  <- "b|y|i"
    title <- "Title"
    sub   <- "Subtitle"
    opts  <- base::paste("option", letters[1:10])
    mssg1 <- "Do you want to continue?"
    mssg2 <- "Why do you want to continue?"
    uj::alert(partA, partB, d = " ")
    uj::alert(partA, partB, title = title, d = " ")
    uj::alert(partA, partB, sub = sub, d = " ")
    uj::alert(partA, partB, title = title, sub = sub, d = " ")
    uj::alert(partA, partB, title = title, sub = sub, d = " ", fm = fmtM)
    uj::alert(partA, partB, title = title, sub = sub, d = " ", fs = fmtS)
    uj::alert(partA, partB, title = title, sub = sub, d = " ", ft = fmtT)
    uj::alert(partA, partB, title = title, sub = sub, d = " ", ft = fmtT, fs = fmtS, fm = fmtM)
    uj::alert(title = title, sub = sub, ft = fmtT, fs = fmtS)
    uj::alert(title = title, sub = sub)
    uj::alert(sub = sub, fs = fmtS)
    uj::acknowledge(partA, partB)
    ch1 <- uj::choose1(opts)
    ch2 <- uj::chooseN(opts)
    ch3 <- uj::chooseN(opts, all = T, none = F, min = 6, max = 10)
    no  <- uj::NO(mssg1)
    ok  <- uj::OK(mssg2)
    yes <- uj::YES(mssg1)
    can <- uj::CANCEL(mssg1)
    ASK <- uj::ask(mssg2)
    NEW <- uj::ask_new(opts)
    dir <- uj::choose_dir(dir.type = "directory for R scripts")
    doc <- uj::choose_doc(doc.type = "document to read")
    base::list(ch1 = ch1, ch2 = ch2, ch3 = ch3, no  = no , ok  = ok, yes = yes,
               can = can, ASK = ASK, NEW = NEW, dir = dir, doc = doc)
  }
  egAlert()
}
}
