% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases_uj.R
\name{aliases_uj}
\alias{aliases_uj}
\alias{dtas}
\alias{dtie}
\alias{dtsub}
\alias{dtcols}
\alias{dtjoin}
\alias{dtrows}
\alias{dtwide}
\alias{dpgrp}
\alias{rdcsv}
\alias{rdclip}
\alias{rdxsv}
\alias{rdtsv}
\alias{up}
\alias{dn}
\alias{spf}
\alias{mid}
\title{Wrappers for functions in other packages}
\usage{
aliases_uj()

dtas(x)

dtie(x)

dtsub(x, ir, ic)

dtcols(x, ic)

dtjoin(x, y, by)

dtrows(x, ir)

dtwide(x, f, v)

dpgrp(x, keys)

rdcsv(file = NULL)

rdclip()

rdxsv(file = NULL, delim = ",")

rdtsv(file = NULL)

up(x)

dn(x)

spf(fmt, ...)

mid(x, first, last)
}
\arguments{
\item{x}{A data.table object (for aliases beginning with \code{dt}. For
aliases of base package functions, as expected for those functions.}

\item{ir, ic}{Indexing vector giving rows, columns of \code{x} to select.}

\item{f}{data.table casting formula}

\item{v}{data.table variable name of column containing values to be cast
wide.}

\item{first, last}{Positive integers indicating first and last characters to
extract.}
}
\description{
Wrappers for functions in other packages
}
\section{Functions}{
\itemize{
\item \code{dtas}: Thin wrapper for
\code{\link[data.table]{as.data.table}}.

\item \code{dtie}: If \code{x} is not already a data.table, coerce it to
a data.table.

\item \code{dtsub}: Extract rows and columns of a data.table using indices
or names rather than the formal data.table syntax for column selection.

\item \code{dtcols}: Extract columns of a data.table using indices or names
rather than the formal data.table syntax for column selection.

\item \code{dtjoin}: merge two data.tables by specified columns.

\item \code{dtrows}: Extract rows of a data.table using indices or names.

\item \code{dtwide}: Reshape data.table \code{x} from long to wide.

\item \code{dpgrp}: Limited functionality thin wrapper for dplyr function
\code{\link[dplyr]{group_by}} (group cases for group-based analyses).

\item \code{rdcsv}: Alias for \code{\link[readr]{read_csv}}.

\item \code{rdclip}: Alias for \code{\link[readr]{clipboard}} to read text
from the system clipboard.

\item \code{rdxsv}: Alias for \code{\link[readr]{read_delim}} to read
(arbitrary) delimiter-separated values from text files.

\item \code{rdtsv}: Alias for \code{\link[readr]{read_tsv}} to read tab
separated values from text files.

\item \code{up}: Alias for \code{base::ceiling}.

\item \code{dn}: Alias for \code{base::floor}.

\item \code{spf}: Alias for \code{base::sprintf}.

\item \code{mid}: Alias for \code{base::substr(x, first, last)}.
}}

\concept{wrappers}
