% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\name{as_mmm.}
\alias{as_mmm.}
\alias{as_clr}
\alias{as_fun}
\alias{as_chr}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to base mode or extended mode.}
\usage{
as_mmm.()

as_clr(x, na = F)

as_fun(x)

as_chr()

as_int()

as_num()

as_lgl()

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{An object}

\item{na}{\link[cmp_lgl_scl]{Complete logical scalar} indicating whether
\code{NA} values qualify as missing color representations.}

\item{levs}{\link[cmp_vec]{Complete atomic vec} giving factor levels.}
}
\description{
Coerce to base mode or extended mode.
}
\section{Functions}{
\itemize{
\item \code{as_clr()}: Coerce a valid ℝ color representation to hexadecimal
RGB character representation.

\item \code{as_fun()}: Return \code{x} if it is a function, otherwise, search
for a function named \code{x} and return that function.

\item \code{as_chr()}: \code{as.character} wrapper.

\item \code{as_int()}: \code{as.integer} wrapper.

\item \code{as_num()}: \code{as.numeric} wrapper.

\item \code{as_lgl()}: \code{as.logical} wrapper.

\item \code{as_ord()}: Coerce to ordered factor.

\item \code{as_uno()}: Coerce to unordered factor.

}}
\seealso{
Other props: 
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{is_ccc.}()},
\code{\link{is_mmm.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()},
\code{\link{xxx.}()}
}
\concept{props}
