% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\name{as_mmm}
\alias{as_mmm}
\alias{as_clr}
\alias{as_fun}
\alias{as_chr}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to base or extended mode.}
\usage{
as_clr(x, na = F)

as_fun(x)

as_chr(x, ...)

as_int(x, ...)

as_num(x, ...)

as_lgl(x, ...)

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{For \code{as_clr}, an object of mode character; for \code{as_fun}, a character scalar function name or a function object; for \code{as_ord} and \code{as_uno}, an atomic object; and for all others, any R object.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values qualify as missing color representations.}

\item{...}{Further arguments passed to or from other methods.}

\item{levs}{A \link[=cmp_vec]{complete atomic vec} of factor levels (ordered factor levels for \code{as_ord}).}
}
\value{
\itemize{
\item \strong{\code{as_fun}}: a function object.
\item \strong{\code{as_int}}: an object of mode \code{'integer'}.
\item \strong{\code{as_num}}: an object of mode \code{'numeric'}.
\item \strong{\code{as_lgl}}: an object of mode \code{'logical'}.
\item \strong{\code{as_ord}}: an object of mode \verb{'ordered'`` (factor). \\item **}as_uno\verb{**: an object of \link[=mmm]\{extended mode\} }'uno'\verb{ (unordered factor). \\item **}as_clr, as_chr\verb{**: an object of mode }'character'`.
}
}
\description{
Functions to coerce objects to \link[base:mode]{base mode} or \link[=mmm]{extended mode}:
\itemize{
\item \strong{\code{as_clr}}: coerces a valid R color representation of mode character to hexadecimal RGB character representation.
\item \strong{\code{as_fun}}: returns \code{x} if it is a function, otherwise, search for a function named \code{x} and return it.
\item \strong{\code{as_chr}}: thinly wraps \code{\link[base:as.character]{base::as.character}}.
\item \strong{\code{as_int}}: thinly wraps \code{\link[base:as.integer]{base::as.integer}}.
\item \strong{\code{as_num}}: thinly wraps \code{\link[base:as.numeric]{base::as.numeric}}.
\item \strong{\code{as_lgl}}: thinly wraps \code{\link[base:as.logical]{base::as.logical}}.
\item \strong{\code{as_ord}}: wraps \code{base::factor(x, levels = levs, ordered = TRUE)}.
\item \strong{\code{as_uno}}: wraps \code{base::factor(x, levels = levs, ordered = FALSE)}.
}
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
