% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\name{as_mmm_uj}
\alias{as_mmm_uj}
\alias{as_clr}
\alias{as_fun}
\alias{as_chr}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to base mode or extended mode.}
\usage{
as_mmm_uj()

as_clr(x, na = F)

as_fun(x)

as_chr()

as_int()

as_num()

as_lgl()

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{An object}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values
qualify as missing color representations.}

\item{levs}{Character vector giving ordered factor levels.}
}
\description{
Coerce to base mode or extended mode.
}
\section{Functions}{
\itemize{
\item \code{as_clr}: Coerce a valid ℝ color representation to hexadecimal RGB
character representation.

\item \code{as_fun}: Return \code{x} if it is a function, otherwise, search for
a function with the name contained in \code{x} and return that function.

\item \code{as_chr}: \code{as.character} wrapper.

\item \code{as_int}: \code{as.integer} wrapper.

\item \code{as_num}: \code{as.numeric} wrapper.

\item \code{as_lgl}: \code{as.logical} wrapper.

\item \code{as_ord}: Coerce to ordered factor.

\item \code{as_uno}: Coerce to unordered factor.
}}

\seealso{
Other props: 
\code{\link{ccc_uj}()},
\code{\link{cmp_ccc_uj}()},
\code{\link{cmp_mmm_ccc_uj}()},
\code{\link{cmp_mmm_uj}()},
\code{\link{comparable_uj}()},
\code{\link{compatible_uj}()},
\code{\link{ddd_uj}()},
\code{\link{eee_uj}()},
\code{\link{fff_uj}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm_uj}()},
\code{\link{meets_uj}()},
\code{\link{mmm_ccc_uj}()},
\code{\link{mmm_uj}()},
\code{\link{sss_uj}()},
\code{\link{ttt_ccc_uj}()},
\code{\link{ttt_mmm_uj}()},
\code{\link{ttt_uj}()},
\code{\link{xxx_uj}()}
}
\concept{props}
