% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{ask_uj}
\alias{ask_uj}
\alias{ask}
\alias{ask1}
\alias{askn}
\alias{asksn}
\alias{asks1}
\alias{asknew}
\title{Ask user for typed input.}
\usage{
ask_uj()

ask(..., def. = v(blank), cancel. = ".stop")

ask1(opts., what., more. = F)

askn(opts., what., all. = T, none. = T, more. = F)

asksn(opts., what., per. = 9)

asks1(opts., what., per. = 9)

asknew(old., what., unique. = T)
}
\arguments{
\item{...}{Objects to be atomized to create the question.}

\item{def.}{Character scalar default answer value.}

\item{cancel.}{Character scalar value to return if user clicks the
\code{cancel} button. The value \code{cancel = '.stop'} is an indicator
that an error should be thrown if the user clicks the \code{cancel} button.}

\item{opts.}{Character vector of options to choose from.}

\item{what.}{Character scalar describing type of option being presented.}

\item{more.}{Logical scalar indicating whether there are more options than
could be presented in a single dialog box that still could be presented.}

\item{all., none.}{Logical scalars indicating whether the user should be able
to select all options or none of the options, respectively.}

\item{per.}{Positive integer scalar indicating the maximum number of options
to present in a single dialog box interaction.}

\item{unique.}{Logical scalar indicating whether new values must be unique.}
}
\value{
Character scalar.
}
\description{
Ask user for typed input.
}
\section{Functions}{
\itemize{
\item \code{ask}: Ask the user for typed input return as is.

\item \code{ask1}: Ask user to select a single option from among a list of
possible options.

\item \code{askn}: Ask user to select a one or more options from among a
list of possible options.

\item \code{asksn}: Ask user to select a one or more options from among a
list of possible options, possibly across multiple rounds of interaction.

\item \code{asks1}: Ask user to select 1 options from among a list of
possible options, possibly across multiple dialog boxes.

\item \code{asknew}: Ask user to enter a space- or comma-separated list of
replacement values for existing values.
}}

\seealso{
Other user_interaction: 
\code{\link{dialogs_uj}()}
}
\concept{user_interaction}
