% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMODE.R
\encoding{UTF-8}
\name{asMODE}
\alias{asMODE}
\alias{asCHR}
\alias{asCLR}
\alias{asFUN}
\alias{asINT}
\alias{asNUM}
\alias{asLGL}
\alias{asORD}
\alias{asUNO}
\title{Coerce to base mode or \code{xmode}}
\usage{
asMODE(x, mode, levs = NULL, na = FALSE)

asCHR(x, na = FALSE)

asCLR(x, na = FALSE)

asFUN(x)

asFUN(x)

asINT(x)

asNUM(x)

asLGL(x)

asORD(x, levs)

asUNO(x, levs)
}
\arguments{
\item{x}{For \code{asCLR}, an object of mode character; for \code{asFUN}, a character scalar function name or a function object; for \code{asORD} and \code{asUNO}, an atomic object; and for all others, any R object.}

\item{levs}{A \link[=cmp_vec]{complete atomic vec} of factor levels (ordered factor levels for \code{asORD}).}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values are acceptable.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\strong{An object of base mode} \code{'character'}                         \cr\cr \code{asCHR}
\cr\cr  \strong{An object of base mode} \code{'integer'}                           \cr\cr \code{asINT}
\cr\cr  \strong{An object of base mode} \code{'logical'}                           \cr\cr \code{asLGL}
\cr\cr  \strong{An object of base mode} \code{'numeric'}                           \cr\cr \code{asNUM}
\cr\cr  \strong{An object of \link[=mmm]{xmode}} \code{'ord'} (ordered factor)     \cr\cr \code{asORD}
\cr\cr  \strong{An object of xmode} \code{'uno'} (unordered factor)                \cr\cr \code{asUNO}
\cr\cr  \strong{An object of xmode} \code{'clr'} (hex color in form \code{'#RRGGBBAA'}) \cr\cr \code{asCLR}
}
\description{
Functions to coerce objects to \link[base:mode]{base mode} or \code{\link[=mmm]{xmode}}.
}
\details{
Functions in this family are as follows:
\tabular{ll}{  \code{asCHR}   \tab Thinly wraps \code{\link[base]{as.character}}.                                                        \cr
  \code{asINT}   \tab Thinly wraps \code{\link[base]{as.integer}}.                                                          \cr
  \code{asNUM}   \tab Thinly wraps \code{\link[base]{as.numeric}}.                                                          \cr
  \code{asLGL}   \tab Thinly wraps \code{\link[base]{as.logical}}.                                             \cr   \tab   \cr
  \code{asORD}   \tab As xmode \code{'ord'} (ordered factor). Wraps \code{factor(x, levels = levs, ordered = TRUE)}.     \cr   \tab   \cr
  \code{asUNO}   \tab As xmode \code{'uno'} (unordered factor). Wraps \code{factor(x, levels = levs, ordered = FALSE)}.  \cr   \tab   \cr
  \code{asCLR}   \tab Coerces character color values to hexadecimal.                                           \cr   \tab   \cr
  \code{asFUN}   \tab If \code{x} is a character scalar function return that function, but if \code{x} is a function object, return it. }
}
\examples{
bins. <- sample(c(0, 1), 10, replace = T)
chrs. <- c("3.14", "2.72", "1.41")
clrs. <- c("red", "#AABBCC", "#AABBCCDD", "blue")
nums. <- c(pi, exp(1), sqrt(2))

asFUN(unique)
asFUN("unique")

bins.
asLGL(bins.)

chrs.
asNUM(chrs.)
asINT(chrs.)

clrs.
asCLR(clrs.)

nums.
asCHR(nums.)

asUNO(clrs., levs = unique(clrs.))
asORD(clrs., levs = sort(unique(clrs.)))
}
\seealso{
Other properties: 
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
