% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{at}
\alias{at}
\alias{at_names}
\alias{is_at}
\alias{add_at}
\alias{set_at}
\alias{rm_at}
\alias{xat}
\title{Manage attributes}
\usage{
at(X, ..., A = FALSE)

at_names(X)

is_at(X, Name)

add_at(X, ...)

set_at(X, ...)

rm_at(X, ...)

xat(X, ...)
}
\arguments{
\item{X}{An R object.}

\item{...}{Optional unquoted names of attributes of \code{X} or character objects. When none are provided, indicates all attributes of \code{X}.}

\item{A}{\code{TRUE} or \code{FALSE} indicating whether to \link[av]{atomize} the result and convert the resulting vector to character.}
}
\value{
\strong{An object}                                   \cr\cr \verb{add_at, set_at, rm_at, xat, at}
\cr\cr  \strong{The} \code{NULL} \strong{object or A character vector} \cr\cr \code{at_names}
}
\description{
Get names of, add new, remove, and extract specific object attributes.
}
\details{
\tabular{ll}{  \verb{rm_at, xat}   \tab Removes attributes of \code{X} with names matching unquoted \code{...} args  \eqn{^{(1)}}. \cr   \tab   \cr
  \code{at_names}     \tab Thinly wraps \code{names(attributes(X))}.                                             \cr   \tab   \cr
  \code{add_at}       \tab Adds named \code{...} args to attributes of \code{X}.                                      \cr   \tab   \cr
  \code{set_at}       \tab Sets attributes of \code{X} to the named values of \code{...}                              \cr
  \code{at}           \tab Gets attributes of \code{X} with names matching unquoted \code{...} args  \eqn{^{(1)}}.    \cr   \tab     }
\tabular{l}{  \eqn{^{(1)}} No \code{...} args indicates all attributes of \code{X}.}
}
