% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\encoding{UTF-8}
\name{atomize}
\alias{atomize}
\alias{atoms}
\alias{av}
\alias{a}
\alias{sort_av}
\alias{sv}
\alias{glue_av}
\alias{gv}
\alias{paste_av}
\alias{pv}
\alias{unique_av}
\alias{uv}
\alias{sort_uv}
\alias{suv}
\alias{which_av}
\alias{wv}
\title{Atomize}
\usage{
atomize(...)

atoms(...)

av(...)

a(...)

sort_av(...)

sv(...)

glue_av(..., g = "")

gv(..., g = "")

paste_av(..., p = "", s = "")

pv(..., p = "", s = "")

unique_av(...)

uv(...)

sort_uv(...)

suv(...)

which_av(...)

wv(...)
}
\arguments{
\item{...}{Arguments to be atomized. Expected to atomize to a logical vector for \code{which_av} and its aliases \code{wav} and \code{wv}.}

\item{g}{Character scalar glue.}

\item{p}{Character scalar prefix.}

\item{s}{Character scalar suffix.}
}
\value{
\strong{A character scalar} \cr\cr \verb{glue_av, gav, gv}
\cr\cr  \strong{A character vector} \cr\cr \verb{paste_av, pav, pv}
\cr\cr  \strong{An integer vector}  \cr\cr \verb{which_av, wav, wv}
\cr\cr  \strong{An atomic vector}   \cr\cr  All others
}
\description{
Get (and maybe process) a vector of all atoms contained in the full set of \code{...} args.
}
\details{
Functions in this family are:
\tabular{ll}{  \verb{atomize, atoms, av, a}   \tab Gets a vector of atomic values in \code{...} args. \cr
  \verb{unique_av, uav, uv}      \tab Calls \code{unique(av(...))}                       \cr
  \verb{which_av, wav, wv}       \tab Calls \code{which(av(...))}                        \cr
  \verb{paste_av, pav, pv}       \tab Calls \code{paste0(p, av(...), s)}                 \cr
  \verb{sort_av, sav, sv}        \tab Calls \code{sort(av(...))}.                        \cr
  \verb{glue_av, gav, gv}        \tab Calls \verb{paste0(av(...) collapse = g)}          \cr
  \verb{sort_uv, suv}            \tab Calls \code{sort(unique(av(...)))}                   }
}
\examples{
x <- list("a", "b", "c")
y <- data.frame(1:3)
x
y
a(x)
av(y)
atoms(x, y)
atomize(x, "d", y, 4)
glue_av(letters, LETTERS, g = "-")
paste_av(letters, p = "\u2022 ", s = ".")
sort_av(letters, LETTERS, 0:9)
sort_uv(letters, LETTERS, 0:9)
unique_av(letters, "d", 0:9, 5)
}
\seealso{
Other extensions: 
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{evals}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{nis}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}

Other values: 
\code{\link{Nth}()},
\code{\link{na}()},
\code{\link{values}()}
}
\concept{extensions}
\concept{values}
