% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{att}
\alias{att}
\alias{atts}
\alias{att_names}
\alias{is_att}
\alias{add_att}
\alias{set_att}
\alias{rm_att}
\alias{xatt}
\title{Manage attributes}
\usage{
att(x, ..., .A = FALSE)

atts(x)

att_names(x)

is_att(x, name)

add_att(x, ...)

set_att(x, ...)

rm_att(x, ...)

xatt(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Optional unquoted names of attributes of \code{x} or character objects. When none are provided, indicates all attributes of \code{x}.}

\item{.A}{\code{TRUE} or \code{FALSE} indicating whether to \link[av]{atomize} the result and convert the resulting vector to character.}
}
\value{
\strong{An object}
\cr\cr \verb{add_att, set_att, rm_att, xatt}
\cr\cr  \strong{THE} \code{NULL} \strong{object or a list}
\cr\cr \verb{atts, att}
\cr\cr  \strong{The} \code{NULL} \strong{object or a character vector}
\cr\cr \code{att_names}
}
\description{
Get names of, add new, remove, and extract specific object attributes.
}
\details{
\tabular{ll}{  \verb{rm_att, xatt}   \tab Removes attributes of \code{x} with names matching unquoted \code{...} args  \eqn{^{(1)}}. \cr   \tab   \cr
  \code{att_names}      \tab Thinly wraps \code{names(attributes(x))}.                                             \cr   \tab   \cr
  \code{add_att}        \tab Adds named \code{...} args to attributes of \code{x}.                                      \cr   \tab   \cr
  \code{set_att}        \tab Sets attributes of \code{x} to the named values of \code{...}                              \cr   \tab   \cr
  \code{atts}           \tab Gets all attributes of \code{x}                                                       \cr   \tab   \cr
  \code{att}            \tab Gets attributes of \code{x} with names matching unquoted \code{...} args  \eqn{^{(1)}}.    \cr   \tab     }
\tabular{l}{  \eqn{^{(1)}} No \code{...} args indicates all attributes of \code{x}.}
}
