% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av.R
\name{av}
\alias{av}
\title{Atomize (reduce to atomic values)}
\usage{
av(...)
}
\arguments{
\item{...}{Arguments to be atomized.}
}
\value{
A single atomic vector containing all constituent atomic values in
the full set of arguments supplied in \code{...}.
}
\description{
Reduce all arguments in \code{...} to a single atomic vector of
atomic elements without any additional attributes.
}
\examples{
Fruits <- c("apples" , "bananas", "oranges")
Digits <- c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9)
Both   <- list(Fruits, Digits)
av(Fruits)
av(Digits)
av(Both)
av("celery", "beets", "lettuce")
av("celery", "beets", "lettuce")
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
