% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av.R
\name{av_uj}
\alias{av_uj}
\alias{av}
\title{Atomize (reduce to atomic values)}
\usage{
av_uj()

av(...)
}
\arguments{
\item{...}{Arguments to be atomized.}
}
\value{
A single atomic vector containing all constituent atomic values in
the full set of arguments supplied in \code{...}.
}
\description{
Atomize (reduce to atomic values)
}
\section{Functions}{
\itemize{
\item \code{av}: Reduce all arguments in \code{...} to a single atomic
vector of atomic elements without any additional attributes.
}}

\examples{
Fruits <- c("apples" , "bananas", "oranges")
Digits <- c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9)
Both   <- list(Fruits, Digits)
av(Fruits)
av(Digits)
av(Both)
av("celery", "beets", "lettuce")
av("celery", "beets", "lettuce")
}
\seealso{
Other meta: 
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
