% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tools.R
\name{base_tools}
\alias{base_tools}
\alias{up}
\alias{dn}
\alias{spf}
\title{Aliases and simple wraps for functions in the package \code{base}.}
\usage{
up()

dn()

spf(fmt, ...)
}
\description{
Alias for \code{ceiling}
}
\section{Functions}{
\itemize{
\item \code{dn}: Alias for \code{floor}.

\item \code{spf}: Alias for \code{sprintf}.
}}

