% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb.R
\name{bbb}
\alias{bbb}
\alias{bbb_props}
\alias{is_bbb_spec}
\alias{ibbb}
\alias{iatm}
\alias{idef}
\alias{ifun}
\alias{inil}
\alias{inll}
\alias{ipop}
\alias{ircr}
\title{Base properties}
\usage{
bbb(x)

bbb_props()

is_bbb_spec(spec)

ibbb(x, spec, ...)

iatm(x, ...)

idef(x, ...)

ifun(x, ...)

inil(x, ...)

inll(x, ...)

ipop(x, ...)

ircr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more base properties (i.e., from \code{bbb_props()}). Base properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_bbb_spec}   \tab A logical
\cr      \code{ibbb}   \tab scalar.
\cr      \code{iBBB}   \tab   
\cr               \tab   
\cr \code{bbb_props}   \tab A character
\cr       \code{bbb}   \tab vector.
}
\eqn{^{2.}} \code{xxx} represents a base property.
}
\description{
\strong{Base property values}
An object's base properties are defined by its most basic structural properties as described in the following table:
\tabular{rll}{
  \code{'atm'}   \tab Atomic      \tab Atomic objects.
\cr \code{'def'}   \tab Defined     \tab Non-\code{NULL} objects.
\cr \code{'nil'}   \tab Nil         \tab \code{0}-length, non-\code{NULL} objects..
\cr \code{'nll'}   \tab Null        \tab The \code{NULL} object.
\cr \code{'pop'}   \tab Populated   \tab Length-\verb{1+} objects.
\cr \code{'rcr'}   \tab Recursive   \tab Data.frames and lists.
\cr           \tab             \tab   
\cr \code{'fun'}   \tab Function    \tab Function objects or character scalar names of functions accessible to the calling function.
}
\strong{Base property functions}
\tabular{rl}{
  \code{is_bbb_spec}   \tab Is \code{spec} is a base property spec?
\cr \code{bbb_props}   \tab Gets all base property values.
\cr      \code{ibbb}   \tab Does \code{x} match base property spec \code{spec}?
\cr      \code{iBBB}   \tab Does \code{x} match property \code{BBB}?
\cr       \code{bbb}   \tab Gets all of \code{x}'s base properties.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_bbb_spec("nil|nll")
bbb_props()
ibbb(NULL, "nil|nll")
ipop(NA)
iatm(list(letters))
bbb(NA)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
