% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_mmm.R
\name{bbb_mmm}
\alias{bbb_mmm}
\alias{bbb_mmm_props}
\alias{atm_ch1}
\alias{atm_ch3}
\alias{atm_chr}
\alias{atm_clr}
\alias{atm_evn}
\alias{atm_fac}
\alias{atm_frc}
\alias{atm_ind}
\alias{atm_lgl}
\alias{atm_neg}
\alias{atm_ngw}
\alias{atm_nng}
\alias{atm_nnw}
\alias{atm_nps}
\alias{atm_npw}
\alias{atm_nst}
\alias{atm_num}
\alias{atm_odd}
\alias{atm_ord}
\alias{atm_pct}
\alias{atm_pos}
\alias{atm_ppn}
\alias{atm_psw}
\alias{atm_srt}
\alias{atm_str}
\alias{atm_uno}
\alias{atm_whl}
\alias{pop_ch1}
\alias{pop_ch3}
\alias{pop_chr}
\alias{pop_clr}
\alias{pop_evn}
\alias{pop_fac}
\alias{pop_frc}
\alias{pop_ind}
\alias{pop_lgl}
\alias{pop_neg}
\alias{pop_ngw}
\alias{pop_nng}
\alias{pop_nnw}
\alias{pop_nps}
\alias{pop_npw}
\alias{pop_nst}
\alias{pop_num}
\alias{pop_odd}
\alias{pop_ord}
\alias{pop_pct}
\alias{pop_pos}
\alias{pop_ppn}
\alias{pop_psw}
\alias{pop_srt}
\alias{pop_str}
\alias{pop_uno}
\alias{pop_whl}
\title{Base + Extended Mode Properties}
\usage{
bbb_mmm(x, bbb, mmm, ...)

bbb_mmm_props()

atm_ch1(x, ...)

atm_ch3(x, ...)

atm_chr(x, ...)

atm_clr(x, ...)

atm_evn(x, ...)

atm_fac(x, ...)

atm_frc(x, ...)

atm_ind(x, ...)

atm_lgl(x, ...)

atm_neg(x, ...)

atm_ngw(x, ...)

atm_nng(x, ...)

atm_nnw(x, ...)

atm_nps(x, ...)

atm_npw(x, ...)

atm_nst(x, ...)

atm_num(x, ...)

atm_odd(x, ...)

atm_ord(x, ...)

atm_pct(x, ...)

atm_pos(x, ...)

atm_ppn(x, ...)

atm_psw(x, ...)

atm_srt(x, ...)

atm_str(x, ...)

atm_uno(x, ...)

atm_whl(x, ...)

pop_ch1(x, ...)

pop_ch3(x, ...)

pop_chr(x, ...)

pop_clr(x, ...)

pop_evn(x, ...)

pop_fac(x, ...)

pop_frc(x, ...)

pop_ind(x, ...)

pop_lgl(x, ...)

pop_neg(x, ...)

pop_ngw(x, ...)

pop_nng(x, ...)

pop_nnw(x, ...)

pop_nps(x, ...)

pop_npw(x, ...)

pop_nst(x, ...)

pop_num(x, ...)

pop_odd(x, ...)

pop_ord(x, ...)

pop_pct(x, ...)

pop_pos(x, ...)

pop_ppn(x, ...)

pop_psw(x, ...)

pop_srt(x, ...)

pop_str(x, ...)

pop_uno(x, ...)

pop_whl(x, ...)
}
\arguments{
\item{x}{An R object}

\item{bbb}{A character scalar containing an
base property from bbb_props().}

\item{mmm}{A character scalar containing an
extended mode property from mmm_props().}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{\code{bbb_mmm_props}}: A character vector.
\cr\cr\strong{bbb_mmm, BBB_MMM}: A logical scalar.
}
\description{
NOTE: \code{BBB} and \code{MMM} are used as placeholders for any
given base and extended mode, respectively.
\cr\cr
\strong{\code{bbb_mmm}}: Evaluates whether \code{x} matches the
base property specified in the argument \code{bbb} and matches the
extended mode specified in the argument \code{mmm}.
(subject to any restrictions in \code{...}).
\cr\cr
\strong{\code{BBB_MMM}}: Evaluates whether \code{x} matches
base property \code{BBB} and extended mode \code{MMM}.
(subject to any restrictions in \code{...}).
\cr\cr
\strong{\code{bbb_mmm_props}}: Gets a character vector of all possible
base + extended mode properties.
}
\section{Specifying Count and Value Restrictions}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar less-than bound on values of \code{x}.        \cr
\code{le}     \tab Scalar less-than-or-equal bound on values of
\code{x}.                                            \cr
\code{ge}     \tab Scalar greater-than-or-equal bound on values of
\code{x}.                                            \cr
\code{gt}     \tab Scalar greater-than bound on values of \code{x}.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc_props}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
