% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfuns.R
\name{bin_uj}
\alias{bin_uj}
\alias{\%IS\%}
\alias{\%!IS\%}
\alias{\%EQ\%}
\alias{\%!EQ\%}
\alias{\%AND\%}
\alias{\%OR\%}
\alias{\%!\%}
\alias{\%1\%}
\alias{\%IN\%}
\alias{\%!IN\%}
\alias{\%HAS\%}
\alias{\%LACKS\%}
\title{Binary functions returning scalar \code{TRUE} or \code{FALSE}.}
\usage{
bin_uj()

x \%IS\% y

x \%!IS\% y

x \%EQ\% y

x \%!EQ\% y

x \%AND\% y

x \%OR\% y

x \%!\% y

x \%1\% y

x \%IN\% y

x \%!IN\% y

x \%HAS\% y

x \%LACKS\% y
}
\arguments{
\item{x, y}{Any objects.}
}
\description{
Binary functions returning scalar \code{TRUE} or \code{FALSE}.
}
\section{Functions}{
\itemize{
\item \code{\%IS\%}: Evaluates whether \code{x} and \code{y} are identical.

\item \code{\%!IS\%}: Evaluates whether \code{x} and \code{y} are not identical.

\item \code{\%EQ\%}: Evaluates whether \code{x} and \code{y} are setequal. If
\code{x} and \code{y} are not compatible, returns \code{FALSE}.

\item \code{\%!EQ\%}: Evaluates whether \code{x} and \code{y} are not setequal.
If \code{x} and \code{y} are not compatible, returns \code{TRUE}.

\item \code{\%AND\%}: Evaluates whether \code{x} and \code{y} are both
\code{TRUE} (that is, a scalar \code{TRUE} value).

\item \code{\%OR\%}: Evaluates whether \code{x} and/or \code{y} is
\code{TRUE}.

\item \code{\%!\%}: Evaluates whether \code{x} and \code{y} are both
\code{FALSE} (that is, a scalar \code{FALSE} value).

\item \code{\%1\%}: Evaluates whether one of \code{x} or \code{y} is
\code{TRUE}.

\item \code{\%IN\%}: Evaluates whether atomic scalar \code{x} is contained in
atomic object \code{y}. If \code{x} is not atomic scalar, \code{y} is not
atomic, or \code{x} and \code{y} are not compatible, returns \code{FALSE}.

\item \code{\%!IN\%}: Evaluates whether atomic scalar \code{x} is contained in
atomic object \code{y}. If \code{x} is not atomic scalar, \code{y} is not
atomic, or \code{x} and \code{y} are not compatible, returns \code{FALSE}.

\item \code{\%HAS\%}: Evaluates whether atomic object \code{x} contains atomic
scalar \code{y}. If \code{x} is not atomic, \code{y} is not atomic scalar,
or \code{x} and \code{y} are not compatible, returns \code{FALSE}.

\item \code{\%LACKS\%}: Evaluates whether atomic object \code{x} lacks (does not
contain) atomic scalar \code{y}. If \code{x} is not atomic, \code{y} is not
atomic scalar, or \code{x} and \code{y} are not compatible, returns
\code{TRUE}.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
