% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank.}
\alias{blank.}
\alias{blank}
\title{Blank strings}
\usage{
blank.()

blank(x)
}
\arguments{
\item{x}{An object}
}
\value{
Logical scalar
}
\description{
Blank strings
}
\section{Functions}{
\itemize{
\item \code{blank()}: Evaluates whether \code{x} is a blank string scalar.

}}
\seealso{
Other strings: 
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}
}
\concept{strings}
