% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank_uj}
\alias{blank_uj}
\alias{blank}
\title{Blank strings}
\usage{
blank_uj()

blank(x)
}
\arguments{
\item{x}{An object}
}
\value{
Logical scalar
}
\description{
Blank strings
}
\section{Functions}{
\itemize{
\item \code{blank}: Evaluates whether \code{x} is a blank string scalar.
}}

\concept{strings_uj}
