% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{build_markdown}
\alias{build_markdown}
\alias{subs}
\alias{sups}
\alias{bfblu}
\alias{blblu}
\alias{bmblu}
\alias{bdblu}
\alias{bfcyn}
\alias{blcyn}
\alias{bmcyn}
\alias{bdcyn}
\alias{bfgrn}
\alias{blgrn}
\alias{bmgrn}
\alias{bdgrn}
\alias{bfmag}
\alias{blmag}
\alias{bmmag}
\alias{bdmag}
\alias{bforn}
\alias{blorn}
\alias{bmorn}
\alias{bdorn}
\alias{bfred}
\alias{blred}
\alias{bmred}
\alias{bdred}
\alias{bfvlt}
\alias{blvlt}
\alias{bmvlt}
\alias{bdvlt}
\alias{bfylw}
\alias{blylw}
\alias{bmylw}
\alias{bdylw}
\alias{bwht}
\alias{bg95}
\alias{bg90}
\alias{bg85}
\alias{bg80}
\alias{bg75}
\alias{bg70}
\alias{bg65}
\alias{bg60}
\alias{bg55}
\alias{bg50}
\alias{bg45}
\alias{bg40}
\alias{bg35}
\alias{bg30}
\alias{bg25}
\alias{bg20}
\alias{bg15}
\alias{bg10}
\alias{bg05}
\alias{bblk}
\alias{binv}
\alias{pfblu}
\alias{plblu}
\alias{pmblu}
\alias{pdblu}
\alias{pfcyn}
\alias{plcyn}
\alias{pmcyn}
\alias{pdcyn}
\alias{pgrn}
\alias{plgrn}
\alias{pmgrn}
\alias{pdgrn}
\alias{pfmag}
\alias{plmag}
\alias{pmmag}
\alias{pdmag}
\alias{pforn}
\alias{plorn}
\alias{pmorn}
\alias{pdorn}
\alias{pfred}
\alias{plred}
\alias{pmred}
\alias{pdred}
\alias{pfvlt}
\alias{plvlt}
\alias{pmvlt}
\alias{pdvlt}
\alias{pfylw}
\alias{plylw}
\alias{pmylw}
\alias{pdylw}
\alias{pwht}
\alias{pg95}
\alias{pg90}
\alias{pg85}
\alias{pg80}
\alias{pg75}
\alias{pg70}
\alias{pg65}
\alias{pg60}
\alias{pg55}
\alias{pg50}
\alias{pg45}
\alias{pg40}
\alias{pg35}
\alias{pg30}
\alias{pg25}
\alias{pg20}
\alias{pg15}
\alias{pg10}
\alias{pg05}
\alias{pblk}
\alias{pinv}
\title{Build markdown expressions for styled and colored ggtext.}
\usage{
subs(x)

sups(x)

bfblu(x)

blblu(x)

bmblu(x)

bdblu(x)

bfcyn(x)

blcyn(x)

bmcyn(x)

bdcyn(x)

bfgrn(x)

blgrn(x)

bmgrn(x)

bdgrn(x)

bfmag(x)

blmag(x)

bmmag(x)

bdmag(x)

bforn(x)

blorn(x)

bmorn(x)

bdorn(x)

bfred(x)

blred(x)

bmred(x)

bdred(x)

bfvlt(x)

blvlt(x)

bmvlt(x)

bdvlt(x)

bfylw(x)

blylw(x)

bmylw(x)

bdylw(x)

bwht(x)

bg95(x)

bg90(x)

bg85(x)

bg80(x)

bg75(x)

bg70(x)

bg65(x)

bg60(x)

bg55(x)

bg50(x)

bg45(x)

bg40(x)

bg35(x)

bg30(x)

bg25(x)

bg20(x)

bg15(x)

bg10(x)

bg05(x)

bblk(x)

binv(x)

pfblu(x)

plblu(x)

pmblu(x)

pdblu(x)

pfcyn(x)

plcyn(x)

pmcyn(x)

pdcyn(x)

pgrn(x)

plgrn(x)

pmgrn(x)

pdgrn(x)

pfmag(x)

plmag(x)

pmmag(x)

pdmag(x)

pforn(x)

plorn(x)

pmorn(x)

pdorn(x)

pfred(x)

plred(x)

pmred(x)

pdred(x)

pfvlt(x)

plvlt(x)

pmvlt(x)

pdvlt(x)

pfylw(x)

plylw(x)

pmylw(x)

pdylw(x)

pwht(x)

pg95(x)

pg90(x)

pg85(x)

pg80(x)

pg75(x)

pg70(x)

pg65(x)

pg60(x)

pg55(x)

pg50(x)

pg45(x)

pg40(x)

pg35(x)

pg30(x)

pg25(x)

pg20(x)

pg15(x)

pg10(x)

pg05(x)

pblk(x)

pinv(x)
}
\arguments{
\item{x}{Character vector of text to be styled.}
}
\value{
Markdown-styled character vector
}
\description{
build_markdown Style text as subscript.
}
\details{
Style text as subscript, superscript, bold color, and plain
color. The following markdown styling function names do not follow the
conventions described later:
\tabular{lll}{
TEXT STYLING   \tab REPRESENTATION\tab NOTE                           \cr
subscript      \tab \code{subs}   \tab —                              \cr
superscript    \tab \code{subs}   \tab —                              \cr
bold black     \tab \code{bblk}   \tab —                              \cr
bold white     \tab \code{bwht}   \tab —                              \cr
bold invisible \tab \code{binv}   \tab 1/256th opaque white           \cr
plain black    \tab \code{pblk}   \tab —                              \cr
plain white    \tab \code{pwht}   \tab —                              \cr
plain invisible\tab \code{pinv}   \tab 1/256th opaque white             }
For all other markdown styling functions, conventions are as
follows:\tabular{lll}{
CONVENTION   \tab VALUE    \tab TEXT STYLING                          \cr
1st character\tab\code{p}  \tab plain.                                \cr
1st character\tab\code{b}  \tab bold.                                 \cr
2nd character\tab\code{f}  \tab full intensity version of color.      \cr
2nd character\tab\code{l}  \tab light version of color.               \cr
2nd character\tab\code{m}  \tab medium version of color.              \cr
2nd character\tab\code{d}  \tab dark version of color.                \cr
2nd character\tab\code{g}  \tab greyscale.                            \cr
suffix       \tab\code{blu}\tab blue.                                 \cr
suffix       \tab\code{cyn}\tab cyan.                                 \cr
suffix       \tab\code{grn}\tab green.                                \cr
suffix       \tab\code{mag}\tab magenta.                              \cr
suffix       \tab\code{orn}\tab orange.                               \cr
suffix       \tab\code{red}\tab red.                                  \cr
suffix       \tab\code{vlt}\tab violet.                               \cr
suffix       \tab\code{ylw}\tab yellow.                               \cr
suffix       \tab\code{05} \tab 5\% of the way from black to white     \cr
suffix       \tab\code{10} \tab 10\% of the way from black to white    \cr
suffix       \tab\code{15} \tab 15\% of the way from black to white    \cr
suffix       \tab\code{20} \tab 20\% of the way from black to white    \cr
suffix       \tab\code{25} \tab 25\% of the way from black to white    \cr
suffix       \tab\code{30} \tab 30\% of the way from black to white    \cr
suffix       \tab\code{35} \tab 35\% of the way from black to white    \cr
suffix       \tab\code{40} \tab 40\% of the way from black to white    \cr
suffix       \tab\code{45} \tab 45\% of the way from black to white    \cr
suffix       \tab\code{50} \tab 50\% of the way from black to white    \cr
suffix       \tab\code{55} \tab 55\% of the way from black to white    \cr
suffix       \tab\code{60} \tab 60\% of the way from black to white    \cr
suffix       \tab\code{65} \tab 65\% of the way from black to white    \cr
suffix       \tab\code{70} \tab 70\% of the way from black to white    \cr
suffix       \tab\code{75} \tab 75\% of the way from black to white    \cr
suffix       \tab\code{80} \tab 80\% of the way from black to white    \cr
suffix       \tab\code{85} \tab 85\% of the way from black to white    \cr
suffix       \tab\code{90} \tab 90\% of the way from black to white    \cr
suffix       \tab\code{95} \tab 95\% of the way from black to white      }
}
\section{Functions}{
\itemize{
\item \code{sups}: Style text as superscript.

\item \code{bfblu}: Style text as bold and full blue.

\item \code{blblu}: Style text as bold and light blue.

\item \code{bmblu}: Style text as bold and medium blue.

\item \code{bdblu}: Style text as bold and dark blue.

\item \code{bfcyn}: Style text as bold and full cyan.

\item \code{blcyn}: Style text as bold and light cyan.

\item \code{bmcyn}: Style text as bold and medium cyan.

\item \code{bdcyn}: Style text as bold and dark cyan.

\item \code{bfgrn}: Style text as bold and full green.

\item \code{blgrn}: Style text as bold and light green.

\item \code{bmgrn}: Style text as bold and medium green.

\item \code{bdgrn}: Style text as bold and dark green.

\item \code{bfmag}: Style text as bold and full magenta.

\item \code{blmag}: Style text as bold and light magenta.

\item \code{bmmag}: Style text as bold and medium magenta.

\item \code{bdmag}: Style text as bold and dark magenta.

\item \code{bforn}: Style text as bold and full orange.

\item \code{blorn}: Style text as bold and light orange.

\item \code{bmorn}: Style text as bold and medium orange.

\item \code{bdorn}: Style text as bold and dark orange.

\item \code{bfred}: Style text as bold and 10\% grey.

\item \code{blred}: Style text as bold and light red.

\item \code{bmred}: Style text as bold and medium red.

\item \code{bdred}: Style text as bold and dark red.

\item \code{bfvlt}: Style text as bold and full violet.

\item \code{blvlt}: Style text as bold and light violet.

\item \code{bmvlt}: Style text as bold and medium violet.

\item \code{bdvlt}: Style text as bold and dark violet.

\item \code{bfylw}: Style text as bold and 10\% grey.

\item \code{blylw}: Style text as bold and light yellow.

\item \code{bmylw}: Style text as bold and medium yellow.

\item \code{bdylw}: Style text as bold and dark yellow.

\item \code{bwht}: Style text as bold and 10\% grey.

\item \code{bg95}: Style text as bold and 95\% grey.

\item \code{bg90}: Style text as bold and 90\% grey.

\item \code{bg85}: Style text as bold and 85\% grey.

\item \code{bg80}: Style text as bold and 80\% grey.

\item \code{bg75}: Style text as bold and 70\% grey.

\item \code{bg70}: Style text as bold and 70\% grey.

\item \code{bg65}: Style text as bold and 65\% grey.

\item \code{bg60}: Style text as bold and 60\% grey.

\item \code{bg55}: Style text as bold and 55\% grey.

\item \code{bg50}: Style text as bold and 50\% grey.

\item \code{bg45}: Style text as bold and 45\% grey.

\item \code{bg40}: Style text as bold and 40\% grey.

\item \code{bg35}: Style text as bold and 35\% grey.

\item \code{bg30}: Style text as bold and 30\% grey.

\item \code{bg25}: Style text as bold and 25\% grey.

\item \code{bg20}: Style text as bold and 20\% grey.

\item \code{bg15}: Style text as bold and 15\% grey.

\item \code{bg10}: Style text as bold and 10\% grey.

\item \code{bg05}: Style text as bold and 10\% grey.

\item \code{bblk}: Style text as bold and black.

\item \code{binv}: Style text as bold and invisible (1/256th opaque).

\item \code{pfblu}: Style text as plain and full blue.

\item \code{plblu}: Style text as plain and light blue.

\item \code{pmblu}: Style text as plain and medium blue.

\item \code{pdblu}: Style text as plain and dark blue.

\item \code{pfcyn}: Style text as plain and full cyan.

\item \code{plcyn}: Style text as plain and light cyan.

\item \code{pmcyn}: Style text as plain and medium cyan.

\item \code{pdcyn}: Style text as plain and dark cyan.

\item \code{pgrn}: Style text as plain and full green.

\item \code{plgrn}: Style text as plain and light green.

\item \code{pmgrn}: Style text as plain and medium green.

\item \code{pdgrn}: Style text as plain and dark green.

\item \code{pfmag}: Style text as plain and full magenta.

\item \code{plmag}: Style text as plain and light magenta.

\item \code{pmmag}: Style text as plain and medium magenta.

\item \code{pdmag}: Style text as plain and dark magenta.

\item \code{pforn}: Style text as plain and full orange.

\item \code{plorn}: Style text as plain and light orange.

\item \code{pmorn}: Style text as plain and medium orange.

\item \code{pdorn}: Style text as plain and dark orange.

\item \code{pfred}: Style text as plain and 10\% grey.

\item \code{plred}: Style text as plain and light red.

\item \code{pmred}: Style text as plain and medium red.

\item \code{pdred}: Style text as plain and dark red.

\item \code{pfvlt}: Style text as plain and full violet.

\item \code{plvlt}: Style text as plain and light violet.

\item \code{pmvlt}: Style text as plain and medium violet.

\item \code{pdvlt}: Style text as plain and dark violet.

\item \code{pfylw}: Style text as plain and 10\% grey.

\item \code{plylw}: Style text as plain and light yellow.

\item \code{pmylw}: Style text as plain and medium yellow.

\item \code{pdylw}: Style text as plain and dark yellow.

\item \code{pwht}: Style text as plain and 10\% grey.

\item \code{pg95}: Style text as plain and 95\% grey.

\item \code{pg90}: Style text as plain and 90\% grey.

\item \code{pg85}: Style text as plain and 85\% grey.

\item \code{pg80}: Style text as plain and 80\% grey.

\item \code{pg75}: Style text as plain and 70\% grey.

\item \code{pg70}: Style text as plain and 70\% grey.

\item \code{pg65}: Style text as plain and 65\% grey.

\item \code{pg60}: Style text as plain and 60\% grey.

\item \code{pg55}: Style text as plain and 55\% grey.

\item \code{pg50}: Style text as plain and 50\% grey.

\item \code{pg45}: Style text as plain and 45\% grey.

\item \code{pg40}: Style text as plain and 40\% grey.

\item \code{pg35}: Style text as plain and 35\% grey.

\item \code{pg30}: Style text as plain and 30\% grey.

\item \code{pg25}: Style text as plain and 25\% grey.

\item \code{pg20}: Style text as plain and 20\% grey.

\item \code{pg15}: Style text as plain and 15\% grey.

\item \code{pg10}: Style text as plain and 10\% grey.

\item \code{pg05}: Style text as plain and 10\% grey.

\item \code{pblk}: Style text as plain and black.

\item \code{pinv}: Style text as plain invisible (1/256th opaque).
}}

