% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callers.R
\name{callers_uj}
\alias{callers_uj}
\alias{callers}
\alias{caller}
\alias{ncallers}
\title{Functions in the Call Stack}
\usage{
callers_uj()

callers(gens. = NULL)

caller()

ncallers()
}
\arguments{
\item{gens.}{The number of generations back in the function call stack to go.}
}
\value{
\code{callers} and \code{caller} return a character scalar or vector.
\code{ncallers} returns a non-negative whole-number scalar.
}
\description{
Functions in the Call Stack
}
\section{Functions}{
\itemize{
\item \code{callers}: Gets the names of all calling functions, with the
immediate calling function in the first position and, if \code{gens.} is
not \code{NULL}, selects elements from the call stack specified by
\code{gens.}.

\item \code{caller}: Get the immediate calling function.

\item \code{ncallers}: Get the number of calling functions in the call stack.
}}

\examples{
fun.a <- function() {A <- 0; fun.b()}
fun.b <- function() {B <- 1; fun.c()}
fun.c <- function() {
  cat("\n callstack(): " , paste0("'", callstack(), "'"))
  cat("\n caller(1)  : '", caller(1), "'")
  cat("\n caller(2)  : '", caller(2), "'")
  cat("\n exist('A', err = F, gens = 1): ", exist('A', err = F, gens = 1))
  cat("\n exist('A', err = F, gens = 2): ", exist('A', err = F, gens = 2))
  cat("\n exist('B', err = F, gens = 1): ", exist('B', err = F, gens = 1))
  cat("\n exist('B', err = F, gens = 2): ", exist('B', err = F, gens = 2))
  cat("\n vget('A', err = F, gens = 2) : ", vget('A', err = F, gens = 2))
  cat("\n vget('A', err = F, gens = 1) : ", vget('A', err = F, gens = 1))
  cat("\n vget('B', err = F, gens = 2) : ", vget('B', err = F, gens = 2))
  cat("\n vget('B', err = F, gens = 1) : ", vget('B', err = F, gens = 1))
  vSet('A', 'A', gens = 2)
  vSet('B', 'B', gens = 1)
  cat("\n vget('A', gens = 2) : ", vget('A', err = F, gens = 2))
  cat("\n vget('B', gens = 1) : ", vget('B', err = F, gens = 1))
}
fun.a()
}
\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}

Other code: 
\code{\link{expr_uj}},
\code{\link{run_uj}()}
}
\concept{code}
\concept{meta}
