% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\encoding{UTF-8}
\name{cb}
\alias{cb}
\alias{rb}
\title{Error-checked row and column binding}
\usage{
cb(...)

rb(...)
}
\arguments{
\item{...}{Multiple \code{\link{compatible}} atomic matrices or multiple compatible \link[=atm_dtf]{atomic data.frames}.}
}
\value{
An \link[=atm_dtf]{atomic data.frame} or an atomic matrix.
}
\description{
Error-checked extensions of \code{\link[base]{rbind}} and/or \code{\link[base]{cbind}}.
}
\details{
\tabular{ll}{  \code{cb}   \tab Column binds an arbitrary number of atomic matrices or \link[=atm_dtf]{atomic data.frames}. \cr   \tab   \cr
  \code{rb}   \tab Row binds an arbitrary number of atomic matrices or atomic data.frames.                                    }
}
\examples{
egM2x4 <- mat(1:8, nr = 2)
egM3x4 <- mat(1:12, nr = 3)
egM4x4 <- mat(1:16, nr = 4)
egM4x3 <- mat(1:12, nc = 3)
egM4x4b <- egM4x4

cb(egM4x3, egM4x4, egM4x4b)
rb(egM2x4, egM3x4 , egM4x4, egM4x4b)

egD4x2 <- dtf(w = 1:4, X = letters[1:4])
egD4x4 <- dtf(w = 1:4, X = letters[1:4], y = NA, z = '.')
egD2x4 <- dtf(w = 1:2, X = letters[1:2], y = NA, z = '.')
egD4x4b <- egD4x4
egD4x4d <- dtf(W = 1:4, X = letters[1:4], Y = NA, Z = '.')

cb(egD4x2, egD4x4, egD4x4d)
rb(egD2x4, egD4x4, egD4x4b)
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}
}
\concept{extensions}
