% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{char_uj}
\alias{char_uj}
\alias{char}
\title{Extract a single character}
\usage{
char_uj()

char(x, n)
}
\arguments{
\item{x}{Atomic character object.}

\item{n}{Positive whole-number scalar or positive whole-number atomic object
of the same dimensions as \code{x}.}
}
\value{
\code{link[=ch1]{onechar}} atomic object of the same dimension as
\code{x}.
}
\description{
Extract a single character
}
\section{Functions}{
\itemize{
\item \code{char}: Get the \code{n}-th character of each element of
\code{x}.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
