% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{charn}
\alias{charn}
\title{Extract a single character}
\usage{
charn(x, n)
}
\arguments{
\item{x}{A \link[=ichr]{character object}.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} or a \link[=cmp_psw]{complete positive whole-number object} of the same dimensions as \code{x}.}
}
\value{
A link[=cmp_ch1]{complete onechar} atomic object of the same dimension as \code{x}.
}
\description{
Get the \code{n}-th character of each element of \code{x}.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{chars}
\concept{strings}
