% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chn.R
\encoding{UTF-8}
\name{chn}
\alias{chn}
\alias{charn}
\title{Get the \code{N}-th character element-wise.}
\usage{
chn(X, N)

charn(X, N)
}
\arguments{
\item{X}{A \link[=pop_chr]{populated character object}.}

\item{N}{A \link[=cmp_psw_scl]{positive whole-number scalar} or a \link[=cmp_psw]{complete positive whole-number object} of the same dimensions as \code{X}.}
}
\value{
A \link[=cmp_ch1]{complete onechar} atomic object of the same dimension as \code{X}.
}
\description{
Get the \code{N}-th character element-wise.
}
\examples{
words <- stringr::fruit[1:10]
words
chn(words, 1)
chn(words, 3)
charN(words, 10)
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{maxnch}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()}
}
\concept{chars}
\concept{strings}
