% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data_help}
\alias{clean_data_help}
\alias{clean_data}
\alias{clean_select}
\alias{clean_rename}
\alias{clean_recode}
\alias{clean_remode}
\alias{clean_na}
\title{Functions to Interactively Clean Data Frames}
\usage{
clean_data_help()

clean_data(x, clear = TRUE)

clean_select(x, clear = TRUE, .log = F)

clean_rename(x, clear = TRUE)

clean_recode(x, clear = TRUE)

clean_remode(x, clear = TRUE)

clean_na(x, clear = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{clear}{A logical scalar indicating whether to clear the console at the start of each data cleaning step.}
}
\value{
A data frame.
}
\description{
This family of functions clean a data frame.
}
\section{Functions}{
\itemize{
\item \code{clean_data()}: Clean a data frame by selecting variables to retain, optionally renaming variables, optionally recoding \code{NA} and other values, optionally coercing variables to a specific mode, and optionally dropping cases with \code{NA} values.

\item \code{clean_select()}: Select variables from a data frame to be retained.

\item \code{clean_rename()}: Optionally rename variables in a data frame.

\item \code{clean_recode()}: Optionally recode \code{NA} and other values of variables.

\item \code{clean_remode()}: Optionally coerce variables in a data frame to a specific mode.

\item \code{clean_na()}: Optionally drop cases with \code{NA} values.

}}
