% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls.R
\encoding{UTF-8}
\name{cls}
\alias{cls}
\alias{xcls}
\alias{set_cls}
\alias{add_cls}
\alias{has_cls}
\title{Manage classes}
\usage{
cls(x)

xcls(x, ...)

set_cls(x, ...)

add_cls(x, ...)

has_cls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unquoted class names and/or names of variables whose contents are \link[=av]{atomized} and converted to character. When \code{...} args are names of variables in the calling environment, they are atomized, otherwise they are treated as unquoted class names.}
}
\value{
\emph{A character vector} \cr   \code{cls}
\cr\cr \emph{A logical scalar} \cr   \code{has_cls}
\cr\cr \emph{An R object} \cr   \verb{add_cls, set_cls, xcls}
}
\description{
\tabular{rl}{
  \code{has_cls}   \tab Evaluate whether class(es) in \code{...} are (all) classes of \code{x}.
\cr \code{add_cls}   \tab Add class(es) in \code{...} to classes of \code{x}.
\cr \code{set_cls}   \tab Set class of \code{x} to class(es) in \code{...}.
\cr    \code{xcls}   \tab Removes class(es) in \code{...} from classes of \code{x}.
\cr     \code{cls}   \tab Gets class(es) of \code{x}.
}
}
\examples{
obj1. <- tb(letters = letters, numbers = 1:26)
obj2. <- set_cls(obj1., c("new", cls(obj1.)))
obj3. <- set_cls(obj1., "new")
cls(obj1.)
cls(obj2.)
cls(obj3.)
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
