% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\name{cmp_mmm.}
\alias{cmp_mmm.}
\alias{cmp_mmm_vals}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{Complete + extended mode properties}
\usage{
cmp_mmm.()

cmp_mmm_vals()

cmp_atm(x)

cmp_ch1(x)

cmp_chr(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE} except for \code{cmp_mmm_vals}, which
returns a character vector.
}
\description{
Check with an object is complete (\code{\link{icmp}}) and of a
specific extended mode \code{\link{mmm}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_mmm_vals()}: Get a character vector of all possible complete +
extended mode values.

\item \code{cmp_atm()}: Is \code{x} a complete atomic object?
(\code{\link{icmp}}, \code{link{iatm}})

\item \code{cmp_ch1()}: Is \code{x} a complete onechar object?
(\code{\link{icmp}}, \code{link{ich1}})

\item \code{cmp_chr()}: Is \code{x} a complete character object?
(\code{\link{icmp}}, \code{link{ichr}})

\item \code{cmp_clr()}: Is \code{x} a complete color object?
(\code{\link{icmp}}, \code{link{iclr}})

\item \code{cmp_evn()}: Is \code{x} a complete even-valued whole-number
object? (\code{\link{icmp}}, \code{link{ievn}})

\item \code{cmp_fac()}: Is \code{x} a complete factor object?
(\code{\link{icmp}}, \code{link{ifac}})

\item \code{cmp_frc()}: Is \code{x} a complete fractional-valued numeric
object? (\code{\link{icmp}}, \code{link{ifrc}})

\item \code{cmp_ind()}: Is \code{x} a complete indexer object?
(\code{\link{icmp}}, \code{link{iind}})

\item \code{cmp_lgl()}: Is \code{x} a complete logical object?
(\code{\link{icmp}}, \code{link{ilgl}})

\item \code{cmp_neg()}: Is \code{x} a complete negative-valued numeric object?
(\code{\link{icmp}}, \code{link{ineg}})

\item \code{cmp_ngw()}: Is \code{x} a complete negative-valued whole-number
object? (\code{\link{icmp}}, \code{link{ingw}})

\item \code{cmp_nng()}: Is \code{x} a complete non-negative-valued
numeric object? (\code{\link{icmp}}, \code{link{inng}})

\item \code{cmp_nnw()}: Is \code{x} a complete non-negative-valued
whole-number object? (\code{\link{icmp}}, \code{link{innw}})

\item \code{cmp_nps()}: Is \code{x} a complete non-positive-valued
numeric object? (\code{\link{icmp}}, \code{link{inps}})

\item \code{cmp_npw()}: Is \code{x} a complete non-positive-valued
whole-number object? (\code{\link{icmp}}, \code{link{inpw}})

\item \code{cmp_nst()}: Is \code{x} a complete non-sortable object?
(\code{\link{icmp}}, \code{link{inst}})

\item \code{cmp_num()}: Is \code{x} a complete numeric object?
(\code{\link{icmp}}, \code{link{inum}})

\item \code{cmp_odd()}: Is \code{x} a complete odd-valued whole-number object?
(\code{\link{icmp}}, \code{link{iodd}})

\item \code{cmp_ord()}: Is \code{x} a complete ordered factor object?
(\code{\link{icmp}}, \code{link{iord}})

\item \code{cmp_pct()}: Is \code{x} a complete percent-valued (0-100) numeric
object? (\code{\link{icmp}}, \code{link{ipct}})

\item \code{cmp_pos()}: Is \code{x} a complete positive-valued numeric object?
(\code{\link{icmp}}, \code{link{ipos}})

\item \code{cmp_ppn()}: Is \code{x} a complete proportion-valued (0-1) numeric
object? (\code{\link{icmp}}, \code{link{ippn}})

\item \code{cmp_psw()}: Is \code{x} a complete positive-valued whole-number
object? (\code{\link{icmp}}, \code{link{ipsw}})

\item \code{cmp_srt()}: Is \code{x} a complete sortable object?
(\code{\link{icmp}}, \code{link{isrt}})

\item \code{cmp_str()}: Is \code{x} a complete string object?
(\code{\link{icmp}}, \code{link{istr}})

\item \code{cmp_uno()}: Is \code{x} a complete unordered factor object?
(\code{\link{icmp}}, \code{link{iuno}})

\item \code{cmp_whl()}: Is \code{x} a complete whole-number object?
(\code{\link{icmp}}, \code{link{iwhl}})

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()},
\code{\link{xxx.}()}
}
\concept{props}
