% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\encoding{UTF-8}
\name{cmp_mmm}
\alias{cmp_mmm}
\alias{cmp_mmm_funs}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_ch3}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{complete + Xmode combination properties}
\usage{
cmp_mmm(X, MMM, ...)

cmp_mmm_funs()

cmp_atm(X, ...)

cmp_ch1(X, ...)

cmp_ch3(X, ...)

cmp_chr(X, ...)

cmp_clr(X, ...)

cmp_evn(X, ...)

cmp_fac(X, ...)

cmp_frc(X, ...)

cmp_ind(X, ...)

cmp_lgl(X, ...)

cmp_neg(X, ...)

cmp_ngw(X, ...)

cmp_nng(X, ...)

cmp_nnw(X, ...)

cmp_nps(X, ...)

cmp_npw(X, ...)

cmp_nst(X, ...)

cmp_num(X, ...)

cmp_odd(X, ...)

cmp_ord(X, ...)

cmp_pct(X, ...)

cmp_pos(X, ...)

cmp_pos(X, ...)

cmp_psw(X, ...)

cmp_srt(X, ...)

cmp_str(X, ...)

cmp_uno(X, ...)

cmp_whl(X, ...)
}
\arguments{
\item{X}{An R object.}

\item{MMM}{A character scalar Xmode property from \code{mmm_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \code{cmp_mmm_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{cmp_\{mmm\}, cmp_mmm}
}
\description{
Check for combination of \link[=CMP]{completeness} and \link[=mmm]{Xmode}.
}
\details{
\tabular{ll}{  \code{cmp_mmm_funs}   \tab What \link[=CMP]{complete} + \link[=mmm]{Xmode} combination \link[=prop_funs]{property functions} are there?                         \cr   \tab   \cr
  \verb{cmp_\{mmm\}}      \tab Is \code{X} both complete and a match to the single Xmode property \code{'{mmm}'} where \code{{mmm}} is a placeholder for any given Xmode property? \cr   \tab   \cr
  \code{cmp_mmm}        \tab Is \code{X} both complete and a match to single Xmode property in arg \code{mmm}?                                                                             }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.max, .maxr, .maxc}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
cmp_mmm_funs()
cmp_mmm(letters, "ch1")
cmp_mmm(c("abc", "def"), "ch3")
cmp_psw(0:10)
cmp_ord(factor(letters, ordered = T))
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp_fast}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
