% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_arr.R
\name{cmp_mmm_arr}
\alias{cmp_mmm_arr}
\alias{cmp_mmm_arr_vals}
\alias{cmp_atm_arr}
\alias{cmp_chr_arr}
\alias{cmp_ch1_arr}
\alias{cmp_clr_arr}
\alias{cmp_evn_arr}
\alias{cmp_fac_arr}
\alias{cmp_frc_arr}
\alias{cmp_ind_arr}
\alias{cmp_lgl_arr}
\alias{cmp_neg_arr}
\alias{cmp_ngw_arr}
\alias{cmp_nng_arr}
\alias{cmp_nnw_arr}
\alias{cmp_nps_arr}
\alias{cmp_npw_arr}
\alias{cmp_nst_arr}
\alias{cmp_num_arr}
\alias{cmp_odd_arr}
\alias{cmp_ord_arr}
\alias{cmp_pct_arr}
\alias{cmp_pos_arr}
\alias{cmp_ppn_arr}
\alias{cmp_psw_arr}
\alias{cmp_srt_arr}
\alias{cmp_str_arr}
\alias{cmp_uno_arr}
\alias{cmp_whl_arr}
\title{Complete + Extended Mode + Atomic Array (cmp + mmm + arr)}
\usage{
cmp_mmm_arr_vals()

cmp_atm_arr(x)

cmp_chr_arr(x)

cmp_ch1_arr(x)

cmp_clr_arr(x)

cmp_evn_arr(x)

cmp_fac_arr(x)

cmp_frc_arr(x)

cmp_ind_arr(x)

cmp_lgl_arr(x)

cmp_neg_arr(x)

cmp_ngw_arr(x)

cmp_nng_arr(x)

cmp_nnw_arr(x)

cmp_nps_arr(x)

cmp_npw_arr(x)

cmp_nst_arr(x)

cmp_num_arr(x)

cmp_odd_arr(x)

cmp_ord_arr(x)

cmp_pct_arr(x)

cmp_pos_arr(x)

cmp_ppn_arr(x)

cmp_psw_arr(x)

cmp_srt_arr(x)

cmp_str_arr(x)

cmp_uno_arr(x)

cmp_whl_arr(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible complete + extended mode + array properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_arr}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_arr}: Is \code{x} a complete atomic array?

\item \code{cmp_chr_arr}: Is \code{x} a complete character atomic array?

\item \code{cmp_ch1_arr}: Is \code{x} a complete onechar atomic array?

\item \code{cmp_clr_arr}: Is \code{x} a complete color atomic array?

\item \code{cmp_evn_arr}: Is \code{x} a complete even-numeric atomic array?

\item \code{cmp_fac_arr}: Is \code{x} a complete factor atomic array?

\item \code{cmp_frc_arr}: Is \code{x} a complete fractional-numeric atomic
array?

\item \code{cmp_ind_arr}: Is \code{x} a complete indexer atomic array?

\item \code{cmp_lgl_arr}: Is \code{x} a complete logical atomic array?

\item \code{cmp_neg_arr}: Is \code{x} a complete negative-numeric atomic array?

\item \code{cmp_ngw_arr}: Is \code{x} a complete negative-whole-numeric atomic
array?

\item \code{cmp_nng_arr}: Is \code{x} a complete non-negative-numeric atomic
array?

\item \code{cmp_nnw_arr}: Is \code{x} a complete non-negative-whole-numeric
atomic array?

\item \code{cmp_nps_arr}: Is \code{x} a complete non-positive-numeric atomic
array?

\item \code{cmp_npw_arr}: Is \code{x} a complete non-positive-whole-numeric
atomic array?

\item \code{cmp_nst_arr}: Is \code{x} a complete non-sortable atomic array?

\item \code{cmp_num_arr}: Is \code{x} a complete numeric atomic array?

\item \code{cmp_odd_arr}: Is \code{x} a complete odd-numeric atomic array?

\item \code{cmp_ord_arr}: Is \code{x} a complete ordered-factor atomic array?

\item \code{cmp_pct_arr}: Is \code{x} a complete percent-numeric atomic array?

\item \code{cmp_pos_arr}: Is \code{x} a complete positive-numeric atomic array?

\item \code{cmp_ppn_arr}: Is \code{x} a complete proportion-numeric atomic
array?

\item \code{cmp_psw_arr}: Is \code{x} a complete positive-whole-numeric atomic
array?

\item \code{cmp_srt_arr}: Is \code{x} a complete sortable atomic array?

\item \code{cmp_str_arr}: Is \code{x} a complete string atomic array?

\item \code{cmp_uno_arr}: Is \code{x} a complete unordered-factor atomic array?

\item \code{cmp_whl_arr}: Is \code{x} a complete whole-numeric atomic array?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
