% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_atb.R
\name{cmp_mmm_atb}
\alias{cmp_mmm_atb}
\alias{cmp_mmm_atb_vals}
\alias{cmp_atm_atb}
\alias{cmp_chr_atb}
\alias{cmp_ch1_atb}
\alias{cmp_clr_atb}
\alias{cmp_evn_atb}
\alias{cmp_fac_atb}
\alias{cmp_frc_atb}
\alias{cmp_ind_atb}
\alias{cmp_lgl_atb}
\alias{cmp_neg_atb}
\alias{cmp_ngw_atb}
\alias{cmp_nng_atb}
\alias{cmp_nnw_atb}
\alias{cmp_nps_atb}
\alias{cmp_npw_atb}
\alias{cmp_nst_atb}
\alias{cmp_num_atb}
\alias{cmp_odd_atb}
\alias{cmp_ord_atb}
\alias{cmp_pct_atb}
\alias{cmp_pos_atb}
\alias{cmp_ppn_atb}
\alias{cmp_psw_atb}
\alias{cmp_srt_atb}
\alias{cmp_str_atb}
\alias{cmp_uno_atb}
\alias{cmp_whl_atb}
\title{Complete + Extended Mode + Atomic Tibble (cmp + mmm + atb)}
\usage{
cmp_mmm_atb_vals()

cmp_atm_atb(x)

cmp_chr_atb(x)

cmp_ch1_atb(x)

cmp_clr_atb(x)

cmp_evn_atb(x)

cmp_fac_atb(x)

cmp_frc_atb(x)

cmp_ind_atb(x)

cmp_lgl_atb(x)

cmp_neg_atb(x)

cmp_ngw_atb(x)

cmp_nng_atb(x)

cmp_nnw_atb(x)

cmp_nps_atb(x)

cmp_npw_atb(x)

cmp_nst_atb(x)

cmp_num_atb(x)

cmp_odd_atb(x)

cmp_ord_atb(x)

cmp_pct_atb(x)

cmp_pos_atb(x)

cmp_ppn_atb(x)

cmp_psw_atb(x)

cmp_srt_atb(x)

cmp_str_atb(x)

cmp_uno_atb(x)

cmp_whl_atb(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic tibble
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{is_cmp_tbl}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_atb}: Is \code{x} a complete atomic tibble?

\item \code{cmp_chr_atb}: Is \code{x} a complete character atomic tibble?

\item \code{cmp_ch1_atb}: Is \code{x} a complete onechar atomic tibble?

\item \code{cmp_clr_atb}: Is \code{x} a complete color atomic tibble?

\item \code{cmp_evn_atb}: Is \code{x} a complete even-numeric atomic tibble?

\item \code{cmp_fac_atb}: Is \code{x} a complete factor atomic tibble?

\item \code{cmp_frc_atb}: Is \code{x} a complete fractional-numeric atomic
tibble?

\item \code{cmp_ind_atb}: Is \code{x} a complete indexer atomic tibble?

\item \code{cmp_lgl_atb}: Is \code{x} a complete logical atomic tibble?

\item \code{cmp_neg_atb}: Is \code{x} a complete negative-numeric atomic
tibble?

\item \code{cmp_ngw_atb}: Is \code{x} a complete negative-whole-numeric atomic
tibble?

\item \code{cmp_nng_atb}: Is \code{x} a complete non-negative-numeric atomic
tibble?

\item \code{cmp_nnw_atb}: Is \code{x} a complete non-negative-numeric atomic
tibble?

\item \code{cmp_nps_atb}: Is \code{x} a complete non-positive-numeric atomic
tibble?

\item \code{cmp_npw_atb}: Is \code{x} a complete non-positive-whole-numeric
atomic tibble?

\item \code{cmp_nst_atb}: Is \code{x} a complete non-sortable atomic tibble?

\item \code{cmp_num_atb}: Is \code{x} a complete numeric atomic tibble?

\item \code{cmp_odd_atb}: Is \code{x} a complete odd-numeric atomic tibble?

\item \code{cmp_ord_atb}: Is \code{x} a complete ordered-factor atomic tibble?

\item \code{cmp_pct_atb}: Is \code{x} a complete percent-numeric atomic tibble?

\item \code{cmp_pos_atb}: Is \code{x} a complete positive-numeric atomic
tibble?

\item \code{cmp_ppn_atb}: Is \code{x} a complete proportion-numeric atomic
tibble?

\item \code{cmp_psw_atb}: Is \code{x} a complete positive-whole-numeric atomic
tibble?

\item \code{cmp_srt_atb}: Is \code{x} a complete sortable atomic tibble?

\item \code{cmp_str_atb}: Is \code{x} a complete string atomic tibble?

\item \code{cmp_uno_atb}: Is \code{x} a complete unordered-factor atomic
tibble?

\item \code{cmp_whl_atb}: Is \code{x} a complete whole-numeric atomic tibble?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
