% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_avl.R
\name{cmp_mmm_avl}
\alias{cmp_mmm_avl}
\alias{cmp_mmm_avl_vals}
\alias{cmp_atm_avl}
\alias{cmp_chr_avl}
\alias{cmp_ch1_avl}
\alias{cmp_clr_avl}
\alias{cmp_evn_avl}
\alias{cmp_fac_avl}
\alias{cmp_frc_avl}
\alias{cmp_ind_avl}
\alias{cmp_lgc_avl}
\alias{cmp_cnb_avl}
\alias{cmp_neg_avl}
\alias{cmp_ngw_avl}
\alias{cmp_nng_avl}
\alias{cmp_nnw_avl}
\alias{cmp_nps_avl}
\alias{cmp_npw_avl}
\alias{cmp_nst_avl}
\alias{cmp_num_avl}
\alias{cmp_odd_avl}
\alias{cmp_ord_avl}
\alias{cmp_pct_avl}
\alias{cmp_pos_avl}
\alias{cmp_ppn_avl}
\alias{cmp_psw_avl}
\alias{cmp_srt_avl}
\alias{cmp_str_avl}
\alias{cmp_uno_avl}
\alias{cmp_whl_avl}
\title{Complete + Extended Mode + Atomic Vlist (cmp + mmm + avl)}
\usage{
cmp_mmm_avl_vals()

cmp_atm_avl(x)

cmp_chr_avl(x)

cmp_ch1_avl(x)

cmp_clr_avl(x)

cmp_evn_avl(x)

cmp_fac_avl(x)

cmp_frc_avl(x)

cmp_ind_avl(x)

cmp_lgc_avl(x)

cmp_cnb_avl(x)

cmp_neg_avl(x)

cmp_ngw_avl(x)

cmp_nng_avl(x)

cmp_nnw_avl(x)

cmp_nps_avl(x)

cmp_npw_avl(x)

cmp_nst_avl(x)

cmp_num_avl(x)

cmp_odd_avl(x)

cmp_ord_avl(x)

cmp_pct_avl(x)

cmp_pos_avl(x)

cmp_ppn_avl(x)

cmp_psw_avl(x)

cmp_srt_avl(x)

cmp_str_avl(x)

cmp_uno_avl(x)

cmp_whl_avl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
See \code{\link{mmm}}, and \code{\link{is_cmp_vlist}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
