% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_avt.R
\name{cmp_mmm_avt}
\alias{cmp_mmm_avt}
\alias{cmp_mmm_avt_vals}
\alias{cmp_atm_avt}
\alias{cmp_chr_avt}
\alias{cmp_ch1_avt}
\alias{cmp_clr_avt}
\alias{cmp_evn_avt}
\alias{cmp_fac_avt}
\alias{cmp_frc_avt}
\alias{cmp_ind_avt}
\alias{cmp_lgc_avt}
\alias{cmp_cnb_avt}
\alias{cmp_neg_avt}
\alias{cmp_ngw_avt}
\alias{cmp_nng_avt}
\alias{cmp_nnw_avt}
\alias{cmp_nps_avt}
\alias{cmp_npw_avt}
\alias{cmp_nst_avt}
\alias{cmp_num_avt}
\alias{cmp_odd_avt}
\alias{cmp_ord_avt}
\alias{cmp_pct_avt}
\alias{cmp_pos_avt}
\alias{cmp_ppn_avt}
\alias{cmp_psw_avt}
\alias{cmp_srt_avt}
\alias{cmp_str_avt}
\alias{cmp_uno_avt}
\alias{cmp_whl_avt}
\title{Complete + Extended Mode + Atomic Vtype (cmp + mmm + avt)}
\usage{
cmp_mmm_avt_vals()

cmp_atm_avt(x)

cmp_chr_avt(x)

cmp_ch1_avt(x)

cmp_clr_avt(x)

cmp_evn_avt(x)

cmp_fac_avt(x)

cmp_frc_avt(x)

cmp_ind_avt(x)

cmp_lgc_avt(x)

cmp_cnb_avt(x)

cmp_neg_avt(x)

cmp_ngw_avt(x)

cmp_nng_avt(x)

cmp_nnw_avt(x)

cmp_nps_avt(x)

cmp_npw_avt(x)

cmp_nst_avt(x)

cmp_num_avt(x)

cmp_odd_avt(x)

cmp_ord_avt(x)

cmp_pct_avt(x)

cmp_pos_avt(x)

cmp_ppn_avt(x)

cmp_psw_avt(x)

cmp_srt_avt(x)

cmp_str_avt(x)

cmp_uno_avt(x)

cmp_whl_avt(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
See \code{\link{mmm}}, and \code{\link{is_cmp_vtype}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
