% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_ccc.R
\name{cmp_mmm_ccc.}
\alias{cmp_mmm_ccc.}
\alias{cmp_mmm_ccc_vals}
\alias{cmp_ch1_arr}
\alias{cmp_ch1_gen}
\alias{cmp_ch1_dtf}
\alias{cmp_ch1_vls}
\alias{cmp_ch1_mat}
\alias{cmp_ch1_mvc}
\alias{cmp_ch1_scl}
\alias{cmp_ch1_vec}
\alias{cmp_chr_arr}
\alias{cmp_chr_gen}
\alias{cmp_chr_dtf}
\alias{cmp_chr_vls}
\alias{cmp_chr_mat}
\alias{cmp_chr_mvc}
\alias{cmp_chr_scl}
\alias{cmp_chr_vec}
\alias{cmp_clr_arr}
\alias{cmp_clr_gen}
\alias{cmp_clr_dtf}
\alias{cmp_clr_vls}
\alias{cmp_clr_mat}
\alias{cmp_clr_mvc}
\alias{cmp_clr_scl}
\alias{cmp_clr_vec}
\alias{cmp_evn_arr}
\alias{cmp_evn_gen}
\alias{cmp_evn_dtf}
\alias{cmp_evn_vls}
\alias{cmp_evn_mat}
\alias{cmp_evn_mvc}
\alias{cmp_evn_scl}
\alias{cmp_evn_vec}
\alias{cmp_fac_arr}
\alias{cmp_fac_gen}
\alias{cmp_fac_dtf}
\alias{cmp_fac_vls}
\alias{cmp_fac_mat}
\alias{cmp_fac_mvc}
\alias{cmp_fac_scl}
\alias{cmp_fac_vec}
\alias{cmp_frc_arr}
\alias{cmp_frc_gen}
\alias{cmp_frc_dtf}
\alias{cmp_frc_vls}
\alias{cmp_frc_mat}
\alias{cmp_frc_mvc}
\alias{cmp_frc_scl}
\alias{cmp_frc_vec}
\alias{cmp_ind_arr}
\alias{cmp_ind_gen}
\alias{cmp_ind_dtf}
\alias{cmp_ind_vls}
\alias{cmp_ind_mat}
\alias{cmp_ind_mvc}
\alias{cmp_ind_scl}
\alias{cmp_ind_vec}
\alias{cmp_lgl_arr}
\alias{cmp_lgl_gen}
\alias{cmp_lgl_dtf}
\alias{cmp_lgl_vls}
\alias{cmp_lgl_mat}
\alias{cmp_lgl_mvc}
\alias{cmp_lgl_scl}
\alias{cmp_lgl_vec}
\alias{cmp_neg_arr}
\alias{cmp_neg_gen}
\alias{cmp_neg_dtf}
\alias{cmp_neg_vls}
\alias{cmp_neg_mat}
\alias{cmp_neg_mvc}
\alias{cmp_neg_scl}
\alias{cmp_neg_vec}
\alias{cmp_ngw_arr}
\alias{cmp_ngw_gen}
\alias{cmp_ngw_dtf}
\alias{cmp_ngw_vls}
\alias{cmp_ngw_mat}
\alias{cmp_ngw_mvc}
\alias{cmp_ngw_scl}
\alias{cmp_ngw_vec}
\alias{cmp_nng_arr}
\alias{cmp_nng_gen}
\alias{cmp_nng_dtf}
\alias{cmp_nng_vls}
\alias{cmp_nng_mat}
\alias{cmp_nng_mvc}
\alias{cmp_nng_scl}
\alias{cmp_nng_vec}
\alias{cmp_nnw_arr}
\alias{cmp_nnw_gen}
\alias{cmp_nnw_dtf}
\alias{cmp_nnw_vls}
\alias{cmp_nnw_mat}
\alias{cmp_nnw_mvc}
\alias{cmp_nnw_scl}
\alias{cmp_nnw_vec}
\alias{cmp_nps_arr}
\alias{cmp_nps_gen}
\alias{cmp_nps_dtf}
\alias{cmp_nps_vls}
\alias{cmp_nps_mat}
\alias{cmp_nps_mvc}
\alias{cmp_nps_scl}
\alias{cmp_nps_vec}
\alias{cmp_npw_arr}
\alias{cmp_npw_gen}
\alias{cmp_npw_dtf}
\alias{cmp_npw_vls}
\alias{cmp_npw_mat}
\alias{cmp_npw_mvc}
\alias{cmp_npw_scl}
\alias{cmp_npw_vec}
\alias{cmp_nst_arr}
\alias{cmp_nst_gen}
\alias{cmp_nst_dtf}
\alias{cmp_nst_vls}
\alias{cmp_nst_mat}
\alias{cmp_nst_mvc}
\alias{cmp_nst_scl}
\alias{cmp_nst_vec}
\alias{cmp_num_arr}
\alias{cmp_num_gen}
\alias{cmp_num_dtf}
\alias{cmp_num_vls}
\alias{cmp_num_mat}
\alias{cmp_num_mvc}
\alias{cmp_num_scl}
\alias{cmp_num_vec}
\alias{cmp_odd_arr}
\alias{cmp_odd_gen}
\alias{cmp_odd_dtf}
\alias{cmp_odd_vls}
\alias{cmp_odd_mat}
\alias{cmp_odd_mvc}
\alias{cmp_odd_scl}
\alias{cmp_odd_vec}
\alias{cmp_ord_arr}
\alias{cmp_ord_gen}
\alias{cmp_ord_dtf}
\alias{cmp_ord_vls}
\alias{cmp_ord_mat}
\alias{cmp_ord_mvc}
\alias{cmp_ord_scl}
\alias{cmp_ord_vec}
\alias{cmp_pct_arr}
\alias{cmp_pct_gen}
\alias{cmp_pct_dtf}
\alias{cmp_pct_vls}
\alias{cmp_pct_mat}
\alias{cmp_pct_mvc}
\alias{cmp_pct_scl}
\alias{cmp_pct_vec}
\alias{cmp_pos_arr}
\alias{cmp_pos_gen}
\alias{cmp_pos_dtf}
\alias{cmp_pos_vls}
\alias{cmp_pos_mat}
\alias{cmp_pos_mvc}
\alias{cmp_pos_scl}
\alias{cmp_pos_vec}
\alias{cmp_ppn_arr}
\alias{cmp_ppn_gen}
\alias{cmp_ppn_dtf}
\alias{cmp_ppn_vls}
\alias{cmp_ppn_mat}
\alias{cmp_ppn_mvc}
\alias{cmp_ppn_scl}
\alias{cmp_ppn_vec}
\alias{cmp_psw_arr}
\alias{cmp_psw_gen}
\alias{cmp_psw_dtf}
\alias{cmp_psw_vls}
\alias{cmp_psw_mat}
\alias{cmp_psw_mvc}
\alias{cmp_psw_scl}
\alias{cmp_psw_vec}
\alias{cmp_srt_arr}
\alias{cmp_srt_gen}
\alias{cmp_srt_dtf}
\alias{cmp_srt_vls}
\alias{cmp_srt_mat}
\alias{cmp_srt_mvc}
\alias{cmp_srt_scl}
\alias{cmp_srt_vec}
\alias{cmp_str_arr}
\alias{cmp_str_gen}
\alias{cmp_str_dtf}
\alias{cmp_str_vls}
\alias{cmp_str_mat}
\alias{cmp_str_mvc}
\alias{cmp_str_scl}
\alias{cmp_str_vec}
\alias{cmp_uno_arr}
\alias{cmp_uno_gen}
\alias{cmp_uno_dtf}
\alias{cmp_uno_vls}
\alias{cmp_uno_mat}
\alias{cmp_uno_mvc}
\alias{cmp_uno_scl}
\alias{cmp_uno_vec}
\alias{cmp_whl_arr}
\alias{cmp_whl_gen}
\alias{cmp_whl_dtf}
\alias{cmp_whl_vls}
\alias{cmp_whl_mat}
\alias{cmp_whl_mvc}
\alias{cmp_whl_scl}
\alias{cmp_whl_vec}
\title{Completeness plus extended mode plus extended class properties}
\usage{
cmp_mmm_ccc.()

cmp_mmm_ccc_vals()

cmp_ch1_arr(x)

cmp_ch1_gen(x)

cmp_ch1_dtf(x)

cmp_ch1_vls(x)

cmp_ch1_mat(x)

cmp_ch1_mvc(x)

cmp_ch1_scl(x)

cmp_ch1_vec(x)

cmp_chr_arr(x)

cmp_chr_gen(x)

cmp_chr_dtf(x)

cmp_chr_vls(x)

cmp_chr_mat(x)

cmp_chr_mvc(x)

cmp_chr_scl(x)

cmp_chr_vec(x)

cmp_clr_arr(x)

cmp_clr_gen(x)

cmp_clr_dtf(x)

cmp_clr_vls(x)

cmp_clr_mat(x)

cmp_clr_mvc(x)

cmp_clr_scl(x)

cmp_clr_vec(x)

cmp_evn_arr(x)

cmp_evn_gen(x)

cmp_evn_dtf(x)

cmp_evn_vls(x)

cmp_evn_mat(x)

cmp_evn_mvc(x)

cmp_evn_scl(x)

cmp_evn_vec(x)

cmp_fac_arr(x)

cmp_fac_gen(x)

cmp_fac_dtf(x)

cmp_fac_vls(x)

cmp_fac_mat(x)

cmp_fac_mvc(x)

cmp_fac_scl(x)

cmp_fac_vec(x)

cmp_frc_arr(x)

cmp_frc_gen(x)

cmp_frc_dtf(x)

cmp_frc_vls(x)

cmp_frc_mat(x)

cmp_frc_mvc(x)

cmp_frc_scl(x)

cmp_frc_vec(x)

cmp_ind_arr(x)

cmp_ind_gen(x)

cmp_ind_dtf(x)

cmp_ind_vls(x)

cmp_ind_mat(x)

cmp_ind_mvc(x)

cmp_ind_scl(x)

cmp_ind_vec(x)

cmp_lgl_arr(x)

cmp_lgl_gen(x)

cmp_lgl_dtf(x)

cmp_lgl_vls(x)

cmp_lgl_mat(x)

cmp_lgl_mvc(x)

cmp_lgl_scl(x)

cmp_lgl_vec(x)

cmp_neg_arr(x)

cmp_neg_gen(x)

cmp_neg_dtf(x)

cmp_neg_vls(x)

cmp_neg_mat(x)

cmp_neg_mvc(x)

cmp_neg_scl(x)

cmp_neg_vec(x)

cmp_ngw_arr(x)

cmp_ngw_gen(x)

cmp_ngw_dtf(x)

cmp_ngw_vls(x)

cmp_ngw_mat(x)

cmp_ngw_mvc(x)

cmp_ngw_scl(x)

cmp_ngw_vec(x)

cmp_nng_arr(x)

cmp_nng_gen(x)

cmp_nng_dtf(x)

cmp_nng_vls(x)

cmp_nng_mat(x)

cmp_nng_mvc(x)

cmp_nng_scl(x)

cmp_nng_vec(x)

cmp_nnw_arr(x)

cmp_nnw_gen(x)

cmp_nnw_dtf(x)

cmp_nnw_vls(x)

cmp_nnw_mat(x)

cmp_nnw_mvc(x)

cmp_nnw_scl(x)

cmp_nnw_vec(x)

cmp_nps_arr(x)

cmp_nps_gen(x)

cmp_nps_dtf(x)

cmp_nps_vls(x)

cmp_nps_mat(x)

cmp_nps_mvc(x)

cmp_nps_scl(x)

cmp_nps_vec(x)

cmp_npw_arr(x)

cmp_npw_gen(x)

cmp_npw_dtf(x)

cmp_npw_vls(x)

cmp_npw_mat(x)

cmp_npw_mvc(x)

cmp_npw_scl(x)

cmp_npw_vec(x)

cmp_nst_arr(x)

cmp_nst_gen(x)

cmp_nst_dtf(x)

cmp_nst_vls(x)

cmp_nst_mat(x)

cmp_nst_mvc(x)

cmp_nst_scl(x)

cmp_nst_vec(x)

cmp_num_arr(x)

cmp_num_gen(x)

cmp_num_dtf(x)

cmp_num_vls(x)

cmp_num_mat(x)

cmp_num_mvc(x)

cmp_num_scl(x)

cmp_num_vec(x)

cmp_odd_arr(x)

cmp_odd_gen(x)

cmp_odd_dtf(x)

cmp_odd_vls(x)

cmp_odd_mat(x)

cmp_odd_mvc(x)

cmp_odd_scl(x)

cmp_odd_vec(x)

cmp_ord_arr(x)

cmp_ord_gen(x)

cmp_ord_dtf(x)

cmp_ord_vls(x)

cmp_ord_mat(x)

cmp_ord_mvc(x)

cmp_ord_scl(x)

cmp_ord_vec(x)

cmp_pct_arr(x)

cmp_pct_gen(x)

cmp_pct_dtf(x)

cmp_pct_vls(x)

cmp_pct_mat(x)

cmp_pct_mvc(x)

cmp_pct_scl(x)

cmp_pct_vec(x)

cmp_pos_arr(x)

cmp_pos_gen(x)

cmp_pos_dtf(x)

cmp_pos_vls(x)

cmp_pos_mat(x)

cmp_pos_mvc(x)

cmp_pos_scl(x)

cmp_pos_vec(x)

cmp_ppn_arr(x)

cmp_ppn_gen(x)

cmp_ppn_dtf(x)

cmp_ppn_vls(x)

cmp_ppn_mat(x)

cmp_ppn_mvc(x)

cmp_ppn_scl(x)

cmp_ppn_vec(x)

cmp_psw_arr(x)

cmp_psw_gen(x)

cmp_psw_dtf(x)

cmp_psw_vls(x)

cmp_psw_mat(x)

cmp_psw_mvc(x)

cmp_psw_scl(x)

cmp_psw_vec(x)

cmp_srt_arr(x)

cmp_srt_gen(x)

cmp_srt_dtf(x)

cmp_srt_vls(x)

cmp_srt_mat(x)

cmp_srt_mvc(x)

cmp_srt_scl(x)

cmp_srt_vec(x)

cmp_str_arr(x)

cmp_str_gen(x)

cmp_str_dtf(x)

cmp_str_vls(x)

cmp_str_mat(x)

cmp_str_mvc(x)

cmp_str_scl(x)

cmp_str_vec(x)

cmp_uno_arr(x)

cmp_uno_gen(x)

cmp_uno_dtf(x)

cmp_uno_vls(x)

cmp_uno_mat(x)

cmp_uno_mvc(x)

cmp_uno_scl(x)

cmp_uno_vec(x)

cmp_whl_arr(x)

cmp_whl_gen(x)

cmp_whl_dtf(x)

cmp_whl_vls(x)

cmp_whl_mat(x)

cmp_whl_mvc(x)

cmp_whl_scl(x)

cmp_whl_vec(x)
}
\arguments{
\item{x}{An ℝ object.}
}
\value{
\code{TRUE} or \code{FALSE} except for \code{cmp_mmm_ccc_vals}, which
returns a character vector.
}
\description{
Combinations of \link[icmp]{completeness}, \link[mmm]{extended
mode}, and \link[ccc]{extended class}.
}
\section{Functions}{
\itemize{
\item \code{cmp_mmm_ccc_vals()}: Get a character vector of all possible completeness
plus extended mode plus extended class properties.

\item \code{cmp_ch1_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[iarr]{array+}?

\item \code{cmp_ch1_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[igen]{generic}?

\item \code{cmp_ch1_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[idtf]{data.frame}?

\item \code{cmp_ch1_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[ivls]{vlist}?

\item \code{cmp_ch1_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[imat]{matrix}?

\item \code{cmp_ch1_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[imvc]{multivec}?

\item \code{cmp_ch1_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[iscl]{scalar}?

\item \code{cmp_ch1_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ich1]{onechar} \link[ivec]{vector+}?

\item \code{cmp_chr_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[iarr]{array+}?

\item \code{cmp_chr_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[igen]{generic}?

\item \code{cmp_chr_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[idtf]{data.frame}?

\item \code{cmp_chr_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[ivls]{vlist}?

\item \code{cmp_chr_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[imat]{matrix}?

\item \code{cmp_chr_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[imvc]{multivec}?

\item \code{cmp_chr_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[iscl]{scalar}?

\item \code{cmp_chr_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} \link[ivec]{vector+}?

\item \code{cmp_clr_arr()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[iarr]{array+}?

\item \code{cmp_clr_gen()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[igen]{generic}?

\item \code{cmp_clr_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[idtf]{data.frame}?

\item \code{cmp_clr_vls()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[ivls]{vlist}?

\item \code{cmp_clr_mat()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[imat]{matrix}?

\item \code{cmp_clr_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[imvc]{multivec}?

\item \code{cmp_clr_scl()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[iscl]{scalar}?

\item \code{cmp_clr_vec()}: Is \code{x} a \link[icmp]{complete}
\link[iclr]{color} \link[ivec]{vector+}?

\item \code{cmp_evn_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[iarr]{array+}?

\item \code{cmp_evn_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[igen]{generic}?

\item \code{cmp_evn_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[idtf]{data.frame}?

\item \code{cmp_evn_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[ivls]{vlist}?

\item \code{cmp_evn_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[imat]{matrix}?

\item \code{cmp_evn_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[imvc]{multivec}?

\item \code{cmp_evn_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[iscl]{scalar}?

\item \code{cmp_evn_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} \link[ivec]{vector+}?

\item \code{cmp_fac_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[iarr]{array+}?

\item \code{cmp_fac_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[igen]{generic}?

\item \code{cmp_fac_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[idtf]{data.frame}?

\item \code{cmp_fac_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[ivls]{vlist}?

\item \code{cmp_fac_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[imat]{matrix}?

\item \code{cmp_fac_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[imvc]{multivec}?

\item \code{cmp_fac_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[iscl]{scalar}?

\item \code{cmp_fac_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ifac]{factor} \link[ivec]{vector+}?

\item \code{cmp_frc_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[iarr]{array+}?

\item \code{cmp_frc_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[igen]{generic}?

\item \code{cmp_frc_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[idtf]{data.frame}?

\item \code{cmp_frc_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[ivls]{vlist}?

\item \code{cmp_frc_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[imat]{matrix}?

\item \code{cmp_frc_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[imvc]{multivec}?

\item \code{cmp_frc_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[iscl]{scalar}?

\item \code{cmp_frc_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} \link[ivec]{vector+}?

\item \code{cmp_ind_arr()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[iarr]{array+}?

\item \code{cmp_ind_gen()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[igen]{generic}?

\item \code{cmp_ind_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[idtf]{data.frame}?

\item \code{cmp_ind_vls()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[ivls]{vlist}?

\item \code{cmp_ind_mat()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[imat]{matrix}?

\item \code{cmp_ind_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[imvc]{multivec}?

\item \code{cmp_ind_scl()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[iscl]{scalar}?

\item \code{cmp_ind_vec()}: Is \code{x} a \link[icmp]{complete}
\link[iind]{indexer} \link[ivec]{vector+}?

\item \code{cmp_lgl_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[iarr]{array+}?

\item \code{cmp_lgl_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[igen]{generic}?

\item \code{cmp_lgl_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[idtf]{data.frame}?

\item \code{cmp_lgl_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[ivls]{vlist}?

\item \code{cmp_lgl_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[imat]{matrix}?

\item \code{cmp_lgl_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[imvc]{multivec}?

\item \code{cmp_lgl_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[iscl]{scalar}?

\item \code{cmp_lgl_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ilgl]{logical} \link[ivec]{vector+}?

\item \code{cmp_neg_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[iarr]{array+}?

\item \code{cmp_neg_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[igen]{generic}?

\item \code{cmp_neg_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[idtf]{data.frame}?

\item \code{cmp_neg_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[ivls]{vlist}?

\item \code{cmp_neg_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[imat]{matrix}?

\item \code{cmp_neg_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[imvc]{multivec}?

\item \code{cmp_neg_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[iscl]{scalar}?

\item \code{cmp_neg_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} \link[ivec]{vector+}?

\item \code{cmp_ngw_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[iarr]{array+}?

\item \code{cmp_ngw_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[igen]{generic}?

\item \code{cmp_ngw_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[idtf]{data.frame}?

\item \code{cmp_ngw_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[ivls]{vlist}?

\item \code{cmp_ngw_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[imat]{matrix}?

\item \code{cmp_ngw_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[imvc]{multivec}?

\item \code{cmp_ngw_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[iscl]{scalar}?

\item \code{cmp_ngw_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} \link[ivec]{vector+}?

\item \code{cmp_nng_arr()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[iarr]{array+}?

\item \code{cmp_nng_gen()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[igen]{generic}?

\item \code{cmp_nng_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[idtf]{data.frame}?

\item \code{cmp_nng_vls()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[ivls]{vlist}?

\item \code{cmp_nng_mat()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[imat]{matrix}?

\item \code{cmp_nng_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[imvc]{multivec}?

\item \code{cmp_nng_scl()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[iscl]{scalar}?

\item \code{cmp_nng_vec()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} \link[ivec]{vector+}?

\item \code{cmp_nnw_arr()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[iarr]{array+}?

\item \code{cmp_nnw_gen()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[igen]{generic}?

\item \code{cmp_nnw_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[idtf]{data.frame}?

\item \code{cmp_nnw_vls()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[ivls]{vlist}?

\item \code{cmp_nnw_mat()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[imat]{matrix}?

\item \code{cmp_nnw_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[imvc]{multivec}?

\item \code{cmp_nnw_scl()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[iscl]{scalar}?

\item \code{cmp_nnw_vec()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} \link[ivec]{vector+}?

\item \code{cmp_nps_arr()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[iarr]{array+}?

\item \code{cmp_nps_gen()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[igen]{generic}?

\item \code{cmp_nps_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[idtf]{data.frame}?

\item \code{cmp_nps_vls()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[ivls]{vlist}?

\item \code{cmp_nps_mat()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[imat]{matrix}?

\item \code{cmp_nps_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[imvc]{multivec}?

\item \code{cmp_nps_scl()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[iscl]{scalar}?

\item \code{cmp_nps_vec()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} \link[ivec]{vector+}?

\item \code{cmp_npw_arr()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[iarr]{array+}?

\item \code{cmp_npw_gen()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[igen]{generic}?

\item \code{cmp_npw_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[idtf]{data.frame}?

\item \code{cmp_npw_vls()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[ivls]{vlist}?

\item \code{cmp_npw_mat()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[imat]{matrix}?

\item \code{cmp_npw_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[imvc]{multivec}?

\item \code{cmp_npw_scl()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[iscl]{scalar}?

\item \code{cmp_npw_vec()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} \link[ivec]{vector+}?

\item \code{cmp_nst_arr()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[iarr]{array+}?

\item \code{cmp_nst_gen()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[igen]{generic}?

\item \code{cmp_nst_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[idtf]{data.frame}?

\item \code{cmp_nst_vls()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[ivls]{vlist}?

\item \code{cmp_nst_mat()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[imat]{matrix}?

\item \code{cmp_nst_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[imvc]{multivec}?

\item \code{cmp_nst_scl()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[iscl]{scalar}?

\item \code{cmp_nst_vec()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} \link[ivec]{vector+}?

\item \code{cmp_num_arr()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[iarr]{array+}?

\item \code{cmp_num_gen()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[igen]{generic}?

\item \code{cmp_num_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[idtf]{data.frame}?

\item \code{cmp_num_vls()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[ivls]{vlist}?

\item \code{cmp_num_mat()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[imat]{matrix}?

\item \code{cmp_num_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[imvc]{multivec}?

\item \code{cmp_num_scl()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[iscl]{scalar}?

\item \code{cmp_num_vec()}: Is \code{x} a \link[icmp]{complete}
\link[inum]{numeric} \link[ivec]{vector+}?

\item \code{cmp_odd_arr()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[iarr]{array+}?

\item \code{cmp_odd_gen()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[igen]{generic}?

\item \code{cmp_odd_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[idtf]{data.frame}?

\item \code{cmp_odd_vls()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[ivls]{vlist}?

\item \code{cmp_odd_mat()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[imat]{matrix}?

\item \code{cmp_odd_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[imvc]{multivec}?

\item \code{cmp_odd_scl()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[iscl]{scalar}?

\item \code{cmp_odd_vec()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} \link[ivec]{vector+}?

\item \code{cmp_ord_arr()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[iarr]{array+}?

\item \code{cmp_ord_gen()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[igen]{generic}?

\item \code{cmp_ord_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[idtf]{data.frame}?

\item \code{cmp_ord_vls()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[ivls]{vlist}?

\item \code{cmp_ord_mat()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[imat]{matrix}?

\item \code{cmp_ord_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[imvc]{multivec}?

\item \code{cmp_ord_scl()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[iscl]{scalar}?

\item \code{cmp_ord_vec()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} \link[ivec]{vector+}?

\item \code{cmp_pct_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[iarr]{array+}?

\item \code{cmp_pct_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[igen]{generic}?

\item \code{cmp_pct_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[idtf]{data.frame}?

\item \code{cmp_pct_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[ivls]{vlist}?

\item \code{cmp_pct_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[imat]{matrix}?

\item \code{cmp_pct_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[imvc]{multivec}?

\item \code{cmp_pct_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[iscl]{scalar}?

\item \code{cmp_pct_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} \link[ivec]{vector+}?

\item \code{cmp_pos_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[iarr]{array+}?

\item \code{cmp_pos_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[igen]{generic}?

\item \code{cmp_pos_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[idtf]{data.frame}?

\item \code{cmp_pos_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[ivls]{vlist}?

\item \code{cmp_pos_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[imat]{matrix}?

\item \code{cmp_pos_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[imvc]{multivec}?

\item \code{cmp_pos_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[iscl]{scalar}?

\item \code{cmp_pos_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} \link[ivec]{vector+}?

\item \code{cmp_ppn_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[iarr]{array+}?

\item \code{cmp_ppn_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[igen]{generic}?

\item \code{cmp_ppn_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[idtf]{data.frame}?

\item \code{cmp_ppn_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[ivls]{vlist}?

\item \code{cmp_ppn_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[imat]{matrix}?

\item \code{cmp_ppn_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[imvc]{multivec}?

\item \code{cmp_ppn_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[iscl]{scalar}?

\item \code{cmp_ppn_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} \link[ivec]{vector+}?

\item \code{cmp_psw_arr()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[iarr]{array+}?

\item \code{cmp_psw_gen()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[igen]{generic}?

\item \code{cmp_psw_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[idtf]{data.frame}?

\item \code{cmp_psw_vls()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[ivls]{vlist}?

\item \code{cmp_psw_mat()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[imat]{matrix}?

\item \code{cmp_psw_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[imvc]{multivec}?

\item \code{cmp_psw_scl()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[iscl]{scalar}?

\item \code{cmp_psw_vec()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} \link[ivec]{vector+}?

\item \code{cmp_srt_arr()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[iarr]{array+}?

\item \code{cmp_srt_gen()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[igen]{generic}?

\item \code{cmp_srt_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[idtf]{data.frame}?

\item \code{cmp_srt_vls()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[ivls]{vlist}?

\item \code{cmp_srt_mat()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[imat]{matrix}?

\item \code{cmp_srt_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[imvc]{multivec}?

\item \code{cmp_srt_scl()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[iscl]{scalar}?

\item \code{cmp_srt_vec()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} \link[ivec]{vector+}?

\item \code{cmp_str_arr()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[iarr]{array+}?

\item \code{cmp_str_gen()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[igen]{generic}?

\item \code{cmp_str_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[idtf]{data.frame}?

\item \code{cmp_str_vls()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[ivls]{vlist}?

\item \code{cmp_str_mat()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[imat]{matrix}?

\item \code{cmp_str_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[imvc]{multivec}?

\item \code{cmp_str_scl()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[iscl]{scalar}?

\item \code{cmp_str_vec()}: Is \code{x} a \link[icmp]{complete}
\link[istr]{string} \link[ivec]{vector+}?

\item \code{cmp_uno_arr()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[iarr]{array+}?

\item \code{cmp_uno_gen()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[igen]{generic}?

\item \code{cmp_uno_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[idtf]{data.frame}?

\item \code{cmp_uno_vls()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[ivls]{vlist}?

\item \code{cmp_uno_mat()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[imat]{matrix}?

\item \code{cmp_uno_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[imvc]{multivec}?

\item \code{cmp_uno_scl()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[iscl]{scalar}?

\item \code{cmp_uno_vec()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} \link[ivec]{vector+}?

\item \code{cmp_whl_arr()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[iarr]{array+}?

\item \code{cmp_whl_gen()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[igen]{generic}?

\item \code{cmp_whl_dtf()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[idtf]{data.frame}?

\item \code{cmp_whl_vls()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[ivls]{vlist}?

\item \code{cmp_whl_mat()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[imat]{matrix}?

\item \code{cmp_whl_mvc()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[imvc]{multivec}?

\item \code{cmp_whl_scl()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[iscl]{scalar}?

\item \code{cmp_whl_vec()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} \link[ivec]{vector+}?

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
