% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_ccc.R
\encoding{UTF-8}
\name{cmp_mmm_ccc}
\alias{cmp_mmm_ccc}
\alias{cmp_mmm_ccc_funs}
\alias{cmp_atm_arr}
\alias{cmp_atm_dtf}
\alias{cmp_atm_gen}
\alias{cmp_atm_mat}
\alias{cmp_atm_mvc}
\alias{cmp_atm_scl}
\alias{cmp_atm_vec}
\alias{cmp_atm_vls}
\alias{cmp_ch1_arr}
\alias{cmp_ch1_dtf}
\alias{cmp_ch1_gen}
\alias{cmp_ch1_mat}
\alias{cmp_ch1_mvc}
\alias{cmp_ch1_scl}
\alias{cmp_ch1_vec}
\alias{cmp_ch1_vls}
\alias{cmp_ch3_arr}
\alias{cmp_ch3_dtf}
\alias{cmp_ch3_gen}
\alias{cmp_ch3_mat}
\alias{cmp_ch3_mvc}
\alias{cmp_ch3_scl}
\alias{cmp_ch3_vec}
\alias{cmp_ch3_vls}
\alias{cmp_chr_arr}
\alias{cmp_chr_dtf}
\alias{cmp_chr_gen}
\alias{cmp_chr_mat}
\alias{cmp_chr_mvc}
\alias{cmp_chr_scl}
\alias{cmp_chr_vec}
\alias{cmp_chr_vls}
\alias{cmp_clr_arr}
\alias{cmp_clr_dtf}
\alias{cmp_clr_gen}
\alias{cmp_clr_mat}
\alias{cmp_clr_mvc}
\alias{cmp_clr_scl}
\alias{cmp_clr_vec}
\alias{cmp_clr_vls}
\alias{cmp_evn_arr}
\alias{cmp_evn_dtf}
\alias{cmp_evn_gen}
\alias{cmp_evn_mat}
\alias{cmp_evn_mvc}
\alias{cmp_evn_scl}
\alias{cmp_evn_vec}
\alias{cmp_evn_vls}
\alias{cmp_fac_arr}
\alias{cmp_fac_dtf}
\alias{cmp_fac_gen}
\alias{cmp_fac_mat}
\alias{cmp_fac_mvc}
\alias{cmp_fac_scl}
\alias{cmp_fac_vec}
\alias{cmp_fac_vls}
\alias{cmp_frc_arr}
\alias{cmp_frc_dtf}
\alias{cmp_frc_gen}
\alias{cmp_frc_mat}
\alias{cmp_frc_mvc}
\alias{cmp_frc_scl}
\alias{cmp_frc_vec}
\alias{cmp_frc_vls}
\alias{cmp_ind_arr}
\alias{cmp_ind_dtf}
\alias{cmp_ind_gen}
\alias{cmp_ind_mat}
\alias{cmp_ind_mvc}
\alias{cmp_ind_scl}
\alias{cmp_ind_vec}
\alias{cmp_ind_vls}
\alias{cmp_lgl_arr}
\alias{cmp_lgl_dtf}
\alias{cmp_lgl_gen}
\alias{cmp_lgl_mat}
\alias{cmp_lgl_mvc}
\alias{cmp_lgl_scl}
\alias{cmp_lgl_vec}
\alias{cmp_lgl_vls}
\alias{cmp_neg_arr}
\alias{cmp_neg_dtf}
\alias{cmp_neg_gen}
\alias{cmp_neg_mat}
\alias{cmp_neg_mvc}
\alias{cmp_neg_scl}
\alias{cmp_neg_vec}
\alias{cmp_neg_vls}
\alias{cmp_ngw_arr}
\alias{cmp_ngw_dtf}
\alias{cmp_ngw_gen}
\alias{cmp_ngw_mat}
\alias{cmp_ngw_mvc}
\alias{cmp_ngw_scl}
\alias{cmp_ngw_vec}
\alias{cmp_ngw_vls}
\alias{cmp_nng_arr}
\alias{cmp_nng_dtf}
\alias{cmp_nng_gen}
\alias{cmp_nng_mat}
\alias{cmp_nng_mvc}
\alias{cmp_nng_scl}
\alias{cmp_nng_vec}
\alias{cmp_nng_vls}
\alias{cmp_nnw_arr}
\alias{cmp_nnw_dtf}
\alias{cmp_nnw_gen}
\alias{cmp_nnw_mat}
\alias{cmp_nnw_mvc}
\alias{cmp_nnw_scl}
\alias{cmp_nnw_vec}
\alias{cmp_nnw_vls}
\alias{cmp_nps_arr}
\alias{cmp_nps_dtf}
\alias{cmp_nps_gen}
\alias{cmp_nps_mat}
\alias{cmp_nps_mvc}
\alias{cmp_nps_scl}
\alias{cmp_nps_vec}
\alias{cmp_nps_vls}
\alias{cmp_npw_arr}
\alias{cmp_npw_dtf}
\alias{cmp_npw_gen}
\alias{cmp_npw_mat}
\alias{cmp_npw_mvc}
\alias{cmp_npw_scl}
\alias{cmp_npw_vec}
\alias{cmp_npw_vls}
\alias{cmp_nst_arr}
\alias{cmp_nst_dtf}
\alias{cmp_nst_gen}
\alias{cmp_nst_mat}
\alias{cmp_nst_mvc}
\alias{cmp_nst_scl}
\alias{cmp_nst_vec}
\alias{cmp_nst_vls}
\alias{cmp_num_arr}
\alias{cmp_num_dtf}
\alias{cmp_num_gen}
\alias{cmp_num_mat}
\alias{cmp_num_mvc}
\alias{cmp_num_scl}
\alias{cmp_num_vec}
\alias{cmp_num_vls}
\alias{cmp_odd_arr}
\alias{cmp_odd_dtf}
\alias{cmp_odd_gen}
\alias{cmp_odd_mat}
\alias{cmp_odd_mvc}
\alias{cmp_odd_scl}
\alias{cmp_odd_vec}
\alias{cmp_odd_vls}
\alias{cmp_ord_arr}
\alias{cmp_ord_dtf}
\alias{cmp_ord_gen}
\alias{cmp_ord_mat}
\alias{cmp_ord_mvc}
\alias{cmp_ord_scl}
\alias{cmp_ord_vec}
\alias{cmp_ord_vls}
\alias{cmp_pct_arr}
\alias{cmp_pct_dtf}
\alias{cmp_pct_gen}
\alias{cmp_pct_mat}
\alias{cmp_pct_mvc}
\alias{cmp_pct_scl}
\alias{cmp_pct_vec}
\alias{cmp_pct_vls}
\alias{cmp_pos_arr}
\alias{cmp_pos_dtf}
\alias{cmp_pos_gen}
\alias{cmp_pos_mat}
\alias{cmp_pos_mvc}
\alias{cmp_pos_scl}
\alias{cmp_pos_vec}
\alias{cmp_pos_vls}
\alias{cmp_ppn_arr}
\alias{cmp_ppn_dtf}
\alias{cmp_ppn_gen}
\alias{cmp_ppn_mat}
\alias{cmp_ppn_mvc}
\alias{cmp_ppn_scl}
\alias{cmp_ppn_vec}
\alias{cmp_ppn_vls}
\alias{cmp_psw_arr}
\alias{cmp_psw_dtf}
\alias{cmp_psw_gen}
\alias{cmp_psw_mat}
\alias{cmp_psw_mvc}
\alias{cmp_psw_scl}
\alias{cmp_psw_vec}
\alias{cmp_psw_vls}
\alias{cmpSRTarr}
\alias{cmpSRTdtf}
\alias{cmpSRTgen}
\alias{cmpSRTmat}
\alias{cmpSRTmvc}
\alias{cmp_srt_scl}
\alias{cmpSRTvec}
\alias{cmpSRTvls}
\alias{cmp_str_arr}
\alias{cmp_str_dtf}
\alias{cmp_str_gen}
\alias{cmp_str_mat}
\alias{cmp_str_mvc}
\alias{cmp_str_scl}
\alias{cmp_str_vec}
\alias{cmp_str_vls}
\alias{cmp_uno_arr}
\alias{cmp_uno_dtf}
\alias{cmp_uno_gen}
\alias{cmp_uno_mat}
\alias{cmp_uno_mvc}
\alias{cmp_uno_scl}
\alias{cmp_uno_vec}
\alias{cmp_uno_vls}
\alias{cmp_whl_arr}
\alias{cmp_whl_dtf}
\alias{cmp_whl_gen}
\alias{cmp_whl_mat}
\alias{cmp_whl_mvc}
\alias{cmp_whl_scl}
\alias{cmp_whl_vec}
\alias{cmp_whl_vls}
\title{Completeness + xmode + xclass combination properties}
\usage{
cmp_mmm_ccc(X, MMM, CCC, ...)

cmp_mmm_ccc_funs()

cmp_atm_arr(X, ...)

cmp_atm_dtf(X, ...)

cmp_atm_gen(X, ...)

cmp_atm_mat(X, ...)

cmp_atm_mvc(X, ...)

cmp_atm_scl(X, ...)

cmp_atm_vec(X, ...)

cmp_atm_vls(X, ...)

cmp_ch1_arr(X, ...)

cmp_ch1_dtf(X, ...)

cmp_ch1_gen(X, ...)

cmp_ch1_mat(X, ...)

cmp_ch1_mvc(X, ...)

cmp_ch1_scl(X, ...)

cmp_ch1_vec(X, ...)

cmp_ch1_vls(X, ...)

cmp_ch3_arr(X, ...)

cmp_ch3_dtf(X, ...)

cmp_ch3_gen(X, ...)

cmp_ch3_mat(X, ...)

cmp_ch3_mvc(X, ...)

cmp_ch3_scl(X, ...)

cmp_ch3_vec(X, ...)

cmp_ch3_vls(X, ...)

cmp_chr_arr(X, ...)

cmp_chr_dtf(X, ...)

cmp_chr_gen(X, ...)

cmp_chr_mat(X, ...)

cmp_chr_mvc(X, ...)

cmp_chr_scl(X, ...)

cmp_chr_vec(X, ...)

cmp_chr_vls(X, ...)

cmp_clr_arr(X, ...)

cmp_clr_dtf(X, ...)

cmp_clr_gen(X, ...)

cmp_clr_mat(X, ...)

cmp_clr_mvc(X, ...)

cmp_clr_scl(X, ...)

cmp_clr_vec(X, ...)

cmp_clr_vls(X, ...)

cmp_evn_arr(X, ...)

cmp_evn_dtf(X, ...)

cmp_evn_gen(X, ...)

cmp_evn_mat(X, ...)

cmp_evn_mvc(X, ...)

cmp_evn_scl(X, ...)

cmp_evn_vec(X, ...)

cmp_evn_vls(X, ...)

cmp_fac_arr(X, ...)

cmp_fac_dtf(X, ...)

cmp_fac_gen(X, ...)

cmp_fac_mat(X, ...)

cmp_fac_mvc(X, ...)

cmp_fac_scl(X, ...)

cmp_fac_vec(X, ...)

cmp_fac_vls(X, ...)

cmp_frc_arr(X, ...)

cmp_frc_dtf(X, ...)

cmp_frc_gen(X, ...)

cmp_frc_mat(X, ...)

cmp_frc_mvc(X, ...)

cmp_frc_scl(X, ...)

cmp_frc_vec(X, ...)

cmp_frc_vls(X, ...)

cmp_ind_arr(X, ...)

cmp_ind_dtf(X, ...)

cmp_ind_gen(X, ...)

cmp_ind_mat(X, ...)

cmp_ind_mvc(X, ...)

cmp_ind_scl(X, ...)

cmp_ind_vec(X, ...)

cmp_ind_vls(X, ...)

cmp_lgl_arr(X, ...)

cmp_lgl_dtf(X, ...)

cmp_lgl_gen(X, ...)

cmp_lgl_mat(X, ...)

cmp_lgl_mvc(X, ...)

cmp_lgl_scl(X, ...)

cmp_lgl_vec(X, ...)

cmp_lgl_vls(X, ...)

cmp_neg_arr(X, ...)

cmp_neg_dtf(X, ...)

cmp_neg_gen(X, ...)

cmp_neg_mat(X, ...)

cmp_neg_mvc(X, ...)

cmp_neg_scl(X, ...)

cmp_neg_vec(X, ...)

cmp_neg_vls(X, ...)

cmp_ngw_arr(X, ...)

cmp_ngw_dtf(X, ...)

cmp_ngw_gen(X, ...)

cmp_ngw_mat(X, ...)

cmp_ngw_mvc(X, ...)

cmp_ngw_scl(X, ...)

cmp_ngw_vec(X, ...)

cmp_ngw_vls(X, ...)

cmp_nng_arr(X, ...)

cmp_nng_dtf(X, ...)

cmp_nng_gen(X, ...)

cmp_nng_mat(X, ...)

cmp_nng_mvc(X, ...)

cmp_nng_scl(X, ...)

cmp_nng_vec(X, ...)

cmp_nng_vls(X, ...)

cmp_nnw_arr(X, ...)

cmp_nnw_dtf(X, ...)

cmp_nnw_gen(X, ...)

cmp_nnw_mat(X, ...)

cmp_nnw_mvc(X, ...)

cmp_nnw_scl(X, ...)

cmp_nnw_vec(X, ...)

cmp_nnw_vls(X, ...)

cmp_nps_arr(X, ...)

cmp_nps_dtf(X, ...)

cmp_nps_gen(X, ...)

cmp_nps_mat(X, ...)

cmp_nps_mvc(X, ...)

cmp_nps_scl(X, ...)

cmp_nps_vec(X, ...)

cmp_nps_vls(X, ...)

cmp_npw_arr(X, ...)

cmp_npw_dtf(X, ...)

cmp_npw_gen(X, ...)

cmp_npw_mat(X, ...)

cmp_npw_mvc(X, ...)

cmp_npw_scl(X, ...)

cmp_npw_vec(X, ...)

cmp_npw_vls(X, ...)

cmp_nst_arr(X, ...)

cmp_nst_dtf(X, ...)

cmp_nst_gen(X, ...)

cmp_nst_mat(X, ...)

cmp_nst_mvc(X, ...)

cmp_nst_scl(X, ...)

cmp_nst_vec(X, ...)

cmp_nst_vls(X, ...)

cmp_num_arr(X, ...)

cmp_num_dtf(X, ...)

cmp_num_gen(X, ...)

cmp_num_mat(X, ...)

cmp_num_mvc(X, ...)

cmp_num_scl(X, ...)

cmp_num_vec(X, ...)

cmp_num_vls(X, ...)

cmp_odd_arr(X, ...)

cmp_odd_dtf(X, ...)

cmp_odd_gen(X, ...)

cmp_odd_mat(X, ...)

cmp_odd_mvc(X, ...)

cmp_odd_scl(X, ...)

cmp_odd_vec(X, ...)

cmp_odd_vls(X, ...)

cmp_ord_arr(X, ...)

cmp_ord_dtf(X, ...)

cmp_ord_gen(X, ...)

cmp_ord_mat(X, ...)

cmp_ord_mvc(X, ...)

cmp_ord_scl(X, ...)

cmp_ord_vec(X, ...)

cmp_ord_vls(X, ...)

cmp_pct_arr(X, ...)

cmp_pct_dtf(X, ...)

cmp_pct_gen(X, ...)

cmp_pct_mat(X, ...)

cmp_pct_mvc(X, ...)

cmp_pct_scl(X, ...)

cmp_pct_vec(X, ...)

cmp_pct_vls(X, ...)

cmp_pos_arr(X, ...)

cmp_pos_dtf(X, ...)

cmp_pos_gen(X, ...)

cmp_pos_mat(X, ...)

cmp_pos_mvc(X, ...)

cmp_pos_scl(X, ...)

cmp_pos_vec(X, ...)

cmp_pos_vls(X, ...)

cmp_ppn_arr(X, ...)

cmp_ppn_dtf(X, ...)

cmp_ppn_gen(X, ...)

cmp_ppn_mat(X, ...)

cmp_ppn_mvc(X, ...)

cmp_ppn_scl(X, ...)

cmp_ppn_vec(X, ...)

cmp_ppn_vls(X, ...)

cmp_psw_arr(X, ...)

cmp_psw_dtf(X, ...)

cmp_psw_gen(X, ...)

cmp_psw_mat(X, ...)

cmp_psw_mvc(X, ...)

cmp_psw_scl(X, ...)

cmp_psw_vec(X, ...)

cmp_psw_vls(X, ...)

cmpSRTarr(X, ...)

cmpSRTdtf(X, ...)

cmpSRTgen(X, ...)

cmpSRTmat(X, ...)

cmpSRTmvc(X, ...)

cmp_srt_scl(X, ...)

cmpSRTvec(X, ...)

cmpSRTvls(X, ...)

cmp_str_arr(X, ...)

cmp_str_dtf(X, ...)

cmp_str_gen(X, ...)

cmp_str_mat(X, ...)

cmp_str_mvc(X, ...)

cmp_str_scl(X, ...)

cmp_str_vec(X, ...)

cmp_str_vls(X, ...)

cmp_uno_arr(X, ...)

cmp_uno_dtf(X, ...)

cmp_uno_gen(X, ...)

cmp_uno_mat(X, ...)

cmp_uno_mvc(X, ...)

cmp_uno_scl(X, ...)

cmp_uno_vec(X, ...)

cmp_uno_vls(X, ...)

cmp_whl_arr(X, ...)

cmp_whl_dtf(X, ...)

cmp_whl_gen(X, ...)

cmp_whl_mat(X, ...)

cmp_whl_mvc(X, ...)

cmp_whl_scl(X, ...)

cmp_whl_vec(X, ...)

cmp_whl_vls(X, ...)
}
\arguments{
\item{X}{An R object.}

\item{CCC}{A character scalar single xclass property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}

\item{mmm}{A character scalar single xmode property from \code{mmm_props()}.}

\item{prop}{A character scalar.}
}
\value{
\strong{A character scalar} \cr\cr \code{cmp_mmm_ccc_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{cmp_\{mmm\}_\{ccc\}, cmp_mmm_ccc}
}
\description{
Check for combinations of \link[=CMP]{completeness}, \link[=mmm]{xmode}, and \link[=ccc]{xclass}.
}
\details{
\tabular{ll}{  \code{cmp_mmm_ccc_funs}   \tab What complete + xmode + xclass combination property functions are there?                                                                                                                                                          \cr   \tab   \cr
  \verb{cmp_\{mmm\}_\{ccc\}}    \tab Is \code{X} both complete and a match to the single xmode property \code{'{mmm}'} and single xclass property \code{'{ccc}'} where \code{{mmm}} and \code{{ccc}} are placeholders for any given xmode property and any given xclass property, respectively. \cr   \tab   \cr
  \code{cmp_mmm_ccc}        \tab Is \code{X} both complete and a match to the single xmode and xclass properties in \code{MMM} and \code{CCC}, respectively?                                                                                                                                     }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.max, .maxr, .maxc}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
cmp_mmm_cccFUNS()
cmp_mmm_ccc(letters, "ch1", "vec")
cmp_mmm_ccc(letters, "str", "scl")
cmp_ch1_vec(letters)
cmp_str_scl("a")
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{ccc}()},
\code{\link{sss_ccc}()},
\code{\link{unq_mmm_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
