% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_ccc.R
\name{cmp_mmm_ccc_uj}
\alias{cmp_mmm_ccc_uj}
\alias{cmp_mmm_ccc_vals}
\alias{cmp_ch1_arr}
\alias{cmp_ch1_gen}
\alias{cmp_ch1_tab}
\alias{cmp_ch1_vls}
\alias{cmp_ch1_vtp}
\alias{cmp_ch1_mat}
\alias{cmp_ch1_mvc}
\alias{cmp_ch1_scl}
\alias{cmp_ch1_vec}
\alias{cmp_chr_arr}
\alias{cmp_chr_gen}
\alias{cmp_chr_tab}
\alias{cmp_chr_vls}
\alias{cmp_chr_vtp}
\alias{cmp_chr_mat}
\alias{cmp_chr_mvc}
\alias{cmp_chr_scl}
\alias{cmp_chr_vec}
\alias{cmp_clr_arr}
\alias{cmp_clr_gen}
\alias{cmp_clr_tab}
\alias{cmp_clr_vls}
\alias{cmp_clr_vtp}
\alias{cmp_clr_mat}
\alias{cmp_clr_mvc}
\alias{cmp_clr_scl}
\alias{cmp_clr_vec}
\alias{cmp_evn_arr}
\alias{cmp_evn_gen}
\alias{cmp_evn_tab}
\alias{cmp_evn_vls}
\alias{cmp_evn_vtp}
\alias{cmp_evn_mat}
\alias{cmp_evn_mvc}
\alias{cmp_evn_scl}
\alias{cmp_evn_vec}
\alias{cmp_fac_arr}
\alias{cmp_fac_gen}
\alias{cmp_fac_tab}
\alias{cmp_fac_vls}
\alias{cmp_fac_vtp}
\alias{cmp_fac_mat}
\alias{cmp_fac_mvc}
\alias{cmp_fac_scl}
\alias{cmp_fac_vec}
\alias{cmp_frc_arr}
\alias{cmp_frc_gen}
\alias{cmp_frc_tab}
\alias{cmp_frc_vls}
\alias{cmp_frc_vtp}
\alias{cmp_frc_mat}
\alias{cmp_frc_mvc}
\alias{cmp_frc_scl}
\alias{cmp_frc_vec}
\alias{cmp_ind_arr}
\alias{cmp_ind_gen}
\alias{cmp_ind_tab}
\alias{cmp_ind_vls}
\alias{cmp_ind_vtp}
\alias{cmp_ind_mat}
\alias{cmp_ind_mvc}
\alias{cmp_ind_scl}
\alias{cmp_ind_vec}
\alias{cmp_lgl_arr}
\alias{cmp_lgl_gen}
\alias{cmp_lgl_tab}
\alias{cmp_lgl_vls}
\alias{cmp_lgl_vtp}
\alias{cmp_lgl_mat}
\alias{cmp_lgl_mvc}
\alias{cmp_lgl_scl}
\alias{cmp_lgl_vec}
\alias{cmp_neg_arr}
\alias{cmp_neg_gen}
\alias{cmp_neg_tab}
\alias{cmp_neg_vls}
\alias{cmp_neg_vtp}
\alias{cmp_neg_mat}
\alias{cmp_neg_mvc}
\alias{cmp_neg_scl}
\alias{cmp_neg_vec}
\alias{cmp_ngw_arr}
\alias{cmp_ngw_gen}
\alias{cmp_ngw_tab}
\alias{cmp_ngw_vls}
\alias{cmp_ngw_vtp}
\alias{cmp_ngw_mat}
\alias{cmp_ngw_mvc}
\alias{cmp_ngw_scl}
\alias{cmp_ngw_vec}
\alias{cmp_nng_arr}
\alias{cmp_nng_gen}
\alias{cmp_nng_tab}
\alias{cmp_nng_vls}
\alias{cmp_nng_vtp}
\alias{cmp_nng_mat}
\alias{cmp_nng_mvc}
\alias{cmp_nng_scl}
\alias{cmp_nng_vec}
\alias{cmp_nnw_arr}
\alias{cmp_nnw_gen}
\alias{cmp_nnw_tab}
\alias{cmp_nnw_vls}
\alias{cmp_nnw_vtp}
\alias{cmp_nnw_mat}
\alias{cmp_nnw_mvc}
\alias{cmp_nnw_scl}
\alias{cmp_nnw_vec}
\alias{cmp_nps_arr}
\alias{cmp_nps_gen}
\alias{cmp_nps_tab}
\alias{cmp_nps_vls}
\alias{cmp_nps_vtp}
\alias{cmp_nps_mat}
\alias{cmp_nps_mvc}
\alias{cmp_nps_scl}
\alias{cmp_nps_vec}
\alias{cmp_npw_arr}
\alias{cmp_npw_gen}
\alias{cmp_npw_tab}
\alias{cmp_npw_vls}
\alias{cmp_npw_vtp}
\alias{cmp_npw_mat}
\alias{cmp_npw_mvc}
\alias{cmp_npw_scl}
\alias{cmp_npw_vec}
\alias{cmp_nst_arr}
\alias{cmp_nst_gen}
\alias{cmp_nst_tab}
\alias{cmp_nst_vls}
\alias{cmp_nst_vtp}
\alias{cmp_nst_mat}
\alias{cmp_nst_mvc}
\alias{cmp_nst_scl}
\alias{cmp_nst_vec}
\alias{cmp_num_arr}
\alias{cmp_num_gen}
\alias{cmp_num_tab}
\alias{cmp_num_vls}
\alias{cmp_num_vtp}
\alias{cmp_num_mat}
\alias{cmp_num_mvc}
\alias{cmp_num_scl}
\alias{cmp_num_vec}
\alias{cmp_odd_arr}
\alias{cmp_odd_gen}
\alias{cmp_odd_tab}
\alias{cmp_odd_vls}
\alias{cmp_odd_vtp}
\alias{cmp_odd_mat}
\alias{cmp_odd_mvc}
\alias{cmp_odd_scl}
\alias{cmp_odd_vec}
\alias{cmp_ord_arr}
\alias{cmp_ord_gen}
\alias{cmp_ord_tab}
\alias{cmp_ord_vls}
\alias{cmp_ord_vtp}
\alias{cmp_ord_mat}
\alias{cmp_ord_mvc}
\alias{cmp_ord_scl}
\alias{cmp_ord_vec}
\alias{cmp_pct_arr}
\alias{cmp_pct_gen}
\alias{cmp_pct_tab}
\alias{cmp_pct_vls}
\alias{cmp_pct_vtp}
\alias{cmp_pct_mat}
\alias{cmp_pct_mvc}
\alias{cmp_pct_scl}
\alias{cmp_pct_vec}
\alias{cmp_pos_arr}
\alias{cmp_pos_gen}
\alias{cmp_pos_tab}
\alias{cmp_pos_vls}
\alias{cmp_pos_vtp}
\alias{cmp_pos_mat}
\alias{cmp_pos_mvc}
\alias{cmp_pos_scl}
\alias{cmp_pos_vec}
\alias{cmp_ppn_arr}
\alias{cmp_ppn_gen}
\alias{cmp_ppn_tab}
\alias{cmp_ppn_vls}
\alias{cmp_ppn_vtp}
\alias{cmp_ppn_mat}
\alias{cmp_ppn_mvc}
\alias{cmp_ppn_scl}
\alias{cmp_ppn_vec}
\alias{cmp_psw_arr}
\alias{cmp_psw_gen}
\alias{cmp_psw_tab}
\alias{cmp_psw_vls}
\alias{cmp_psw_vtp}
\alias{cmp_psw_mat}
\alias{cmp_psw_mvc}
\alias{cmp_psw_scl}
\alias{cmp_psw_vec}
\alias{cmp_srt_arr}
\alias{cmp_srt_gen}
\alias{cmp_srt_tab}
\alias{cmp_srt_vls}
\alias{cmp_srt_vtp}
\alias{cmp_srt_mat}
\alias{cmp_srt_mvc}
\alias{cmp_srt_scl}
\alias{cmp_srt_vec}
\alias{cmp_str_arr}
\alias{cmp_str_gen}
\alias{cmp_str_tab}
\alias{cmp_str_vls}
\alias{cmp_str_vtp}
\alias{cmp_str_mat}
\alias{cmp_str_mvc}
\alias{cmp_str_scl}
\alias{cmp_str_vec}
\alias{cmp_uno_arr}
\alias{cmp_uno_gen}
\alias{cmp_uno_tab}
\alias{cmp_uno_vls}
\alias{cmp_uno_vtp}
\alias{cmp_uno_mat}
\alias{cmp_uno_mvc}
\alias{cmp_uno_scl}
\alias{cmp_uno_vec}
\alias{cmp_whl_arr}
\alias{cmp_whl_gen}
\alias{cmp_whl_tab}
\alias{cmp_whl_vls}
\alias{cmp_whl_vtp}
\alias{cmp_whl_mat}
\alias{cmp_whl_mvc}
\alias{cmp_whl_scl}
\alias{cmp_whl_vec}
\title{Complete + Extended Mode + Extended Class (cmp + mmm)}
\usage{
cmp_mmm_ccc_uj()

cmp_mmm_ccc_vals()

cmp_ch1_arr(x)

cmp_ch1_gen(x)

cmp_ch1_tab(x)

cmp_ch1_vls(x)

cmp_ch1_vtp(x)

cmp_ch1_mat(x)

cmp_ch1_mvc(x)

cmp_ch1_scl(x)

cmp_ch1_vec(x)

cmp_chr_arr(x)

cmp_chr_gen(x)

cmp_chr_tab(x)

cmp_chr_vls(x)

cmp_chr_vtp(x)

cmp_chr_mat(x)

cmp_chr_mvc(x)

cmp_chr_scl(x)

cmp_chr_vec(x)

cmp_clr_arr(x)

cmp_clr_gen(x)

cmp_clr_tab(x)

cmp_clr_vls(x)

cmp_clr_vtp(x)

cmp_clr_mat(x)

cmp_clr_mvc(x)

cmp_clr_scl(x)

cmp_clr_vec(x)

cmp_evn_arr(x)

cmp_evn_gen(x)

cmp_evn_tab(x)

cmp_evn_vls(x)

cmp_evn_vtp(x)

cmp_evn_mat(x)

cmp_evn_mvc(x)

cmp_evn_scl(x)

cmp_evn_vec(x)

cmp_fac_arr(x)

cmp_fac_gen(x)

cmp_fac_tab(x)

cmp_fac_vls(x)

cmp_fac_vtp(x)

cmp_fac_mat(x)

cmp_fac_mvc(x)

cmp_fac_scl(x)

cmp_fac_vec(x)

cmp_frc_arr(x)

cmp_frc_gen(x)

cmp_frc_tab(x)

cmp_frc_vls(x)

cmp_frc_vtp(x)

cmp_frc_mat(x)

cmp_frc_mvc(x)

cmp_frc_scl(x)

cmp_frc_vec(x)

cmp_ind_arr(x)

cmp_ind_gen(x)

cmp_ind_tab(x)

cmp_ind_vls(x)

cmp_ind_vtp(x)

cmp_ind_mat(x)

cmp_ind_mvc(x)

cmp_ind_scl(x)

cmp_ind_vec(x)

cmp_lgl_arr(x)

cmp_lgl_gen(x)

cmp_lgl_tab(x)

cmp_lgl_vls(x)

cmp_lgl_vtp(x)

cmp_lgl_mat(x)

cmp_lgl_mvc(x)

cmp_lgl_scl(x)

cmp_lgl_vec(x)

cmp_neg_arr(x)

cmp_neg_gen(x)

cmp_neg_tab(x)

cmp_neg_vls(x)

cmp_neg_vtp(x)

cmp_neg_mat(x)

cmp_neg_mvc(x)

cmp_neg_scl(x)

cmp_neg_vec(x)

cmp_ngw_arr(x)

cmp_ngw_gen(x)

cmp_ngw_tab(x)

cmp_ngw_vls(x)

cmp_ngw_vtp(x)

cmp_ngw_mat(x)

cmp_ngw_mvc(x)

cmp_ngw_scl(x)

cmp_ngw_vec(x)

cmp_nng_arr(x)

cmp_nng_gen(x)

cmp_nng_tab(x)

cmp_nng_vls(x)

cmp_nng_vtp(x)

cmp_nng_mat(x)

cmp_nng_mvc(x)

cmp_nng_scl(x)

cmp_nng_vec(x)

cmp_nnw_arr(x)

cmp_nnw_gen(x)

cmp_nnw_tab(x)

cmp_nnw_vls(x)

cmp_nnw_vtp(x)

cmp_nnw_mat(x)

cmp_nnw_mvc(x)

cmp_nnw_scl(x)

cmp_nnw_vec(x)

cmp_nps_arr(x)

cmp_nps_gen(x)

cmp_nps_tab(x)

cmp_nps_vls(x)

cmp_nps_vtp(x)

cmp_nps_mat(x)

cmp_nps_mvc(x)

cmp_nps_scl(x)

cmp_nps_vec(x)

cmp_npw_arr(x)

cmp_npw_gen(x)

cmp_npw_tab(x)

cmp_npw_vls(x)

cmp_npw_vtp(x)

cmp_npw_mat(x)

cmp_npw_mvc(x)

cmp_npw_scl(x)

cmp_npw_vec(x)

cmp_nst_arr(x)

cmp_nst_gen(x)

cmp_nst_tab(x)

cmp_nst_vls(x)

cmp_nst_vtp(x)

cmp_nst_mat(x)

cmp_nst_mvc(x)

cmp_nst_scl(x)

cmp_nst_vec(x)

cmp_num_arr(x)

cmp_num_gen(x)

cmp_num_tab(x)

cmp_num_vls(x)

cmp_num_vtp(x)

cmp_num_mat(x)

cmp_num_mvc(x)

cmp_num_scl(x)

cmp_num_vec(x)

cmp_odd_arr(x)

cmp_odd_gen(x)

cmp_odd_tab(x)

cmp_odd_vls(x)

cmp_odd_vtp(x)

cmp_odd_mat(x)

cmp_odd_mvc(x)

cmp_odd_scl(x)

cmp_odd_vec(x)

cmp_ord_arr(x)

cmp_ord_gen(x)

cmp_ord_tab(x)

cmp_ord_vls(x)

cmp_ord_vtp(x)

cmp_ord_mat(x)

cmp_ord_mvc(x)

cmp_ord_scl(x)

cmp_ord_vec(x)

cmp_pct_arr(x)

cmp_pct_gen(x)

cmp_pct_tab(x)

cmp_pct_vls(x)

cmp_pct_vtp(x)

cmp_pct_mat(x)

cmp_pct_mvc(x)

cmp_pct_scl(x)

cmp_pct_vec(x)

cmp_pos_arr(x)

cmp_pos_gen(x)

cmp_pos_tab(x)

cmp_pos_vls(x)

cmp_pos_vtp(x)

cmp_pos_mat(x)

cmp_pos_mvc(x)

cmp_pos_scl(x)

cmp_pos_vec(x)

cmp_ppn_arr(x)

cmp_ppn_gen(x)

cmp_ppn_tab(x)

cmp_ppn_vls(x)

cmp_ppn_vtp(x)

cmp_ppn_mat(x)

cmp_ppn_mvc(x)

cmp_ppn_scl(x)

cmp_ppn_vec(x)

cmp_psw_arr(x)

cmp_psw_gen(x)

cmp_psw_tab(x)

cmp_psw_vls(x)

cmp_psw_vtp(x)

cmp_psw_mat(x)

cmp_psw_mvc(x)

cmp_psw_scl(x)

cmp_psw_vec(x)

cmp_srt_arr(x)

cmp_srt_gen(x)

cmp_srt_tab(x)

cmp_srt_vls(x)

cmp_srt_vtp(x)

cmp_srt_mat(x)

cmp_srt_mvc(x)

cmp_srt_scl(x)

cmp_srt_vec(x)

cmp_str_arr(x)

cmp_str_gen(x)

cmp_str_tab(x)

cmp_str_vls(x)

cmp_str_vtp(x)

cmp_str_mat(x)

cmp_str_mvc(x)

cmp_str_scl(x)

cmp_str_vec(x)

cmp_uno_arr(x)

cmp_uno_gen(x)

cmp_uno_tab(x)

cmp_uno_vls(x)

cmp_uno_vtp(x)

cmp_uno_mat(x)

cmp_uno_mvc(x)

cmp_uno_scl(x)

cmp_uno_vec(x)

cmp_whl_arr(x)

cmp_whl_gen(x)

cmp_whl_tab(x)

cmp_whl_vls(x)

cmp_whl_vtp(x)

cmp_whl_mat(x)

cmp_whl_mvc(x)

cmp_whl_scl(x)

cmp_whl_vec(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE} except for \code{cmp_mmm_ccc_vals}, which
returns a character vector.
}
\description{
Check whether an object is complete, of a specific extended
mode, and of a specific extended class. See \code{\link{icmp}}, and
\code{\link{immm}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_mmm_ccc_vals}: Get a character vector of all possible complete +
extended mode + extended class properties.

\item \code{cmp_ch1_arr}: Is \code{x} a complete onechar array?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{iarr}})

\item \code{cmp_ch1_gen}: Is \code{x} a complete onechar
generic?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{igen}})

\item \code{cmp_ch1_tab}: Is \code{x} a complete onechar
tabular?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{itab}})

\item \code{cmp_ch1_vls}: Is \code{x} a complete onechar
vlist?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{ivls}})

\item \code{cmp_ch1_vtp}: Is \code{x} a complete onechar
vtype?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{ivtp}})

\item \code{cmp_ch1_mat}: Is \code{x} a complete onechar
matrix?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{imat}})

\item \code{cmp_ch1_mvc}: Is \code{x} a complete onechar
mvector?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{imvc}})

\item \code{cmp_ch1_scl}: Is \code{x} a complete onechar
scalar?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{iscl}})

\item \code{cmp_ch1_vec}: Is \code{x} a complete onechar
vector?
(\code{\link{icmp}}, \code{\link{ich1}}, \code{\link{ivec}})

\item \code{cmp_chr_arr}: Is \code{x} a complete character
array?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{iarr}})

\item \code{cmp_chr_gen}: Is \code{x} a complete character
generic?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{igen}})

\item \code{cmp_chr_tab}: Is \code{x} a complete character
tabular?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{itab}})

\item \code{cmp_chr_vls}: Is \code{x} a complete character
vlist?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{ivls}})

\item \code{cmp_chr_vtp}: Is \code{x} a complete character
vtype?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{ivtp}})

\item \code{cmp_chr_mat}: Is \code{x} a complete character
matrix?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{imat}})

\item \code{cmp_chr_mvc}: Is \code{x} a complete character
mvector?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{imvc}})

\item \code{cmp_chr_scl}: Is \code{x} a complete character
scalar?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{iscl}})

\item \code{cmp_chr_vec}: Is \code{x} a complete character
vector?
(\code{\link{icmp}}, \code{\link{ichr}}, \code{\link{ivec}})

\item \code{cmp_clr_arr}: Is \code{x} a complete color
array?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{iarr}})

\item \code{cmp_clr_gen}: Is \code{x} a complete color
generic?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{igen}})

\item \code{cmp_clr_tab}: Is \code{x} a complete color
tabular?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{itab}})

\item \code{cmp_clr_vls}: Is \code{x} a complete color
vlist?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{ivls}})

\item \code{cmp_clr_vtp}: Is \code{x} a complete color
vtype?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{ivtp}})

\item \code{cmp_clr_mat}: Is \code{x} a complete color
matrix?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{imat}})

\item \code{cmp_clr_mvc}: Is \code{x} a complete color
mvector?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{imvc}})

\item \code{cmp_clr_scl}: Is \code{x} a complete color
scalar?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{iscl}})

\item \code{cmp_clr_vec}: Is \code{x} a complete color
vector?
(\code{\link{icmp}}, \code{\link{iclr}}, \code{\link{ivec}})

\item \code{cmp_evn_arr}: Is \code{x} a complete even-valued
whole-number array?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{iarr}})

\item \code{cmp_evn_gen}: Is \code{x} a complete even-valued
whole-number generic?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{igen}})

\item \code{cmp_evn_tab}: Is \code{x} a complete even-valued
whole-number tabular?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{itab}})

\item \code{cmp_evn_vls}: Is \code{x} a complete even-valued
whole-number vlist?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{ivls}})

\item \code{cmp_evn_vtp}: Is \code{x} a complete even-valued
whole-number vtype?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{ivtp}})

\item \code{cmp_evn_mat}: Is \code{x} a complete even-valued
whole-number matrix?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{imat}})

\item \code{cmp_evn_mvc}: Is \code{x} a complete even-valued
whole-number mvector?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{imvc}})

\item \code{cmp_evn_scl}: Is \code{x} a complete even-valued
whole-number scalar?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{iscl}})

\item \code{cmp_evn_vec}: Is \code{x} a complete even-valued
whole-number vector?
(\code{\link{icmp}}, \code{\link{ievn}}, \code{\link{ivec}})

\item \code{cmp_fac_arr}: Is \code{x} a complete factor
array?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{iarr}})

\item \code{cmp_fac_gen}: Is \code{x} a complete factor
generic?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{igen}})

\item \code{cmp_fac_tab}: Is \code{x} a complete factor
tabular?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{itab}})

\item \code{cmp_fac_vls}: Is \code{x} a complete factor
vlist?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{ivls}})

\item \code{cmp_fac_vtp}: Is \code{x} a complete factor
vtype?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{ivtp}})

\item \code{cmp_fac_mat}: Is \code{x} a complete factor
matrix?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{imat}})

\item \code{cmp_fac_mvc}: Is \code{x} a complete factor
mvector?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{imvc}})

\item \code{cmp_fac_scl}: Is \code{x} a complete factor
scalar?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{iscl}})

\item \code{cmp_fac_vec}: Is \code{x} a complete factor
vector?
(\code{\link{icmp}}, \code{\link{ifac}}, \code{\link{ivec}})

\item \code{cmp_frc_arr}: Is \code{x} a complete fractional-valued
numeric array?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{iarr}})

\item \code{cmp_frc_gen}: Is \code{x} a complete fractional-valued
numeric generic?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{igen}})

\item \code{cmp_frc_tab}: Is \code{x} a complete fractional-valued
numeric tabular?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{itab}})

\item \code{cmp_frc_vls}: Is \code{x} a complete fractional-valued
numeric vlist?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{ivls}})

\item \code{cmp_frc_vtp}: Is \code{x} a complete fractional-valued
numeric vtype?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{ivtp}})

\item \code{cmp_frc_mat}: Is \code{x} a complete fractional-valued
numeric matrix?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{imat}})

\item \code{cmp_frc_mvc}: Is \code{x} a complete fractional-valued
numeric mvector?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{imvc}})

\item \code{cmp_frc_scl}: Is \code{x} a complete fractional-valued
numeric scalar?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{iscl}})

\item \code{cmp_frc_vec}: Is \code{x} a complete fractional-valued
numeric vector?
(\code{\link{icmp}}, \code{\link{ifrc}}, \code{\link{ivec}})

\item \code{cmp_ind_arr}: Is \code{x} a complete indexer
array?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{iarr}})

\item \code{cmp_ind_gen}: Is \code{x} a complete indexer
generic?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{igen}})

\item \code{cmp_ind_tab}: Is \code{x} a complete indexer
tabular?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{itab}})

\item \code{cmp_ind_vls}: Is \code{x} a complete indexer
vlist?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{ivls}})

\item \code{cmp_ind_vtp}: Is \code{x} a complete indexer
vtype?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{ivtp}})

\item \code{cmp_ind_mat}: Is \code{x} a complete indexer
matrix?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{imat}})

\item \code{cmp_ind_mvc}: Is \code{x} a complete indexer
mvector?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{imvc}})

\item \code{cmp_ind_scl}: Is \code{x} a complete indexer
scalar?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{iscl}})

\item \code{cmp_ind_vec}: Is \code{x} a complete indexer
vector?
(\code{\link{icmp}}, \code{\link{iind}}, \code{\link{ivec}})

\item \code{cmp_lgl_arr}: Is \code{x} a complete logical
array?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{iarr}})

\item \code{cmp_lgl_gen}: Is \code{x} a complete logical
generic?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{igen}})

\item \code{cmp_lgl_tab}: Is \code{x} a complete logical
tabular?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{itab}})

\item \code{cmp_lgl_vls}: Is \code{x} a complete logical
vlist?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{ivls}})

\item \code{cmp_lgl_vtp}: Is \code{x} a complete logical
vtype?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{ivtp}})

\item \code{cmp_lgl_mat}: Is \code{x} a complete logical
matrix?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{imat}})

\item \code{cmp_lgl_mvc}: Is \code{x} a complete logical
mvector?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{imvc}})

\item \code{cmp_lgl_scl}: Is \code{x} a complete logical
scalar?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{iscl}})

\item \code{cmp_lgl_vec}: Is \code{x} a complete logical
vector?
(\code{\link{icmp}}, \code{\link{ilgl}}, \code{\link{ivec}})

\item \code{cmp_neg_arr}: Is \code{x} a complete negative-valued
numeric array?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{iarr}})

\item \code{cmp_neg_gen}: Is \code{x} a complete negative-valued
numeric generic?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{igen}})

\item \code{cmp_neg_tab}: Is \code{x} a complete negative-valued
numeric tabular?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{itab}})

\item \code{cmp_neg_vls}: Is \code{x} a complete negative-valued
numeric vlist?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{ivls}})

\item \code{cmp_neg_vtp}: Is \code{x} a complete negative-valued
numeric vtype?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{ivtp}})

\item \code{cmp_neg_mat}: Is \code{x} a complete negative-valued
numeric matrix?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{imat}})

\item \code{cmp_neg_mvc}: Is \code{x} a complete negative-valued
numeric mvector?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{imvc}})

\item \code{cmp_neg_scl}: Is \code{x} a complete negative-valued
numeric scalar?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{iscl}})

\item \code{cmp_neg_vec}: Is \code{x} a complete negative-valued
numeric vector?
(\code{\link{icmp}}, \code{\link{ineg}}, \code{\link{ivec}})

\item \code{cmp_ngw_arr}: Is \code{x} a complete negative-valued
whole-number array?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{iarr}})

\item \code{cmp_ngw_gen}: Is \code{x} a complete negative-valued
whole-number generic?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{igen}})

\item \code{cmp_ngw_tab}: Is \code{x} a complete negative-valued
whole-number tabular?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{itab}})

\item \code{cmp_ngw_vls}: Is \code{x} a complete negative-valued
whole-number vlist?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{ivls}})

\item \code{cmp_ngw_vtp}: Is \code{x} a complete negative-valued
whole-number vtype?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{ivtp}})

\item \code{cmp_ngw_mat}: Is \code{x} a complete negative-valued
whole-number matrix?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{imat}})

\item \code{cmp_ngw_mvc}: Is \code{x} a complete negative-valued
whole-number mvector?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{imvc}})

\item \code{cmp_ngw_scl}: Is \code{x} a complete negative-valued
whole-number scalar?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{iscl}})

\item \code{cmp_ngw_vec}: Is \code{x} a complete negative-valued
whole-number vector?
(\code{\link{icmp}}, \code{\link{ingw}}, \code{\link{ivec}})

\item \code{cmp_nng_arr}: Is \code{x} a complete non-negative-valued
numeric array?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{iarr}})

\item \code{cmp_nng_gen}: Is \code{x} a complete non-negative-valued
numeric generic?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{igen}})

\item \code{cmp_nng_tab}: Is \code{x} a complete non-negative-valued
numeric tabular?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{itab}})

\item \code{cmp_nng_vls}: Is \code{x} a complete non-negative-valued
numeric vlist?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{ivls}})

\item \code{cmp_nng_vtp}: Is \code{x} a complete non-negative-valued
numeric vtype?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{ivtp}})

\item \code{cmp_nng_mat}: Is \code{x} a complete non-negative-valued
numeric matrix?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{imat}})

\item \code{cmp_nng_mvc}: Is \code{x} a complete non-negative-valued
numeric mvector?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{imvc}})

\item \code{cmp_nng_scl}: Is \code{x} a complete non-negative-valued
numeric scalar?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{iscl}})

\item \code{cmp_nng_vec}: Is \code{x} a complete non-negative-valued
numeric vector?
(\code{\link{icmp}}, \code{\link{inng}}, \code{\link{ivec}})

\item \code{cmp_nnw_arr}: Is \code{x} a complete non-negative-valued
whole-number array?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{iarr}})

\item \code{cmp_nnw_gen}: Is \code{x} a complete non-negative-valued
whole-number generic?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{igen}})

\item \code{cmp_nnw_tab}: Is \code{x} a complete non-negative-valued
whole-number tabular?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{itab}})

\item \code{cmp_nnw_vls}: Is \code{x} a complete non-negative-valued
whole-number vlist?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{ivls}})

\item \code{cmp_nnw_vtp}: Is \code{x} a complete non-negative-valued
whole-number vtype?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{ivtp}})

\item \code{cmp_nnw_mat}: Is \code{x} a complete non-negative-valued
whole-number matrix?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{imat}})

\item \code{cmp_nnw_mvc}: Is \code{x} a complete non-negative-valued
whole-number mvector?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{imvc}})

\item \code{cmp_nnw_scl}: Is \code{x} a complete non-negative-valued
whole-number scalar?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{iscl}})

\item \code{cmp_nnw_vec}: Is \code{x} a complete non-negative-valued
whole-number vector?
(\code{\link{icmp}}, \code{\link{innw}}, \code{\link{ivec}})

\item \code{cmp_nps_arr}: Is \code{x} a complete non-positive-valued
numeric array?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{iarr}})

\item \code{cmp_nps_gen}: Is \code{x} a complete non-positive-valued
numeric generic?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{igen}})

\item \code{cmp_nps_tab}: Is \code{x} a complete non-positive-valued
numeric tabular?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{itab}})

\item \code{cmp_nps_vls}: Is \code{x} a complete non-positive-valued
numeric vlist?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{ivls}})

\item \code{cmp_nps_vtp}: Is \code{x} a complete non-positive-valued
numeric vtype?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{ivtp}})

\item \code{cmp_nps_mat}: Is \code{x} a complete non-positive-valued
numeric matrix?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{imat}})

\item \code{cmp_nps_mvc}: Is \code{x} a complete non-positive-valued
numeric mvector?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{imvc}})

\item \code{cmp_nps_scl}: Is \code{x} a complete non-positive-valued
numeric scalar?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{iscl}})

\item \code{cmp_nps_vec}: Is \code{x} a complete non-positive-valued
numeric vector?
(\code{\link{icmp}}, \code{\link{inps}}, \code{\link{ivec}})

\item \code{cmp_npw_arr}: Is \code{x} a complete non-positive-valued
whole-number array?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{iarr}})

\item \code{cmp_npw_gen}: Is \code{x} a complete non-positive-valued
whole-number generic?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{igen}})

\item \code{cmp_npw_tab}: Is \code{x} a complete non-positive-valued
whole-number tabular?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{itab}})

\item \code{cmp_npw_vls}: Is \code{x} a complete non-positive-valued
whole-number vlist?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{ivls}})

\item \code{cmp_npw_vtp}: Is \code{x} a complete non-positive-valued
whole-number vtype?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{ivtp}})

\item \code{cmp_npw_mat}: Is \code{x} a complete non-positive-valued
whole-number matrix?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{imat}})

\item \code{cmp_npw_mvc}: Is \code{x} a complete non-positive-valued
whole-number mvector?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{imvc}})

\item \code{cmp_npw_scl}: Is \code{x} a complete non-positive-valued
whole-number scalar?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{iscl}})

\item \code{cmp_npw_vec}: Is \code{x} a complete non-positive-valued
whole-number vector?
(\code{\link{icmp}}, \code{\link{inpw}}, \code{\link{ivec}})

\item \code{cmp_nst_arr}: Is \code{x} a complete non-sortable
array?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{iarr}})

\item \code{cmp_nst_gen}: Is \code{x} a complete non-sortable
generic?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{igen}})

\item \code{cmp_nst_tab}: Is \code{x} a complete non-sortable
tabular?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{itab}})

\item \code{cmp_nst_vls}: Is \code{x} a complete non-sortable
vlist?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{ivls}})

\item \code{cmp_nst_vtp}: Is \code{x} a complete non-sortable
vtype?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{ivtp}})

\item \code{cmp_nst_mat}: Is \code{x} a complete non-sortable
matrix?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{imat}})

\item \code{cmp_nst_mvc}: Is \code{x} a complete non-sortable
mvector?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{imvc}})

\item \code{cmp_nst_scl}: Is \code{x} a complete non-sortable
scalar?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{iscl}})

\item \code{cmp_nst_vec}: Is \code{x} a complete non-sortable
vector?
(\code{\link{icmp}}, \code{\link{inst}}, \code{\link{ivec}})

\item \code{cmp_num_arr}: Is \code{x} a complete numeric
array?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{iarr}})

\item \code{cmp_num_gen}: Is \code{x} a complete numeric
generic?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{igen}})

\item \code{cmp_num_tab}: Is \code{x} a complete numeric
tabular?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{itab}})

\item \code{cmp_num_vls}: Is \code{x} a complete numeric
vlist?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{ivls}})

\item \code{cmp_num_vtp}: Is \code{x} a complete numeric
vtype?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{ivtp}})

\item \code{cmp_num_mat}: Is \code{x} a complete numeric
matrix?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{imat}})

\item \code{cmp_num_mvc}: Is \code{x} a complete numeric
mvector?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{imvc}})

\item \code{cmp_num_scl}: Is \code{x} a complete numeric
scalar?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{iscl}})

\item \code{cmp_num_vec}: Is \code{x} a complete numeric
vector?
(\code{\link{icmp}}, \code{\link{inum}}, \code{\link{ivec}})

\item \code{cmp_odd_arr}: Is \code{x} a complete odd-valued
whole-number array?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{iarr}})

\item \code{cmp_odd_gen}: Is \code{x} a complete odd-valued
whole-number generic?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{igen}})

\item \code{cmp_odd_tab}: Is \code{x} a complete odd-valued
whole-number tabular?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{itab}})

\item \code{cmp_odd_vls}: Is \code{x} a complete odd-valued
whole-number vlist?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{ivls}})

\item \code{cmp_odd_vtp}: Is \code{x} a complete odd-valued
whole-number vtype?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{ivtp}})

\item \code{cmp_odd_mat}: Is \code{x} a complete odd-valued
whole-number matrix?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{imat}})

\item \code{cmp_odd_mvc}: Is \code{x} a complete odd-valued
whole-number mvector?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{imvc}})

\item \code{cmp_odd_scl}: Is \code{x} a complete odd-valued
whole-number scalar?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{iscl}})

\item \code{cmp_odd_vec}: Is \code{x} a complete odd-valued
whole-number vector?
(\code{\link{icmp}}, \code{\link{iodd}}, \code{\link{ivec}})

\item \code{cmp_ord_arr}: Is \code{x} a complete ordered-factor
array?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{iarr}})

\item \code{cmp_ord_gen}: Is \code{x} a complete ordered-factor
generic?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{igen}})

\item \code{cmp_ord_tab}: Is \code{x} a complete ordered-factor
tabular?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{itab}})

\item \code{cmp_ord_vls}: Is \code{x} a complete ordered-factor
vlist?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{ivls}})

\item \code{cmp_ord_vtp}: Is \code{x} a complete ordered-factor
vtype?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{ivtp}})

\item \code{cmp_ord_mat}: Is \code{x} a complete ordered-factor
matrix?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{imat}})

\item \code{cmp_ord_mvc}: Is \code{x} a complete ordered-factor
mvector?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{imvc}})

\item \code{cmp_ord_scl}: Is \code{x} a complete ordered-factor
scalar?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{iscl}})

\item \code{cmp_ord_vec}: Is \code{x} a complete ordered-factor
vector?
(\code{\link{icmp}}, \code{\link{iord}}, \code{\link{ivec}})

\item \code{cmp_pct_arr}: Is \code{x} a complete percent-valued
(0-100) numeric array?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{iarr}})

\item \code{cmp_pct_gen}: Is \code{x} a complete percent-valued
(0-100) numeric generic?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{igen}})

\item \code{cmp_pct_tab}: Is \code{x} a complete percent-valued
(0-100) numeric tabular?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{itab}})

\item \code{cmp_pct_vls}: Is \code{x} a complete percent-valued
(0-100) numeric vlist?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{ivls}})

\item \code{cmp_pct_vtp}: Is \code{x} a complete percent-valued
(0-100) numeric vtype?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{ivtp}})

\item \code{cmp_pct_mat}: Is \code{x} a complete percent-valued
(0-100) numeric matrix?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{imat}})

\item \code{cmp_pct_mvc}: Is \code{x} a complete percent-valued
(0-100) numeric mvector?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{imvc}})

\item \code{cmp_pct_scl}: Is \code{x} a complete percent-valued
(0-100) numeric scalar?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{iscl}})

\item \code{cmp_pct_vec}: Is \code{x} a complete percent-valued
(0-100) numeric vector?
(\code{\link{icmp}}, \code{\link{ipct}}, \code{\link{ivec}})

\item \code{cmp_pos_arr}: Is \code{x} a complete positive-valued
numeric array?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{iarr}})

\item \code{cmp_pos_gen}: Is \code{x} a complete positive-valued
numeric generic?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{igen}})

\item \code{cmp_pos_tab}: Is \code{x} a complete positive-valued
numeric tabular?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{itab}})

\item \code{cmp_pos_vls}: Is \code{x} a complete positive-valued
numeric vlist?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{ivls}})

\item \code{cmp_pos_vtp}: Is \code{x} a complete positive-valued
numeric vtype?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{ivtp}})

\item \code{cmp_pos_mat}: Is \code{x} a complete positive-valued
numeric matrix?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{imat}})

\item \code{cmp_pos_mvc}: Is \code{x} a complete positive-valued
numeric mvector?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{imvc}})

\item \code{cmp_pos_scl}: Is \code{x} a complete positive-valued
numeric scalar?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{iscl}})

\item \code{cmp_pos_vec}: Is \code{x} a complete positive-valued
numeric vector?
(\code{\link{icmp}}, \code{\link{ipos}}, \code{\link{ivec}})

\item \code{cmp_ppn_arr}: Is \code{x} a complete proportion-valued
(0-1) numeric array?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{iarr}})

\item \code{cmp_ppn_gen}: Is \code{x} a complete proportion-valued
(0-1) numeric generic?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{igen}})

\item \code{cmp_ppn_tab}: Is \code{x} a complete proportion-valued
(0-1) numeric tabular?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{itab}})

\item \code{cmp_ppn_vls}: Is \code{x} a complete proportion-valued
(0-1) numeric vlist?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{ivls}})

\item \code{cmp_ppn_vtp}: Is \code{x} a complete proportion-valued
(0-1) numeric vtype?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{ivtp}})

\item \code{cmp_ppn_mat}: Is \code{x} a complete proportion-valued
(0-1) numeric matrix?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{imat}})

\item \code{cmp_ppn_mvc}: Is \code{x} a complete proportion-valued
(0-1) numeric mvector?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{imvc}})

\item \code{cmp_ppn_scl}: Is \code{x} a complete proportion-valued
(0-1) numeric scalar?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{iscl}})

\item \code{cmp_ppn_vec}: Is \code{x} a complete proportion-valued
(0-1) numeric vector?
(\code{\link{icmp}}, \code{\link{ippn}}, \code{\link{ivec}})

\item \code{cmp_psw_arr}: Is \code{x} a complete positive-valued
whole-number array?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{iarr}})

\item \code{cmp_psw_gen}: Is \code{x} a complete positive-valued
whole-number generic?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{igen}})

\item \code{cmp_psw_tab}: Is \code{x} a complete positive-valued
whole-number tabular?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{itab}})

\item \code{cmp_psw_vls}: Is \code{x} a complete positive-valued
whole-number vlist?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{ivls}})

\item \code{cmp_psw_vtp}: Is \code{x} a complete positive-valued
whole-number vtype?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{ivtp}})

\item \code{cmp_psw_mat}: Is \code{x} a complete positive-valued
whole-number matrix?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{imat}})

\item \code{cmp_psw_mvc}: Is \code{x} a complete positive-valued
whole-number mvector?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{imvc}})

\item \code{cmp_psw_scl}: Is \code{x} a complete positive-valued
whole-number scalar?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{iscl}})

\item \code{cmp_psw_vec}: Is \code{x} a complete positive-valued
whole-number vector?
(\code{\link{icmp}}, \code{\link{ipsw}}, \code{\link{ivec}})

\item \code{cmp_srt_arr}: Is \code{x} a complete sortable
array?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{iarr}})

\item \code{cmp_srt_gen}: Is \code{x} a complete sortable
generic?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{igen}})

\item \code{cmp_srt_tab}: Is \code{x} a complete sortable
tabular?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{itab}})

\item \code{cmp_srt_vls}: Is \code{x} a complete sortable
vlist?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{ivls}})

\item \code{cmp_srt_vtp}: Is \code{x} a complete sortable
vtype?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{ivtp}})

\item \code{cmp_srt_mat}: Is \code{x} a complete sortable
matrix?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{imat}})

\item \code{cmp_srt_mvc}: Is \code{x} a complete sortable
mvector?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{imvc}})

\item \code{cmp_srt_scl}: Is \code{x} a complete sortable
scalar?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{iscl}})

\item \code{cmp_srt_vec}: Is \code{x} a complete sortable
vector?
(\code{\link{icmp}}, \code{\link{isrt}}, \code{\link{ivec}})

\item \code{cmp_str_arr}: Is \code{x} a complete string
array?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{iarr}})

\item \code{cmp_str_gen}: Is \code{x} a complete string
generic?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{igen}})

\item \code{cmp_str_tab}: Is \code{x} a complete string
tabular?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{itab}})

\item \code{cmp_str_vls}: Is \code{x} a complete string
vlist?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{ivls}})

\item \code{cmp_str_vtp}: Is \code{x} a complete string
vtype?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{ivtp}})

\item \code{cmp_str_mat}: Is \code{x} a complete string
matrix?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{imat}})

\item \code{cmp_str_mvc}: Is \code{x} a complete string
mvector?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{imvc}})

\item \code{cmp_str_scl}: Is \code{x} a complete string
scalar?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{iscl}})

\item \code{cmp_str_vec}: Is \code{x} a complete string
vector?
(\code{\link{icmp}}, \code{\link{istr}}, \code{\link{ivec}})

\item \code{cmp_uno_arr}: Is \code{x} a complete unordered-factor
array?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{iarr}})

\item \code{cmp_uno_gen}: Is \code{x} a complete unordered-factor
generic?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{igen}})

\item \code{cmp_uno_tab}: Is \code{x} a complete unordered-factor
tabular?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{itab}})

\item \code{cmp_uno_vls}: Is \code{x} a complete unordered-factor
vlist?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{ivls}})

\item \code{cmp_uno_vtp}: Is \code{x} a complete unordered-factor
vtype?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{ivtp}})

\item \code{cmp_uno_mat}: Is \code{x} a complete unordered-factor
matrix?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{imat}})

\item \code{cmp_uno_mvc}: Is \code{x} a complete unordered-factor
mvector?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{imvc}})

\item \code{cmp_uno_scl}: Is \code{x} a complete unordered-factor
scalar?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{iscl}})

\item \code{cmp_uno_vec}: Is \code{x} a complete unordered-factor
vector?
(\code{\link{icmp}}, \code{\link{iuno}}, \code{\link{ivec}})

\item \code{cmp_whl_arr}: Is \code{x} a complete whole-number array?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{iarr}})

\item \code{cmp_whl_gen}: Is \code{x} a complete whole-number generic?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{igen}})

\item \code{cmp_whl_tab}: Is \code{x} a complete whole-number tabular?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{itab}})

\item \code{cmp_whl_vls}: Is \code{x} a complete whole-number vlist?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{ivls}})

\item \code{cmp_whl_vtp}: Is \code{x} a complete whole-number vtype?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{ivtp}})

\item \code{cmp_whl_mat}: Is \code{x} a complete whole-number matrix?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{imat}})

\item \code{cmp_whl_mvc}: Is \code{x} a complete whole-number mvector?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{imvc}})

\item \code{cmp_whl_scl}: Is \code{x} a complete whole-number scalar?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{iscl}})

\item \code{cmp_whl_vec}: Is \code{x} a complete whole-number vector?
(\code{\link{icmp}}, \code{\link{iwhl}}, \code{\link{ivec}})
}}

\seealso{
Other props: 
\code{\link{as_mmm_uj}()},
\code{\link{ccc_uj}()},
\code{\link{cmp_ccc_uj}()},
\code{\link{cmp_mmm_uj}()},
\code{\link{comparable_uj}()},
\code{\link{compatible_uj}()},
\code{\link{ddd_uj}()},
\code{\link{eee_uj}()},
\code{\link{fff_uj}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm_uj}()},
\code{\link{meets_uj}()},
\code{\link{mmm_ccc_uj}()},
\code{\link{mmm_uj}()},
\code{\link{sss_uj}()},
\code{\link{ttt_ccc_uj}()},
\code{\link{ttt_mmm_uj}()},
\code{\link{ttt_uj}()},
\code{\link{xxx_uj}()}
}
\concept{props}
