% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_gen.R
\name{cmp_mmm_gen}
\alias{cmp_mmm_gen}
\alias{cmp_mmm_gen_vals}
\alias{cmp_atm_gen}
\alias{cmp_chr_gen}
\alias{cmp_ch1_gen}
\alias{cmp_clr_gen}
\alias{cmp_evn_gen}
\alias{cmp_fac_gen}
\alias{cmp_frc_gen}
\alias{cmp_ind_gen}
\alias{cmp_lgc_gen}
\alias{cmp_neg_gen}
\alias{cmp_ngw_gen}
\alias{cmp_nng_gen}
\alias{cmp_nnw_gen}
\alias{cmp_nps_gen}
\alias{cmp_npw_gen}
\alias{cmp_nst_gen}
\alias{cmp_num_gen}
\alias{cmp_odd_gen}
\alias{cmp_ord_gen}
\alias{cmp_pct_gen}
\alias{cmp_pos_gen}
\alias{cmp_ppn_gen}
\alias{cmp_psw_gen}
\alias{cmp_srt_gen}
\alias{cmp_str_gen}
\alias{cmp_uno_gen}
\alias{cmp_whl_gen}
\title{Complete + Extended Mode + Generic Atomic (cmp + mmm + avl; includes
atomic tibbles and atomic vlists)}
\usage{
cmp_mmm_gen_vals()

cmp_atm_gen(x)

cmp_chr_gen(x)

cmp_ch1_gen(x)

cmp_clr_gen(x)

cmp_evn_gen(x)

cmp_fac_gen(x)

cmp_frc_gen(x)

cmp_ind_gen(x)

cmp_lgc_gen(x)

cmp_neg_gen(x)

cmp_ngw_gen(x)

cmp_nng_gen(x)

cmp_nnw_gen(x)

cmp_nps_gen(x)

cmp_npw_gen(x)

cmp_nst_gen(x)

cmp_num_gen(x)

cmp_odd_gen(x)

cmp_ord_gen(x)

cmp_pct_gen(x)

cmp_pos_gen(x)

cmp_ppn_gen(x)

cmp_psw_gen(x)

cmp_srt_gen(x)

cmp_str_gen(x)

cmp_uno_gen(x)

cmp_whl_gen(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_cmp_generic}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
