% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_mat.R
\name{cmp_mmm_mat}
\alias{cmp_mmm_mat}
\alias{cmp_mmm_mat_vals}
\alias{cmp_atm_mat}
\alias{cmp_chr_mat}
\alias{cmp_ch1_mat}
\alias{cmp_clr_mat}
\alias{cmp_evn_mat}
\alias{cmp_fac_mat}
\alias{cmp_frc_mat}
\alias{cmp_ind_mat}
\alias{cmp_lgl_mat}
\alias{cmp_neg_mat}
\alias{cmp_ngw_mat}
\alias{cmp_nng_mat}
\alias{cmp_nnw_mat}
\alias{cmp_nps_mat}
\alias{cmp_npw_mat}
\alias{cmp_nst_mat}
\alias{cmp_num_mat}
\alias{cmp_odd_mat}
\alias{cmp_ord_mat}
\alias{cmp_pct_mat}
\alias{cmp_pos_mat}
\alias{cmp_ppn_mat}
\alias{cmp_psw_mat}
\alias{cmp_srt_mat}
\alias{cmp_str_mat}
\alias{cmp_uno_mat}
\alias{cmp_whl_mat}
\title{Complete + Extended Mode + Atomic Matrix (comp + mmm + mat)}
\usage{
cmp_mmm_mat_vals()

cmp_atm_mat(x)

cmp_chr_mat(x)

cmp_ch1_mat(x)

cmp_clr_mat(x)

cmp_evn_mat(x)

cmp_fac_mat(x)

cmp_frc_mat(x)

cmp_ind_mat(x)

cmp_lgl_mat(x)

cmp_neg_mat(x)

cmp_ngw_mat(x)

cmp_nng_mat(x)

cmp_nnw_mat(x)

cmp_nps_mat(x)

cmp_npw_mat(x)

cmp_nst_mat(x)

cmp_num_mat(x)

cmp_odd_mat(x)

cmp_ord_mat(x)

cmp_pct_mat(x)

cmp_pos_mat(x)

cmp_ppn_mat(x)

cmp_psw_mat(x)

cmp_srt_mat(x)

cmp_str_mat(x)

cmp_uno_mat(x)

cmp_whl_mat(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic matrix
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_mat}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_mat}: Is \code{x} a complete atomic matrix?

\item \code{cmp_chr_mat}: Is \code{x} a complete character atomic matrix?

\item \code{cmp_ch1_mat}: Is \code{x} a complete onechar atomic matrix?

\item \code{cmp_clr_mat}: Is \code{x} a complete color atomic matrix?

\item \code{cmp_evn_mat}: Is \code{x} a complete even-number atomic matrix?

\item \code{cmp_fac_mat}: Is \code{x} a complete factor atomic matrix?

\item \code{cmp_frc_mat}: Is \code{x} a complete fractional-number atomic
matrix?

\item \code{cmp_ind_mat}: Is \code{x} a complete indexer atomic matrix?

\item \code{cmp_lgl_mat}: Is \code{x} a complete logical atomic matrix?

\item \code{cmp_neg_mat}: Is \code{x} a complete negative-number atomic matrix?

\item \code{cmp_ngw_mat}: Is \code{x} a complete negative-whole-number atomic
matrix?

\item \code{cmp_nng_mat}: Is \code{x} a complete non-negative-number atomic
matrix?

\item \code{cmp_nnw_mat}: Is \code{x} a complete non-negative-whole-number
atomic matrix?

\item \code{cmp_nps_mat}: Is \code{x} a complete non-positive-number atomic
matrix?

\item \code{cmp_npw_mat}: Is \code{x} a complete non-positive-whole-number
atomic matrix?

\item \code{cmp_nst_mat}: Is \code{x} a complete non-sortable atomic matrix?

\item \code{cmp_num_mat}: Is \code{x} a complete number atomic matrix?

\item \code{cmp_odd_mat}: Is \code{x} a complete odd-number atomic matrix?

\item \code{cmp_ord_mat}: Is \code{x} a complete ordered-factor atomic matrix?

\item \code{cmp_pct_mat}: Is \code{x} a complete percent-number atomic matrix?

\item \code{cmp_pos_mat}: Is \code{x} a complete positive-number atomic matrix?

\item \code{cmp_ppn_mat}: Is \code{x} a complete proportion-number atomic
matrix?

\item \code{cmp_psw_mat}: Is \code{x} a complete positive-whole-number atomic
matrix?

\item \code{cmp_srt_mat}: Is \code{x} a complete sortable atomic matrix?

\item \code{cmp_str_mat}: Is \code{x} a complete string atomic matrix?

\item \code{cmp_uno_mat}: Is \code{x} a complete unordered-factor atomic
matrix?

\item \code{cmp_whl_mat}: Is \code{x} a complete whole-number atomic matrix?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
