% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_mvc.R
\name{cmp_mmm_mvc}
\alias{cmp_mmm_mvc}
\alias{cmp_mmm_mvc_vals}
\alias{cmp_atm_mvc}
\alias{cmp_chr_mvc}
\alias{cmp_ch1_mvc}
\alias{cmp_clr_mvc}
\alias{cmp_evn_mvc}
\alias{cmp_fac_mvc}
\alias{cmp_frc_mvc}
\alias{cmp_ind_mvc}
\alias{cmp_lgl_mvc}
\alias{cmp_neg_mvc}
\alias{cmp_ngw_mvc}
\alias{cmp_nng_mvc}
\alias{cmp_nnw_mvc}
\alias{cmp_nps_mvc}
\alias{cmp_npw_mvc}
\alias{cmp_nst_mvc}
\alias{cmp_num_mvc}
\alias{cmp_odd_mvc}
\alias{cmp_ord_mvc}
\alias{cmp_pct_mvc}
\alias{cmp_pos_mvc}
\alias{cmp_ppn_mvc}
\alias{cmp_psw_mvc}
\alias{cmp_srt_mvc}
\alias{cmp_str_mvc}
\alias{cmp_uno_mvc}
\alias{cmp_whl_mvc}
\title{Complete + Extended Mode + Atomic Mvect (cmp + mmm + mvc)}
\usage{
cmp_mmm_mvc_vals()

cmp_atm_mvc(x)

cmp_chr_mvc(x)

cmp_ch1_mvc(x)

cmp_clr_mvc(x)

cmp_evn_mvc(x)

cmp_fac_mvc(x)

cmp_frc_mvc(x)

cmp_ind_mvc(x)

cmp_lgl_mvc(x)

cmp_neg_mvc(x)

cmp_ngw_mvc(x)

cmp_nng_mvc(x)

cmp_nnw_mvc(x)

cmp_nps_mvc(x)

cmp_npw_mvc(x)

cmp_nst_mvc(x)

cmp_num_mvc(x)

cmp_odd_mvc(x)

cmp_ord_mvc(x)

cmp_pct_mvc(x)

cmp_pos_mvc(x)

cmp_ppn_mvc(x)

cmp_psw_mvc(x)

cmp_srt_mvc(x)

cmp_str_mvc(x)

cmp_uno_mvc(x)

cmp_whl_mvc(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic mvect
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_mvc}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_mvc}: Is \code{x} a complete atomic mvect?

\item \code{cmp_chr_mvc}: Is \code{x} a complete character atomic mvect?

\item \code{cmp_ch1_mvc}: Is \code{x} a complete onechar atomic mvect?

\item \code{cmp_clr_mvc}: Is \code{x} a complete color atomic mvect?

\item \code{cmp_evn_mvc}: Is \code{x} a complete even-number atomic mvect?

\item \code{cmp_fac_mvc}: Is \code{x} a complete factor atomic mvect?

\item \code{cmp_frc_mvc}: Is \code{x} a complete fractional-number atomic
mvect?

\item \code{cmp_ind_mvc}: Is \code{x} a complete indexer atomic mvect?

\item \code{cmp_lgl_mvc}: Is \code{x} a complete logical atomic mvect?

\item \code{cmp_neg_mvc}: Is \code{x} a complete negative-number atomic mvect?

\item \code{cmp_ngw_mvc}: Is \code{x} a complete negative-whole-number atomic
mvect?

\item \code{cmp_nng_mvc}: Is \code{x} a complete non-negative-number atomic
mvect?

\item \code{cmp_nnw_mvc}: Is \code{x} a complete non-negative-whole-number
atomic mvect?

\item \code{cmp_nps_mvc}: Is \code{x} a complete non-positive-number atomic
mvect?

\item \code{cmp_npw_mvc}: Is \code{x} a complete non-positive-whole-number
atomic mvect?

\item \code{cmp_nst_mvc}: Is \code{x} a complete non-sortable atomic mvect?

\item \code{cmp_num_mvc}: Is \code{x} a complete number atomic mvect?

\item \code{cmp_odd_mvc}: Is \code{x} a complete odd-number atomic mvect?

\item \code{cmp_ord_mvc}: Is \code{x} a complete ordered-factor atomic mvect?

\item \code{cmp_pct_mvc}: Is \code{x} a complete percent-number atomic mvect?

\item \code{cmp_pos_mvc}: Is \code{x} a complete positive-number atomic mvect?

\item \code{cmp_ppn_mvc}: Is \code{x} a complete proportion-number atomic
mvect?

\item \code{cmp_psw_mvc}: Is \code{x} a complete positive-whole-number atomic
mvect?

\item \code{cmp_srt_mvc}: Is \code{x} a complete sortable atomic mvect?

\item \code{cmp_str_mvc}: Is \code{x} a complete string atomic mvect?

\item \code{cmp_uno_mvc}: Is \code{x} a complete unordered-factor atomic mvect?

\item \code{cmp_whl_mvc}: Is \code{x} a complete whole-number atomic mvect?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
