% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_vec.R
\name{cmp_mmm_vec}
\alias{cmp_mmm_vec}
\alias{cmp_mmm_vec_vals}
\alias{cmp_atm_vec}
\alias{cmp_chr_vec}
\alias{cmp_ch1_vec}
\alias{cmp_clr_vec}
\alias{cmp_evn_vec}
\alias{cmp_fac_vec}
\alias{cmp_frc_vec}
\alias{cmp_ind_vec}
\alias{cmp_lgl_vec}
\alias{cmp_neg_vec}
\alias{cmp_ngw_vec}
\alias{cmp_nng_vec}
\alias{cmp_nnw_vec}
\alias{cmp_nps_vec}
\alias{cmp_npw_vec}
\alias{cmp_nst_vec}
\alias{cmp_num_vec}
\alias{cmp_odd_vec}
\alias{cmp_ord_vec}
\alias{cmp_pct_vec}
\alias{cmp_pos_vec}
\alias{cmp_ppn_vec}
\alias{cmp_psw_vec}
\alias{cmp_srt_vec}
\alias{cmp_str_vec}
\alias{cmp_uno_vec}
\alias{cmp_whl_vec}
\title{Complete + Extended Mode + Atomic Vect (cmp + mmm + vec)}
\usage{
cmp_mmm_vec_vals()

cmp_atm_vec(x)

cmp_chr_vec(x)

cmp_ch1_vec(x)

cmp_clr_vec(x)

cmp_evn_vec(x)

cmp_fac_vec(x)

cmp_frc_vec(x)

cmp_ind_vec(x)

cmp_lgl_vec(x)

cmp_neg_vec(x)

cmp_ngw_vec(x)

cmp_nng_vec(x)

cmp_nnw_vec(x)

cmp_nps_vec(x)

cmp_npw_vec(x)

cmp_nst_vec(x)

cmp_num_vec(x)

cmp_odd_vec(x)

cmp_ord_vec(x)

cmp_pct_vec(x)

cmp_pos_vec(x)

cmp_ppn_vec(x)

cmp_psw_vec(x)

cmp_srt_vec(x)

cmp_str_vec(x)

cmp_uno_vec(x)

cmp_whl_vec(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic vect
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_vect}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_vec}: Is \code{x} a complete atomic vect?

\item \code{cmp_chr_vec}: Is \code{x} a complete character atomic vect?

\item \code{cmp_ch1_vec}: Is \code{x} a complete onechar atomic vect?

\item \code{cmp_clr_vec}: Is \code{x} a complete color  atomic vect?

\item \code{cmp_evn_vec}: Is \code{x} a complete even-numeric atomic vect?

\item \code{cmp_fac_vec}: Is \code{x} a complete factor atomic vect?

\item \code{cmp_frc_vec}: Is \code{x} a complete fractional-numeric atomic
vect?

\item \code{cmp_ind_vec}: Is \code{x} a complete indexer atomic vect?

\item \code{cmp_lgl_vec}: Is \code{x} a complete logical atomic vect?

\item \code{cmp_neg_vec}: Is \code{x} a complete negative-numeric atomic vect?

\item \code{cmp_ngw_vec}: Is \code{x} a complete negative-whole-numeric atomic
vect?

\item \code{cmp_nng_vec}: Is \code{x} a complete non-negative-numeric atomic
vect?

\item \code{cmp_nnw_vec}: Is \code{x} a complete non-negative-whole-numeric
atomic vect?

\item \code{cmp_nps_vec}: Is \code{x} a complete non-positive-numeric atomic
vect?

\item \code{cmp_npw_vec}: Is \code{x} a complete non-positive-whole-numeric
atomic vect?

\item \code{cmp_nst_vec}: Is \code{x} a complete non-sortable atomic vect?

\item \code{cmp_num_vec}: Is \code{x} a complete numeric atomic vect?

\item \code{cmp_odd_vec}: Is \code{x} a complete odd-numeric atomic vect?

\item \code{cmp_ord_vec}: Is \code{x} a complete ordered-factor atomic vect?

\item \code{cmp_pct_vec}: Is \code{x} a complete percent-numeric atomic vect?

\item \code{cmp_pos_vec}: Is \code{x} a complete positive-numeric atomic vect?

\item \code{cmp_ppn_vec}: Is \code{x} a complete proportion-numeric atomic
vect?

\item \code{cmp_psw_vec}: Is \code{x} a complete positive-whole-numeric atomic
vect?

\item \code{cmp_srt_vec}: Is \code{x} a complete sortable atomic vect?

\item \code{cmp_str_vec}: Is \code{x} a complete string atomic vect?

\item \code{cmp_uno_vec}: Is \code{x} a complete unordered factor atomic vect?

\item \code{cmp_whl_vec}: Is \code{x} a complete whole-numeric atomic vect?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
