% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color}
\alias{color}
\alias{rgba}
\alias{hsva}
\alias{blend}
\alias{red}
\alias{grn}
\alias{blu}
\alias{cyn}
\alias{mag}
\alias{vlt}
\alias{ylw}
\alias{gry}
\alias{blk}
\alias{wht}
\alias{ppl}
\alias{orn}
\alias{blu_grn}
\alias{blu_ppl}
\alias{blu_vlt}
\alias{ylw_grn}
\alias{ylw_orn}
\alias{red_orn}
\alias{red_ppl}
\alias{red_vlt}
\alias{pal_cb}
\alias{pal_swatch}
\title{Simplified Color Creation}
\usage{
color(x, lighten = 0, darken = 0, a = 1, comp = F)

rgba(r = 1, g = 1, b = 1, a = 1)

hsva(h = 1, s = 1, v = 1, a = 1)

blend(x, y, wx = 1, wy = 1, a = NA)

red(p = 1, s = 0, a = 1)

grn(p = 1, s = 0, a = 1)

blu(p = 1, s = 0, a = 1)

cyn(p = 1, s = 0, a = 1)

mag(p = 1, s = 0, a = 1)

vlt(p = 1, s = 0, a = 1)

ylw(p = 1, s = 0, a = 1)

gry(p = 0.5, a = 1)

blk(a = 1)

wht(a = 1)

ppl(p = 1, s = 0, a = 1)

orn(p = 1, s = 0, a = 1)

blu_grn(p = 1, s = 0, a = 1)

blu_ppl(p = 1, s = 0, a = 1)

blu_vlt(p = 1, s = 0, a = 1)

ylw_grn(p = 1, s = 0, a = 1)

ylw_orn(p = 1, s = 0, a = 1)

red_orn(p = 1, s = 0, a = 1)

red_ppl(p = 1, s = 0, a = 1)

red_vlt(p = 1, s = 0, a = 1)

pal_cb(ngroups, ncolors = NULL, a = 1)

pal_swatch(x)
}
\arguments{
\item{x}{Vector containing valid ℝ color representations. Recycled with
\code{y}, \code{lighten}, \code{darken}, \code{comp}, \code{a}, \code{p},
\code{s}, \code{wx}, and \code{wy} where these are valid arguments.}

\item{lighten}{A vector of proportions by which to lighten the colors given
in \code{x} toward white.}

\item{darken}{A vector of proportions by which to darken the colors given in
\code{x} toward black.}

\item{a}{A numeric vector of 1 or more alpha level in the range \code{[0, 1]}
(0 = transparent, 0.5 = translucent, 1 = opaque). \code{NA} values in
\code{a} are only applicable with \code{color(.)} and \code{blend(.)}) to
indicate keeping existing alpha values.}

\item{comp}{A logical vector indicating whether to return complementary
colors.}

\item{r, g, b}{Numeric vectors of values in the range \code{[0, 1]}
giving intensities of red, blue, and green from the RGB color space.
Recycled with each other and \code{a}.}

\item{h, s, v}{Numeric vectors of values in the range \code{[0, 1]} giving hue,
saturation, and value/brightness from the HSV colorspace. Recycled with
each other and \code{a}.}

\item{y}{Vector containing valid ℝ color representations.}

\item{wx, wy}{Weight to apply to \code{x} and \code{y} in color blending.}

\item{p, s}{Numeric vectors of 1 or more primary and secondary color
intensities (respectively) in the range \code{[0, 1]}.}

\item{na}{Logical scalar indicating whether \code{NA} counts as a valid color
alpha level.}
}
\value{
A character vector of 1 or more hexadecimal RGB + alpha color
representations in the form \code{'#RRGGBBAA'}.
}
\description{
Simple color creation with a wide variety of options (intensity,
blending, lightening, darkening, opacity), get flexible colorblind
palettes, and plot color swatches.
}
\details{
\strong{\code{color}}
\cr Takes any valid R color representation and lightens, darkens, adds
opacity levels, and/or gets the complementary color.
\cr \cr
\strong{\code{rgba}}
\cr Creates colors using red (\code{r}), green (\code{g}), blue (\code{b}),
and alpha (\code{a}) intensities between \code{0} and \code{1}, inclusive.
\cr \cr
\strong{\code{hsva}}
\cr Creates colors using hue (\code{h}), staturation (\code{s}),
value/brightness (\code{v}), and alpha (\code{a}) intensities between
\code{0} and \code{1}, inclusive.
\cr \cr
\strong{\code{blend}}
\cr Blends two colors.
\cr \cr
\strong{\code{blk}, \code{wht}, and \code{gry}}
\cr Create variations on black, white, and grey.
\cr \cr
\strong{\code{red}, \code{grn}, and \code{blu}}
\cr Create variations on RGB primary colors red, green, and blue.
\cr \cr
\strong{\code{cyn}, \code{mag}, \code{vlt}, and \code{ylw}}
\cr Create variations on CMYK primary colors cyan, magenta, violet, and
yellow.
\cr \cr
\strong{\code{prp} and \code{orn}}
\cr Create variations on purple (slightly more blue than with violet) and
orange.
\cr \cr
\strong{\code{blu_grn}, \code{blu_prp}, and \code{blu_vlt}}
\cr Create variations on traditional blue blends.
\cr \cr
\strong{\code{ylw_grn} and \code{ylw_orn}}
\cr Create variations on traditional yellow blends.
\cr \cr
\strong{\code{red_orn}, \code{red_prp}, and \code{red_vlt}}
\cr Create variations on traditional red blends.
\cr \cr
\strong{\code{pal_cb}}
\cr Get colors from a flexible colorblind palette with 2 to 20 colors.
Colors 1-10 are bright. Colors 11-20 are 50\% darkened versions of colors
1-10. The difference in brightness is maximized to allow for distinguishing
bright from darkened versions, but the differences may be insufficient for
some audiences. Using only colors 1-10 is advisable.
\cr \cr
\strong{\code{pal_swatch}}
\cr Draws swatches for a palette.
\cr\cr
The arguments \code{r}, \code{g}, \code{b}, \code{a}, \code{p}, and
\code{s} are recyclable. Functions returns a vector of colors of length
equal to the length of the longest vector passed in as an parameter.
Parameters are recycled until they are as long as they longest parameter.
An error is thrown if each parameter's length is not a whole-number divisor
of the length of the maximum-length parameter.
\cr \cr
Red, green, and blue colors that make up a color are primary (\code{p}).
Those that do not are secondary (\code{s}) and function in the following
manner:
\tabular{ll}{
\code{p = 1, s = 0}\tab The most intense color
\cr\code{p > s > 0}   \tab A lighter, muted color
\cr\code{p < 1}       \tab A darker color
\cr\code{p = s}       \tab A grey color
}
Treating \code{s} as primary and \code{p} as secondary creates complemtary
colors in the following manner:
\tabular{ll}{
\code{s = 1, p = 0}\tab The most intense complementary color
\cr\code{s > p > 0}   \tab A lighter, muted complementary color
\cr\code{s < 1}       \tab A darker color
}
}
\seealso{
Other colors: 
\code{\link{is_color}()}
}
\concept{colors}
