% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\encoding{UTF-8}
\name{color_funs}
\alias{color_funs}
\alias{color}
\alias{rgba}
\alias{hsva}
\alias{blend}
\alias{red}
\alias{grn}
\alias{blu}
\alias{cyn}
\alias{mag}
\alias{vlt}
\alias{ylw}
\alias{gry}
\alias{blk}
\alias{wht}
\alias{ppl}
\alias{orn}
\alias{blu_grn}
\alias{blu_ppl}
\alias{blu_vlt}
\alias{ylw_grn}
\alias{ylw_orn}
\alias{red_orn}
\alias{red_ppl}
\alias{red_vlt}
\alias{pal_cb}
\alias{pal_swatch}
\title{Simple color creation}
\usage{
color_funs()

color(x, lighten = 0, darken = 0, a = 1, comp = F)

rgba(r = 1, g = 1, b = 1, a = 1)

hsva(h = 1, s = 1, v = 1, a = 1)

blend(x, y, wx = 1, wy = 1, a = NA)

red(p = 1, s = 0, a = 1)

grn(p = 1, s = 0, a = 1)

blu(p = 1, s = 0, a = 1)

cyn(p = 1, s = 0, a = 1)

mag(p = 1, s = 0, a = 1)

vlt(p = 1, s = 0, a = 1)

ylw(p = 1, s = 0, a = 1)

gry(p = 0.5, a = 1)

blk(a = 1)

wht(a = 1)

ppl(p = 1, s = 0, a = 1)

orn(p = 1, s = 0, a = 1)

blu_grn(p = 1, s = 0, a = 1)

blu_ppl(p = 1, s = 0, a = 1)

blu_vlt(p = 1, s = 0, a = 1)

ylw_grn(p = 1, s = 0, a = 1)

ylw_orn(p = 1, s = 0, a = 1)

red_orn(p = 1, s = 0, a = 1)

red_ppl(p = 1, s = 0, a = 1)

red_vlt(p = 1, s = 0, a = 1)

pal_cb(ng, nc = min(ng, 20), a = 1)

pal_swatch(x)
}
\arguments{
\item{x}{An \link[=atm_vec]{atomic vec} containing valid color representations. Recycled with \verb{y, lighten, darken, comp, a, p, s, wx, wy} where these are valid arguments.}

\item{lighten}{A \link[=cmp_ppn_vec]{complete proportion numeric vec} of proportions by which to lighten the colors given in \code{x} toward white.}

\item{darken}{A complete proportion numeric vec of proportions by which to darken the colors given in \code{x} toward black.}

\item{a}{A complete proportion numeric vec of 1 or more alpha levels (\code{0} = transparent, \code{0.5} = translucent, \code{1} = opaque). \code{NA} values in \code{a} are only applicable with \code{color(.)} and \code{blend(.)}) to indicate keeping existing alpha values.}

\item{comp}{A \link[=cmp_lgl_vec]{complete logical vec} indicating whether to return complementary colors.}

\item{r, g, b}{Complete proportion numeric vecs giving intensities of red, blue, and green from the RGB color space. Recycled with each other and \code{a}.}

\item{h, s, v}{Complete proportion numeric vecs giving hue, saturation, and value/brightness from the HSV color space. Recycled√with each other and \code{a}.}

\item{y}{An atomic vec containing valid r color representations.}

\item{wx, wy}{\link[=cmp_nnw_vec]{Complete non-negative numeric vecs} giving weights to apply to \code{x} and \code{y} in color blending.}

\item{p, s}{Complete poportion numeric vecs of 1 or more primary and secondary color intensities (respectively) in the range \verb{[0, 1]}.}

\item{nc, ng}{\link[=cmp_psw_scl]{Complete positive whole-number scalars} giving the number of colors and groups, respectively.}

\item{.na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} counts as a valid color alpha level.}
}
\value{
A character vector of \code{1} or more hexadecimal RGB + alpha color representations in the form \code{'#RRGGBBAA'}.
}
\description{
Simple color creation with a wide variety of options (intensity, blending, lightening, darkening, opacity). This family of functions uses three-letter codes for common colors as follows.
}
\details{
\tabular{lll}{  \emph{Color}   \tab \emph{Color}   \tab \emph{RGB channel values of}        \cr
  \emph{code}.   \tab \emph{name}    \tab \emph{default color version}        \cr
  \code{wht}     \tab white     \tab \code{c(r = 1  , g = 1  , b = 1  )} \cr
  \code{blk}     \tab black     \tab \code{c(r = 0  , g = 0  , b = 0  )} \cr
  \code{gry}     \tab grey      \tab \code{c(r = 0.5, g = 0.5, b = 0.5)} \cr
  \code{red}     \tab red       \tab \code{c(r = 1  , g = 0  , b = 0  )} \cr
  \code{grn}     \tab green     \tab \code{c(r = 0  , g = 1  , b = 0  )} \cr
  \code{blu}     \tab blue      \tab \code{c(r = 0  , g = 0  , b = 1  )} \cr
  \code{ylw}     \tab yellow    \tab \code{c(r = 1  , g = 1  , b = 0  )} \cr
  \code{cyn}     \tab cyan      \tab \code{c(r = 0  , g = 1  , b = 1  )} \cr
  \code{mag}     \tab magenta   \tab \code{c(r = 1  , g = 0  , b = 1  )} \cr
  \code{vlt}     \tab violet    \tab \code{c(r = 1  , g = 0  , b = 1  )} \cr
  \code{ppl}     \tab purple    \tab a \code{1:2} blend of red:blue.     \cr
  \code{orn}     \tab orange    \tab a \code{33:51} blend of red:yellow.   }
}
\section{Functions}{
\itemize{
\item \code{color()}: Lighten, darken, set alpha opacity, and/or take the complement of a color argument.

\item \code{rgba()}: Creates colors using proportional red, green, blue, and alpha channel values.

\item \code{hsva()}: Creates colors using proportional hue, saturation, value, and alpha channel values.

\item \code{blend()}: Blends colors \code{x} and \code{y} using weights \code{wx} and \code{wy}, and optionally adds proportional alpha opacity \code{a}. NOTE that \code{wx} and \code{wy} are rescaled to sum to \code{1}.

\item \code{red()}: Creates variations on red based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (red), secondary colors (green and blue), and alpha opacity, respectively.

\item \code{grn()}: Creates variations on green based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (green), secondary colors (red and blue), and alpha opacity, respectively.

\item \code{blu()}: Creates variations on blue based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (blue), secondary colors (red and green), and alpha opacity, respectively.

\item \code{cyn()}: Creates variations on cyan based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (green and blue), secondary colors (red), and alpha opacity, respectively.

\item \code{mag()}: Creates variations on magenta based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (red and blue), secondary colors (green), and alpha opacity, respectively.

\item \code{vlt()}: Creates variations on violet based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (red and blue), secondary colors (green), and alpha opacity, respectively.

\item \code{ylw()}: Creates variations on cyan based on proportional weights \code{p}, \code{s}, and \code{a} for primary color (red and green), secondary colors (blue), and alpha opacity, respectively.

\item \code{gry()}: Creates variations on grey based on proportional weights \code{p} and \code{a} for black and alpha opacity, respectively.

\item \code{blk()}: Creates variations on black base on proportional weight \code{a} for alpha opacity.

\item \code{wht()}: Creates variations on white based on proportional weight \code{a} for alpha opacity.

\item \code{ppl()}: Creates variations on purple via \code{blend(red(p, s), blu(p, s), 1, 2, a)}

\item \code{orn()}: Creates variations on orange via \code{blend(red(p, s), ylw(p, s), 33, 51, a)}

\item \code{blu_grn()}: Creates variations on blue-green via \code{blend(blu(p, s), grn(p, s), a)}.

\item \code{blu_ppl()}: Creates variations on blue-purple via \code{blend(blu(p, s), ppl(p, s), a)}.

\item \code{blu_vlt()}: Creates variations on blue-violet via \code{blend(blu(p, s), vlt(p, s), a)}.

\item \code{ylw_grn()}: Creates variations on yellow-green via \code{blend(ylw(p, s), grn(p, s), a)}.

\item \code{ylw_orn()}: Creates variations on yellow-orange via \code{blend(ylw(p, s), orn(p, s), a)}.

\item \code{red_orn()}: Creates variations on red orange via \code{blend(red(p, s), orange(p, s), a)}.

\item \code{red_ppl()}: Creates variations on red-purple via \code{blend(red(p, s), ppl(p, s), a)}.

\item \code{red_vlt()}: Creates variations on red-violet via \code{blend(red(p, s), vlt(p, s), a)}.

\item \code{pal_cb()}: Generates a colorblind palette based on number of groups to be represented (\code{ng}), number of colors to include (\code{nc}) and alpha opacity values.

\item \code{pal_swatch()}: Draws a swatch of the color palette in \code{x} (a vector of character color values).

}}
\section{Recycling}{
 The only arguments not recycled are \verb{.na, nc}, and \code{ng}.
}

\section{The Arguments \code{p} and \code{s}}{
 \code{p} is used to indicate the proportion of the RGB values of the most intense version of a hue should be present in a resulting color. \code{s} is used to indicate how much of complement of each RGB value should be present in a resulting color. The table below gives a helpful heuristic for setting values of \code{p} and \code{s}:
\tabular{ll}{  \strong{Argument values}   \tab \strong{Resulting color}    \cr
  \verb{p = 1, s = 0}        \tab intense                \cr
  \verb{p > s > 0}           \tab lighter                \cr
  \verb{s < p < 1}           \tab darker                 \cr
  \code{p = s = 1}           \tab white                  \cr
  \code{p = s = 0}           \tab black                  \cr
  \code{p = s = 0.5}         \tab grey                   \cr
  \verb{s = 1, p = 0}        \tab intense, complementary \cr
  \verb{s > p > 0}           \tab lighter, complementary \cr
  \verb{p < s < 1}           \tab darker, complementary    }
a concrete numeric example is that calling

\if{html}{\out{<div class="sourceCode">}}\preformatted{   ylw(p = 0.7, s = 0.3)
}\if{html}{\out{</div>}}

starts with primary and secondary RGB values of

\if{html}{\out{<div class="sourceCode">}}\preformatted{   prm = c(r = 1, g = 1, b = 0)
   sec = c(r = 0, g = 0, b = 1)
}\if{html}{\out{</div>}}

To get final RGB values, the primary RGB values are multiplied by \code{p = 0.7}, secondary RGB values are multiple by \code{s = 0.3}, and the results are summed pairwise giving

\if{html}{\out{<div class="sourceCode">}}\preformatted{   final = c(r = 0.7, g = 0.7, b = 0.3)
}\if{html}{\out{</div>}}
}

\examples{
color("red")
color("#00FF80")
color("#FF0000AA")
color(c("red", "#00FF80", "#FF0000AA"))
rgba(1/3, 2/3, 3/3, 1/2)
hsva(1/3, 2/3, 3/3, 1/2)
rgba(1/3, 2/3, 3/3, 1/2)
blend("red", "blue", wx = 2, wy = 1, a = 0.5)
blk(a = 1 / 2)
wht(a = 1 / 2)
gry(p = 2 /3, a = 0.5)
red(p = 2 / 3, s = 1 / 3, a = 1 / 2)
grn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
blu(p = 2 / 3, s = 1 / 3, a = 1 / 2)
cyn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
mag(p = 2 / 3, s = 1 / 3, a = 1 / 2)
vlt(p = 2 / 3, s = 1 / 3, a = 1 / 2)
ylw(p = 2 / 3, s = 1 / 3, a = 1 / 2)
ppl(p = 2 / 3, s = 1 / 3, a = 1 / 2)
orn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
blu_grn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
blu_ppl(p = 2 / 3, s = 1 / 3, a = 1 / 2)
blu_vlt(p = 2 / 3, s = 1 / 3, a = 1 / 2)
ylw_grn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
ylw_orn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
red_orn(p = 2 / 3, s = 1 / 3, a = 1 / 2)
red_ppl(p = 2 / 3, s = 1 / 3, a = 1 / 2)
red_vlt(p = 2 / 3, s = 1 / 3, a = 1 / 2)
pal_cb(12)
pal_cb(3)
pal_swatch(pal_cb(12))
pal_swatch(pal_cb(3))
}
\seealso{
Other plots: 
\code{\link{gg_help}()},
\code{\link{gr}},
\code{\link{markdown_help}()},
\code{\link{pals_help}()},
\code{\link{u2u_help}()}
}
\concept{color}
\concept{plots}
