% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_prop_funs.R
\encoding{UTF-8}
\name{combo2props}
\alias{combo2props}
\title{Split a Property Combos into a Vector of Single Properties}
\usage{
combo2props(x)
}
\arguments{
\item{x}{A character scalar.}
}
\value{
Either \code{NULL} or a character vector (returns \code{NULL} when \code{x} is not of mode \code{'character'} or is of length zero).
}
\description{
Splits a single property combo (underscore separated) into its constituent single properties.
}
\seealso{
Other utils: 
\code{\link{check_valid}()},
\code{\link{compat}()},
\code{\link{meets_errs}()},
\code{\link{ox_vals}()},
\code{\link{spec2combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()}
}
\concept{utils}
