% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparable.R
\name{comparable.}
\alias{comparable.}
\alias{comparable}
\title{Are objects comparable?}
\usage{
comparable.()

comparable(..., recycle. = T)
}
\arguments{
\item{...}{An arbitrary number of arguments to be checked for comparability
with each other.}

\item{recycle.}{\link[=cmp_lgl_scl]{Complete logical scalar} indicating
whether arguments in \code{...} must be recyclable to be comparable.}
}
\value{
A logical scalar.
}
\description{
Determines whether modes of all arguments in \code{...} are
comparable (i.e., sortable and compatible with each other), meaning that
the are either all character, all logical, all numeric, or all ordered
factor with the same set of levels (in the same order).
}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
