% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparable.R
\name{comparable_uj}
\alias{comparable_uj}
\alias{comparable}
\title{Are objects comparable?}
\usage{
comparable_uj()

comparable(..., recycle. = T)
}
\arguments{
\item{...}{An arbitrary number of arguments to be checked for comparability
with each other.}

\item{recycle.}{A logical scalar indicating whether arguments in \code{...}
must be recyclable to be comparable.}
}
\value{
A logical scalar.
}
\description{
Are objects comparable?
}
\section{Functions}{
\itemize{
\item \code{comparable}: Determines whether modes of all arguments in
\code{...} are comparable (i.e., sortable and compatible with each other),
meaning that all are character, logical, numeric, or ordered factor with
the same set of levels (in the same order).
}}

\seealso{
Other props: 
\code{\link{as_mmm_uj}()},
\code{\link{ccc_uj}()},
\code{\link{cmp_ccc_uj}()},
\code{\link{cmp_mmm_ccc_uj}()},
\code{\link{cmp_mmm_uj}()},
\code{\link{compatible_uj}()},
\code{\link{ddd_uj}()},
\code{\link{eee_uj}()},
\code{\link{fff_uj}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm_uj}()},
\code{\link{meets_uj}()},
\code{\link{mmm_ccc_uj}()},
\code{\link{mmm_uj}()},
\code{\link{sss_uj}()},
\code{\link{ttt_ccc_uj}()},
\code{\link{ttt_mmm_uj}()},
\code{\link{ttt_uj}()},
\code{\link{xxx_uj}()}
}
\concept{props}
