% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{console.}
\alias{console.}
\alias{say}
\alias{cat0}
\alias{continue}
\title{Interact with the user via the console}
\usage{
console.()

say(...)

cat0(...)

continue()
}
\arguments{
\item{...}{An arbitrary number of atomic arguments to be atomized and glued
together in a message.}
}
\value{
\code{NULL} (called for side effects of updating user or pausing
execution for user review).
}
\description{
Interact with the user via the console
}
\section{Functions}{
\itemize{
\item \code{say()}: Collapse \code{...} into a character scalar and print
it to the console.

\item \code{cat0()}: Collapse \code{...} into a character scalar, prefix
and post-fix a newline, and print the result to the console.

\item \code{continue()}: Print \code{'Hit [enter] or [return] to continue'}
to the console, and pause execution until user hits one of those keys.

}}
