% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{console}
\alias{console}
\alias{say}
\alias{cat0}
\alias{continue}
\title{Interact with the user via the console}
\usage{
say(...)

cat0(...)

continue()
}
\arguments{
\item{...}{An arbitrary number of atomic arguments to be atomized and glued
together in a message.}
}
\value{
\code{NULL} (called for side effects of updating user or pausing
execution for user review).
}
\description{
Say something to the user without asking for something.
}
\section{Functions}{
\itemize{
\item \code{cat0}: Prefix a newline and postfix a newline to message.

\item \code{continue}: Pause execution until user hits \code{\[enter\]} or
\code{\[return\]}.
}}

