% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\encoding{UTF-8}
\name{constants_help}
\alias{constants_help}
\alias{.BoldGreekVals}
\alias{.BoldGreekDefs}
\alias{.BoldGREEKvals}
\alias{.BoldGREEKdefs}
\alias{.ItalicGreekVals}
\alias{.ItalicGreekDefs}
\alias{.ItalicGREEKvals}
\alias{.ItalicGREEKdefs}
\alias{.PlainGreekVals}
\alias{.PlainGreekDefs}
\alias{.PlainGREEKvals}
\alias{.PlainGREEKdefs}
\alias{.EnclosureVals}
\alias{.EnclosureDefs}
\alias{.BasicMarkdownVals}
\alias{.BasicMarkdownDefs}
\alias{.HexColorVals}
\alias{.HexColorDefs}
\alias{.OpenMarkdownColorVals}
\alias{.OpenMarkdownColorDefs}
\alias{.MiscVals}
\alias{.MiscDefs}
\alias{.UTF8vals}
\alias{.UTF8defs}
\alias{.LineTypeVals}
\alias{.LineTypeDefs}
\alias{.ShapeVals}
\alias{.ShapeDefs}
\alias{.CharSetVals}
\alias{.CharSetDefs}
\alias{.pkgVals}
\alias{.pkgDefs}
\title{Package \code{uj} atomic scalar and vector constants}
\format{
An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 21.

An object of class \code{character} of length 21.

An object of class \code{list} of length 9.

An object of class \code{character} of length 9.

An object of class \code{list} of length 53.

An object of class \code{character} of length 53.

An object of class \code{list} of length 150.

An object of class \code{character} of length 150.

An object of class \code{list} of length 18.

An object of class \code{character} of length 18.

An object of class \code{list} of length 73.

An object of class \code{character} of length 73.

An object of class \code{list} of length 1.

An object of class \code{character} of length 1.

An object of class \code{list} of length 10.

An object of class \code{character} of length 10.

An object of class \code{list} of length 14.

An object of class \code{character} of length 14.

An object of class \code{list} of length 493.

An object of class \code{character} of length 493.
}
\usage{
constants_help()

.BoldGreekVals

.BoldGreekDefs

.BoldGREEKvals

.BoldGREEKdefs

.ItalicGreekVals

.ItalicGreekDefs

.ItalicGREEKvals

.ItalicGREEKdefs

.PlainGreekVals

.PlainGreekDefs

.PlainGREEKvals

.PlainGREEKdefs

.EnclosureVals

.EnclosureDefs

.BasicMarkdownVals

.BasicMarkdownDefs

.HexColorVals

.HexColorDefs

.OpenMarkdownColorVals

.OpenMarkdownColorDefs

.MiscVals

.MiscDefs

.UTF8vals

.UTF8defs

.LineTypeVals

.LineTypeDefs

.ShapeVals

.ShapeDefs

.CharSetVals

.CharSetDefs

.pkgVals

.pkgDefs
}
\value{
Either a named list or a named character vector.
}
\description{
Lists of named scalars and/or vectors include the following lists and named character vectors defining elements of the associated lists:
\tabular{ll}{
                             \tab \strong{Name of Associated}
\cr \strong{Name of List}             \tab \strong{Definitions Vector}
\cr \code{.pkgVals}                   \tab \code{.pkgDefs}
\cr \code{.MiscVals}                  \tab \code{.MiscDefs}
\cr \code{.UTF8vals}                  \tab \code{.UTF8defs}
\cr \code{.ShapeVals}                 \tab \code{.ShapeDefs}
\cr \code{.CharSetVals}               \tab \code{.CharsetDefs}
\cr \code{.LineTypeVals}              \tab \code{.LineTypeDefs}
\cr \code{.HexColorVals}              \tab \code{.HexColorDefs}
\cr \code{.EnclosureVals}             \tab \code{.EnclosureDefs}
\cr \code{.BoldGreekVals}             \tab \code{.BoldGreekDefs}
\cr \code{.BoldGREEKvals}             \tab \code{.BoldGREEKdefs}
\cr \code{.PlainGreekVals}            \tab \code{.PlainGreekDefs}
\cr \code{.PlainGREEKvals}            \tab \code{.PlainGREEKdefs}
\cr \code{.ItalicGreekVals}           \tab \code{.ItalicGreekDefs}
\cr \code{.ItalicGREEKvals}           \tab \code{.ItalicGREEKdefs}
\cr \code{.BasicMarkdownVals}         \tab \code{.BasicMarkdownDefs}
\cr \code{.OpenMarkdownColorVals}     \tab \code{.OpenMarkdownColorDefs}
}
}
\section{Functions}{
\itemize{
\item \code{.BoldGreekVals}: Markdown-formatted bold lowercase Greek letters in UTF-8 encoding.

\item \code{.BoldGreekDefs}: Plain language definitions of \code{.BoldGreekVals} constants.

\item \code{.BoldGREEKvals}: Markdown-formatted bold uppercase Greek letters in UTF-8 encoding.

\item \code{.BoldGREEKdefs}: Plain language definitions of \code{.BoldGREEKvals} constants.

\item \code{.ItalicGreekVals}: Markdown-formatted italic lowercase Greek letters in UTF-8 encoding.

\item \code{.ItalicGreekDefs}: Plain language definitions of \code{.ItalicGreekVals} constants.

\item \code{.ItalicGREEKvals}: Markdown-formatted italic uppercase Greek letters in UTF-8 encoding.

\item \code{.ItalicGREEKdefs}: Plain language definitions of \code{.ItalicGREEKvals} constants.

\item \code{.PlainGreekVals}: Markdown-formatted plain-font lowercase Greek letters in UTF-8 encoding.

\item \code{.PlainGreekDefs}: Plain language definitions of \code{.PlainGreekVals} constants.

\item \code{.PlainGREEKvals}: Markdown-formatted plain-font uppercase Greek letters in UTF-8 encoding.

\item \code{.PlainGREEKdefs}: Plain language definitions of \code{.PlainGREEKvals} constants.

\item \code{.EnclosureVals}: UTF-8 representations of enclosing characters (quotes, parents, braces, and brackets).

\item \code{.EnclosureDefs}: Plain language definitions of \code{.EnclosureVals} constants.

\item \code{.BasicMarkdownVals}: Basic markdown-formatting constants.

\item \code{.BasicMarkdownDefs}: Plain language definitions of \code{.BasicMarkdownVals} constants.

\item \code{.HexColorVals}: Basic hexadecimal color strings and sets.

\item \code{.HexColorDefs}: Plain language definitions of \code{.HexColorVals} constants.

\item \code{.OpenMarkdownColorVals}: Markdown span opening color specification constants.

\item \code{.OpenMarkdownColorDefs}: Plain language definitions of \code{.OpenMarkdownColorVals} constants.

\item \code{.MiscVals}: Miscellaneous atomic constants.

\item \code{.MiscDefs}: Plain language definitions of \code{.MiscVals} constants

\item \code{.UTF8vals}: UTF-8 character constants.

\item \code{.UTF8defs}: Plain language definitions of \code{.UTF8vals} constants.

\item \code{.LineTypeVals}: Linetype definition palette.

\item \code{.LineTypeDefs}: Plain language definition of \code{.LineTypeVals} linetype palette constant.

\item \code{.ShapeVals}: Plotting shape palette constants as numerals for use in ggplot and standard R plots.

\item \code{.ShapeDefs}: Plain language definitions of \code{.ShapeVals} plotting shape palette constants.

\item \code{.CharSetVals}: Character set constants.

\item \code{.CharSetDefs}: Plain language definition of \code{.CharSetVals} constants.

\item \code{.pkgVals}: List of all package \code{uj} constants.

\item \code{.pkgDefs}: Plain language definition of all package \code{uj} constants.

}}
\examples{
egConstants <- function() {
  print(.LineTypeVals)
  print(.LineTypeDefs)
  print(data.frame(name = names(.MiscVals), value = uj::av(.MiscVals), definition = uj::av(.MiscDefs)))
  print(data.frame(name = names(.UTF8vals), value = uj::av(.UTF8vals), definition = uj::av(.UTF8defs)))
  print(data.frame(name = names(.EnclosureVals), value = uj::av(.EnclosureVals), definition = uj::av(.EnclosureDefs)))
  print(data.frame(name = names(.BoldGreekVals), value = uj::av(.BoldGreekVals), definition = uj::av(.BoldGreekDefs)))
  print(data.frame(name = names(.BoldGREEKvals), value = uj::av(.BoldGREEKvals), definition = uj::av(.BoldGREEKdefs)))
  print(data.frame(name = names(.PlainGreekVals), value = uj::av(.PlainGreekVals), definition = uj::av(.PlainGreekDefs)))
  print(data.frame(name = names(.PlainGREEKvals), value = uj::av(.PlainGREEKvals), definition = uj::av(.PlainGREEKdefs)))
  print(data.frame(name = names(.ItalicGreekVals), value = uj::av(.ItalicGreekVals), definition = uj::av(.ItalicGreekDefs)))
  print(data.frame(name = names(.ItalicGREEKvals), value = uj::av(.ItalicGREEKvals), definition = uj::av(.ItalicGREEKdefs)))
  print(data.frame(name = names(.BasicMarkdownVals), value = uj::av(.BasicMarkdownVals), definition = uj::av(.BasicMarkdownDefs)))
  print(data.frame(name = names(.OpenMarkdownColorVals), value = uj::av(.OpenMarkdownColorVals), definition = uj::av(.OpenMarkdownColorDefs)))
  print(uj::name_vals(.HexColorVals, .HexColorDefs))
  print(uj::name_vals(.ShapeVals, .ShapeDefs))
  print(uj::name_vals(.CharSetVals, .CharSetDefs))
  print(uj::name_vals(.pkgVals, .pkgDefs))
}

egConstants()
}
\keyword{datasets}
