% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants_uj}
\alias{constants_uj}
\alias{scl_constants}
\alias{vec_constants}
\alias{v}
\alias{l}
\title{Character scalar and atomic vector constants.}
\usage{
constants_uj()

scl_constants()

vec_constants()

v(...)

l(...)
}
\arguments{
\item{...}{unquoted, comma-separated list of the names of constants to
return. If multiple constants are specified, they are coerced into a single
atomic vector result.}
}
\value{
An atomic vector of length 1 or greater.
}
\description{
Character scalar and atomic vector constants.
}
\section{Functions}{
\itemize{
\item \code{scl_constants}: Get a tibble containing all character scalar
constants of package \code{uj}.

\item \code{vec_constants}: Get a tibble containing all vector constants of
package \code{uj}.

\item \code{v}: Get one or more character scalar constants.

\item \code{l}: Get one or more vector constants.
}}

\concept{constants}
