% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayons.R
\name{crayons}
\alias{crayons}
\alias{st_vals}
\alias{bg_vals}
\alias{fg_vals}
\alias{st_bld}
\alias{st_itl}
\alias{st_pln}
\alias{st_res}
\alias{st_und}
\alias{st}
\alias{fg_blk}
\alias{fg_blu}
\alias{fg_cyn}
\alias{fg_grn}
\alias{fg_gry}
\alias{fg_sil}
\alias{fg_mag}
\alias{fg_red}
\alias{fg_wht}
\alias{fg_ylw}
\alias{fg}
\alias{bg_blk}
\alias{bg_blu}
\alias{bg_cyn}
\alias{bg_grn}
\alias{bg_mag}
\alias{bg_red}
\alias{bg_wht}
\alias{bg_ylw}
\alias{bg}
\alias{txt}
\title{Wrappers for package \code{\link[crayon]{crayon}} functions to styled and color console text}
\usage{
st_vals(st = NULL)

bg_vals(bg = NULL)

fg_vals(fg = NULL)

st_bld(...)

st_itl(...)

st_pln(...)

st_res(...)

st_und(...)

st(st, ...)

fg_blk(...)

fg_blu(...)

fg_cyn(...)

fg_grn(...)

fg_gry(...)

fg_sil(...)

fg_mag(...)

fg_red(...)

fg_wht(...)

fg_ylw(...)

fg(fg, ...)

bg_blk(...)

bg_blu(...)

bg_cyn(...)

bg_grn(...)

bg_mag(...)

bg_red(...)

bg_wht(...)

bg_ylw(...)

bg(bg, ...)

txt(..., bg = NULL, fg = NULL, st = NULL)
}
\arguments{
\item{st}{A quoted or unquoted, case-insensitive, character scalar text style from \code{st_vals()}. May also be \code{NULL} for function \code{st_vals}.}

\item{bg}{A quoted or unquoted, case-insensitive, character scalar text background color from \code{bg_vals()}. May also be \code{NULL} for function \code{bg_vals}.}

\item{fg}{A quoted or unquoted, case-insensitive, character scalar text foreground color from \code{fg_vals()}. May also be \code{NULL} for function \code{fg_vals}.}

\item{...}{An arbitrary number of objects to be \link[=collapse_dots]{collapsed} to a character scalar to be styled.}
}
\value{
\emph{A character vector} \cr   \verb{st_vals, fg_vals, bg_vals}
\cr\cr \emph{A character scalar} \cr   All others
}
\description{
Functions in this family call \code{\link{collapse_dots}(...)} before applying styles or colors.
\tabular{rl}{
  \code{txt}   \tab Processes text using \code{st}, \code{bg}, \code{fg}.
\cr \code{st}   \tab Styles text using arg \code{st}.
\cr \code{fg}   \tab Colors text foreground using arg \code{fg}.
\cr \code{bg}   \tab Colors text background using arg \code{bg}.
\cr        \tab  
} \tabular{rl}{
  \code{st_bld}   \tab Calls \code{\link[crayon:bold]{crayon::bold}}.
\cr \code{st_itl}   \tab Calls \code{\link[crayon:italic]{crayon::italic}}.
\cr \code{st_und}   \tab Calls \code{\link[crayon:underline]{crayon::underline}}.
\cr            \tab  
\cr \code{st_pln}   \tab Calls \code{\link[crayon:reset]{crayon::reset}}.
\cr \code{st_res}   \tab (\code{crayon}'s alias for plain text)
\cr            \tab  
\cr \code{fg_blk}   \tab Calls \code{\link[crayon:black]{crayon::black}}
\cr \code{fg_blk}   \tab Calls \code{\link[crayon:black]{crayon::black}}.
\cr \code{fg_blu}   \tab Calls \code{\link[crayon:blue]{crayon::blue}}.
\cr \code{fg_cyn}   \tab Calls \code{\link[crayon:cyan]{crayon::cyan}}.
\cr \code{fg_grn}   \tab Calls \code{\link[crayon:green]{crayon::green}}.
\cr \code{fg_mag}   \tab Calls \code{\link[crayon:magenta]{crayon::magenta}}.
\cr \code{fg_red}   \tab Calls \code{\link[crayon:red]{crayon::red}}
\cr \code{fg_wht}   \tab Calls \code{\link[crayon:white]{crayon::white}}.
\cr \code{fg_ylw}   \tab Calls \code{\link[crayon:yellow]{crayon::yellow}}.
\cr            \tab  
\cr \code{fg_gry}   \tab Calls \code{\link[crayon:silver]{crayon::silver}}.
\cr \code{fg_sil}   \tab (\code{crayon}'s alias for grey)
\cr            \tab  
\cr \code{bg_blk}   \tab Calls \code{\link[crayon:bgBlack]{crayon::bgBlack}}.
\cr \code{bg_blu}   \tab Calls \code{\link[crayon:bgBlue]{crayon::bgBlue}}.
\cr \code{bg_cyn}   \tab Calls \code{\link[crayon:bgCyan]{crayon::bgCyan}}.
\cr \code{bg_grn}   \tab Calls \code{\link[crayon:bgGreen]{crayon::bgGreen}}.
\cr \code{bg_mag}   \tab Calls \code{\link[crayon:bgMagenta]{crayon::bgMagenta}}.
\cr \code{bg_red}   \tab Calls \code{\link[crayon:bgRed]{crayon::bgRed}}.
\cr \code{bg_wht}   \tab Calls \code{\link[crayon:bgWhite]{crayon::bgWhite}}.
\cr \code{bg_ylw}   \tab Calls \code{\link[crayon:bgYellow]{crayon::bgYellow}}.
\cr            \tab  
} \tabular{rl}{
  \code{st_vals}   \tab Gets all valid values of arg \code{st}.
\cr \code{bg_vals}   \tab Gets all valid values of arg \code{bg}.
\cr \code{fg_vals}   \tab Gets all valid values of arg \code{fg}.
}
See \emph{details} for style-specific and color-specific valid indicator values.
}
\details{
The following table lists all valid values of \code{bg} for allowed colors:
\tabular{rl}{
  magenta   \tab \code{'m'}, \code{'mag'}, \code{'magenta'}
\cr yellow   \tab \code{'y'}, \code{'ylw'}, \code{'yellow'}
\cr  black   \tab \code{'k'}, \code{'blk'}, \code{'black'}
\cr  white   \tab \code{'w'}, \code{'wht'}, \code{'white'}
\cr  green   \tab \code{'g'}, \code{'grn'}, \code{'green'}
\cr   blue   \tab \code{'b'}, \code{'blu'}, \code{'blue'}
\cr   cyan   \tab \code{'c'}, \code{'cyn'}, \code{'cyan'}
\cr    red   \tab \code{'r'}, \code{'red'}
}
\cr Valid values of \code{fg} are the valid values of \code{bg} plus the following indicators for grey: \code{'gry'}, \code{'gray'}, \code{'grey'}, \code{'s'}, \code{'sil'}, \code{'silver'}.
\cr\cr The following table lists all valid values of \code{st} for allowed styles:
\tabular{rl}{
  underline   \tab \code{'u'}, \code{'un'}, \code{'und'}, \code{'under'}, \code{'underline'}, \code{'underlined'}
\cr    italic   \tab \code{'i'}, \code{'it'}, \code{'itl'}, \code{'ital'}, \code{'italic'}, \code{'italics'}, \code{'italicized'}, \code{'e'}, \code{'em'}, \code{'emp'}, \code{'emph'}, \code{'emphasis'}, \code{'emphasized'}
\cr     plain   \tab \code{'p'},\code{ 'pl'}, \code{'pln'}, \code{'plain'}, \code{'r'}, \code{'re'}, \verb{'res}', \code{'reset'}
\cr      bold   \tab \code{'b'}, \code{'bo'}, \code{'bld'}, \code{'bold'}, \code{'bolded'}, \code{'s'}, \code{'st'}, \code{'str'}, \code{'strong'}
}
\cr Calling \code{st_vals(st)}, \code{bg_vals(bg)}, and \code{fg_vals(fg)} gets a character vector of all indicator values synonymous with the values in \code{st}, \code{bg} and \code{fg}, respectively.
}
