% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\name{ddd.}
\alias{ddd.}
\alias{ddd}
\alias{nddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{ddd_vals}
\alias{iddd}
\title{Defined dimensionality properties}
\usage{
ddd.()

ddd(x)

nddd(x)

id0D(x)

id1D(x)

id2D(x)

idHD(x)

ddd_vals()

iddd(x, ddd, ...)
}
\arguments{
\item{x}{An object.}

\item{ddd}{\link[cmp_chr_scl]{Complete character scalar} containing one or
more values from \code{ddd_vals()} separated by pipes and/or underscores.
Combinations of defined dimensionality properties can be specified by
separating them with underscores. Separating defined dimensionality
properties or combinations of defined dimensionality properties with pipes
will result in a value of \code{TRUE} if any of them applies to \code{x}.}

\item{...}{Additional arguments to \code{\link{meets}} containing value and
element/row/column count restrictions.}
}
\value{
\code{nddd} returns a non-negative whole-number scalar.
\code{ddd_vals} and \code{ddd} return a character scalar or vector. All
others return either \code{TRUE} or \code{FALSE}.
}
\description{
An object's defined dimensionality (def. dim.) is the number of
dimensions on which its components can be indexed. The following table
gives levels of defined dimensionality, property values assigned to levels
of defined dimensionality, property names, and a definition of each value
of defined dimensionality.\tabular{llll}{
          \tab DEFINED          \tab DEFINED          \tab                \cr
NUMBER    \tab DIMENSIONALITY   \tab DIMENSIONALITY   \tab CHARACTERISTICS\cr
OF DEFINED\tab PROPERTY         \tab PROPERTY         \tab OF QUALIFYING  \cr
DIMENSIONS   \tab VALUE         \tab NAME             \tab OBJECTS        \cr
\code{0}     \tab\code{'d0D'}   \tab 0-dimensional structure.
\tab \code{NULL}.                         \cr
\code{1}     \tab\code{'d1D'}   \tab 1-dimensional structure.
\tab Vector, \link[ivls]{vlist},
1-dimensional array.                 \cr
\code{2}     \tab\code{'d2D'}   \tab 2-dimensional structure.
\tab \code{\link[idtf]{data.frame}} or
matrix.                              \cr
\code{≥ 3}   \tab\code{'dHD'}   \tab Hyper-dimensional structure.
\tab Array with 3+ dimensions.              }
Functions related to defined dimensionality are described in the following
table:\tabular{ll}{
FUNCTION          \tab WHAT THE                                           \cr
FORMAT            \tab FUNCTION DOES                                      \cr
\code{i***}       \tab Evaluates whether an object is of the defined
dimensionality represented by \code{***}.          \cr
\code{ddd}        \tab Gets a character scalar containing the defined
dimensionality of an object.                       \cr
\code{iddd}       \tab Evaluates an object for a specific defined
dimensionality and any additional properties specified
in \code{...}.                                     \cr
\code{ddd_vals}   \tab Gets a character vector of all possible defined
dimensionality property values.                      }
}
\section{Functions}{
\itemize{
\item \code{ddd()}: Gets a vector of all defined dimensionality properties
from \code{ddd_vals()} that are applicable to \code{x}.

\item \code{nddd()}: Gets the \strong{number} of defined dimensions of
\code{x} (i.e., not a character scalar defined dimensionality property
from \code{ddd_vals}).

\item \code{id0D()}: Evaluate whether \code{x} has \code{0} defined
dimensions.

\item \code{id1D()}: Evaluate whether \code{x} has \code{1} defined dimension.

\item \code{id2D()}: Evaluate whether \code{x} has \code{2} defined
dimensions.

\item \code{idHD()}: Evaluate whether \code{x} has \code{3+} defined
dimensions.

\item \code{ddd_vals()}: Get a character vector of all possible defined
dimensionality property values.

\item \code{iddd()}: Evaluates whether at least one of the defined
dimensionality properties in \code{ddd} is applicable to \code{x}.

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
