% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd_prop_funs.R
\encoding{UTF-8}
\name{ddd_prop_funs}
\alias{ddd_prop_funs}
\alias{ddd}
\alias{ddd_funs}
\alias{ddd_props}
\alias{is_ddd_spec}
\alias{DDD}
\alias{D0D}
\alias{D1D}
\alias{D2D}
\alias{DHD}
\alias{nddd}
\title{Defined Dimensionality Droperties}
\usage{
ddd_prop_funs()

ddd(x)

ddd_funs()

ddd_props()

is_ddd_spec(spec)

DDD(x, spec, ...)

D0D(x, ...)

D1D(x, ...)

D2D(x, ...)

DHD(x, ...)

nddd(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more defined.D properties from \code{ddd_props()}. Defined.D properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a Match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \verb{ddd_props, ddd_funs, ddd}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{is_ddd_spec, DDD, \{DDD\}}
\cr\cr  \strong{A numeric scalar}   \cr\cr \code{nddd}
}
\description{
An object's defined dimensionality is the number of dimensions on which its elements can be indexed. The associated property values are:
\tabular{lll}{  \verb{'d0d', 'D0D'}   \tab \verb{0D}   \tab The \code{NULL} object.                          \cr
  \verb{'d1d', 'D1D'}   \tab \verb{1D}   \tab Vectors, \link[=VLS]{vlists}, \verb{1D} arrays.  \cr
  \verb{'d2d', 'D2D'}   \tab \verb{2D}   \tab Data.frames and matrices.                   \cr
  \verb{'dhd', 'DHD'}   \tab \code{HD}   \tab Hyper-dimensional arrays (of 3+ dimensions).  }
}
\section{Functions}{
\itemize{
\item \code{ddd()}: Lists all defined dimensionality properties possessed by \code{x}. Returns a lowercase character scalar.

\item \code{ddd_funs()}: Lists all defined dimensionality property checking functions. Returns a sorted, uppercase, character vector.

\item \code{ddd_props()}: Lists all defined dimensionality properties. Returns a sorted, lowercase, character vector.

\item \code{is_ddd_spec()}: Checks whether \code{spec} is a defined-dimensionality property spec. Returns a logical scalar. See \code{\link{ppp}} for a definition of a property spec.

\item \code{DDD()}: Checks whether \code{spec} is a defined-dimensionality property spec subject to any count or value restrictions in \code{...}. Returns a logical scalar. See \code{\link{ppp}} for a definition of a property spec.

\item \code{D0D()}: Checks \code{x} for zero defined dimensions subject to any count or value restrictions in \code{...}. Returns a logical scalar.

\item \code{D1D()}: Checks \code{x} for one defined dimensions subject to any count or value restrictions in \code{...}. Returns a logical scalar.

\item \code{D2D()}: Checks \code{x} for two defined dimensions subject to any count or value restrictions in \code{...}. Returns a logical scalar.

\item \code{DHD()}: Checks \code{x} for three or more defined dimensions subject to any count or value restrictions in \code{...}. Returns a logical scalar.

\item \code{nddd()}: Gets the defined dimensionality of \code{x}. Returns a non-negative integer scalar.

}}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.max, .maxr, .maxc}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                 \cr   \tab   \cr
  \verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                 \cr   \tab   \cr
  \verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
ddd_funs()
ddd_props()
is_ddd_spec("d1D|d2D")
nddd(matrix(1))
nddd(letters)
nddd(1)
DDD(data.frame(letters), "d2d|dhd")
D0D(NULL)
D1D(NULL)
ddd(letters)
}
