% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{dialogs_uj}
\alias{dialogs_uj}
\alias{msgbox}
\alias{dirbox}
\alias{docbox}
\alias{ansbox}
\alias{no}
\alias{msg}
\alias{okx}
\alias{dir}
\alias{doc}
\alias{yes}
\title{Dialog boxes using package \code{svDialogs}.}
\usage{
dialogs_uj()

msgbox(..., t. = "ok")

dirbox(..., def. = getwd())

docbox(...)

ansbox(x. = "Enter a value", def. = "")

no(...)

msg(x., t. = "ok")

okx(x.)

dir(type. = "file")

doc(type. = "document")

yes(...)
}
\arguments{
\item{...}{Arguments to be atomized and glued together into a message.}

\item{t.}{Character scalar type of dialog box (valid values are \code{c('ok',
'okcancel', 'yesno', 'yesnocancel'}).}

\item{def.}{Character scalar default value.}

\item{x.}{Character scalar message.}

\item{type., }{Character scalar type of directory or document to select.}
}
\value{
Character scalar
}
\description{
Launch a dialog box.
}
\section{Functions}{
\itemize{
\item \code{msgbox}: Collapse \code{...} into a character scalar message to
the user and launch a simple dialog box to give the message.

\item \code{dirbox}: Ask user to select a directory/folder.
Fail safe feature launches a message dialog box prompting the use to take
an action in the next dialog box. After user acknowledges, launches the
directory/folder selection dialog to avoid problems with the prompt not
showing up on all operating systems.

\item \code{docbox}: Ask user to select a document. Fail safe feature
launches a message dialog box prompting the use to take an action in the
next dialog box. After user acknowledges, launches the file selection
dialog to avoid problems with the prompt not showing up on all operating
systems.

\item \code{ansbox}: Ask user to input a text response.

\item \code{no}: Does user respond 'no'?

\item \code{msg}: Display a message in a dialog box, returning
the value of the button clicked.

\item \code{okx}: Display a message in a dialog box, asking
user whether to cancel (or continue).

\item \code{dir}: Ask user to select a directory.

\item \code{doc}: Ask user to select a document.

\item \code{yes}: Does user respond 'yes'?
}}

\seealso{
Other user_interaction: 
\code{\link{ask_uj}()}
}
\concept{user_interaction}
