% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\encoding{UTF-8}
\name{.choose_from}
\alias{.choose_from}
\title{Choose from a List of Options with a Given Message and a Variety of Options}
\usage{
.choose_from(
  opts,
  mssg,
  all,
  none,
  min,
  max,
  ft,
  fs,
  fun,
  stack,
  clear,
  cancel = T
)
}
\arguments{
\item{opts}{An atomic vector of options to choose from.}

\item{mssg}{Character scalar message to prompt the selection.}

\item{all}{Logical scalar indicating whether it is valid to select all options.}

\item{none}{Logical scalar indicating whether it is valid to select none of the options.}

\item{min}{Non-negative integer scalar minimum number of options to select.}

\item{max}{Positive integer scalar maximum numver of options to select.}

\item{ft}{Character scalar title formatting spec.}

\item{fs}{Character scalar subtitle formatting spec.}

\item{fun}{Character scalar name of calling function.}

\item{stack}{Character vector call stack.}

\item{clear}{Logical scalar indicating whether to clear the console before alerting the user to the selection.}

\item{cancel}{Logical scalar indicating whether cancellation of selection is valid.}
}
\value{
An atomic vector.
}
\description{
Choose from a List of Options with a Given Message and a Variety of Options
}
\examples{
\dontrun{
  uj::.choose_from(letters, "What letter?", T, N, 0, 26, "b|s|u", "w|b|p", "console", "console", F)
  uj::.choose_from(0:9    , "What number?", F, F, 1, 1 , "r|y|b", "y|r|i", "console", "console", T, cancel = F)
}

}
\seealso{
Other alert_utils: 
\code{\link{.fmt_errs}()},
\code{\link{.ok_fmt}()},
\code{\link{match_alias}()}
}
\concept{alert_utils}
