% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\encoding{UTF-8}
\name{.fmt_errs}
\alias{.fmt_errs}
\title{Validate Formatting Argument Values}
\usage{
.fmt_errs(
  st = NULL,
  bg = NULL,
  fg = NULL,
  d = " ",
  nullst = TRUE,
  nullbg = TRUE,
  nullfg = TRUE
)
}
\arguments{
\item{st}{Optional style spec.}

\item{bg}{Optional background color spec.}

\item{fg}{Optional foreground color spec.}

\item{d}{Character scalar text delimiter.}

\item{nullst}{Logical scalar indicating whether a \code{NULL} style spec is valid.}

\item{nullbg}{Logical scalar indicating whether a \code{NULL} background color spec is valid.}

\item{nullfg}{Logical scalar indicating whether a \code{NULL} foreground color spec is valid.}
}
\description{
Validate Formatting Argument Values
}
\examples{
\dontrun{
  uj::.fmt_errs(),
  uj::.fmt_errs(d = 100),
  uj::.fmt_errs(st = "", bg = "", fg = ""),
  uj::.fmt_errs(st = "q", bg = "q", fg = "q"),
  uj::.fmt_errs(nullst = F, nullbg = F, nullfg = F))
}
}
\seealso{
Other alert_utils: 
\code{\link{.choose_from}()},
\code{\link{.ok_fmt}()},
\code{\link{match_alias}()}
}
\concept{alert_utils}
