% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\encoding{UTF-8}
\name{.ok_fmt}
\alias{.ok_fmt}
\title{Validate Formatting Argument Values}
\usage{
.ok_fmt(x)
}
\arguments{
\item{x}{Character vector format spec (element 1 is style, element 2 is foreground color, element 3 is background color). May be a pipe-delimited character scalar.}
}
\description{
Validate Formatting Argument Values
}
\examples{
egOkFmt <- function() {
  base::list(blank  = uj::.ok_fmt(""),
             null   = uj::.ok_fmt(NULL),
             bad.st = uj::.ok_fmt("q|r|b"),
             ok.all = uj::.ok_fmt("y|r|b"),
             padded = uj::.ok_fmt("  yellow | red | bold  "))
}
egOkFmt()
}
\seealso{
Other alert_utils: 
\code{\link{.choose_from}()},
\code{\link{.fmt_errs}()},
\code{\link{match_alias}()}
}
\concept{alert_utils}
