% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\alias{dot}
\alias{dot_names}
\alias{named_dots}
\alias{unnamed_dots}
\title{Get named arguments from \code{...} with default values if missing,
named and unnamed arguments from \code{...}, and names of arguments in
\code{...}.}
\usage{
dots(names., defs., ...)

dot(name., def., ...)

dot_names(..., subs. = NULL, req. = T, blank. = F, u. = T)

named_dots(...)

unnamed_dots(...)
}
\arguments{
\item{names.}{\code{NULL} or an atomic scalar/vector (may include \code{NA})
values). Is split along the delimiter \code{'|'} to allow for compactness
in submitting multiple names. \code{NULL} will match an argument in
\code{...} or element of \code{defs.} with the name \code{'NULL'}.
\code{NA} values will match an argument in \code{...} or an element of
\code{defs.} with the name \code{'NA'}.}

\item{defs.}{List of default objects/values to return if the specified
arguments are not in \code{...}. Elements of \code{defs.} must be uniquely
named. If \code{defs.} is a tibble, columns with matching names are
returned.}

\item{...}{An arbitrary number of arguments.}

\item{name.}{\code{NULL} or an atomic scalar. \code{NULL} is replaced with
\code{'NULL'} and \code{NA} is replaced with \code{'NA'}.}

\item{def.}{Default object/value to return if a specified argument is not in
\code{...}. Can be, but does not need to be, a list.}

\item{subs.}{Either \code{NULL} or a character scalar/vector. If it is a
character scalar/vector, it is split using pipes \code{'|'} as a delimiter.
If there are no pipes contained in \code{names.}, it remains unchanged.
When this argument is not \code{NULL}, it is substituted for the names of
arguments in \code{...}; thus, after splitting, its length must equal the
number of arguments in \code{...}. For example, the value \code{names. =
  c("one", "two", "three|four|five")} indicates that there should be five
arguments in \code{...} and the vector \code{c('one', 'two', 'three',
  'four', 'five')} is substituted for their names.}

\item{req.}{\code{TRUE} or \code{FALSE} indicating whether names are
required.}

\item{blank.}{\code{TRUE} or \code{FALSE} indicating whether blank names are
allowed.}

\item{u.}{\code{TRUE} or \code{FALSE} indicating whether names must be
unique.}
}
\description{
Get named arguments from \code{...} with default values if missing,
named and unnamed arguments from \code{...}, and names of arguments in
\code{...}.
}
\details{
\strong{\code{dots}}
\cr Extract one or more arguments from those in \code{...} based on their
names matching values supplied in \code{names.}. If a supplied name matches
the name of an arguments in \code{...}, that argument is returned.
Otherwise, the element of \code{defs.} with a matching name is returned.
\cr\cr
\code{names.} must be an atomic scalar/vector, and is coerced to character
mode. The exceptions are \code{NULL} and \code{NA}, which are converted to
\code{'NULL'} and \code{'NA'}. Any values of \code{names.} that are reserved
words or non-valid object names are backtick quoted for matching.
\cr\cr
\code{defs.} is a named list with default values to return if there is no
matching argument for a value in \code{names.}.
\cr\cr
Each value in \code{names.} must have a matching argument in \code{...} or
a matching element in \code{defs.}.
\cr\cr
\strong{\code{dot}}
\cr A simplified version for extracting a single named argument from
\code{...} or, if a matching argument is not found, its default value.
\cr\cr
\strong{\code{dot_names}}
\cr If \code{names.} is \code{NULL}, the return value is \code{...names()},
otherwise, \code{names.} i returned. Throws an error in the following
cases
\itemize{
\item \code{0} Arguments are supplied in \code{...}, \code{names.} is not
\code{NULL} and its length is not equal to the number of arguments
in \code{...}.
\item \code{req.} is \code{TRUE}, \code{names.} is \code{NULL}, and none
of the arguments in \code{...} are named.
\item \code{names.} is \code{NULL} and either (a) \code{blank.} is
\code{TRUE} and any argument in \code{...} is either unnamed or its
name is \code{''} or (b) \code{u.} is \code{TRUE} and multiple
arguments in \code{...} share a name.
\item \code{names.} is not \code{NULL} and either (a) \code{blank.} is
\code{TRUE} and \code{names.} contains \code{''} or (b) \code{u.}
is \code{TRUE} and \code{names.} contains duplicate values.
}
\cr\cr
\strong{\code{named_args(...)}}
\cr Extracts named arguments from \code{...} as a named list (does not
include arguments named with blank strings).
\cr\cr
\strong{\code{unnamed_args(...)}}
\cr Extract unnamed arguments from \code{...} as an unnamed list (includes
any arguments named with blank strings).
}
\concept{dot_args}
