% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\encoding{UTF-8}
\name{dp}
\alias{dp}
\alias{dpagg}
\alias{dpn}
\alias{dpall}
\alias{dpgrp}
\alias{dpjoin}
\alias{dplab}
\alias{dpmod}
\title{Thin wrappers of \code{dplyr} functions}
\usage{
dpagg(.data, ..., .groups = NULL)

dpn(...)

dpall(...)

dpgrp(x, keys)

dpjoin(...)

dplab(...)

dpmod(...)
}
\arguments{
\item{x}{A data.frame}

\item{keys}{A \link[=cmp_chr_vec]{complete character vec} naming grouping variables in \code{x}.}
}
\value{
A \code{data.frame}.
}
\description{
Thin wrappers of \code{dplyr} functions
}
\details{
\tabular{ll}{  \code{dpjoin}   \tab Thinly wraps \code{\link[dplyr]{left_join}}.               \cr
  \code{dpagg}    \tab Thinly wraps \code{\link[dplyr]{summarize}}.               \cr
  \code{dpall}    \tab Thinly wraps \code{\link[dplyr]{all_of}}.                  \cr
  \code{dplab}    \tab Thinly wraps \code{\link[dplyr]{rename}}.                  \cr
  \code{dpmod}    \tab Thinly wraps \code{\link[dplyr]{mutate}}.                  \cr
  \code{dpgrp}    \tab Limitedly wraps \code{\link[dplyr]{group_by}}\eqn{^{(1)}}  \cr
  \code{dpn}      \tab Thinly wraps \code{\link[dplyr]{n}}.                         }
\tabular{l}{  \eqn{^{(1)}} Accepts as argument only a data.frame (\code{x}) and a character vector naming grouping variables (\code{keys}).}
}
\seealso{
Other wraps: 
\code{\link{basics}()},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{rd}},
\code{\link{sca}()},
\code{\link{tb}()}
}
\concept{wraps}
