% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_tools.R
\name{dplyr_tools}
\alias{dplyr_tools}
\alias{dpn}
\alias{dpall}
\alias{dpmod}
\alias{dpsum}
\alias{dpjoin}
\alias{dpcount}
\alias{dpinner}
\alias{dpungrp}
\alias{dprename}
\alias{dpgrp}
\title{Aliases and wrappers for functions in package \code{dplyr}.}
\usage{
dpn()

dpall(x)

dpmod(.data, ...)

dpsum(.data, ..., .groups = NULL)

dpjoin(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

dpcount(x, ..., wt = NULL, sort = FALSE, name = NULL)

dpinner(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

dpungrp(x, ...)

dprename(.data, ...)

dpgrp(x, keys)
}
\description{
Alias for \code{\link[dplyr]{n}} (frequency).
}
\section{Functions}{
\itemize{
\item \code{dpall}: Alias for \code{\link[dplyr]{all_of}} (strict
selection; throws error with mismatched names).

\item \code{dpmod}: Alias for \code{\link[dplyr]{mutate}} (create new
variables).

\item \code{dpsum}: Alias for \code{\link[dplyr]{summarize}}
(aggregation).

\item \code{dpjoin}: Alias for \code{\link[dplyr]{left_join}}
(merge variables from \code{y} into rows of \code{x}).

\item \code{dpcount}: Alias for \code{\link[dplyr]{count}}
(counting observations).

\item \code{dpinner}: Alias for \code{\link[dplyr]{inner_join}} (merge
variables from \code{y} into rows of \code{x} and variables from \code{x}
into rows of \code{y}).

\item \code{dpungrp}: Alias for \code{\link[dplyr]{ungroup}} (undo case
groupings).

\item \code{dprename}: Alias for \code{\link[dplyr]{rename}} (rename
variables).

\item \code{dpgrp}: Limited functionality thin wrapper for
\code{\link[dplyr]{group_by}} (group cases for group-based analyses).
}}

