% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_tools.R
\name{dt_tools}
\alias{dt_tools}
\alias{dtmk}
\alias{dtis}
\alias{dtcast}
\alias{dtmerge}
\alias{dtas}
\alias{dtie}
\alias{dtsub}
\alias{dtcols}
\alias{dtjoin}
\alias{dtrows}
\alias{dtwide}
\title{Aliases and wrappers for functions in package \code{data.table}.}
\usage{
dtmk(
  ...,
  keep.rownames = FALSE,
  check.names = FALSE,
  key = NULL,
  stringsAsFactors = FALSE
)

dtis(x)

dtcast(
  data,
  formula,
  fun.aggregate = NULL,
  ...,
  margins = NULL,
  subset = NULL,
  fill = NULL,
  value.var = guess(data)
)

dtmerge(
  x,
  y,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  all = FALSE,
  all.x = all,
  all.y = all,
  sort = TRUE,
  suffixes = c(".x", ".y"),
  no.dups = TRUE,
  allow.cartesian = getOption("datatable.allow.cartesian"),
  ...
)

dtas(x)

dtie(x)

dtsub(x, ir, ic)

dtcols(x, ic)

dtjoin(x, y, by)

dtrows(x, ir)

dtwide(x, f, v)
}
\description{
Alias for \code{\link[data.table]{data.table}}.
}
\section{Functions}{
\itemize{
\item \code{dtis}: Alias for \code{\link[data.table]{is.data.table}}.

\item \code{dtcast}: Alias for \code{\link[data.table]{dcast}} (long-to-wide
reshaping function).

\item \code{dtmerge}: Alias for \code{\link[data.table]{merge}}.

\item \code{dtas}: Thin wrapper for
\code{\link[data.table]{as.data.table}}.

\item \code{dtie}: If \code{x} is not already a data.table, coerce it to
a data.table.

\item \code{dtsub}: Extract rows and columns of a data table using indices
or names rather than the formal data.table syntax for column selection.

\item \code{dtcols}: Extract columns of a data table using indices or names
rather than the formal data.table syntax for column selection.

\item \code{dtjoin}: merge two tables by specified columns.

\item \code{dtrows}: Extract rows of a data table using indices or names.

\item \code{dtwide}: Reshape \code{x} from long to wide.
}}

