% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_xxx.R
\name{emp_xxx}
\alias{emp_xxx}
\alias{emp_xxx_vals}
\alias{emp_arr}
\alias{emp_atm}
\alias{emp_gen}
\alias{emp_mat}
\alias{emp_rcr}
\alias{emp_tib}
\alias{emp_vls}
\alias{emp_vtp}
\title{Empty + Fundamental Type or Extended Class (emp + fff or ccc)}
\usage{
emp_xxx_vals()

emp_arr(x)

emp_atm(x)

emp_gen(x)

emp_mat(x)

emp_rcr(x)

emp_tib(x)

emp_vls(x)

emp_vtp(x)
}
\arguments{
\item{x}{An object}
}
\description{
See \code{\link{xemp}}, \code{link{xatm}}, \code{link{xrcr}},
\code{link{is_vtype}}, \code{link{is.matrix}},
\code{link[tibble]{is_tibble}}, and \code{\link{is_vlist}}.
}
