% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ex.}
\alias{ex.}
\alias{ex}
\title{Extract rows, columns, and/or elements}
\usage{
ex.()

ex(x, r = NULL, c = NULL, e = NULL)
}
\arguments{
\item{x}{Any valid R object.}

\item{r}{\link[cmp_ind_vec]{Complete indexer vec} indicating elements to
extract from \code{x} of \link[id1D]{defined dimensionality 1}.}

\item{c}{\link[cmp_ind_vec]{Complete indexer vec} indicating columns to
extract from \code{x} of \link[id2D]{defined dimensionality 2}.}
}
\description{
Extract rows, columns, and/or elements
}
\section{Functions}{
\itemize{
\item \code{ex()}: Extract specific rows and/or columns of matrices and
\link[idtf]{dtfs} or extract specific elements of \link[ivec]{vecs} or
\link[ivls]{vlists}.

}}
\examples{
Mat <- matrix(1:100, nrow = 10)
Dat   <- as_tibble(Mat)
iR <- 1:10 \%in\% 1:5
iC <- 1:10 \%in\% 6:10
iE <- which(1:26 \%in\% 1:13)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, iR = 1:5, iC = 6:10)
ex(mat, r = 20)
ex(Dat, c = 6:10)
ex(Dat, e = 1:10)
ex(letters, e = 1:13)
ex(letters, e = iE)
ex(letters, r = iR, c = iC)
}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
