% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ex}
\alias{ex}
\title{Extract rows, columns, and/or elements}
\usage{
ex(x, r = NULL, c = NULL, e = NULL)
}
\arguments{
\item{x}{Any valid R object.}

\item{r, c}{\link[=cmp_ind_vec]{Complete indexer vecs} indicating rows and columns to extract from from matrix/data.frame \code{x}.}

\item{e}{A complete indexer vec indicating elements to extract from vector/vlist \code{x}.}
}
\description{
Extract rows and/or columns of matrices/data.frames or elements of vectors/\link[=ivls]{vlists}.
}
\examples{
Mat <- matrix(1:100, nrow = 10)
Dat   <- as_tibble(Mat)
iR <- 1:10 \%in\% 1:5
iC <- 1:10 \%in\% 6:10
iE <- which(1:26 \%in\% 1:13)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, iR = 1:5, iC = 6:10)
ex(mat, r = 20)
ex(Dat, c = 6:10)
ex(Dat, e = 1:10)
ex(letters, e = 1:13)
ex(letters, e = iE)
ex(letters, r = iR, c = iC)
}
