% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ex_uj}
\alias{ex_uj}
\alias{ex}
\title{Extract rows, columns, and/or elements}
\usage{
ex_uj()

ex(x, r = NULL, c = NULL, e = NULL)
}
\arguments{
\item{x}{Any valid R object.}

\item{r}{A logical or integer vector indexing rows to extract from \code{x}
of defined dimensionality 2 (\code{\link{id2D}}).}

\item{c}{A logical or integer vector indexing columns to extract from
\code{x} of defined dimensionality 2 (\code{\link{id2D}}).}

\item{e}{A logical or integer vector indexing elements to extract from
\code{x} of defined dimensionality 1 (\code{\link{id1D}}).}
}
\description{
Extract rows, columns, and/or elements
}
\section{Functions}{
\itemize{
\item \code{ex}: Extract specific rows and/or columns of matrices and
tabulars (\code{\link{itab}}) or extract specific elements of vecs or
vlists (\code{\link{ivec}} and \code{\link{ivls}}).
}}

\examples{
Mat <- matrix(1:100, nrow = 10)
Dat   <- as.tibble(MATRIX)
iR <- 1:10 \%in\% 1:5
iC <- 1:10 \%in\% 6:10
iE <- which(1:26 \%in\% 1:13)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, iR = 1:5, iC = 6:10)
ex(mat, r = 20)
ex(Dat, c = 6:10)
ex(Dat, e = 1:10)
ex(letters, e = 1:13)
ex(letters, e = iE)
ex(letters, r = iR, c = iC)
}
\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
