% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpr.R
\name{expr_uj}
\alias{expr_uj}
\alias{expr_ux}
\alias{xbold}
\alias{xrun}
\title{Build and/or evaluate expressions}
\usage{
expr_ux()

xbold(...)

xrun(...)
}
\arguments{
\item{...}{An arbitrary number of objects atomized and glued to create the
expression.}
}
\value{
Either a character scalar or the result of calling the expression.
}
\description{
Build and/or evaluate expressions
}
\section{Functions}{
\itemize{
\item \code{xbold}: Build an expression containing bolded text.

\item \code{xrun}: Build an expression and execute the resulting command.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}

Other code: 
\code{\link{callers_uj}()},
\code{\link{run_uj}()}
}
\concept{code}
\concept{strings}
