% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failsafe.R
\name{failsafe.}
\alias{failsafe.}
\alias{failsafe}
\alias{isERR}
\alias{notERR}
\alias{msgERR}
\title{Evaluate objects, flag any errors}
\usage{
failsafe.()

failsafe(x)

isERR(x)

notERR(x)

msgERR(x)
}
\arguments{
\item{x}{An object or a call to evaluate in the environment of a parent
function where the initial call was made.}
}
\value{
\code{failsafe} returns either \code{x} or an object of class
\code{'simpleError'}. \code{isERR, notERR} returns a logical scalar.
\code{msgERR} returns either \code{NULL} or a character scalar.
}
\description{
Evaluate objects, flag any errors
}
\section{Functions}{
\itemize{
\item \code{failsafe()}: Attempts to call \code{identity(x)}. If an error is
produced in doing so, returns an object of class \code{'error'} or of class
\code{'simpleError'}. If an error is not generated, returns the safely
evaluated value of \code{x}.

\item \code{isERR()}: Evaluates whether calling \code{identity(x)} produced
an error.

\item \code{notERR()}: Evaluates whether calling \code{identity(x)} did not
produce an error.

\item \code{msgERR()}: Gets any error message associated with calling
\code{identity(x)}. If there is none, returns \code{NULL}.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}

Other errs: 
\code{\link{errs.}()}

Other failsafe: 
\code{\link{is_failsafe.}()}
}
\concept{errs}
\concept{extensions}
\concept{failsafe}
