% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fff.R
\name{fff.}
\alias{fff.}
\alias{fff}
\alias{iemp}
\alias{ipnt}
\alias{ilin}
\alias{irow}
\alias{icol}
\alias{irct}
\alias{isqr}
\alias{isld}
\alias{fff_vals}
\alias{ifff}
\title{Object form properties}
\usage{
fff.()

fff(x)

iemp(x)

ipnt(x)

ilin(x)

irow(x)

icol(x)

irct(x)

isqr(x)

isld(x)

fff_vals()

ifff(x, fff, ...)
}
\arguments{
\item{x}{An object.}

\item{fff}{\link[=cmp_chr_scl]{Complete character scalar} containing one or
more values from \code{fff_vals()} separated by pipes and/or underscores.
Combinations of forms can be specified by separating them with underscores.
Separating forms or combinations of forms with pipes will result in a value
of \code{TRUE} if any of them applies to \code{x}.}

\item{...}{
  Arguments passed on to \code{\link[=meets.]{meets.}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{lll}{
\code{fff_vals} and \code{fff}\tab  \tab A character vector.            \cr
\code{ifff} and \code{i[fff]} \tab  \tab A logical scalar.                }
}
\description{
Gets a vector of form properties from \code{fff_vals()} that are
applicable to \code{x}. An object's form properties have to do with the
number of \link[=ddd]{defined dimensions} and/or \link[=eee]{effective
dimensions} plus some additional restrictions:\tabular{lll}{
FUNDAMENTAL   \tab FUNDAMENTAL   \tab FUNDAMENTAL                       \cr
TYPE VALUE    \tab TYPE NAME     \tab TYPE DEFINITION                   \cr
\code{'emp'}\tab empty         \tab Non-\code{NULL}, length = 0.        \cr
\code{'pop'}\tab populated     \tab Length > 0.                         \cr
\code{'pnt'}\tab point         \tab Containing 1 element (includes 1-by-1
data.frames).                       \cr
\code{'lin'}\tab linear        \tab Effectively 1-dimensional.          \cr
\code{'row'}\tab 1-row         \tab Row matrix or row data.frame.       \cr
\code{'col'}\tab 1-column      \tab Column matrix or column data.frame. \cr
\code{'rct'}\tab rectangular   \tab 2+ by 2+ matrix or data.frame  .    \cr
\code{'sqr'}\tab square        \tab Square atomic matrix.               \cr
\code{'sld'}\tab solid         \tab Array with 2+ index positions in 3+
dimensions.                           }
Functions related to form properties are described in the following
table:\tabular{ll}{
FUNCTION          \tab WHAT THE                                         \cr
FORMAT            \tab FUNCTION DOES                                    \cr
\code{i[fff]}       \tab Evaluates whether an object is of the form
represented by \code{***}.                       \cr
\code{fff}        \tab Gets a character scalar containing the forms of an
object.                                          \cr
\code{ifff}       \tab Evaluates an object for a specific form and any
additional properties specified in \code{...}.   \cr
\code{fff_vals}   \tab Gets a character vector of all possible form
property values.                                   }
}
\section{Specifying Additional Property Requirements}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar value all values of \code{x} must be less
than.                                                \cr
\code{le}     \tab Scalar value all values of \code{x} must be less than or
equal to.                                            \cr
\code{ge}     \tab Scalar value all values of \code{x} must be greater than
or equal to.                                         \cr
\code{gt}     \tab Scalar value all values of \code{x} must be greater
than.                                                  }
}

\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
