% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fff.R
\name{fff}
\alias{fff}
\alias{fff_vals}
\alias{xemp}
\alias{xpnt}
\alias{xlin}
\alias{xrow}
\alias{xcol}
\alias{xrct}
\alias{xsqr}
\alias{xsld}
\alias{is_fff}
\title{Form (fff) Property Family}
\usage{
fff_vals()

xemp(x)

xpnt(x)

xlin(x)

xrow(x)

xcol(x)

xrct(x)

xsqr(x)

xsld(x)

fff(x)

is_fff(x, xxx, ...)
}
\arguments{
\item{x}{An object.}

\item{xxx}{A character scalar containing one or more values from
\code{fff_vals()} separated by pipes and/or underscores (".").
Combinations of forms can be specified by separating them with
underscores. Separating forms or combinations of forms with pipes will
result in a value of \code{TRUE} if any of them applies to \code{x}.}

\item{...}{Additional arguments to \code{\link{meets}} containing value and
element/row/column count restrictions.}
}
\value{
\code{fff_vals} and \code{fff} return a character vector. All
others return either \code{TRUE} or \code{FALSE}.
}
\description{
An object's form (fff) properties have to do with the number of
\link[=ddd]{defined dimensions} and/or \link[=eee]{effective dimensions}
plus some additional restrictions.
\tabular{lll}{
\emph{Value}\tab\emph{Name}\tab\emph{Requirements}
\cr\code{'emp'}\tab empty      \tab non-\code{NULL}, length = 0.
\cr\code{'pop'}\tab populated  \tab length > 0
\cr\code{'pnt'}\tab point      \tab containing 1 element (for tibbles
this means \code{1 × 1} in
dimension).
\cr\code{'lin'}\tab linear     \tab effectively 1-dimensional.
\cr\code{'row'}\tab 1-row      \tab row matrix or tibble.
\cr\code{'col'}\tab 1-column   \tab column matrix or tibbles.
\cr\code{'rct'}\tab rectangular\tab matrix or tibble with multiple
rows and multiple columns.
\cr\code{'sqr'}\tab square     \tab square atomic matrix.
\cr\code{'sld'}\tab solid      \tab effectively hyper-dimensional.
}
}
\details{
\strong{\code{fff_vals}}
\cr Gets all valid form properties.
\cr\cr
\strong{\code{fff}}
\cr Gets a vector of form properties from \code{fff_vals()} that are
applicable to \code{x}.
\cr\cr
\strong{\code{is_fff}}
\cr Evaluates whether any property in \code{xxx} is a form property
applicable to \code{x}).
\cr\cr
\strong{\code{xfff}}
\cr Evaluates whether \code{x} has the form property represented by
\code{fff}.
\cr\cr
\strong{Additional Arguments in \code{...}}
\cr Submitting additional arguments to \code{is_fff} via \code{...}
allows for checking not just the form but whether length, number of rows,
number of columns, and element values meet flexible criteria.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
