% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_bind.R
\encoding{UTF-8}
\name{flex_rbind}
\alias{flex_rbind}
\title{Flexibly row bind two data frames}
\usage{
flex_rbind(x, y)
}
\arguments{
\item{x, y}{Atomic data.frames.}
}
\value{
A data.frame.
}
\description{
Checks for compatibility of variables with the same name in both variables. Adds \code{NA}-populated variables to each data frame as needed to allow for row binding. Variables in the resulting data.frame are sorted left to right in ascending order of variable name.
}
