% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{fmt_errs}
\alias{fmt_errs}
\title{Validate Formatting Argument Values}
\usage{
fmt_errs(
  st = NULL,
  bg = NULL,
  fg = NULL,
  d = " ",
  nullst = TRUE,
  nullbg = TRUE,
  nullfg = TRUE
)
}
\arguments{
\item{st}{Optional style spec.}

\item{bg}{Optional background color spec.}

\item{fg}{Optional foreground color spec.}

\item{d}{Character scalar text delimiter.}

\item{nullst}{Logical scalar indicating whether a \code{NULL} style spec is valid.}

\item{nullbg}{Logical scalar indicating whether a \code{NULL} background color spec is valid.}

\item{nullfg}{Logical scalar indicating whether a \code{NULL} foreground color spec is valid.}
}
\description{
Validate Formatting Argument Values
}
\examples{
fmt_errs()
fmt_errs(d = 100)
fmt_errs(st = "", bg = "", fg = "")
fmt_errs(st = "q", bg = "q", fg = "q")
fmt_errs(nullst = F, nullbg = F, nullfg = F)
}
\seealso{
Other alert_utils: 
\code{\link{choose_from}()},
\code{\link{fmt_stack}()},
\code{\link{glu}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{ssplit}()}
}
\concept{alert_utils}
