% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{fmt_stack}
\alias{fmt_stack}
\title{Conver the Call Stack to a Character Vector of Function Names}
\usage{
fmt_stack(stack)
}
\arguments{
\item{stack}{A character vector call stack}
}
\value{
A character vector
}
\description{
Conver the Call Stack to a Character Vector of Function Names
}
\seealso{
Other alert_utils: 
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{glu}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{ssplit}()}
}
\concept{alert_utils}
