% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\name{fsub.}
\alias{fsub.}
\alias{fsub}
\title{Fixed-value string substitution}
\usage{
fsub.()

fsub(x, pats, subs)
}
\arguments{
\item{x}{Atomic character object.}

\item{pats}{Character vect containing string patterns to be replaced.}

\item{subs}{Character vect containing replacement strings.}
}
\value{
Atomic character object.
}
\description{
Fixed-value string substitution
}
\section{Functions}{
\itemize{
\item \code{fsub()}: Replace strings in \code{pats} with corresponding
(possibly recycled) strings in \code{subs}.

}}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{has_pat.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}

Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
\concept{strings}
