% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\name{fsub_uj}
\alias{fsub_uj}
\alias{fsub}
\title{Fixed-value string substitution}
\usage{
fsub_uj()

fsub(x, pats, subs)
}
\arguments{
\item{x}{Atomic character object.}

\item{pats}{Character vect containing string patterns to be replaced.}

\item{subs}{Character vect containing replacement strings.}
}
\value{
Atomic character object.
}
\description{
Fixed-value string substitution
}
\section{Functions}{
\itemize{
\item \code{fsub}: Replace strings in \code{pats} with corresponding
(possibly recycled) strings in \code{subs}.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
