% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{fun_pkg_stack}
\alias{fun_pkg_stack}
\title{List the Calling Function, the Calling Package, and the Call Stack as a Vector of Function Names}
\usage{
fun_pkg_stack(fun, pkg, stack, .fun, .stack)
}
\arguments{
\item{fun, pkg}{A character scalar.}

\item{stack}{A character vector.}

\item{.fun}{An alternate function specification.}

\item{.stack}{An alternate call stack specification.}
}
\value{
A character vector.
}
\description{
List the Calling Function, the Calling Package, and the Call Stack as a Vector of Function Names
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{not_err}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}
}
\concept{utils}
