% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx.R
\name{fx_uj}
\alias{fx_uj}
\alias{f0}
\alias{f1}
\alias{fx}
\title{Robust Fork}
\usage{
fx_uj()

f0(test, yes, no)

f1(test, yes, no, na = no, err = no)

fx(test, yes, no, na = "err")
}
\arguments{
\item{test}{For \code{f0(.)}, \code{test = TRUE} or \code{test = FALSE} is
expected, but any other value of \code{test} will return \code{no}. For
\code{f1(.)}, \code{test = TRUE} or \code{test = FALSE} is also expected,
but allows for specification of how \code{test = NA} should be handled and
separate specification of how any other value of \code{test} should be
handled. For \code{fx(.)}, expects \code{test} to be a logical scalar or
vector containing no \code{NA} values.}

\item{yes}{For \code{f0(.)} and \code{f1(.)}, an object to be returned when
\code{test = TRUE}. For \code{fx(.)}, an atomic vector of length 1 or of
the same length as \code{x}.}

\item{no}{For \code{f0(.)} and \code{f1(.)}, an object to be returned when
\code{test = TRUE}. For \code{fx(.)}, an atomic vector of length 1 or of
the same length as \code{x} \link[=compatible]{compatible} with \code{yes}.}

\item{na}{For \code{f1(.)}, \code{'err'} or an object to be returned when
\code{test = NA}. For \code{fx(.)}, \code{'err'} or an atomic scalar
\link[=compatible]{compatible} with \code{yes} and \code{no}..}

\item{err}{\code{'err'} or an object to be returned when \code{test} is not
an atomic scalar in \code{c(TRUE, FALSE, NA)}.}
}
\value{
An object.
}
\description{
Robust extended functionality for \code{\link[base]{ifelse}}.
}
\section{Functions}{
\itemize{
\item \code{f0}: If \code{test = TRUE}, returns \code{yes}. If \code{test}
is anything else, returns \code{no}.

\item \code{f1}: If \code{test = TRUE}, returns \code{yes}. If \code{test =
  FALSE}, returns \code{no}. If \code{test = NA} and \code{na = 'err'},
throws an error, otherwise returns \code{na}. If \code{test} is anything
else and \code{err = 'err'}, throws an error, otherwise returns \code{err}.

\item \code{fx}: Evaluates \code{test} (which must be a logical scalar or
vector containing no \code{NA} values) and returns an object of the same
dimension as \code{test} where \code{TRUE} values are exchanged for the
associated value of \code{yes} and \code{FALSE} values are exchanged for
the associated value of \code{no}. \code{yes} and \code{no} must be
recyclable with \code{test}. \code{yes} and \code{no} must also be
\link[=compatible]{compatible} with each other.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}

Other fork: 
\code{\link{case_uj}()},
\code{\link{fork_uj.}()}
}
\concept{fork}
\concept{meta}
