% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{ggplot_tools}
\alias{ggplot_tools}
\alias{gp}
\alias{gaes}
\alias{gsv}
\alias{gvars}
\alias{ggrid}
\alias{gwrap}
\alias{gann}
\alias{gexp}
\alias{gds}
\alias{gleg}
\alias{gmar}
\alias{glabs}
\alias{gxlab}
\alias{gxlim}
\alias{gylab}
\alias{gylim}
\alias{gunit}
\alias{gnudge}
\alias{garea}
\alias{gdot}
\alias{ghln}
\alias{gjit}
\alias{glab}
\alias{glin}
\alias{grct}
\alias{gtxt}
\alias{gvln}
\alias{gacon}
\alias{gcstp}
\alias{gcman}
\alias{gdman}
\alias{gfman}
\alias{glman}
\alias{gsman}
\alias{gxcon}
\alias{gxlog}
\alias{gxsqt}
\alias{gycon}
\alias{gylog}
\alias{gysqt}
\alias{gelin}
\alias{gemkd}
\alias{genll}
\alias{gerct}
\alias{getxt}
\alias{gtbw}
\alias{gthm}
\alias{gtmin}
\title{Aliases and wraps for functions in package \code{ggplot2}.}
\usage{
gp(data = NULL, mapping = aes(), ..., environment = parent.frame())

gaes(x, y, ...)

gsv(
  filename,
  plot = last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  ...
)

gvars(...)

ggrid(
  rows = NULL,
  cols = NULL,
  scales = "fixed",
  space = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = TRUE,
  margins = FALSE,
  facets = NULL
)

gwrap(
  facets,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = TRUE,
  dir = "h",
  strip.position = "top"
)

gann(
  geom,
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xend = NULL,
  yend = NULL,
  ...,
  na.rm = FALSE
)

gexp(mult = 0, add = 0)

gds(...)

gleg(
  title = waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  keywidth = NULL,
  keyheight = NULL,
  direction = NULL,
  default.unit = "line",
  override.aes = list(),
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  reverse = FALSE,
  order = 0,
  ...
)

gmar(t = 0, r = 0, b = 0, l = 0, unit = "pt")

glabs(
  ...,
  title = waiver(),
  subtitle = waiver(),
  caption = waiver(),
  tag = waiver(),
  alt = waiver(),
  alt_insight = waiver()
)

gxlab(label)

gxlim(...)

gylab(label)

gylim(...)

gunit(x, units, data = NULL)

gnudge(x = 0, y = 0)

garea(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "stack",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  outline.type = "upper"
)

gdot(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

ghln(
  mapping = NULL,
  data = NULL,
  ...,
  yintercept,
  na.rm = FALSE,
  show.legend = NA
)

gjit(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "jitter",
  ...,
  width = NULL,
  height = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

glab(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  label.padding = unit(0.25, "lines"),
  label.r = unit(0.15, "lines"),
  label.size = 0.25,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

glin(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

grct(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

gtxt(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

gvln(
  mapping = NULL,
  data = NULL,
  ...,
  xintercept,
  na.rm = FALSE,
  show.legend = NA
)

gacon(..., range = c(0.1, 1))

gcstp(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps",
  aesthetics = "colour"
)

gcman(
  ...,
  values,
  aesthetics = "colour",
  breaks = waiver(),
  na.value = "grey50"
)

gdman(..., values, breaks = waiver(), na.value = NA)

gfman(..., values, aesthetics = "fill", breaks = waiver(), na.value = "grey50")

glman(..., values, breaks = waiver(), na.value = "blank")

gsman(..., values, breaks = waiver(), na.value = NA)

gxcon(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  n.breaks = NULL,
  labels = waiver(),
  limits = NULL,
  expand = waiver(),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  guide = waiver(),
  position = "bottom",
  sec.axis = waiver()
)

gxlog(...)

gxsqt(...)

gycon(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  n.breaks = NULL,
  labels = waiver(),
  limits = NULL,
  expand = waiver(),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  guide = waiver(),
  position = "left",
  sec.axis = waiver()
)

gylog(...)

gysqt(...)

gelin(
  colour = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE
)

gemkd(
  family = NULL,
  face = NULL,
  size = NULL,
  colour = NULL,
  fill = NULL,
  box.colour = NULL,
  linetype = NULL,
  linewidth = NULL,
  hjust = NULL,
  vjust = NULL,
  halign = NULL,
  valign = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL,
  padding = NULL,
  r = NULL,
  color = NULL,
  box.color = NULL,
  align_widths = NULL,
  align_heights = NULL,
  rotate_margins = NULL,
  debug = FALSE,
  inherit.blank = FALSE
)

genll()

gerct(
  fill = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE
)

getxt(
  family = NULL,
  face = NULL,
  colour = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  color = NULL,
  margin = NULL,
  debug = NULL,
  inherit.blank = FALSE
)

gtbw(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

gthm(
  line,
  rect,
  text,
  title,
  aspect.ratio,
  axis.title,
  axis.title.x,
  axis.title.x.top,
  axis.title.x.bottom,
  axis.title.y,
  axis.title.y.left,
  axis.title.y.right,
  axis.text,
  axis.text.x,
  axis.text.x.top,
  axis.text.x.bottom,
  axis.text.y,
  axis.text.y.left,
  axis.text.y.right,
  axis.ticks,
  axis.ticks.x,
  axis.ticks.x.top,
  axis.ticks.x.bottom,
  axis.ticks.y,
  axis.ticks.y.left,
  axis.ticks.y.right,
  axis.ticks.length,
  axis.ticks.length.x,
  axis.ticks.length.x.top,
  axis.ticks.length.x.bottom,
  axis.ticks.length.y,
  axis.ticks.length.y.left,
  axis.ticks.length.y.right,
  axis.line,
  axis.line.x,
  axis.line.x.top,
  axis.line.x.bottom,
  axis.line.y,
  axis.line.y.left,
  axis.line.y.right,
  legend.background,
  legend.margin,
  legend.spacing,
  legend.spacing.x,
  legend.spacing.y,
  legend.key,
  legend.key.size,
  legend.key.height,
  legend.key.width,
  legend.text,
  legend.text.align,
  legend.title,
  legend.title.align,
  legend.position,
  legend.direction,
  legend.justification,
  legend.box,
  legend.box.just,
  legend.box.margin,
  legend.box.background,
  legend.box.spacing,
  panel.background,
  panel.border,
  panel.spacing,
  panel.spacing.x,
  panel.spacing.y,
  panel.grid,
  panel.grid.major,
  panel.grid.minor,
  panel.grid.major.x,
  panel.grid.major.y,
  panel.grid.minor.x,
  panel.grid.minor.y,
  panel.ontop,
  plot.background,
  plot.title,
  plot.title.position,
  plot.subtitle,
  plot.caption,
  plot.caption.position,
  plot.tag,
  plot.tag.position,
  plot.margin,
  strip.background,
  strip.background.x,
  strip.background.y,
  strip.placement,
  strip.text,
  strip.text.x,
  strip.text.y,
  strip.switch.pad.grid,
  strip.switch.pad.wrap,
  ...,
  complete = FALSE,
  validate = TRUE
)

gtmin(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\description{
Alias for \code{\link[ggplot2]{ggplot}}.
}
\section{Functions}{
\itemize{
\item \code{gaes}: Alias for \code{\link[ggplot2]{aes}}.

\item \code{gsv}: Alias for \code{\link[ggplot2]{ggsave}}.

\item \code{gvars}: Alias for \code{\link[ggplot2]{vars}}.

\item \code{ggrid}: Alias for \code{\link[ggplot2]{facet_grid}}.

\item \code{gwrap}: Alias for \code{\link[ggplot2]{facet_wrap}}.

\item \code{gann}: Alias for \code{\link[ggplot2]{annotate}}.

\item \code{gexp}: Alias for \code{\link[ggplot2]{expansion}}.

\item \code{gds}: Alias for \code{\link[ggplot2]{guides}}.

\item \code{gleg}: Alias for \code{\link[ggplot2]{guide_legend}}.

\item \code{gmar}: Alias for \code{\link[ggplot2]{margin}}.

\item \code{glabs}: Alias for \code{\link[ggplot2]{labs}}.

\item \code{gxlab}: Alias for \code{\link[ggplot2]{xlab}}.

\item \code{gxlim}: Alias for \code{\link[ggplot2]{xlim}}.

\item \code{gylab}: Alias for \code{\link[ggplot2]{ylab}}.

\item \code{gylim}: Alias for \code{\link[ggplot2]{ylim}}.

\item \code{gunit}: Alias for \code{\link[ggplot2]{unit}}.

\item \code{gnudge}: Alias for \code{\link[ggplot2]{position_nudge}}.

\item \code{garea}: Alias for \code{\link[ggplot2]{geom_area}}.

\item \code{gdot}: Alias for \code{\link[ggplot2]{geom_point}}.

\item \code{ghln}: Alias for \code{\link[ggplot2]{geom_hline}}.

\item \code{gjit}: Alias for \code{\link[ggplot2]{geom_jitter}}.

\item \code{glab}: Alias for \code{\link[ggplot2]{geom_label}}.

\item \code{glin}: Alias for \code{\link[ggplot2]{geom_line}}.

\item \code{grct}: Alias for \code{\link[ggplot2]{geom_rect}}.

\item \code{gtxt}: Alias for \code{\link[ggplot2]{geom_text}}.

\item \code{gvln}: Alias for \code{\link[ggplot2]{geom_vline}}.

\item \code{gacon}: Alias for
\code{\link[ggplot2]{scale_alpha_continuous}}.

\item \code{gcstp}: Alias for \code{\link[ggplot2]{scale_color_steps}}.

\item \code{gcman}: Alias for \code{\link[ggplot2]{scale_color_manual}}.

\item \code{gdman}: Alias for \code{\link[ggplot2]{scale_shape_manual}}.

\item \code{gfman}: Alias for \code{\link[ggplot2]{scale_fill_manual}}.

\item \code{glman}: Alias for
\code{\link[ggplot2]{scale_linetype_manual}}.

\item \code{gsman}: Alias for \code{\link[ggplot2]{scale_size_manual}}.

\item \code{gxcon}: Alias for \code{\link[ggplot2]{scale_x_continuous}}.

\item \code{gxlog}: Alias for \code{\link[ggplot2]{scale_x_log10}}.

\item \code{gxsqt}: Alias for \code{\link[ggplot2]{scale_x_sqrt}}.

\item \code{gycon}: Alias for \code{\link[ggplot2]{scale_y_continuous}}.

\item \code{gylog}: Alias for \code{\link[ggplot2]{scale_y_log10}}.

\item \code{gysqt}: Alias for \code{\link[ggplot2]{scale_y_sqrt}}.

\item \code{gelin}: Alias for \code{\link[ggplot2]{element_line}}.

\item \code{gemkd}: Alias for \code{\link[ggplot2]{element_markdown}}.

\item \code{genll}: Alias for \code{\link[ggplot2]{element_blank}}.

\item \code{gerct}: Alias for \code{\link[ggplot2]{element_rect}}.

\item \code{getxt}: Alias for \code{\link[ggplot2]{element_text}}.

\item \code{gtbw}: Alias for \code{\link[ggplot2]{theme_bw}}.

\item \code{gthm}: Alias for \code{\link[ggplot2]{theme}}.

\item \code{gtmin}: Alias for \code{\link[ggplot2]{theme_minimal}}.
}}

