% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{gitPUSH}
\alias{gitPUSH}
\title{Git commands}
\usage{
gitPUSH(message, add.all = T, console = T)
}
\arguments{
\item{message}{Character scalar git commit message.}

\item{add.all}{Non-\code{NA} logical scalar indicating whether to call \verb{git add --all} before calling \verb{git commit -am "\{message\}"}.}

\item{console}{Non-\code{NA} logical scalar indicating whether to bring the console to the front upon conclusion.}
}
\value{
\code{NULL}
}
\description{
Run git commands from the console or programmatically, activating and bringing a terminal to the front, executing the command, and returning focus to the console.
}
