% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_push}
\alias{git_push}
\title{Git commands}
\usage{
git_push(Message, AddAll = T, Console = T)
}
\arguments{
\item{Message}{Character scalar git commit message.}

\item{AddAll}{Non-\code{NA} logical scalar indicating whether to call \verb{git add --all} before calling \verb{git commit -am "\{Message\}"}.}

\item{Console}{Non-\code{NA} logical scalar indicating whether to bring the console to the front upon conclusion.}
}
\value{
\code{NULL}
}
\description{
Run git commands from the console or programmatically, activating and bringing a terminal to the front, executing the command, and returning focus to the console.
}
