% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{glu}
\alias{glu}
\title{Collapse and Glue \code{...} Arguments with Space as Default Delimiter}
\usage{
glu(..., d = " ")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=g]{g}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A character scalar.
}
\description{
Collapse and Glue \code{...} Arguments with Space as Default Delimiter
}
\examples{
glu(letters)
glu(letters, 0:9, d = "")
}
\seealso{
Other alert_utils: 
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{ssplit}()}
}
\concept{alert_utils}
