% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\alias{reopen}
\title{Open or reopen an R project}
\usage{
go(...)

reopen()
}
\arguments{
\item{...}{Either nothing or a single unquoted R project name.}
}
\value{
\code{NULL}
}
\description{
Open or reopen an R project
}
\details{
\tabular{ll}{  \code{reopen}    \tab Reopens the current R project.  \cr   \tab   \cr
  \code{go}        \tab Opens another project and loads the associated library if it exists. When \code{name = NULL}, launches a document selection dialog box and prompts user to select an existing R project. Otherwise, searches for an R project with name equal to the value of \code{name} in the parent directory of the current working directory, and if there is such a project, opens that project.}
}
