% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_tools.R
\name{graphics_tools}
\alias{graphics_tools}
\alias{strh}
\alias{strw}
\title{Aliases and wraps for functions in package \code{graphics}.}
\usage{
strh(s, units = "user", cex = NULL, font = NULL, vfont = NULL, ...)

strw(s, units = "user", cex = NULL, font = NULL, vfont = NULL, ...)
}
\description{
Alias for \code{\link[graphics]{strheight}}
}
\section{Functions}{
\itemize{
\item \code{strw}: Alias for \code{\link[graphics]{strwidth}}
}}

