% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haspat.R
\name{has_pat.}
\alias{has_pat.}
\alias{has_pat}
\title{Whether a string has (contains) a substring}
\usage{
has_pat.()

has_pat(x, pat)
}
\arguments{
\item{x}{\link[chr_vec]{Character vec} of strings to search.}

\item{pat}{\link[cmp_chr_scl]{Complete character scalar} string to search
for.}
}
\value{
Logical scalar or vector.
}
\description{
Whether a string has (contains) a substring
}
\section{Functions}{
\itemize{
\item \code{has_pat()}: Evaluate whether each element of \code{x} contains the
string in \code{pat}.

}}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}
}
\concept{strings}
