% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haspat.R
\name{haspat}
\alias{haspat}
\title{Whether a string has a substring.}
\usage{
haspat(x, pat)
}
\arguments{
\item{x}{Character scalar or vector of strings to search.}

\item{pat}{Character scalar string to search for.}
}
\value{
Logical scalar or vector.
}
\description{
Whether a string has a substring.
}
