% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haspat.R
\name{haspat_uj}
\alias{haspat_uj}
\alias{haspat}
\title{Whether a string has a substring.}
\usage{
haspat_uj()

haspat(x, pat)
}
\arguments{
\item{x}{Character scalar or vector of strings to search.}

\item{pat}{Character scalar string to search for.}
}
\value{
Logical scalar or vector.
}
\description{
Whether a string has a substring.
}
\section{Functions}{
\itemize{
\item \code{haspat}: Evaluate whether each element of \code{x} contains the
string in \code{pat}.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
