% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\encoding{UTF-8}
\name{hex_dec_funs}
\alias{hex_dec_funs}
\alias{to_dec}
\alias{to_hex}
\alias{todec}
\alias{tohex}
\title{Convert non-negative whole numbers between decimal and hexadecimal}
\usage{
hex_dec_funs()

to_dec(x)

to_hex(x)

todec(x)

tohex(x)
}
\arguments{
\item{x}{A \link[=cmp_nnw]{complete non-negative whole-number object} or a \link[=cmp_chr]{complete character vec} containing only non-negative whole-number hexademical values.}
}
\value{
\strong{A non-negative whole number object}                                 \cr \verb{to_dec, todec}
\cr\cr  \strong{A character object of non-negative whole number hexadecimal values} \cr \verb{to_hex, tohex}
}
\description{
\tabular{ll}{  \verb{to_dec, todec}   \tab Converts hexadecimal \link[=cmp_chr]{complete character object} \code{x} from hexadecimal to decimal. \code{x} may be formatted as either \code{'hhh...'} or \code{'#hhh...'} where \code{h} is a placeholder for a hexidecimal digit in either upper or lower case. \cr   \tab   \cr
  \verb{to_hex, tohex}   \tab Converts \link[=cmp_psw]{complete positive whole-number} \code{x} to hexadecimal formatted as \code{'#HHH...'} where \code{H} is a placeholder for an uppercase hexadecimal digit.                                                                                        }
}
\section{Functions}{
\itemize{
\item \code{to_dec()}: Convert hexadecimal numerals to decimal numerals.

\item \code{to_hex()}: Convert decimal numerals to hexadecimal numerals (\emph{not} prefixed by \code{'#'}).

\item \code{todec()}: An alias for to_dec

\item \code{tohex()}: An aliax for to_hex

}}
\examples{
egHexDec <- function() {
  RGBA   <- col2rgb("orange", alpha = 0.5)
  HEX1   <- rgb(RGBA[1] / 255, RGBA[2] / 255, RGBA[3] / 255, RGBA[4] / 255)
  HEX1   <- c(substr(HEX1, 2, 3), substr(HEX1, 4, 5), substr(HEX1, 6, 7), substr(HEX1, 8, 9))
  DEC1   <- to_dec(HEX1)
  HEX2   <- to_hex(DEC1)
  HEX12A <- tohex(12)
  DEC12A <- todec(HEX12A)
  HEX12B <- tohex(DEC12A)
  list(RGBA = RGBA, HEX1 = HEX2, HEX12A = HEX12A, HEX12B = HEX12B, DEC1 = DEC1, DEC12A = DEC12A)
}
egHexDec <- egHexDec()
egHexDec
}
\seealso{
Other conversions: 
\code{\link{rng2lev}()},
\code{\link{tocase}()},
\code{\link{u2u}()}
}
\concept{conversions}
