% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\name{hexdec.}
\alias{hexdec.}
\alias{todec}
\alias{tohex}
\title{Convert between integer and hexadecimal}
\usage{
hexdec.()

todec(x)

tohex(x)
}
\arguments{
\item{x}{\link[psw_vec]{Positive whole number vec} or
\link[chr_vec]{character vec} of hexademical number representations.}
}
\description{
Hexademical representation is of mode character and may begin
with "#".
}
\section{Functions}{
\itemize{
\item \code{todec()}: Convert from hexadecimal to decimal.

\item \code{tohex()}: Convert from decimal to hexadecimal.

}}
\seealso{
Other math: 
\code{\link{inc.}()},
\code{\link{stats.}()}
}
\concept{math}
