% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\name{hexdec_uj}
\alias{hexdec_uj}
\alias{hexdex_uj}
\alias{todec}
\alias{tohex}
\title{Convert between integer and hexadecimal}
\usage{
hexdex_uj()

todec(x)

tohex(x)
}
\arguments{
\item{x}{Vector of positive whole numbers or strings with hexademical number
representations.}
}
\description{
Convert between integer and hexadecimal
}
\details{
Hexademical representation is of mode character and may begin with
"#".
}
\section{Functions}{
\itemize{
\item \code{todec}: Convert from hexadecimal to decimal.

\item \code{tohex}: Convert from decimal to hexadecimal.
}}

\seealso{
Other math: 
\code{\link{inc_uj}()},
\code{\link{stats_uj}()}
}
\concept{math}
