% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R
\name{inc_uj}
\alias{inc_uj}
\alias{inc}
\alias{dec}
\title{Increment or Decrement a Variable}
\usage{
inc_uj()

inc(x, i = 1, na = F)

dec(x, d = 1, na = F)
}
\arguments{
\item{x}{Any atomic numeric object.}

\item{i}{A numeric scalar giving the value by which to increment.}

\item{na}{A logical scalar indicating whether to throw an error if any values
of \code{x} are \code{NA}.}

\item{d}{A numeric scalar giving the value by which to decrement.}
}
\value{
\code{x} or \code{list(...)}, but with each element incremented or
decremented.
}
\description{
Increment or Decrement a Variable
}
\section{Functions}{
\itemize{
\item \code{inc}: Increments \code{x} by \code{i}.

\item \code{dec}: Decrements \code{x} by \code{d}.
}}

\seealso{
Other math: 
\code{\link{hexdec_uj}},
\code{\link{stats_uj}()}
}
\concept{math}
