% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\encoding{UTF-8}
\name{insert_before_pos}
\alias{insert_before_pos}
\alias{insert_before_col}
\alias{insert_before_row}
\alias{insert_after_pos}
\alias{insert_after_col}
\alias{insert_after_row}
\title{Manipulate positions, rows, and columns}
\usage{
insert_before_pos(X, V, N)

insert_before_col(X, Y, N)

insert_before_row(X, Y, N)

insert_after_pos(X, V, N)

insert_after_col(X, Y, N)

insert_after_row(X, Y, N)
}
\arguments{
\item{X}{For \code{move_pos} and \code{move_val} an atomic vector. For \code{move_row} and \code{move_col}, a data.frame or matrix.}

\item{V}{A non-\code{NA} atomic vector mode-compatible with \code{X}.}

\item{N}{A complete whole number scalar identifying a position, row, or column of \code{X}. Negative values index from the last position, row, or column rather than the first.}

\item{Y}{A matrix or data frame \code{cbind} or \code{rbind} compatible with \code{X}.}
}
\value{
An object of the same class as \code{X} with increased dimension.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \code{insert_after_pos}    \tab Insert value \code{V} after the \code{N}-the position of \code{X}.   \cr   \tab   \cr
  \code{insert_after_col}    \tab Insert \code{Y} into \code{X} after the \code{N}-th column of \code{X}    \cr   \tab   \cr
  \code{insert_after_row}    \tab Insert \code{Y} into \code{X} after the \code{N}-th row of \code{X}       \cr   \tab   \cr
  \code{insert_before_pos}   \tab Insert value \code{V} before the \code{N}-the positions of \code{X}. \cr   \tab   \cr
  \code{insert_before_col}   \tab Insert \code{Y} into \code{X} before the \code{N}-th column of \code{X}   \cr   \tab   \cr
  \code{insert_before_row}   \tab Insert \code{Y} into \code{X} before the \code{N}-th row of \code{X}                     }
\code{N < 0} index from the last position, row, or column rather than the first.
}
\seealso{
Other function_form: 
\code{\link{move_pos}()},
\code{\link{put}()}
}
\concept{function_form}
