% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\encoding{UTF-8}
\name{insert_elts_before}
\alias{insert_elts_before}
\alias{insert_cols_before}
\alias{insert_rows_before}
\alias{insert_elt_after}
\alias{insert_col_after}
\alias{insert_after_row}
\title{Manipulate positions, rows, and columns}
\usage{
insert_elts_before(X, NewElts, Elt)

insert_cols_before(X, NewCols, Col)

insert_rows_before(X, NewRows, Row)

insert_elt_after(X, NewElts, Elt)

insert_col_after(X, NewCols, Col)

insert_after_row(X, NewRows, Row)
}
\arguments{
\item{X}{For \code{insert_elt_before} and \code{insert_elt_after} a non-empty atomic vector. A non-empty atomic matrix or atomic data frame (?uj::atm_dtf) for all others.}

\item{NewElts}{An atomic vector compatible with \code{X} (?uj::compatible).}

\item{Elt}{A whole number scalar indexing a position of \code{X}. Negative values index from the last position.}

\item{NewCols}{A matrix or data.frame compatible witht \code{Y} for col binding.}

\item{Col}{A whole number scalar indexing a column of \code{X}. Negative values index from the last column.}

\item{NewRows}{A matrix or data.frame compatible with \code{X} for row binding.}

\item{Row}{A whole number scalar indexing a row of \code{X}. Negative values index from the last row.}
}
\value{
An object of the same class as \code{X} with increased dimension.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \code{insert_elts_before}   \tab Insert \code{Elts} into \code{X} before the \code{Elt}-the positions of \code{X}. \cr   \tab   \cr
  \code{insert_cols_before}   \tab Insert \code{Cols} into \code{X} before the \code{Col}-th column of \code{X}.     \cr   \tab   \cr
  \code{insert_rows_before}   \tab Insert \code{Rows} into \code{X} before the \code{Row}-th row of \code{X}.        \cr   \tab   \cr
  \code{insert_elts_after}    \tab Insert \code{Elts} into \code{X} after the \code{Elt}-the position of \code{X}.   \cr   \tab   \cr
  \code{insert_cols_after}    \tab Insert \code{Cols} into \code{X} after the \code{Col}-th column of \code{X}.      \cr   \tab   \cr
  \code{insert_rows_after}    \tab Insert \code{Rows} into \code{X} after the \code{Row}-th row of \code{X}.                        }
\code{Elt < 0}, \code{Col < 0}, and \code{Row < 0} index from the last position, row, or column rather than the first.
}
\seealso{
Other function_form: 
\code{\link{move_elt}()},
\code{\link{put}()},
\code{\link{rm_vals}()}
}
\concept{function_form}
