% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\encoding{UTF-8}
\name{insert_help}
\alias{insert_help}
\alias{insert_elts_before}
\alias{insert_cols_before}
\alias{insert_rows_before}
\alias{insert_elt_after}
\alias{insert_col_after}
\alias{insert_after_row}
\title{Manipulate positions, rows, and columns}
\usage{
insert_help()

insert_elts_before(x, new.elts, elt)

insert_cols_before(x, new.cols, col)

insert_rows_before(x, new.rows, row)

insert_elt_after(x, new.elts, elt)

insert_col_after(x, new.cols, col)

insert_after_row(x, new.rows, row)
}
\arguments{
\item{x}{For \code{insert_elt_before} and \code{insert_elt_after} a non-empty atomic vector. A non-empty atomic matrix or atomic data frame (?atm_dtf) for all others.}

\item{new.elts}{An atomic vector compatible with \code{x} (?uj::compatible).}

\item{elt}{A whole number scalar indexing a position of \code{x}. Negative values index from the last position.}

\item{new.cols}{A matrix or data.frame compatible witht \code{Y} for col binding.}

\item{col}{A whole number scalar indexing a column of \code{x}. Negative values index from the last column.}

\item{new.rows}{A matrix or data.frame compatible with \code{x} for row binding.}

\item{row}{A whole number scalar indexing a row of \code{x}. Negative values index from the last row.}
}
\value{
An object of the same class as \code{x} with increased dimension.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\code{elt < 0}, \code{col < 0}, and \code{row < 0} index from the last position, row, or column rather than the first.
}
\section{Functions}{
\itemize{
\item \code{insert_elts_before()}: Insert \code{new.elts} into \code{x} before the \code{elt}-th positions of \code{x}.

\item \code{insert_cols_before()}: Insert \code{new.cols} into \code{x} before the \code{col}-th column of \code{x}.

\item \code{insert_rows_before()}: Insert \code{new.rows} into \code{x} before the \code{row}-th row of \code{x}.

\item \code{insert_elt_after()}: Insert \code{new.elts} into \code{x} after the \code{elt}-the position of \code{x}.

\item \code{insert_col_after()}: Insert \code{new.cols} into \code{x} after the \code{col}-th column of \code{x}.

\item \code{insert_after_row()}: Insert \code{new.rows} into \code{x} after the \code{row}-th row of \code{x}.

}}
