% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipat.R
\name{ipat.}
\alias{ipat.}
\alias{has_pat}
\alias{ipat}
\alias{wpat}
\alias{npat}
\alias{ppat}
\alias{ppats}
\title{Simplified string indexing for fixed patterns}
\usage{
ipat.()

has_pat(x, pat)

ipat(x, pat)

wpat(x, pat)

npat(x, pat)

ppat(x, pat)

ppats(x, pat)
}
\arguments{
\item{x}{\link[=chr_vec]{Character vector}.}

\item{pat}{A fixed (not regular expression) \link[=cmp_chr_scl]{complete
character scalar} pattern to search for in the atomic vector resulting from
collapsing all elements of every argument supplied via \code{...}.}
}
\value{
\tabular{lll}{
\code{has_pat}, \code{ipat}          \tab   
\tab Logical vector.                                              \cr
\code{wpat}, \code{npat}, \code{ppat}\tab  
\tab \link[=whl_vec]{Whole-number vector}.                        \cr
\code{ppats}                         \tab  
\tab \link[=whl_vls]{Whole-number vlist}.                           }
}
\description{
\tabular{ll}{
\code{has_pat}   \tab Evaluate whether each element of \code{x} contains
the string in \code{pat}.                         \cr
\code{ipat}      \tab Logically indexes character vector elements
containing a pattern as \code{TRUE} or
\code{FALSE}.                                     \cr
\code{wpat}      \tab Numerically indexes character vector elements
containing a pattern as integer element numbers.  \cr
\code{npat}      \tab Counts the number of times a pattern appears in each
element of a character vector.                    \cr
\code{ppat}      \tab Identifies the position of \emph{the first} character
of the first instance of a pattern in each element of
a character vector.                               \cr
\code{ppats}     \tab Identifies the position of the first character of
\emph{every} instance of a pattern in each element of
a character vector.                                 }
}
\details{
Always \link[=a]{atomizes} \code{...} via \code{av(...)} to create a
single character vector to search for a fixed pattern.
}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()}
}
\concept{strings}
