% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipat.R
\name{ipat}
\alias{ipat}
\alias{wpat}
\alias{npat}
\alias{ppat}
\alias{ppats}
\title{Simplified string indexing for fixed patterns}
\usage{
ipat(x, pat)

wpat(x, pat)

npat(x, pat)

ppat(x, pat)

ppats(x, pat)
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A fixed (not regular expression) character scalar pattern to
search for in the atomic vector resulting from collapsing all elements of
every argument supplied via \code{...}.}
}
\value{
A logical vector (\code{ipat}), an empty or numeric vector
(\code{wpat, npat, ppat}), or a list of numeric vectors (\code{ppats}).
}
\description{
Simplified string indexing for fixed patterns
}
\details{
\strong{\code{ipat}}
\cr Logically indexes character vector elements containing a pattern as
\code{TRUE} or \code{FALSE}.
\cr\cr
\strong{\code{wpat}}
\cr Numerically indexes character vector elements containing a pattern as
integer element numbers.
\cr\cr
\strong{\code{npat}}
\cr Counts the number of times a pattern appears in each element of a
character vector.
\cr\cr
\strong{\code{ppat}}
\cr Identifies the position of \emph{the first} character of the first
instance of a pattern in each element of a character vector.
\cr\cr
\strong{\code{ppats}}
\cr Identifies the position of the first character of \emph{every} instance
of a pattern in each element of a character vector.

Always atomizes \code{...} via \code{a(...)} to create a single
character vector to search for a fixed pattern.
}
\seealso{
Other strings: 
\code{\link{Oxford}},
\code{\link{delim}},
\code{\link{pgrid}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
