% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipat.R
\name{ipat_uj}
\alias{ipat_uj}
\alias{ipat}
\alias{wpat}
\alias{npat}
\alias{ppat}
\alias{ppats}
\title{Simplified string indexing for fixed patterns}
\usage{
ipat_uj()

ipat(x, pat)

wpat(x, pat)

npat(x, pat)

ppat(x, pat)

ppats(x, pat)
}
\arguments{
\item{x}{A character vector.}

\item{pat}{A fixed (not regular expression) character scalar pattern to
search for in the atomic vector resulting from collapsing all elements of
every argument supplied via \code{...}.}
}
\value{
A logical vector (\code{ipat}), an empty or numeric vector
(\code{wpat, npat, ppat}), or a list of numeric vectors (\code{ppats}).
}
\description{
Simplified string indexing for fixed patterns
}
\details{
Always atomizes \code{...} via \code{av(...)} to create a single
character vector to search for a fixed pattern.
}
\section{Functions}{
\itemize{
\item \code{ipat}: Logically indexes character vector elements containing a
pattern as \code{TRUE} or \code{FALSE}.

\item \code{wpat}: Numerically indexes character vector elements containing
a pattern as integer element numbers.

\item \code{npat}: Counts the number of times a pattern appears in each
element of a character vector.

\item \code{ppat}: Identifies the position of \emph{the first} character of
the first instance of a pattern in each element of a character vector.

\item \code{ppats}: Identifies the position of the first character of
\emph{every} instance of a pattern in each element of a character vector.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
