% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_color.R
\name{is_color}
\alias{is_color}
\title{Is an object a vect of valid character color representations?}
\usage{
is_color(x, na = FALSE)
}
\arguments{
\item{x}{An object}

\item{na}{\code{TRUE} or \code{FALSE}, indicating whether \code{NA} values
qualify.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is an object a vect of valid character color representations?
}
\seealso{
Other colors: 
\code{\link{color}}

Other is_functions: 
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{colors}
\concept{is_functions}
