% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mmm.R
\name{is_mmm.}
\alias{is_mmm.}
\alias{is_atm}
\alias{is_chr}
\alias{is_ch1}
\alias{is_clr}
\alias{is_fac}
\alias{is_lgl}
\alias{is_num}
\alias{is_evn}
\alias{is_odd}
\alias{is_frc}
\alias{is_neg}
\alias{is_pos}
\alias{is_nps}
\alias{is_nng}
\alias{is_whl}
\alias{is_ngw}
\alias{is_nnw}
\alias{is_npw}
\alias{is_psw}
\alias{is_pct}
\alias{is_ppn}
\alias{is_ord}
\alias{is_str}
\alias{is_nst}
\alias{is_srt}
\alias{is_uno}
\alias{is_ind}
\title{Class-Agnostic Extended Atomic Modes}
\usage{
is_mmm.()

is_atm(x, na = FALSE, emp = FALSE)

is_chr(x, na = FALSE, emp = FALSE)

is_ch1(x, na = FALSE, emp = FALSE)

is_clr(x, na = FALSE, emp = FALSE)

is_fac(x, na = FALSE, emp = FALSE)

is_lgl(x, na = FALSE, emp = FALSE)

is_num(x, na = FALSE, emp = FALSE)

is_evn(x, na = FALSE, emp = FALSE)

is_odd(x, na = FALSE, emp = FALSE)

is_frc(x, na = FALSE, emp = FALSE)

is_neg(x, na = FALSE, emp = FALSE)

is_pos(x, na = FALSE, emp = FALSE)

is_nps(x, na = FALSE, emp = FALSE)

is_nng(x, na = FALSE, emp = FALSE)

is_whl(x, na = FALSE, emp = FALSE)

is_ngw(x, na = FALSE, emp = FALSE)

is_nnw(x, na = FALSE, emp = FALSE)

is_npw(x, na = FALSE, emp = FALSE)

is_psw(x, na = FALSE, emp = FALSE)

is_pct(x, na = FALSE, emp = FALSE)

is_ppn(x, na = FALSE, emp = FALSE)

is_ord(x, na = FALSE, emp = FALSE)

is_str(x)

is_nst(x, na = FALSE, emp = FALSE)

is_srt(x, na = FALSE, emp = FALSE)

is_uno(x, na = FALSE, emp = FALSE)

is_ind(x, na = FALSE, emp = FALSE)
}
\arguments{
\item{x}{An object to check for properties.}

\item{na, emp}{\link[cmp_lgl_scl]{Complete logical scalars} indicating whether
\code{NA} values are allowed and whether empty atomic objects qualify,
respectively.}
}
\value{
Logical scalar

Logical scalar.
}
\description{
This family of functions adds missing value and emptiness checks
and defines new dynamically-evaluated modes. The full set of extended modes
is given in the following tables.
\strong{Character Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                         \cr
MODE VALUE   \tab MODE NAME     \tab CHARACTERISTICS                    \cr
\code{'chr'}\tab Character      \tab Character.                         \cr
\code{'clr'}\tab Xolor          \tab Hex color value or color name.     \cr
\code{'ch1'}\tab Onechar        \tab Single characters.                 \cr
\code{'str'}\tab String         \tab No blank ("") strings                }
\strong{Categorical Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                         \cr
MODE VALUE   \tab MODE NAME     \tab CHARACTERISTICS                    \cr
\code{'fac'}\tab Factor         \tab Factor.                            \cr
\code{'lgl'}\tab Logical        \tab Logical.                           \cr
\code{'ord'}\tab Ordered        \tab Ordered factor.                    \cr
\code{'uno'}\tab Unordered      \tab Unordered factor.                    }
\strong{Combination Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                         \cr
MODE VALUE   \tab MODE NAME     \tab CHARACTERISTICS                    \cr
\code{'ind'}\tab Indexer        \tab Logical or positive whole number.  \cr
\code{'srt'}\tab Sortable       \tab Character, logical, numeric, or
ordered factor.                    \cr
\code{'nst'}\tab non-sortable   \tab Atomic, but not sortable.            }
\strong{Numeric Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                         \cr
MODE VALUE   \tab MODE NAME     \tab CHARACTERISTICS                    \cr
\code{'num'}\tab Numeric        \tab Numeric.                           \cr
\code{'frc'}\tab Fractional     \tab Fractionally numeric.              \cr
\code{'pct'}\tab Percent        \tab Percentage (in [0, 100]).        \cr
\code{'ppn'}\tab Proportion     \tab Proportion (in [0, 1]).          \cr
\code{'pos'}\tab Positive       \tab Positive numeric.                  \cr
\code{'nng'}\tab Non-negative   \tab Non-negative numeric.              \cr
\code{'nps'}\tab Non-positive   \tab Non-positive numeric.             \cr
\code{'neg'}\tab Negative       \tab Negative numeric.                  \cr
\code{'whl'}\tab Whole          \tab Whole number.                      \cr
\code{'evn'}\tab Even           \tab Even (whole) number.               \cr
\code{'odd'}\tab Odd            \tab Odd (whole) number.                \cr
\code{'psw'}\tab Positive whole \tab Positive whole-number.             \cr
\code{'nnw'}\tab Non-neg whole  \tab Non-negative whole-number.         \cr
\code{'npw'}\tab Non-pos whole  \tab Non-positive whole-number.          \cr
\code{'ngw'}\tab Negative whole \tab Negative whole-number.                 }
}
\section{Functions}{
\itemize{
\item \code{is_atm()}: Check for atomic mode.

\item \code{is_chr()}: Check for character mode.

\item \code{is_ch1()}: Check for onechar (single character valued) mode.

\item \code{is_clr()}: Check for valid character color mode.

\item \code{is_fac()}: Check for factor mode.

\item \code{is_lgl()}: Check for logical mode.

\item \code{is_num()}: Check for numeric mode.

\item \code{is_evn()}: Check for even-number valued mode.

\item \code{is_odd()}: Check for odd number valued mode.

\item \code{is_frc()}: Check for at least one fractional numeric value mode.

\item \code{is_neg()}: Check for negative numeric mode.

\item \code{is_pos()}: Check for positive numeric mode.

\item \code{is_nps()}: Check for non-positive numeric mode.

\item \code{is_nng()}: Check for non-negative numeric mode.

\item \code{is_whl()}: Check for whole-number numeric mode.

\item \code{is_ngw()}: Check for negative whole-number numeric mode.

\item \code{is_nnw()}: Check for non-negative whole-number numeric mode.

\item \code{is_npw()}: Check for non-positive whole-number numeric mode.

\item \code{is_psw()}: Check for positive whole-number numeric mode.

\item \code{is_pct()}: Check for positive whole-number numeric mode.

\item \code{is_ppn()}: Check for positive whole-number numeric mode.

\item \code{is_ord()}: Check for ordered factor mode.

\item \code{is_str()}: Check for string mode (no \code{NA} values, no blanks)..

\item \code{is_nst()}: Check for non-sortable atomic mode. Non-sortable objects
are atomic objects of modes that are not sortable. Sortable objects are
non-empty atomic objects of mode 'character', 'logical', 'numeric', or
'ordered'.

\item \code{is_srt()}: Check for sortable atomic mode. Sortable objects are
non-empty atomic objects of mode 'character', 'logical', 'numeric', or
'ordered'. Non-sortable objects are atomic objects of modes that are not
sortable.

\item \code{is_uno()}: Check for unordered factor mode.

\item \code{is_ind()}: Check for indexer mode (logical or integer).

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{is_ccc.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()},
\code{\link{xxx.}()}
}
\concept{props}
