% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mmm.R
\name{is_mmm_uj}
\alias{is_mmm_uj}
\alias{is_atm}
\alias{is_chr}
\alias{is_ch1}
\alias{is_clr}
\alias{is_fac}
\alias{is_lgl}
\alias{is_num}
\alias{is_evn}
\alias{is_odd}
\alias{is_frc}
\alias{is_neg}
\alias{is_pos}
\alias{is_nps}
\alias{is_nng}
\alias{is_whl}
\alias{is_ngw}
\alias{is_nnw}
\alias{is_npw}
\alias{is_psw}
\alias{is_pct}
\alias{is_ppn}
\alias{is_ord}
\alias{is_str}
\alias{is_nst}
\alias{is_srt}
\alias{is_uno}
\alias{is_ind}
\title{Class-Agnostic Extended Atomic Modes}
\usage{
is_mmm_uj()

is_atm(x, na = FALSE, emp = FALSE)

is_chr(x, na = FALSE, emp = FALSE)

is_ch1(x, na = FALSE, emp = FALSE)

is_clr(x, na = FALSE, emp = FALSE)

is_fac(x, na = FALSE, emp = FALSE)

is_lgl(x, na = FALSE, emp = FALSE)

is_num(x, na = FALSE, emp = FALSE)

is_evn(x, na = FALSE, emp = FALSE)

is_odd(x, na = FALSE, emp = FALSE)

is_frc(x, na = FALSE, emp = FALSE)

is_neg(x, na = FALSE, emp = FALSE)

is_pos(x, na = FALSE, emp = FALSE)

is_nps(x, na = FALSE, emp = FALSE)

is_nng(x, na = FALSE, emp = FALSE)

is_whl(x, na = FALSE, emp = FALSE)

is_ngw(x, na = FALSE, emp = FALSE)

is_nnw(x, na = FALSE, emp = FALSE)

is_npw(x, na = FALSE, emp = FALSE)

is_psw(x, na = FALSE, emp = FALSE)

is_pct(x, na = FALSE, emp = FALSE)

is_ppn(x, na = FALSE, emp = FALSE)

is_ord(x, na = FALSE, emp = FALSE)

is_str(x)

is_nst(x, na = FALSE, emp = FALSE)

is_srt(x, na = FALSE, emp = FALSE)

is_uno(x, na = FALSE, emp = FALSE)

is_ind(x, na = FALSE, emp = FALSE)
}
\arguments{
\item{x}{An object to check for properties.}

\item{na, emp}{Logical scalars indicating whether \code{NA} values are allowed
and whether empty atomic objects qualify, respectively.}
}
\value{
Logical scalar

Logical scalar.
}
\description{
This family of functions adds missing value and emptiness checks
and defines new dynamically-evaluated modes. The full set of extended modes
is given in the following tables.
\strong{Character Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                        \cr
MODE VALUE  \tab MODE NAME      \tab CHARACTERISTICS                   \cr
\code{'chr'}\tab character      \tab character                         \cr
\code{'clr'}\tab color          \tab hex color value or color name     \cr
\code{'ch1'}\tab onechar        \tab single characters                 \cr
\code{'str'}\tab string         \tab no blank ("") values                }
\strong{Categorical Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                        \cr
MODE VALUE  \tab MODE NAME      \tab CHARACTERISTICS                   \cr
\code{'fac'}\tab factor         \tab factor                            \cr
\code{'lgl'}\tab logical        \tab logical                           \cr
\code{'ord'}\tab ordered        \tab ordered factor                    \cr
\code{'uno'}\tab unordered      \tab unordered factor                    }
\strong{Combination Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                        \cr
MODE VALUE  \tab MODE NAME      \tab CHARACTERISTICS                   \cr
\code{'ind'}\tab indexer        \tab logical or positive whole number  \cr
\code{'srt'}\tab sortable       \tab character, logical, numeric, or
ordered factor                     \cr
\code{'nst'}\tab non-sortable    tab atomic, but not sortable            }
\strong{Numeric Extended Modes}\tabular{lll}{
EXTENDED    \tab EXTENDED       \tab QUALIFYING                        \cr
MODE VALUE  \tab MODE NAME      \tab CHARACTERISTICS                   \cr
\code{'num'}\tab numeric        \tab numeric                           \cr
\code{'frc'}\tab fractional     \tab fractionally valued numeric       \cr
\code{'pct'}\tab percent        \tab percentage valued (in [0, 100]) \cr
\code{'ppn'}\tab proportion     \tab proportion valued (in [0, 1])   \cr
\code{'pos'}\tab positive       \tab positive valued numeric           \cr
\code{'nng'}\tab non-negative   \tab non-negative valued numeric       \cr
\code{'nps'}\tab non-positive   \tab non-positive valued numeric       \cr
\code{'neg'}\tab negative       \tab negative valued numeric           \cr
\code{'whl'}\tab whole          \tab whole number valued               \cr
\code{'evn'}\tab even           \tab even (whole) number-valued        \cr
\code{'odd'}\tab odd            \tab odd (whole) number-valued         \cr
\code{'psw'}\tab positive whole \tab positive whole-number valued      \cr
\code{'nnw'}\tab non-neg whole  \tab non-negative whole-number valued  \cr
\code{'npw'}\tab non-pos whole  \tab non-positive whole-number valued  \cr
\code{'ngw'}\tab negative whole \tab negative whole-number valued        }
}
\section{Functions}{
\itemize{
\item \code{is_atm}: Check for atomic mode.

\item \code{is_chr}: Check for character mode.

\item \code{is_ch1}: Check for onechar (single character valued) mode.

\item \code{is_clr}: Check for valid character color mode.

\item \code{is_fac}: Check for factor mode.

\item \code{is_lgl}: Check for logical mode.

\item \code{is_num}: Check for numeric mode.

\item \code{is_evn}: Check for even-number valued mode.

\item \code{is_odd}: Check for odd number valued mode.

\item \code{is_frc}: Check for at least one fractional numeric value mode.

\item \code{is_neg}: Check for negative numeric mode.

\item \code{is_pos}: Check for positive numeric mode.

\item \code{is_nps}: Check for non-positive numeric mode.

\item \code{is_nng}: Check for non-negative numeric mode.

\item \code{is_whl}: Check for whole-number numeric mode.

\item \code{is_ngw}: Check for negative whole-number numeric mode.

\item \code{is_nnw}: Check for non-negative whole-number numeric mode.

\item \code{is_npw}: Check for non-positive whole-number numeric mode.

\item \code{is_psw}: Check for positive whole-number numeric mode.

\item \code{is_pct}: Check for positive whole-number numeric mode.

\item \code{is_ppn}: Check for positive whole-number numeric mode.

\item \code{is_ord}: Check for ordered factor mode.

\item \code{is_str}: Check for string mode (no \code{NA} values, no blanks)..

\item \code{is_nst}: Check for non-sortable atomic mode. Non-sortable
objects are atomic objects of modes that are not sortable. Sortable objects
are non-empty atomic objects of mode 'character', 'logical', 'numeric', or
'ordered'.

\item \code{is_srt}: Check for sortable atomic mode. Sortable objects are
non-empty atomic objects of mode 'character', 'logical', 'numeric', or
'ordered'. Non-sortable objects are atomic objects of modes that are not
sortable.

\item \code{is_uno}: Check for unordered factor mode.

\item \code{is_ind}: Check for indexer mode (logical or integer).
}}

\seealso{
Other props: 
\code{\link{as_mmm_uj}()},
\code{\link{ccc_uj}()},
\code{\link{cmp_ccc_uj}()},
\code{\link{cmp_mmm_ccc_uj}()},
\code{\link{cmp_mmm_uj}()},
\code{\link{comparable_uj}()},
\code{\link{compatible_uj}()},
\code{\link{ddd_uj}()},
\code{\link{eee_uj}()},
\code{\link{fff_uj}()},
\code{\link{is_ccc_uj}()},
\code{\link{meets_uj}()},
\code{\link{mmm_ccc_uj}()},
\code{\link{mmm_uj}()},
\code{\link{sss_uj}()},
\code{\link{ttt_ccc_uj}()},
\code{\link{ttt_mmm_uj}()},
\code{\link{ttt_uj}()},
\code{\link{xxx_uj}()}
}
\concept{props}
