% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\encoding{UTF-8}
\name{is_unique}
\alias{is_unique}
\alias{is_unq}
\alias{isUNQ}
\title{Does an atomic object only contain unique values?}
\usage{
is_unique(x, a = T, nas = F)

is_unq(x, a = T, nas = F)

isUNQ(x, a = T, nas = F)
}
\arguments{
\item{x}{An atomic object.}

\item{a}{\code{TRUE} or \code{FALSE} indicating whether to reduce \code{x} to an atomic vector containing all of its atomic values. When \code{FALSE} and \code{x} is not atomic, throws an error.}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values are allowed.}
}
\value{
Scalar \code{TRUE} or scalar \code{FALSE}.
}
\description{
Evaluates whether \code{x} contains only only unique atomic values, with options for for atomizing before evaluation and for handling \code{NA} values.
}
\examples{
is_unique(letters)
is_unq(sample(letters, 27, replace = T))
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{evals}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{nis}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}

Other properties: 
\code{\link{asMODE}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{extensions}
\concept{properties}
