% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unq.}
\alias{is_unq.}
\alias{is_unq}
\title{Does an object have only unique atomic values?}
\usage{
is_unq.()

is_unq(x, a = T, na = F)
}
\arguments{
\item{x}{An atomic object.}

\item{a}{A non-\code{NA} logical scalar indicating whether to reduce \code{x}
to an atomic vector containing all of its atomic values. When \code{FALSE}
and \code{x} is not atomic, throws an error.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values
are allowed.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Evaluates whether \code{x} contains only only unique atomic
values, with options for for atomizing before evaluation and for handling
\code{NA} values.
}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
