% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unq.R
\encoding{UTF-8}
\name{is_unq}
\alias{is_unq}
\title{Evaluate whether an object has only unique atomic values}
\usage{
is_unq(x, a = T, na = F)
}
\arguments{
\item{x}{An atomic object.}

\item{a}{A non-\code{NA} logical scalar indicating whether to reduce \code{x} to an atomic vector containing all of its atomic values. When \code{FALSE} and \code{x} is not atomic, throws an error.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values are allowed.}
}
\value{
Scalar \code{TRUE} or scalar \code{FALSE}.
}
\description{
Evaluates whether \code{x} contains only only unique atomic values, with options for for atomizing before evaluation and for handling \code{NA} values.
}
\examples{
is_unq(letters)
is_unq(sample(letters, 27, replace = T))
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()},
\code{\link{x}()}

Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{reclass}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{extensions}
\concept{properties}
