% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx.R
\docType{data}
\name{is_xxx}
\alias{is_xxx}
\alias{is_atm}
\alias{is_chr}
\alias{is_ch1}
\alias{is_clr}
\alias{is_fac}
\alias{is_lgl}
\alias{is_num}
\alias{is_evn}
\alias{is_odd}
\alias{is_frc}
\alias{is_neg}
\alias{is_pos}
\alias{is_nps}
\alias{is_nng}
\alias{is_whl}
\alias{is_ngw}
\alias{is_nnw}
\alias{is_npw}
\alias{is_psw}
\alias{is_pct}
\alias{is_ppn}
\alias{is_ord}
\alias{is_str}
\alias{is_nst}
\alias{is_srt}
\alias{is_uno}
\alias{is_ind}
\title{Three-character is_xxx aliases for basic atomic is.mode functions.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
is_xxx

is_atm(x, na = FALSE, emp = FALSE)

is_chr(x, na = FALSE, emp = FALSE)

is_ch1(x, na = FALSE, emp = FALSE)

is_clr(x, na = FALSE, emp = FALSE)

is_fac(x, na = FALSE, emp = FALSE)

is_lgl(x, na = FALSE, emp = FALSE)

is_num(x, na = FALSE, emp = FALSE)

is_evn(x, na = FALSE, emp = FALSE)

is_odd(x, na = FALSE, emp = FALSE)

is_frc(x, na = FALSE, emp = FALSE)

is_neg(x, na = FALSE, emp = FALSE)

is_pos(x, na = FALSE, emp = FALSE)

is_nps(x, na = FALSE, emp = FALSE)

is_nng(x, na = FALSE, emp = FALSE)

is_whl(x, na = FALSE, emp = FALSE)

is_ngw(x, na = FALSE, emp = FALSE)

is_nnw(x, na = FALSE, emp = FALSE)

is_npw(x, na = FALSE, emp = FALSE)

is_psw(x, na = FALSE, emp = FALSE)

is_pct(x, na = FALSE, emp = FALSE)

is_ppn(x, na = FALSE, emp = FALSE)

is_ord(x, na = FALSE, emp = FALSE)

is_str(x)

is_nst(x, na = FALSE, emp = FALSE)

is_srt(x, na = FALSE, emp = FALSE)

is_uno(x, na = FALSE, emp = FALSE)

is_ind(x, na = FALSE, emp = FALSE)
}
\arguments{
\item{x}{An object to check for properties.}

\item{na, emp}{Logical scalars indicating whether \code{NA} values are allowed
and whether empty atomic objects qualify, respectively.}
}
\value{
Logical scalar

Logical scalar.
}
\description{
Check for 3-character object property.
}
\details{
This family of functions adds missing value and emptiness checks and
defines new dynamically-evaluated modes, including one-char
(single-character valued), colors (valid character color values), unordered
factor, sortable, and non-sortable,
}
\section{Functions}{
\itemize{
\item \code{is_atm}: Check for atomic mode.

\item \code{is_chr}: Check for character mode.

\item \code{is_ch1}: Check for onechar (single character valued) mode.

\item \code{is_clr}: Check for valid character color mode.

\item \code{is_fac}: Check for factor mode.

\item \code{is_lgl}: Check for logical mode.

\item \code{is_num}: Check for numeric mode.

\item \code{is_evn}: Check for even-number valued mode.

\item \code{is_odd}: Check for odd number valued mode.

\item \code{is_frc}: Check for at least one fractional numeric value mode.

\item \code{is_neg}: Check for negative numeric mode.

\item \code{is_pos}: Check for positive numeric mode.

\item \code{is_nps}: Check for non-positive numeric mode.

\item \code{is_nng}: Check for non-negative numeric mode.

\item \code{is_whl}: Check for whole-number numeric mode.

\item \code{is_ngw}: Check for negative whole-number numeric mode.

\item \code{is_nnw}: Check for non-negative whole-number numeric mode.

\item \code{is_npw}: Check for non-positive whole-number numeric mode.

\item \code{is_psw}: Check for positive whole-number numeric mode.

\item \code{is_pct}: Check for positive whole-number numeric mode.

\item \code{is_ppn}: Check for positive whole-number numeric mode.

\item \code{is_ord}: Check for ordered factor mode.

\item \code{is_str}: Check for string mode (no \code{NA} values, no blanks)..

\item \code{is_nst}: Check for non-sortable atomic mode. Non-sortable objects
are atomic objects of modes that are not sortable. Sortable objects are
non-empty atomic objects of mode 'character', 'logical', 'numeric', or
'ordered'.

\item \code{is_srt}: Check for sortable atomic mode. Sortable objects are
non-empty atomic objects of mode 'character', 'logical', 'numeric', or
'ordered'. Non-sortable objects are atomic objects of modes that are not
sortable.

\item \code{is_uno}: Check for unordered factor mode.

\item \code{is_ind}: Check for indexer mode (logical or integer).
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
\keyword{datasets}
