% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_array.R
\name{is_xxx_array}
\alias{is_xxx_array}
\alias{is_any_array}
\alias{is_pop_array}
\alias{is_emp_array}
\alias{is_atm_array}
\alias{is_rcr_array}
\alias{is_cmp_array}
\title{Is an object an array (maybe with another property as well)?}
\usage{
is_any_array(x)

is_pop_array(x)

is_emp_array(x)

is_atm_array(x, xxx = NULL)

is_rcr_array(x)

is_cmp_array(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks for the following types of arrays:
\tabular{lll}{
\strong{Type}\tab\strong{Function}  \tab\strong{Definition}
\cr Any      \tab\code{is_any_array}\tab Any array.
\cr Populated\tab\code{is_pop_array}\tab A non-empty array.
\cr Empty    \tab\code{is_emp_array}\tab An empty array.
\cr Atomic   \tab\code{is_atm_array}\tab An atomic array.
\cr Recursive\tab\code{is_rcr_array}\tab A recursive array.
\cr Complete \tab\code{is_cmp_array}\tab An atomic array containing no
\code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
