% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_generic.R
\name{is_xxx_generic}
\alias{is_xxx_generic}
\alias{is_generic}
\alias{is_any_generic}
\alias{is_pop_generic}
\alias{is_emp_generic}
\alias{is_atm_generic}
\alias{is_rcr_generic}
\alias{is_cmp_generic}
\title{Is an object a generic (maybe with another property as well)?}
\usage{
is_generic(x)

is_any_generic(x)

is_pop_generic(x)

is_emp_generic(x)

is_atm_generic(x, xxx = NULL)

is_rcr_generic(x)

is_cmp_generic(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A generic is an array, \link[=is_vtype]{vtype},
\link[tibble:is_tibble]{tibble}, or \link[=is_vlist]{vlist}.
\cr\cr
Also checks for the following types of generics:
\tabular{lll}{
\strong{Type}  \tab\strong{Function}\tab\strong{Definition}
\cr Any      \tab\code{is_any_generic}\tab Any generic.
\cr Populated\tab\code{is_pop_generic}\tab Length is ≥ 1.
\cr Empty    \tab\code{is_emp_generic}\tab Length is 0.
\cr Atomic   \tab\code{is_atm_generic}\tab A populated atomic array, a
populated \link[=is_atm_vtype]{atomic vtype}, a
populated \link[=is_atm_tibble]{atomic tibble}, or a
populated \link[=is_atm_vlist]{atomic vlist}).
\cr Recursive\tab\code{is_rcr_generic}\tab A populated
\link[=is_rcr_array]{recursive array}, a proulated
\link[=is_rcr_vtype]{recursive vtype}, a populated
\link[=is_rcr_tibble]{recursive tibble}, or a
populated \link[=is_rcr_vlist]{recursive vlist}.
\cr Complete \tab\code{is_cmp_generic}\tab An atomic generic containing
no \code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
