% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_matrix.R
\name{is_xxx_matrix}
\alias{is_xxx_matrix}
\alias{is_any_matrix}
\alias{is_pop_matrix}
\alias{is_emp_matrix}
\alias{is_atm_matrix}
\alias{is_rcr_matrix}
\alias{is_cmp_matrix}
\title{Is an object a matrix?}
\usage{
is_any_matrix(x)

is_pop_matrix(x)

is_emp_matrix(x)

is_atm_matrix(x, xxx = NULL)

is_rcr_matrix(x)

is_cmp_matrix(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks for the following types of matrices:
\tabular{lll}{
\strong{Type}\tab\strong{Function}  \tab\strong{Definition}
\cr Any      \tab\code{is_any_matrix}\tab Any matrix.
\cr Populated\tab\code{is_pop_matrix}\tab A non-empty matrix.
\cr Empty    \tab\code{is_emp_matrix}\tab An empty matrix.
\cr Atomic   \tab\code{is_atm_matrix}\tab A populated, atomic matrix.
\cr Recursive\tab\code{is_rcr_matrix}\tab A populated matrix containing at
least one element that is not an
atomic scalar.
\cr Complete \tab\code{is_cmp_matrix}\tab A populated atomic matrix
containing no \code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
