% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_mvect.R
\name{is_xxx_mvect}
\alias{is_xxx_mvect}
\alias{is_mvect}
\alias{is_any_mvect}
\alias{is_atm_mvect}
\alias{is_rcr_mvect}
\alias{is_cmp_mvect}
\title{Is an object an mvect (maybe with another property as well)?}
\usage{
is_mvect(x)

is_any_mvect(x)

is_atm_mvect(x, xxx = NULL)

is_rcr_mvect(x)

is_cmp_mvect(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
An mvect is any non-data.frame object that is effectively
linear/effectively 1D. This includes vectors and
\link[=is_vlist]{vlists} of length ≥ 2, and effectively linear or
\link[=eee]{effectively 1D} arrays (populated arrays with multiple
index positions in exactly 1 dimension).
\cr\cr
Also checks for the following types of mvects:
\tabular{lll}{
\strong{Type}\tab\strong{Function}  \tab\strong{Definition}
\cr Any      \tab\code{is_any_mvect}\tab Any mvect.
\cr Atomic   \tab\code{is_atm_mvect}\tab An atomic mvect.
\cr Recursive\tab\code{is_rcr_mvect}\tab A vlist of length ≥ 2 or an
effectively linear array
containing at least 1 element that
is not an atomic scalar.
\cr Complete \tab\code{is_cmp_mvect}\tab An atomic mvect containing no
\code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
