% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_scalar.R
\name{is_xxx_scalar}
\alias{is_xxx_scalar}
\alias{is_scalar}
\alias{is_any_scalar}
\alias{is_atm_scalar}
\alias{is_rcr_scalar}
\alias{is_cmp_scalar}
\title{Is an object scalar?}
\usage{
is_scalar(x)

is_any_scalar(x)

is_atm_scalar(x, xxx = NULL)

is_rcr_scalar(x)

is_cmp_scalar(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A scalar is defined as a vector, array, or
\link[is_vlist]{vlist} of length \code{1}.
\cr\cr
Also checks for the following types of scalars:
\tabular{lll}{
\strong{Type}\tab\strong{Function}   \tab\strong{Definition}
\cr Any      \tab\code{is_any_scalar}\tab Any scalar.
\cr Atomic   \tab\code{is_atm_scalar}\tab An atomic scalar.
\cr Recursive\tab\code{is_rcr_scalar}\tab A scalar of recursive value.
\cr Complete \tab\code{is_cmp_scalar}\tab A non-\code{NA} atomic scalar.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
