% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_tibble.R
\name{is_xxx_tibble}
\alias{is_xxx_tibble}
\alias{is_any_tibble}
\alias{is_pop_tibble}
\alias{is_emp_tibble}
\alias{is_atm_tibble}
\alias{is_rcr_tibble}
\alias{is_cmp_tibble}
\title{Is an object a type of tibble (maybe with another property as well)?}
\usage{
is_any_tibble(x)

is_pop_tibble(x)

is_emp_tibble(x)

is_atm_tibble(x, xxx = NULL)

is_rcr_tibble(x)

is_cmp_tibble(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks for the following types of tibbles:
\tabular{lll}{
\strong{Type}\tab\strong{Function}   \tab\strong{Definition}
\cr Any      \tab\code{is_any_tibble}\tab A
\link[tibble:is_tibble]{tibble}.
\cr Populated\tab\code{is_pop_tibble}\tab A tibble of length ≥ 1.
\cr Empty    \tab\code{is_emp_tibble}\tab A tibble of length 0.
\cr Atomic   \tab\code{is_atm_tibble}\tab A populated tibble whose cells
all contain atomic scalars.
\cr Recursive\tab\code{is_rcr_tibble}\tab A populated tibble with one or
more cells containing a value
that is not an atomic scalar.
\cr Complete \tab\code{is_cmp_tibble}\tab A populated atomic tibble
containing no \code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
