% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_vect.R
\name{is_xxx_vect}
\alias{is_xxx_vect}
\alias{is_vect}
\alias{is_any_vect}
\alias{is_atm_vect}
\alias{is_rcr_vect}
\alias{is_cmp_vect}
\title{Is an object a vect (maybe with another property as well)?}
\usage{
is_vect(x)

is_any_vect(x)

is_atm_vect(x, xxx = NULL)

is_rcr_vect(x)

is_cmp_vect(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A vect is any object that is effectively a vector of length 1 or
greater. This includes \link[=is_scalar]{scalars}, vectors,
\link[=is_vlist]{vlists}, length-1 arrays, and arrays of length ≥ 2 with
multiple index positions in only 1 dimension (i.e.,
\link[=eee]{effectively 1D}).
\cr\cr
Also checks for the following types of vects:
\tabular{lll}{
\strong{Type}\tab\strong{Function} \tab\strong{Definition}
\cr Any      \tab\code{is_any_vect}\tab Any vect.
\cr Atomic   \tab\code{is_atm_vect}\tab An atomic vect.
\cr Recursive\tab\code{is_rcr_vect}\tab A \link[=is_vlist]{vlist} or an
array vect containing one or more
recursive element.
\cr Complete \tab\code{is_cmp_vect}\tab An atomic vect containing no
\code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vlist}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
