% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_vlist.R
\name{is_xxx_vlist}
\alias{is_xxx_vlist}
\alias{is_vlist}
\alias{is_any_vlist}
\alias{is_pop_vlist}
\alias{is_emp_vlist}
\alias{is_atm_vlist}
\alias{is_rcr_vlist}
\alias{is_cmp_vlist}
\title{Is an object a vlist?}
\usage{
is_vlist(x)

is_any_vlist(x)

is_pop_vlist(x)

is_emp_vlist(x)

is_atm_vlist(x, xxx = NULL)

is_rcr_vlist(x)

is_cmp_vlist(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific mmm.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A vlist (vector-list) is a non-data.frame list.
\cr\cr Also checks for the following types of vlists:
\tabular{lll}{
\strong{Type}\tab\strong{Function}  \tab\strong{Definition}
\cr Any      \tab\code{is_any_vlist}\tab Any vlist.
\cr Populated\tab\code{is_pop_vlist}\tab A non-empty vlist.
\cr Empty    \tab\code{is_emp_vlist}\tab A length-0 vlist.
\cr Atomic   \tab\code{is_atm_vlist}\tab A populated vlist, all of whose
elements are populated and atomic.
\cr Recursive\tab\code{is_rcr_vlist}\tab A populated vlist, all of whose
elements are populated and at
least one of which is recursive.
\cr Complete \tab\code{is_cmp_vlist}\tab A populated atomic vlist
containing no \code{NA} values.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vtype}},
\code{\link{logicals}}
}
\concept{is_functions}
