% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_vtype.R
\name{is_xxx_vtype}
\alias{is_xxx_vtype}
\alias{is_vtype}
\alias{is_any_vtype}
\alias{is_pop_vtype}
\alias{is_emp_vtype}
\alias{is_atm_vtype}
\alias{is_rcr_vtype}
\alias{is_cmp_vtype}
\title{Is an object a vtype (maybe with another property as well)?}
\usage{
is_vtype(x)

is_any_vtype(x)

is_pop_vtype(x)

is_emp_vtype(x)

is_atm_vtype(x, xxx = NULL)

is_rcr_vtype(x)

is_cmp_vtype(x, xxx = NULL)
}
\arguments{
\item{x}{An object}

\item{xxx}{\code{NULL} or a character scalar from \code{mmm_vals}
indicating an optional check for a specific extended mode.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
A vtype is a non-data.frame object that is empty, effectively
scalar, or effectively linear (having ≥ 2 elements). This includes any
vector or vlist, empty arrays, scalar, and effectively
linear/\link[=eee]{effectively 1D} arrays (populated arrays having
multiple index positions in only 1 dimension).
\cr\cr
Also checks for the following types of vtypes:
\tabular{lll}{
\strong{Type}\tab\strong{Function}  \tab\strong{Definition}
\cr Any      \tab\code{is_any_vtype}\tab Any vtype.
\cr Populated\tab\code{is_pop_vtype}\tab A vtype of length ≥ 1.
\cr Empty    \tab\code{is_emp_vtype}\tab An empty vtype.
\cr Atomic   \tab\code{is_atm_vtype}\tab A populated atomic vtype.
\cr Recursive\tab\code{is_rcr_vtype}\tab A populated vtype that is either a
vlist or an effectively scalar or
effectively linear array
containing one or more recursive
elements.
\cr Complete \tab\code{is_cmp_vtype}\tab A populated atomic vtype
containing no \code{NA} elements.
}
}
\seealso{
Other is_functions: 
\code{\link{is_color}()},
\code{\link{is_failsafe}},
\code{\link{is_sortable}()},
\code{\link{is_unique}()},
\code{\link{is_xxx_array}},
\code{\link{is_xxx_generic}},
\code{\link{is_xxx_matrix}},
\code{\link{is_xxx_mvect}},
\code{\link{is_xxx_scalar}},
\code{\link{is_xxx_tibble}},
\code{\link{is_xxx_vect}},
\code{\link{is_xxx_vlist}},
\code{\link{logicals}}
}
\concept{is_functions}
