% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{l}
\alias{l}
\title{Get non-scalar or non-character constants.}
\usage{
l(...)
}
\arguments{
\item{...}{unquoted, comma-separated list of the names of constants to
return. If multiple constants are specified, they are coerced into a single
atomic vector result.}
}
\value{
An atomic vector of length 1 or greater.
}
\description{
Get non-scalar or non-character constants.
}
\seealso{
Other constants: 
\code{\link{v}()}
}
\concept{constants}
