% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\alias{vec}
\alias{vec_na}
\alias{mat}
\alias{atb0}
\alias{atb_na}
\alias{atb}
\alias{vls}
\alias{mk_diag}
\title{Make atomic vectors, atomic matrices, atomic tibbles, and atomic or
recursive vlists}
\usage{
vec(..., r. = 1, en. = NULL)

vec_na(r.)

mat(..., nr. = 1, nc. = NULL, br. = F, rn. = NULL, cn. = NULL)

atb0(cn.)

atb_na(cn., nr.)

atb(..., rn. = NULL, cn. = NULL)

vls(..., en. = NULL)

mk_diag(x = 1, n = 1)
}
\arguments{
\item{...}{Objects to placed in an atomic vect, atomic matrix, atomib tibble,
vlist, or square atomic diagonal matrix.}

\item{r.}{Number of replications.}

\item{en.}{Optional character vector of names to apply to vector or list
elements. May be pipe-delimited in character vector.}

\item{nr.}{Number of rows.}

\item{nc.}{Number of columns.}

\item{br.}{Whether to fill matrices by row.}

\item{rn.}{Optional character vector of names to apply to rows of a matrix or
tibble. May be pipe-delimited in character vector.}

\item{cn.}{Optional character vector of names to apply to columns of a matrix
or tibble. May be pipe-delimited in character vector. For tibbles, this
argument can be used only if the arguments in \code{...} are not named.}

\item{x}{An numeric, logical, or character object of length 1 or greater.}

\item{n}{A positive whole number indicating the number of replications of
\code{x}.}
}
\value{
An \link[=is_atm_tibble]{atomic tibble}, an
\link[=is_atm_vect]{atomic vect}, an \link[=is_atm_matrix]{atomic matrix}
(\code{mat.}), or a \link[=is_vlist]{vlist.} All others return the value of
function they are thin wrappers for.
}
\description{
Extended functionality for creating vectors, matrices,
tibbles and vlists.
}
\details{
\strong{\code{vec(..., r, en.)}}
\cr Creates an atomic vector from the atomic objects in \code{...}
collapsed into a vector, replicated \code{r.} times, with optional element
names.
\cr\cr
\strong{\code{vec_na(r.)}}
\cr Creates a vector of \code{r.} \code{NA} values.
\cr\cr
\strong{\code{mat(..., nr., nc., br., rn., cn.)}}
\cr Creates an atomic matrix from the atomic elements in \code{...} with
\code{r.} rows, \code{c.} columns, optionally filling by row, with optional
row and column names.
\cr\cr
\strong{\code{atb0(cn.)}}
\cr Creates an \link[=is_atm_tibble]{atomic tibble} with 0 rows and columns
with names from \code{cn.}, which may be pipe delimited.
\cr\cr
\strong{\code{atb_na(cn., nr.)}}
\cr Creates an \link[=is_atm_tibble]{atomic tibble} with \code{nr.} rows
and columns with names from \code{cn.} (which may be pipe delimited). All
cells of the resulting tibble are populated with \code{NA}.
\cr\cr
\strong{\code{atb(..., rn., cn.)}}
\cr Creates an \link[=is_atm_tibble]{atomic tibble} from the atomic vect
arguments in \code{...} with optional row names, plus optional column names
to replace any names of arguments in \code{...}.
\cr\cr
\strong{\code{vls(..., en. = NULL)}}
\cr Creates a \link[=is_vlist]{vlist} from the arguments in \code{...} with
optional element names to replace any names of arguments in \code{...}.
\cr\cr
\strong{\code{mk_diag(x = 1, n = 1)}}
\cr Creates a square, atomic diagonal matrix. For numeric \code{x},
off-diagonals are \code{0}. For logical \code{x}, off diagonals are
\code{FALSE}. For character \code{x}, off diagonals are blank strings
(\code{""}). For all others, off diagonals are \code{NA}.
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
