% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variables.R
\name{map_variables}
\alias{map_variables}
\title{Map Variables in a Data Frame to Specified Roles}
\usage{
map_variables(
  x,
  roles,
  defs,
  mmms,
  .ROLES = NULL,
  .DEFS = NULL,
  .MMMS = NULL,
  .TRIM = TRUE,
  .RENAME = TRUE,
  .CLEAR = TRUE
)
}
\arguments{
\item{x}{A \link[uj:rct_atm_dtf]{rectangular atomic x.frame} whose variables are to be mapped.}

\item{roles}{\link[uj::unq_str_mvc]{Unique string multivec} containing labels for \emph{required} variable roles (that are also valid object names without back-quoting).}

\item{defs}{\link[uj::unq_str_mvc]{Unique string multivec} with \code{length(defs) == length(roles)} containing definitions/descriptions associated with corresponding values of \code{roles}.}

\item{mmms}{\link[uj::cmp_str_mvc]{Complete string multivec} with \code{length(mmms) == length(roles)} containing \link[uj::mmm]{extended modes} associated with corresponding values of \code{roles}.}

\item{.ROLES}{Optional \link[uj::unq_str_mvc]{unique string multivec} containing labels for \emph{optional} variable roles (that are also valid object names without back-quoting).}

\item{.DEFS}{Optional \link[uj::unq_str_mvc]{unique string multivec} with \code{length(.DEFS) == length(.ROLES)} containing definitions/descriptions associated with corresponding values of \code{.ROLES}.}

\item{.MMMS}{Optional \link[uj::cmp_str_mvc]{Complete string multivec} with \code{length(.MMMS) == length(.ROLES)} containing \link[uj::mmm]{extended modes} associated with corresponding values of \code{.ROLES}.}

\item{.TRIM, .RENAME, .CLEAR}{Logical scalar indicating, respectively, whether to trim un-mapped variables from \code{x}, whether to rename mapped variables using the associated values of \code{roles} and \code{.ROLES}, and whether to clear the console with each stop in the variable mapping process.}
}
\value{
A \link[uj:rct_atm_dtf]{rectangular atomic x.frame}.
}
\description{
Utility to map variables in a x frame to specific roles, with options to trim variables not assigned to roles, to rename variables to match variable roles (vs. attaching a variable map as an attribute), and to clear the console with each step in variable mapping.
}
\details{
Variable mapping is conducted interactively via the console for all variable roles in \code{roles} and \code{.ROLES} that do not have matching variable names in \code{x}. For variable roles in \code{roles} and \code{.ROLES} that have matching variable names in \code{x}, matching is done automatically.
}
\examples{
x <- x.frame(order = 1:26,
                   letters = letters,
                   LETTERS = LETTERS,
                   fac.letters = factor(letters),
                   ord.LETTERS = ordered(LETTERS),
                   lgl.sampled = sample(c(TRUE, FALSE), 26, replace = TRUE),
                   stringsAsFactors = FALSE)
roles  <- c("order"     , "letters"          , "LETTERS"          )
defs   <- c("case order", "lowercase letters", "uppercase letters")
mmms   <- c("psw"       , "ch1"              , "ch1"              )
.ROLES <- c("fac.letters"                          , "ord.LETTERS"                        )
.DEFS  <- c("lowercase letters as unordered factor", "uppercase letters as ordered factor")
.MMMS  <- c("uno"                                  , "ord"                                )
map_variables(x, roles, defs, mmms)
map_variables(x, roles, defs, mmms, .TRIM = FALSE)
map_variables(x, roles, defs, mmms, .TRIM = FALSE, .RENAME = FALSE)
map_variables(x, roles, defs, mmms, .TRIM = FALSE, .RENAME = FALSE, .CLEAR = FALSE)
map_variables(x, roles, defs, mmms, .ROLES = .ROLES, .DEFS = .DEFS, .MMMS = .MMMS)
map_variables(x, roles, defs, mmms, .ROLES = .ROLES, .DEFS = .DEFS, .MMMS = .MMMS, .TRIM = FALSE)
map_variables(x, roles, defs, mmms, .ROLES = .ROLES, .DEFS = .DEFS, .MMMS = .MMMS, .TRIM = FALSE, .RENAME = FALSE)
map_variables(x, roles, defs, mmms, .ROLES = .ROLES, .DEFS = .DEFS, .MMMS = .MMMS, .TRIM = FALSE, .RENAME = FALSE, .CLEAR = FALSE)
}
