% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\encoding{UTF-8}
\name{markdown}
\alias{markdown}
\alias{subs}
\alias{sups}
\alias{bfblu}
\alias{blblu}
\alias{bmblu}
\alias{bdblu}
\alias{bfcyn}
\alias{blcyn}
\alias{bmcyn}
\alias{bdcyn}
\alias{bfgrn}
\alias{blgrn}
\alias{bmgrn}
\alias{bdgrn}
\alias{bfmag}
\alias{blmag}
\alias{bmmag}
\alias{bdmag}
\alias{bforn}
\alias{blorn}
\alias{bmorn}
\alias{bdorn}
\alias{bfred}
\alias{blred}
\alias{bmred}
\alias{bdred}
\alias{bfvlt}
\alias{blvlt}
\alias{bmvlt}
\alias{bdvlt}
\alias{bfylw}
\alias{blylw}
\alias{bmylw}
\alias{bdylw}
\alias{bwht}
\alias{bg95}
\alias{bg90}
\alias{bg85}
\alias{bg80}
\alias{bg75}
\alias{bg70}
\alias{bg65}
\alias{bg60}
\alias{bg55}
\alias{bg50}
\alias{bg45}
\alias{bg40}
\alias{bg35}
\alias{bg30}
\alias{bg25}
\alias{bg20}
\alias{bg15}
\alias{bg10}
\alias{bg05}
\alias{bblk}
\alias{binv}
\alias{pfblu}
\alias{plblu}
\alias{pmblu}
\alias{pdblu}
\alias{pfcyn}
\alias{plcyn}
\alias{pmcyn}
\alias{pdcyn}
\alias{pfgrn}
\alias{plgrn}
\alias{pmgrn}
\alias{pdgrn}
\alias{pfmag}
\alias{plmag}
\alias{pmmag}
\alias{pdmag}
\alias{pforn}
\alias{plorn}
\alias{pmorn}
\alias{pdorn}
\alias{pfred}
\alias{plred}
\alias{pmred}
\alias{pdred}
\alias{pfvlt}
\alias{plvlt}
\alias{pmvlt}
\alias{pdvlt}
\alias{pfylw}
\alias{plylw}
\alias{pmylw}
\alias{pdylw}
\alias{pwht}
\alias{pg95}
\alias{pg90}
\alias{pg85}
\alias{pg80}
\alias{pg75}
\alias{pg70}
\alias{pg65}
\alias{pg60}
\alias{pg55}
\alias{pg50}
\alias{pg45}
\alias{pg40}
\alias{pg35}
\alias{pg30}
\alias{pg25}
\alias{pg20}
\alias{pg15}
\alias{pg10}
\alias{pg05}
\alias{pblk}
\alias{pinv}
\alias{ifblu}
\alias{ilblu}
\alias{imblu}
\alias{idblu}
\alias{ifcyn}
\alias{ilcyn}
\alias{imcyn}
\alias{idcyn}
\alias{ifgrn}
\alias{ilgrn}
\alias{imgrn}
\alias{idgrn}
\alias{ifmag}
\alias{ilmag}
\alias{immag}
\alias{idmag}
\alias{iforn}
\alias{ilorn}
\alias{imorn}
\alias{idorn}
\alias{ifred}
\alias{ilred}
\alias{imred}
\alias{idred}
\alias{ifvlt}
\alias{ilvlt}
\alias{imvlt}
\alias{idvlt}
\alias{ifylw}
\alias{ilylw}
\alias{imylw}
\alias{idylw}
\alias{iwht}
\alias{ig95}
\alias{ig90}
\alias{ig85}
\alias{ig80}
\alias{ig75}
\alias{ig70}
\alias{ig65}
\alias{ig60}
\alias{ig55}
\alias{ig50}
\alias{ig45}
\alias{ig40}
\alias{ig35}
\alias{ig30}
\alias{ig25}
\alias{ig20}
\alias{ig15}
\alias{ig10}
\alias{ig05}
\alias{iblk}
\alias{iinv}
\title{Build markdown expressions for styled + colored \code{ggtext} strings}
\usage{
markdown()

subs(X)

sups(X)

bfblu(X)

blblu(X)

bmblu(X)

bdblu(X)

bfcyn(X)

blcyn(X)

bmcyn(X)

bdcyn(X)

bfgrn(X)

blgrn(X)

bmgrn(X)

bdgrn(X)

bfmag(X)

blmag(X)

bmmag(X)

bdmag(X)

bforn(X)

blorn(X)

bmorn(X)

bdorn(X)

bfred(X)

blred(X)

bmred(X)

bdred(X)

bfvlt(X)

blvlt(X)

bmvlt(X)

bdvlt(X)

bfylw(X)

blylw(X)

bmylw(X)

bdylw(X)

bwht(X)

bg95(X)

bg90(X)

bg85(X)

bg80(X)

bg75(X)

bg70(X)

bg65(X)

bg60(X)

bg55(X)

bg50(X)

bg45(X)

bg40(X)

bg35(X)

bg30(X)

bg25(X)

bg20(X)

bg15(X)

bg10(X)

bg05(X)

bblk(X)

binv(X)

pfblu(X)

plblu(X)

pmblu(X)

pdblu(X)

pfcyn(X)

plcyn(X)

pmcyn(X)

pdcyn(X)

pfgrn(X)

plgrn(X)

pmgrn(X)

pdgrn(X)

pfmag(X)

plmag(X)

pmmag(X)

pdmag(X)

pforn(X)

plorn(X)

pmorn(X)

pdorn(X)

pfred(X)

plred(X)

pmred(X)

pdred(X)

pfvlt(X)

plvlt(X)

pmvlt(X)

pdvlt(X)

pfylw(X)

plylw(X)

pmylw(X)

pdylw(X)

pwht(X)

pg95(X)

pg90(X)

pg85(X)

pg80(X)

pg75(X)

pg70(X)

pg65(X)

pg60(X)

pg55(X)

pg50(X)

pg45(X)

pg40(X)

pg35(X)

pg30(X)

pg25(X)

pg20(X)

pg15(X)

pg10(X)

pg05(X)

pblk(X)

pinv(X)

ifblu(X)

ilblu(X)

imblu(X)

idblu(X)

ifcyn(X)

ilcyn(X)

imcyn(X)

idcyn(X)

ifgrn(X)

ilgrn(X)

imgrn(X)

idgrn(X)

ifmag(X)

ilmag(X)

immag(X)

idmag(X)

iforn(X)

ilorn(X)

imorn(X)

idorn(X)

ifred(X)

ilred(X)

imred(X)

idred(X)

ifvlt(X)

ilvlt(X)

imvlt(X)

idvlt(X)

ifylw(X)

ilylw(X)

imylw(X)

idylw(X)

iwht(X)

ig95(X)

ig90(X)

ig85(X)

ig80(X)

ig75(X)

ig70(X)

ig65(X)

ig60(X)

ig55(X)

ig50(X)

ig45(X)

ig40(X)

ig35(X)

ig30(X)

ig25(X)

ig20(X)

ig15(X)

ig10(X)

ig05(X)

iblk(X)

iinv(X)
}
\arguments{
\item{X}{A \link[=cmp_chr_vec]{character vector} of text to be styled.}
}
\value{
A markdown-styled character vector.
}
\description{
Style markdown text with subscripts, superscripts, and colors.
}
\details{
\strong{Subscript/superscript} functions are:
\tabular{ll}{  \code{subs}   \tab subscript   \cr
  \code{sups}   \tab superscript   }
\cr\cr \strong{All remaining functions} in this family start with a prefix character as follows:
\tabular{ll}{  prefix = \code{b}   \tab bold   \cr
  prefix = \code{i}   \tab italic \cr
  prefix = \code{p}   \tab plain    }
\cr\cr \strong{Black, white, and invisible text} functions suffixes are:
\tabular{ll}{  \code{blk}   \tab black   \cr
  \code{wht}   \tab white   \cr
  \code{inv}   \tab invisible }
\cr For example, the function \code{bwht} styles markdown text as bold and white.
\cr\cr \strong{Greyscale} text suffixes take the form:
\tabular{ll}{  \code{g05}   \tab \code{ 5:95} white:black ratio \cr
  \code{g10}   \tab \code{10:90} white:black ratio \cr
  \code{...}   \tab \code{...}                     \cr
  \code{g90}   \tab \code{90:10} white:black ratio \cr
  \code{g95}   \tab \code{95:5 } white:black ratio   }
\cr For example, the function \code{pg45} styles markdown text as plain and grey with a \code{45:55} white:black ratio.
\cr\cr \strong{Other common-color text} function names append an infix and a suffix to the selected prefix, as follows:
\tabular{ll}{  infix = \code{f}      \tab full intensity    \cr
  infix = \code{l}      \tab light             \cr
  infix = \code{m}      \tab medium            \cr
  infix = \code{d}      \tab dark \cr   \tab   \cr
  suffix = \code{blu}   \tab blue              \cr
  suffix = \code{cyn}   \tab cyan              \cr
  suffix = \code{grn}   \tab green             \cr
  suffix = \code{mag}   \tab magenta           \cr
  suffix = \code{orn}   \tab orange            \cr
  suffix = \code{red}   \tab red               \cr
  suffix = \code{vlt}   \tab violet            \cr
  suffix = \code{ylw}   \tab yellow              }

\cr For example, the function \code{pfred} styles markdown text as plain and full-intensity red.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other colors: 
\code{\link{sca}()}

Other plots: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{pals}()},
\code{\link{sca}()},
\code{\link{u2u}()}
}
\concept{colors}
\concept{plots}
\concept{strings}
