% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{match_bg}
\alias{match_bg}
\alias{match_fg}
\alias{match_st}
\alias{match_alias}
\title{Match Color or Style Aliases}
\usage{
match_bg(x)

match_fg(x)

match_st(x)

match_alias()
}
\arguments{
\item{x}{A character scalar alias.}
}
\value{
A 3-letter character scalar.
}
\description{
Returns the Official Color or Style Designation from Any Alias
}
\section{Functions}{
\itemize{
\item \code{match_bg()}: Matches any valid alias for background color.

\item \code{match_fg()}: Matches any valid alias for foreground color.

\item \code{match_st()}: Matches any valid alias for style.

}}
\examples{
cat("\n\nBACKGROUND COLORS"); for (b in 1:bgs()) {cat("\n'", match_bg(b), "' << '", b, "'", sep = "")}
cat("\n\nFOREGROUND COLORS"); for (f in 1:bgs()) {cat("\n'", match_fg(f), "' << '", f, "'", sep = "")}
cat("\n\nTEXT STYLES"      ); for (s in 1:sts()) {cat("\n'", match_st(s), "' << '", s, "'", sep = "")}
}
\seealso{
Other alert_utils: 
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{glu}()},
\code{\link{ok_fmt}()},
\code{\link{ssplit}()}
}
\concept{alert_utils}
