% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_nch.R
\encoding{UTF-8}
\name{max_nch}
\alias{max_nch}
\title{Length of longest string}
\usage{
max_nch(...)
}
\arguments{
\item{...}{An arbitrary number of objects of mode character.}
}
\value{
An integer scalar.
}
\description{
Gets the number of characters in the longest string in \code{x}.
}
\examples{
max_nch(letters)
max_nch("a", "collection", "of", "words")
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}},
\code{\link{markdown}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other counts: 
\code{\link{meets}()},
\code{\link{n_is}()}

Other chars: 
\code{\link{chn}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()}
}
\concept{base_extensions}
\concept{chars}
\concept{counts}
\concept{strings}
