% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnch.R
\encoding{UTF-8}
\name{maxnch}
\alias{maxnch}
\alias{maxlen}
\title{Length of Longest String}
\usage{
maxnch(...)

maxlen(...)
}
\arguments{
\item{...}{An arbitrary number of objects of mode character.}
}
\value{
An integer scalar.
}
\description{
Gets the number of characters in the longest string in \code{...}.
}
\examples{
max_nch(letters)
maxlen("a", "collection", "of", "words")
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown_help}()},
\code{\link{ox}()},
\code{\link{ox_vals}()},
\code{\link{pgrid_help}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss_help}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other counts: 
\code{\link{N}()}

Other chars: 
\code{\link{chn}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss_help}()}
}
\concept{chars}
\concept{counts}
\concept{strings}
