% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnch.R
\name{maxnch_uj}
\alias{maxnch_uj}
\alias{maxnch}
\title{Length of longest string}
\usage{
maxnch_uj()

maxnch(x)
}
\arguments{
\item{x}{Atomic character object.}
}
\value{
Integer scalar.
}
\description{
Length of longest string
}
\section{Functions}{
\itemize{
\item \code{maxnch}: Gets the number of characters in the longest string in
\code{x}.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
