% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meets.R
\name{meets.}
\alias{meets.}
\title{Does an object meet count and/or value restrictions?}
\usage{
meets.()
}
\arguments{
\item{x}{An object.}

\item{...}{Optional named arguments specifying (additional) property
requirements for \code{x} in terms of element values and/or
element/row/column counts. See the \emph{Specifying Additional Property
Requirements} section.}
}
\value{
A logical scalar.
}
\description{
Count and value restrictions are provided in \code{...}. If none
are provided
}
\section{Specifying Additional Property Requirements}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar value all values of \code{x} must be less
than.                                                \cr
\code{le}     \tab Scalar value all values of \code{x} must be less than or
equal to.                                            \cr
\code{ge}     \tab Scalar value all values of \code{x} must be greater than
or equal to.                                         \cr
\code{gt}     \tab Scalar value all values of \code{x} must be greater
than.                                                  }
}

\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
