% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meets.R
\name{meets}
\alias{meets}
\title{Evaluate whether an object meets count and/or value restrictions}
\usage{
meets(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Optional named arguments count and/or value restrictions for \code{x}. See the \emph{specifying count and value restrictions} section.}
}
\value{
A logical scalar.
}
\description{
Count and value restrictions are provided in \code{...}. If none are provided, returns \code{TRUE}.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\itemize{
\item \strong{\code{n}}: a vector of valid lengths/numbers of elements.
\item \strong{\code{nr}}: a vector of valid numbers of rows.
\item \strong{\code{nc}}: a vector of valid numbers of columns.
\item \strong{\code{min}}: a scalar minimum valid length/number of element.
\item \strong{\code{minr}}: a scalar minimum valid number of rows.
\item \strong{\code{minc}}: a scalar minimum valid number of columns.
\item \strong{\code{max}}: a scalar maximum valid length/number of element.
\item \strong{\code{maxr}}: a scalar maximum valid number of rows.
\item \strong{\code{maxc}}: a scalar maximum valid number of columns.
\item \strong{\code{vals}}: a vector of valid values.
\item \strong{\code{lt}}: a scalar less-than (exclusive upper) bound.
\item \strong{\code{le}}: a scalar less-than-or-equal (inclusive upper) bound
\item \strong{\code{ge}}: a scalar greater-than-or-equal (inclusive lower) bound.
\item \strong{\code{gt}}: a scalar greater-than bound (exclusive lower) bound.
}
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
