% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meets_errs.R
\encoding{UTF-8}
\name{meets_errs}
\alias{meets_errs}
\title{Checks for Meta-Errors in Count and/or Value-Restriction Arguments, If There Are Any}
\usage{
meets_errs(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
No return value. Called for the side effect of evaluating the validity of any count- or value-restriction arguments in \code{...}.
}
\description{
Checks whether any count- or value-restrictions are carried in \code{...} arguments. If so, evaluates whether those restriction arguments are valid. If not, throws errors describing invalid restriction arguments.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.max, .maxr, .maxc}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                 \cr   \tab   \cr
  \verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                 \cr   \tab   \cr
  \verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{not_err}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}
}
\concept{utils}
