% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_tools.R
\name{mgcv_tools}
\alias{mgcv_tools}
\alias{mgbam}
\alias{mgocat}
\alias{mgpred}
\title{Aliases and wrappers for functions in package \code{mgcv}.}
\usage{
mgbam(
  formula,
  family = gaussian(),
  data = list(),
  weights = NULL,
  subset = NULL,
  na.action = na.omit,
  offset = NULL,
  method = "fREML",
  control = list(),
  select = FALSE,
  scale = 0,
  gamma = 1,
  knots = NULL,
  sp = NULL,
  min.sp = NULL,
  paraPen = NULL,
  chunk.size = 10000,
  rho = 0,
  AR.start = NULL,
  discrete = FALSE,
  cluster = NULL,
  nthreads = 1,
  gc.level = 0,
  use.chol = FALSE,
  samfrac = 1,
  coef = NULL,
  drop.unused.levels = TRUE,
  G = NULL,
  fit = TRUE,
  drop.intercept = NULL,
  ...
)

mgocat(theta = NULL, link = "identity", R = NULL)

mgpred(
  object,
  newdata,
  type = "link",
  se.fit = FALSE,
  terms = NULL,
  exclude = NULL,
  block.size = 50000,
  newdata.guaranteed = FALSE,
  na.action = na.pass,
  cluster = NULL,
  discrete = TRUE,
  n.threads = 1,
  gc.level = 0,
  ...
)
}
\description{
Alias for \code{\link[mgcv]{bam}} (big-data generalized additive
models).
}
\section{Functions}{
\itemize{
\item \code{mgocat}: Alias for \code{\link[mgcv]{ocat}} (ordered category
logistic regression family).

\item \code{mgpred}: Alias for \code{\link[mgcv]{predict.bam}} (fitted
value prediction for new data).
}}

