% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{mid}
\alias{mid}
\title{Substring extraction by position.}
\usage{
mid(x, f, l)
}
\arguments{
\item{x}{Character scalar or vector of strings to search.}

\item{f, l}{Positive integer scalar first and last character positions.}
}
\value{
Character scalar or vector.
}
\description{
Substring extraction by position.
}
