% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkd.R
\docType{data}
\name{mkd}
\alias{mkd}
\alias{subs}
\alias{sups}
\alias{bfblu}
\alias{blblu}
\alias{bmblu}
\alias{bdblu}
\alias{bfcyn}
\alias{blcyn}
\alias{bmcyn}
\alias{bdcyn}
\alias{bfgrn}
\alias{blgrn}
\alias{bmgrn}
\alias{bdgrn}
\alias{bfmag}
\alias{blmag}
\alias{bmmag}
\alias{bdmag}
\alias{bforn}
\alias{blorn}
\alias{bmorn}
\alias{bdorn}
\alias{bfred}
\alias{blred}
\alias{bmred}
\alias{bdred}
\alias{bfvlt}
\alias{blvlt}
\alias{bmvlt}
\alias{bdvlt}
\alias{bfylw}
\alias{blylw}
\alias{bmylw}
\alias{bdylw}
\alias{bwht}
\alias{bg95}
\alias{bg90}
\alias{bg85}
\alias{bg80}
\alias{bg75}
\alias{bg70}
\alias{bg65}
\alias{bg60}
\alias{bg55}
\alias{bg50}
\alias{bg45}
\alias{bg40}
\alias{bg35}
\alias{bg30}
\alias{bg25}
\alias{bg20}
\alias{bg15}
\alias{bg10}
\alias{bg05}
\alias{bblk}
\alias{binv}
\alias{pfblu}
\alias{plblu}
\alias{pmblu}
\alias{pdblu}
\alias{pfcyn}
\alias{plcyn}
\alias{pmcyn}
\alias{pdcyn}
\alias{pfgrn}
\alias{plgrn}
\alias{pmgrn}
\alias{pdgrn}
\alias{pfmag}
\alias{plmag}
\alias{pmmag}
\alias{pdmag}
\alias{pforn}
\alias{plorn}
\alias{pmorn}
\alias{pdorn}
\alias{pfred}
\alias{plred}
\alias{pmred}
\alias{pdred}
\alias{pfvlt}
\alias{plvlt}
\alias{pmvlt}
\alias{pdvlt}
\alias{pfylw}
\alias{plylw}
\alias{pmylw}
\alias{pdylw}
\alias{pwht}
\alias{pg95}
\alias{pg90}
\alias{pg85}
\alias{pg80}
\alias{pg75}
\alias{pg70}
\alias{pg65}
\alias{pg60}
\alias{pg55}
\alias{pg50}
\alias{pg45}
\alias{pg40}
\alias{pg35}
\alias{pg30}
\alias{pg25}
\alias{pg20}
\alias{pg15}
\alias{pg10}
\alias{pg05}
\alias{pblk}
\alias{pinv}
\title{Build Markdown Expressions for Styled and Colored \code{ggtext}}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
mkd

subs(x)

sups(x)

bfblu(x)

blblu(x)

bmblu(x)

bdblu(x)

bfcyn(x)

blcyn(x)

bmcyn(x)

bdcyn(x)

bfgrn(x)

blgrn(x)

bmgrn(x)

bdgrn(x)

bfmag(x)

blmag(x)

bmmag(x)

bdmag(x)

bforn(x)

blorn(x)

bmorn(x)

bdorn(x)

bfred(x)

blred(x)

bmred(x)

bdred(x)

bfvlt(x)

blvlt(x)

bmvlt(x)

bdvlt(x)

bfylw(x)

blylw(x)

bmylw(x)

bdylw(x)

bwht(x)

bg95(x)

bg90(x)

bg85(x)

bg80(x)

bg75(x)

bg70(x)

bg65(x)

bg60(x)

bg55(x)

bg50(x)

bg45(x)

bg40(x)

bg35(x)

bg30(x)

bg25(x)

bg20(x)

bg15(x)

bg10(x)

bg05(x)

bblk(x)

binv(x)

pfblu(x)

plblu(x)

pmblu(x)

pdblu(x)

pfcyn(x)

plcyn(x)

pmcyn(x)

pdcyn(x)

pfgrn(x)

plgrn(x)

pmgrn(x)

pdgrn(x)

pfmag(x)

plmag(x)

pmmag(x)

pdmag(x)

pforn(x)

plorn(x)

pmorn(x)

pdorn(x)

pfred(x)

plred(x)

pmred(x)

pdred(x)

pfvlt(x)

plvlt(x)

pmvlt(x)

pdvlt(x)

pfylw(x)

plylw(x)

pmylw(x)

pdylw(x)

pwht(x)

pg95(x)

pg90(x)

pg85(x)

pg80(x)

pg75(x)

pg70(x)

pg65(x)

pg60(x)

pg55(x)

pg50(x)

pg45(x)

pg40(x)

pg35(x)

pg30(x)

pg25(x)

pg20(x)

pg15(x)

pg10(x)

pg05(x)

pblk(x)

pinv(x)
}
\arguments{
\item{x}{A \link[=cmp_chr_vec]{character vector} of text to be styled.}
}
\value{
Markdown-styled character vector
}
\description{
Style text as subscript, superscript, bold color, and plain
color. The following markdown styling function names do not follow the
conventions described later:
\tabular{lll}{
TEXT STYLING      \tab REPRESENTATION   \tab NOTE                     \cr
subscript         \tab \code{subs}           \tab —                        \cr
superscript       \tab \code{subs}           \tab —                        \cr
bold black        \tab \code{bblk}           \tab —                        \cr
bold white        \tab \code{bwht}           \tab —                        \cr
bold invisible    \tab \code{binv}           \tab 1/256th opaque white.    \cr
plain black       \tab \code{pblk}           \tab —                        \cr
plain white       \tab \code{pwht}           \tab —                        \cr
plain invisible   \tab \code{pinv}           \tab 1/256th opaque white.      }
For all other markdown styling functions, conventions are as
follows:\tabular{lll}{
CONVENTION      \tab VALUE   \tab TEXT STYLING                        \cr
1st character   \tab \code{p}     \tab Plain.                              \cr
1st character   \tab \code{b}     \tab Bold.                               \cr
2nd character   \tab \code{f}     \tab Full intensity version of color.    \cr
2nd character   \tab \code{l}     \tab Light version of color.             \cr
2nd character   \tab \code{m}     \tab Medium version of color.            \cr
2nd character   \tab \code{d}     \tab Dark version of color.              \cr
2nd character   \tab \code{g}     \tab Grey.                               \cr
suffix          \tab \code{blu}   \tab Blue.                               \cr
suffix          \tab \code{cyn}   \tab Cyan.                               \cr
suffix          \tab \code{grn}   \tab Green.                              \cr
suffix          \tab \code{mag}   \tab Magenta.                            \cr
suffix          \tab \code{orn}   \tab Orange.                             \cr
suffix          \tab \code{red}   \tab Red.                                \cr
suffix          \tab \code{vlt}   \tab Violet.                             \cr
suffix          \tab \code{ylw}   \tab Yellow.                             \cr
suffix          \tab \code{05}    \tab 5\% of the way from black to white.  \cr
suffix          \tab \code{10}    \tab 10\% of the way from black to white. \cr
suffix          \tab \code{15}    \tab 15\% of the way from black to white. \cr
suffix          \tab \code{20}    \tab 20\% of the way from black to white. \cr
suffix          \tab \code{25}    \tab 25\% of the way from black to white. \cr
suffix          \tab \code{30}    \tab 30\% of the way from black to white. \cr
suffix          \tab \code{35}    \tab 35\% of the way from black to white. \cr
suffix          \tab \code{40}    \tab 40\% of the way from black to white. \cr
suffix          \tab \code{45}    \tab 45\% of the way from black to white. \cr
suffix          \tab \code{50}    \tab 50\% of the way from black to white. \cr
suffix          \tab \code{55}    \tab 55\% of the way from black to white. \cr
suffix          \tab \code{60}    \tab 60\% of the way from black to white. \cr
suffix          \tab \code{65}    \tab 65\% of the way from black to white. \cr
suffix          \tab \code{70}    \tab 70\% of the way from black to white. \cr
suffix          \tab \code{75}    \tab 75\% of the way from black to white. \cr
suffix          \tab \code{80}    \tab 80\% of the way from black to white. \cr
suffix          \tab \code{85}    \tab 85\% of the way from black to white. \cr
suffix          \tab \code{90}    \tab 90\% of the way from black to white. \cr
suffix          \tab \code{95}    \tab 95\% of the way from black to white.   }
}
\section{Functions}{
\itemize{
\item \code{pg15()}: Style text as plain and 15\% grey.

\item \code{pg10()}: Style text as plain and 10\% grey.

\item \code{pg05()}: Style text as plain and 10\% grey.

\item \code{pblk()}: Style text as plain and black.

\item \code{pinv()}: Style text as plain invisible (1/256th opaque).

}}
\seealso{
Other plots: 
\code{\link{pals_uj}}

Other colors: 
\code{\link{color}},
\code{\link{pals_uj}}
}
\concept{colors}
\concept{plots}
\keyword{datasets}
