% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_agn.R
\name{mmm_agn}
\alias{mmm_agn}
\alias{mmm_agn_vals}
\alias{chr_agn}
\alias{ch1_agn}
\alias{clr_agn}
\alias{evn_agn}
\alias{fac_agn}
\alias{frc_agn}
\alias{ind_agn}
\alias{lgl_agn}
\alias{neg_agn}
\alias{ngw_agn}
\alias{nng_agn}
\alias{nnw_agn}
\alias{nps_agn}
\alias{npw_agn}
\alias{nst_agn}
\alias{num_agn}
\alias{odd_agn}
\alias{ord_agn}
\alias{pct_agn}
\alias{pos_agn}
\alias{ppn_agn}
\alias{psw_agn}
\alias{srt_agn}
\alias{str_agn}
\alias{uno_agn}
\alias{whl_agn}
\title{Extended Mode + Atomic Generic (mmm + agn)}
\usage{
mmm_agn_vals()

chr_agn(x)

ch1_agn(x)

clr_agn(x)

evn_agn(x)

fac_agn(x)

frc_agn(x)

ind_agn(x)

lgl_agn(x)

neg_agn(x)

ngw_agn(x)

nng_agn(x)

nnw_agn(x)

nps_agn(x)

npw_agn(x)

nst_agn(x)

num_agn(x)

odd_agn(x)

ord_agn(x)

pct_agn(x)

pos_agn(x)

ppn_agn(x)

psw_agn(x)

srt_agn(x)

str_agn(x)

uno_agn(x)

whl_agn(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic generic properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_gen}}.
}
\section{Functions}{
\itemize{
\item \code{chr_agn}: Is \code{x} a character atomic generic?

\item \code{ch1_agn}: Is \code{x} a onechar atomic generic?

\item \code{clr_agn}: Is \code{x} a color atomic

\item \code{evn_agn}: Is \code{x} an even-numeric atomic generic?

\item \code{fac_agn}: Is \code{x} a factor atomic generic?

\item \code{frc_agn}: Is \code{x} a fractional-numeric atomic generic?

\item \code{ind_agn}: Is \code{x} an indexer atomic generic?

\item \code{lgl_agn}: Is \code{x} a logical atomic generic?

\item \code{neg_agn}: Is \code{x} a negative-numeric atomic generic?

\item \code{ngw_agn}: Is \code{x} a negative-whole-numeric atomic generic?

\item \code{nng_agn}: Is \code{x} a non-negative-numeric atomic generic?

\item \code{nnw_agn}: Is \code{x} a non-negative-whole-numeric atomic generic?

\item \code{nps_agn}: Is \code{x} a non-positive-numeric atomic generic?

\item \code{npw_agn}: Is \code{x} a non-positive-whole-numeric atomic generic?

\item \code{nst_agn}: Is \code{x} a non-sortable atomic generic?

\item \code{num_agn}: Is \code{x} a numeric atomic generic?

\item \code{odd_agn}: Is \code{x} an odd-numeric atomic generic?

\item \code{ord_agn}: Is \code{x} an ordered-factor atomic generic?

\item \code{pct_agn}: Is \code{x} a percent-numeric atomic generic?

\item \code{pos_agn}: Is \code{x} a positive-numeric atomic generic?

\item \code{ppn_agn}: Is \code{x} a proportion-numeric atomic generic?

\item \code{psw_agn}: Is \code{x} a positive-whole-numeric atomic generic?

\item \code{srt_agn}: Is \code{x} a sortable atomic generic?

\item \code{str_agn}: Is \code{x} a string atomic generic?

\item \code{uno_agn}: Is \code{x} an unordered-factor atomic generic?

\item \code{whl_agn}: Is \code{x} a whole-numeric atomic generic?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
