% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_arr.R
\name{mmm_arr}
\alias{mmm_arr}
\alias{mmm_arr_vals}
\alias{atm_arr}
\alias{chr_arr}
\alias{ch1_arr}
\alias{clr_arr}
\alias{evn_arr}
\alias{fac_arr}
\alias{frc_arr}
\alias{ind_arr}
\alias{lgc_arr}
\alias{neg_arr}
\alias{ngw_arr}
\alias{nng_arr}
\alias{nnw_arr}
\alias{nps_arr}
\alias{npw_arr}
\alias{nst_arr}
\alias{num_arr}
\alias{odd_arr}
\alias{ord_arr}
\alias{pct_arr}
\alias{pos_arr}
\alias{psw_arr}
\alias{ppn_arr}
\alias{srt_arr}
\alias{str_arr}
\alias{uno_arr}
\alias{whl_arr}
\title{Extended Mode + Atomic array (mmm + arr)}
\usage{
mmm_arr_vals()

atm_arr(x)

chr_arr(x)

ch1_arr(x)

clr_arr(x)

evn_arr(x)

fac_arr(x)

frc_arr(x)

ind_arr(x)

lgc_arr(x)

neg_arr(x)

ngw_arr(x)

nng_arr(x)

nnw_arr(x)

nps_arr(x)

npw_arr(x)

nst_arr(x)

num_arr(x)

odd_arr(x)

ord_arr(x)

pct_arr(x)

pos_arr(x)

psw_arr(x)

ppn_arr(x)

srt_arr(x)

str_arr(x)

uno_arr(x)

whl_arr(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}} and \code{\link{is_atm_array}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
