% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_atb.R
\name{mmm_atb}
\alias{mmm_atb}
\alias{mmm_atb_vals}
\alias{chr_atb}
\alias{ch1_atb}
\alias{clr_atb}
\alias{evn_atb}
\alias{fac_atb}
\alias{frc_atb}
\alias{ind_atb}
\alias{lgl_atb}
\alias{neg_atb}
\alias{ngw_atb}
\alias{nng_atb}
\alias{nnw_atb}
\alias{nps_atb}
\alias{npw_atb}
\alias{nst_atb}
\alias{num_atb}
\alias{odd_atb}
\alias{ord_atb}
\alias{pct_atb}
\alias{pos_atb}
\alias{psw_atb}
\alias{ppn_atb}
\alias{srt_atb}
\alias{str_atb}
\alias{uno_atb}
\alias{whl_atb}
\title{Extended Mode + Atomic tibble (mmm + atb)}
\usage{
mmm_atb_vals()

chr_atb(x)

ch1_atb(x)

clr_atb(x)

evn_atb(x)

fac_atb(x)

frc_atb(x)

ind_atb(x)

lgl_atb(x)

neg_atb(x)

ngw_atb(x)

nng_atb(x)

nnw_atb(x)

nps_atb(x)

npw_atb(x)

nst_atb(x)

num_atb(x)

odd_atb(x)

ord_atb(x)

pct_atb(x)

pos_atb(x)

psw_atb(x)

ppn_atb(x)

srt_atb(x)

str_atb(x)

uno_atb(x)

whl_atb(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic tibble properties.
}
\details{
See \code{\link{mmm}} and \code{\link{atm_tbl}}.
}
\section{Functions}{
\itemize{
\item \code{chr_atb}: Is \code{x} a character atomic tibble?

\item \code{ch1_atb}: Is \code{x} a onechar atomic tibble?

\item \code{clr_atb}: Is \code{x} a color atomic tibble?

\item \code{evn_atb}: Is \code{x} a even-numeric atomic tibble?

\item \code{fac_atb}: Is \code{x} a factor atomic tibble?

\item \code{frc_atb}: Is \code{x} a fractional-numeric atomic tibble?

\item \code{ind_atb}: Is \code{x} an indexer atomic tibble?

\item \code{lgl_atb}: Is \code{x} a logical atomic tibble?

\item \code{neg_atb}: Is \code{x} a negative-numeric atomic tibble?

\item \code{ngw_atb}: Is \code{x} a non-negative-whole-numeric atomic tibble?

\item \code{nng_atb}: Is \code{x} a non-negative-numeric atomic tibble?

\item \code{nnw_atb}: Is \code{x} a non-negative-whole-numeric atomic tibble?

\item \code{nps_atb}: Is \code{x} a non-positive-numeric atomic tibble?

\item \code{npw_atb}: Is \code{x} a non-positive-whole-numeric atomic tibble?

\item \code{nst_atb}: Is \code{x} a non-sortable atomic tibble?

\item \code{num_atb}: Is \code{x} a numeric atomic tibble?

\item \code{odd_atb}: Is \code{x} an odd-numeric atomic tibble?

\item \code{ord_atb}: Is \code{x} an ordered-factor atomic tibble?

\item \code{pct_atb}: Is \code{x} a percent-numeric atomic tibble?

\item \code{pos_atb}: Is \code{x} a positive-numeric atomic tibble?

\item \code{psw_atb}: Is \code{x} a positive-whole-numeric atomic tibble?

\item \code{ppn_atb}: Is \code{x} a proportion-numeric atomic tibble?

\item \code{srt_atb}: Is \code{x} a sortable atomic tibble?

\item \code{str_atb}: Is \code{x} a string atomic tibble?

\item \code{uno_atb}: Is \code{x} an unordered-factor atomic tibble?

\item \code{whl_atb}: Is \code{x} a whole-numeric atomic tibble?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
