% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_avl.R
\name{mmm_avl}
\alias{mmm_avl}
\alias{mmm_avl_vals}
\alias{atm_avl}
\alias{chr_avl}
\alias{ch1_avl}
\alias{clr_avl}
\alias{evn_avl}
\alias{fac_avl}
\alias{frc_avl}
\alias{ind_avl}
\alias{lgc_avl}
\alias{cnb_avl}
\alias{neg_avl}
\alias{ngw_avl}
\alias{nng_avl}
\alias{nnw_avl}
\alias{nps_avl}
\alias{npw_avl}
\alias{nst_avl}
\alias{num_avl}
\alias{odd_avl}
\alias{ord_avl}
\alias{pct_avl}
\alias{pos_avl}
\alias{ppn_avl}
\alias{psw_avl}
\alias{srt_avl}
\alias{str_avl}
\alias{uno_avl}
\alias{whl_avl}
\title{Extended Mode + Atomic vlist (mmm + avl)}
\usage{
mmm_avl_vals()

atm_avl(x)

chr_avl(x)

ch1_avl(x)

clr_avl(x)

evn_avl(x)

fac_avl(x)

frc_avl(x)

ind_avl(x)

lgc_avl(x)

cnb_avl(x)

neg_avl(x)

ngw_avl(x)

nng_avl(x)

nnw_avl(x)

nps_avl(x)

npw_avl(x)

nst_avl(x)

num_avl(x)

odd_avl(x)

ord_avl(x)

pct_avl(x)

pos_avl(x)

ppn_avl(x)

psw_avl(x)

srt_avl(x)

str_avl(x)

uno_avl(x)

whl_avl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_vlist}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
