% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_avt.R
\name{mmm_avt}
\alias{mmm_avt}
\alias{mmm_avt_vals}
\alias{chr_avt}
\alias{ch1_avt}
\alias{clr_avt}
\alias{evn_avt}
\alias{fac_avt}
\alias{frc_avt}
\alias{ind_avt}
\alias{lgl_avt}
\alias{neg_avt}
\alias{ngw_avt}
\alias{nng_avt}
\alias{nnw_avt}
\alias{nps_avt}
\alias{npw_avt}
\alias{nst_avt}
\alias{num_avt}
\alias{odd_avt}
\alias{ord_avt}
\alias{pct_avt}
\alias{pos_avt}
\alias{ppn_avt}
\alias{psw_avt}
\alias{srt_avt}
\alias{str_avt}
\alias{uno_avt}
\alias{whl_avt}
\title{Extended Mode + Atomic vtype (mmm + avt)}
\usage{
mmm_avt_vals()

chr_avt(x)

ch1_avt(x)

clr_avt(x)

evn_avt(x)

fac_avt(x)

frc_avt(x)

ind_avt(x)

lgl_avt(x)

neg_avt(x)

ngw_avt(x)

nng_avt(x)

nnw_avt(x)

nps_avt(x)

npw_avt(x)

nst_avt(x)

num_avt(x)

odd_avt(x)

ord_avt(x)

pct_avt(x)

pos_avt(x)

ppn_avt(x)

psw_avt(x)

srt_avt(x)

str_avt(x)

uno_avt(x)

whl_avt(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic vtype properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_vtp}}.
}
\section{Functions}{
\itemize{
\item \code{chr_avt}: Is \code{x} a character atomic vtype?

\item \code{ch1_avt}: Is \code{x} a onechar atomic vtype?

\item \code{clr_avt}: Is \code{x} a color atomic vtype?

\item \code{evn_avt}: Is \code{x} an even-numeric atomic vtype?

\item \code{fac_avt}: Is \code{x} a factor atomic vtype?

\item \code{frc_avt}: Is \code{x} a fractional-numeric atomic vtype?

\item \code{ind_avt}: Is \code{x} an indexer atomic vtype?

\item \code{lgl_avt}: Is \code{x} a logical atomic vtype?

\item \code{neg_avt}: Is \code{x} a negative-numeric atomic vtype?

\item \code{ngw_avt}: Is \code{x} a non-negative-whole-numeric atomic vtype?

\item \code{nng_avt}: Is \code{x} a non-negative-numeric atomic vtype?

\item \code{nnw_avt}: Is \code{x} a non-negative-whole-numeric atomic vtype?

\item \code{nps_avt}: Is \code{x} a non-positive-numeric atomic vtype?

\item \code{npw_avt}: Is \code{x} a non-positive-whole-numeric atomic vtype?

\item \code{nst_avt}: Is \code{x} a non-sortable atomic atomic vtype?

\item \code{num_avt}: Is \code{x} a numeric atomic vtype?

\item \code{odd_avt}: Is \code{x} an odd-numeric atomic vtype?

\item \code{ord_avt}: Is \code{x} an ordered-factor atomic vtype?

\item \code{pct_avt}: Is \code{x} a percent-numeric atomic vtype?

\item \code{pos_avt}: Is \code{x} a positive-numeric atomic vtype?

\item \code{ppn_avt}: Is \code{x} a proportion-numeric atomic vtype?

\item \code{psw_avt}: Is \code{x} a positive-whole-numeric atomic vtype?

\item \code{srt_avt}: Is \code{x} a sortable atomic atomic vtype?

\item \code{str_avt}: Is \code{x} a string atomic vtype?

\item \code{uno_avt}: Is \code{x} an unordered-factor atomic vtype?

\item \code{whl_avt}: Is \code{x} a whole-numeric atomic vtype?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
