% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_mat.R
\name{mmm_mat}
\alias{mmm_mat}
\alias{mmm_mat_vals}
\alias{atm_mat}
\alias{chr_mat}
\alias{ch1_mat}
\alias{clr_mat}
\alias{evn_mat}
\alias{fac_mat}
\alias{frc_mat}
\alias{ind_mat}
\alias{lgc_mat}
\alias{cnb_mat}
\alias{neg_mat}
\alias{ngw_mat}
\alias{nng_mat}
\alias{nnw_mat}
\alias{nps_mat}
\alias{npw_mat}
\alias{nst_mat}
\alias{num_mat}
\alias{odd_mat}
\alias{ord_mat}
\alias{pct_mat}
\alias{pos_mat}
\alias{ppn_mat}
\alias{psw_mat}
\alias{srt_mat}
\alias{str_mat}
\alias{uno_mat}
\alias{whl_mat}
\title{Extended Mode + Atomic matrix (mmm + mat)}
\usage{
mmm_mat_vals()

atm_mat(x)

chr_mat(x)

ch1_mat(x)

clr_mat(x)

evn_mat(x)

fac_mat(x)

frc_mat(x)

ind_mat(x)

lgc_mat(x)

cnb_mat(x)

neg_mat(x)

ngw_mat(x)

nng_mat(x)

nnw_mat(x)

nps_mat(x)

npw_mat(x)

nst_mat(x)

num_mat(x)

odd_mat(x)

ord_mat(x)

pct_mat(x)

pos_mat(x)

ppn_mat(x)

psw_mat(x)

srt_mat(x)

str_mat(x)

uno_mat(x)

whl_mat(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_matrix}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
