% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_mvc.R
\name{mmm_mvc}
\alias{mmm_mvc}
\alias{mmm_mvc_vals}
\alias{chr_mvc}
\alias{ch1_mvc}
\alias{clr_mvc}
\alias{evn_mvc}
\alias{fac_mvc}
\alias{frc_mvc}
\alias{ind_mvc}
\alias{lgl_mvc}
\alias{neg_mvc}
\alias{ngw_mvc}
\alias{nng_mvc}
\alias{nnw_mvc}
\alias{nps_mvc}
\alias{npw_mvc}
\alias{nst_mvc}
\alias{num_mvc}
\alias{odd_mvc}
\alias{ord_mvc}
\alias{pct_mvc}
\alias{pos_mvc}
\alias{ppn_mvc}
\alias{psw_mvc}
\alias{srt_mvc}
\alias{str_mvc}
\alias{uno_mvc}
\alias{whl_mvc}
\title{Extended Mode + Atomic mvect (mmm + mvc)}
\usage{
mmm_mvc_vals()

chr_mvc(x)

ch1_mvc(x)

clr_mvc(x)

evn_mvc(x)

fac_mvc(x)

frc_mvc(x)

ind_mvc(x)

lgl_mvc(x)

neg_mvc(x)

ngw_mvc(x)

nng_mvc(x)

nnw_mvc(x)

nps_mvc(x)

npw_mvc(x)

nst_mvc(x)

num_mvc(x)

odd_mvc(x)

ord_mvc(x)

pct_mvc(x)

pos_mvc(x)

ppn_mvc(x)

psw_mvc(x)

srt_mvc(x)

str_mvc(x)

uno_mvc(x)

whl_mvc(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic mvect properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_mvc}}.
}
\section{Functions}{
\itemize{
\item \code{chr_mvc}: Is \code{x} a character mvect?

\item \code{ch1_mvc}: Is \code{x} a onechar mvect?

\item \code{clr_mvc}: Is \code{x} a color mvect?

\item \code{evn_mvc}: Is \code{x} an even-numeric mvect?

\item \code{fac_mvc}: Is \code{x} a factor mvect?

\item \code{frc_mvc}: Is \code{x} a fractional-numeric mvect?

\item \code{ind_mvc}: Is \code{x} an indexer mvect?

\item \code{lgl_mvc}: Is \code{x} a logical mvect?

\item \code{neg_mvc}: Is \code{x} a negative-numeric mvect?

\item \code{ngw_mvc}: Is \code{x} a negative-whole-numeric mvect?

\item \code{nng_mvc}: Is \code{x} a non-negative-numeric mvect?

\item \code{nnw_mvc}: Is \code{x} a non-negative-whole-numeric mvect?

\item \code{nps_mvc}: Is \code{x} a non-positive-numeric mvect?

\item \code{npw_mvc}: Is \code{x} a non-positive-whole-numeric mvect?

\item \code{nst_mvc}: Is \code{x} a non-sortable atomic mvect?

\item \code{num_mvc}: Is \code{x} a numeric mvect?

\item \code{odd_mvc}: Is \code{x} an odd-numeric mvect?

\item \code{ord_mvc}: Is \code{x} an ordered-factor mvect?

\item \code{pct_mvc}: Is \code{x} a percent-numeric mvect?

\item \code{pos_mvc}: Is \code{x} a positive-numeric mvect?

\item \code{ppn_mvc}: Is \code{x} a proportion-numeric mvect?

\item \code{psw_mvc}: Is \code{x} a positive-whole-numeric mvect?

\item \code{srt_mvc}: Is \code{x} a sortable atomic mvect?

\item \code{str_mvc}: Is \code{x} a string mvect?

\item \code{uno_mvc}: Is \code{x} an unordered-factor mvect?

\item \code{whl_mvc}: Is \code{x} a whole-numeric mvect?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
