% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_scl.R
\name{mmm_scl}
\alias{mmm_scl}
\alias{mmm_scl_vals}
\alias{chr_scl}
\alias{ch1_scl}
\alias{clr_scl}
\alias{evn_scl}
\alias{fac_scl}
\alias{frc_scl}
\alias{ind_scl}
\alias{lgl_scl}
\alias{neg_scl}
\alias{ngw_scl}
\alias{nng_scl}
\alias{nnw_scl}
\alias{nps_scl}
\alias{npw_scl}
\alias{nst_scl}
\alias{num_scl}
\alias{odd_scl}
\alias{ord_scl}
\alias{pct_scl}
\alias{pos_scl}
\alias{ppn_scl}
\alias{psw_scl}
\alias{srt_scl}
\alias{chr_str}
\alias{uno_scl}
\alias{whl_scl}
\title{Extended Mode + Atomic scalar (mmm + scl)}
\usage{
mmm_scl_vals()

chr_scl(x)

ch1_scl(x)

clr_scl(x)

evn_scl(x)

fac_scl(x)

frc_scl(x)

ind_scl(x)

lgl_scl(x)

neg_scl(x)

ngw_scl(x)

nng_scl(x)

nnw_scl(x)

nps_scl(x)

npw_scl(x)

nst_scl(x)

num_scl(x)

odd_scl(x)

ord_scl(x)

pct_scl(x)

pos_scl(x)

ppn_scl(x)

psw_scl(x)

srt_scl(x)

chr_str(x)

uno_scl(x)

whl_scl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic scalar properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_scl}}.
}
\section{Functions}{
\itemize{
\item \code{chr_scl}: Is \code{x} a character scalar?

\item \code{ch1_scl}: Is \code{x} a onechar scalar?

\item \code{clr_scl}: Is \code{x} a color scalar?

\item \code{evn_scl}: Is \code{x} an even-numeric scalar?

\item \code{fac_scl}: Is \code{x} a factor scalar?

\item \code{frc_scl}: Is \code{x} a fractional-numeric scalar?

\item \code{ind_scl}: Is \code{x} an indexer scalar?

\item \code{lgl_scl}: Is \code{x} a logical scalar?

\item \code{neg_scl}: Is \code{x} a negative-numeric scalar?

\item \code{ngw_scl}: Is \code{x} a negative-whole-numeric scalar?

\item \code{nng_scl}: Is \code{x} a non-negative-numeric scalar?

\item \code{nnw_scl}: Is \code{x} an non-negative-whole-numeric scalar?

\item \code{nps_scl}: Is \code{x} a non-positive-numeric scalar?

\item \code{npw_scl}: Is \code{x} a non-positive-whole-numeric scalar?

\item \code{nst_scl}: Is \code{x} a non-sortable scalar?

\item \code{num_scl}: Is \code{x} a numeric scalar?

\item \code{odd_scl}: Is \code{x} an odd-numeric scalar?

\item \code{ord_scl}: Is \code{x} an ordered-factor scalar?

\item \code{pct_scl}: Is \code{x} a percent-numeric scalar?

\item \code{pos_scl}: Is \code{x} a positive-numeric scalar?

\item \code{ppn_scl}: Is \code{x} a proportion-numeric scalar?

\item \code{psw_scl}: Is \code{x} a positive-whole-numeric scalar?

\item \code{srt_scl}: Is \code{x} a sortable scalar?

\item \code{chr_str}: Is \code{x} a string scalar?

\item \code{uno_scl}: Is \code{x} an unordered-factor scalar?

\item \code{whl_scl}: Is \code{x} a whole-numeric scalar?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
