% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_vec.R
\name{mmm_vec}
\alias{mmm_vec}
\alias{mmm_vec_vals}
\alias{chr_vec}
\alias{clr_vec}
\alias{evn_vec}
\alias{fac_vec}
\alias{frc_vec}
\alias{ind_vec}
\alias{lgl_vec}
\alias{neg_vec}
\alias{ngw_vec}
\alias{nng_vec}
\alias{nnw_vec}
\alias{nps_vec}
\alias{npw_vec}
\alias{nst_vec}
\alias{num_vec}
\alias{odd_vec}
\alias{ord_vec}
\alias{pct_vec}
\alias{pos_vec}
\alias{ppn_vec}
\alias{psw_vec}
\alias{srt_vec}
\alias{str_vec}
\alias{uno_vec}
\alias{whl_vec}
\title{Extended Mode + Atomic vec (mmm + vec)}
\usage{
mmm_vec_vals()

chr_vec(x)

clr_vec(x)

evn_vec(x)

fac_vec(x)

frc_vec(x)

ind_vec(x)

lgl_vec(x)

neg_vec(x)

ngw_vec(x)

nng_vec(x)

nnw_vec(x)

nps_vec(x)

npw_vec(x)

nst_vec(x)

num_vec(x)

odd_vec(x)

ord_vec(x)

pct_vec(x)

pos_vec(x)

ppn_vec(x)

psw_vec(x)

srt_vec(x)

str_vec(x)

uno_vec(x)

whl_vec(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic vec properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_vec}}.
}
\section{Functions}{
\itemize{
\item \code{chr_vec}: Is \code{x} a character atomic vec?

\item \code{clr_vec}: Is \code{x} a onechar atomic vec?

\item \code{evn_vec}: Is \code{x} an even-number atomic vec?

\item \code{fac_vec}: Is \code{x} a factor atomic vec?

\item \code{frc_vec}: Is \code{x} a fractional-number atomic vec?

\item \code{ind_vec}: Is \code{x} an indexer atomic vec?

\item \code{lgl_vec}: Is \code{x} a logical atomic vec?

\item \code{neg_vec}: Is \code{x} a negative-number atomic vec?

\item \code{ngw_vec}: Is \code{x} a negative-whole-number atomic vec?

\item \code{nng_vec}: Is \code{x} a non-negative-number atomic vec?

\item \code{nnw_vec}: Is \code{x} a non-negative-whole-number atomic vec?

\item \code{nps_vec}: Is \code{x} a non-positive-number atomic vec?

\item \code{npw_vec}: Is \code{x} a non-positive-whole-number atomic vec?

\item \code{nst_vec}: Is \code{x} a non-sortable atomic vec?

\item \code{num_vec}: Is \code{x} a number atomic vec?

\item \code{odd_vec}: Is \code{x} an odd-number atomic vec?

\item \code{ord_vec}: Is \code{x} an ordered-factor atomic vec?

\item \code{pct_vec}: Is \code{x} a percent-number atomic vec?

\item \code{pos_vec}: Is \code{x} a positive-number atomic vec?

\item \code{ppn_vec}: Is \code{x} a atomic proportion-number vec?

\item \code{psw_vec}: Is \code{x} a positive-whole-number atomic vec?

\item \code{srt_vec}: Is \code{x} a sortable atomic vec?

\item \code{str_vec}: Is \code{x} a string atomic vec?

\item \code{uno_vec}: Is \code{x} a unordered-factor atomic vec?

\item \code{whl_vec}: Is \code{x} a whole-number atomic vec?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
