% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\encoding{UTF-8}
\name{move_elt}
\alias{move_elt}
\alias{move_val}
\alias{move_row}
\alias{move_col}
\title{Manipulate positions, rows, and columns}
\usage{
move_elt(X, From, To)

move_val(X, Val, To)

move_row(X, From, To)

move_col(X, From, To)
}
\arguments{
\item{X}{For \code{move_elt} and \code{move_val} an atomic vector. A data.frame or matrix for all others.}

\item{From}{Whole-number scalar indexing the element, row, or column to be moved. Negative values index from the last position, row or column.}

\item{To}{Whole-number scalar indexing the element, row, or column to be moved into. Negative values index from the last position, row or column.}

\item{Val}{A non-\code{NA} atomic scalar value found in \code{X}.}
}
\value{
An object of the same class and dimension as \code{X}.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \code{move_val}     \tab Move value \code{Val} of \code{X} into the \code{From}-th element of \code{X}. Assumes \code{length(which(X == Val)) == 1}. \cr   \tab   \cr
  \code{move_elt}     \tab Move the value in the \code{From}-th position of \code{X} to the \code{To}-th position of \code{X}.                    \cr   \tab   \cr
  \code{move_row}     \tab Move the \code{From}-th row of \code{X} to the \code{To}-th row of \code{X}.                                           \cr   \tab   \cr
  \code{move_col}     \tab Move the \code{From}-th column of \code{X} to the \code{To}-th column of \code{X}.                                                    }
\code{From < 0} and \code{To < 0} index from the last position, row, or column rather than the first.
}
\seealso{
Other function_form: 
\code{\link{insert_elts_before}()},
\code{\link{put}()},
\code{\link{rm_vals}()}
}
\concept{function_form}
