% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\encoding{UTF-8}
\name{move_pos}
\alias{move_pos}
\alias{move_val}
\alias{move_row}
\alias{move_col}
\alias{swap_pos}
\alias{swap_rows}
\alias{swap_cols}
\title{Manipulate positions, rows, and columns}
\usage{
move_pos(X, I, J)

move_val(X, V, J)

move_row(X, I, J)

move_col(X, I, J)

swap_pos(X, I, J)

swap_rows(X, I, J)

swap_cols(X, I, J)
}
\arguments{
\item{X}{For \code{move_val}, \code{move_pos}, and \code{swap_pos} an atomic vector. A data.frame or matrix for all others.}

\item{I, J}{Whole-number scalars indexing positions, row, or columns of \code{X}. Negatives index from the last position, row, or column.}

\item{V}{A non-\code{NA} atomic scalar value found in \code{X}.}
}
\value{
An object of the same class and dimension as \code{X}.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \code{move_val}     \tab Move value \code{V} of \code{X} into the \code{I}-th element of \code{X}. Assumes \code{length(which(X == V)) == 1}. \cr   \tab   \cr
  \code{move_pos}     \tab Move the value in the \code{I}-th position of \code{X} to the \code{J}-th position of \code{X}.                 \cr   \tab   \cr
  \code{move_row}     \tab Move the \code{I}-th row of \code{X} to the \code{J}-th row of \code{X}.                                        \cr   \tab   \cr
  \code{move_col}     \tab Move the \code{I}-th column of \code{X} to the \code{J}-th column of \code{X}.                                  \cr   \tab   \cr
  \code{swap_pos}     \tab Move the \code{I}-th value of \code{X} to the \code{J}-th position and vice versa.                         \cr   \tab   \cr
  \code{swap_row}     \tab Move the \code{I}-th row of \code{X} to the \code{J}-th row and vice versa.                                \cr   \tab   \cr
  \code{swap_col}     \tab Move the \code{I}-th column of \code{X} to the \code{J}-th column and vice versa.                                         }
\code{I < 0} and \code{J < 0} index from the last position, row, or column rather than the first.
}
\seealso{
Other function_form: 
\code{\link{insert_before_pos}()},
\code{\link{put}()}
}
\concept{function_form}
