% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\name{n_th.}
\alias{n_th.}
\alias{n_th}
\alias{first_n}
\alias{last_n}
\alias{n_th_last}
\title{Extract elements by index position}
\usage{
n_th.()

n_th(x, n)

first_n(x, n)

last_n(x, n)

n_th_last(x, n)
}
\arguments{
\item{x}{Vector to extract elements from.}

\item{n}{\link[cmp_psw_scl]{Complete positive whole-number scalar}.}
}
\value{
Subset of \code{x}
}
\description{
Extract elements by index position
}
\section{Functions}{
\itemize{
\item \code{n_th()}: Get the \code{n}-th element(s) of a vector (from the
front).

\item \code{first_n()}: Get the first \code{n} elements of a vector.

\item \code{last_n()}: Get the last \code{n} elements of a vector.

\item \code{n_th_last()}: Get the \code{n}-th from last element(s).

}}
\examples{
first_n(letters, 5)
last_n(letters, 5)
enth(letters, 5)
enth_last(letters, )
}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
