% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\name{n_th_uj}
\alias{n_th_uj}
\alias{n_th}
\alias{first_n}
\alias{last_n}
\alias{n_th_last}
\title{First \code{n}, last \code{n}, \code{n}th of, or \code{n}-th from
last.}
\usage{
n_th_uj()

n_th(x, n)

first_n(x, n)

last_n(x, n)

n_th_last(x, n)
}
\arguments{
\item{x}{Vector to extract elements from}

\item{n}{Positive integer vector of indexing values to extract.}
}
\value{
Subset of \code{x}
}
\description{
First \code{n}, last \code{n}, \code{n}th of, or \code{n}-th from
last.
}
\section{Functions}{
\itemize{
\item \code{n_th}: Get the \code{n}-th element(s) of a vector (from the
front).

\item \code{first_n}: Get the first \code{n} elements of a vector.

\item \code{last_n}: Get the last \code{n} elements of a vector.

\item \code{n_th_last}: Get the \code{n}-th from last element(s).
}}

\examples{
first_n(letters, 5)
last_n(letters, 5)
n_th(letters, 5)
n_th_rev(letters, )
}
