% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_vals.R
\name{na.}
\alias{na.}
\alias{na}
\alias{ok}
\alias{sub_na}
\alias{nas}
\alias{oks}
\alias{rm_na}
\title{Manage \code{NA} and non-\code{NA} values}
\usage{
na.()

na(x)

ok(x)

sub_na(x, s)

nas(x)

oks(x)

rm_na(x)
}
\arguments{
\item{x}{The argument to be inspected/managed.}

\item{s}{\link[atm_scl]{Atomic scalar} to replace \code{NA} values. Mode must
be \code{\link{compatible}} with \code{x}.}
}
\value{
An atomic scalar, an atomic vector, an atomic array, an
\link[idtf]{atomic dtf}, or an \link[ivls]{atomic vlist}.
}
\description{
All functions in this group take objects of atomic mode,
\link[idtf]{atomic dtfs}, or \link[ivls]{atomic vlists}.
}
\section{Functions}{
\itemize{
\item \code{na()}: Index \code{NA} values of \code{x}. Returns an object with
the same dimensions as \code{x} where all atomic values are replaced by
either \code{TRUE} or \code{FALSE} indicating whether the original atomic
value was \code{NA}.

\item \code{ok()}: Index non-\code{NA} values of \code{x}. Returns an object
with the same dimensions as \code{x} where all atomic values are replaced
by either \code{TRUE} or \code{FALSE} indicating whether the original
atomic value was OK (i.e., not \code{NA}).

\item \code{sub_na()}: Substitute \code{s} for \code{NA} values of \code{x}.
Returns an object with the same dimensions as \code{x} where all \code{NA}
values are replaced by \code{s}. Note that the modes of \code{x} and
\code{s} must be compatible (\code{\link{compatible}}).

\item \code{nas()}: Is \code{x} an \code{NA} scalar? Returns \code{TRUE} or
\code{FALSE} according to whether \code{x} is an \code{NA} scalar.

\item \code{oks()}: Is \code{x} a non-\code{NA} scalar? Returns \code{TRUE} or
\code{FALSE} according to whether \code{x} is a non-\code{NA} scalar.

\item \code{rm_na()}: Remove \code{NA} values of \code{x}. If there are no
\code{NA} values in \code{x}, this function returns \code{x} unchanged.
\strong{However}, if there are \code{NA} values in \code{x}, by removing
\code{NA} values from \code{x}, this function changes the dimensions of
and/or class of \code{x}. If \code{x} is an array or an
\code{\link[itab]{atomic tabular}}, its class is reduced to an atomic
vector sans \code{NA} values. If \code{x} is an \code{\link[ivls]{atomic
  vlist}}, any elements with \code{NA} values are reduced to atomic vectors
sans their \code{NA} values.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
