% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\encoding{UTF-8}
\name{na_help}
\alias{na_help}
\alias{na}
\alias{ok}
\alias{sub_na}
\alias{na0}
\alias{ok0}
\alias{rm_na}
\title{Manage \code{NA} and non-\code{NA} values}
\usage{
na_help()

na(x)

ok(x)

sub_na(x, sub)

na0(x)

ok0(x)

rm_na(x)
}
\arguments{
\item{x}{The argument to be inspected/managed.}

\item{sub}{An \link[=atm_scl]{atomic scalar} to replace \code{NA} values. Mode must be \link[=compatible]{compatible} with\code{x}}
}
\description{
All functions in this group except \code{na0} and \code{ok0} take objects of atomic mode, \link[=atm_dtf]{atomic dtfs}, or \link[=atm_vls]{atomic vlists}.
}
\section{Functions}{
\itemize{
\item \code{na()}: Logically indexes \code{NA} values of \code{x}. Returns a logical object of the same dimension as \code{x}.

\item \code{ok()}: Logically indexes non-\code{NA} values of \code{x}. Returns a logical object of the same dimension as \code{x}.

\item \code{sub_na()}: Substitutes \code{sub} for \code{NA} values of\code{x}. Modes of \code{x} and \code{sub} must be \link[=compatible]{compatible}. Returns \code{x} \emph{with} \code{NA} \emph{values replaced}.

\item \code{na0()}: Evaluates whether \code{x} is an atomic \code{NA} scalar. Returns a logical scalar.

\item \code{ok0()}: Evaluates whether \code{x} is an atomic, non-\code{NA} scalar. Returns a logical scalar.

\item \code{rm_na()}: Removes \code{NA} values of \code{x} If there are no \code{NA} values in \code{x}, returns \code{x} unchanged. If there are \code{NA} values in \code{x}, removing them may change the dimensions and/or class of \code{x}.  Returns \code{x} \emph{with} \code{NA} \emph{values removed}.

}}
\examples{
egNA0 <- NA
egOK0 <- "a"
egVec <- c(egNA0, egOK0)
egMat <- sample(vec, 10, replace = T)
egMat
na(egVec)
ok(egVec)
na(egMat)
ok(egMat)
na0(egNAS)
na0(egOK0)
na0(egVec)
ok0(egOK0)
ok0(egNAS)
ok0(egVec)
sub_na(egMat, "x")
rm_na(egMat)
}
\seealso{
Other missingness: 
\code{\link{stats0_help}()}
}
\concept{missingness}
