% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_vals.R
\name{na_vals_uj}
\alias{na_vals_uj}
\alias{na}
\alias{ok}
\alias{sub_na}
\alias{nas}
\alias{oks}
\alias{rm_na}
\title{Manage NA and Non-NA Values}
\usage{
na_vals_uj()

na(x)

ok(x)

sub_na(x, s)

nas(x)

oks(x)

rm_na(x)
}
\arguments{
\item{x}{The argument to be inspected/managed.}

\item{s}{Atomic scalar to replace \code{NA} values. Mode must be
\link[=compatible]{compatible} with \code{x}.}
}
\value{
An atomic scalar, an atomic vector, an atomic array, an
atomic tabular (\code{\link{iatb}}) or atomic vlist (\code{\link{iavl}}).
}
\description{
All functions in this group take atomic objects, atomic tibbles
(\code{\link{iatm}}) and atomic vlists (\code{\link{avl}}).
\cr\cr
\strong{\code{na}}
\cr Returns an object with the same dimensions as \code{x} where all atomic
values are replaced by either \code{TRUE} or \code{FALSE} indicating
whether the original atomic value was \code{NA}.
\cr\cr
\strong{\code{ok}}
\cr Returns an object with the same dimensions as \code{x} where all atomic
values are replaced by either \code{TRUE} or \code{FALSE} indicating
whether the original atomic value was OK (i.e., not \code{NA}).
\cr\cr
\strong{\code{nas}}
\cr Returns \code{TRUE} or \code{FALSE} according to whether \code{x} is an
\code{NA} scalar.
\cr\cr
\strong{\code{oks}}
\cr Returns \code{TRUE} or \code{FALSE} according to whether \code{x} is a
non-\code{NA} scalar.
\cr\cr
\strong{\code{sub_na}}
\cr Returns an object with the same dimensions as \code{x} where all
\code{NA} values are replaced by \code{s}. Note that the modes of \code{x}
and \code{s} must be compatible (\code{\link{compatible}}).
\cr\cr
\strong{\code{rm_na}}
If there are no \code{NA} values in \code{x}, this function returns
\code{x} unchanged.
\cr\cr
However, if there are \code{NA} values in \code{x}, by removing \code{NA}
values from \code{x}, this function changes the dimensions of and/or class
of \code{x}. If \code{x} is an array or an atomic tibble, its class is
reduced to an atomic vector sans \code{NA} values. If \code{x} is an atomic
vlist (\code{\link{iavl}}), any elements with \code{NA} values are reduced
to atomic vectors sans their \code{NA} values.
}
\section{Functions}{
\itemize{
\item \code{na}: Index \code{NA} values of \code{x}.

\item \code{ok}: Index non-\code{NA} values of \code{x}.

\item \code{sub_na}: Substitute \code{s} for \code{NA} values of \code{x}.

\item \code{nas}: Is \code{x} an \code{NA} scalar?

\item \code{oks}: Is \code{x} a non-\code{NA} scalar?

\item \code{rm_na}: Remove \code{NA} values of \code{x}.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
