% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{naming.}
\alias{naming.}
\alias{namev}
\alias{namer}
\alias{namec}
\alias{namel}
\alias{namerc}
\alias{named}
\alias{enamed}
\alias{rnamed}
\alias{cnamed}
\alias{rcnamed}
\alias{named.}
\alias{getnames}
\alias{en}
\alias{rn}
\alias{cn}
\alias{rcn}
\title{Get or set element, row, and/or column names.}
\usage{
naming.()

namev(x., v.)

namer(x., r.)

namec(x., c.)

namel(x., n.)

namerc(x., r., c.)

named(x., dim. = 0, u. = T, blank. = F)

enamed(x., u. = T, blank. = F)

rnamed(x., u. = T, blank. = F)

cnamed(x., u. = T, blank. = F)

rcnamed(x., u. = T, blank. = F)

named.(..., u. = T, blank. = F)

getnames(x., d. = 0, u. = T, err. = F)

en(x., u. = T, err. = F)

rn(x., u. = T, err. = F)

cn(x., u. = T, err. = F)

rcn(x., u. = T, err. = F)
}
\arguments{
\item{x.}{Argument to be inspected for names or to be named..}

\item{v.}{Character vector of element names.}

\item{r., c.}{Character vectors of element, row, or column names,
respectively.}

\item{n.}{Character scalar element-name.}

\item{dim.}{Dimension \code{0} for elements (of atomic or list vectors only),
\code{1} for rows of matrices or tibbles, \code{2} for columns of matrices
or tibbles, or \code{12} for rows and columns of matrices or tibbles.}

\item{u.}{Logical scalar indicating whether names must be unique.}

\item{blank.}{Logical scalar indicating whether blank strings (\code{""}) are
allowed as names.}

\item{...}{Arbitrary number of arguments to be inspected for names.}

\item{d.}{Integer scalar indicating dimensions to get names for. Valid values
are \code{c(1, 2, 12)} for row, column, and both.}

\item{err.}{Logical scalar indicating whether names must be non-missing.}

\item{x}{Vector or vlist for \code{namev} and \code{vn}; matrix or data.frame
for \code{rn}, \code{namer}, \code{cn}, \code{namec}, \code{namerc}; or any
object for \code{namel}.}
}
\value{
Character vector or a named/renamed version of \code{x}.
}
\description{
Get or set element, row, and/or column names.
}
\section{Functions}{
\itemize{
\item \code{namev()}: Name the elements of a vector.

\item \code{namer()}: Name the rows of a matrix or data frame.

\item \code{namec()}: Name the columns of a matrix or data frame.

\item \code{namel()}: Create a 1-element list from \code{x} and name that
element with the value of \code{e}

\item \code{namerc()}: Name the rows and columns of a matrix or data frame.

\item \code{named()}: Determine whether an objects elements, rows, and/or
columns are named, accounting for whether names must be unique, whether the
object may contain empty elements, and whether names may be blank strings
("").

\item \code{enamed()}: Are elements of \code{x} named?

\item \code{rnamed()}: Are rows of \code{x} named?

\item \code{cnamed()}: Are columns of \code{x} named?

\item \code{rcnamed()}: Are both rows and columns of \code{x} named?

\item \code{named.()}: Are all \code{...} arguments named?

\item \code{getnames()}: Get element, row, and/or column names with optional
restrictions.

\item \code{en()}: Get element names.

\item \code{rn()}: Get row names.

\item \code{cn()}: Get column names.

\item \code{rcn()}: Get row and column names (in a 2-element list).

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
