% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\encoding{UTF-8}
\name{naming_help}
\alias{naming_help}
\alias{name_vals}
\alias{name_vs}
\alias{name_elts}
\alias{name_es}
\alias{name_rows}
\alias{name_rs}
\alias{name_cols}
\alias{name_cs}
\alias{name_scl_vls}
\alias{name_sv}
\alias{name_rows_cols}
\alias{name_rcs}
\alias{name}
\alias{named}
\alias{named_elts}
\alias{elts_named}
\alias{named_es}
\alias{es_named}
\alias{named_vals}
\alias{vals_named}
\alias{named_vs}
\alias{vs_named}
\alias{named_rows}
\alias{rows_named}
\alias{named_rs}
\alias{rs_named}
\alias{named_cols}
\alias{cols_named}
\alias{named_cs}
\alias{cs_named}
\alias{named_rows_cols}
\alias{rows_cols_named}
\alias{named_rcs}
\alias{rcs_named}
\alias{all_dots_named}
\alias{get_names}
\alias{enames}
\alias{en}
\alias{vnames}
\alias{vn}
\alias{rnames}
\alias{rn}
\alias{cnames}
\alias{cn}
\alias{row_col_names}
\alias{rcnames}
\alias{rcn}
\title{Naming utilities}
\usage{
naming_help()

name_vals(x, vn)

name_vs(x, vn)

name_elts(x, en)

name_es(x, en)

name_rows(x, rn)

name_rs(x, rn)

name_cols(x, cn)

name_cs(x, cn)

name_scl_vls(x, en)

name_sv(x, en)

name_rows_cols(x, rn, cn)

name_rcs(x, rn, cn)

name(x, en = NULL, vn = NULL, rn = NULL, cn = NULL)

named(x, d = 0, u = T, bl = F)

named_elts(x, u = T, bl = F)

elts_named(x, u = T, bl = F)

named_es(x, u = T, bl = F)

es_named(x, u = T, bl = F)

named_vals(x, u = T, bl = F)

vals_named(x, u = T, bl = F)

named_vs(x, u = T, bl = F)

vs_named(x, u = T, bl = F)

named_rows(x, u = T, bl = F)

rows_named(x, u = T, bl = F)

named_rs(x, u = T, bl = F)

rs_named(x, u = T, bl = F)

named_cols(x, u = T, bl = F)

cols_named(x, u = T, bl = F)

named_cs(x, u = T, bl = F)

cs_named(x, u = T, bl = F)

named_rows_cols(x, u = T, bl = F)

rows_cols_named(x, u = T, bl = F)

named_rcs(x, u = T, bl = F)

rcs_named(x, u = T, bl = F)

all_dots_named(...)

get_names(x, d = 0, u = T, err = F)

enames(x, u = T, err = F)

en(x, u = T, err = F)

vnames()

vn(x, u = T, err = F)

rnames(x, u = T, err = F)

rn(x, u = T, err = F)

cnames(x, u = T, err = F)

cn(x, u = T, err = F)

row_col_names(x, u = T, err = F)

rcnames(x, u = T, err = F)

rcn(x, u = T, err = F)
}
\arguments{
\item{x}{Vector or list for \code{name_vals}/\code{name_elts} and \code{val_names / elt_names / vn / en}; matrix or data.frame for \code{rnames / rn}, \code{cnames / cn}, \code{rcnames / rcn}, \code{name_rcs}, \code{name_cols}, and \code{name_rows}; or any object for \code{name_scl_vls}.}

\item{en, vn, rn, cn}{A \link[=cmp_chr_vec]{complete character vec} element, value, row, and column named, respectively.}

\item{d}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the dimension(s) to name: \code{0} for elements (of atomic or list vectors only), \code{1} for rows of matrices or tibbles, \code{2} for columns of matrices or data.frames, or \code{12} for rows and columns of matrices or data.frames.}

\item{u}{\code{TRUE} or \code{FALSE} indicating whether names and \code{...} argument names must be unique, respectively.}

\item{bl}{\code{TRUE} or \code{FALSE} indicating whether blank strings (\code{""}) are allowed as names and \code{...} argument names, respectively.}

\item{...}{An arbitrary number of arguments.}

\item{err}{\code{TRUE} or \code{FALSE} indicating whether names and \code{...} argument names must exist, respectively.}

\item{name}{A non-\code{NA} character scalar to name the single element of the resulting \link[=VLS]{vlist}.}
}
\description{
Get object names (with optional restrictions), check whether names exist (with optional restrictions), and apply names.
}
\section{Functions}{
\itemize{
\item \code{name_vals()}: Named the values/elements of \code{x} as \code{vn}. Returns \code{x} with names applied.

\item \code{name_vs()}: An alias for \code{name_vals}.

\item \code{name_elts()}: Name the values/elements of \code{x} as \code{en}

\item \code{name_es()}: An alias for \code{name_elts}.

\item \code{name_rows()}: Name the rows of \code{x} as \code{rn}. Returns \code{x} with row names applied.

\item \code{name_rs()}: An alias for \code{name_rows}.

\item \code{name_cols()}: Name the columns of \code{x} as \code{cn}. Returns \code{x} with column names applied.

\item \code{name_cs()}: An alias for \code{name_cols}.

\item \code{name_scl_vls()}: Store \code{x} in a single-element list and name that element \code{en}. Returns a named, single-element list.

\item \code{name_sv()}: An alias for \code{name_scl_vls}.

\item \code{name_rows_cols()}: Name the rows of \code{x} as \code{rn} and the columns of \code{x} as \code{cn}. Returns \code{x} with row and column names applied.

\item \code{name_rcs()}: An alias for \code{name_rows_cols}.

\item \code{name()}: Name the values/elements of, rows of, and/or columns of \code{x}. Returns \code{x} with specified names applied.

\item \code{named()}: Evaluates whether elements/rows/columns/drawers/... of \code{x} are named, optionally checking for uniqueness and non-blankness. Returns a logical scalar or throws an error.

\item \code{named_elts()}: Evaluates whether elements/values of \code{x} are named, optionally checking for uniqueness and non-blankness. Returns a logical scalar or throws an error.

\item \code{elts_named()}: An alias for \code{named_elts}.

\item \code{named_es()}: An alias for \code{named_elts}.

\item \code{es_named()}: An alias for \code{named_elts}.

\item \code{named_vals()}: An alias for \code{named_elts}.

\item \code{vals_named()}: An alias for \code{named_elts}.

\item \code{named_vs()}: An alias for \code{named_elts}.

\item \code{vs_named()}: An alias for \code{named_elts}.

\item \code{named_rows()}: Evaluates whether rows of \code{x} are named, optionally checking for uniqueness and non-blankness. Returns a logical scalar or throws an error.

\item \code{rows_named()}: An alias for \code{named_rows}.

\item \code{named_rs()}: An alias for \code{named_rows}.

\item \code{rs_named()}: An alias for \code{named_rows}.

\item \code{named_cols()}: Evaluates whether columns of \code{x} are named, optionally checking for uniqueness and non-blankness. Returns a logical scalar or throws an error.

\item \code{cols_named()}: An alias for \code{named_cols}.

\item \code{named_cs()}: An alias for \code{named_cols}.

\item \code{cs_named()}: An alias for \code{named_cols}.

\item \code{named_rows_cols()}: Evaluates whether rows and columns of \code{x} are named, optionally checking for uniqueness and non-blankness. Returns a logical scalar or throws an error.

\item \code{rows_cols_named()}: An alias for \code{named_rows_cols}.

\item \code{named_rcs()}: An alias for \code{named_rows_cols}.

\item \code{rcs_named()}: An alias for \code{named_rows_cols}.

\item \code{all_dots_named()}: Evaluates whether all \code{...} arguments are named. Returns a logical scalar or throws an error.

\item \code{get_names()}: Generalized name retrieval (elements/values/rows/cols/drawers/...), optionally checking for uniqueness. Returns either a character vector of a list of two character vectors, or throws an error.

\item \code{enames()}: Gets names of the elements/values of \code{x}, optionally checking for uniqueness. Returns a character vector or throws an error.

\item \code{en()}: An alias for \code{enames}.

\item \code{vnames()}: An alias for \code{enames}.

\item \code{vn()}: An alias for \code{enames}.

\item \code{rnames()}: Gets names of the rows of \code{x}, optionally checking for uniqueness. Returns a character vector or throws an error.

\item \code{rn()}: An alias for \code{rnames}.

\item \code{cnames()}: Gets names of the columns of \code{x}, optionally checking for uniqueness. Returns a character vector or throws an error.

\item \code{cn()}: An alias for \code{cnames}.

\item \code{row_col_names()}: Gets names of the rows and columns of \code{x}, optionally checking for uniqueness. Returns a character vector or throws an error.

\item \code{rcnames()}: An alias for \code{row_col_names}.

\item \code{rcn()}: An alias for \code{row_col_names}.

}}
\examples{
egDotsNamed <- function(...) {dnamed(...)}

egVec <- 1:5
egVls <- list(num = 1:5, ae = letters[1:5])
egMat <- matrix(letters[1:25], nrow = 5)
egDtf <- data.frame(ae = letters[1:5], AE = LETTERS[1:5])

egVec
egVls
egMat
egDtf

named(egVec)
named(egVls)
named(egMat, d = 12)
named(egDtf, d = 2)

vals_are_named(egVls)
rows_are_named(egMat)
cols_are_named(egDtf)
rcs_are_named(egDtf)

dots_are_named(egVec, egVls)
dots_are_named(var1 = egVec, var2 = egVls)

get_names(egVls)
get_names(egMat, d = 1)
get_names(egDtf, d = 2)

en(egVec)
en(egVls)

rn(egMat)
cn(egDtf)
rcn(egDtf)

name_scl_vls(egDtf, "a data.frame")
name_vals(egVec, letters[egVec])
name_rows(egMat, letters[egVec])
name_cols(egDtf, c("letters", "LETTERS"))
name_rcs(egMat, letters[egVec], egVec)
}
