% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{not_err}
\alias{not_err}
\title{Does Evaluating \code{x} NOT Create an Error?}
\usage{
not_err(x)
}
\arguments{
\item{x}{Object to be evaluated.}
}
\value{
A logical scalar.
}
\description{
Does Evaluating \code{x} NOT Create an Error?
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}
}
\concept{utils}
