% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nth.R
\encoding{UTF-8}
\name{Nth}
\alias{Nth}
\alias{firstN}
\alias{lastN}
\alias{Nth_from_last}
\alias{Nth_plus}
\title{Extract elements by numeric position}
\usage{
Nth(x, N)

firstN(x, N)

lastN(x, N)

Nth_from_last(x, N)

Nth_plus(x, N)
}
\arguments{
\item{x}{A \link[=pop_vec]{populated vector} to extract elements from.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}.}
}
\value{
A scalar, vector, or \link[=VLS]{vlist}.
}
\description{
Extract the \code{N}-th, \code{n}-th from last, first \code{n}, and last \code{n} elements of a vector of list.
\tabular{ll}{  \code{Nth_from_last}   \tab Gets the \code{N}th from last element(s). \cr
  \code{Nth_plus}        \tab Gets element \code{N} and onwards.        \cr
  \code{firstN}          \tab Gets the first \code{N} elements.         \cr
  \code{lastN}           \tab Gets the last \code{N} elements.          \cr
  \code{Nth}             \tab Gets the \code{N}-th element(s).            }
}
\examples{
Nth_from_last(letters, 5)
Nth_plus(letters, 5)
firstN(letters, 5)
lastN(letters, 5)
Nth(letters, 5:7)
}
\seealso{
Other values: 
\code{\link{atomize}()},
\code{\link{na}()},
\code{\link{values}()}
}
\concept{to_std_fun_form}
\concept{values}
