% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\encoding{UTF-8}
\name{nth_help}
\alias{nth_help}
\alias{.nth}
\alias{.first_n}
\alias{.last_n}
\alias{.nth_from_last}
\alias{.nth_plus}
\title{Extract elements by numeric position}
\usage{
nth_help()

.nth(x, n)

.first_n(x, n)

.last_n(x, n)

.nth_from_last(x, n)

.nth_plus(x, n)
}
\arguments{
\item{x}{A \link[=pop_vec]{populated vector} to extract elements from.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}.}
}
\value{
A scalar, vector, or \link[=VLS]{vlist}.
}
\description{
Extract the \code{n}-th, \code{n}-th from last, first \code{n}, and last \code{n} elements of a vector of list.
}
\section{Functions}{
\itemize{
\item \code{.nth()}: Gets the \code{n}-th element(s) of \code{x}.

\item \code{.first_n()}: Gets the first \code{n} elements of \code{x}.

\item \code{.last_n()}: Gets the last \code{n} elements of \code{x}.

\item \code{.nth_from_last()}: Gets the \code{n}th from last element(s) of \code{x}.

\item \code{.nth_plus()}: Gets element \code{n} and onwards of \code{x}.

}}
\examples{
.nth_from_last(letters, 5)
.nth_plus(letters, 5)
.first_n(letters, 5)
.last_n(letters, 5)
.nth(letters, 5:7)
}
