% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{ok_fmt}
\alias{ok_fmt}
\title{Validate a Formatting Spec (of form "\link{bg.color}|\link{fg.color}|\link{style}")}
\usage{
ok_fmt(x)
}
\arguments{
\item{x}{A character scalar formatting spec.}
}
\value{
A logical scalar
}
\description{
Validate a Formatting Spec (of form "\link{bg.color}|\link{fg.color}|\link{style}")
}
\examples{
.ok_fmt("")
.ok_fmt(NULL)
.ok_fmt("q|r|b")
.ok_fmt("y|r|b")
.ok_fmt("  yellow | red | bold  ")
}
\seealso{
Other alert_utils: 
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{glu}()},
\code{\link{match_bg}()},
\code{\link{ssplit}()}
}
\concept{alert_utils}
