% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os.}
\alias{os.}
\alias{os}
\title{Identify the current operating system}
\usage{
os.()

os()
}
\value{
Character scalar
}
\description{
Identify the current operating system
}
\section{Functions}{
\itemize{
\item \code{os()}: Identifies the operating system as 'win', 'mac', or 'unix'.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}

Other environments: 
\code{\link{callers.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{values.}()}
}
\concept{environments}
\concept{extensions}
