% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os}
\alias{os}
\title{Identify current operating system}
\usage{
os()
}
\value{
Character scalar
}
\description{
Identifies the operating system as 'win', 'mac', or 'unix'.
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
