% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/ox.R
\encoding{UTF-8}
\name{ox_vals}
\alias{ox_vals}
\title{Simple Oxford-Comma Lists}
\usage{
ox_vals(x, join)

ox_vals(x, join)
}
\arguments{
\item{x}{A character vector to be formatted as an Oxford-comma list.}

\item{join}{A character scalar conjunction for between the next to last and last elements of \code{x}.}
}
\value{
A character scalar.
egOxVals <- function() {
base::list(and.letters = uj::ox_vals(letters, "and"),
or.digits   = uj::ox_vals(0:9    , "or" ))
}

egOxVals()

A character scalar.
egOxVals <- function() {
base::list(and.letters = uj::ox_vals(letters, "and"),
or.digits   = uj::ox_vals(0:9    , "or" ))
}

egOxVals()
}
\description{
Simple Oxford-Comma Lists

Simple Oxford-Comma Lists
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown_help}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid_help}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss_help}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
