% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\docType{data}
\name{pals_uj}
\alias{pals_uj}
\alias{pal_colors}
\alias{pal_lines}
\alias{pal_fdots}
\alias{pal_ldots}
\alias{pal_edots}
\alias{pal_sdots}
\title{Plotting Palettes}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
pals_uj

pal_colors(n)

pal_lines(n)

pal_fdots(n)

pal_ldots(n)

pal_edots(n)

pal_sdots(n)
}
\arguments{
\item{n}{\link[=cmp_psw_scl]{Complete positive whole-number scalar}
indicating the number of unique values to return.}
}
\value{
\strong{\code{pal_colors, pal_lines}}
\cr A character vector.
\cr\cr
\strong{\code{pal_edots, pal_fdots, pal_ldots, pal_sdots}}
\cr A positive whole number vector.
}
\description{
Plotting Palettes
}
\section{Functions in This Family}{

\strong{\code{pal_colors}}
\cr Creates a color palette with the first \code{n} color values from
\code{v(colors)}.
\cr\cr
\strong{\code{pal_lines}}
\cr Creates a line-type palette with the first \code{n} line-type values
from \code{v(lines)}.
\cr\cr
\strong{\code{pal_edots}}
\cr Creates an empty-dots/shapes palette with the first \code{n} empty-dot
values from \code{v(edots)}. Empty dots are closed shapes with an unfilled
interior.
\cr\cr
\strong{\code{pal_fdots}}
\cr Creates a fillable-dots/shapes palette with the first \code{n}
empty-dot values from \code{v(fdots)}. Fillable dots are closed shapes with
a fillable interior.
\cr\cr
\strong{\code{pal_ldots}}
\cr Creates a line-only dots/shapes palette with the first \code{n}
line-only dot values from \code{v(ldots)}. Line-only dots are non-enclosing
shapes.
\cr\cr
\strong{\code{pal_sdots}}
\cr Creates a solid-dots/shapes palette with the first first \code{n}
solid-dot values from \code{v(sdots)}.
\cr\cr
NOTE: If \code{n > v(.)}, values are recycled.
}

\seealso{
Other plots: 
\code{\link{mkd}}

Other colors: 
\code{\link{color}},
\code{\link{mkd}}
}
\concept{colors}
\concept{plots}
\keyword{datasets}
