% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{is_path}
\alias{as_path}
\alias{object_path}
\alias{parent_path}
\alias{object_name}
\alias{parent_name}
\alias{object_dirs}
\alias{parent_dirs}
\alias{newdirs}
\title{Evaluate and manipulate paths to files or folders}
\usage{
is_path(..., err = F)

as_path(...)

object_path(path, err = T)

parent_path(path, err = T)

object_name(path, err = T)

parent_name(path, err = T)

object_dirs(path, err = T)

parent_dirs(path, err = T)

newdirs(dirs, path = NULL)
}
\arguments{
\item{...}{Atomic arguments pasted to create}

\item{err}{A non-\code{NA} logical scalar indicating whether an error should be thrown if \code{...} is not a valid object path when collapsed to a character scalar.}

\item{path}{A \link[=cmp_chr_scl]{complete character scalar} path to either a directory or a file.}

\item{dirs}{A \link[=cmp_chr_vec]{complete character vec} of new sub-directory names.}
}
\value{
\itemize{
\item \strong{\code{is_path}}: a logical scalar.
\item \strong{\verb{object_dirs, folder_dirs}}: a character vector.
\item \strong{\verb{object_path, folder_path, object_name, folder_name}}: a character scalar.
}
}
\description{
\itemize{
\item \strong{\code{is_path}}: checks whether \code{...} resolves to a valid path for an object (file or folder) when \code{...} arguments are collapsed into a character scalar.
\item \strong{\code{as_path}}: collapses \code{...} into a path using the current platform file path separator \code{.Platform$file.sep}.
\item \strong{\code{new_dirs}}: creates sub-directories within an existing directory, optionally asking the user to choose that existing directory.
\item \strong{\code{object_path}}: collapses \code{...} into a character scalar and expand the path to the object indicated by the result to account for relative paths.
\item \strong{\code{parent_path}}: calls \code{object_path} and extract from the resulting character scalar just the path to the parent folder of the object specified by \code{...} (i.e., discarding the name of the object itself).
\item \strong{\code{object_name}}: calls \code{object_path} and extract from the resulting character scalar just the name of the object, discarding the path to its parent.
\item \strong{\code{parent_name}}: calls \code{parent_path} and extract from the resulting character scalar just the name of last folder in the path.
\item \strong{\code{object_dirs}}: calls \code{object_path} and split the resulting character scalar into a character vector containing the name of the root folder, the names of any intermediate folders, the name of the parent folder of the object, and the name of the object.
\item \strong{\code{parent_dirs}}: calls \code{parent_path} and split the resulting character scalar into a character vector containing the name of the root folder, the names of any intermediate folders, and the name of the parent folder of the object specified by \code{...}.
}
}
\details{
In most cases, if (optionally) specified, functions throw an error if \code{...} does not resolve to a valid object path.
}
\seealso{
Other files: 
\code{\link{rd}}
}
\concept{files}
