% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause.}
\alias{pause.}
\alias{pause}
\title{Pause execution.}
\usage{
pause.()

pause(x = 1e-07)
}
\arguments{
\item{x}{A \link[=cmp_num_scl]{complete mumeric scalar} indicating the number
of seconds to pause.}
}
\value{
\code{NULL} (called for its side effect).
}
\description{
Pause for \code{x} seconds, then resume execution.
}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
