% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause..R
\name{pause}
\alias{pause}
\title{Pause execution.}
\usage{
pause(x = 1e-07)
}
\arguments{
\item{x}{A \link[=cmp_num_scl]{complete mumeric scalar} indicating the number of seconds to pause.}
}
\value{
\code{NULL} (called for its side effect).
}
\description{
Pause for \code{x} seconds, then resume execution.
}
\examples{
cat("\n pause(0.01)")
pause(0.01)
cat("\n pause(0.1)")
pause(0.1)
cat("\n pause(0.2)")
pause(0.2)
cat("\n pause(0.5)")
pause(0.5)
cat("\n pause(1)")
pause(1)
cat("\n done")
}
