% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause_uj}
\alias{pause_uj}
\alias{pause}
\title{Pause execution.}
\usage{
pause_uj()

pause(x = 1e-07)
}
\arguments{
\item{x}{Numeric scalar number of seconds to pause.}
}
\value{
\code{NULL} (called for side effect).
}
\description{
Pause execution.
}
\section{Functions}{
\itemize{
\item \code{pause}: Pause for \code{x} seconds, then resume execution.
}}

