% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrid.R
\name{pgrid.}
\alias{pgrid.}
\alias{pgrid}
\alias{pgrid0}
\title{\code{expand.grid} for \code{paste} and \code{paste0}}
\usage{
pgrid.()

pgrid(..., p. = " ", ch. = F, blank. = F, na.err. = T)

pgrid0(..., ch. = F, blank. = F, na.err. = T)
}
\arguments{
\item{...}{Non-empty atomic objects.}

\item{p.}{\link[cmp_chr_scl]{Complete character scalar} to use as the
'paste'.}

\item{ch., blank., na.err.}{\link[cmp_lgl_scl]{Complete logical scalars}
indicating, respectively, whether to split arguments in \code{...} into
constituent characters after conversion to mode 'character', whether to add
a blank string element to each argument in \code{...}, and whether to throw
an error if an argument in \code{...} contains an \code{NA} value.}
}
\value{
A character vector.
}
\description{
\code{expand.grid} for \code{paste} and \code{paste0}
}
\section{Functions}{
\itemize{
\item \code{pgrid()}: Convert arguments in \code{...} to character, separating
them into their constituent characters (if \code{ch = T}), add a blank
string (if \code{add.blank = T}), and create a character vector containing
all possible combinations of elements in \code{...} pasted together using
the 'paste' \code{p}. The first part of the character scalar in each
element of the result is from \code{...1}, the second from \code{..2}, and
so on.

\item \code{pgrid0()}: Call \code{pgrid} with \code{p. = ""} (a blank string).

}}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}
}
\concept{strings}
