% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrid.R
\name{pgrid_uj}
\alias{pgrid_uj}
\alias{pgrid}
\alias{pgrid0}
\title{\code{expand.grid} for \code{paste} and \code{paste0}}
\usage{
pgrid_uj()

pgrid(..., p. = " ", ch. = F, add.blank. = F)

pgrid0(..., ch. = F, add.blank. = F)
}
\arguments{
\item{...}{Non-empty atomic objects.}

\item{p.}{A character scalar to use as the 'paste'.}

\item{ch.}{A logical scalar indicating whether to split arguments in
\code{...} into constituent characters after conversion to mode
'character'.}

\item{add.blank.}{Whether to add a blank ("") to each argument in \code{...}.}
}
\value{
A character vector.
}
\description{
\code{expand.grid} for \code{paste} and \code{paste0}
}
\section{Functions}{
\itemize{
\item \code{pgrid}: Convert arguments in \code{...} to character, separating
them into their constituent characters (if \code{ch = T}), add a blank
string (if \code{add.blank = T}), and create a character vector containing
all possible combinations of elements in \code{...} pasted together using
the 'paste' \code{p}. The first part of the character scalar in each
element of the result is from \code{...1}, the second from \code{..2}, and
so on.

\item \code{pgrid0}: Call \code{pgrid} with \code{p. = ""} (a blank string).
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{revstr_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
