% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_xxx.R
\name{pop_xxx}
\alias{pop_xxx}
\alias{pop_xxx_vals}
\alias{pop_arr}
\alias{pop_atm}
\alias{pop_mat}
\alias{pop_rcr}
\alias{pop_tib}
\alias{pop_vls}
\alias{pop_vtp}
\title{Populated + Fundamental Type or Extended Class (pop + ttt or ccc)}
\usage{
pop_xxx_vals()

pop_arr(x)

pop_atm(x)

pop_mat(x)

pop_rcr(x)

pop_tib(x)

pop_vls(x)

pop_vtp(x)
}
\arguments{
\item{x}{An object}
}
\description{
See \code{\link{xpop}}, \code{link{xatm}}, \code{link{xrcr}},
\code{link{is_vtype}}, \code{link{is.matrix}},
\code{link[tibble]{is_tibble}}, and \code{\link{is_vlist}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
