% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_fast.R
\name{ppp_fast}
\alias{ppp_fast}
\alias{.ATM}
\alias{.DEF}
\alias{.FUN}
\alias{.NIL}
\alias{.NLL}
\alias{.POP}
\alias{.RCR}
\alias{.ARR}
\alias{.DTF}
\alias{.GEN}
\alias{.MAT}
\alias{.MVC}
\alias{.SCL}
\alias{.VLS}
\alias{.VEC}
\alias{.D0D}
\alias{.D1D}
\alias{.D2D}
\alias{.DHD}
\alias{.EUD}
\alias{.E2D}
\alias{.EHD}
\alias{.E0D}
\alias{.E1D}
\alias{.NA0}
\alias{.OK0}
\alias{.DUP}
\alias{.UNQ}
\alias{.CMP}
\alias{.MSS}
\alias{.PRT}
\alias{.CH1}
\alias{.CH3}
\alias{.CHR}
\alias{.STR}
\alias{.CLR}
\alias{.FAC}
\alias{.ORD}
\alias{.UNO}
\alias{.IND}
\alias{.NST}
\alias{.SRT}
\alias{.LGL}
\alias{.NUM}
\alias{.WHL}
\alias{.ODD}
\alias{.EVN}
\alias{.NGW}
\alias{.NPW}
\alias{.PSW}
\alias{.FRC}
\alias{.NEG}
\alias{.POS}
\alias{.NNW}
\alias{.NPS}
\alias{.PCT}
\alias{.PPN}
\alias{.COL}
\alias{.ROW}
\alias{.PNT}
\alias{.SLD}
\alias{.SQR}
\alias{.EMP}
\alias{.RCT}
\alias{.LIN}
\alias{.atm_dtf}
\alias{.atm_mat}
\alias{.atm_scl}
\alias{.atm_vec}
\alias{.atm_vls}
\alias{.chr_arr}
\alias{.chr_dtf}
\alias{.chr_vec}
\alias{.chr_vls}
\alias{.cmp_atm}
\alias{.cmp_chr}
\alias{.cmp_num}
\alias{.cmp_scl}
\alias{.cmp_vec}
\alias{.lgl_scl}
\alias{.nnw_vec}
\alias{.pop_atm}
\alias{.pop_chr}
\alias{.pop_dtf}
\alias{.pop_mat}
\alias{.pop_srt}
\alias{.pop_vec}
\alias{.pop_vls}
\alias{.cmp_atm_vec}
\alias{.cmp_ch1_scl}
\alias{.cmp_ch1_vec}
\alias{.cmp_ch3_scl}
\alias{.cmp_ch3_vec}
\alias{.cmp_chr_arr}
\alias{.cmp_chr_vec}
\alias{.cmp_chr_vls}
\alias{.cmp_chr_gen}
\alias{.cmp_chr_scl}
\alias{.cmp_clr_vec}
\alias{.cmp_lgl_scl}
\alias{.cmp_lgl_vec}
\alias{.cmp_nng_scl}
\alias{.cmp_nnw_scl}
\alias{.cmp_nnw_vec}
\alias{.cmp_num_scl}
\alias{.cmp_num_vec}
\alias{.cmp_pos_vec}
\alias{.cmp_ppn_vec}
\alias{.cmp_psw_scl}
\alias{.cmp_psw_vec}
\alias{.cmp_srt_vec}
\alias{.cmp_str_scl}
\alias{.cmp_str_vec}
\alias{.cmp_whl_scl}
\alias{.cmp_atm_mvc}
\alias{.unq_atm_mvc}
\alias{.unq_atm_vec}
\alias{.unq_ch1_vec}
\alias{.unq_ch3_vec}
\alias{.unq_chr_vec}
\alias{.unq_nnw_vec}
\alias{.unq_num_vec}
\alias{.unq_psw_vec}
\alias{.unq_str_vec}
\alias{.atm_rct_dtf}
\title{Fast property checking for common checks}
\usage{
ppp_fast()

.ATM(x, .VALID = NULL)

.DEF(x, .VALID = NULL)

.FUN(x, .VALID = NULL)

.NIL(x)

.NLL(x)

.POP(x, .VALID = NULL)

.RCR(x, .VALID = NULL)

.ARR(x, .VALID = NULL)

.DTF(x, .VALID = NULL)

.GEN(x, .VALID = NULL)

.MAT(x, .VALID = NULL)

.MVC(x, .VALID = NULL)

.SCL(x, .VALID = NULL)

.VLS(x, .VALID = NULL)

.VEC(x, .VALID = NULL)

.D0D(x)

.D1D(x, .VALID = NULL)

.D2D(x, .VALID = NULL)

.DHD(x, .VALID = NULL)

.EUD(x)

.E2D(x, .VALID = NULL)

.EHD(x, .VALID = NULL)

.E0D(x, .VALID = NULL)

.E1D(x, .VALID = NULL)

.NA0(x)

.OK0(x, .VALID = NULL)

.DUP(x, .VALID = NULL)

.UNQ(x, .VALID = NULL)

.CMP(x, .VALID = NULL)

.MSS(x)

.PRT(x, .VALID = NULL)

.CH1(x, .VALID = NULL)

.CH3(x, .VALID = NULL)

.CHR(x, .VALID = NULL)

.STR(x, .VALID = NULL)

.CLR(x, .VALID = NULL)

.FAC(x, .VALID = NULL)

.ORD(x, .VALID = NULL)

.UNO(x, .VALID = NULL)

.IND(x, .VALID = NULL)

.NST(x, .VALID = NULL)

.SRT(x, .VALID = NULL)

.LGL(x, .VALID = NULL)

.NUM(x, .VALID = NULL)

.WHL(x, .VALID = NULL)

.ODD(x, .VALID = NULL)

.EVN(x, .VALID = NULL)

.NGW(x, .VALID = NULL)

.NPW(x, .VALID = NULL)

.PSW(x, .VALID = NULL)

.FRC(x, .VALID = NULL)

.NEG(x, .VALID = NULL)

.POS(x, .VALID = NULL)

.NNW(x, .VALID = NULL)

.NPS(x, .VALID = NULL)

.PCT(x, .VALID = NULL)

.PPN(x, .VALID = NULL)

.COL(x, .VALID = NULL)

.ROW(x, .VALID = NULL)

.PNT(x, .VALID = NULL)

.SLD(x, .VALID = NULL)

.SQR(x, .VALID = NULL)

.EMP(x, .VALID = NULL)

.RCT(x, .VALID = NULL)

.LIN(x, .VALID = NULL)

.atm_dtf(x, .VALID = NULL)

.atm_mat(x, .VALID = NULL)

.atm_scl(x, .VALID = NULL)

.atm_vec(x, .VALID = NULL)

.atm_vls(x, .VALID = NULL)

.chr_arr(x, .VALID = NULL)

.chr_dtf(x, .VALID = NULL)

.chr_vec(x, .VALID = NULL)

.chr_vls(x, .VALID = NULL)

.cmp_atm(x, .VALID = NULL)

.cmp_chr(x, .VALID = NULL)

.cmp_num(x, .VALID = NULL)

.cmp_scl(x, .VALID = NULL)

.cmp_vec(x, .VALID = NULL)

.lgl_scl(x, .VALID = NULL)

.nnw_vec(x, .VALID = NULL)

.pop_atm(x, .VALID = NULL)

.pop_chr(x, .VALID = NULL)

.pop_dtf(x, .VALID = NULL)

.pop_mat(x, .VALID = NULL)

.pop_srt(x, .VALID = NULL)

.pop_vec(x, .VALID = NULL)

.pop_vls(x, .VALID = NULL)

.cmp_atm_vec(x, .VALID = NULL)

.cmp_ch1_scl(x, .VALID = NULL)

.cmp_ch1_vec(x, .VALID = NULL)

.cmp_ch3_scl(x, .VALID = NULL)

.cmp_ch3_vec(x, .VALID = NULL)

.cmp_chr_arr(x, .VALID = NULL)

.cmp_chr_vec(x, .VALID = NULL)

.cmp_chr_vls(x, .VALID = NULL)

.cmp_chr_gen(x, .VALID = NULL)

.cmp_chr_scl(x, .VALID = NULL)

.cmp_clr_vec(x, .VALID = NULL)

.cmp_lgl_scl(x, .VALID = NULL)

.cmp_lgl_vec(x, .VALID = NULL)

.cmp_nng_scl(x, .VALID = NULL)

.cmp_nnw_scl(x, .VALID = NULL)

.cmp_nnw_vec(x, .VALID = NULL)

.cmp_num_scl(x, .VALID = NULL)

.cmp_num_vec(x, .VALID = NULL)

.cmp_pos_vec(x, .VALID = NULL)

.cmp_ppn_vec(x, .VALID = NULL)

.cmp_psw_scl(x, .VALID = NULL)

.cmp_psw_vec(x, .VALID = NULL)

.cmp_srt_vec(x, .VALID = NULL)

.cmp_str_scl(x, .VALID = NULL)

.cmp_str_vec(x, .VALID = NULL)

.cmp_whl_scl(x, .VALID = NULL)

.cmp_atm_mvc(x, .VALID = NULL)

.unq_atm_mvc(x, .VALID = NULL)

.unq_atm_vec(x, .VALID = NULL)

.unq_ch1_vec(x, .VALID = NULL)

.unq_ch3_vec(x, .VALID = NULL)

.unq_chr_vec(x, .VALID = NULL)

.unq_nnw_vec(x, .VALID = NULL)

.unq_num_vec(x, .VALID = NULL)

.unq_psw_vec(x, .VALID = NULL)

.unq_str_vec(x, .VALID = NULL)

.atm_rct_dtf(x, .VALID = NULL)
}
\arguments{
\item{x}{Any object. See \emph{details}.}

\item{.VALID}{An optional object containing valid atomic values to check against atomic values in \code{x}. See \emph{details}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Fast property checking for \code{\link{ppp}} functions without the dot prefix.
}
\details{
When \code{.VALID} is supplied, both \code{x} and \code{.VALID} are \link[uj:av]{atomized} before checking whether all atomic elements of \code{x} are contained in the atomic elements of \code{.VALID}.
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
