% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\encoding{UTF-8}
\name{purge_help}
\alias{purge_help}
\alias{purge}
\alias{x}
\alias{purge_plots}
\alias{rm_plot}
\alias{xplot}
\alias{purge_plot}
\alias{rm_plots}
\alias{xplots}
\alias{purge_warnings}
\alias{purge_warn}
\alias{xwarnings}
\alias{xwarn}
\alias{purge_console}
\alias{purge_con}
\alias{xconsole}
\alias{xcon}
\alias{purge_global}
\alias{purge_all}
\alias{rm_global}
\alias{rm_all}
\alias{xglobal}
\alias{xall}
\title{Purge objects and parts of objects}
\usage{
purge_help()

purge(...)

x(...)

purge_plots(x = NULL)

rm_plot(x = NULL)

xplot(x = NULL)

purge_plot(x = NULL)

rm_plots(x = NULL)

xplots(x = NULL)

purge_warnings()

purge_warn()

xwarnings()

xwarn()

purge_console()

purge_con()

xconsole()

xcon()

purge_global()

purge_all()

rm_global()

rm_all()

xglobal()

xall()
}
\arguments{
\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs} containing names of objects to purge from the global environment.}

\item{x}{Anticipated to be a plot object, but any object can be handled.}

\item{funs}{\link[=cmp_chr_vec]{A complete character vec} of function names, possibly with package prefixes.}
}
\value{
\strong{A character vector}    \cr\cr \verb{purge_pref, rm_pref, xpref}
\cr\cr  \strong{The} \code{NULL} \strong{object} \cr\cr all others
}
\description{
Purge objects and parts of objects
}
\details{
All purge (\code{x}) functions include garbage collection.
\tabular{ll}{  \verb{purge_plot, rm_plot, xplot}   \tab                \cr
  \verb{purge_warn, rm_warn, xwarn}   \tab Purge warnings                        \cr
  \verb{purge_con, rm_con, xcon}      \tab Purge console                         \cr
  \verb{purge, x}                     \tab                   }
\tabular{l}{  \eqn{^{(1)}} Converts \code{'base::paste'} to \code{'paste'}, for example.           \cr
  \eqn{^{(2)}} Followed by printing the object \code{x} if not \code{NULL}.            \cr   }
\tabular{ll}{  \verb{purge_global, purge_all}      \tab \strong{WARNING: Use with care!}            \cr
  \verb{rm_global, rm_all}            \tab    *These functions purge *            \cr
  \verb{xglobal, xall}                \tab    \emph{the global environment}              }
}
\section{Functions}{
\itemize{
\item \code{purge()}: Purge specific objects as named in \code{...} (unquoted object names).

\item \code{x()}: An aliase for \code{purge}.

\item \code{purge_plots()}: Closes all open plots/graphic devices, followed by printing \code{x} if not \code{NULL}.

\item \code{rm_plot()}: An alias for \code{purge_plot}.

\item \code{xplot()}: An alias for \code{purge_plot}.

\item \code{purge_plot()}: An alias for \code{purge_plot}.

\item \code{rm_plots()}: An alias for \code{purge_plot}.

\item \code{xplots()}: An alias for \code{purge_plot}.

\item \code{purge_warnings()}: Purges warning messages.

\item \code{purge_warn()}: An alias for \code{purge_warnings}.

\item \code{xwarnings()}: An alias for \code{purge_warnings}.

\item \code{xwarn()}: An alias for \code{purge_warnings}.

\item \code{purge_console()}: Purges the contents of the console.

\item \code{purge_con()}: An alias for \code{purge_console}.

\item \code{xconsole()}: An alias for \code{purge_console}.

\item \code{xcon()}: An alias for \code{purge_console}.

\item \code{purge_global()}: Purges the global environment.

\item \code{purge_all()}: An alias for \code{purge_global}.

\item \code{rm_global()}: An alias for \code{purge_global}.

\item \code{rm_all()}: An alias for \code{purge_global}.

\item \code{xglobal()}: An alias for \code{purge_global}.

\item \code{xall()}: An alias for \code{purge_global}.

}}
\examples{
egFunNames <- c("base::c", "uj::cmp_chr_scl", "run")
xpref(egFunNames)
\dontrun{
  ## build a ggplot scatterplot object (y.)
  library(ggplot2)
  egObj <- data.frame(x = 0:100, y = sqrt(0:100)
  egGgp <- ggplot(egObj, aes(x, y)) + geom_point()

  ## print to console, wait for user, purge console
  say("\n A message")
  continue()
  xcon()

  ## display a scatterplot, wait for user
  plot(egObj$x, egObj$y)
  continue()

  ## purge plots and print ggplot
  xplot(egGgp)

  ## wait for user, purge plots again
  continue()
  xplot()

  ## generate warning, print warnings
  sqrt(-1)
  warnings()

  ## purge warnings, validate that warnings are purged
  xwarn()
  warnings()
}
}
\seealso{
Other meta: 
\code{\link{callers}()},
\code{\link{console_help}()},
\code{\link{dot_args}()},
\code{\link{pause}()},
\code{\link{ply_help}()},
\code{\link{recycling_help}()},
\code{\link{run}()}

Other purge: 
\code{\link{rm_vals_help}()}
}
\concept{meta}
\concept{purge}
