% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put.}
\alias{put.}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put values into objects}
\usage{
put.()

put(x, v, e)

putr(x, v, r)

putc(x, v, c)
}
\arguments{
\item{x}{Object into which values will be placed. May be a vector, list,
matrix, or data.frame.}

\item{v}{Vector of replacement values.}

\item{e}{Logical vector, integer vector, or integer scalar indexing
element(s) to replace.}

\item{r}{Logical vector or integer scalar indexing row to replace.}

\item{c}{Logical vector or integer scalar indexing column to replace.}
}
\value{
An object of the same dimension as \code{x}.
}
\description{
Put values into objects
}
\section{Functions}{
\itemize{
\item \code{put()}: Put values in \code{v} into \code{e}-indexed elements of
the vector or list \code{x}.

\item \code{putr()}: Put the vector \code{v} into the \code{r}-indexed row of
the matrix or data frame \code{x}.

\item \code{putc()}: Put the vector \code{v} into the \code{c}-indexed of the
matrix or data frame \code{x}.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
