% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put values into objects}
\usage{
put(x, v, e)

putr(x, v, r)

putc(x, v, c)
}
\arguments{
\item{x}{Object into which values will be placed. May be a \link[=ivec]{vec},
\link[ivls]{vlist}, matrix, or data.frame.}

\item{v}{A vector of replacement values.}

\item{e}{A \link[=cmp_ind_vec]{complete indexer vec} or
\link[=cmp_psw_scl]{complete positive whole-number scalar} indexing
element(s) to replace.}

\item{r, c}{A \link[=cmp_ind_scl]{complete indexer scalar} indicating row or
column to replace, respectively.}
}
\value{
An object of the same class and dimension as \code{x}.
}
\description{
\tabular{ll}{
FUNCTION   \tab WHAT IT DOES                                            \cr
\code{put}      \tab Put values of \code{v} into \code{e}-indexed elements of the vector
or \link[=ivls]{vlist} \code{x}.                             \cr
\code{putr}     \tab Put the vector \code{v} into the \code{r}-indexed row of the matrix
or data frame \code{x}.                                      \cr
\code{putc}     \tab or \link[=ivls]{vlist} \code{x}.                             \cr
\code{putr}     \tab Put the vector \code{v} into the \code{c}-indexed column of the
matrix or data frame \code{x}.                                 }
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
