% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put_uj}
\alias{put_uj}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put Values into Objects}
\usage{
put_uj()

put(x, rep, elt)

putr(x, rep, row)

putc(x, rep, col)
}
\arguments{
\item{x}{Object into which values will be placed. May be a vector, list,
matrix, or data.frame.}

\item{rep}{Vector of replacement values.}

\item{elt}{Logical vector, integer vector, or integer scalar indexing
element(s) to replace.}

\item{row}{Logical vector or integer scalar indexing row to replace.}

\item{col}{Logical vector or integer scalar indexing column to replace.}
}
\value{
An object of the same dimension as \code{x}.
}
\description{
Put Values into Objects
}
\section{Functions}{
\itemize{
\item \code{put}: Put values into indexed elements of vector or list
\code{x}.

\item \code{putr}: Put vector into indexed row of matrix or data frame
\code{x}.

\item \code{putc}: Put vector into indexed col of matrix or data frame
\code{x}.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
