% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r}
\alias{r}
\alias{e}
\alias{re}
\title{Error-checked wrappers for \code{base::rep}}
\usage{
r(r, ...)

e(e, ...)

re(e, r, ...)

re(e, r, ...)
}
\arguments{
\item{r}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the number of replications of entire vectors.}

\item{...}{One or more atomic vectors to be replicated, reduced to unique values, or reduced to unique duplicated values.}

\item{e}{A complete positive whole-number scalar giving the number of times each value is to be replicated in place.}
}
\value{
An atomic vector.
}
\description{
These functions \link[=av]{atomize} \code{...} arguments, converting them to an atomic vector (represented here as \code{dots}) before replication.
\itemize{
\item \strong{\code{r}}: calls \code{rep.int(dots, r)}.
\item \strong{\code{e}}: calls \code{rep.int(dots, each = e)}.
\item \strong{\verb{re, er}}: call \code{rep(dots, times = r, each = e)}.
}
NOTE: Although \code{er} and \code{re} produce identical results when their arguments are the same, both are included for convenience.
}
\examples{
r(2, 0:4, 5:9)
e(3, 0:4, 5:9)
re(2, 3, 0:4, 5:9)
er(3, 2, 0:4, 5:9)
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
