% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\encoding{UTF-8}
\name{r_help}
\alias{r_help}
\alias{r}
\alias{e}
\alias{re}
\alias{er}
\title{Error-checked wrappers for \code{base::rep}}
\usage{
r_help()

r(r, ...)

e(e, ...)

re(r, e, ...)

er(e, r, ...)
}
\arguments{
\item{r}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the number of replications of entire vectors.}

\item{...}{One or more objects containing atomic values to be replicated (\link[=a]{atomized} before replication).}

\item{e}{A complete positive whole-number scalar giving the number of times each value is to be replicated in place.}
}
\value{
An atomic vector.
}
\description{
These functions \link[=av]{atomize} \code{...} arguments, converting them to an atomic vector before replication.
}
\section{Functions}{
\itemize{
\item \code{r()}: Calls \code{rep.int(av(...), r)}.

\item \code{e()}: Calls \code{rep.int(av(...), each = e)}.

\item \code{re()}: Calls \code{rep(av(...), times = r, each = e)}.

\item \code{er()}: Calls \code{rep(av(...), times = r, each = e)}.

}}
\examples{
r(2, 0:2, 3)
e(3, 0:2, 3)
re(2, 3, 0:2, 3)
}
