% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\encoding{UTF-8}
\name{rd}
\alias{rd}
\alias{rdCSV}
\alias{rdCLIP}
\alias{rdXSV}
\alias{rdTSV}
\title{Thin and extended functionality wrappers of \code{readr} functions.}
\usage{
rdCSV(
  file = NULL,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = base::c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = base::min(1000, n_max),
  name_repair = "unique",
  num_threads = readr::readr_threads(),
  progress = readr::show_progress(),
  show_col_types = readr::should_show_types(),
  skip_empty_rows = TRUE,
  lazy = readr::should_read_lazy()
)

rdCLIP()

rdXSV(
  file = NULL,
  delim = NULL,
  quote = "\\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = base::c("", "NA"),
  quoted_na = TRUE,
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  n_max = Inf,
  guess_max = base::min(1000, n_max),
  name_repair = "unique",
  num_threads = readr::readr_threads(),
  progress = readr::show_progress(),
  show_col_types = sreadr::should_show_types(),
  skip_empty_rows = TRUE,
  lazy = readr::should_read_lazy()
)

rdTSV(
  file = NULL,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = base::c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = base::min(1000, n_max),
  progress = readr::show_progress(),
  name_repair = "unique",
  num_threads = readr::readr_threads(),
  show_col_types = readr::should_show_types(),
  skip_empty_rows = TRUE,
  lazy = readr::should_read_lazy()
)
}
\arguments{
\item{file}{Either \code{NULL} or a \link[=cmp_chr_vec]{complete character vector} that resolves to a file path. When \code{file = NULL} the user is asked to select a file using a system dialog box.}
}
\value{
\strong{Varies}       \cr\cr \code{rdCLIP}
\cr\cr  \strong{A data.frame} \cr\cr \verb{rdCSV, rdTSV, rdXSV}
}
\description{
Read the clipboard and text files, but prompting user to choose a file if no file name/path is provided.
}
\details{
\tabular{ll}{  \code{rdCLIP}   \tab Thinly wraps \code{\link[readr]{clipboard}}         \cr   \tab   \cr
  \code{rdCSV}    \tab Extends \code{\link[readr]{read_csv}}\eqn{^{(1)}}   \cr
  \code{rdTSV}    \tab Extends \code{\link[readr]{read_tsv}}\eqn{^{(1)}}   \cr
  \code{rdXSV}    \tab Extends \code{\link[readr]{read_delim}}\eqn{^{(1)}}   }
\tabular{l}{  \eqn{^{(1)}} \code{file = NULL} prompts the user to select a file.         }
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{evals}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{nis}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}

Other wraps: 
\code{\link{basics}()},
\code{\link{dp}},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{sca}()},
\code{\link{tb}()}

Other files: 
\code{\link{paths}}
}
\concept{extensions}
\concept{files}
\concept{wraps}
