% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rct_delim.R
\name{read_rct_delim}
\alias{read_rct_delim}
\title{Interactively Select and Read a Rectangular Delimited Text Data File}
\usage{
read_rct_delim(type, .D = ",", .CLEAR = TRUE, ...)
}
\arguments{
\item{type}{A \link[uj:cmp_str_scl]{complete string scalar} description of the type of data file to be read.}

\item{.D}{A \link[uj:cmp_ch1_scl]{complete onechar scalar} text delimiter.}

\item{.CLEAR}{A logical scalar indicating whether to clear the console to denote a new user interaction.}
}
\value{
A data.frame.
}
\description{
Interactively Select and Read a Rectangular Delimited Text Data File
}
