% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr_tools.R
\name{readr_tools}
\alias{readr_tools}
\alias{rdcsv}
\alias{rdclip}
\alias{rddsv}
\alias{rdtsv}
\title{Aliases and wraps for functions in package \code{readr}.}
\usage{
rdcsv(
  file,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = readr_threads(),
  progress = show_progress(),
  show_col_types = should_show_types(),
  skip_empty_rows = TRUE,
  lazy = should_read_lazy()
)

rdclip()

rddsv(
  file,
  delim = NULL,
  quote = "\\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = readr_threads(),
  progress = show_progress(),
  show_col_types = should_show_types(),
  skip_empty_rows = TRUE,
  lazy = should_read_lazy()
)

rdtsv(
  file,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = show_progress(),
  name_repair = "unique",
  num_threads = readr_threads(),
  show_col_types = should_show_types(),
  skip_empty_rows = TRUE,
  lazy = should_read_lazy()
)
}
\description{
Alias for \code{\link[readr]{read_csv}} to read comma-separated
values from text files.
}
\section{Functions}{
\itemize{
\item \code{rdclip}: Alias for \code{\link[readr]{clipboard}} to read text
from the system clipboard.

\item \code{rddsv}: Alias for \code{\link[readr]{read_delim}} to read
(arbitrary) delimiter-separated values from text files.

\item \code{rdtsv}: Alias for \code{\link[readr]{read_tsv}} to read tab
separated values from text files.
}}

