% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass.}
\alias{reclass.}
\alias{reclass}
\title{Set object class}
\usage{
reclass.()

reclass(x, cls)
}
\arguments{
\item{x}{Any object.}

\item{cls}{Character vector of classes.}
}
\value{
An object of class \code{c}.
}
\description{
Set object class
}
\section{Functions}{
\itemize{
\item \code{reclass()}: Sets class(es) of \code{x} to \code{cls}.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
