% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Set object class}
\usage{
reclass(x, cls)
}
\arguments{
\item{x}{Any object.}

\item{cls}{Character vector of classes.}
}
\value{
An object of class \code{c}.
}
\description{
Set object class
}
