% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass_uj}
\alias{reclass_uj}
\alias{reclass}
\title{Set object class}
\usage{
reclass_uj()

reclass(x, cls)
}
\arguments{
\item{x}{Any object.}

\item{cls}{Character vector of classes.}
}
\value{
An object of class \code{c}.
}
\description{
Set object class
}
\section{Functions}{
\itemize{
\item \code{reclass}: Sets class(es) of \code{x} to \code{cls}.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
