% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling.R
\name{recycling.}
\alias{recycling.}
\alias{recyclable_n}
\alias{recyclable}
\alias{recycle}
\title{Recycling and recyclability}
\usage{
recycling.()

recyclable_n(n, targ = max(n))

recyclable(..., n. = NULL, min. = 1, max. = NULL, targ. = NULL, err. = F)

recycle(..., n. = NULL, min. = 1, max. = NULL, targ. = NULL)
}
\arguments{
\item{...}{Named arguments to be recycled in the environment of the calling
function.}

\item{n.}{For \code{recyclable_n}, a \link[cmp_psw_vec]{complete positive
whole-number vec} giving the lengths of arguments to be recycled; for
\code{recyclable} and \code{recycle}, either \code{NULL} or a
\link[cmp_psw_vec]{complete positive whole-number vec} giving the set of
valid recycled argument lengths.}

\item{min.}{\code{NULL} or \link[cmp_psw_scl]{complete positive whole-number
scalar} giving the minimum valid recycled argument length.}

\item{max.}{\code{NULL} or \link[cmp_psw_scl]{complete positive whole-number
scalar} giving the maximum valid recycled argument length.}

\item{targ.}{\link[cmp_psw_scl]{Complete positive whole-number scalar} giving
the target length of recycled arguments.}

\item{err.}{\link[cmp_lgl_scl]{Complete logical scalar} indicating whether to
throw an error if the arguments in \code{...} are not recyclable.}

\item{lengths.}{\link[cmp_psw_vec]{Complete positive whole-number vec} of
lengths to check for recyclability.}
}
\value{
\code{recyclable_n} and \code{recyclable} return \code{TRUE} or
\code{FALSE}. \code{recycle} returns \code{NULL} as it is called for the
side effect of recycling arguments in the environment of a parent function.
}
\description{
Recycling and recyclability
}
\section{Functions}{
\itemize{
\item \code{recyclable_n()}: Check whether the vector of lengths in \code{n.}
represents recyclable arguments subject to the setting in \code{targ.}.

\item \code{recyclable()}: Evaluate whether arguments in \code{...} are
recyclable subject to settings in the arguments \code{n.}, \code{min.},
\code{max.}, and \code{targ.}.

\item \code{recycle()}: Recycle arguments in \code{...} in the environment of
the calling function subject to settings in the arguments \code{n.},
\code{min.}, \code{max.}, and \code{targ.}.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}

Other environments: 
\code{\link{callers.}()},
\code{\link{os.}()},
\code{\link{run.}()},
\code{\link{values.}()}
}
\concept{environments}
\concept{extensions}
