% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling.R
\name{recycling_uj}
\alias{recycling_uj}
\alias{recyclable_n}
\alias{recyclable}
\alias{recycle}
\title{Recycling and recyclability}
\usage{
recycling_uj()

recyclable_n(n., targ. = max(n.))

recyclable(..., n. = NULL, min. = 1, max. = NULL, targ. = NULL, err. = F)

recycle(..., n. = NULL, min. = 1, max. = NULL, targ. = NULL)
}
\arguments{
\item{n.}{For \code{recyclable_n}, the lengths of arguments to be recycled;
for \code{recyclable} and \code{recycle}, \code{NULL} or the set of valid
recycled argument lengths.}

\item{targ.}{Target length of recycled arguments.}

\item{...}{Named arguments to be recycled in the environment of the calling
function.}

\item{min.}{\code{NULL} or the minimum valid recycled argument length.}

\item{max.}{\code{NULL} or the maximum valid recycled argument length.}

\item{err.}{\code{TRUE} or \code{FALSE} indicating whether to throw an error
if the arguments in \code{...} are not recyclable.}

\item{lengths.}{Vector of lengths to check for recyclability.}
}
\value{
\code{recyclable_n} and \code{recyclable} return \code{TRUE} or
\code{FALSE}. \code{recycle} returns \code{NULL} as it is called for the
side effect of recycling arguments in the environment of a parent function.
}
\description{
Recycling and recyclability
}
\section{Functions}{
\itemize{
\item \code{recyclable_n}: Check whether the vector of lengths in \code{n.}
represents recyclable arguments subject to the setting in \code{targ.}.

\item \code{recyclable}: Evaluate whether arguments in \code{...} are
recyclable subject to settings in the arguments \code{n.}, \code{min.},
\code{max.}, and \code{targ.}.

\item \code{recycle}: Recycle arguments in \code{...} in the environment
of the calling function subject to settings in the arguments \code{n.},
\code{min.}, \code{max.}, and \code{targ.}.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
