% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{removal}
\alias{removal}
\alias{xat}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{Remove various objects or characteristics.}
\usage{
xat(x., ...)

xcls(x., ...)

xwarn()

xplot(x. = NULL)

xcon()
}
\arguments{
\item{x.}{An object.}

\item{...}{Character vectors containing attribute names or classes to remove.}
}
\value{
An object (\code{xat} and \code{xcls}) or \code{NULL} (\code{xwarn}
and \code{xcls}).
}
\description{
Remove from \code{x} the attributes named in \code{...}.
}
\section{Functions}{
\itemize{
\item \code{xcls}: Remove from \code{x.} the classes named in \code{...}.

\item \code{xwarn}: Purge all warnings.

\item \code{xplot}: Purge all plots and if optional \code{x.} is provided,
print it (\code{x.} is intended to be a plot object).

\item \code{xcon}: Clear the console.
}}

