% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revstr.R
\name{rev_str.}
\alias{rev_str.}
\alias{rev_str}
\title{Reverse strings}
\usage{
rev_str.()

rev_str(x)
}
\arguments{
\item{x}{A \link[=cmp_chr_vec]{complete character vec}.}
}
\value{
Character vector.
}
\description{
Reverse the order of characters in each element of a character
vector.
}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()}
}
\concept{strings}
