% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revstr.R
\name{revstr}
\alias{revstr}
\title{Reverse the order of characters in each element of a character vector.}
\usage{
revstr(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Reverse the order of characters in each element of a character vector.
}
