% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revstr.R
\name{revstr_uj}
\alias{revstr_uj}
\alias{revstr}
\title{Reverse strings}
\usage{
revstr_uj()

revstr(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Reverse strings
}
\section{Functions}{
\itemize{
\item \code{revstr}: Reverse the order of characters in each element of a
character vector.
}}

\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{spacing_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
