% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\encoding{UTF-8}
\name{rm_vals}
\alias{rm_vals}
\alias{rm_elts}
\alias{rm_rows}
\alias{rm_cols}
\alias{xvals}
\alias{xelts}
\alias{xrows}
\alias{xcols}
\title{Manipulate positions, rows, and columns}
\usage{
rm_vals(X, Vals)

rm_elts(X, Elts)

rm_rows(X, Rows)

rm_cols(X, Cols)

xvals(X, Vals)

xelts(X, Elts)

xrows(X, Rows)

xcols(X, Cols)
}
\arguments{
\item{X}{For \verb{*vals} and \verb{*elts} an atomic vector. For \verb{*cols} and \verb{*rows}, a atomic data.frame (?uj::atm_dtf) or an atomic matrix.}

\item{Vals}{A unique-valued atomic vector (?uj::unq_atm_vec) mode-compatible with \code{X}.}

\item{Elts, Rows, Cols}{Atomic vector indexing elements, rows, or columns of \code{X} to be deleted, respectively. These are either
\itemize{\item A uniquely-valued positive whole-number vector.
\item A complete logical vector (?uj::cmp_lgl_vec) of length \code{length(X)}.
\item A uniquely-valued character vector of element/column/row names.   }}
}
\value{
An object of the same class as \code{X}.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \verb{purge_vals, rm_vals, xvals}   \tab Remove values in atomic vector \code{Vals} from atomic vector \code{X}. \cr   \tab   \cr
  \verb{purge_elts, rm_elts, xelts}   \tab Delete elements of \code{X} indexed in \code{Elts}                      \cr   \tab   \cr
  \verb{purge_rows, rm_rows, xrows}   \tab Delete rows of \code{X} indexed in \code{Rows}.                         \cr   \tab   \cr
  \verb{purge_cols, rm_cols, xcols}   \tab Delete columns of \code{X} indexed in \code{Cols}                                      }
}
\seealso{
Other function_form: 
\code{\link{insert_elts_before}()},
\code{\link{move_elt}()},
\code{\link{put}()}

Other purge: 
\code{\link{purge}()}
}
\concept{function_form}
\concept{purge}
