% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_vals.R
\encoding{UTF-8}
\name{rm_vals_help}
\alias{rm_vals_help}
\alias{rm_vals}
\alias{rm_elts}
\alias{rm_rows}
\alias{rm_cols}
\alias{xvals}
\alias{xelts}
\alias{xrows}
\alias{xcols}
\title{Manipulate positions, rows, and columns}
\usage{
rm_vals_help()

rm_vals(x, vals)

rm_elts(x, elts)

rm_rows(x, rows)

rm_cols(x, cols)

xvals(x, vals)

xelts(x, elts)

xrows(x, rows)

xcols(x, cols)
}
\arguments{
\item{x}{For \verb{*vals} and \verb{*elts} an atomic vector. For \verb{*cols} and \verb{*rows}, a atomic data.frame (?atm_dtf) or an atomic matrix.}

\item{vals}{A unique-valued atomic vector (?uj::unq_atm_vec) mode-compatible with \code{x}.}

\item{elts, rows, cols}{Atomic vector indexing elements, rows, or columns of \code{x} to be deleted, respectively. These are either
\itemize{\item A uniquely-valued positive whole-number vector.
\item A complete logical vector (?uj::cmp_lgl_vec) of length \code{length(x)}.
\item A uniquely-valued character vector of element/column/row names.   }}
}
\value{
An object of the same class as \code{x}.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \verb{purge_vals, rm_vals, xvals}   \tab  \cr   \tab   \cr
  \verb{purge_elts, rm_elts, xelts}   \tab                       \cr   \tab   \cr
  \verb{purge_rows, rm_rows, xrows}   \tab                          \cr   \tab   \cr
  \verb{purge_cols, rm_cols, xcols}   \tab                                       }
}
\section{Functions}{
\itemize{
\item \code{rm_vals()}: Remove values in atomic vector \code{vals} from atomic vector \code{x}.

\item \code{rm_elts()}: Delete elements of \code{x} indexed in \code{elts}.

\item \code{rm_rows()}: Delete rows of \code{x} indexed in \code{rows}.

\item \code{rm_cols()}: Delete columns of \code{x} indexed in \code{cols}.

\item \code{xvals()}: An alias for \code{rm_vals}.

\item \code{xelts()}: An alias for \code{rm_elts}.

\item \code{xrows()}: An alias for \code{rm_rows}.

\item \code{xcols()}: An alias for \code{rm_cols}.

}}
\seealso{
Other purge: 
\code{\link{purge_help}()}
}
\concept{purge}
