% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run.}
\alias{run.}
\alias{run}
\title{Build and evaluate code}
\usage{
run.()

run(...)
}
\arguments{
\item{...}{Arguments to be pasted into a character scalar command to be run
(i.e., parsed and evaluated) in the environment where \code{run(...)} was
called.}
}
\value{
\code{NULL}. Called for its side effect of executing code in the
environment where \code{run(...)} was called.
}
\description{
Build and evaluate code
}
\section{Functions}{
\itemize{
\item \code{run()}: Collapses \code{...} into a character scalar and evaluates
it as code in the calling environment.

}}
\examples{
Join <- ""
CallVector <- c("paste0(", 0:3, ", collapse = Join)")
CallScalar <- paste0(CallVector, collapse = Join)
CallVector <- c("Vector <- ", CallVector)
CallScalar <- paste0("Scalar <- ", CallScalar)
run(CallVector)
run(CallScalar)
print(CallVector)
print(CallScalar)
}
\seealso{
Other code: 
\code{\link{xpr.}()}

Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}

Other environments: 
\code{\link{callers.}()},
\code{\link{os.}()},
\code{\link{recycling.}()},
\code{\link{values.}()}
}
\concept{code}
\concept{environments}
\concept{extensions}
