% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Evaluate constructed code}
\usage{
run(...)
}
\arguments{
\item{...}{Arguments to be pasted into a character scalar command to be run
(i.e., parsed and evaluated) in the environment where \code{run(...)} was
called.}
}
\value{
\code{NULL}. Called for its side effect of executing code in the
environment where \code{run(...)} was called.
}
\description{
Code to build and run code.
}
\examples{
Join <- ""
CallVector <- c("paste0(", 0:3, ", collapse = Join)")
CallScalar <- paste0(CallVector, collapse = Join)
CallVector <- c("Vector <- ", CallVector)
CallScalar <- paste0("Scalar <- ", CallScalar)
run(CallVector)
run(CallScalar)
print(CallVector)
print(CallScalar)
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
