% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_uj}
\alias{run_uj}
\alias{run}
\title{Build and evaluate code}
\usage{
run_uj()

run(...)
}
\arguments{
\item{...}{Arguments to be pasted into a character scalar command to be run
(i.e., parsed and evaluated) in the environment where \code{run(...)} was
called.}
}
\value{
\code{NULL}. Called for its side effect of executing code in the
environment where \code{run(...)} was called.
}
\description{
Build and evaluate code
}
\section{Functions}{
\itemize{
\item \code{run}: Collapses \code{...} into a character scalar and evaluates
it as code in the calling environment.
}}

\examples{
Join <- ""
CallVector <- c("paste0(", 0:3, ", collapse = Join)")
CallScalar <- paste0(CallVector, collapse = Join)
CallVector <- c("Vector <- ", CallVector)
CallScalar <- paste0("Scalar <- ", CallScalar)
run(CallVector)
run(CallScalar)
print(CallVector)
print(CallScalar)
}
\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}

Other code: 
\code{\link{callers_uj}()},
\code{\link{expr_uj}}
}
\concept{code}
\concept{meta}
