% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc.}
\alias{sc.}
\alias{sca}
\title{Wraps of functions from package \code{sca;es}.}
\usage{
sc.()

sca(color, alpha = NA)
}
\arguments{
\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA},
existing alpha values are preserved.}
}
\description{
Vectorised in both colour and alpha.
}
\section{Functions}{
\itemize{
\item \code{sca()}: Thin wrap of \code{\link[scales]{alpha}}.

}}
\examples{
alpha("red", 0.1)
alpha(colours(), 0.5)
alpha("red", seq(0, 1, length.out = 10))
alpha(c("first" = "gold", "second" = "lightgray", "third" = "#cd7f32"), .5)
}
\seealso{
Other wraps: 
\code{\link{bs.}()},
\code{\link{dp.}()},
\code{\link{dt.}()},
\code{\link{ggx.}()},
\code{\link{rd.}()},
\code{\link{tb.}()}
}
\concept{wraps}
