% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_tools.R
\name{scales_tools}
\alias{scales_tools}
\alias{sca}
\title{Aliases and wraps for functions in package \code{scales}}
\usage{
sca(colour, alpha = NA)
}
\description{
Alias for \code{\link[scales]{alpha}} (apply an alpha opacity
values to colors).
}
