% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_block_to_matrix.R
\encoding{UTF-8}
\name{score_block_to_matrix}
\alias{score_block_to_matrix}
\title{Convert an Item Score Block to an Item Score Matrix}
\usage{
score_block_to_matrix(x)
}
\arguments{
\item{x}{A data frame with a single item score string variable with one row per person.}
}
\value{
An \eqn{n_{person} \times n_{item}} item score matrix.
}
\description{
Takes an item score block (a data frame with a single item score string variable) to an item score matrix.
}
