% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sets}
\alias{sets}
\title{Sets from combinations}
\usage{
sets(n.)
}
\arguments{
\item{n}{Size of the complete set each subset is drawn from.}
}
\description{
Create a list containing elements of indices to create all
possible combinations of 1 or more elements of an object of size \code{n}.
}
