% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\encoding{UTF-8}
\name{spaces}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{String spacing functions}
\usage{
spaces(n)

pad(..., n. = 0, s. = "r", p. = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n}{A link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the number of spaces.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{n.}{A complete positive whole-number scalar indicating the number of pad characters.}

\item{s.}{A \link[=cmp_ch1_scl]{complete onechar scalar} indicating side(s) to pad: \code{'l'}, \code{'r'}, or \code{'b'} for left, right, or both, respectively.}

\item{p.}{A complete onechar scalar containing a single character used to pad strings.}
}
\value{
\emph{A character scalar} \cr   \code{spaces}
\cr\cr \emph{A character object} \cr   \verb{pad, trm, sqz}
}
\description{
String spacing functions
}
\details{
\tabular{rl}{
  \code{spaces}   \tab Creates a character scalar of \code{n} spaces.
\cr  \code{pad}   \tab Pads strings with leading, trailing white space using \code{\link[stringr:str_pad]{stringr::str_pad}}.
\cr  \code{sqz}   \tab Trims leading, trailing, and extra internal white space using \code{\link[stringr:str_squish]{stringr::str_squish}}.
\cr  \code{trm}   \tab Trims leading and trailing white space using \code{\link[stringr:str_trim]{stringr::str_trim}}.
}
}
\examples{
spaces(0)
spaces(4)
pad(0:9, n. = 3)
pad(0:9, n. = 3, s. = "r")
pad(0:9, n. = 3, s. = "l")
pad(0:9, n. = 3, s. = "b")
pad(0:9, n. = 3, s. = "b", p. = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}},
\code{\link{markdown}},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{chn}()},
\code{\link{max_nch}()},
\code{\link{revstr}()},
\code{\link{ss}()}
}
\concept{chars}
\concept{strings}
