% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.R
\name{spacing.}
\alias{spacing.}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{Simple spacing functions}
\usage{
spacing.()

spaces(n.)

pad(..., n. = 0, s. = "r", p. = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n.}{Positive whole-number scalar indicating the number of spaces or pad
characters.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{s.}{Character scalar "side" padded ("l", "r", or "b" for left, right,
or both).}

\item{p.}{Character scalar containing a single character used to pad strings.}
}
\description{
Simple spacing functions
}
\section{Functions}{
\itemize{
\item \code{spaces()}: Create a character vector of \code{n.} spaces.

\item \code{pad()}: Pad a string with leading and/or trailing spaces
using \code{\link[stringr]{pad}}.

\item \code{sqz()}: Trim leading, trailing, and extra internal spaces
using \code{\link[stringr]{str_squish}}.

\item \code{trm()}: Trim leading and trailing spaces using
\code{\link[stringr]{str_trim}}.

}}
\examples{
spaces(0)
spaces(4)
pad(0:9, n. = 3)
pad(0:9, n. = 3, s. = "r")
pad(0:9, n. = 3, s. = "l")
pad(0:9, n. = 3, s. = "b")
pad(0:9, n. = 3, s. = "b", p. = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{has_pat.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}
}
\concept{strings}
