% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.R
\name{spacing}
\alias{spacing}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{Error-checked \code{stringr}-based spacing functions}
\usage{
spaces(n)

pad(..., n. = 0, s. = "r", p. = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n}{A link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the number of spaces.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{n.}{A complete positive whole-number scalar indicating the number of pad characters.}

\item{s.}{A \link[=cmp_ch1_scl]{complete onechar scalar} indicating side(s) to pad: \code{'l'}, \code{'r'}, or \code{'b'} for left, right, or both, respectively.}

\item{p.}{A complete onechar scalar containing a single character used to pad strings.}
}
\value{
\itemize{
\item \strong{\code{spaces}}: a character scalar.
\item \strong{\code{pad}, \code{trm}, \code{sqz}}: a character object.
}
}
\description{
\itemize{
\item \strong{\code{spaces}}: creates a character vector of \code{n} spaces.
\item \strong{\code{pad}}: pads a string with leading and/or trailing spaces with \code{\link[stringr:pad]{stringr::pad}}.
\item \strong{\code{sqz}}: trims leading, trailing, and extra internal spaces with \code{\link[stringr:str_squish]{stringr::str_squish}}.
\item \strong{\code{trm}}: trims leading and trailing spaces with \code{\link[stringr:str_trim]{stringr::str_trim}}.
}
}
\examples{
spaces(0)
spaces(4)
pad(0:9, n. = 3)
pad(0:9, n. = 3, s. = "r")
pad(0:9, n. = 3, s. = "l")
pad(0:9, n. = 3, s. = "b")
pad(0:9, n. = 3, s. = "b", p. = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{charn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
