% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.R
\name{spacing_uj}
\alias{spacing_uj}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{Simple spacing functions}
\usage{
spacing_uj()

spaces(n.)

pad(..., n. = 0, s. = "r", p. = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n.}{Whole-number scalar indicating the number of spaces or pad
characters.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{s.}{Character scalar "side" padded ("l", "r", or "b" for left, right,
or both).}

\item{p.}{Character scalar containing a single character used to pad strings.}
}
\description{
Simple spacing functions
}
\section{Functions}{
\itemize{
\item \code{spaces}: Create a character vector of \code{n.} spaces.

\item \code{pad}: Pad a string with leading and/or trailing spaces
using \code{\link[stringr]{pad}}.

\item \code{sqz}: Trim leading, trailing, and extra internal spaces
spaces using \code{\link[stringr]{str_squish}}.

\item \code{trm}: Trim leading and trailing spaces using
\code{\link[stringr]{str_trim}}.
}}

\examples{
spaces(0)
spaces(4)
pad(0:9, n. = 3)
pad(0:9, n. = 3, s. = "r")
pad(0:9, n. = 3, s. = "l")
pad(0:9, n. = 3, s. = "b")
pad(0:9, n. = 3, s. = "b", p. = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{build_strings_uj}()},
\code{\link{char_uj}()},
\code{\link{delim_uj}()},
\code{\link{expr_uj}},
\code{\link{fsub_uj}()},
\code{\link{haspat_uj}()},
\code{\link{ipat_uj}()},
\code{\link{maxnch_uj}()},
\code{\link{ox_uj}()},
\code{\link{pgrid_uj}()},
\code{\link{revstr_uj}()},
\code{\link{ss_uj}()},
\code{\link{tocase_uj}()},
\code{\link{weave_uj}()}
}
\concept{strings}
