% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_prop_funs.R
\encoding{UTF-8}
\name{spec2combos}
\alias{spec2combos}
\title{Convert a Property Spec to a Vector of Property Combos}
\usage{
spec2combos(x)
}
\arguments{
\item{x}{A character vector or scalar.}
}
\value{
A character vector or \code{NULL} (returns \code{NULL} when \code{x} is not of mode \code{'character'} or is of length zero).
}
\description{
Splits a property spec along pipes to place each single property or property combo in a separate element (property combos are multiple properties separated by underscores).
}
\seealso{
Other utils: 
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{meets_errs}()},
\code{\link{ox_vals}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()}
}
\concept{utils}
