% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{ssplit}
\alias{ssplit}
\title{Split Strings along Pipes and Trim Whitespace from Both Sides of All Resulting Elements}
\usage{
ssplit(x, d = "|")
}
\arguments{
\item{x}{An object of mode character.}

\item{d}{Character scalar text delimiter.}
}
\value{
A character vector
}
\description{
Split Strings along Pipes and Trim Whitespace from Both Sides of All Resulting Elements
}
\examples{
ssplit(" Just | a | little | thing ")
ssplit("  Just  |  a  |  little  |  thing  ")
}
\seealso{
Other alert_utils: 
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{glu}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()}
}
\concept{alert_utils}
