% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss.}
\alias{sss.}
\alias{sss}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\alias{sss_vals}
\alias{isss}
\title{State of completeness properties}
\usage{
sss.()

sss(x)

icmp(x)

imss(x)

inas(x)

ioks(x)

iprt(x)

sss_vals()

isss(x, sss, ...)
}
\arguments{
\item{x}{An object.}

\item{sss}{\link[=cmp_chr_scl]{Complete character scalar} containing one or
more values from \code{sss_vals()} separated by pipes.}

\item{...}{
  Arguments passed on to \code{\link[=meets.]{meets.}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{lll}{
\code{sss_vals()}            \tab   \tab A character vector.            \cr
\code{sss(x)}                \tab   \tab A character scalar or
\code{NULL}.                   \cr
\code{isss} and \code{i[sss]}\tab   \tab A logical scalar.                }
}
\description{
An object's state of completeness is defined for
\link[=ipop]{populated} \link[=atm_dtf]{atomic data.frame}, populated
\link[=atm_vls]{atomic vlists}, populated atomic vectors, and populated
atomic arrays. For all others, state is undefined and returned as
\code{NULL}. The following table summarizes valid state properties.
\tabular{lll}{
STATE       \tab STATE      \tab CHARACTERISTICS                          \cr
PROPERTY    \tab PROPERTY   \tab OF QUALIFYING                            \cr
VALUE       \tab NAME       \tab OBJECTS                                  \cr
\code{'cmp'}\tab Complete   \tab Non-empty, atomic, no \code{NA} values.  \cr
\code{'mss'}\tab Missing    \tab Non-empty, atomic, only \code{NA} values.\cr
\code{'prt'}\tab Partial    \tab Non-empty, atomic, both \code{NA} and
non-\code{NA} values.                    \cr
\code{'nas'}\tab NA scalar  \tab Atomic \code{NA} scalar.                 \cr
\code{'oks'}\tab OK scalar  \tab Atomic non-\code{NA} scalar.               }
Functions related to state of completeness are described in the following
table:\tabular{ll}{
FUNCTION          \tab WHAT THE                                           \cr
FORMAT            \tab FUNCTION DOES                                      \cr
\code{i[sss]}     \tab Evaluates whether an object is of the state of
completeness represented by \code{[sss]}.          \cr
\code{sss}        \tab Gets a character vector containing all state of
completeness properties of an object.              \cr
\code{isss}       \tab Evaluates an object for a specific state of
completeness and any additional properties specified
in \code{...}.                                     \cr
\code{sss_vals}   \tab Gets a character vector of all possible state of
completeness property values.                        }
}
\section{Specifying Additional Property Requirements}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar value all values of \code{x} must be less
than.                                                \cr
\code{le}     \tab Scalar value all values of \code{x} must be less than or
equal to.                                            \cr
\code{ge}     \tab Scalar value all values of \code{x} must be greater than
or equal to.                                         \cr
\code{gt}     \tab Scalar value all values of \code{x} must be greater
than.                                                  }
}

\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
