% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss}
\alias{sss}
\alias{sss_props}
\alias{is_sss_spec}
\alias{isss}
\alias{icol}
\alias{iemp}
\alias{ilin}
\alias{ipnt}
\alias{irct}
\alias{irow}
\alias{isld}
\alias{isqr}
\title{Shape properties}
\usage{
sss(x)

sss_props()

is_sss_spec(spec)

isss(x, spec, ...)

icol(x, ...)

iemp(x, ...)

ilin(x, ...)

ipnt(x, ...)

irct(x, ...)

irow(x, ...)

isld(x, ...)

isqr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec}
containing one or more shape properties (i.e., from
\code{sss_vals()}). \strong{NOTE}: properties may be pipe-separated. If
If there are multiple properties in \code{spec}, \code{x} is inspected for
a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{\code{sss_vals}}: A character vector.
\cr\cr\strong{\code{sss}}: A character scalar or character vector.
\cr\cr\strong{\code{iSSS, isss, is_sss_spec}}: A logical scalar.
}
\description{
NOTE: \code{SSS} is used as a placeholder representing any given
shape property.
\cr\cr
An object's shape properties have to do with the
number of \link[=ddd]{defined dimensions} and/or \link[=eee]{effective
dimensions} with some additional restrictions. All shape properties are
described in the following table:\tabular{lll}{
PROPERTY      \tab PROPERTY   \tab QUALIFYING                         \cr
NAME          \tab VALUE      \tab OBJECTS                            \cr
Empty         \tab\code{'emp'}\tab Non-\code{NULL}, length-1 objects. \cr
Point         \tab\code{'pnt'}\tab Length-1 vectors, length-1
\link[=ivls]{vlists}, length-1 arrays,
and \code{1 x 1} data.frames.      \cr
Linear        \tab\code{'lin'}\tab Length 2+ vectors, length 2+
\link[=ivls]{vlists}, length 2+ arrays
with multiple index positions in just
1 dimension, and data.frames with 1
row and multiple columns or 1 column
and multiple rows.                 \cr
Row           \tab\code{'row'}\tab Matrices and data.frames with 1 row
and multiple columns.             \cr
Column        \tab\code{'col'}\tab Matrices and data.frames with 1
column and multiple rows          \cr
Rectangular   \tab\code{'rct'}\tab Matrices and data.frames with multiple
rows and multiple columns.         \cr
Square        \tab\code{'sqr'}\tab Matrices and data.frames with multiple
rows, multiple columns, and the same
number of rows and columns.
Solid         \tab\code{'sld'}\tab Arrays with multiple index positions
in 3+ dimensions.                    }
Functions related to shape properties are as follows:
\tabular{ll}{
SSS FUNCTION         \tab WHAT IT DOES                                \cr
\code{iSSS}          \tab Evaluates whether \code{x} matches the
shape property \code{SSS}
(subject to any restrictions in \code{...}).\cr
\code{sss}           \tab Gets a character vector containing all
shape properties matching
\code{x}.                                   \cr
\code{isss}          \tab Evaluates \code{x} against the
shape property specification
in \code{spec} (subject to any restrictions in
\code{...}).                                \cr
\code{sss_props}     \tab Gets a character vector of all possible
shape property values.                      \cr
\code{is_sss_spec}   \tab Evaluates whether \code{spec} is a valid
shape property
specification.                                }
}
\section{Specifying Count and Value Restrictions}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar less-than bound on values of \code{x}.        \cr
\code{le}     \tab Scalar less-than-or-equal bound on values of
\code{x}.                                            \cr
\code{ge}     \tab Scalar greater-than-or-equal bound on values of
\code{x}.                                            \cr
\code{gt}     \tab Scalar greater-than bound on values of \code{x}.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc_props}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm}()},
\code{\link{ppp}()}
}
\concept{props}
