% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss_ccc.R
\encoding{UTF-8}
\name{sss_ccc}
\alias{sss_ccc}
\title{Shape + xclass combination properties}
\usage{
sss_ccc(x, sss, ccc, ...)
}
\arguments{
\item{x}{An R object.}

\item{sss}{A character scalar single basic property from \code{link{sss_props}()}.}

\item{ccc}{A character scalar single xclass property from \code{link{ccc_props}()}}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector}
\cr   \code{sss_ccc_funs}
\cr\cr \emph{A logical scalar}
\cr   \code{SSS_CCC}
\cr   \code{sss_ccc}
}
\description{
\tabular{rl}{
  \code{sss_ccc_funs}   \tab What \link[=sss]{shape} + \link[=ccc]{xclass} combination \link[=prop_funs]{property functions} are there?
\cr                \tab  
\cr    \code{sss_ccc}   \tab Is \code{x} a match to the single shape and xclass properties in \code{bbb} and \code{ccc}, respectively?
\cr                \tab  
\cr    \code{SSS_CCC}   \tab Is \code{x} a match to single shape and xclass properties \code{'SSS'} and \code{'CCC'}, respectively?
}
Some combinations of shape and xclass are nonsensical. That includes:\itemize{
\item Shapes \code{emp} and \code{pnt} combined with xclasses \code{mvc} or \code{scl}.
\item Shapes \code{row}, \code{col}, or \code{rct} combined with xclasses other than \code{dtf} or \code{mat}
\item Shape \code{sqr} combined with any xclass other than \code{mat}.
}
Those nonsensical combinations do not have corresponding \code{SSS_CCC} property functions.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab  
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab  
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
sss_ccc_funs()
sss_ccc_props()
sss_ccc(letters, "lin", "vec")
sss_ccc(1, "emp", "vec")
lin_gen(letters)
pnt_vec(1)
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{unq_mmm_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
