% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss_uj}
\alias{sss_uj}
\alias{sss}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\alias{sss_vals}
\alias{isss}
\title{State of an object}
\usage{
sss_uj()

sss(x.)

icmp(x.)

imss(x.)

inas(x.)

ioks(x.)

iprt(x.)

sss_vals()

isss(x., xxx., ...)
}
\arguments{
\item{x.}{An object.}

\item{xxx.}{A character scalar containing one or more values from
\code{sss_vals()} separated by pipes.}

\item{...}{Additional arguments to \code{\link{meets}} containing value and
element/row/column count restrictions.}
}
\value{
\code{sss_vals()} returns a character vector. \code{sss(x.)}
returns a character scalar or \code{NULL}. All others return either
\code{TRUE} or \code{FALSE}.
}
\description{
An object's state (of completeness) is defined for non-empty
atomic tibbles, non-empty atomic lists, non-empty atomic vectors, and
non-empty atomic arrays. For all others, state is undefined and returned as
\code{NULL}. The following table summarizes valid state properties.
\tabular{lll}{
STATE       \tab STATE           \tab CHARACTERISTICS                 \cr
PROPERTY    \tab PROPERTY        \tab OF QUALIFYING                   \cr
VALUE       \tab NAME            \tab OBJECTS                         \cr
\code{'cmp'}\tab complete        \tab non-empty, atomic, no \code{NA}
values                          \cr
\code{'mss'}\tab missing         \tab non-empty, atomic, only \code{NA}
values                          \cr
\code{'prt'}\tab partial         \tab non-empty, atomic, both \code{NA}
and non-\code{NA} values        \cr
\code{'nas'}\tab\code{NA} scalar \tab atomic \code{NA} scalar         \cr
\code{'oks'}\tab\code{OK} scalar \tab atomic non-\code{NA} scalar       }
}
\section{Functions}{
\itemize{
\item \code{sss}: Get a character scalar containing all state properties
from \code{sss_vals()} that is applicable to \code{x.}, if defined. If not
defined, return \code{NULL}.

\item \code{icmp}: Is \code{x.} complete (atomic, of length 1 or greater, and
containing no \code{NA} values)?

\item \code{imss}: Is \code{x.} a missing object (atomic, of length 1 or
greater, and containing only \code{NA} values)?

\item \code{inas}: Is \code{x.} an atomic scalar \code{NA} value?

\item \code{ioks}: Is \code{x.} an atomic scalar non-\code{NA} value?

\item \code{iprt}: Is \code{x.} an atomic object with some \code{NA} and some
non-\code{NA} values?

\item \code{sss_vals}: Get a character vector of all possible state property
values.

\item \code{isss}: Evaluate whether \code{x.} has a state property
represented in \code{xxx.}.
}}

\section{Additional Arguments in \code{...}}{

Submitting additional arguments to \code{is_sss} via \code{...}
allows for checking not just the state but whether length, number of rows,
number of columns, and element values meet flexible criteria.
}

\seealso{
Other props: 
\code{\link{as_mmm_uj}()},
\code{\link{ccc_uj}()},
\code{\link{cmp_ccc_uj}()},
\code{\link{cmp_mmm_ccc_uj}()},
\code{\link{cmp_mmm_uj}()},
\code{\link{comparable_uj}()},
\code{\link{compatible_uj}()},
\code{\link{ddd_uj}()},
\code{\link{eee_uj}()},
\code{\link{fff_uj}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm_uj}()},
\code{\link{meets_uj}()},
\code{\link{mmm_ccc_uj}()},
\code{\link{mmm_uj}()},
\code{\link{ttt_ccc_uj}()},
\code{\link{ttt_mmm_uj}()},
\code{\link{ttt_uj}()},
\code{\link{xxx_uj}()}
}
\concept{props}
