% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callers.R
\encoding{UTF-8}
\name{stack2funs}
\alias{stack2funs}
\title{Reduce the Call Stack to Just a Vector of Function Names}
\usage{
stack2funs(stack)
}
\arguments{
\item{stack}{A character vector.}
}
\value{
A character vector.
}
\description{
Reduce the Call Stack to Just a Vector of Function Names
}
\examples{
egStack2Funs <- function() {
  egStack2FunsF <- function() {
    stack <- uj::callers()
    uj::stack2funs(stack)
  }
  egStack2FunsE <- function() {egStack2FunsF()}
  egStack2FunsD <- function() {egStack2FunsE()}
  egStack2FunsC <- function() {egStack2FunsD()}
  egStack2FunsB <- function() {egStack2FunsC()}
  egStack2FunsA <- function() {egStack2FunsB()}
  egStack2FunsA()
}
egStack2Funs()
}
\seealso{
Other utils: 
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{meets_errs}()},
\code{\link{spec2combos}()},
\code{\link{spec2props}()}
}
\concept{utils}
