% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{stack2pkgs}
\alias{stack2pkgs}
\title{Reduce the Call Stack to Just a Vector of Package Names}
\usage{
stack2pkgs(stack)
}
\arguments{
\item{stack}{A character vector.}
}
\value{
A character vector.
}
\description{
Reduce the Call Stack to Just a Vector of Package Names
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{not_err}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()}
}
\concept{utils}
