% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\docType{data}
\name{stats0}
\alias{stats0}
\alias{sum0}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\title{Compute Stats Ignoring \code{NA} Values}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
stats0

sum0(...)

sd0(...)

cor0(x, y = NULL)

cov0(x, y = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x)

rsds0(x)

csums0(x)

rsums0(x)

cmeans0(x)

rmeans0(x)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic
values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x}{A \link[=cmp_num_vec]{complete numeric vec} or a
\link[=cmp_num_mat]{complete numeric matrix}.}

\item{y}{An optional \link[=cmp_num_vec]{complete numeric vec} or a
\link[=cmp_num_mat]{complete numeric matrix}.}

\item{n}{Positive whole-number scalar size of the complete set each subset is
drawn from.}
}
\description{
Compute Stats Ignoring \code{NA} Values
}
\section{Functions}{
\itemize{
\item \code{sum0()}: Standard deviation of collective non-\code{NA} atomic elements
in \code{...}.

\item \code{sd0()}: Standard deviation of collective non-\code{NA} atomic elements
in \code{...}.

\item \code{cor0()}: Correlation of complete cases in \code{x} and/or \code{y}.

\item \code{cov0()}: Covariance of complete cases in \code{x} and/or \code{y}.

\item \code{min0()}: Minimum value of collective non-\code{NA} atomic elements in
\code{...}.

\item \code{max0()}: Maximum value of collective non-\code{NA} atomic elements in
\code{...}.

\item \code{var0()}: Variance of collective non-\code{NA} atomic elements in \code{...}.

\item \code{pmin0()}: Pairwise minimum values ignoring pairs containing any \code{NA}
values.

\item \code{pmax0()}: Pairwise maximum values ignoring pairs containing any \code{NA}
values.

\item \code{mean0()}: Mean value of collective non-\code{NA} atomic elements in
\code{...}.

\item \code{csds0()}: Column standard deviations of \code{x} removing non-\code{NA} values
separately for each column.

\item \code{rsds0()}: Row standard deviations of \code{x} removing non-\code{NA} values
separately for each row.

\item \code{csums0()}: Column sums of \code{x} removing non-\code{NA} values separately for
each column.

\item \code{rsums0()}: Row sums of \code{x} removing non-\code{NA} values separately for
each row.

\item \code{cmeans0()}: Column means of \code{x} removing non-\code{NA} values separately
for each column.

\item \code{rmeans0()}: Row means of \code{x} removing non-\code{NA} values separately for
each row.

}}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}

Other na_values: 
\code{\link{na}()}
}
\concept{extensions}
\concept{na_values}
\keyword{datasets}
