% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats0.R
\encoding{UTF-8}
\name{stats0_help}
\alias{stats0_help}
\alias{sum0}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{cvars0}
\alias{rvars0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\title{Compute Statistics Ignoring \code{NA} Calues}
\usage{
stats0_help()

sum0(...)

sd0(...)

cor0(x, y = NULL)

cov0(x, y = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x)

rsds0(x)

cvars0(x)

rvars0(x)

csums0(x)

rsums0(x)

cmeans0(x)

rmeans0(x)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x}{A \link[=num_vec]{numeric vector} or a \link[=num_mat]{numeric matrix}.}

\item{y}{An optional numeric vector or numeric matrix.}
}
\value{
\strong{A numeric vector}           \cr\cr \verb{cmeans0, csums0, cvars0, csds0} \cr \verb{rmeans0, rsums0, rvars0, rsds0} \cr \verb{mean0, sum0, var0, sd0} \cr \verb{pmin0, min0} \cr \verb{pmax0, max0}
\cr\cr  \strong{A numeric vector or matrix} \cr\cr \verb{cor0, cov0}
}
\description{
Simple wrappers for calculating statistics ignoring any \code{NA} values.
}
\section{Functions}{
\itemize{
\item \code{sum0()}: Calculate sum ignoring \code{NA} values.

\item \code{sd0()}: Calculate standard deviation ignoring \code{NA} values.

\item \code{cor0()}: Calculate correlation ignoring \code{NA} values.

\item \code{cov0()}: Calculate covariance ignoring \code{NA} values.

\item \code{min0()}: Calculate minimum value ignoring \code{NA} values.

\item \code{max0()}: Calculate maximum ignoring \code{NA} values.

\item \code{var0()}: Calculate variance ignoring \code{NA} values.

\item \code{pmin0()}: Calculate pairmin values ignoring \code{NA} values.

\item \code{pmax0()}: Calculate pairmax values ignoring \code{NA} values.

\item \code{mean0()}: Calculate mean value ignoring \code{NA} values.

\item \code{csds0()}: Calculate column standard deviations ignoring \code{NA} values.

\item \code{rsds0()}: Calculate row standard deviations ignoring \code{NA} values.

\item \code{cvars0()}: Calculate column variances ignoring \code{NA} values.

\item \code{rvars0()}: Calculate row variances ignoring \code{NA} values.

\item \code{csums0()}: Calculate column sums ignoring \code{NA} values.

\item \code{rsums0()}: Calculate row sums ignoring \code{NA} values.

\item \code{cmeans0()}: Calculate column means ignoring \code{NA} values.

\item \code{rmeans0()}: Calculate row means ignoring \code{NA} values.

}}
\examples{
egVec1 <- sample(0:99, 10)
egVec2 <- sample(0:99, 20)
egVec3 <- sample(0:99, 20)
egMat1 <- sample(0:99, 100)

egVec1[sample(1:10, 1)] <- NA
egVec2[sample(1:20, 1)] <- NA
egVec3[sample(1:20, 2)] <- NA
egMat1[sample(1:100, 5)] <- NA
egMat1 <- matrix(egMat1, nrow = 10)
rownames(egMat1) <- paste0("R", 1:10)
colnames(egMat1) <- paste0("C", 1:10)
egDtf1 <- as.data.frame(egMat1)
egX <- list(v1 = egVec1, v2 = egVec2, v3 = egVec3, m = egMat1, d = egDtf1)

egVec1
egVec2
egVec3
egMat1
egDtf1
egX

matrix(c(cmeans0(egMat1), cmeans0(egDtf1)), ncol = 2)
matrix(c(rmeans0(egMat1), rmeans0(egDtf1)), ncol = 2)
matrix(c(csums0( egMat1), csums0( egDtf1)), ncol = 2)
matrix(c(rsums0( egMat1), rsums0( egDtf1)), ncol = 2)
matrix(c(csds0(  egMat1), csds0(  egDtf1)), ncol = 2)
matrix(c(rsds0(  egMat1), rsds0(  egDtf1)), ncol = 2)

pmin0(egVec1, egVec2, egVec3)
pmax0(egVec2, egVec2, egVec3)

cor0(egVec2, egVec3)
cov0(egVec2, egVec3)

cor0(egMat1)
cov0(egMat1)

list(mean = mean0(egX), min = min0(egX), max = max0(egX),
      sum = sum0( egX), var = var0(egX), sd  = sd0( egX))
}
\seealso{
Other missingness: 
\code{\link{na_help}()}
}
\concept{missingness}
