% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_tools}
\alias{stats_tools}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\title{Compute basic statistics by dropping \code{NA} values (complete cases
only for \code{cor0}, and \code{cov0}).}
\usage{
sd0(...)

cor0(x., y. = NULL)

cov0(x., y. = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x.)

rsds0(x.)

csums0(x.)

rsums0(x.)

cmeans0(x.)

rmeans0(x.)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic
values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x.}{Numeric vector or matrix.}

\item{y.}{Optional numeric vector of matrix.}
}
\description{
Standard deviation of non-\code{NA} atomic elements contained in
\code{...} as a whole.
}
\section{Functions}{
\itemize{
\item \code{cor0}: Correlation of complete cases in \code{x.}
and/or \code{y.}.

\item \code{cov0}: Covariance of complete cases in \code{x.}
and/or \code{y.}.

\item \code{min0}: Minimum value of non-\code{NA} atomic
elements contained in \code{...} as a whole.

\item \code{max0}: Maximum value of non-\code{NA} atomic
elements contained in \code{...} as a whole.

\item \code{var0}: Variance of non-\code{NA} atomic elements
contained in \code{...} as a whole.

\item \code{pmin0}: Pairwise minimum values ignoring \code{NA}
values.

\item \code{pmax0}: Pairwise maximum values ignoring \code{NA}
values.

\item \code{mean0}: Mean value of non-\code{NA} atomic elements
contained in \code{...} as a whole.

\item \code{csds0}: Column standard deviations of \code{x}
removing non-\code{NA} values separately for each column.

\item \code{rsds0}: Row standard deviations of \code{x} removing
non-\code{NA} values separately for each row.

\item \code{csums0}: Column sums of \code{x} removing
non-\code{NA} values separately for each column.

\item \code{rsums0}: Row sums of \code{x} removing non-\code{NA}
values separately for each row.

\item \code{cmeans0}: Column means of \code{x} removing
non-\code{NA} values separately for each column.

\item \code{rmeans0}: Row means of \code{x} removing non-\code{NA}
values separately for each row.
}}

