% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_uj}
\alias{stats_uj}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\alias{sets}
\title{Compute stats ignoring NA}
\usage{
stats_uj()

sd0(...)

cor0(x., y. = NULL)

cov0(x., y. = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x.)

rsds0(x.)

csums0(x.)

rsums0(x.)

cmeans0(x.)

rmeans0(x.)

sets(n.)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic
values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x.}{Numeric vector or matrix.}

\item{y.}{Optional numeric vector of matrix.}

\item{n.}{Size of the complete set each subset is drawn from.}
}
\description{
Compute stats ignoring NA
}
\section{Functions}{
\itemize{
\item \code{sd0}: Standard deviation of non-\code{NA} atomic elements
contained in \code{...} as a whole.

\item \code{cor0}: Correlation of complete cases in \code{x.} and/or
\code{y.}.

\item \code{cov0}: Covariance of complete cases in \code{x.} and/or
\code{y.}.

\item \code{min0}: Minimum value of non-\code{NA} atomic elements contained
in \code{...} as a whole.

\item \code{max0}: Maximum value of non-\code{NA} atomic elements contained
in \code{...} as a whole.

\item \code{var0}: Variance of non-\code{NA} atomic elements contained in
\code{...} as a whole.

\item \code{pmin0}: Pairwise minimum values ignoring \code{NA} values.

\item \code{pmax0}: Pairwise maximum values ignoring \code{NA} values.

\item \code{mean0}: Mean value of non-\code{NA} atomic elements contained in
\code{...} as a whole.

\item \code{csds0}: Column standard deviations of \code{x} removing
non-\code{NA} values separately for each column.

\item \code{rsds0}: Row standard deviations of \code{x} removing
non-\code{NA} values separately for each row.

\item \code{csums0}: Column sums of \code{x} removing non-\code{NA} values
separately for each column.

\item \code{rsums0}: Row sums of \code{x} removing non-\code{NA} values
separately for each row.

\item \code{cmeans0}: Column means of \code{x} removing non-\code{NA} values
separately for each column.

\item \code{rmeans0}: Row means of \code{x} removing non-\code{NA}
values separately for each row.

\item \code{sets}: Create a list containing elements of indices to create
all possible combinations of 1 or more elements of an object of size
\code{n}.
}}

\seealso{
Other math: 
\code{\link{hexdec_uj}},
\code{\link{inc_uj}()}
}
\concept{math}
