% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopperr.R
\encoding{UTF-8}
\name{stopperr}
\alias{stopperr}
\alias{getterr}
\alias{purgerr}
\alias{stopper}
\alias{getter}
\alias{purger}
\alias{err_if}
\alias{err_if_not}
\alias{errs_ifs}
\alias{errs_if_nots}
\alias{err_if_pop}
\alias{errs_if_pop}
\alias{errs}
\alias{err}
\alias{banked_errs}
\alias{checkerr}
\alias{checker}
\alias{bankerr}
\alias{bankerrs}
\title{Error checking, banking, and processing}
\usage{
stopperr(..., fun = "", stack = "")

getterr()

purgerr()

stopper(..., fun = "", stack = "")

getter()

purger()

err_if(test, ..., fun = "", stack = "", d = " ")

err_if_not(test, ..., fun = "", stack = "", d = " ")

errs_ifs(..., fun = "", stack = "", d = " ")

errs_if_nots(..., fun = "", stack = "", d = " ")

err_if_pop(..., fun = "", stack = "", d = " ")

errs_if_pop(..., fun = "", stack = "")

errs(errs, stack = "")

err(..., stack = "", d = " ")

banked_errs(gens = 0)

checkerr(gens = 0)

checker(gens = 0)

bankerr(..., gens = 0, d = "")

bankerrs(..., gens = 0)
}
\arguments{
\item{...}{\strong{Unnamed} arguments.}

\item{fun}{A character scalar naming the function generating an error or errors.}

\item{stack}{An optional character vector naming the lineage of the function generating the error. If \code{NULL}, retrieves the stack under the assumption that the immediate calling function is Where the error is generated.}

\item{test}{A non-\code{NA} logical scalar.}

\item{d}{A non-\code{NA} character scalar delimiter for collapsing \code{...} into a an error message.}

\item{gens}{A \link[=cmp_nnw_scl]{complete non-negative whole-number scalar} indicating the number of generations back in the call stack in which to bank and/or check for error messages.}

\item{funs}{A \link[=cmp_chr_vec]{complete character vec} containing \code{1} or more \link[=prop_funs]{property function} names.}
}
\value{
**A **\code{\link[base]{simpleError}} \strong{object} \cr\cr \code{getterr}
\cr\cr  \strong{A character vector}                           \cr\cr \code{banked_Errs}
\cr\cr  All others are called for their side effects.
}
\description{
Bank error messages in the immediate environment of a function to allow for exhaustive error checking before throwing an exception. Results in a possibly multiple-error, accumulated message to be processed upon completion of error checking.
}
\details{
\strong{Primary stopping functions}
\tabular{ll}{  \code{stopperr} or \code{stopper}   \tab Stops execution by:                                                                                                                                                                                                             \cr
                            \tab \enumerate{
\item Posting an \code{\link{alert}} header the console with the following components: an error \code{'ERROR'} title and a subheader identifying the function Where the error originated (from \code{FUN}).
\item Posting one or more error message to the console following the header, each preceded by a bullet, Where each element of each \code{...} arg is a separate error message.
\item Creating a \code{\link[base]{simpleError}} object with an error message consisting of an \link[=lineage]{abbreviated function call lineage} given as a vector in \code{stack}.
\item Attaching the function identity, package identity, and associated message(s) to the \code{simpleError} object as attributes.
\item Archiving the \code{simpleError} object in the global variable \code{.ppp_LAST_ERR_ppp.} (which can be retrieved by calling \code{getter()} and can be purged by calling \code{purger()}). Allows for error tracing in the circumstance that R purges the last error in final error processing.
\item Calling \code{stop} with the \code{simpleError} object as the argument.                                                                                                                                           } \cr   \tab   \cr
  \code{checkerr} or \code{checker}   \tab Calls \code{stopper} with any error messages banked by the functions described in sections \emph{error banking} \emph{utilities functions} and \emph{condition-based error-checking functions}. If none are banked, does nothing.      \cr   \tab   \cr
  \code{purgerr} or \code{purger}     \tab Purges the most recent \code{\link[base]{simpleError}} object generated by this family of functions.                                                                                                               \cr   \tab   \cr
  \code{getterr} or \code{getter}     \tab Gets the most recent \code{simpleError} object generated by this family of functions.                                                                                                                                                  }
\cr Both \code{stopperr / stopper} and \code{checkerr / checker} can identify error-generating functions further up the call Stack than the function in which they are called:
\itemize{\item \code{stopperr / stopper} uses args \code{fun}, and \code{stack} to identify the error-generating function.
\item \code{checkerr / checker} uses arg \code{gens} to identify the error-generating function.}
\cr\cr \strong{Secondary stopping functions}
\cr\cr These functions are designed to be called directly from the function Where an error is generated. It gathers the name of the function generating the error rather than requiring the user to provide the function name. These functions operate as follows:
\tabular{ll}{  \code{Errs}            \tab Calls \code{stopperr} treating each element of each \code{...} arg as a separate error message, allowing for compiling multiple error message before processing.                                                              \cr   \tab   \cr
  \code{err}             \tab Calls \code{stopperr} with a single error message constructed by \link[=collapse_dots]{collapsing} all elements of all \code{...} args into a character scalar error message before processing.                                              }
\cr\cr \strong{Multiple-error conditional stopping functions}
\cr\cr These functions conditionally compile multiple errors, and if any are compiled, they notify the user and stop execution.
\tabular{ll}{  \code{errs_if_nots}    \tab Conditionally compiles errors, treating each odd-numbered \code{...} arg as a test and each even-numbered \code{...} arg as the corresponding error message if the test is \code{FALSE}, and calls \code{stopperr} if any are compiled. \cr   \tab   \cr
  \code{errs_if_pop}     \tab Calls \code{stopperr} if there are any \code{...} args, treating each \code{...} arg as a separate error message.                                                                                                                  \cr   \tab   \cr
  \code{errs_ifs}        \tab Conditionally compiles errors, treating each odd-numbered \code{...} arg as a test and each even-numbered \code{...} arg as the corresponding error message if the test is \code{TRUE}, and calls \code{stopperr} if any are compiled.                 }
\cr\cr \strong{Single-error conditional stopping functions}
\cr\cr These functions conditionally construct a single error, and if one is constructed, they notify the user and stop execution.
\tabular{ll}{  \code{err_if_pop}      \tab If there are any \code{...} args, collapsing them into a character scalar error message, and calls \code{stopperr}.                                                                                                           \cr   \tab   \cr
  \code{err_if_not}      \tab If \code{TEST = FALSE}, collapses \code{...} args to a character scalar error message and calls \code{stopperr}.                                                                                                                   \cr   \tab   \cr
  \code{err_if}          \tab If \code{TEST = TRUE}, collapses \code{...} args to a character scalar error message and calls \code{stopperr}.                                                                                                                                   }
\cr\cr \strong{Utility functions}
\cr\cr These functions purge and retrieve the most recent error generated by this family of functions.
\cr\cr \strong{Primary error banking functions}
\cr\cr Error banking utility functions: These functions are utilities for banking user-defined error messages within a function to allow for checking for multiple errors in separate statements and banking those error messages as they are checked, waiting to process banked error messages until an error checking block is completed. These functions also allow for generating and checking for error messages further up the call Stack than the function in which the error banking/processing occurs by specifying the number of generations back in the call Stack Where error banking/processing occurs in \code{gens}:
\tabular{ll}{  \code{banked_errs}     \tab Retrieves the bank of error message stored in the environment of the function \code{gens} generations back in the call Stack.                                                                                            \cr   \tab   \cr
  \code{bankerrs}        \tab Banks each element of \link[=cmp_chr_vec]{complete character vec} as an individual error message.                                                                                                                   \cr   \tab   \cr
  \code{bankerr}         \tab Banks an arbitrary error message (built by \link[=collapse_dots]{collapsing} \code{...} args) in the environment of the function \code{gens} generations back in the call Stack.                                                             }
All functions in family \code{\link{check_xxx}} also check for specific types of errors and incrementally bank errors if any are found.
\cr\cr \strong{Error checking / conditional error banking functions}
\cr\cr \strong{\code{checkerr}} checks for any banked error messages. If there are any, processes them and stops execution. Otherwise, does nothing.
}
\examples{
egStopper <- function() {stopperr('stopper demo')}
egErrs    <- function() {Errs('Errs demo1', 'Errs demo2')}
egErr     <- function() {err('err', 'demo')}
egErrors  <- function(..., tf = NA, lgl = 42, not = FALSE, pop = NULL,
                           fail = simpleError('error'), funs = 2:4, spec = 42,
                           vals = 42, class = 42, nas.or = NULL, nll.or = NA,
                           chars = '5', when.a = "error.a", when.b = "error.b") {
  bankerr(...elt(1))
  bankErrs(...elt(2), ...elt(3))
  check_tf(tf = tf)
  check_lgl(lgl = lgl)
  check_t(not = not)
  check_pop(pop = pop)
  check_fail(fail = fail)
  check_funs(c('cmp_ch1_vec', 'cmp_ngw_vec'), funs = funs)
  check_spec('cmp_ch1_vec|nll|nas', spec = spec)
  check_vals(letters, vals = vals)
  check_cls('data.frame', class)
  check_nas_or(c('cmp_ch1_vec', 'cmp_ngw_vec'), nas.or = nas.or)
  check_nll_or(c('cmp_ch1_vec', 'cmp_ngw_vec'), nll.or = nll.or)
  check_chars(letters, chars = chars)
  check_when(when.a = when.a, when.b = when.b, c('error.a', ''), c('error.b', ''))
  checkerr()
}
\dontrun{
  egstopperr()
  getterr()
  purgerr()
  getter()
  egErrs()
  egErrs()
  egErrors()
}
}
\seealso{
Other Errs: 
\code{\link{check_xxx_funs}()}
}
\concept{Errs}
