% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap.}
\alias{swap.}
\alias{swap}
\title{Swap atomic values.}
\usage{
swap.()

swap(x, old, new, all = F)
}
\arguments{
\item{x}{An atomic object.}

\item{old, new}{\link[=atm_vec]{Atomic vecs} of unique values to be replaced
and replacement values, respectively. \code{length(new)} must be in
\code{c(1, length(old))}.}

\item{all}{link[=cmp_lgl_scl]{Complete logical scalar} indicating whether all
values of \code{x} must be contained in \code{old}.}
}
\value{
An atomic object of the same dimensions as \code{x}.
}
\description{
Swap atomic values.
}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
