% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap atomic values.}
\usage{
swap(x, old, new, all = F)
}
\arguments{
\item{x}{An atomic object.}

\item{old}{Atomic vect of unique values to be replaced.}

\item{new}{Atomic vect of replacement values. \code{length(new)} must be in
\code{c(1, length(old))}.}

\item{all}{\code{TRUE} or \code{FALSE} indicating whether all values of
\code{x} must be contained in \code{old}.}
}
\value{
An atomic object of the same dimensions as \code{x}.
}
\description{
Swaps atomic values.
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
