% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\encoding{UTF-8}
\name{swap_help}
\alias{swap_help}
\alias{swap}
\alias{swap_vals}
\alias{swap_elts}
\alias{swap_rows}
\alias{swap_cols}
\title{Swap by Row Index, Column Index, Element Index, and Atomic Value.}
\usage{
swap_help()

swap(x, s1, s2, type = "i")

swap_vals(x, v1, v2)

swap_elts(x, i1, i2)

swap_rows(x, i1, i2)

swap_cols(x, i1, i2)
}
\arguments{
\item{x}{A non-empty data.frame, matrix, vlist (?VLS), or an atomic object of any dimension.}

\item{from, to}{Uniquely valued non-zero whole number vectors indexing source (\code{from}) and destination (\code{to}) positions, rows, or columns. \code{from < 0} and \code{to < 0} index from the last value rather than the first.}

\item{old}{\link[=unq_atm_vec]{A uniquely valued atomic vec} of values to be replaced.}

\item{new}{\link[=atm_vec]{An atomic vec} of replacement values. \code{length(new)} must be in \code{c(1, length(old))}.}

\item{dim}{Dimension on which swapping occurs. \code{0} indicates swapping by element index of a non-empty vector, vlist, or other atomic object. \code{1} and \code{2} indicate, respectively, swapping rows and columns of a data.frame or matrix.}

\item{all}{\code{TRUE} or \code{FALSE} indicating whether all values of \code{x} must be contained in \code{old} (ignored when \code{old} and \code{new} are \code{NULL}).}
}
\value{
An object of the same class and dimensions as \code{x}.
}
\description{
Swap by Row Index, Column Index, Element Index, and Atomic Value.
}
\section{Functions}{
\itemize{
\item \code{swap()}: Swap elements of \code{x} either by index (when \code{type == "i"}) or by value (when \code{type == "v"}).

\item \code{swap_vals()}: Swap elements of \code{x} either by value, meaning that elements of \code{x} matching \code{v1[i]} are replaced by \code{v2[i]} and vice versa.

\item \code{swap_elts()}: Swap elements of \code{x} by index, meaning that element \code{i1[i]} of \code{x} is replaced by element \code{i2[i]} of and vice versa.

\item \code{swap_rows()}: Swap rows of \code{x} by index, meaning that row \code{i1[i]} of \code{x} is replaced by row \code{i2[i]} of \code{x} and vice versa.

\item \code{swap_cols()}: Swap columns of \code{x} by index, meaning that column \code{i1[i]} of \code{x} is replaced by column \code{i2[i]} of \code{x} and vice versa.

}}
\examples{
AF <- LETTERS[1:6]
M6 <- matrix(1:36, nrow = 6)

i1 <- 1:3
i2 <- 6:4

v1 <- c("A", "B", "C")
v2 <- c("a", "b", "c")

AF
swap_vals(AF, v1, v2)
swap_elts(AF, i1, i2)
swap(AF, v1, v2, type = "v")
swap(AF, i1, i2, type = "i")

M6
swap_rows(M6, i1, i2)
swap_cols(M6, i1, i2)
}
\seealso{
Other forks: 
\code{\link{case}()},
\code{\link{fork_help}()}
}
\concept{forks}
