% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap_uj}
\alias{swap_uj}
\alias{swap}
\title{Swap atomic values.}
\usage{
swap_uj()

swap(x, old, new, all = F)
}
\arguments{
\item{x}{An atomic object.}

\item{old}{Atomic vect of unique values to be replaced.}

\item{new}{Atomic vect of replacement values. \code{length(new)} must be in
\code{c(1, length(old))}.}

\item{all}{\code{TRUE} or \code{FALSE} indicating whether all values of
\code{x} must be contained in \code{old}.}
}
\value{
An atomic object of the same dimensions as \code{x}.
}
\description{
Swap atomic values.
}
\section{Functions}{
\itemize{
\item \code{swap}: Swaps atomic values.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
