% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\encoding{UTF-8}
\name{tb}
\alias{tb}
\alias{trb}
\alias{tbr}
\alias{as_tb}
\alias{is_tb}
\title{\code{tibble} wrappers}
\usage{
tb(
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

trb(...)

tbr(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

as_tb(
  x,
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

is_tb(x)
}
\value{
A tibble.
}
\description{
Thin wrappers of functions from package \code{tibble}. See links in \emph{details} for function-specific documentation.
}
\details{
\tabular{ll}{  \code{as_tb}   \tab Thinly wraps \code{\link[tibble]{as_tibble}}.  \cr
  \code{is_tb}   \tab Thinly wraps \code{\link[tibble]{is_tibble}}.  \cr
  \code{trb}     \tab Thinly wraps \code{\link[tibble]{tribble}}.    \cr
  \code{tbr}     \tab Thinly wraps \code{\link[tibble]{tibble_row}}. \cr
  \code{tb}      \tab Thinly wraps \code{\link[tibble]{tibble}}.       }
}
\examples{
tb(letters = letters, numbers = 1:26, constant = pi)
trb(~letters, ~numbers, ~constant,
    "a"     , 1,      , pi       ,
    "b"     , 2,      , pi       ,
    "c"     , 3,      , pi       )
tbr("a", 1, pi)
as_tb(list(letters = letters, numbers = 1:26, constant = pi))
is_tb(tb(letters))
}
\seealso{
Other wraps: 
\code{\link{basics}()},
\code{\link{dp}()},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{rd}()},
\code{\link{sca}()}
}
\concept{wraps}
