% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\encoding{UTF-8}
\name{tb_help}
\alias{tb_help}
\alias{tb}
\alias{trb}
\alias{tbr}
\alias{as_tb}
\alias{is_tb}
\title{\code{tibble} wrappers}
\usage{
tb_help()

tb(
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

trb(...)

tbr(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

as_tb(
  x,
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

is_tb(x)
}
\value{
A tibble.
}
\description{
Thin wrappers of functions from package \code{tibble}.
}
\section{Functions}{
\itemize{
\item \code{tb()}: Thinly wraps \code{\link[tibble]{tibble}}.

\item \code{trb()}: Thinly wraps \code{\link[tibble]{tribble}}.

\item \code{tbr()}: Thinly wraps \code{\link[tibble]{tibble_row}}.

\item \code{as_tb()}: Thinly wraps \code{\link[tibble]{as_tibble}}.

\item \code{is_tb()}: Thinly wraps \code{\link[tibble]{is_tibble}}.

}}
\examples{
tb(letters = letters, numbers = 1:26, constant = pi)
trb(~letters, ~numbers, ~constant,
    "a"     , 1       , pi       ,
    "b"     , 2       , pi       ,
    "c"     , 3       , pi       )
tbr("a", 1, pi)
as_tb(list(letters = letters, numbers = 1:26, constant = pi))
is_tb(tb(letters))
}
\seealso{
Other wraps: 
\code{\link{basics_help}()},
\code{\link{dp}()},
\code{\link{dt}()},
\code{\link{gg_help}()},
\code{\link{gr}},
\code{\link{rd_help}()}
}
\concept{wraps}
