% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_tools.R
\name{tibble_tools}
\alias{tibble_tools}
\alias{tb}
\alias{rtb}
\alias{trb}
\alias{astb}
\title{Aliases and wraps for functions in the package \code{tibble}.}
\usage{
tb(
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

rtb(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

trb(...)

astb(
  x,
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)
}
\description{
Alias for \code{\link[tibble]{tibble}} (create a tibble).
}
\section{Functions}{
\itemize{
\item \code{rtb}: Alias for \code{\link[tibble]{tibble_row}} (create a
row tibble).

\item \code{trb}: Alias for \code{\link[tibble]{tribble}} (create a
by-row parsed
tibble).

\item \code{astb}: Alias for \code{\link[tibble]{as_tibble}} (coerce to
tibble).
}}

