% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocase.R
\name{tocase.}
\alias{tocase.}
\alias{tocase}
\alias{tosentence}
\alias{totitle}
\alias{lc}
\alias{sc}
\alias{tc}
\alias{uc}
\title{Convert string case}
\usage{
tocase.()

tocase(case., ..., a. = T)

tosentence(..., a. = T)

totitle(..., a. = T)

lc(..., a. = T)

sc(..., a. = T)

tc(..., a. = T)

uc(..., a. = T)
}
\arguments{
\item{case.}{\link[cmp_ch1_scl]{Complete onechar scalar}: \code{'l'},
\code{'s'}, \code{'t'}, or \code{'u'} to indicate lower, sentence, title,
or upper case, respectively. Case insensitive.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{a.}{\link[cmp_lgl_scl]{Complete logical scalar} indicating whether to
\link[a]{atomize} \code{...} before processing.}
}
\description{
Convert strings to lower, upper, sentence, and title cases
(supplements \code{tolower} and \code{toupper}).

Convert to the case specified in \code{case}.
}
\section{Functions}{
\itemize{
\item \code{tocase()}: Convert among lower, upper, sentence, and title case.

\item \code{tosentence()}: Convert to sentence case via
\code{\link[stringr]{str_to_sentence}}.

\item \code{totitle()}: Convert to sentence case via
\code{\link[stringr]{str_to_title}}.

\item \code{lc()}: Converts to lower case via \code{\link[base]{tolower}}.

\item \code{sc()}: Convert to sentence case via
\code{\link[stringr]{str_to_sentence}}.

\item \code{tc()}: Convert to title case via
\code{\link[stringr]{str_to_title}}.

\item \code{uc()}: Converts to upper case via \code{\link[base]{tolower}}.

}}
\examples{
TestTitle    <- "This is a Title"
TestSentence <- "This is a sentence."
TestWords    <- "this is an ordered group of words"
TestSet      <- c("a", "set", "of", "words")
TestVector   <- c(TestTitle, TestSentence, TestWords, TestSet)
lc(TestVector)
lc(TestTitle, TestSentence, TestWords, TestSet)
sc(TestTitle, TestSentence, TestWords, TestSet)
tc(TestTitle, TestSentence, TestWords, TestSet)
uc(TestTitle, TestSentence, TestWords, TestSet)
scase(TestTitle, TestSentence, TestWords, TestSet, case = "LOWER")
scase(TestVector, case = "l")
scase(TestVector, case = "s")
scase(TestTitle, TestSentence, TestWords, TestSet, case = "T")
scase(TestVector, case = "UPPER")
}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}
}
\concept{strings}
