% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_ccc.R
\name{ttt_ccc.}
\alias{ttt_ccc.}
\alias{ttt_ccc_vals}
\alias{atm_arr}
\alias{atm_gen}
\alias{atm_mat}
\alias{atm_mvc}
\alias{atm_scl}
\alias{atm_dtf}
\alias{atm_vec}
\alias{atm_vls}
\alias{pop_arr}
\alias{pop_gen}
\alias{pop_mat}
\alias{pop_mvc}
\alias{pop_scl}
\alias{pop_dtf}
\alias{pop_vec}
\alias{pop_vls}
\alias{cmp_arr}
\alias{cmp_gen}
\alias{cmp_mat}
\alias{cmp_mvc}
\alias{cmp_scl}
\alias{cmp_dtf}
\alias{cmp_vec}
\alias{cmp_vls}
\alias{emp_arr}
\alias{emp_gen}
\alias{emp_mat}
\alias{emp_mvc}
\alias{emp_scl}
\alias{emp_dtf}
\alias{emp_vec}
\alias{emp_vls}
\alias{rcr_arr}
\alias{rcr_gen}
\alias{rcr_mat}
\alias{rcr_mvc}
\alias{rcr_scl}
\alias{rcr_dtf}
\alias{rcr_vec}
\alias{rcr_vls}
\title{Fundamental Type + Extended Class Properties}
\usage{
ttt_ccc.()

ttt_ccc_vals()

atm_arr(x, mmm = NULL)

atm_gen(x, mmm = NULL)

atm_mat(x, mmm = NULL)

atm_mvc(x, mmm = NULL)

atm_scl(x, mmm = NULL)

atm_dtf(x, mmm = NULL)

atm_vec(x, mmm = NULL)

atm_vls(x, mmm = NULL)

pop_arr(x, mmm = NULL)

pop_gen(x, mmm = NULL)

pop_mat(x, mmm = NULL)

pop_mvc(x, mmm = NULL)

pop_scl(x, mmm = NULL)

pop_dtf(x, mmm = NULL)

pop_vec(x, mmm = NULL)

pop_vls(x, mmm = NULL)

cmp_arr(x)

cmp_gen(x)

cmp_mat(x)

cmp_mvc(x)

cmp_scl(x)

cmp_dtf(x)

cmp_vec(x)

cmp_vls(x)

emp_arr(x)

emp_gen(x)

emp_mat(x)

emp_mvc(x)

emp_scl(x)

emp_dtf(x)

emp_vec(x)

emp_vls(x)

rcr_arr(x)

rcr_gen(x)

rcr_mat(x)

rcr_mvc(x)

rcr_scl(x)

rcr_dtf(x)

rcr_vec(x)

rcr_vls(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\tabular{ll}{
\code{ttt_ccc_vals}   \tab A character vector.                          \cr
\code{[ttt]_[ccc]}    \tab A logical scalar.                              }
}
\description{
\tabular{ll}{
\code{ttt_ccc_vals}   \tab Gets a character vector of all possible
\link[=ttt]{fundamental type} +
\link[=ccc]{extended class} properties.      \cr
\code{[ttt]_[ccc]}    \tab Evaluates whether \code{x} is of the
\link[=ttt]{fundamental type} represented by the
placeholder \code{[ttt]} and of the
\link[=ccc]{extended class} represented by the
placeholder \code{[ccc]}.                      }
}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
