% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_ccc.R
\docType{data}
\name{ttt_ccc}
\alias{ttt_ccc}
\alias{ttt_ccc_vals}
\alias{any_arr}
\alias{any_gen}
\alias{any_mat}
\alias{any_mvc}
\alias{any_scl}
\alias{any_tbl}
\alias{any_vec}
\alias{any_vls}
\alias{any_vtp}
\alias{atm_arr}
\alias{atm_gen}
\alias{atm_mat}
\alias{atm_mvc}
\alias{atm_scl}
\alias{atm_tbl}
\alias{atm_vec}
\alias{atm_vls}
\alias{atm_vtp}
\alias{pop_arr}
\alias{pop_gen}
\alias{pop_mat}
\alias{pop_mvc}
\alias{pop_scl}
\alias{pop_tbl}
\alias{pop_vec}
\alias{pop_vls}
\alias{pop_vtp}
\alias{cmp_arr}
\alias{cmp_gen}
\alias{cmp_mat}
\alias{cmp_mvc}
\alias{cmp_scl}
\alias{cmp_tbl}
\alias{cmp_vec}
\alias{cmp_vtp}
\alias{cmp_vls}
\alias{emp_arr}
\alias{emp_gen}
\alias{emp_mat}
\alias{emp_mvc}
\alias{emp_scl}
\alias{emp_tbl}
\alias{emp_vec}
\alias{emp_vls}
\alias{emp_vtp}
\alias{rcr_arr}
\alias{rcr_gen}
\alias{rcr_mat}
\alias{rcr_mvc}
\alias{rcr_scl}
\alias{rcr_tbl}
\alias{rcr_vec}
\alias{rcr_vls}
\alias{rcr_vtp}
\title{Check whether \code{x} of a certain \link[=ccc]{extended class} + a
certain \link[=ttt]{fundamental type}?}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ttt_ccc

ttt_ccc_vals()

any_arr(x)

any_gen(x)

any_mat(x)

any_mvc(x)

any_scl(x)

any_tbl(x)

any_vec(x)

any_vls(x)

any_vtp(x)

atm_arr(x, xxx = NULL)

atm_gen(x, xxx = NULL)

atm_mat(x, xxx = NULL)

atm_mvc(x, xxx = NULL)

atm_scl(x, xxx = NULL)

atm_tbl(x, xxx = NULL)

atm_vec(x, xxx = NULL)

atm_vls(x, xxx = NULL)

atm_vtp(x, xxx = NULL)

pop_arr(x, xxx = NULL)

pop_gen(x, xxx = NULL)

pop_mat(x, xxx = NULL)

pop_mvc(x, xxx = NULL)

pop_scl(x, xxx = NULL)

pop_tbl(x, xxx = NULL)

pop_vec(x, xxx = NULL)

pop_vls(x, xxx = NULL)

pop_vtp(x, xxx = NULL)

cmp_arr(x, xxx = NULL)

cmp_gen(x, xxx = NULL)

cmp_mat(x, xxx = NULL)

cmp_mvc(x, xxx = NULL)

cmp_scl(x, xxx = NULL)

cmp_tbl(x, xxx = NULL)

cmp_vec(x, xxx = NULL)

cmp_vtp(x, xxx = NULL)

cmp_vls(x, xxx = NULL)

emp_arr(x)

emp_gen(x)

emp_mat(x)

emp_mvc(x)

emp_scl(x)

emp_tbl(x)

emp_vec(x)

emp_vls(x)

emp_vtp(x)

rcr_arr(x)

rcr_gen(x)

rcr_mat(x)

rcr_mvc(x)

rcr_scl(x)

rcr_tbl(x)

rcr_vec(x)

rcr_vls(x)

rcr_vtp(x)
}
\arguments{
\item{x}{An object.}

\item{xxx}{Optional character scalar extended mode (see
\code{\link{mmm_vals}}).}
}
\value{
Logical scalar.
}
\description{
Check whether \code{x} of a certain \link[=ccc]{extended class} + a
certain \link[=ttt]{fundamental type}?
}
\section{Functions}{
\itemize{
\item \code{ttt_ccc_vals}: Get a character vector of all possible fundamental
type plus extended class properties.

\item \code{any_arr}: Is \code{x} an array of any type?

\item \code{any_gen}: Is \code{x} any type of generic?

\item \code{any_mat}: Is \code{x} a matrix of any type?

\item \code{any_mvc}: Is \code{x} an mvect of any type?

\item \code{any_scl}: Is \code{x} is any type of scalar?

\item \code{any_tbl}: Is \code{x} any type of tibble?

\item \code{any_vec}: Is \code{x} any type of vec?

\item \code{any_vls}: Is \code{x} any type of vlist?

\item \code{any_vtp}: Is \code{x} any type of vtype?

\item \code{atm_arr}: Is \code{x} an atomic array?

\item \code{atm_gen}: Is \code{x} an atomic generic?

\item \code{atm_mat}: Is \code{x} an atomic matrix?

\item \code{atm_mvc}: Is \code{x} an atomic mvect?

\item \code{atm_scl}: Is \code{x} is an atomic scalar?

\item \code{atm_tbl}: Is \code{x} an atomic tibble?

\item \code{atm_vec}: Is \code{x} an atomic vec?

\item \code{atm_vls}: Is \code{x} an atomic vlist?

\item \code{atm_vtp}: Is \code{x} an atomic \code{vtype}?

\item \code{pop_arr}: Is \code{x} a populated array?

\item \code{pop_gen}: Is \code{x} a populated generic?

\item \code{pop_mat}: Is \code{x} a populated generic?

\item \code{pop_mvc}: Is \code{x} a populated generic?

\item \code{pop_scl}: Is \code{x} a populated generic?

\item \code{pop_tbl}: Is \code{x} a populated tibble?

\item \code{pop_vec}: Is \code{x} a populated tibble?

\item \code{pop_vls}: Is \code{x} a populated vlist?

\item \code{pop_vtp}: Is \code{x} a populated \code{vtype}?

\item \code{cmp_arr}: Is \code{x} a complete array?

\item \code{cmp_gen}: Is \code{x} a complete generic?

\item \code{cmp_mat}: Is \code{x} a complete atomic matrix?

\item \code{cmp_mvc}: Is \code{x} a complete atomic mvect?

\item \code{cmp_scl}: Is \code{x} is an complete atomic scalar?

\item \code{cmp_tbl}: Is \code{x} a complete atomic tibble?

\item \code{cmp_vec}: Is \code{x} a complete atomic vec?

\item \code{cmp_vtp}: Is \code{x} a complete atomic \code{vtype}?

\item \code{cmp_vls}: Is \code{x} a complete vlist?

\item \code{emp_arr}: Is \code{x} an empty array?

\item \code{emp_gen}: Is \code{x} an empty generic?

\item \code{emp_mat}: Is \code{x} an empty generic?

\item \code{emp_mvc}: Is \code{x} an empty generic?

\item \code{emp_scl}: Is \code{x} an empty generic?

\item \code{emp_tbl}: Is \code{x} an empty tibble?

\item \code{emp_vec}: Is \code{x} an empty generic?

\item \code{emp_vls}: Is \code{x} an empty vlist?

\item \code{emp_vtp}: Is \code{x} an empty \code{vtype}?

\item \code{rcr_arr}: Is \code{x} a recursive array?

\item \code{rcr_gen}: Is \code{x} a recursive generic?

\item \code{rcr_mat}: Is \code{x} a recursive matrix?

\item \code{rcr_mvc}: Is \code{x} a recursive mvect?

\item \code{rcr_scl}: Is \code{x} is a recursive scalar?

\item \code{rcr_tbl}: Is \code{x} a recursive tibble?

\item \code{rcr_vec}: Is \code{x} a recursive vec?

\item \code{rcr_vls}: Is \code{x} a recursive vlist?

\item \code{rcr_vtp}: Is \code{x} a recursive \code{vtype}?
}}

\keyword{datasets}
