% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_mmm.R
\name{ttt_mmm.}
\alias{ttt_mmm.}
\alias{ttt_mmm_vals}
\alias{atm_ch1}
\alias{atm_chr}
\alias{atm_clr}
\alias{atm_evn}
\alias{atm_fac}
\alias{atm_frc}
\alias{atm_ind}
\alias{atm_lgl}
\alias{atm_neg}
\alias{atm_ngw}
\alias{atm_nng}
\alias{atm_nnw}
\alias{atm_nps}
\alias{atm_npw}
\alias{atm_nst}
\alias{atm_num}
\alias{atm_odd}
\alias{atm_ord}
\alias{atm_pct}
\alias{atm_pos}
\alias{atm_ppn}
\alias{atm_psw}
\alias{atm_srt}
\alias{atm_str}
\alias{atm_uno}
\alias{atm_whl}
\alias{pop_atm}
\alias{pop_ch1}
\alias{pop_chr}
\alias{pop_clr}
\alias{pop_evn}
\alias{pop_fac}
\alias{pop_frc}
\alias{pop_ind}
\alias{pop_lgl}
\alias{pop_neg}
\alias{pop_ngw}
\alias{pop_nng}
\alias{pop_nnw}
\alias{pop_nps}
\alias{pop_npw}
\alias{pop_nst}
\alias{pop_num}
\alias{pop_odd}
\alias{pop_ord}
\alias{pop_pct}
\alias{pop_pos}
\alias{pop_ppn}
\alias{pop_psw}
\alias{pop_srt}
\alias{pop_str}
\alias{pop_uno}
\alias{pop_whl}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\alias{emp_atm}
\alias{emp_ch1}
\alias{emp_chr}
\alias{emp_clr}
\alias{emp_evn}
\alias{emp_fac}
\alias{emp_frc}
\alias{emp_ind}
\alias{emp_lgl}
\alias{emp_neg}
\alias{emp_ngw}
\alias{emp_nng}
\alias{emp_nnw}
\alias{emp_nps}
\alias{emp_npw}
\alias{emp_nst}
\alias{emp_num}
\alias{emp_odd}
\alias{emp_ord}
\alias{emp_pct}
\alias{emp_pos}
\alias{emp_ppn}
\alias{emp_psw}
\alias{emp_srt}
\alias{emp_str}
\alias{emp_uno}
\alias{emp_whl}
\title{Fundamental Type + Extended Mode Properties}
\usage{
ttt_mmm.()

ttt_mmm_vals()

atm_ch1(x)

atm_chr(x)

atm_clr(x)

atm_evn(x)

atm_fac(x)

atm_frc(x)

atm_ind(x)

atm_lgl(x)

atm_neg(x)

atm_ngw(x)

atm_nng(x)

atm_nnw(x)

atm_nps(x)

atm_npw(x)

atm_nst(x)

atm_num(x)

atm_odd(x)

atm_ord(x)

atm_pct(x)

atm_pos(x)

atm_ppn(x)

atm_psw(x)

atm_srt(x)

atm_str(x)

atm_uno(x)

atm_whl(x)

pop_atm(x)

pop_ch1(x)

pop_chr(x)

pop_clr(x)

pop_evn(x)

pop_fac(x)

pop_frc(x)

pop_ind(x)

pop_lgl(x)

pop_neg(x)

pop_ngw(x)

pop_nng(x)

pop_nnw(x)

pop_nps(x)

pop_npw(x)

pop_nst(x)

pop_num(x)

pop_odd(x)

pop_ord(x)

pop_pct(x)

pop_pos(x)

pop_ppn(x)

pop_psw(x)

pop_srt(x)

pop_str(x)

pop_uno(x)

pop_whl(x)

cmp_atm(x)

cmp_ch1(x)

cmp_chr(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)

emp_atm(x)

emp_ch1(x)

emp_chr(x)

emp_clr(x)

emp_evn(x)

emp_fac(x)

emp_frc(x)

emp_ind(x)

emp_lgl(x)

emp_neg(x)

emp_ngw(x)

emp_nng(x)

emp_nnw(x)

emp_nps(x)

emp_npw(x)

emp_nst(x)

emp_num(x)

emp_odd(x)

emp_ord(x)

emp_pct(x)

emp_pos(x)

emp_ppn(x)

emp_psw(x)

emp_srt(x)

emp_str(x)

emp_uno(x)

emp_whl(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\tabular{ll}{
\code{ttt_mmm_vals}   \tab A character vector.                          \cr
\code{[ttt]_[mmm]}    \tab A logical scalar.                              }
}
\description{
\tabular{ll}{
\code{ttt_mmm_vals}   \tab Gets a character vector of all possible
\link[=ttt]{fundamental type} +
\link[=mmm]{extended mode} properties.       \cr
\code{[ttt]_[mmm]}    \tab Evaluates whether \code{x} is of the
\link[=ttt]{fundamental type} represented by the
placeholder \code{[ttt]} and of the
\link[=mmm]{extended mode} represented by the
placeholder \code{[mmm]}.                      }
}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()}
}
\concept{props}
