% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_mmm.R
\name{ttt_mmm_uj}
\alias{ttt_mmm_uj}
\alias{ttt_mmm_vals}
\alias{pop_atm}
\alias{pop_ch1}
\alias{pop_chr}
\alias{pop_clr}
\alias{pop_evn}
\alias{pop_fac}
\alias{pop_frc}
\alias{pop_ind}
\alias{pop_lgl}
\alias{pop_neg}
\alias{pop_ngw}
\alias{pop_nng}
\alias{pop_nnw}
\alias{pop_nps}
\alias{pop_npw}
\alias{pop_nst}
\alias{pop_num}
\alias{pop_odd}
\alias{pop_ord}
\alias{pop_pct}
\alias{pop_pos}
\alias{pop_ppn}
\alias{pop_psw}
\alias{pop_srt}
\alias{pop_str}
\alias{pop_uno}
\alias{pop_whl}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\alias{emp_atm}
\alias{emp_ch1}
\alias{emp_chr}
\alias{emp_clr}
\alias{emp_evn}
\alias{emp_fac}
\alias{emp_frc}
\alias{emp_ind}
\alias{emp_lgl}
\alias{emp_neg}
\alias{emp_ngw}
\alias{emp_nng}
\alias{emp_nnw}
\alias{emp_nps}
\alias{emp_npw}
\alias{emp_nst}
\alias{emp_num}
\alias{emp_odd}
\alias{emp_ord}
\alias{emp_pct}
\alias{emp_pos}
\alias{emp_ppn}
\alias{emp_psw}
\alias{emp_srt}
\alias{emp_str}
\alias{emp_uno}
\alias{emp_whl}
\alias{rcr_atm}
\alias{rcr_ch1}
\alias{rcr_chr}
\alias{rcr_clr}
\alias{rcr_evn}
\alias{rcr_fac}
\alias{rcr_frc}
\alias{rcr_ind}
\alias{rcr_lgl}
\alias{rcr_neg}
\alias{rcr_ngw}
\alias{rcr_nng}
\alias{rcr_nnw}
\alias{rcr_nps}
\alias{rcr_npw}
\alias{rcr_nst}
\alias{rcr_num}
\alias{rcr_odd}
\alias{rcr_ord}
\alias{rcr_pct}
\alias{rcr_pos}
\alias{rcr_ppn}
\alias{rcr_psw}
\alias{rcr_srt}
\alias{rcr_str}
\alias{rcr_uno}
\alias{rcr_whl}
\title{Fundamental type + extended mode}
\usage{
ttt_mmm_uj()

ttt_mmm_vals()

pop_atm(x)

pop_ch1(x)

pop_chr(x)

pop_clr(x)

pop_evn(x)

pop_fac(x)

pop_frc(x)

pop_ind(x)

pop_lgl(x)

pop_neg(x)

pop_ngw(x)

pop_nng(x)

pop_nnw(x)

pop_nps(x)

pop_npw(x)

pop_nst(x)

pop_num(x)

pop_odd(x)

pop_ord(x)

pop_pct(x)

pop_pos(x)

pop_ppn(x)

pop_psw(x)

pop_srt(x)

pop_str(x)

pop_uno(x)

pop_whl(x)

cmp_atm(x)

cmp_ch1(x)

cmp_chr(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)

emp_atm(x)

emp_ch1(x)

emp_chr(x)

emp_clr(x)

emp_evn(x)

emp_fac(x)

emp_frc(x)

emp_ind(x)

emp_lgl(x)

emp_neg(x)

emp_ngw(x)

emp_nng(x)

emp_nnw(x)

emp_nps(x)

emp_npw(x)

emp_nst(x)

emp_num(x)

emp_odd(x)

emp_ord(x)

emp_pct(x)

emp_pos(x)

emp_ppn(x)

emp_psw(x)

emp_srt(x)

emp_str(x)

emp_uno(x)

emp_whl(x)

rcr_atm(x)

rcr_ch1(x)

rcr_chr(x)

rcr_clr(x)

rcr_evn(x)

rcr_fac(x)

rcr_frc(x)

rcr_ind(x)

rcr_lgl(x)

rcr_neg(x)

rcr_ngw(x)

rcr_nng(x)

rcr_nnw(x)

rcr_nps(x)

rcr_npw(x)

rcr_nst(x)

rcr_num(x)

rcr_odd(x)

rcr_ord(x)

rcr_pct(x)

rcr_pos(x)

rcr_ppn(x)

rcr_psw(x)

rcr_srt(x)

rcr_str(x)

rcr_uno(x)

rcr_whl(x)
}
\arguments{
\item{x}{An object.}

\item{xxx}{Optional character scalar extended mode (see
\code{\link{mmm_vals}}).}
}
\value{
Logical scalar.
}
\description{
Fundamental type + extended mode
}
\section{Functions}{
\itemize{
\item \code{ttt_mmm_vals}: Get a character vector of all possible fundamental
type plus extended class properties.

\item \code{pop_atm}: Is \code{x} a populated atomic object?
(\code{\link{ipop}}, \code{\link{iatm}}).

\item \code{pop_ch1}: Is \code{x} a populated onechar object?
(\code{\link{ipop}}, \code{\link{ich1}}).

\item \code{pop_chr}: Is \code{x} a populated character object?
(\code{\link{ipop}}, \code{\link{ichr}}).

\item \code{pop_clr}: Is \code{x} a populated color object?
(\code{\link{ipop}}, \code{\link{iclr}}).

\item \code{pop_evn}: Is \code{x} a populated even-valued
whole-number object?
(\code{\link{ipop}}, \code{\link{ievn}}).

\item \code{pop_fac}: Is \code{x} a populated factor object?
(\code{\link{ipop}}, \code{\link{ifac}}).

\item \code{pop_frc}: Is \code{x} a populated fractional-valued
numeric object?
(\code{\link{ipop}}, \code{\link{ifrc}}).

\item \code{pop_ind}: Is \code{x} a populated indexer object?
(\code{\link{ipop}}, \code{\link{iind}}).

\item \code{pop_lgl}: Is \code{x} a populated logical object?
(\code{\link{ipop}}, \code{\link{ilgl}}).

\item \code{pop_neg}: Is \code{x} a populated negative-valued
numeric object?
(\code{\link{ipop}}, \code{\link{ineg}}).

\item \code{pop_ngw}: Is \code{x} a populated negative-valued
whole-number object?
(\code{\link{ipop}}, \code{\link{ingw}}).

\item \code{pop_nng}: Is \code{x} a populated non-negative-valued
numeric object?
(\code{\link{ipop}}, \code{\link{inng}}).

\item \code{pop_nnw}: Is \code{x} a populated non-negative-valued
whole-number object?
(\code{\link{ipop}}, \code{\link{innw}}).

\item \code{pop_nps}: Is \code{x} a populated non-positive-valued
numeric object?
(\code{\link{ipop}}, \code{\link{inps}}).

\item \code{pop_npw}: Is \code{x} a populated non-positive-valued
whole-number object?
(\code{\link{ipop}}, \code{\link{inpw}}).

\item \code{pop_nst}: Is \code{x} a populated non-sortable object?
(\code{\link{ipop}}, \code{\link{inst}}).

\item \code{pop_num}: Is \code{x} a populated numeric object?
(\code{\link{ipop}}, \code{\link{inum}}).

\item \code{pop_odd}: Is \code{x} a populated odd-valued
whole-number object?
(\code{\link{ipop}}, \code{\link{iodd}}).

\item \code{pop_ord}: Is \code{x} a populated ordered-factor object?
(\code{\link{ipop}}, \code{\link{iord}}).

\item \code{pop_pct}: Is \code{x} a populated percent-valued
(0-100) numeric object?
(\code{\link{ipop}}, \code{\link{ipct}}).

\item \code{pop_pos}: Is \code{x} a populated positive-valued
numeric object?
(\code{\link{ipop}}, \code{\link{ipos}}).

\item \code{pop_ppn}: Is \code{x} a populated proportion-valued
(0-1) numeric object?
(\code{\link{ipop}}, \code{\link{ippn}}).

\item \code{pop_psw}: Is \code{x} a populated positive-valued
whole-number object?
(\code{\link{ipop}}, \code{\link{ipsw}}).

\item \code{pop_srt}: Is \code{x} a populated sortable object?
(\code{\link{ipop}}, \code{\link{isrt}}).

\item \code{pop_str}: Is \code{x} a populated string object?
(\code{\link{ipop}}, \code{\link{istr}}).

\item \code{pop_uno}: Is \code{x} a populated unordered-factor object?
(\code{\link{ipop}}, \code{\link{iuno}}).

\item \code{pop_whl}: Is \code{x} a populated whole-number object?
(\code{\link{ipop}}, \code{\link{iwhl}}).

\item \code{cmp_atm}: Is \code{x} a complete atomic object?
(\code{\link{icmp}}, \code{\link{iatm}}).

\item \code{cmp_ch1}: Is \code{x} a complete onechar object?
(\code{\link{icmp}}, \code{\link{ich1}}).

\item \code{cmp_chr}: Is \code{x} a complete character object?
(\code{\link{icmp}}, \code{\link{ichr}}).

\item \code{cmp_clr}: Is \code{x} a complete color object?
(\code{\link{icmp}}, \code{\link{iclr}}).

\item \code{cmp_evn}: Is \code{x} a complete even-valued
whole-number object?
(\code{\link{icmp}}, \code{\link{ievn}}).

\item \code{cmp_fac}: Is \code{x} a complete factor object?
(\code{\link{icmp}}, \code{\link{ifac}}).

\item \code{cmp_frc}: Is \code{x} a complete fractional-valued
numeric object?
(\code{\link{icmp}}, \code{\link{ifrc}}).

\item \code{cmp_ind}: Is \code{x} a complete indexer object?
(\code{\link{icmp}}, \code{\link{iind}}).

\item \code{cmp_lgl}: Is \code{x} a complete logical object?
(\code{\link{icmp}}, \code{\link{ilgl}}).

\item \code{cmp_neg}: Is \code{x} a complete negative-valued
numeric object?
(\code{\link{icmp}}, \code{\link{ineg}}).

\item \code{cmp_ngw}: Is \code{x} a complete negative-valued
whole-number object?
(\code{\link{icmp}}, \code{\link{ingw}}).

\item \code{cmp_nng}: Is \code{x} a complete non-negative-valued
numeric object?
(\code{\link{icmp}}, \code{\link{inng}}).

\item \code{cmp_nnw}: Is \code{x} a complete non-negative-valued
whole-number object?
(\code{\link{icmp}}, \code{\link{innw}}).

\item \code{cmp_nps}: Is \code{x} a complete non-positive-valued
numeric object?
(\code{\link{icmp}}, \code{\link{inps}}).

\item \code{cmp_npw}: Is \code{x} a complete non-positive-valued
whole-number object?
(\code{\link{icmp}}, \code{\link{inpw}}).

\item \code{cmp_nst}: Is \code{x} a complete non-sortable object?
(\code{\link{icmp}}, \code{\link{inst}}).

\item \code{cmp_num}: Is \code{x} a complete numeric object?
(\code{\link{icmp}}, \code{\link{inum}}).

\item \code{cmp_odd}: Is \code{x} a complete odd-valued
whole-number object?
(\code{\link{icmp}}, \code{\link{iodd}}).

\item \code{cmp_ord}: Is \code{x} a complete ordered-factor object?
(\code{\link{icmp}}, \code{\link{iord}}).

\item \code{cmp_pct}: Is \code{x} a complete percent-valued
(0-100) numeric object?
(\code{\link{icmp}}, \code{\link{ipct}}).

\item \code{cmp_pos}: Is \code{x} a complete positive-valued
numeric object?
(\code{\link{icmp}}, \code{\link{ipos}}).

\item \code{cmp_ppn}: Is \code{x} a complete proportion-valued
(0-1) numeric object?
(\code{\link{icmp}}, \code{\link{ippn}}).

\item \code{cmp_psw}: Is \code{x} a complete positive-valued
whole-number object?
(\code{\link{icmp}}, \code{\link{ipsw}}).

\item \code{cmp_srt}: Is \code{x} a complete sortable object?
(\code{\link{icmp}}, \code{\link{isrt}}).

\item \code{cmp_str}: Is \code{x} a complete string object?
(\code{\link{icmp}}, \code{\link{istr}}).

\item \code{cmp_uno}: Is \code{x} a complete unordered-factor object?
(\code{\link{icmp}}, \code{\link{iuno}}).

\item \code{cmp_whl}: Is \code{x} a complete whole-number object?
(\code{\link{icmp}}, \code{\link{iwhl}}).

\item \code{emp_atm}: Is \code{x} an empty atomic object?
(\code{\link{iemp}}, \code{\link{iatm}}).

\item \code{emp_ch1}: Is \code{x} an empty onechar object?
(\code{\link{iemp}}, \code{\link{ich1}}).

\item \code{emp_chr}: Is \code{x} an empty character object?
(\code{\link{iemp}}, \code{\link{ichr}}).

\item \code{emp_clr}: Is \code{x} an empty color object?
(\code{\link{iemp}}, \code{\link{iclr}}).

\item \code{emp_evn}: Is \code{x} an empty even-valued
whole-number object?
(\code{\link{iemp}}, \code{\link{ievn}}).

\item \code{emp_fac}: Is \code{x} an empty factor object?
(\code{\link{iemp}}, \code{\link{ifac}}).

\item \code{emp_frc}: Is \code{x} an empty fractional-valued
numeric object?
(\code{\link{iemp}}, \code{\link{ifrc}}).

\item \code{emp_ind}: Is \code{x} an empty indexer object?
(\code{\link{iemp}}, \code{\link{iind}}).

\item \code{emp_lgl}: Is \code{x} an empty logical object?
(\code{\link{iemp}}, \code{\link{ilgl}}).

\item \code{emp_neg}: Is \code{x} an empty negative-valued
numeric object?
(\code{\link{iemp}}, \code{\link{ineg}}).

\item \code{emp_ngw}: Is \code{x} an empty negative-valued
whole-number object?
(\code{\link{iemp}}, \code{\link{ingw}}).

\item \code{emp_nng}: Is \code{x} an empty non-negative-valued
numeric object?
(\code{\link{iemp}}, \code{\link{inng}}).

\item \code{emp_nnw}: Is \code{x} an empty non-negative-valued
whole-number object?
(\code{\link{iemp}}, \code{\link{innw}}).

\item \code{emp_nps}: Is \code{x} an empty non-positive-valued
numeric object?
(\code{\link{iemp}}, \code{\link{inps}}).

\item \code{emp_npw}: Is \code{x} an empty non-positive-valued
whole-number object?
(\code{\link{iemp}}, \code{\link{inpw}}).

\item \code{emp_nst}: Is \code{x} an empty non-sortable object?
(\code{\link{iemp}}, \code{\link{inst}}).

\item \code{emp_num}: Is \code{x} an empty numeric object?
(\code{\link{iemp}}, \code{\link{inum}}).

\item \code{emp_odd}: Is \code{x} an empty odd-valued
whole-number object?
(\code{\link{iemp}}, \code{\link{iodd}}).

\item \code{emp_ord}: Is \code{x} an empty ordered-factor object?
(\code{\link{iemp}}, \code{\link{iord}}).

\item \code{emp_pct}: Is \code{x} an empty percent-valued
(0-100) numeric object?
(\code{\link{iemp}}, \code{\link{ipct}}).

\item \code{emp_pos}: Is \code{x} an empty positive-valued
numeric object?
(\code{\link{iemp}}, \code{\link{ipos}}).

\item \code{emp_ppn}: Is \code{x} an empty proportion-valued
(0-1) numeric object?
(\code{\link{iemp}}, \code{\link{ippn}}).

\item \code{emp_psw}: Is \code{x} an empty positive-valued
whole-number object?
(\code{\link{iemp}}, \code{\link{ipsw}}).

\item \code{emp_srt}: Is \code{x} an empty sortable object?
(\code{\link{iemp}}, \code{\link{isrt}}).

\item \code{emp_str}: Is \code{x} an empty string object?
(\code{\link{iemp}}, \code{\link{istr}}).

\item \code{emp_uno}: Is \code{x} an empty unordered-factor object?
(\code{\link{iemp}}, \code{\link{iuno}}).

\item \code{emp_whl}: Is \code{x} an empty whole-number object?
(\code{\link{iemp}}, \code{\link{iwhl}}).

\item \code{rcr_atm}: Is \code{x} a recursive atomic object?
(\code{\link{ircr}}, \code{\link{iatm}}).

\item \code{rcr_ch1}: Is \code{x} a recursive onechar object?
(\code{\link{ircr}}, \code{\link{ich1}}).

\item \code{rcr_chr}: Is \code{x} a recursive character object?
(\code{\link{ircr}}, \code{\link{ichr}}).

\item \code{rcr_clr}: Is \code{x} a recursive color object?
(\code{\link{ircr}}, \code{\link{iclr}}).

\item \code{rcr_evn}: Is \code{x} a recursive even-valued
whole-number object?
(\code{\link{ircr}}, \code{\link{ievn}}).

\item \code{rcr_fac}: Is \code{x} a recursive factor object?
(\code{\link{ircr}}, \code{\link{ifac}}).

\item \code{rcr_frc}: Is \code{x} a recursive fractional-valued
numeric object?
(\code{\link{ircr}}, \code{\link{ifrc}}).

\item \code{rcr_ind}: Is \code{x} a recursive indexer object?
(\code{\link{ircr}}, \code{\link{iind}}).

\item \code{rcr_lgl}: Is \code{x} a recursive logical object?
(\code{\link{ircr}}, \code{\link{ilgl}}).

\item \code{rcr_neg}: Is \code{x} a recursive negative-valued
numeric object?
(\code{\link{ircr}}, \code{\link{ineg}}).

\item \code{rcr_ngw}: Is \code{x} a recursive negative-valued
whole-number object?
(\code{\link{ircr}}, \code{\link{ingw}}).

\item \code{rcr_nng}: Is \code{x} a recursive non-negative-valued
numeric object?
(\code{\link{ircr}}, \code{\link{inng}}).

\item \code{rcr_nnw}: Is \code{x} a recursive non-negative-valued
whole-number object?
(\code{\link{ircr}}, \code{\link{innw}}).

\item \code{rcr_nps}: Is \code{x} a recursive non-positive-valued
numeric object?
(\code{\link{ircr}}, \code{\link{inps}}).

\item \code{rcr_npw}: Is \code{x} a recursive non-positive-valued
whole-number object?
(\code{\link{ircr}}, \code{\link{inpw}}).

\item \code{rcr_nst}: Is \code{x} a recursive non-sortable object?
(\code{\link{ircr}}, \code{\link{inst}}).

\item \code{rcr_num}: Is \code{x} a recursive numeric object?
(\code{\link{ircr}}, \code{\link{inum}}).

\item \code{rcr_odd}: Is \code{x} a recursive odd-valued
whole-number object?
(\code{\link{ircr}}, \code{\link{iodd}}).

\item \code{rcr_ord}: Is \code{x} a recursive ordered-factor object?
(\code{\link{ircr}}, \code{\link{iord}}).

\item \code{rcr_pct}: Is \code{x} a recursive percent-valued
(0-100) numeric object?
(\code{\link{ircr}}, \code{\link{ipct}}).

\item \code{rcr_pos}: Is \code{x} a recursive positive-valued
numeric object?
(\code{\link{ircr}}, \code{\link{ipos}}).

\item \code{rcr_ppn}: Is \code{x} a recursive proportion-valued
(0-1) numeric object?
(\code{\link{ircr}}, \code{\link{ippn}}).

\item \code{rcr_psw}: Is \code{x} a recursive positive-valued
whole-number object?
(\code{\link{ircr}}, \code{\link{ipsw}}).

\item \code{rcr_srt}: Is \code{x} a recursive sortable object?
(\code{\link{ircr}}, \code{\link{isrt}}).

\item \code{rcr_str}: Is \code{x} a recursive string object?
(\code{\link{ircr}}, \code{\link{istr}}).

\item \code{rcr_uno}: Is \code{x} a recursive unordered-factor object?
(\code{\link{ircr}}, \code{\link{iuno}}).

\item \code{rcr_whl}: Is \code{x} a recursive whole-number object?
(\code{\link{ircr}}, \code{\link{iwhl}}).
}}

\seealso{
Other props: 
\code{\link{as_mmm_uj}()},
\code{\link{ccc_uj}()},
\code{\link{cmp_ccc_uj}()},
\code{\link{cmp_mmm_ccc_uj}()},
\code{\link{cmp_mmm_uj}()},
\code{\link{comparable_uj}()},
\code{\link{compatible_uj}()},
\code{\link{ddd_uj}()},
\code{\link{eee_uj}()},
\code{\link{fff_uj}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm_uj}()},
\code{\link{meets_uj}()},
\code{\link{mmm_ccc_uj}()},
\code{\link{mmm_uj}()},
\code{\link{sss_uj}()},
\code{\link{ttt_ccc_uj}()},
\code{\link{ttt_uj}()},
\code{\link{xxx_uj}()}
}
\concept{props}
