% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{u2u.}
\alias{u2u.}
\alias{u2u}
\alias{u2cm}
\alias{u2in}
\alias{u2mm}
\alias{u2pt}
\alias{cm2u}
\alias{in2u}
\alias{mm2u}
\alias{pt2u}
\alias{cm2in}
\alias{cm2mm}
\alias{cm2pt}
\alias{in2cm}
\alias{in2mm}
\alias{in2pt}
\alias{mm2cm}
\alias{mm2in}
\alias{mm2pt}
\alias{pt2cm}
\alias{pt2in}
\alias{pt2mm}
\title{Convert units from among plotting units}
\usage{
u2u.()

u2u(x, old, new)

u2cm(x, old)

u2in(x, old)

u2mm(x, old)

u2pt(x, old)

cm2u(x, new)

in2u(x, new)

mm2u(x, new)

pt2u(x, new)

cm2in(x)

cm2mm(x)

cm2pt(x)

in2cm(x)

in2mm(x)

in2pt(x)

mm2cm(x)

mm2in(x)

mm2pt(x)

pt2cm(x)

pt2in(x)

pt2mm(x)
}
\arguments{
\item{x}{\link[atm_num]{Atomic, numeric object}.}

\item{old, new}{\link[cmp_chr_scl]{Complete character scalars} giving old and
new units of length. Valid values are \code{c('cm', 'in', 'mm', 'pt')} for
centimeters, inches, millimeters, and points, respectively.}
}
\value{
An atomic, numeric object.
}
\description{
Convert units from among plotting units
}
\section{Functions}{
\itemize{
\item \code{u2u()}: Convert units from units specifying both original and new
units.

\item \code{u2cm()}: Convert units from to centimeter units.

\item \code{u2in()}: Convert units from to inch units.

\item \code{u2mm()}: Convert units from to millimeter units.

\item \code{u2pt()}: Convert units from to point units.

\item \code{cm2u()}: Convert units from centimeters to some other units.

\item \code{in2u()}: Convert units from inches to some other units.

\item \code{mm2u()}: Convert units from millimeters to some other units.

\item \code{pt2u()}: Convert units from points to some other units.

\item \code{cm2in()}: Convert units from centimeters to inches.

\item \code{cm2mm()}: Convert units from centimeters to millimeters.

\item \code{cm2pt()}: Convert units from centimeters to points.

\item \code{in2cm()}: Convert units from inches to centimeters

\item \code{in2mm()}: Convert units from inches to millimeters.

\item \code{in2pt()}: Convert units from inches to points.

\item \code{mm2cm()}: Convert units from millimeters to centimeters.

\item \code{mm2in()}: Convert units from millimeters to inches.

\item \code{mm2pt()}: Convert units from millimeters to points.

\item \code{pt2cm()}: Convert units from points to centimeters.

\item \code{pt2in()}: Convert units from points to inches.

\item \code{pt2mm()}: Convert units from points to millimeters.

}}
