% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u2u.R
\encoding{UTF-8}
\name{u2u_help}
\alias{u2u_help}
\alias{u2u}
\alias{u2cm}
\alias{u2in}
\alias{u2mm}
\alias{u2pt}
\alias{cm2u}
\alias{in2u}
\alias{mm2u}
\alias{pt2u}
\alias{cm2in}
\alias{cm2mm}
\alias{cm2pt}
\alias{in2cm}
\alias{in2mm}
\alias{in2pt}
\alias{mm2cm}
\alias{mm2in}
\alias{mm2pt}
\alias{pt2cm}
\alias{pt2in}
\alias{pt2mm}
\title{Convert plotting units}
\usage{
u2u_help()

u2u(x, old, new)

u2cm(x, old)

u2in(x, old)

u2mm(x, old)

u2pt(x, old)

cm2u(x, new)

in2u(x, new)

mm2u(x, new)

pt2u(x, new)

cm2in(x)

cm2mm(x)

cm2pt(x)

in2cm(x)

in2mm(x)

in2pt(x)

mm2cm(x)

mm2in(x)

mm2pt(x)

pt2cm(x)

pt2in(x)

pt2mm(x)
}
\arguments{
\item{x}{An \link[=atm_num]{atomic, numeric object}.}

\item{old, new}{\link[=cmp_chr_scl]{Complete character scalars} giving old and new units of distance, respectively: \code{'cm'} for centimeters, \code{'in'} for inches, \code{'mm'} for millimeters, and \code{'pt'} for points.}
}
\value{
An atomic numeric object.
}
\description{
Functions in this family convert among common plotting units. Available unit codes are:
\tabular{ll}{  \code{'cm'}   \tab centimeters    \cr
  \code{'in'}   \tab inches         \cr
  \code{'mm'}   \tab millimeters    \cr
  \code{'pt'}   \tab points (72/inch) }
}
\section{Functions}{
\itemize{
\item \code{u2u()}: Converts \code{x} from \code{old} units to \code{new} units

\item \code{u2cm()}: Converts \code{x} from \code{old} units to centimeters.

\item \code{u2in()}: Converts \code{x} from \code{old} units to inches.

\item \code{u2mm()}: Converts \code{x} from \code{old} units to millimeters.

\item \code{u2pt()}: Converts \code{x} from \code{old} units to points.

\item \code{cm2u()}: Converts \code{x} from centimeters to \code{new} units.

\item \code{in2u()}: Converts \code{x} from inches to \code{new} units.

\item \code{mm2u()}: Converts \code{x} from millimeters to \code{new} units.

\item \code{pt2u()}: Converts \code{x} from points to \code{new} units.

\item \code{cm2in()}: Converts \code{x} from centimeters to inches.

\item \code{cm2mm()}: Converts \code{x} from centimeters to millimeters.

\item \code{cm2pt()}: Converts \code{x} from centimeters to points.

\item \code{in2cm()}: Converts \code{x} from inches to centimeters.

\item \code{in2mm()}: Converts \code{x} from inches to millimeters.

\item \code{in2pt()}: Converts \code{x} from inches to points.

\item \code{mm2cm()}: Converts \code{x} from millimeters to centimeters.

\item \code{mm2in()}: Converts \code{x} from millimeters to inches.

\item \code{mm2pt()}: Converts \code{x} from millimeters to points.

\item \code{pt2cm()}: Converts \code{x} from points to centimeters.

\item \code{pt2in()}: Converts \code{x} from points to inches.

\item \code{pt2mm()}: Converts \code{x} from points to millimeters.

}}
\examples{
u2u(1:3, "cm", "pt")
u2u(1:3, "in", "mm")
u2cm(1:3, "in")
u2in(1:3, "cm")
u2pt(1:3, "cm")
cm2u(1:3, "in")
in2u(1:3, "cm")
pt2u(1:3, "cm")
cm2in(1:3)
cm2mm(1:3)
cm2pt(1:3)
in2cm(1:3)
in2mm(1:3)
in2pt(1:3)
mm2cm(1:3)
mm2in(1:3)
mm2pt(1:3)
pt2cm(1:3)
pt2in(1:3)
pt2mm(1:3)
}
\seealso{
Other conversions: 
\code{\link{hex_dec_help}()},
\code{\link{rng2lev}()},
\code{\link{tocase}()}

Other plots: 
\code{\link{color_funs}()},
\code{\link{gg_help}()},
\code{\link{gr}},
\code{\link{markdown_help}()},
\code{\link{pals_help}()}
}
\concept{conversions}
\concept{plots}
