% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{uj_constants.}
\alias{uj_constants.}
\alias{uj_constants}
\alias{v}
\title{Package constants for \code{uj}}
\usage{
uj_constants.()

uj_constants()

v(...)
}
\arguments{
\item{...}{unquoted, comma-separated list of the names of constants to
return. If multiple constants are specified, they are coerced into a single
atomic vector result.}
}
\value{
An atomic vector of length 1 or greater.
}
\description{
Package constants for \code{uj}
}
\section{Functions}{
\itemize{
\item \code{uj_constants()}: Get a named \link[ivls]{atomic vlist} containing
all package \code{uj} constants with an associated attribute describing
each.

\item \code{v()}: Get one or more \code{uj} package constants.

}}
