% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\docType{data}
\encoding{UTF-8}
\name{ujPkgConstants}
\alias{ujPkgConstants}
\alias{.PkgPropFamilies}
\alias{.PkgAllPropVals}
\alias{.PkgCrayonSynonyms}
\alias{.PkgCrayonFamilies}
\alias{.BoldGreekVals}
\alias{.BoldGreekDefs}
\alias{.BoldGREEKvals}
\alias{.BoldGREEKdefs}
\alias{.ItalicGreekVals}
\alias{.ItalicGreekDefs}
\alias{.ItalicGREEKvals}
\alias{.ItalicGREEKdefs}
\alias{.PlainGreekVals}
\alias{.PlainGreekDefs}
\alias{.PlainGREEKvals}
\alias{.PlainGREEKdefs}
\alias{.EnclosureVals}
\alias{.EnclosureDefs}
\alias{.BasicMarkdownVals}
\alias{.BasicMarkdownDefs}
\alias{.OpenMarkdownColorVals}
\alias{.OpenMarkdownColorDefs}
\alias{.MiscVals}
\alias{.MiscDefs}
\alias{.UTF8vals}
\alias{.UTF8defs}
\alias{.PkgColorVals}
\alias{.PkgColorDefs}
\alias{.PkgLinetypeVals}
\alias{.PkgLinetypeDefs}
\alias{.PkgShapeVals}
\alias{.PkgShapeDefs}
\alias{.PkgCharSetVals}
\alias{.PkgCharSetDefs}
\alias{.PkgPropVals}
\alias{.PkgPropDefs}
\alias{.PkgCrayonVals}
\alias{.PkgCrayonDefs}
\alias{.PkgVals}
\alias{.PkgDefs}
\title{Package \code{uj} atomic scalar and vector constants}
\format{
An object of class \code{list} of length 14.

An object of class \code{character} of length 132.

An object of class \code{list} of length 14.

An object of class \code{list} of length 3.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{list} of length 24.

An object of class \code{character} of length 24.

An object of class \code{character} of length 21.

An object of class \code{character} of length 21.

An object of class \code{list} of length 9.

An object of class \code{character} of length 9.

An object of class \code{list} of length 150.

An object of class \code{character} of length 150.

An object of class \code{list} of length 18.

An object of class \code{character} of length 18.

An object of class \code{list} of length 71.

An object of class \code{character} of length 73.

An object of class \code{list} of length 52.

An object of class \code{character} of length 52.

An object of class \code{list} of length 1.

An object of class \code{character} of length 1.

An object of class \code{list} of length 10.

An object of class \code{character} of length 10.

An object of class \code{list} of length 14.

An object of class \code{character} of length 14.

An object of class \code{list} of length 17.

An object of class \code{character} of length 17.

An object of class \code{list} of length 17.

An object of class \code{character} of length 17.

An object of class \code{list} of length 524.

An object of class \code{character} of length 526.
}
\usage{
ujPkgConstants()

.PkgPropFamilies

.PkgAllPropVals

.PkgCrayonSynonyms

.PkgCrayonFamilies

.BoldGreekVals

.BoldGreekDefs

.BoldGREEKvals

.BoldGREEKdefs

.ItalicGreekVals

.ItalicGreekDefs

.ItalicGREEKvals

.ItalicGREEKdefs

.PlainGreekVals

.PlainGreekDefs

.PlainGREEKvals

.PlainGREEKdefs

.EnclosureVals

.EnclosureDefs

.BasicMarkdownVals

.BasicMarkdownDefs

.OpenMarkdownColorVals

.OpenMarkdownColorDefs

.MiscVals

.MiscDefs

.UTF8vals

.UTF8defs

.PkgColorVals

.PkgColorDefs

.PkgLinetypeVals

.PkgLinetypeDefs

.PkgShapeVals

.PkgShapeDefs

.PkgCharSetVals

.PkgCharSetDefs

.PkgPropVals

.PkgPropDefs

.PkgCrayonVals

.PkgCrayonDefs

.PkgVals

.PkgDefs
}
\value{
An object of class \code{list} containing atomic scalar or atomic vector elements (\code{.xxxVals} objects) or a character vector describing the constants (\code{.xxxDefs} objects).
}
\description{
Constants defined by package \code{uj} are provided in a number of package values with associated descriptions. Constants are organized into the following types:
\tabular{ll}{        Properties   \tab Object properties defined by \code{\link{ppp}()} functions;
\cr   \tab   \cr   Crayons      \tab Synonyms for \code{\link[crayon:red]{crayon::colors}} and \code{\link[crayon:styles]{crayon::styles}} used by \code{\link[uj:st_vals]{uj::crayons}} functions;
\cr   \tab   \cr   Markdown     \tab Basic, subscript, superscript, and bold, plain, and italic text color open tags;
\cr   \tab   \cr   Colors       \tab Named color codes and color palettes;
\cr   \tab   \cr   Linetypes    \tab Named line types (extending those available by name in base R), and
\cr   \tab   \cr   Markers      \tab ASCII codes for plotting characters and base R symbols as value markers rather than as text including:
\cr   \tab \itemize{
\item Lowercase letters,
\item Uppercase letters,
\item Digits (including 0),
\item Digits (excluding 0),
\item Distinctive non-enclosing shapes,
\item Distinctive punctuation characters,
\item Base R empty/non-fillable shapes,
\item Base R fillable shapes,
\item Base R solid shapes, and
\item A blank string shape.
}
\cr   Charsets    \tab Characters sets including:
\cr   \tab \itemize{
\item Plain-font UTF-8 upper and lower Greek in markdown-compatible configuration
\item Plain-font UTF-8 upper and lower Greek in markdown-compatible configuration
\item Escaped UTF-8 codes for a variety of special or functional characters.
\item Characters used for enclosing text (e.g., quotes, brackets);
\item Lowercase and/or uppercase letters,
\item Digits as characters,
\item Characters valid for file names,
\item Upper and lowercase vowels,
\item Characters for plotting labels without special formatting,
\item Characters valid for R model or formulas,
\item Characters valid for the right side of R models or formulas,
\item Characters valid for R object names without back ticks,
\item A restricted set for file names,
\item A restricted set for R models and formulas,
\item A restricted set for the right side of R models and formulas, and
\item A restricted set for R object names.
}}
}
\keyword{datasets}
