% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\encoding{UTF-8}
\name{uniqueness}
\alias{uniqueness}
\alias{is_unique}
\alias{is_unique0}
\alias{not_unique}
\alias{not_unique0}
\alias{is_unq}
\alias{is_unq0}
\alias{not_unq}
\alias{not_unq0}
\title{Does an atomic object only contain unique values?}
\usage{
uniqueness()

is_unique(x, a = FALSE, na = FALSE)

is_unique0(...)

not_unique(x, a = FALSE, na = FALSE)

not_unique0(...)

is_unq(x, a = FALSE, na = FALSE)

is_unq0(...)

not_unq(x, a = FALSE, na = FALSE)

not_unq0(...)
}
\arguments{
\item{x}{An atomic object.}

\item{a}{\code{TRUE} or \code{FALSE} indicating whether to reduce \code{x} to an atomic vector containing all of its atomic values. When \code{FALSE} and \code{x} is not atomic, throws an error.}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values are allowed.}
}
\value{
A logical scalar.
}
\description{
Evaluates whether \code{x} contains only only unique atomic values, with options for for atomizing before evaluation and for handling \code{NA} values.
}
\details{
\tabular{ll}{  \verb{not_unique0, not_unq0}   \tab whether \code{x} \emph{is not} an atomic vector containing any duplicate value.                            \cr   \tab   \cr
  \verb{not_unique, not_unq}     \tab whether \code{x} \emph{is not} a unique data.frame, list, or vector depending on values of \code{a} and \code{na}. \cr   \tab   \cr
  \verb{is_unique0, is_unq0}     \tab whether \code{x} \emph{is} an atomic vector containing only unique values.                                 \cr   \tab   \cr
  \verb{is_unique, is_unq}       \tab whether \code{x} \emph{is} a unique data.frame, list, or vector depending on values of \code{a} and \code{na}.                    }
}
\section{Functions}{
\itemize{
\item \code{is_unique()}: Evaluate whether \code{x} \emph{is} a unique data.frame, list, or vector depending on values of \code{a} and \code{na}.

\item \code{is_unique0()}: Evaluates whether \code{x} \emph{is} an atomic vector containing only unique values.

\item \code{not_unique()}: Evaluates whether \code{x} \emph{is not} a unique data.frame, list, or vector depending on values of \code{a} and \code{na}.

\item \code{not_unique0()}: Evaluates whether \code{x} \emph{is not} an atomic vector containing any duplicate value.

\item \code{is_unq()}: An alias for \code{is_unique}

\item \code{is_unq0()}: An alias for \code{is_unique0}

\item \code{not_unq()}: An alias for \code{not_unique}

\item \code{not_unq0()}: An alias for \code{not_unique0}

}}
\examples{
is_unique(letters)
is_unq(sample(letters, 27, replace = T))
}
