% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_ccc.R
\encoding{UTF-8}
\name{unq_ccc}
\alias{unq_ccc}
\alias{unq_ccc_funs}
\alias{unq_arr}
\alias{unq_dtf}
\alias{unq_gen}
\alias{unq_mat}
\alias{unq_mvc}
\alias{unq_scl}
\alias{unq_vec}
\alias{unq_vls}
\title{unique + xclass combination properties}
\usage{
unq_ccc(x, ccc, ...)

unq_ccc_funs()

unq_arr(x, ...)

unq_dtf(x, ...)

unq_gen(x, ...)

unq_mat(x, ...)

unq_mvc(x, ...)

unq_scl(x, ...)

unq_vec(x, ...)

unq_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{ccc}{A character scalar single xclass property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector}
\cr    \code{unq_ccc_funs}
\cr
\cr \emph{A logical scalar}
\cr    \code{unq_ccc}
\cr    \code{unq_CCC}
}
\description{
\tabular{rl}{
  \code{unq_ccc_funs}   \tab What \link[=iunq]{unique} + \link[=ccc]{xclass} combination \link[=prop_funs]{property functions} are there?
\cr                \tab  
\cr    \code{unq_ccc}   \tab Is \code{x} both unique and a match to the single xclass property in argument \code{ccc}?
\cr                \tab  
\cr    \code{unq_CCC}   \tab Is \code{x} both unique and a match to the single xclass property \code{'CCC'}?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab  
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab  
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
unq_ccc_fun()
unq_ccc(letters, "mvc")
unq_ccc(1, "scl")
unq_ccc(NA, "gen")
unq_mvc(letters)
unq_scl(1)
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unq}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{reclass}()},
\code{\link{sss}()}
}
\concept{properties}
