% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_mmm_ccc.R
\encoding{UTF-8}
\name{unq_mmm_ccc}
\alias{unq_mmm_ccc}
\alias{unq_mmm_ccc_funs}
\alias{unq_atm_arr}
\alias{unq_atm_dtf}
\alias{unq_atm_gen}
\alias{unq_atm_mat}
\alias{unq_atm_mvc}
\alias{unq_atm_scl}
\alias{unq_atm_vec}
\alias{unq_atm_vls}
\alias{unq_ch1_arr}
\alias{unq_ch1_dtf}
\alias{unq_ch1_gen}
\alias{unq_ch1_mat}
\alias{unq_ch1_mvc}
\alias{unq_ch1_scl}
\alias{unq_ch1_vec}
\alias{unq_ch1_vls}
\alias{unq_ch3_arr}
\alias{unq_ch3_dtf}
\alias{unq_ch3_gen}
\alias{unq_ch3_mat}
\alias{unq_ch3_mvc}
\alias{unq_ch3_scl}
\alias{unq_ch3_vec}
\alias{unq_ch3_vls}
\alias{unq_chr_arr}
\alias{unq_chr_dtf}
\alias{unq_chr_gen}
\alias{unq_chr_mat}
\alias{unq_chr_mvc}
\alias{unq_chr_scl}
\alias{unq_chr_vec}
\alias{unq_chr_vls}
\alias{unq_clr_arr}
\alias{unq_clr_dtf}
\alias{unq_clr_gen}
\alias{unq_clr_mat}
\alias{unq_clr_mvc}
\alias{unq_clr_scl}
\alias{unq_clr_vec}
\alias{unq_clr_vls}
\alias{unq_evn_arr}
\alias{unq_evn_dtf}
\alias{unq_evn_gen}
\alias{unq_evn_mat}
\alias{unq_evn_mvc}
\alias{unq_evn_scl}
\alias{unq_evn_vec}
\alias{unq_evn_vls}
\alias{unq_fac_arr}
\alias{unq_fac_dtf}
\alias{unq_fac_gen}
\alias{unq_fac_mat}
\alias{unq_fac_mvc}
\alias{unq_fac_scl}
\alias{unq_fac_vec}
\alias{unq_fac_vls}
\alias{unq_frc_arr}
\alias{unq_frc_dtf}
\alias{unq_frc_gen}
\alias{unq_frc_mat}
\alias{unq_frc_mvc}
\alias{unq_frc_scl}
\alias{unq_frc_vec}
\alias{unq_frc_vls}
\alias{unq_ind_arr}
\alias{unq_ind_dtf}
\alias{unq_ind_gen}
\alias{unq_ind_mat}
\alias{unq_ind_mvc}
\alias{unq_ind_scl}
\alias{unq_ind_vec}
\alias{unq_ind_vls}
\alias{unq_lgl_arr}
\alias{unq_lgl_dtf}
\alias{unq_lgl_gen}
\alias{unq_lgl_mat}
\alias{unq_lgl_mvc}
\alias{unq_lgl_scl}
\alias{unq_lgl_vec}
\alias{unq_lgl_vls}
\alias{unq_neg_arr}
\alias{unq_neg_dtf}
\alias{unq_neg_gen}
\alias{unq_neg_mat}
\alias{unq_neg_mvc}
\alias{unq_neg_scl}
\alias{unq_neg_vec}
\alias{unq_neg_vls}
\alias{unq_ngw_arr}
\alias{unq_ngw_dtf}
\alias{unq_ngw_gen}
\alias{unq_ngw_mat}
\alias{unq_ngw_mvc}
\alias{unq_ngw_scl}
\alias{unq_ngw_vec}
\alias{unq_ngw_vls}
\alias{unq_nng_arr}
\alias{unq_nng_dtf}
\alias{unq_nng_gen}
\alias{unq_nng_mat}
\alias{unq_nng_mvc}
\alias{unq_nng_scl}
\alias{unq_nng_vec}
\alias{unq_nng_vls}
\alias{unq_nnw_arr}
\alias{unq_nnw_dtf}
\alias{unq_nnw_gen}
\alias{unq_nnw_mat}
\alias{unq_nnw_mvc}
\alias{unq_nnw_scl}
\alias{unq_nnw_vec}
\alias{unq_nnw_vls}
\alias{unq_nps_arr}
\alias{unq_nps_dtf}
\alias{unq_nps_gen}
\alias{unq_nps_mat}
\alias{unq_nps_mvc}
\alias{unq_nps_scl}
\alias{unq_nps_vec}
\alias{unq_nps_vls}
\alias{unq_npw_arr}
\alias{unq_npw_dtf}
\alias{unq_npw_gen}
\alias{unq_npw_mat}
\alias{unq_npw_mvc}
\alias{unq_npw_scl}
\alias{unq_npw_vec}
\alias{unq_npw_vls}
\alias{unq_nst_arr}
\alias{unq_nst_dtf}
\alias{unq_nst_gen}
\alias{unq_nst_mat}
\alias{unq_nst_mvc}
\alias{unq_nst_scl}
\alias{unq_nst_vec}
\alias{unq_nst_vls}
\alias{unq_num_arr}
\alias{unq_num_dtf}
\alias{unq_num_gen}
\alias{unq_num_mat}
\alias{unq_num_mvc}
\alias{unq_num_scl}
\alias{unq_num_vec}
\alias{unq_num_vls}
\alias{unq_odd_arr}
\alias{unq_odd_dtf}
\alias{unq_odd_gen}
\alias{unq_odd_mat}
\alias{unq_odd_mvc}
\alias{unq_odd_scl}
\alias{unq_odd_vec}
\alias{unq_odd_vls}
\alias{unq_ord_arr}
\alias{unq_ord_dtf}
\alias{unq_ord_gen}
\alias{unq_ord_mat}
\alias{unq_ord_mvc}
\alias{unq_ord_scl}
\alias{unq_ord_vec}
\alias{unq_ord_vls}
\alias{unq_pct_arr}
\alias{unq_pct_dtf}
\alias{unq_pct_gen}
\alias{unq_pct_mat}
\alias{unq_pct_mvc}
\alias{unq_pct_scl}
\alias{unq_pct_vec}
\alias{unq_pct_vls}
\alias{unq_pos_arr}
\alias{unq_pos_dtf}
\alias{unq_pos_gen}
\alias{unq_pos_mat}
\alias{unq_pos_mvc}
\alias{unq_pos_scl}
\alias{unq_pos_vec}
\alias{unq_pos_vls}
\alias{unq_ppn_arr}
\alias{unq_ppn_dtf}
\alias{unq_ppn_gen}
\alias{unq_ppn_mat}
\alias{unq_ppn_mvc}
\alias{unq_ppn_scl}
\alias{unq_ppn_vec}
\alias{unq_ppn_vls}
\alias{unq_psw_arr}
\alias{unq_psw_dtf}
\alias{unq_psw_gen}
\alias{unq_psw_mat}
\alias{unq_psw_mvc}
\alias{unq_psw_scl}
\alias{unq_psw_vec}
\alias{unq_psw_vls}
\alias{unq_srt_arr}
\alias{unq_srt_dtf}
\alias{unq_srt_gen}
\alias{unq_srt_mat}
\alias{unq_srt_mvc}
\alias{unq_srt_scl}
\alias{unq_srt_vec}
\alias{unq_srt_vls}
\alias{unq_str_arr}
\alias{unq_str_dtf}
\alias{unq_str_gen}
\alias{unq_str_mat}
\alias{unq_str_mvc}
\alias{unq_str_scl}
\alias{unq_str_vec}
\alias{unq_str_vls}
\alias{unq_uno_arr}
\alias{unq_uno_dtf}
\alias{unq_uno_gen}
\alias{unq_uno_mat}
\alias{unq_uno_mvc}
\alias{unq_uno_scl}
\alias{unq_uno_vec}
\alias{unq_uno_vls}
\alias{unq_whl_arr}
\alias{unq_whl_dtf}
\alias{unq_whl_gen}
\alias{unq_whl_mat}
\alias{unq_whl_mvc}
\alias{unq_whl_scl}
\alias{unq_whl_vec}
\alias{unq_whl_vls}
\title{unique + xmode + xclass combination properties}
\usage{
unq_mmm_ccc(X, MMM, CCC, ...)

unq_mmm_ccc_funs()

unq_atm_arr(X, ...)

unq_atm_dtf(X, ...)

unq_atm_gen(X, ...)

unq_atm_mat(X, ...)

unq_atm_mvc(X, ...)

unq_atm_scl(X, ...)

unq_atm_vec(X, ...)

unq_atm_vls(X, ...)

unq_ch1_arr(X, ...)

unq_ch1_dtf(X, ...)

unq_ch1_gen(X, ...)

unq_ch1_mat(X, ...)

unq_ch1_mvc(X, ...)

unq_ch1_scl(X, ...)

unq_ch1_vec(X, ...)

unq_ch1_vls(X, ...)

unq_ch3_arr(X, ...)

unq_ch3_dtf(X, ...)

unq_ch3_gen(X, ...)

unq_ch3_mat(X, ...)

unq_ch3_mvc(X, ...)

unq_ch3_scl(X, ...)

unq_ch3_vec(X, ...)

unq_ch3_vls(X, ...)

unq_chr_arr(X, ...)

unq_chr_dtf(X, ...)

unq_chr_gen(X, ...)

unq_chr_mat(X, ...)

unq_chr_mvc(X, ...)

unq_chr_scl(X, ...)

unq_chr_vec(X, ...)

unq_chr_vls(X, ...)

unq_clr_arr(X, ...)

unq_clr_dtf(X, ...)

unq_clr_gen(X, ...)

unq_clr_mat(X, ...)

unq_clr_mvc(X, ...)

unq_clr_scl(X, ...)

unq_clr_vec(X, ...)

unq_clr_vls(X, ...)

unq_evn_arr(X, ...)

unq_evn_dtf(X, ...)

unq_evn_gen(X, ...)

unq_evn_mat(X, ...)

unq_evn_mvc(X, ...)

unq_evn_scl(X, ...)

unq_evn_vec(X, ...)

unq_evn_vls(X, ...)

unq_fac_arr(X, ...)

unq_fac_dtf(X, ...)

unq_fac_gen(X, ...)

unq_fac_mat(X, ...)

unq_fac_mvc(X, ...)

unq_fac_scl(X, ...)

unq_fac_vec(X, ...)

unq_fac_vls(X, ...)

unq_frc_arr(X, ...)

unq_frc_dtf(X, ...)

unq_frc_gen(X, ...)

unq_frc_mat(X, ...)

unq_frc_mvc(X, ...)

unq_frc_scl(X, ...)

unq_frc_vec(X, ...)

unq_frc_vls(X, ...)

unq_ind_arr(X, ...)

unq_ind_dtf(X, ...)

unq_ind_gen(X, ...)

unq_ind_mat(X, ...)

unq_ind_mvc(X, ...)

unq_ind_scl(X, ...)

unq_ind_vec(X, ...)

unq_ind_vls(X, ...)

unq_lgl_arr(X, ...)

unq_lgl_dtf(X, ...)

unq_lgl_gen(X, ...)

unq_lgl_mat(X, ...)

unq_lgl_mvc(X, ...)

unq_lgl_scl(X, ...)

unq_lgl_vec(X, ...)

unq_lgl_vls(X, ...)

unq_neg_arr(X, ...)

unq_neg_dtf(X, ...)

unq_neg_gen(X, ...)

unq_neg_mat(X, ...)

unq_neg_mvc(X, ...)

unq_neg_scl(X, ...)

unq_neg_vec(X, ...)

unq_neg_vls(X, ...)

unq_ngw_arr(X, ...)

unq_ngw_dtf(X, ...)

unq_ngw_gen(X, ...)

unq_ngw_mat(X, ...)

unq_ngw_mvc(X, ...)

unq_ngw_scl(X, ...)

unq_ngw_vec(X, ...)

unq_ngw_vls(X, ...)

unq_nng_arr(X, ...)

unq_nng_dtf(X, ...)

unq_nng_gen(X, ...)

unq_nng_mat(X, ...)

unq_nng_mvc(X, ...)

unq_nng_scl(X, ...)

unq_nng_vec(X, ...)

unq_nng_vls(X, ...)

unq_nnw_arr(X, ...)

unq_nnw_dtf(X, ...)

unq_nnw_gen(X, ...)

unq_nnw_mat(X, ...)

unq_nnw_mvc(X, ...)

unq_nnw_scl(X, ...)

unq_nnw_vec(X, ...)

unq_nnw_vls(X, ...)

unq_nps_arr(X, ...)

unq_nps_dtf(X, ...)

unq_nps_gen(X, ...)

unq_nps_mat(X, ...)

unq_nps_mvc(X, ...)

unq_nps_scl(X, ...)

unq_nps_vec(X, ...)

unq_nps_vls(X, ...)

unq_npw_arr(X, ...)

unq_npw_dtf(X, ...)

unq_npw_gen(X, ...)

unq_npw_mat(X, ...)

unq_npw_mvc(X, ...)

unq_npw_scl(X, ...)

unq_npw_vec(X, ...)

unq_npw_vls(X, ...)

unq_nst_arr(X, ...)

unq_nst_dtf(X, ...)

unq_nst_gen(X, ...)

unq_nst_mat(X, ...)

unq_nst_mvc(X, ...)

unq_nst_scl(X, ...)

unq_nst_vec(X, ...)

unq_nst_vls(X, ...)

unq_num_arr(X, ...)

unq_num_dtf(X, ...)

unq_num_gen(X, ...)

unq_num_mat(X, ...)

unq_num_mvc(X, ...)

unq_num_scl(X, ...)

unq_num_vec(X, ...)

unq_num_vls(X, ...)

unq_odd_arr(X, ...)

unq_odd_dtf(X, ...)

unq_odd_gen(X, ...)

unq_odd_mat(X, ...)

unq_odd_mvc(X, ...)

unq_odd_scl(X, ...)

unq_odd_vec(X, ...)

unq_odd_vls(X, ...)

unq_ord_arr(X, ...)

unq_ord_dtf(X, ...)

unq_ord_gen(X, ...)

unq_ord_mat(X, ...)

unq_ord_mvc(X, ...)

unq_ord_scl(X, ...)

unq_ord_vec(X, ...)

unq_ord_vls(X, ...)

unq_pct_arr(X, ...)

unq_pct_dtf(X, ...)

unq_pct_gen(X, ...)

unq_pct_mat(X, ...)

unq_pct_mvc(X, ...)

unq_pct_scl(X, ...)

unq_pct_vec(X, ...)

unq_pct_vls(X, ...)

unq_pos_arr(X, ...)

unq_pos_dtf(X, ...)

unq_pos_gen(X, ...)

unq_pos_mat(X, ...)

unq_pos_mvc(X, ...)

unq_pos_scl(X, ...)

unq_pos_vec(X, ...)

unq_pos_vls(X, ...)

unq_ppn_arr(X, ...)

unq_ppn_dtf(X, ...)

unq_ppn_gen(X, ...)

unq_ppn_mat(X, ...)

unq_ppn_mvc(X, ...)

unq_ppn_scl(X, ...)

unq_ppn_vec(X, ...)

unq_ppn_vls(X, ...)

unq_psw_arr(X, ...)

unq_psw_dtf(X, ...)

unq_psw_gen(X, ...)

unq_psw_mat(X, ...)

unq_psw_mvc(X, ...)

unq_psw_scl(X, ...)

unq_psw_vec(X, ...)

unq_psw_vls(X, ...)

unq_srt_arr(X, ...)

unq_srt_dtf(X, ...)

unq_srt_gen(X, ...)

unq_srt_mat(X, ...)

unq_srt_mvc(X, ...)

unq_srt_scl(X, ...)

unq_srt_vec(X, ...)

unq_srt_vls(X, ...)

unq_str_arr(X, ...)

unq_str_dtf(X, ...)

unq_str_gen(X, ...)

unq_str_mat(X, ...)

unq_str_mvc(X, ...)

unq_str_scl(X, ...)

unq_str_vec(X, ...)

unq_str_vls(X, ...)

unq_uno_arr(X, ...)

unq_uno_dtf(X, ...)

unq_uno_gen(X, ...)

unq_uno_mat(X, ...)

unq_uno_mvc(X, ...)

unq_uno_scl(X, ...)

unq_uno_vec(X, ...)

unq_uno_vls(X, ...)

unq_whl_arr(X, ...)

unq_whl_dtf(X, ...)

unq_whl_gen(X, ...)

unq_whl_mat(X, ...)

unq_whl_mvc(X, ...)

unq_whl_scl(X, ...)

unq_whl_vec(X, ...)

unq_whl_vls(X, ...)
}
\arguments{
\item{X}{An R object.}

\item{MMM}{A character scalar single xmode property from \code{\link{mmm_props}()}.}

\item{CCC}{A character scalar single xclass property from \code{\link{ccc_props}()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \code{unq_mmm_ccc_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{unq_\{mmm\}_\{ccc\}} \cr \code{unq_mmm_ccc}
}
\description{
Functions for checking combinations of \link[=UNQ]{uniqueness}, \link[=mmm]{xmode}, and \link[=ccc]{xclass}.
}
\details{
\tabular{ll}{  \code{unq_mmm_ccc_funs}   \tab What \link[=iunq]{unique} + \link[=mmm]{xmode} + \link[=ccc]{xclass} combination \link[=prop_funs]{property functions} are there? \cr   \tab   \cr
  \verb{unq_\{mmm\}_\{ccc\}}    \tab Is \code{X} both unique and a match to the single xmode property \code{'{mmm}'} and single xclass property \code{'{ccc}'}? (where \code{{mmm}} and
\code{{ccc}} are placeholders for any given xmode property and any given xclass property, respectively).                               \cr   \tab   \cr
  \code{unq_mmm_ccc}        \tab Is \code{X} both unique and a match to the single xmode and xclass properties in arguments \code{MMM} and \code{CCC}, respectively?                             }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.max, .maxr, .maxc}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
unq_mmm_ccc_funs()
unq_mmm_ccc(letters, "ch1", "vec")
unq_mmm_ccc(letters, "str", "scl")
unq_ch1_vec(letters)
unq_str_scl("a")
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{sss_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
