% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_mmm_ccc_prop_funs.R
\encoding{UTF-8}
\name{unq_mmm_ccc_prop_funs}
\alias{unq_mmm_ccc_prop_funs}
\alias{unq_mmm_ccc}
\alias{unq_mmm_ccc_funs}
\alias{unq_mmm_ccc_props}
\alias{unq_atm_arr}
\alias{unq_atm_dtf}
\alias{unq_atm_gen}
\alias{unq_atm_mat}
\alias{unq_atm_mvc}
\alias{unq_atm_scl}
\alias{unq_atm_vec}
\alias{unq_atm_vls}
\alias{unq_ch1_arr}
\alias{unq_ch1_dtf}
\alias{unq_ch1_gen}
\alias{unq_ch1_mat}
\alias{unq_ch1_mvc}
\alias{unq_ch1_scl}
\alias{unq_ch1_vec}
\alias{unq_ch1_vls}
\alias{unq_ch3_arr}
\alias{unq_ch3_dtf}
\alias{unq_ch3_gen}
\alias{unq_ch3_mat}
\alias{unq_ch3_mvc}
\alias{unq_ch3_scl}
\alias{unq_ch3_vec}
\alias{unq_ch3_vls}
\alias{unq_chr_arr}
\alias{unq_chr_dtf}
\alias{unq_chr_gen}
\alias{unq_chr_mat}
\alias{unq_chr_mvc}
\alias{unq_chr_scl}
\alias{unq_chr_vec}
\alias{unq_chr_vls}
\alias{unq_clr_arr}
\alias{unq_clr_dtf}
\alias{unq_clr_gen}
\alias{unq_clr_mat}
\alias{unq_clr_mvc}
\alias{unq_clr_scl}
\alias{unq_clr_vec}
\alias{unq_clr_vls}
\alias{unq_evn_arr}
\alias{unq_evn_dtf}
\alias{unq_evn_gen}
\alias{unq_evn_mat}
\alias{unq_evn_mvc}
\alias{unq_evn_scl}
\alias{unq_evn_vec}
\alias{unq_evn_vls}
\alias{unq_fac_arr}
\alias{unq_fac_dtf}
\alias{unq_fac_gen}
\alias{unq_fac_mat}
\alias{unq_fac_mvc}
\alias{unq_fac_scl}
\alias{unq_fac_vec}
\alias{unq_fac_vls}
\alias{unq_frc_arr}
\alias{unq_frc_dtf}
\alias{unq_frc_gen}
\alias{unq_frc_mat}
\alias{unq_frc_mvc}
\alias{unq_frc_scl}
\alias{unq_frc_vec}
\alias{unq_frc_vls}
\alias{unq_ind_arr}
\alias{unq_ind_dtf}
\alias{unq_ind_gen}
\alias{unq_ind_mat}
\alias{unq_ind_mvc}
\alias{unq_ind_scl}
\alias{unq_ind_vec}
\alias{unq_ind_vls}
\alias{unq_lgl_arr}
\alias{unq_lgl_dtf}
\alias{unq_lgl_gen}
\alias{unq_lgl_mat}
\alias{unq_lgl_mvc}
\alias{unq_lgl_scl}
\alias{unq_lgl_vec}
\alias{unq_lgl_vls}
\alias{unq_neg_arr}
\alias{unq_neg_dtf}
\alias{unq_neg_gen}
\alias{unq_neg_mat}
\alias{unq_neg_mvc}
\alias{unq_neg_scl}
\alias{unq_neg_vec}
\alias{unq_neg_vls}
\alias{unq_ngw_arr}
\alias{unq_ngw_dtf}
\alias{unq_ngw_gen}
\alias{unq_ngw_mat}
\alias{unq_ngw_mvc}
\alias{unq_ngw_scl}
\alias{unq_ngw_vec}
\alias{unq_ngw_vls}
\alias{unq_nng_arr}
\alias{unq_nng_dtf}
\alias{unq_nng_gen}
\alias{unq_nng_mat}
\alias{unq_nng_mvc}
\alias{unq_nng_scl}
\alias{unq_nng_vec}
\alias{unq_nng_vls}
\alias{unq_nnw_arr}
\alias{unq_nnw_dtf}
\alias{unq_nnw_gen}
\alias{unq_nnw_mat}
\alias{unq_nnw_mvc}
\alias{unq_nnw_scl}
\alias{unq_nnw_vec}
\alias{unq_nnw_vls}
\alias{unq_nps_arr}
\alias{unq_nps_dtf}
\alias{unq_nps_gen}
\alias{unq_nps_mat}
\alias{unq_nps_mvc}
\alias{unq_nps_scl}
\alias{unq_nps_vec}
\alias{unq_nps_vls}
\alias{unq_npw_arr}
\alias{unq_npw_dtf}
\alias{unq_npw_gen}
\alias{unq_npw_mat}
\alias{unq_npw_mvc}
\alias{unq_npw_scl}
\alias{unq_npw_vec}
\alias{unq_npw_vls}
\alias{unq_nst_arr}
\alias{unq_nst_dtf}
\alias{unq_nst_gen}
\alias{unq_nst_mat}
\alias{unq_nst_mvc}
\alias{unq_nst_scl}
\alias{unq_nst_vec}
\alias{unq_nst_vls}
\alias{unq_num_arr}
\alias{unq_num_dtf}
\alias{unq_num_gen}
\alias{unq_num_mat}
\alias{unq_num_mvc}
\alias{unq_num_scl}
\alias{unq_num_vec}
\alias{unq_num_vls}
\alias{unq_odd_arr}
\alias{unq_odd_dtf}
\alias{unq_odd_gen}
\alias{unq_odd_mat}
\alias{unq_odd_mvc}
\alias{unq_odd_scl}
\alias{unq_odd_vec}
\alias{unq_odd_vls}
\alias{unq_ord_arr}
\alias{unq_ord_dtf}
\alias{unq_ord_gen}
\alias{unq_ord_mat}
\alias{unq_ord_mvc}
\alias{unq_ord_scl}
\alias{unq_ord_vec}
\alias{unq_ord_vls}
\alias{unq_pct_arr}
\alias{unq_pct_dtf}
\alias{unq_pct_gen}
\alias{unq_pct_mat}
\alias{unq_pct_mvc}
\alias{unq_pct_scl}
\alias{unq_pct_vec}
\alias{unq_pct_vls}
\alias{unq_pos_arr}
\alias{unq_pos_dtf}
\alias{unq_pos_gen}
\alias{unq_pos_mat}
\alias{unq_pos_mvc}
\alias{unq_pos_scl}
\alias{unq_pos_vec}
\alias{unq_pos_vls}
\alias{unq_ppn_arr}
\alias{unq_ppn_dtf}
\alias{unq_ppn_gen}
\alias{unq_ppn_mat}
\alias{unq_ppn_mvc}
\alias{unq_ppn_scl}
\alias{unq_ppn_vec}
\alias{unq_ppn_vls}
\alias{unq_psw_arr}
\alias{unq_psw_dtf}
\alias{unq_psw_gen}
\alias{unq_psw_mat}
\alias{unq_psw_mvc}
\alias{unq_psw_scl}
\alias{unq_psw_vec}
\alias{unq_psw_vls}
\alias{unq_srt_arr}
\alias{unq_srt_dtf}
\alias{unq_srt_gen}
\alias{unq_srt_mat}
\alias{unq_srt_mvc}
\alias{unq_srt_scl}
\alias{unq_srt_vec}
\alias{unq_srt_vls}
\alias{unq_str_arr}
\alias{unq_str_dtf}
\alias{unq_str_gen}
\alias{unq_str_mat}
\alias{unq_str_mvc}
\alias{unq_str_scl}
\alias{unq_str_vec}
\alias{unq_str_vls}
\alias{unq_uno_arr}
\alias{unq_uno_dtf}
\alias{unq_uno_gen}
\alias{unq_uno_mat}
\alias{unq_uno_mvc}
\alias{unq_uno_scl}
\alias{unq_uno_vec}
\alias{unq_uno_vls}
\alias{unq_whl_arr}
\alias{unq_whl_dtf}
\alias{unq_whl_gen}
\alias{unq_whl_mat}
\alias{unq_whl_mvc}
\alias{unq_whl_scl}
\alias{unq_whl_vec}
\alias{unq_whl_vls}
\title{Combo Uniqueness Plus Extended Mode Plus Extended Class Combination Properties}
\usage{
unq_mmm_ccc_prop_funs()

unq_mmm_ccc(x, mmm, ccc, ...)

unq_mmm_ccc_funs()

unq_mmm_ccc_props()

unq_atm_arr(x, ...)

unq_atm_dtf(x, ...)

unq_atm_gen(x, ...)

unq_atm_mat(x, ...)

unq_atm_mvc(x, ...)

unq_atm_scl(x, ...)

unq_atm_vec(x, ...)

unq_atm_vls(x, ...)

unq_ch1_arr(x, ...)

unq_ch1_dtf(x, ...)

unq_ch1_gen(x, ...)

unq_ch1_mat(x, ...)

unq_ch1_mvc(x, ...)

unq_ch1_scl(x, ...)

unq_ch1_vec(x, ...)

unq_ch1_vls(x, ...)

unq_ch3_arr(x, ...)

unq_ch3_dtf(x, ...)

unq_ch3_gen(x, ...)

unq_ch3_mat(x, ...)

unq_ch3_mvc(x, ...)

unq_ch3_scl(x, ...)

unq_ch3_vec(x, ...)

unq_ch3_vls(x, ...)

unq_chr_arr(x, ...)

unq_chr_dtf(x, ...)

unq_chr_gen(x, ...)

unq_chr_mat(x, ...)

unq_chr_mvc(x, ...)

unq_chr_scl(x, ...)

unq_chr_vec(x, ...)

unq_chr_vls(x, ...)

unq_clr_arr(x, ...)

unq_clr_dtf(x, ...)

unq_clr_gen(x, ...)

unq_clr_mat(x, ...)

unq_clr_mvc(x, ...)

unq_clr_scl(x, ...)

unq_clr_vec(x, ...)

unq_clr_vls(x, ...)

unq_evn_arr(x, ...)

unq_evn_dtf(x, ...)

unq_evn_gen(x, ...)

unq_evn_mat(x, ...)

unq_evn_mvc(x, ...)

unq_evn_scl(x, ...)

unq_evn_vec(x, ...)

unq_evn_vls(x, ...)

unq_fac_arr(x, ...)

unq_fac_dtf(x, ...)

unq_fac_gen(x, ...)

unq_fac_mat(x, ...)

unq_fac_mvc(x, ...)

unq_fac_scl(x, ...)

unq_fac_vec(x, ...)

unq_fac_vls(x, ...)

unq_frc_arr(x, ...)

unq_frc_dtf(x, ...)

unq_frc_gen(x, ...)

unq_frc_mat(x, ...)

unq_frc_mvc(x, ...)

unq_frc_scl(x, ...)

unq_frc_vec(x, ...)

unq_frc_vls(x, ...)

unq_ind_arr(x, ...)

unq_ind_dtf(x, ...)

unq_ind_gen(x, ...)

unq_ind_mat(x, ...)

unq_ind_mvc(x, ...)

unq_ind_scl(x, ...)

unq_ind_vec(x, ...)

unq_ind_vls(x, ...)

unq_lgl_arr(x, ...)

unq_lgl_dtf(x, ...)

unq_lgl_gen(x, ...)

unq_lgl_mat(x, ...)

unq_lgl_mvc(x, ...)

unq_lgl_scl(x, ...)

unq_lgl_vec(x, ...)

unq_lgl_vls(x, ...)

unq_neg_arr(x, ...)

unq_neg_dtf(x, ...)

unq_neg_gen(x, ...)

unq_neg_mat(x, ...)

unq_neg_mvc(x, ...)

unq_neg_scl(x, ...)

unq_neg_vec(x, ...)

unq_neg_vls(x, ...)

unq_ngw_arr(x, ...)

unq_ngw_dtf(x, ...)

unq_ngw_gen(x, ...)

unq_ngw_mat(x, ...)

unq_ngw_mvc(x, ...)

unq_ngw_scl(x, ...)

unq_ngw_vec(x, ...)

unq_ngw_vls(x, ...)

unq_nng_arr(x, ...)

unq_nng_dtf(x, ...)

unq_nng_gen(x, ...)

unq_nng_mat(x, ...)

unq_nng_mvc(x, ...)

unq_nng_scl(x, ...)

unq_nng_vec(x, ...)

unq_nng_vls(x, ...)

unq_nnw_arr(x, ...)

unq_nnw_dtf(x, ...)

unq_nnw_gen(x, ...)

unq_nnw_mat(x, ...)

unq_nnw_mvc(x, ...)

unq_nnw_scl(x, ...)

unq_nnw_vec(x, ...)

unq_nnw_vls(x, ...)

unq_nps_arr(x, ...)

unq_nps_dtf(x, ...)

unq_nps_gen(x, ...)

unq_nps_mat(x, ...)

unq_nps_mvc(x, ...)

unq_nps_scl(x, ...)

unq_nps_vec(x, ...)

unq_nps_vls(x, ...)

unq_npw_arr(x, ...)

unq_npw_dtf(x, ...)

unq_npw_gen(x, ...)

unq_npw_mat(x, ...)

unq_npw_mvc(x, ...)

unq_npw_scl(x, ...)

unq_npw_vec(x, ...)

unq_npw_vls(x, ...)

unq_nst_arr(x, ...)

unq_nst_dtf(x, ...)

unq_nst_gen(x, ...)

unq_nst_mat(x, ...)

unq_nst_mvc(x, ...)

unq_nst_scl(x, ...)

unq_nst_vec(x, ...)

unq_nst_vls(x, ...)

unq_num_arr(x, ...)

unq_num_dtf(x, ...)

unq_num_gen(x, ...)

unq_num_mat(x, ...)

unq_num_mvc(x, ...)

unq_num_scl(x, ...)

unq_num_vec(x, ...)

unq_num_vls(x, ...)

unq_odd_arr(x, ...)

unq_odd_dtf(x, ...)

unq_odd_gen(x, ...)

unq_odd_mat(x, ...)

unq_odd_mvc(x, ...)

unq_odd_scl(x, ...)

unq_odd_vec(x, ...)

unq_odd_vls(x, ...)

unq_ord_arr(x, ...)

unq_ord_dtf(x, ...)

unq_ord_gen(x, ...)

unq_ord_mat(x, ...)

unq_ord_mvc(x, ...)

unq_ord_scl(x, ...)

unq_ord_vec(x, ...)

unq_ord_vls(x, ...)

unq_pct_arr(x, ...)

unq_pct_dtf(x, ...)

unq_pct_gen(x, ...)

unq_pct_mat(x, ...)

unq_pct_mvc(x, ...)

unq_pct_scl(x, ...)

unq_pct_vec(x, ...)

unq_pct_vls(x, ...)

unq_pos_arr(x, ...)

unq_pos_dtf(x, ...)

unq_pos_gen(x, ...)

unq_pos_mat(x, ...)

unq_pos_mvc(x, ...)

unq_pos_scl(x, ...)

unq_pos_vec(x, ...)

unq_pos_vls(x, ...)

unq_ppn_arr(x, ...)

unq_ppn_dtf(x, ...)

unq_ppn_gen(x, ...)

unq_ppn_mat(x, ...)

unq_ppn_mvc(x, ...)

unq_ppn_scl(x, ...)

unq_ppn_vec(x, ...)

unq_ppn_vls(x, ...)

unq_psw_arr(x, ...)

unq_psw_dtf(x, ...)

unq_psw_gen(x, ...)

unq_psw_mat(x, ...)

unq_psw_mvc(x, ...)

unq_psw_scl(x, ...)

unq_psw_vec(x, ...)

unq_psw_vls(x, ...)

unq_srt_arr(x, ...)

unq_srt_dtf(x, ...)

unq_srt_gen(x, ...)

unq_srt_mat(x, ...)

unq_srt_mvc(x, ...)

unq_srt_scl(x, ...)

unq_srt_vec(x, ...)

unq_srt_vls(x, ...)

unq_str_arr(x, ...)

unq_str_dtf(x, ...)

unq_str_gen(x, ...)

unq_str_mat(x, ...)

unq_str_mvc(x, ...)

unq_str_scl(x, ...)

unq_str_vec(x, ...)

unq_str_vls(x, ...)

unq_uno_arr(x, ...)

unq_uno_dtf(x, ...)

unq_uno_gen(x, ...)

unq_uno_mat(x, ...)

unq_uno_mvc(x, ...)

unq_uno_scl(x, ...)

unq_uno_vec(x, ...)

unq_uno_vls(x, ...)

unq_whl_arr(x, ...)

unq_whl_dtf(x, ...)

unq_whl_gen(x, ...)

unq_whl_mat(x, ...)

unq_whl_mvc(x, ...)

unq_whl_scl(x, ...)

unq_whl_vec(x, ...)

unq_whl_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{mmm}{A character scalar single extended mode property from \code{\link{mmm_props}()}.}

\item{ccc}{A character scalar single extended class property from \code{\link{ccc_props}()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Combo uniqueness plus extended mode plus extended class combination properties are the combination of \link[=UNQ]{uniqueness}, a single \link[=mmm]{extended mode} property, and a single \link[=ccc]{extended class} property.
}
\section{Functions}{
\itemize{
\item \code{unq_mmm_ccc()}: Checks \code{x} for uniqueness, for extended mode \code{mmm}, and for extended class \code{ccc} subject to any count or value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_mmm_ccc_funs()}: Lists all combo uniqueness plus extended mode plus extended class property checking functions. Returns a character vector.

\item \code{unq_mmm_ccc_props()}: Lists all combo uniqueness plus extended mode plus extended class properties.

\item \code{unq_atm_arr()}: Checks \code{x} for uniqueness, for atomicness-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_dtf()}: Checks \code{x} for uniqueness, for atomicness-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_gen()}: Checks \code{x} for uniqueness, for atomicness-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_mat()}: Checks \code{x} for uniqueness, for atomicness-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_mvc()}: Checks \code{x} for uniqueness, for atomicness-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_scl()}: Checks \code{x} for uniqueness, for atomicness-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_vec()}: Checks \code{x} for uniqueness, for atomicness-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_atm_vls()}: Checks \code{x} for uniqueness, for atomicness-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_arr()}: Checks \code{x} for uniqueness, for onechar-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_dtf()}: Checks \code{x} for uniqueness, for onechar-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_gen()}: Checks \code{x} for uniqueness, for onechar-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_mat()}: Checks \code{x} for uniqueness, for onechar-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_mvc()}: Checks \code{x} for uniqueness, for onechar-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_scl()}: Checks \code{x} for uniqueness, for onechar-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_vec()}: Checks \code{x} for uniqueness, for onechar-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch1_vls()}: Checks \code{x} for uniqueness, for onechar-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_arr()}: Checks \code{x} for uniqueness, for threechar-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_dtf()}: Checks \code{x} for uniqueness, for threechar-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_gen()}: Checks \code{x} for uniqueness, for threechar-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_mat()}: Checks \code{x} for uniqueness, for threechar-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_mvc()}: Checks \code{x} for uniqueness, for threechar-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_scl()}: Checks \code{x} for uniqueness, for threechar-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_vec()}: Checks \code{x} for uniqueness, for threechar-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ch3_vls()}: Checks \code{x} for uniqueness, for threechar-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_arr()}: Checks \code{x} for uniqueness, for character-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_dtf()}: Checks \code{x} for uniqueness, for character-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_gen()}: Checks \code{x} for uniqueness, for character-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_mat()}: Checks \code{x} for uniqueness, for character-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_mvc()}: Checks \code{x} for uniqueness, for character-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_scl()}: Checks \code{x} for uniqueness, for character-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_vec()}: Checks \code{x} for uniqueness, for character-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_chr_vls()}: Checks \code{x} for uniqueness, for character-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_arr()}: Checks \code{x} for uniqueness, for color-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_dtf()}: Checks \code{x} for uniqueness, for color-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_gen()}: Checks \code{x} for uniqueness, for color-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_mat()}: Checks \code{x} for uniqueness, for color-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_mvc()}: Checks \code{x} for uniqueness, for color-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_scl()}: Checks \code{x} for uniqueness, for color-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_vec()}: Checks \code{x} for uniqueness, for color-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_clr_vls()}: Checks \code{x} for uniqueness, for color-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_arr()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_dtf()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_gen()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_mat()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_mvc()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_scl()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_vec()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_evn_vls()}: Checks \code{x} for uniqueness, for even-whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_arr()}: Checks \code{x} for uniqueness, for factor-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_dtf()}: Checks \code{x} for uniqueness, for factor-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_gen()}: Checks \code{x} for uniqueness, for factor-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_mat()}: Checks \code{x} for uniqueness, for factor-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_mvc()}: Checks \code{x} for uniqueness, for factor-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_scl()}: Checks \code{x} for uniqueness, for factor-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_vec()}: Checks \code{x} for uniqueness, for factor-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_fac_vls()}: Checks \code{x} for uniqueness, for factor-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_arr()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_dtf()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_gen()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_mat()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_mvc()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_scl()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_vec()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_frc_vls()}: Checks \code{x} for uniqueness, for fractional-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_arr()}: Checks \code{x} for uniqueness, for indexer-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_dtf()}: Checks \code{x} for uniqueness, for indexer-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_gen()}: Checks \code{x} for uniqueness, for indexer-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_mat()}: Checks \code{x} for uniqueness, for indexer-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_mvc()}: Checks \code{x} for uniqueness, for indexer-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_scl()}: Checks \code{x} for uniqueness, for indexer-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_vec()}: Checks \code{x} for uniqueness, for indexer-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ind_vls()}: Checks \code{x} for uniqueness, for indexer-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_arr()}: Checks \code{x} for uniqueness, for logical-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_dtf()}: Checks \code{x} for uniqueness, for -logicalness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_gen()}: Checks \code{x} for uniqueness, for logical-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_mat()}: Checks \code{x} for uniqueness, for logical-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_mvc()}: Checks \code{x} for uniqueness, for logical-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_scl()}: Checks \code{x} for uniqueness, for logical-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_vec()}: Checks \code{x} for uniqueness, for logical-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_lgl_vls()}: Checks \code{x} for uniqueness, for logical-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_arr()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_dtf()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_gen()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_mat()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_mvc()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_scl()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_vec()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_neg_vls()}: Checks \code{x} for uniqueness, for negative-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_arr()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_dtf()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_gen()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_mat()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_mvc()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_scl()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_vec()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ngw_vls()}: Checks \code{x} for uniqueness, for negative-whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_arr()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_dtf()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_gen()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_mat()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_mvc()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_scl()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_vec()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nng_vls()}: Checks \code{x} for uniqueness, for non-negative-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_arr()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_dtf()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_gen()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_mat()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_mvc()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_scl()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_vec()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nnw_vls()}: Checks \code{x} for uniqueness, for non-negative-whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_arr()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_dtf()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_gen()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_mat()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_mvc()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_scl()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_vec()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nps_vls()}: Checks \code{x} for uniqueness, for non-positive-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_arr()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_dtf()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_gen()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_mat()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_mvc()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_scl()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_vec()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_npw_vls()}: Checks \code{x} for uniqueness, for non-positive-whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_arr()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_dtf()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_gen()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_mat()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_mvc()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_scl()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_vec()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_nst_vls()}: Checks \code{x} for uniqueness, for nonsortable-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_arr()}: Checks \code{x} for uniqueness, for numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_dtf()}: Checks \code{x} for uniqueness, for numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_gen()}: Checks \code{x} for uniqueness, for numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_mat()}: Checks \code{x} for uniqueness, for numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_mvc()}: Checks \code{x} for uniqueness, for numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_scl()}: Checks \code{x} for uniqueness, for numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_vec()}: Checks \code{x} for uniqueness, for numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_num_vls()}: Checks \code{x} for uniqueness, for numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_arr()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_dtf()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_gen()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_mat()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_mvc()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_scl()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_vec()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_odd_vls()}: Checks \code{x} for uniqueness, for odd-whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_arr()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_dtf()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_gen()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_mat()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_mvc()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_scl()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_vec()}: Checks \code{x} for uniqueness, for -ordered-factorness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ord_vls()}: Checks \code{x} for uniqueness, for ordered-factor-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_arr()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_dtf()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_gen()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_mat()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_mvc()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_scl()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_vec()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pct_vls()}: Checks \code{x} for uniqueness, for percent-valued-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_arr()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_dtf()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_gen()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_mat()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_mvc()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_scl()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_vec()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_pos_vls()}: Checks \code{x} for uniqueness, for positive-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_arr()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_dtf()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_gen()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_mat()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_mvc()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_scl()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_vec()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_ppn_vls()}: Checks \code{x} for uniqueness, for proportion-valued-numeric-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_arr()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_dtf()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_gen()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_mat()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_mvc()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_scl()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_vec()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_psw_vls()}: Checks \code{x} for uniqueness, for positive-whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_arr()}: Checks \code{x} for uniqueness, for sortable-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_dtf()}: Checks \code{x} for uniqueness, for sortable-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_gen()}: Checks \code{x} for uniqueness, for sortable-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_mat()}: Checks \code{x} for uniqueness, for sortable-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_mvc()}: Checks \code{x} for uniqueness, for sortable-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_scl()}: Checks \code{x} for uniqueness, for sortable-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_vec()}: Checks \code{x} for uniqueness, for sortable-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_srt_vls()}: Checks \code{x} for uniqueness, for sortable-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_arr()}: Checks \code{x} for uniqueness, for string-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_dtf()}: Checks \code{x} for uniqueness, for string-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_gen()}: Checks \code{x} for uniqueness, for string-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_mat()}: Checks \code{x} for uniqueness, for string-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_mvc()}: Checks \code{x} for uniqueness, for string-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_scl()}: Checks \code{x} for uniqueness, for string-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_vec()}: Checks \code{x} for uniqueness, for string-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_str_vls()}: Checks \code{x} for uniqueness, for string-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_arr()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_dtf()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_gen()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_mat()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_mvc()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_scl()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_vec()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_uno_vls()}: Checks \code{x} for uniqueness, for unordered-factor-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_arr()}: Checks \code{x} for uniqueness, for whole-number-ness, and for array-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_dtf()}: Checks \code{x} for uniqueness, for whole-number-ness, and for data.frame-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_gen()}: Checks \code{x} for uniqueness, for whole-number-ness, and for generic-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_mat()}: Checks \code{x} for uniqueness, for whole-number-ness, and for matrix-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_mvc()}: Checks \code{x} for uniqueness, for whole-number-ness, and for multivec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_scl()}: Checks \code{x} for uniqueness, for whole-number-ness, and for scalar-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_vec()}: Checks \code{x} for uniqueness, for whole-number-ness, and for vec-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

\item \code{unq_whl_vls()}: Checks \code{x} for uniqueness, for whole-number-ness, and for vector-list-ness subject to any count/value restrictions in \code{...}. Returns a logical scalar.

}}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{\verb{.max, .maxr, .maxc} \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                 \cr  \tab  \cr
\verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                 \cr  \tab  \cr
\verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr  \tab  \cr
\verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr  \tab  \cr
\code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
unq_mmm_ccc_funs()
unq_mmm_ccc(letters, "ch1", "vec")
unq_mmm_ccc(letters, "str", "scl")
unq_ch1_vec(letters)
unq_str_scl("a")
}
\seealso{
Other props: 
\code{\link{cmp_mmm_ccc_prop_funs}()},
\code{\link{mmm_ccc_prop_funs}()},
\code{\link{sss_ccc_prop_funs}()}
}
\concept{props}
