% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.r
\name{uv.}
\alias{uv.}
\alias{uv}
\alias{uav}
\title{Reduce \code{...} to a vector of unique atomic values}
\usage{
uv.()

uv(...)

uav(...)
}
\arguments{
\item{...}{One or more objects to be reduced to a single atomic vector and
then further reduced to unique atomic values.}
}
\value{
An atomic vector.
}
\description{
\link[av]{Atomizes} \code{...}, reducing the set of arguments to
a single vector of unique constituent atomic values.
}
\examples{
uv(0:5, 5:10, 10:15, 15:20)
}
