% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{v}
\alias{v}
\title{Get one or more character scalar constants.}
\usage{
v(...)
}
\arguments{
\item{...}{unquoted, comma-separated list of the names of constants to
return. If multiple constants are specified, they are coerced into a single
atomic vector result. If no constants are specified (i.e., \code{v()}), a
tibble is returned giving}
}
\value{
An atomic vector of length 1 or greater.
}
\description{
Get one or more character scalar constants.
}
\seealso{
Other constants: 
\code{\link{l}()}
}
\concept{constants}
