% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\alias{vnames}
\alias{vtable}
\alias{v}
\title{Package \code{uj} defined values}
\usage{
values()

vnames()

vtable()

v(...)
}
\arguments{
\item{...}{Unquoted, comma-separated list of the names of constants to return. If multiple constants are specified, they are coerced into a single atomic vector result.}
}
\value{
\itemize{
\item \strong{\code{values}}: a \link[=ivls]{vlist}.
\item \strong{\code{vtable}}: a data.frame.
\item \strong{\verb{v, vnames}}: a character vector.
}
}
\description{
\itemize{
\item \strong{\code{values}}: gets a named list where names are the arguments submitted to \code{v(...)}, element values are the values of the contents of the named element, and the attribute \code{'descriptions'} gives verbose descriptions if needed.
\item \strong{\code{vnames}}: gets a vector of named values available.
\item \strong{\code{vtable}}: gets a tibble with three columns: \code{name}, \code{value}, and \code{description}.
\item \strong{\code{v}}: retrieves one or more \code{uj} package values by name.
}
}
