% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wraps_scales.R
\docType{data}
\name{wrap_scales}
\alias{wrap_scales}
\alias{sca}
\title{Wraps of Functions from the \code{scales} Package}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
wrap_scales

sca(color, alpha = NA)
}
\arguments{
\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA},
existing alpha values are preserved.}
}
\description{
\tabular{ll}{
\code{sca}   \tab Thin wrap of \code{\link[scales]{alpha}}.               }
}
\examples{
alpha("red", 0.1)
alpha(colours(), 0.5)
alpha("red", seq(0, 1, length.out = 10))
alpha(c("first" = "gold", "second" = "lightgray", "third" = "#cd7f32"), .5)
}
\seealso{
Other wraps: 
\code{\link{wrap_base}},
\code{\link{wrap_data.table}},
\code{\link{wrap_dplyr}},
\code{\link{wrap_gg}},
\code{\link{wrap_readr}},
\code{\link{wrap_tibble}}
}
\concept{wraps}
\keyword{datasets}
