% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wraps_base.R
\name{wraps_base}
\alias{wraps_base}
\alias{u}
\alias{dn}
\alias{up}
\alias{mid}
\alias{spf}
\alias{levs}
\title{Thin wrappers for \code{base} functions}
\usage{
u(x)

dn(x)

up(x)

mid(x, start, stop)

spf(fmt, ...)

levs(x)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}

\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{arguments for particular methods.}
}
\value{
For a vector, an object of the same type of \code{x}, but with only
  one copy of each duplicated element.  No attributes are copied (so
  the result has no names).

  For a data frame, a data frame is returned with the same columns but
  possibly fewer rows (and with row names from the first occurrences of
  the unique rows).

  A matrix or array is subsetted by \code{[, drop = FALSE]}, so
  dimensions and dimnames are copied appropriately, and the result
  always has the same number of dimensions as \code{x}.
}
\description{
\itemize{
\item \strong{\code{u}}: thinly wraps code{\link[base]{unique}}.
\item \strong{\code{up}}: thinly wraps \code{\link[base]{ceiling}}.
\item \strong{\code{dn}}: thinly wraps \code{\link[base]{floor}}.
\item \strong{\code{mid}}: thinly wraps \code{\link[base]{substr}}.
\item \strong{\code{spf}}: thinly wraps \code{\link[base]{sprintf}}.
\item \strong{\code{levs}}: thinly wraps \code{\link[base]{levels}}.
}
NOTE: \code{u} differs from \code{\link{uv}} in that \code{uv} \link[=a]{atomizes} an arbitrary number of \code{...} arguments before calling \code{unique}.
}
\details{
This is a generic function with methods for vectors, data frames and
  arrays (including matrices).

  The array method calculates for each element of the dimension
  specified by \code{MARGIN} if the remaining dimensions are identical
  to those for an earlier element (in row-major order).  This would most
  commonly be used for matrices to find unique rows (the default) or columns
  (with \code{MARGIN = 2}).

  Note that unlike the Unix command \code{uniq} this omits
  \emph{duplicated} and not just \emph{repeated} elements/rows.  That
  is, an element is omitted if it is equal to any previous element and
  not just if it is equal the immediately previous one.  (For the
  latter, see \code{\link[base]{rle}}).

  Missing values (\code{"\link[base]{NA}"}) are regarded as equal, numeric and
  complex ones differing from \code{NaN}; character strings will be compared in a
  \dQuote{common encoding}; for details, see \code{\link[base]{match}} (and
  \code{\link[base]{duplicated}}) which use the same concept.

  Values in \code{incomparables} will never be marked as duplicated.
  This is intended to be used for a fairly small set of values and will
  not be efficient for a very large set.

  When used on a data frame with more than one column, or an array or
  matrix when comparing dimensions of length greater than one, this
  tests for identity of character representations.  This will
  catch people who unwisely rely on exact equality of floating-point
  numbers!
}
\note{
The S version of \code{substring<-} ignores \code{last}; this version
  does not.

  These functions are often used with \code{\link[base]{nchar}} to truncate a
  display.  That does not really work (you want to limit the width, not
  the number of characters, so it would be better to use
  \code{\link[base]{strtrim}}), but at least make sure you use the default
  \code{nchar(type = "c")}.
}
\section{Warning}{

  Using this for lists is potentially slow, especially if the elements
  are not atomic vectors (see \code{\link[base]{vector}}) or differ only
  in their attributes.  In the worst case it is \eqn{O(n^2)}.



  The realities of computer arithmetic can cause unexpected results,
  especially with \code{floor} and \code{ceiling}.  For example, we
  \sQuote{know} that \code{floor(log(x, base = 8))} for \code{x = 8} is
  \code{1}, but \code{0} has been seen on an \R platform.  It is
  normally necessary to use a tolerance.

  Rounding to decimal digits in binary arithmetic is non-trivial (when
  \code{digits != 0}) and may be surprising.  Be aware that most decimal
  fractions are \emph{not} exactly representable in binary double precision.
  In \R 4.0.0, the algorithm for \code{round(x, d)}, for \eqn{d > 0}, has
  been improved to \emph{measure} and round \dQuote{to nearest even},
  contrary to earlier versions of \R (or also to \code{\link[base]{sprintf}()}
  or \code{\link[base]{format}()} based rounding).



  The realities of computer arithmetic can cause unexpected results,
  especially with \code{floor} and \code{ceiling}.  For example, we
  \sQuote{know} that \code{floor(log(x, base = 8))} for \code{x = 8} is
  \code{1}, but \code{0} has been seen on an \R platform.  It is
  normally necessary to use a tolerance.

  Rounding to decimal digits in binary arithmetic is non-trivial (when
  \code{digits != 0}) and may be surprising.  Be aware that most decimal
  fractions are \emph{not} exactly representable in binary double precision.
  In \R 4.0.0, the algorithm for \code{round(x, d)}, for \eqn{d > 0}, has
  been improved to \emph{measure} and round \dQuote{to nearest even},
  contrary to earlier versions of \R (or also to \code{\link[base]{sprintf}()}
  or \code{\link[base]{format}()} based rounding).



  The format string is passed down the OS's \code{sprintf} function, and
  incorrect formats can cause the latter to crash the \R process .  \R
  does perform sanity checks on the format, but not all possible user
  errors on all platforms have been tested, and some might be terminal.

  The behaviour on inputs not documented here is \sQuote{undefined},
  which means it is allowed to differ by platform.

}

\section{S4 methods}{

  These are all (internally) S4 generic.

  \code{ceiling}, \code{floor} and \code{trunc} are members of the
  \code{\link[=S4groupGeneric]{Math}} group generic.  As an S4
  generic, \code{trunc} has only one argument.

  \code{round} and \code{signif} are members of the
  \code{\link[=S4groupGeneric]{Math2}} group generic.



  These are all (internally) S4 generic.

  \code{ceiling}, \code{floor} and \code{trunc} are members of the
  \code{\link[=S4groupGeneric]{Math}} group generic.  As an S4
  generic, \code{trunc} has only one argument.

  \code{round} and \code{signif} are members of the
  \code{\link[=S4groupGeneric]{Math2}} group generic.

}

\examples{
x <- c(3:5, 11:8, 8 + 0:5)
(ux <- unique(x))
(u2 <- unique(x, fromLast = TRUE)) # different order
stopifnot(identical(sort(ux), sort(u2)))

length(unique(sample(100, 100, replace = TRUE)))
## approximately 100(1 - 1/e) = 63.21

unique(iris)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
\seealso{
Other wraps: 
\code{\link{wraps_data_table}},
\code{\link{wraps_dplyr}},
\code{\link{wraps_gg}},
\code{\link{wraps_graphics}},
\code{\link{wraps_readr}},
\code{\link{wraps_scales}},
\code{\link{wraps_tibble}}
}
\author{
Original code by Jonathan Rougier.
}
\concept{wraps}
