% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv.R
\encoding{UTF-8}
\name{wv}
\alias{wv}
\title{Which \link[=av]{atomized} values are \code{TRUE}?}
\usage{
wv(...)
}
\description{
Calls \code{which(av(x))}.
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unq}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{x}()}

Other values: 
\code{\link{atomize}()},
\code{\link{n_th}()},
\code{\link{na}()},
\code{\link{uv}()},
\code{\link{values}()}
}
\concept{extensions}
\concept{values}
