% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{x.}
\alias{x.}
\alias{xat}
\alias{xat0}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{Remove various objects or characteristics.}
\usage{
x.()

xat(x, ...)

xat0(x)

xcls(x, ...)

xwarn()

xplot(x = NULL)

xcon()
}
\arguments{
\item{x}{An R object.}

\item{...}{One or more \link[cmp_chr_vec]{complete character vecs} containing
names of attributes or classes to remove.}
}
\value{
An object (\code{xat} and \code{xcls}) or \code{NULL} (\code{xwarn}
and \code{xcls}).
}
\description{
Remove various objects or characteristics.
}
\section{Functions}{
\itemize{
\item \code{xat()}: Remove from \code{x.} the attributes named in \code{...}.

\item \code{xat0()}: Remove all attributes from \code{x.}.

\item \code{xcls()}: Remove from \code{x.} the classes named in \code{...}.

\item \code{xwarn()}: Purge all warnings.

\item \code{xplot()}: Purge all plots. Then if optional \code{x.} is provided, print
it (\code{x.} is intended to be a plot object).

\item \code{xcon()}: Clear the console.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{xb.}()}
}
\concept{extensions}
