% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats.R, R/x.R
\encoding{UTF-8}
\name{xat}
\alias{xat}
\alias{x}
\alias{removal}
\alias{xplot}
\alias{xall}
\alias{xwarn}
\alias{xcon}
\title{'\code{x}' out objects}
\usage{
xat(x, ...)

x(...)

xplot(x = NULL)

xall()

xwarn()

xcon()
}
\arguments{
\item{x}{An R object.}

\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs} containing names of attributes or classes to remove.}
}
\value{
\code{NULL}
\cr   \code{xplot}
\cr   \code{xwarn}
\cr   \code{xcon}
\cr   \code{xall}
\cr   \code{x}
}
\description{
All \code{x} (removal) functions are followed by garbage collection.
\tabular{rl}{
\code{xall}   \tab \strong{WARNING: Use with care!}
\cr             \tab \strong{Purges the global environment.}
\cr             \tab  
\cr   \code{xplot}   \tab Purges plots and prints \code{x}.
\cr   \code{xwarn}   \tab Purges warnings.
\cr    \code{xcon}   \tab Purges console.
\cr       \code{x}   \tab Deletes \code{...} args in the calling environment.
}
}
\examples{
\dontrun{
## build a ggplot scatterplot object (y.)
library(ggplot2)
ggp. <- ggplot(obj., aes(x, y)) + geom_point()

## print to console, wait for user, purge console
say("\n A message")
continue()
xcon()

## display a scatterplot, wait for user
plot(obj.$x, obj.$y)
continue()

## purge plots and print ggplot
xplot(ggp.)

## wait for user, purge plots again
continue()
xplot()

## generate warning, print warnings
sqrt(-1)
warnings()

## purge warnings, validate that warnings are purged
xwarn()
warnings()
}
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()}

Other meta: 
\code{\link{callers}()},
\code{\link{dot_args}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{say}()}
}
\concept{extensions}
\concept{meta}
