% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{x_uj}
\alias{x_uj}
\alias{xat}
\alias{xAT}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{Remove various objects or characteristics.}
\usage{
x_uj()

xat(x., ...)

xAT(x.)

xcls(x., ...)

xwarn()

xplot(x. = NULL)

xcon()
}
\arguments{
\item{x.}{An object.}

\item{...}{Character vectors containing attribute names or classes to remove.}
}
\value{
An object (\code{xat} and \code{xcls}) or \code{NULL} (\code{xwarn}
and \code{xcls}).
}
\description{
Remove various objects or characteristics.
}
\section{Functions}{
\itemize{
\item \code{xat}: Remove from \code{x.} the attributes listed in \code{...}.

\item \code{xAT}: Remove all attributes from \code{x.}.

\item \code{xcls}: Remove from \code{x.} the classes named in \code{...}.

\item \code{xwarn}: Purge all warnings.

\item \code{xplot}: Purge all plots and if optional \code{x.} is provided,
print it (\code{x.} is intended to be a plot object).

\item \code{xcon}: Clear the console.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
