% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xb.R
\name{xb.}
\alias{xb.}
\alias{xb}
\alias{cb}
\alias{rb}
\title{Error-checked row and column binding}
\usage{
xb.()

xb(..., bind = NA)

cb(...)

rb(...)
}
\arguments{
\item{...}{Multiple \code{\link{compatible}} atomic matrices or multiple
compatible \link[idtf]{atomic data.frames}.}

\item{bind}{\link[ch1_scl]{Onechar scalar} indicating how to bind: \code{'r'}
for row binding, \code{'c'} for column-binding, \code{NA} to identify
whether either row or column binding is successful and return whichever was
successful.}
}
\value{
An \code{\link[idtf]{atomic dtf}} or an atomic matrix.
}
\description{
Row and/or column binds arguments in \code{...}. Requires all
arguments in \code{...} to be \code{\link{compatible}} atomic matrices or
\link[idtf]{atomic data.frames}.
}
\section{Functions}{
\itemize{
\item \code{xb()}: Column or row bind arguments in \code{...}. If \code{bind =
  NA}, \code{xb} attempts to both row and column bind arguments in
\code{...}. If both are successful, processes an error indicating ambiguous
dimension for binding. If neither is successful, processes an associated
error. Otherwise, returns the result that was successful.

\item \code{cb()}: Column bind an arbitrary number of matrices or data frames.

\item \code{rb()}: Row bind an arbitrary number of matrices or data frames.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()}
}
\concept{extensions}
