% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xb.R
\name{xb}
\alias{xb}
\alias{cb}
\alias{rb}
\alias{cb_unq}
\title{Robustly Error-Checked Row and Column Binding for Atomic Matrices and
Atomic Tibbles.}
\usage{
xb(..., bind. = NA)

cb(...)

rb(...)

cb_unq(x., y.)
}
\arguments{
\item{..., x., y.}{Either \link[=compatible_mats]{atomic matrices compatible
for binding} or \link[=compatible_atbs]{atomic tibbles compatible for
binding}.}

\item{bind.}{\code{'r'} to indicate row binding, \code{'c'} to indicate
column, or \code{NA} to identify whether either row or column binding is
successful and return whichever was successful.}
}
\value{
An \link[=atm_tbl]{atomic tibble} or an atomic matrix.
}
\description{
Requires all arguments in \code{...} to be either (a)
\link[=compatible_mats]{compatible atomic matrices} or (b)
\link[=compatible_atbs]{compatible atomic tibbles}.
}
\details{
If \code{bind. = NA}, \code{xb} attempts to row bind and column bind
arguments in \code{...}. If both are successful, processes an error
indicating ambiguous dimension for binding. If neither is successful,
processes an associated error. Otherwise, returns the result that was
successful.
}
\section{Functions}{
\itemize{
\item \code{cb}: Column bind an arbitrary number of matrices or data frames.

\item \code{rb}: Row bind an arbitrary number of matrices or data frames.

\item \code{cb_unq}: Column bind two data frames or matrices keeping only the
intersection of columns/variables they share.
}}

\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}}
}
\concept{meta}
