% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xb.R
\name{xb_uj}
\alias{xb_uj}
\alias{xb}
\alias{cb}
\alias{rb}
\title{Error-checked row and column binding}
\usage{
xb_uj()

xb(..., bind. = NA)

cb(...)

rb(...)
}
\arguments{
\item{..., x., y.}{Either atomic matrices compatible for binding or atomic
tabulars compatible for binding (see \code{\link{compatible}}).}

\item{bind.}{\code{'r'} to indicate row binding, \code{'c'} to indicate
column, or \code{NA} to identify whether either row or column binding is
successful and return whichever was successful.}
}
\value{
An atomic tabular or an atomic matrix.
}
\description{
Requires all arguments in \code{...} to be either (a) compatible
atomic matrices or (b) compatible atomic tabulars (see
\code{\link{compatible}}).
}
\details{
If \code{bind. = NA}, \code{xb} attempts to row bind and column bind
arguments in \code{...}. If both are successful, processes an error
indicating ambiguous dimension for binding. If neither is successful,
processes an associated error. Otherwise, returns the result that was
successful.
}
\section{Functions}{
\itemize{
\item \code{xb}: Column bind an arbitrary number of matrices or data frames.

\item \code{cb}: Column bind an arbitrary number of matrices or data frames.

\item \code{rb}: Row bind an arbitrary number of matrices or data frames.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()}
}
\concept{meta}
