% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpr.R
\name{xpr.}
\alias{xpr.}
\alias{xbold}
\alias{xrun}
\title{Build and/or Evaluate Expressions}
\usage{
xpr.()

xbold(...)

xrun(...)
}
\arguments{
\item{...}{An arbitrary number of objects atomized into a vector and then
collapsed to create an expression.}
}
\value{
Either a character scalar or the result of calling the expression.
}
\description{
Build and/or Evaluate Expressions
}
\section{Functions}{
\itemize{
\item \code{xbold()}: Build an expression bolding the character scalar produced
by collapsing \code{...}.

\item \code{xrun()}: Build an expression from the character scalar produced by
collapsing \code{...} and then execute the resulting command.

}}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()}

Other code: 
\code{\link{run.}()}
}
\concept{code}
\concept{strings}
