% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpr.R
\name{xpr}
\alias{xpr}
\alias{xbold}
\alias{xrun}
\title{Build expressions as character scalars and/or run resulting commands.}
\usage{
xbold(...)

xrun(...)
}
\arguments{
\item{...}{An arbitrary number of objects atomized and glued to create the
expression.}
}
\value{
Either a character scalar or the result of calling the expression.
}
\description{
Build expression containing bold text.
}
\section{Functions}{
\itemize{
\item \code{xrun}: Build an expression and execute the resulting command.
}}

