% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\name{as_mmm}
\alias{as_mmm}
\alias{as_clr}
\alias{as_fun}
\alias{as_chr}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to Base Mode or Extended Mode.}
\usage{
as_clr(x, na = F)

as_fun(x)

as_chr(x, ...)

as_int(x, ...)

as_num(x, ...)

as_lgl(x, ...)

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{For \code{as_clr}, an object of mode character; for \code{as_fun}, a
character scalar function name or a function object; for \code{as_ord} and
\code{as_uno}, an atomic object; and for all others, any R object.}

\item{na}{\link[=cmp_lgl_scl]{Complete logical scalar} indicating whether
\code{NA} values qualify as missing color representations.}

\item{...}{Further arguments passed to or from other methods.}

\item{levs}{\link[=cmp_vec]{Complete atomic vec} giving factor levels.}
}
\value{
\strong{\code{as_clr, as_chr}}
\cr An object of mode \code{'character'}.
\cr\cr
\strong{\code{as_fun}}
\cr A function object.
\cr\cr
\strong{\code{as_int}}
\cr An object of mode \code{'integer'}.
\cr\cr
\strong{\code{as_num}}
\cr An object of mode \code{'numeric'}.
\cr\cr
\strong{\code{as_lgl}}
\cr An object of mode \code{'logical'}.
\cr\cr
\strong{\code{as_ord}}
\cr An object of mode \code{'ordered'} (factor).
\cr\cr
\strong{\code{as_uno}}
\cr An object of \link[=mmm]{extended mode} \code{'uno'} (unordered
factor).
}
\description{
Coerce to Base Mode or Extended Mode.
}
\section{Functions in This Family}{

\strong{\code{as_clr}}
\cr Coerce a valid R color representation of mode character to hexadecimal
RGB character representation.
\cr\cr
\strong{\code{as_fun}}
\cr Return \code{x} if it is a function, otherwise, search for a function
named \code{x} and return it.
\cr\cr
\strong{\code{as_chr}}
\cr Thin wrapper of \code{\link[base]{as.character}}.
\cr\cr
\strong{\code{as_int}}
\cr Thin wrapper of \code{\link[base]{as.integer}}.
\cr\cr
\strong{\code{as_num}}
\cr Thin wrapper of \code{\link[base]{as.numeric}}.
\cr\cr
\strong{\code{as_lgl}}
\cr Thin wrapper of \code{\link[base]{as.logical}}.
\cr\cr
\strong{\code{as_ord}}
\cr Wrapper for \code{factor(x, levels = levs, ordered = TRUE)}.
\cr\cr
\strong{\code{as_uno}}
\cr Wrapper for \code{factor(x, levels = levs, ordered = FALSE)}.
}

\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
