% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\encoding{UTF-8}
\name{as_mmm}
\alias{as_mmm}
\alias{as_clr}
\alias{as_fun}
\alias{as_chr}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to base mode or \code{xmode}}
\usage{
as_clr(x, na = FALSE)

as_fun(x)

as_chr(x, ...)

as_int(x, ...)

as_num(x, ...)

as_lgl(x, ...)

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{For \code{as_clr}, an object of mode character; for \code{as_fun}, a character scalar function name or a function object; for \code{as_ord} and \code{as_uno}, an atomic object; and for all others, base::any R object.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values are acceptable.}

\item{...}{Further arguments passed to or from other methods.}

\item{levs}{A \link[=cmp_vec]{complete atomic vec} of factor levels (ordered factor levels for \code{as_ord}).}
}
\value{
\emph{An object of base mode} \code{'character'}
\cr   \code{as_chr}
\cr\cr \emph{An object of base mode} \code{'integer'}
\cr   \code{as_int}
\cr\cr \emph{An object of base mode} \code{'logical'}
\cr   \code{as_lgl}
\cr\cr \emph{An object of base mode} \code{'numeric'}
\cr   \code{as_num}
\cr\cr \emph{An object of \link[=mmm]{xmode}} \code{'ord'}\eqn{^1}
\cr   \code{as_ord}
\cr\cr \emph{An object of xmode} \code{'uno'}\eqn{^2}
\cr   \code{as_uno}
\cr\cr \emph{An object of xmode} \code{'clr'}\eqn{^3}
\cr   \code{as_clr}
\cr\cr \verb{ }\eqn{^{1.}} Ordered factor.
\cr \verb{ }\eqn{^{2.}} Unordered factor.
\cr \verb{ }\eqn{^{3.}} Character hexadecimal RGB color values in the form \code{'#RRGGBBAA'}.
}
\description{
Functions to coerce objects to \link[base:mode]{base mode} or \code{\link[=mmm]{xmode}}:
\tabular{rl}{
  \code{as_chr}   \tab Thinly wraps \code{\link[base:as.character]{base::as.character}}.
\cr \code{as_int}   \tab Thinly wraps \code{\link[base:as.integer]{base::as.integer}}.
\cr \code{as_num}   \tab Thinly wraps \code{\link[base:as.numeric]{base::as.numeric}}.
\cr \code{as_lgl}   \tab Thinly wraps \code{\link[base:as.logical]{base::as.logical}}.
\cr \code{as_clr}   \tab Coerces R character colors to hexadecimal character RGB colors.
\cr \code{as_fun}   \tab If \code{x} is a character scalar function return that function, but if \code{x} is a function object, return it.
\cr \code{as_ord}   \tab As xmode \code{'ord'} (ordered factor). Wraps \code{base::factor(x, levels = levs, ordered = TRUE)}.
\cr \code{as_uno}   \tab As xmode \code{'uno'} (unordered factor). Wraps \code{base::factor(x, levels = levs, ordered = FALSE)}.
}
}
\examples{
bins. <- sample(c(0, 1), 10, replace = T)
chrs. <- c("3.14", "2.72", "1.41")
clrs. <- c("red", "#AABBCC", "#AABBCCDD", "blue")
nums. <- c(pi, exp(1), sqrt(2))

as_fun(unique)
as_fun("unique")

bins.
as_lgl(bins.)

chrs.
as_num(chrs.)
as_int(chrs.)

clrs.
as_clr(clrs.)

nums.
as_chr(nums.)

as_uno(clrs., levs = unique(clrs.))
as_ord(clrs., levs = sort(unique(clrs.)))
}
\seealso{
Other properties: 
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
