% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\name{as_mmm}
\alias{as_mmm}
\alias{as_clr}
\alias{as_fun}
\alias{as_chr}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to Base Mode or Extended Mode.}
\usage{
as_clr(x, na = F)

as_fun(x)

as_chr(x, ...)

as_int(x, ...)

as_num(x, ...)

as_lgl(x, ...)

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{For \code{as_clr}, an object of mode character; for \code{as_fun}, a
character scalar function name or a function object; for \code{as_ord} and
\code{as_uno}, an atomic object; and for all others, any R object.}

\item{na}{A \link[=cmp_lgl_scl]{complete logical scalar} indicating whether
\code{NA} values qualify as missing color representations.}

\item{...}{Further arguments passed to or from other methods.}

\item{levs}{\link[=cmp_vec]{Complete atomic vec} giving factor levels.}
}
\value{
\tabular{ll}{
FUNCTION             \tab RETURN VALUE                                  \cr
\code{as_clr}, \code{as_chr}   \tab An object of mode \code{'character'}.              \cr
\code{as_fun}             \tab A function object.                            \cr
\code{as_int}             \tab An object of mode \code{'integer'}.                \cr
\code{as_num}             \tab An object of mode \code{'numeric'}.                \cr
\code{as_lgl}             \tab An object of mode \code{'logical'}.                \cr
\code{as_ord}             \tab An object of mode \verb{'ordered'`` (factor).      \\cr }as_uno\verb{            \\tab An object of \link[=mmm]\{extended mode\}}'uno'`
(unordered factor).                             }
}
\description{
\tabular{ll}{
FUNCTION    \tab WHAT IT DOES                                            \cr
\code{as_clr}    \tab Coerce a valid R color representation of mode character to
hexadecimal RGB character representation.              \cr
\code{as_fun}    \tab Return \code{x} if it is a function, otherwise, search for a
function named \code{x} and return it.                      \cr
\code{as_chr}    \tab Thin wrapper of \code{as.character}.                        \cr
\code{as_int}    \tab Thin wrapper of \code{as.integer}.                          \cr
\code{as_num}    \tab Thin wrapper of \code{as.numeric}.                          \cr
\code{as_lgl}    \tab Thin wrapper of \code{as.logical}.                          \cr
\code{as_ord}    \tab Wrapper for \code{factor(x, levels = levs, ordered = TRUE)}.\cr
\code{as_uno}    \tab Wrapper for \code{factor(x, levels = levs, ordered = FALSE)}. }
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
