% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mmm.R
\encoding{UTF-8}
\name{as_mmm}
\alias{as_mmm}
\alias{as_chr}
\alias{as_clr}
\alias{as_fun}
\alias{as_int}
\alias{as_num}
\alias{as_lgl}
\alias{as_ord}
\alias{as_uno}
\title{Coerce to base mode or \code{\link[mmm]{extended mode}}}
\usage{
as_mmm(x, mode, levs = NULL, na = FALSE)

as_chr(x, na = FALSE)

as_clr(x, na = FALSE)

as_fun(x)

as_int(x)

as_num(x)

as_lgl(x)

as_ord(x, levs)

as_uno(x, levs)
}
\arguments{
\item{x}{For \code{as_clr}, an object of mode character; for \code{as_fun}, a character scalar function name or a function object; for \code{as_ord} and \code{as_uno}, an atomic object; and for all others, any R object.}

\item{mode}{A \link[mmm]{threechar scalar} containing an \link[mmm]{extended mode} value.}

\item{levs}{Optional \link[=cmp_vec]{complete atomic vec} of factor levels (ordered factor levels for \code{as_ord}).}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values are acceptable.}
}
\value{
\strong{An object of base mode} \code{'character'}                                 \cr\cr \code{as_chr}
\cr\cr  \strong{An object of base mode} \code{'integer'}                                   \cr\cr \code{as_int}
\cr\cr  \strong{An object of base mode} \code{'logical'}                                   \cr\cr \code{as_lgl}
\cr\cr  \strong{An object of base mode} \code{'numeric'}                                   \cr\cr \code{as_num}
\cr\cr  \strong{An object of \link[=mode]{extended mode}} \code{'ord'} (ordered factor)    \cr\cr \code{as_ord}
\cr\cr  \strong{An object of extended mode} \code{'uno'} (unordered factor)                \cr\cr \code{as_uno}
\cr\cr  \strong{An object of extended mode} \code{'clr'} (hex color in form \code{'#RRGGBBAA'}) \cr\cr \code{as_clr}
}
\description{
Functions to coerce objects to \link[base:mode]{base mode} or \code{\link[=mode]{extended mode}}.
}
\details{
Functions in this family are as follows:
\tabular{ll}{  \code{as_chr}   \tab Thinly wraps \code{\link[base]{as.character}}.                                                          \cr
  \code{as_int}   \tab Thinly wraps \code{\link[base]{as.integer}}.                                                            \cr
  \code{as_num}   \tab Thinly wraps \code{\link[base]{as.numeric}}.                                                            \cr
  \code{as_lgl}   \tab Thinly wraps \code{\link[base]{as.logical}}.                                                            \cr   \tab   \cr
  \code{as_ord}   \tab As extended mode \code{'ord'} (ordered factor). Wraps \code{factor(x, levels = levs, ordered = TRUE)}.            \cr   \tab   \cr
  \code{as_uno}   \tab As extended mode \code{'uno'} (unordered factor). Wraps \code{factor(x, levels = levs, ordered = FALSE)}.         \cr   \tab   \cr
  \code{as_clr}   \tab Coerces character color values to hexadecimal.                                                          \cr   \tab   \cr
  \code{as_fun}   \tab If \code{x} is a character scalar function return that function, but if \code{x} is a function object, return it.                }
}
\examples{
bins. <- sample(c(0, 1), 10, replace = T)
chrs. <- c("3.14", "2.72", "1.41")
clrs. <- c("red", "#AABBCC", "#AABBCCDD", "blue")
nums. <- c(pi, exp(1), sqrt(2))

as_fun(unique)
as_fun("unique")

bins.
as_lgl(bins.)

chrs.
as_num(chrs.)
as_int(chrs.)

clrs.
as_clr(clrs.)

nums.
as_chr(nums.)

as_uno(clrs., levs = unique(clrs.))
as_ord(clrs., levs = sort(unique(clrs.)))
}
\seealso{
Other properties: 
\code{\link{bbb_ccc_help}()},
\code{\link{bbb_help}()},
\code{\link{bbb_mmm_prop_funs}()},
\code{\link{ccc_help}()},
\code{\link{cls}()},
\code{\link{cmp_ccc_help}()},
\code{\link{cmp_mmm_ccc_help}()},
\code{\link{cmp_mmm_help}()},
\code{\link{compatible}()},
\code{\link{ddd_prop_funs}()},
\code{\link{eee_help}()},
\code{\link{iii_help}()},
\code{\link{meets}()},
\code{\link{mmm_ccc_help}()},
\code{\link{mmm_help}()},
\code{\link{ppp_fast_help}()},
\code{\link{ppp_help}()},
\code{\link{sss_ccc_help}()},
\code{\link{sss_prop_funs}()},
\code{\link{unq_ccc_prop_funs}()},
\code{\link{unq_mmm_ccc_help}()},
\code{\link{unq_mmm_help}()}
}
\concept{properties}
