% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xxx.R
\name{as_xxx}
\alias{as_xxx}
\alias{as_color}
\alias{as_fun}
\title{Extended \code{as} Property (xxx) Functions}
\usage{
as_color(x, na = F)

as_fun(x)
}
\arguments{
\item{x}{An object}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values
qualify as missing color representations.}
}
\description{
Extended \code{as} Property (xxx) Functions
}
\details{
\strong{\code{as_color}}
\cr Coerce a valid ℝ color representation to hexadecimal RGB
character representation.
\cr\cr
\strong{\code{as_fun}}
\cr Return \code{x} if it is a function, otherwise, search for a function
with the name contained in \code{x} and return that function.
}
\seealso{
Other meta: 
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
