% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\encoding{UTF-8}
\name{atomize}
\alias{atomize}
\alias{atoms}
\alias{av}
\alias{a}
\alias{sav}
\title{Atomize (reduce arguments to an atomic vector)}
\usage{
atomize(...)

atoms(...)

av(...)

a(...)

sav(...)
}
\arguments{
\item{...}{Arguments to be atomized.}
}
\value{
A single atomic vector containing all constituent atomic values in the full set of arguments supplied in \code{...}.
}
\description{
\tabular{rl}{
  \verb{atomize, a}   \tab Gets a vector of all atomic
\cr \verb{atoms, av}   \tab values in all \code{...} arguments.
\cr               \tab   
\cr       \code{sav}   \tab Gets \code{sort(a(...))}.
}
}
\examples{
x <- list("a", "b", "c")
y <- data.frame(1:3)

x
y

a(x)
av(y)
atoms(x, y)
atomize(x, "d", y, 4)
sav(x, "d", y, 4)
}
\seealso{
Other extensions: 
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()},
\code{\link{xat}()}

Other values: 
\code{\link{n_th}()},
\code{\link{na}()},
\code{\link{uv}()},
\code{\link{values}()},
\code{\link{wv}()}
}
\concept{extensions}
\concept{values}
