% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize}
\alias{atomize}
\alias{a}
\alias{av}
\title{Atomize to a Vector of Atomic Values}
\usage{
a(...)

av(...)
}
\arguments{
\item{...}{Arguments to be atomized.}
}
\value{
A single atomic vector containing all constituent atomic values in
the full set of arguments supplied in \code{...}.
}
\description{
Reduce all \code{...} arguments to a single atomic vector of
atomic elements without any additional attributes.
}
\examples{
Fruits <- c("apples" , "bananas", "oranges")
Digits <- c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9)
Both   <- list(Fruits, Digits)
av(Fruits)
av(Digits)
av(Both)
av("celery", "beets", "lettuce")
av("celery", "beets", "lettuce")
}
\seealso{
Other extensions: 
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
