% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize}
\alias{atomize}
\alias{atoms}
\alias{av}
\alias{a}
\alias{sav}
\title{Atomize (reduce arguments to an atomic vector)}
\usage{
atomize(...)

atoms(...)

av(...)

a(...)

sav(...)
}
\arguments{
\item{...}{Arguments to be atomized.}
}
\value{
A single atomic vector containing all constituent atomic values in the full set of arguments supplied in \code{...}.
}
\description{
\tabular{rl}{
\code{atomize}   \tab Gets a vector of all
\cr \code{atoms}   \tab atomic values in all
\cr    \code{av}   \tab \code{...} arguments.
\cr     \code{a}   \tab   
\cr           \tab   
\cr   \code{sav}   \tab Gets a sorted vector of
\cr           \tab all atomic values in all
\cr           \tab \code{...} arguments.
}
}
\examples{
x <- list("a", "b", "c")
y <- data.frame(1:3)

x
y

a(x)
av(y)
atoms(x, y)
atomize(x, "d", y, 4)
sav(x, "d", y, 4)
}
