% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats.R
\name{ats}
\alias{ats}
\alias{at}
\alias{at_names}
\alias{add_at}
\alias{xat}
\title{Manage attributes}
\usage{
at(x, ..., a = FALSE)

at_names(x)

add_at(x, ...)

xat(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Optional unquoted names of attributes of \code{x} or character objects. When none are provided, indicates all attributes of \code{x}.}

\item{a}{\code{TRUE} or \code{FALSE} indicating whether to \link[av]{atomize} the result and convert the resulting vector to character.}
}
\value{
\emph{An R object} \cr   \verb{add_ats, xats, ats}
\cr\cr \code{NULL} \emph{or a character vector} \cr   \code{at_names}
}
\description{
\tabular{rl}{
  \code{at_names}   \tab Thinly wraps \code{names(attributes(x))}.
\cr  \code{add_at}   \tab Adds named \code{...} args to attributes of \code{x}.
\cr     \code{xat}   \tab Removes attributes of \code{x} with names matching unquoted \code{...} args\eqn{^1}.
\cr      \code{at}   \tab Gets attributes of \code{x} with names matching unquoted \code{...} args\eqn{^1}.
}
\verb{ }\eqn{^{1.}} No \code{...} args indicates all attributes of \code{x}.
}
