% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb.R
\name{bbb}
\alias{bbb}
\alias{bbb_props}
\alias{is_bbb_spec}
\alias{ibbb}
\alias{iatm}
\alias{idef}
\alias{ifun}
\alias{inil}
\alias{inll}
\alias{ipop}
\alias{ircr}
\title{Base Properties}
\usage{
bbb(x)

bbb_props()

is_bbb_spec(spec)

ibbb(x, spec, ...)

iatm(x, ...)

idef(x, ...)

ifun(x, ...)

inil(x, ...)

inll(x, ...)

ipop(x, ...)

ircr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec}
containing one or more base properties (i.e., from
\code{bbb_vals()}). \strong{NOTE}: properties may be pipe-separated. If
If there are multiple properties in \code{spec}, \code{x} is inspected for
a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{\code{bbb_vals}}: A character vector.
\cr\cr\strong{\code{bbb}}: A character scalar or character vector.
\cr\cr\strong{\code{iBBB, ibbb, is_bbb_spec}}: A logical scalar.
}
\description{
NOTE: \code{BBB} is used as a wildcard representing any given
based property.
\cr\cr
An object's base properties are defined by its most basic
structural properties as described in the following table: \tabular{lll}{
PROPERTY     \tab PROPERTY   \tab QUALIFYING                          \cr
NAME         \tab VALUE      \tab OBJECTS                             \cr
atomic       \tab\code{'atm'}\tab Atomic objects.                     \cr
defined      \tab\code{'def'}\tab Non-\code{NULL} objects.            \cr
function     \tab\code{'fun'}\tab Function objects or character scalars
containing a valid function name
accessible to the calling function. \cr
nil          \tab\code{'nil'}\tab Objects of nil (or zero) length.    \cr
null         \tab\code{'nll'}\tab The \code{NULL} object.             \cr
populated    \tab\code{'pop'}\tab Objects of length 1 or greater.     \cr
recursive    \tab\code{'rcr'}\tab Recursive objects (data.frames and
lists).                               }
Functions related to base properties are as follows:
\tabular{ll}{
BBB FUNCTION         \tab WHAT IT DOES                                \cr
\code{iBBB}          \tab Evaluates whether \code{x} matches the
base property \code{BBB}
(subject to any restrictions in \code{...}).\cr
\code{bbb}           \tab Gets a character vector containing all
base properties matching
\code{x}.                                   \cr
\code{ibbb}          \tab Evaluates \code{x} against the
base property specification
in \code{spec} (subject to any restrictions in
\code{...}).                                \cr
\code{bbb_props}     \tab Gets a character vector of all possible
base property values.                       \cr
\code{is_bbb_spec}   \tab Evaluates whether \code{spec} is a valid
base property
specification.                                }
}
\section{Specifying Count and Value Restrictions}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar less-than bound on values of \code{x}.        \cr
\code{le}     \tab Scalar less-than-or-equal bound on values of
\code{x}.                                            \cr
\code{ge}     \tab Scalar greater-than-or-equal bound on values of
\code{x}.                                            \cr
\code{gt}     \tab Scalar greater-than bound on values of \code{x}.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc_props}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
