% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb.R
\name{bbb}
\alias{bbb}
\alias{bbb_props}
\alias{is_bbb_spec}
\alias{ibbb}
\alias{iatm}
\alias{idef}
\alias{ifun}
\alias{inil}
\alias{inll}
\alias{ipop}
\alias{ircr}
\title{Base properties}
\usage{
bbb(x)

bbb_props()

is_bbb_spec(spec)

ibbb(x, spec, ...)

iatm(x, ...)

idef(x, ...)

ifun(x, ...)

inil(x, ...)

inll(x, ...)

ipop(x, ...)

ircr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more base properties (i.e., from \code{bbb_props()}). Base properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\itemize{
\item \strong{\code{bbb}}: a character scalar/vector.
\item \strong{\code{bbb_vals}}: a character vector.
\item \strong{\verb{ibbb, ixxx, is_bbb_spec}}: a logical scalar.
}
}
\description{
An object's base properties are defined by its most basic structural properties as described in the following table:
\itemize{
\item \strong{\code{'atm'}} (atomic): atomic objects.
\item \strong{\code{'def'}} (defined): non-\code{NULL} objects.
\item \strong{\code{'fun'}} (function): function objects or character scalar names of functions accessible to the calling function.
\item \strong{\code{'nil'}} (nil): objects of nil length (i.e., \code{0}-length).
\item \strong{\code{'nll'}} (null): the \code{NULL} object.
\item \strong{\code{'pop'}} (populated): objects of length \code{1} or greater.
\item \strong{\code{'rcr'}} (recursive): data.frames and lists.
}
\strong{Functions}
\itemize{
\item \strong{\code{bbb}}: gets a character vector containing all base properties possessed by \code{x}.
\item \strong{\code{ixxx}}: evaluates whether \code{x} possesses the base property \code{xxx} (a placeholder for any given base property value), subject to any restrictions in \code{...}.
\item \strong{\code{ibbb}}: evaluates whether \code{x} possesses one or more (possibly pipe-delimited) base properties in \code{spec}, subject to any restrictions in \code{...}.
\item \strong{\code{bbbs_props}}: gets a character vector of all possible base property values.
\item \strong{\code{is_bbb_spec}}: evaluates whether \code{spec} is a valid base property specification.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\itemize{
\item \strong{\code{n}}: a vector of valid lengths/numbers of elements.
\item \strong{\code{nr}}: a vector of valid numbers of rows.
\item \strong{\code{nc}}: a vector of valid numbers of columns.
\item \strong{\code{min}}: a scalar minimum valid length/number of element.
\item \strong{\code{minr}}: a scalar minimum valid number of rows.
\item \strong{\code{minc}}: a scalar minimum valid number of columns.
\item \strong{\code{max}}: a scalar maximum valid length/number of element.
\item \strong{\code{maxr}}: a scalar maximum valid number of rows.
\item \strong{\code{maxc}}: a scalar maximum valid number of columns.
\item \strong{\code{vals}}: a vector of valid values.
\item \strong{\code{lt}}: a scalar less-than (exclusive upper) bound.
\item \strong{\code{le}}: a scalar less-than-or-equal (inclusive upper) bound
\item \strong{\code{ge}}: a scalar greater-than-or-equal (inclusive lower) bound.
\item \strong{\code{gt}}: a scalar greater-than bound (exclusive lower) bound.
}
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
