% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb.R
\encoding{UTF-8}
\name{bbb}
\alias{bbb}
\alias{bbb_funs}
\alias{bbb_props}
\alias{is_bbb_spec}
\alias{BBB}
\alias{ATM}
\alias{DEF}
\alias{FUN}
\alias{NIL}
\alias{NLL}
\alias{POP}
\alias{RCR}
\title{Basic properties}
\usage{
bbb(x)

bbb_funs()

bbb_props()

is_bbb_spec(spec)

BBB(x, spec, ...)

ATM(x, ...)

DEF(x, ...)

FUN(x, ...)

NIL(x)

NLL(x)

POP(x, ...)

RCR(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more basic properties from \code{bbb_props()}. basic properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \verb{bbb_props, bbb_funs, bbb}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{is_bbb_spec, \{BBB\}, BBB}
}
\description{
An object's basic properties are defined by its fundamental structure as follows:
\tabular{lll}{  \verb{'atm', 'ATM'}   \tab \code{atomic}      \tab Atomic (but not \code{NULL}). Both a basic and an \link[=mmm]{xmode} property.                                    \cr   \tab   \cr
  \verb{'fun', 'FUN'}   \tab \code{function}    \tab A function or a character scalar name of a function accessible from the environment of the calling function. \cr   \tab   \cr
  \verb{'rcr', 'RCR'}   \tab \code{recursive}   \tab data.frame or list.                                                                                                       \cr
  \verb{'pop', 'POP'}   \tab \code{populated}   \tab Length \code{1} or greater.                                                                                                    \cr
  \verb{'def', 'DEF'}   \tab \code{defined}     \tab Not \code{NULL}.                                                                                                               \cr
  \verb{'nll', 'NLL'}   \tab \code{null}        \tab \code{NULL}.                                                                                                                   \cr
  \verb{'nil', 'NIL'}   \tab \code{nil}         \tab Length \code{0} but not \code{NULL}.                                                                                                  }
}
\details{
\tabular{ll}{  \code{is_bbb_spec}   \tab Is \code{spec} a basic property specification?                                                                   \cr   \tab   \cr
  \code{bbb_props}     \tab What basic properties are there?                                                                            \cr   \tab   \cr
  \code{bbb_funs}      \tab What basic properties functions are there?                                                                  \cr   \tab   \cr
  \code{{BBB}}         \tab Is \code{x} a match to the basic property \code{'{BBB}'} where \code{{BBB}} is a placeholder for any given basic property? \cr   \tab   \cr
  \code{BBB}           \tab Is \code{x} a match to the basic property specification \code{spec}?                                                  \cr   \tab   \cr
  \code{bbb}           \tab What are \code{x}'s basic properties?                                                                                           }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
bbb_funs()
bbb_props()
is_bbb_spec("nil|nll")
BBB(NULL, "nil|nll")
POP(NA)
ATM(list(letters))
bbb(NA)
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
