% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_ccc.R
\name{bbb_ccc}
\alias{bbb_ccc}
\alias{bbb_ccc_props}
\alias{atm_arr}
\alias{atm_dtf}
\alias{atm_gen}
\alias{atm_mat}
\alias{atm_mvc}
\alias{atm_scl}
\alias{atm_vec}
\alias{atm_vls}
\alias{nil_arr}
\alias{nil_dtf}
\alias{nil_gen}
\alias{nil_mat}
\alias{nil_mvc}
\alias{nil_scl}
\alias{nil_vec}
\alias{nil_vls}
\alias{pop_arr}
\alias{pop_dtf}
\alias{pop_gen}
\alias{pop_mat}
\alias{pop_mvc}
\alias{pop_scl}
\alias{pop_vec}
\alias{pop_vls}
\alias{rcr_arr}
\alias{rcr_dtf}
\alias{rcr_gen}
\alias{rcr_mat}
\alias{rcr_mvc}
\alias{rcr_scl}
\alias{rcr_vec}
\alias{rcr_vls}
\title{Base + Extended Class Properties}
\usage{
bbb_ccc(x, bbb, ccc, ...)

bbb_ccc_props()

atm_arr(x, ...)

atm_dtf(x, ...)

atm_gen(x, ...)

atm_mat(x, ...)

atm_mvc(x, ...)

atm_scl(x, ...)

atm_vec(x, ...)

atm_vls(x, ...)

nil_arr(x, ...)

nil_dtf(x, ...)

nil_gen(x, ...)

nil_mat(x, ...)

nil_mvc(x, ...)

nil_scl(x, ...)

nil_vec(x, ...)

nil_vls(x, ...)

pop_arr(x, ...)

pop_dtf(x, ...)

pop_gen(x, ...)

pop_mat(x, ...)

pop_mvc(x, ...)

pop_scl(x, ...)

pop_vec(x, ...)

pop_vls(x, ...)

rcr_arr(x, ...)

rcr_dtf(x, ...)

rcr_gen(x, ...)

rcr_mat(x, ...)

rcr_mvc(x, ...)

rcr_scl(x, ...)

rcr_vec(x, ...)

rcr_vls(x, ...)
}
\arguments{
\item{x}{An R object}

\item{bbb}{A character scalar containing an
base property from bbb_props().}

\item{ccc}{A character scalar containing an
extended class property from ccc_props().}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{\code{bbb_ccc_props}}
\cr A character vector.
\cr\cr
\strong{\code{bbb_ccc, BBB_CCC}}
\cr A logical scalar.
}
\description{
Base + Extended Class Properties
}
\section{Functions in this Family}{

NOTE: \code{BBB} and \code{CCC} are used as placeholders for any
given base and extended class, respectively.
\cr\cr
\strong{\code{bbb_ccc}}
\cr Evaluates whether \code{x} matches the base property specified in the
argument \code{bbb} and matches the extended class specified in the
argument \code{ccc}. (subject to any restrictions in \code{...}).
\cr\cr
\strong{\code{BBB_CCC}}
\cr Evaluates whether \code{x} matches base property \code{BBB} and
extended class \code{CCC}. (subject to any restrictions in \code{...}).
\cr\cr
\strong{\code{bbb_ccc_props}}
\cr Gets a character vector of all possible base + extended class
properties.
}

\section{Specifying Count and Value Restrictions}{
 Specifying additional
requirements in \code{...} is optional. The full set of recognized
arguments names are defined in the following table along with the
properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                      \cr
\code{n}      \tab Vector of valid lengths/numbers of elements.           \cr
\code{nr}     \tab Vector of valid numbers of rows.                       \cr
\code{nc}     \tab Vector of valid numbers of columns.                    \cr
\code{min}    \tab Scalar minimum valid length/number of element.         \cr
\code{minr}   \tab Scalar minimum valid number of rows.                   \cr
\code{minc}   \tab Scalar minimum valid number of columns.                \cr
\code{max}    \tab Scalar maximum valid length/number of element.         \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                   \cr
\code{maxc}   \tab Scalar maximum valid number of columns.                \cr
\code{vals}   \tab Vector of valid values.                                \cr
\code{lt}     \tab Scalar less-than (exclusive upper) bound.              \cr
\code{le}     \tab Scalar less-than-or-equal (inclusive upper) bound      \cr
\code{ge}     \tab Scalar greater-than-or-equal (inclusive lower) bound.  \cr
\code{gt}     \tab Scalar greater-than bound (exclusive lower) bound.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
