% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_ccc.R, R/sss_ccc.R
\encoding{UTF-8}
\name{bbb_ccc}
\alias{bbb_ccc}
\alias{bbb_ccc_funs}
\alias{atm_arr}
\alias{atm_dtf}
\alias{atm_gen}
\alias{atm_mat}
\alias{atm_mvc}
\alias{atm_scl}
\alias{atm_vec}
\alias{atm_vls}
\alias{nil_arr}
\alias{nil_dtf}
\alias{nil_gen}
\alias{nil_mat}
\alias{nil_vec}
\alias{nil_vls}
\alias{pop_arr}
\alias{pop_dtf}
\alias{pop_gen}
\alias{pop_mat}
\alias{pop_mvc}
\alias{pop_scl}
\alias{pop_vec}
\alias{pop_vls}
\alias{sss_ccc_funs}
\alias{emp_arr}
\alias{emp_dtf}
\alias{emp_gen}
\alias{emp_mat}
\alias{emp_vec}
\alias{emp_vls}
\alias{pnt_arr}
\alias{pnt_dtf}
\alias{pnt_gen}
\alias{pnt_mat}
\alias{pnt_scl}
\alias{pnt_vec}
\alias{pnt_vls}
\alias{lin_arr}
\alias{lin_dtf}
\alias{lin_gen}
\alias{lin_mat}
\alias{lin_mvc}
\alias{lin_vec}
\alias{lin_vls}
\alias{col_dtf}
\alias{col_mat}
\alias{row_dtf}
\alias{row_mat}
\alias{rct_dtf}
\alias{rct_mat}
\alias{sqr_mat}
\alias{sld_arr}
\title{basic + xclass combination properties}
\usage{
bbb_ccc(x, bbb, ccc, ...)

bbb_ccc_funs()

atm_arr(x, ...)

atm_dtf(x, ...)

atm_gen(x, ...)

atm_mat(x, ...)

atm_mvc(x, ...)

atm_scl(x, ...)

atm_vec(x, ...)

atm_vls(x, ...)

nil_arr(x, ...)

nil_dtf(x, ...)

nil_gen(x, ...)

nil_mat(x, ...)

nil_vec(x, ...)

nil_vls(x, ...)

pop_arr(x, ...)

pop_dtf(x, ...)

pop_gen(x, ...)

pop_mat(x, ...)

pop_mvc(x, ...)

pop_scl(x, ...)

pop_vec(x, ...)

pop_vls(x, ...)

sss_ccc_funs()

emp_arr(x, ...)

emp_dtf(x, ...)

emp_gen(x, ...)

emp_mat(x, ...)

emp_vec(x, ...)

emp_vls(x, ...)

pnt_arr(x, ...)

pnt_dtf(x, ...)

pnt_gen(x, ...)

pnt_mat(x, ...)

pnt_scl(x, ...)

pnt_vec(x, ...)

pnt_vls(x, ...)

lin_arr(x, ...)

lin_dtf(x, ...)

lin_gen(x, ...)

lin_mat(x, ...)

lin_mvc(x, ...)

lin_vec(x, ...)

lin_vls(x, ...)

col_dtf(x, ...)

col_mat(x, ...)

row_dtf(x, ...)

row_mat(x, ...)

rct_dtf(x, ...)

rct_mat(x, ...)

sqr_mat(x, ...)

sld_arr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{bbb}{A character scalar single basic property from \code{\link{bbb_props}()}.}

\item{ccc}{A character scalar single xclass property from \code{\link{ccc_props}()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector} \cr   \code{bbb_ccc_funs}
\cr\cr \emph{A logical scalar} \cr   \verb{BBB_CCC, bbb_ccc}
}
\description{
\tabular{rl}{
  \code{bbb_ccc_funs}   \tab What \link[=bbb]{basic} + \link[=ccc]{xclass} combination \link[=prop_funs]{property functions} are there?
\cr    \code{bbb_ccc}   \tab Is \code{x} a match to the single basic and xclass properties in \code{bbb} and \code{ccc}, respectively?
\cr    \code{BBB_CCC}   \tab Is \code{x} a match to single basic and xclass properties \code{'BBB'} and \code{'CCC'}, respectively?
}
Some combinations of basic + xclass properties are non-sensical. For this reason, the basic properties represented in this family of functions are \code{c('atm', 'nil',  'pop')}.
\cr\cr In addition, the base property \code{'nil'} is nonsensical in combination with xclasses \code{'mvc'} and \code{'scl'} (which thus do not have combined \code{BBB_CCC} property functions).
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
bbb_ccc_funs()
bbb_ccc_props()
bbb_ccc(letters, "atm", "mvc")
bbb_ccc(1, "nil", "vec")
atm_gen(letters)
atm_scl(1)
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{sss_ccc}()},
\code{\link{unq_mmm_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
