% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_ccc.R
\name{bbb_ccc}
\alias{bbb_ccc}
\alias{bbb_ccc_props}
\alias{atm_arr}
\alias{atm_dtf}
\alias{atm_gen}
\alias{atm_mat}
\alias{atm_mvc}
\alias{atm_scl}
\alias{atm_vec}
\alias{atm_vls}
\alias{nil_arr}
\alias{nil_dtf}
\alias{nil_gen}
\alias{nil_mat}
\alias{nil_mvc}
\alias{nil_scl}
\alias{nil_vec}
\alias{nil_vls}
\alias{pop_arr}
\alias{pop_dtf}
\alias{pop_gen}
\alias{pop_mat}
\alias{pop_mvc}
\alias{pop_scl}
\alias{pop_vec}
\alias{pop_vls}
\alias{rcr_arr}
\alias{rcr_dtf}
\alias{rcr_gen}
\alias{rcr_mat}
\alias{rcr_mvc}
\alias{rcr_scl}
\alias{rcr_vec}
\alias{rcr_vls}
\title{Base + extended class properties}
\usage{
bbb_ccc(x, bbb, ccc, ...)

bbb_ccc_props()

atm_arr(x, ...)

atm_dtf(x, ...)

atm_gen(x, ...)

atm_mat(x, ...)

atm_mvc(x, ...)

atm_scl(x, ...)

atm_vec(x, ...)

atm_vls(x, ...)

nil_arr(x, ...)

nil_dtf(x, ...)

nil_gen(x, ...)

nil_mat(x, ...)

nil_mvc(x, ...)

nil_scl(x, ...)

nil_vec(x, ...)

nil_vls(x, ...)

pop_arr(x, ...)

pop_dtf(x, ...)

pop_gen(x, ...)

pop_mat(x, ...)

pop_mvc(x, ...)

pop_scl(x, ...)

pop_vec(x, ...)

pop_vls(x, ...)

rcr_arr(x, ...)

rcr_dtf(x, ...)

rcr_gen(x, ...)

rcr_mat(x, ...)

rcr_mvc(x, ...)

rcr_scl(x, ...)

rcr_vec(x, ...)

rcr_vls(x, ...)
}
\arguments{
\item{x}{An R object}

\item{bbb}{A character scalar containing a base property from \code{bbb_props()}.}

\item{ccc}{A character scalar containing an extended class property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\itemize{
\item \strong{\code{ccc_mmm_props}}: a character vector.
\item \strong{\verb{ccc_mmm, xxx_yyy}}: a logical scalar.
}
}
\description{
\itemize{
\item \strong{\code{bbb_ccc}}: evaluates whether \code{x} matches the base property specified in the argument \code{ccc} and matches the extended class specified in the argument \code{ccc} subject to any restrictions in \code{...}.
\item \strong{\code{xxx_yyy}}: evaluates whether \code{x} matches base property \code{xxx} and extended class \code{yyy} subject to any restrictions in \code{...} where \code{xxx} and \code{yyy} are placeholders for any given base property and any given extended class property, respectively.
\item \strong{\code{bbb_ccc_props}}: gets a character vector of all possible base + extended class properties.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\itemize{
\item \strong{\code{n}}: a vector of valid lengths/numbers of elements.
\item \strong{\code{nr}}: a vector of valid numbers of rows.
\item \strong{\code{nc}}: a vector of valid numbers of columns.
\item \strong{\code{min}}: a scalar minimum valid length/number of element.
\item \strong{\code{minr}}: a scalar minimum valid number of rows.
\item \strong{\code{minc}}: a scalar minimum valid number of columns.
\item \strong{\code{max}}: a scalar maximum valid length/number of element.
\item \strong{\code{maxr}}: a scalar maximum valid number of rows.
\item \strong{\code{maxc}}: a scalar maximum valid number of columns.
\item \strong{\code{vals}}: a vector of valid values.
\item \strong{\code{lt}}: a scalar less-than (exclusive upper) bound.
\item \strong{\code{le}}: a scalar less-than-or-equal (inclusive upper) bound
\item \strong{\code{ge}}: a scalar greater-than-or-equal (inclusive lower) bound.
\item \strong{\code{gt}}: a scalar greater-than bound (exclusive lower) bound.
}
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
