% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_ccc.R
\name{bbb_ccc}
\alias{bbb_ccc}
\alias{bbb_ccc_props}
\alias{atm_arr}
\alias{atm_dtf}
\alias{atm_gen}
\alias{atm_mat}
\alias{atm_mvc}
\alias{atm_scl}
\alias{atm_vec}
\alias{atm_vls}
\alias{nil_arr}
\alias{nil_dtf}
\alias{nil_gen}
\alias{nil_mat}
\alias{nil_mvc}
\alias{nil_scl}
\alias{nil_vec}
\alias{nil_vls}
\alias{pop_arr}
\alias{pop_dtf}
\alias{pop_gen}
\alias{pop_mat}
\alias{pop_mvc}
\alias{pop_scl}
\alias{pop_vec}
\alias{pop_vls}
\alias{rcr_arr}
\alias{rcr_dtf}
\alias{rcr_gen}
\alias{rcr_mat}
\alias{rcr_mvc}
\alias{rcr_scl}
\alias{rcr_vec}
\alias{rcr_vls}
\title{Base + xclass properties}
\usage{
bbb_ccc(x, bbb, ccc, ...)

bbb_ccc_props()

atm_arr(x, ...)

atm_dtf(x, ...)

atm_gen(x, ...)

atm_mat(x, ...)

atm_mvc(x, ...)

atm_scl(x, ...)

atm_vec(x, ...)

atm_vls(x, ...)

nil_arr(x, ...)

nil_dtf(x, ...)

nil_gen(x, ...)

nil_mat(x, ...)

nil_mvc(x, ...)

nil_scl(x, ...)

nil_vec(x, ...)

nil_vls(x, ...)

pop_arr(x, ...)

pop_dtf(x, ...)

pop_gen(x, ...)

pop_mat(x, ...)

pop_mvc(x, ...)

pop_scl(x, ...)

pop_vec(x, ...)

pop_vls(x, ...)

rcr_arr(x, ...)

rcr_dtf(x, ...)

rcr_gen(x, ...)

rcr_mat(x, ...)

rcr_mvc(x, ...)

rcr_scl(x, ...)

rcr_vec(x, ...)

rcr_vls(x, ...)
}
\arguments{
\item{x}{An R object}

\item{bbb}{A character scalar containing a \code{base} property from \code{bbb_props()}.}

\item{ccc}{A character scalar containing an \code{xclass} property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{bbb_ccc_props}   \tab A character vector.
\cr                 \tab   
\cr     \code{bbb_ccc}   \tab A logical scalar
\cr     \code{BBB_CCC}   \tab   
}
}
\description{
\tabular{rl}{
  \code{bbb_ccc_props}   \tab Gets a character vector of all possible \link[=bbb]{\code{base}} + \link[=ccc]{\code{xclass}} properties.
\cr                 \tab  
\cr     \code{bbb_ccc}   \tab Evaluates whether \code{x} matches \code{base} and \code{xclass} specs in \code{bbb} and \code{ccc}
\cr                 \tab  
\cr     \code{BBB_CCC}   \tab Evaluates whether \code{x} matches \code{base} and \code{xclass} properties \code{'BBB'} and \code{'CCC'}.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
bbb_ccc_props()
bbb_ccc(letters, "atm", "mvc")
bbb_ccc(1, "nil", "vec")
atm_gen(letters)
atm_scl(1)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
