% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_mmm.R
\encoding{UTF-8}
\name{bbb_mmm}
\alias{bbb_mmm}
\alias{bbb_mmm_funs}
\alias{atm_ch1}
\alias{atm_ch3}
\alias{atm_chr}
\alias{atm_clr}
\alias{atm_evn}
\alias{atm_fac}
\alias{atm_frc}
\alias{atm_ind}
\alias{atm_lgl}
\alias{atm_neg}
\alias{atm_ngw}
\alias{atm_nng}
\alias{atm_nnw}
\alias{atm_nps}
\alias{atm_npw}
\alias{atm_nst}
\alias{atm_num}
\alias{atm_odd}
\alias{atm_ord}
\alias{atm_pct}
\alias{atm_pos}
\alias{atm_ppn}
\alias{atm_psw}
\alias{atm_srt}
\alias{atm_str}
\alias{atm_uno}
\alias{atm_whl}
\alias{pop_atm}
\alias{pop_ch1}
\alias{pop_ch3}
\alias{pop_chr}
\alias{pop_clr}
\alias{pop_evn}
\alias{pop_fac}
\alias{pop_frc}
\alias{pop_ind}
\alias{pop_lgl}
\alias{pop_neg}
\alias{pop_ngw}
\alias{pop_nng}
\alias{pop_nnw}
\alias{pop_nps}
\alias{pop_npw}
\alias{pop_nst}
\alias{pop_num}
\alias{pop_odd}
\alias{pop_ord}
\alias{pop_pct}
\alias{pop_pos}
\alias{pop_ppn}
\alias{pop_psw}
\alias{pop_srt}
\alias{pop_str}
\alias{pop_uno}
\alias{pop_whl}
\title{basic + xmode combination properties}
\usage{
bbb_mmm(X, BBB, MMM, ...)

bbb_mmm_funs()

atm_ch1(X, ...)

atm_ch3(X, ...)

atm_chr(X, ...)

atm_clr(X, ...)

atm_evn(X, ...)

atm_fac(X, ...)

atm_frc(X, ...)

atm_ind(X, ...)

atm_lgl(X, ...)

atm_neg(X, ...)

atm_ngw(X, ...)

atm_nng(X, ...)

atm_nnw(X, ...)

atm_nps(X, ...)

atm_npw(X, ...)

atm_nst(X, ...)

atm_num(X, ...)

atm_odd(X, ...)

atm_ord(X, ...)

atm_pct(X, ...)

atm_pos(X, ...)

atm_ppn(X, ...)

atm_psw(X, ...)

atm_srt(X, ...)

atm_str(X, ...)

atm_uno(X, ...)

atm_whl(X, ...)

pop_atm(X, ...)

pop_ch1(X, ...)

pop_ch3(X, ...)

pop_chr(X, ...)

pop_clr(X, ...)

pop_evn(X, ...)

pop_fac(X, ...)

pop_frc(X, ...)

pop_ind(X, ...)

pop_lgl(X, ...)

pop_neg(X, ...)

pop_ngw(X, ...)

pop_nng(X, ...)

pop_nnw(X, ...)

pop_nps(X, ...)

pop_npw(X, ...)

pop_nst(X, ...)

pop_num(X, ...)

pop_odd(X, ...)

pop_ord(X, ...)

pop_pct(X, ...)

pop_pos(X, ...)

pop_ppn(X, ...)

pop_psw(X, ...)

pop_srt(X, ...)

pop_str(X, ...)

pop_uno(X, ...)

pop_whl(X, ...)
}
\arguments{
\item{X}{An R object.}

\item{BBB}{A character scalar single basic property from \code{c('atm', 'pop')}.}

\item{MMM}{A character scalar single xmode property from \code{\link{mmm_props}()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \code{bbb_mmm_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{\{bbb\}_\{mmm\}, bbb_mmm}
}
\description{
Check for combinations of \link[=bbb]{basic} and \link[=mmm]{xmode} properties.
}
\details{
\tabular{ll}{  \code{bbb_mmm_funs}   \tab What \link[=bbb]{basic} + \link[=mmm]{xmode} combination property functions are there?                    \cr   \tab   \cr
  \verb{\{bbb\}_\{mmm\}}    \tab Is \code{X} a match to the single basic property \code{'{bbb}'} and single xmode property \code{'{mmm}'}? (where \code{{bbb}}
and \code{{mmm}} are placeholders for any given basic property and any given xmode property, respectively).    \cr   \tab   \cr
  \code{bbb_mmm}        \tab Is \code{X} a match to single basic and xmode properties in \code{BBB} and \code{MMM}, respectively?                                    }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.max, .maxr, .maxc}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.min, .minr, .minc}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.lt, .le, .ge, .gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.n, .nr, .nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.vals}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
bbb_mmm_funs()
bbb_mmm_props()
bbb_mmm(letters, "atm", "ch1")
atm_ch1(letters)
pop_psw(1:10)
pop_psw(0:10)

}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp_fast}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
