% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_mmm.R
\encoding{UTF-8}
\name{bbb_mmm}
\alias{bbb_mmm}
\alias{bbb_mmm_funs}
\alias{atm_ch1}
\alias{atm_ch3}
\alias{atm_chr}
\alias{atm_clr}
\alias{atm_evn}
\alias{atm_fac}
\alias{atm_frc}
\alias{atm_ind}
\alias{atm_lgl}
\alias{atm_neg}
\alias{atm_ngw}
\alias{atm_nng}
\alias{atm_nnw}
\alias{atm_nps}
\alias{atm_npw}
\alias{atm_nst}
\alias{atm_num}
\alias{atm_odd}
\alias{atm_ord}
\alias{atm_pct}
\alias{atm_pos}
\alias{atm_ppn}
\alias{atm_psw}
\alias{atm_srt}
\alias{atm_str}
\alias{atm_uno}
\alias{atm_whl}
\alias{pop_atm}
\alias{pop_ch1}
\alias{pop_ch3}
\alias{pop_chr}
\alias{pop_clr}
\alias{pop_evn}
\alias{pop_fac}
\alias{pop_frc}
\alias{pop_ind}
\alias{pop_lgl}
\alias{pop_neg}
\alias{pop_ngw}
\alias{pop_nng}
\alias{pop_nnw}
\alias{pop_nps}
\alias{pop_npw}
\alias{pop_nst}
\alias{pop_num}
\alias{pop_odd}
\alias{pop_ord}
\alias{pop_pct}
\alias{pop_pos}
\alias{pop_ppn}
\alias{pop_psw}
\alias{pop_srt}
\alias{pop_str}
\alias{pop_uno}
\alias{pop_whl}
\title{basic + xmode combination properties}
\usage{
bbb_mmm(x, bbb, mmm, ...)

bbb_mmm_funs()

atm_ch1(x, ...)

atm_ch3(x, ...)

atm_chr(x, ...)

atm_clr(x, ...)

atm_evn(x, ...)

atm_fac(x, ...)

atm_frc(x, ...)

atm_ind(x, ...)

atm_lgl(x, ...)

atm_neg(x, ...)

atm_ngw(x, ...)

atm_nng(x, ...)

atm_nnw(x, ...)

atm_nps(x, ...)

atm_npw(x, ...)

atm_nst(x, ...)

atm_num(x, ...)

atm_odd(x, ...)

atm_ord(x, ...)

atm_pct(x, ...)

atm_pos(x, ...)

atm_ppn(x, ...)

atm_psw(x, ...)

atm_srt(x, ...)

atm_str(x, ...)

atm_uno(x, ...)

atm_whl(x, ...)

pop_atm(x, ...)

pop_ch1(x, ...)

pop_ch3(x, ...)

pop_chr(x, ...)

pop_clr(x, ...)

pop_evn(x, ...)

pop_fac(x, ...)

pop_frc(x, ...)

pop_ind(x, ...)

pop_lgl(x, ...)

pop_neg(x, ...)

pop_ngw(x, ...)

pop_nng(x, ...)

pop_nnw(x, ...)

pop_nps(x, ...)

pop_npw(x, ...)

pop_nst(x, ...)

pop_num(x, ...)

pop_odd(x, ...)

pop_ord(x, ...)

pop_pct(x, ...)

pop_pos(x, ...)

pop_ppn(x, ...)

pop_psw(x, ...)

pop_srt(x, ...)

pop_str(x, ...)

pop_uno(x, ...)

pop_whl(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{bbb}{A character scalar single basic property from \code{c('atm', 'pop')}.}

\item{mmm}{A character scalar single xmode property from \code{mmm_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector}
\cr    \code{bbb_mmm_funs}
\cr\cr \emph{A logical scalar}
\cr    \code{BBB_MMM}
\cr    \code{bbb_mmm}
}
\description{
\tabular{rl}{
  \code{bbb_mmm_funs}   \tab What \link[=bbb]{basic} + \link[=mmm]{xmode} combination property functions are there?
\cr                \tab  
\cr    \code{bbb_mmm}   \tab Is \code{x} a match to the single basic and xmode properties in \code{bbb} and \code{mmm}, respectively?
\cr                \tab  
\cr    \code{BBB_MMM}   \tab Is \code{x} a match to single basic and xmode properties \code{'BBB'} and \code{'MMM'}, respectively?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab  
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab  
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
bbb_mmm_funs()
bbb_mmm_props()
bbb_mmm(letters, "atm", "ch1")
atm_ch1(letters)
pop_psw(1:10)
pop_psw(0:10)
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unq}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{reclass}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
