% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfuns.R
\name{binfuns}
\alias{binfuns}
\alias{\%IS\%}
\alias{\%!IS\%}
\alias{\%EQ\%}
\alias{\%!EQ\%}
\alias{\%&\%}
\alias{\%|\%}
\alias{\%!\%}
\alias{\%1\%}
\alias{\%IN\%}
\alias{\%!IN\%}
\alias{\%HAS\%}
\alias{\%LACKS\%}
\title{Binary Functions Returning Only a Scalar \code{TRUE} or \code{FALSE}.}
\usage{
x \%IS\% y

x \%!IS\% y

x \%EQ\% y

x \%!EQ\% y

x \%&\% y

x \%|\% y

x \%!\% y

x \%1\% y

x \%IN\% y

x \%!IN\% y

x \%HAS\% y

x \%LACKS\% y
}
\arguments{
\item{x, y}{Any objects.}
}
\description{
Binary Functions Returning Only a Scalar \code{TRUE} or \code{FALSE}.
}
\details{
\strong{\code{x \%IS\% y}}
\cr Evaluates whether \code{x} and \code{y} are identical.
\cr\cr
\strong{\code{x \%!IS\% y}}
\cr Evaluates whether \code{x} and \code{y} are not identical.
\cr\cr
\strong{\code{x \%EQ\% y}}
\cr Evaluates whether \code{x} and \code{y} are \code{\link{set.equal}}. If
\code{x} and \code{y} are not \link[=compatible]{compatible}, returns
\code{FALSE}.
\cr\cr
\strong{\code{x \%!EQ\% y}}
\cr Evaluates whether \code{x} and \code{y} are not
\code{\link{set.equal}}. If \code{x} and \code{y} are not
\link[=compatible]{compatible}, returns \code{TRUE}.
\cr\cr
\strong{\code{x \%&\% y}}
\cr Evaluates whether \code{x} and \code{y} are both \code{TRUE} (that is,
a scalar \code{TRUE} value).
\cr\cr
\strong{\code{x \%|\% y}}
\cr Evaluates whether \code{x} and/or \code{y} is \code{TRUE}.
\cr\cr
\strong{\code{x \%!\% y}}
\cr Evaluates whether \code{x} and \code{y} are both \code{FALSE} (that is,
a scalar \code{FALSE} value).
\cr\cr
\strong{\code{x \%1\% y}}
\cr Evaluates whether one of \code{x} or \code{y} is \code{TRUE}.
\cr\cr
\strong{\code{x \%IN\% y}}
\cr Evaluates whether atomic scalar \code{x} is contained in atomic object
\code{y}. If \code{x} is not atomic scalar, \code{y} is not atomic, or
\code{x} and \code{y} are not compatible, returns \code{FALSE}.
\cr\cr
\strong{\code{x \%!IN\% y}}
\cr Evaluates whether atomic scalar \code{x} is contained in atomic object
\code{y}. If \code{x} is not atomic scalar, \code{y} is not atomic, or
\code{x} and \code{y} are not compatible, returns \code{FALSE}.
\cr\cr
\strong{\code{x \%HAS\% y}}
\cr Evaluates whether atomic object \code{x} contains atomic scalar
\code{y}. If \code{x} is not atomic, \code{y} is not atomic scalar, or
\code{x} and \code{y} are not compatible, returns \code{FALSE}.
\cr\cr
\strong{\code{x \%LACKS\% y}}
\cr Evaluates whether atomic object \code{x} lacks (does not contain)
atomic scalar \code{y}. If \code{x} is not atomic, \code{y} is not atomic
scalar, or \code{x} and \code{y} are not compatible, returns
\code{TRUE}.
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
