% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Blank Strings}
\usage{
blank(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Logical scalar
}
\description{
Evaluates whether \code{x} is a blank string scalar (i.e.,
\code{""}).
}
\seealso{
Other strings: 
\code{\link{charn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
