% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Is an object a character scalar blank string.}
\usage{
blank(x)
}
\arguments{
\item{x}{An object}
}
\value{
Logical scalar
}
\description{
Is an object a character scalar blank string.
}
