% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callers.R
\name{callers}
\alias{callers}
\alias{caller}
\alias{ncallers}
\title{Functions in the Call Stack}
\usage{
callers(gens. = NULL)

caller()

ncallers()
}
\arguments{
\item{gens.}{The number of generations back in the function call stack to go.}
}
\value{
\code{callers} and \code{caller} return a character scalar or vector.
\code{ncallers} returns a non-negative whole-number scalar.
}
\description{
Gets the names of all calling functions, with the immediate
calling function in the first position and, if \code{gens.} is not
\code{NULL}, selects elements from the call stack specified by
\code{gens.}.
}
\section{Functions}{
\itemize{
\item \code{caller}: Get the immediate calling function.

\item \code{ncallers}: Get the number of calling functions in the call stack.
}}

\examples{
fun.a <- function() {A <- 0; fun.b()}
fun.b <- function() {B <- 1; fun.c()}
fun.c <- function() {
  cat("\n callstack(): " , paste0("'", callstack(), "'"))
  cat("\n caller(1)  : '", caller(1), "'")
  cat("\n caller(2)  : '", caller(2), "'")
  cat("\n exist('A', err = F, gens = 1): ", exist('A', err = F, gens = 1))
  cat("\n exist('A', err = F, gens = 2): ", exist('A', err = F, gens = 2))
  cat("\n exist('B', err = F, gens = 1): ", exist('B', err = F, gens = 1))
  cat("\n exist('B', err = F, gens = 2): ", exist('B', err = F, gens = 2))
  cat("\n vget('A', err = F, gens = 2) : ", vget('A', err = F, gens = 2))
  cat("\n vget('A', err = F, gens = 1) : ", vget('A', err = F, gens = 1))
  cat("\n vget('B', err = F, gens = 2) : ", vget('B', err = F, gens = 2))
  cat("\n vget('B', err = F, gens = 1) : ", vget('B', err = F, gens = 1))
  vSet('A', 'A', gens = 2)
  vSet('B', 'B', gens = 1)
  cat("\n vget('A', gens = 2) : ", vget('A', err = F, gens = 2))
  cat("\n vget('B', gens = 1) : ", vget('B', err = F, gens = 1))
}
fun.a()
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
