% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\encoding{UTF-8}
\name{cb}
\alias{cb}
\alias{rb}
\title{Error-checked row and column binding}
\usage{
cb(...)

rb(...)
}
\arguments{
\item{...}{Multiple \code{\link{compatible}} atomic matrices or multiple compatible \link[=atm_dtf]{atomic data.frames}.}
}
\value{
An \link[=atm_dtf]{atomic data.frame} or an atomic matrix.
}
\description{
Error-checked extensions of \code{\link[base:rbind]{base::rbind}} and/or \code{\link[base:cbind]{base::cbind}}:
\tabular{rl}{
  \code{cb}   \tab Column binds an arbitrary number of atomic matrices or \link[=atm_dtf]{atomic data.frames}.
\cr        \tab  
\cr \code{rb}   \tab Row binds an arbitrary number of atomic matrices or atomic data.frames.
}
}
\examples{
m2x4 <- mat(1:8, nr = 2)
m3x4 <- mat(1:12, nr = 3)
m4x4 <- mat(1:16, nr = 4)
m4x3 <- mat(1:12, nc = 3)
m4x4b <- m4x4

cb(m4x3, m4x4, m4x4b)
rb(m2x4, m3x4 , m4x4, m4x4b)

d4x2 <- dtf(w = 1:4, x = letters[1:4])
d4x4 <- dtf(w = 1:4, x = letters[1:4], y = NA, z = '.')
d2x4 <- dtf(w = 1:2, x = letters[1:2], y = NA, z = '.')
d4x4b <- d4x4
d4x4d <- dtf(W = 1:4, X = letters[1:4], Y = NA, Z = '.')

cb(d4x2, d4x4, d4x4d)
rb(d2x4, d4x4, d4x4b)
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()},
\code{\link{xat}()}
}
\concept{extensions}
