% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc..R
\name{ccc}
\alias{ccc}
\alias{ccc_props}
\alias{is_ccc_spec}
\alias{iccc}
\alias{iarr}
\alias{idtf}
\alias{igen}
\alias{imat}
\alias{imvc}
\alias{iscl}
\alias{ivec}
\alias{ivls}
\title{Extended class properties}
\usage{
ccc(x)

ccc_props()

is_ccc_spec(spec)

iccc(x, spec, ...)

iarr(x, ...)

idtf(x, ...)

igen(x, ...)

imat(x, ...)

imvc(x, ...)

iscl(x, ...)

ivec(x, ...)

ivls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more extended class properties (i.e., from \code{ccc_props()}). Extended class properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_ccc_spec}   \tab A logical scalar.
\cr \code{ccc_props}   \tab A character vector.
\cr      \code{iccc}   \tab A logical scalar.
\cr      \code{ixxx}   \tab A logical scalar\eqn{^2}.
\cr       \code{ccc}   \tab A character scalar or vector.
}
\eqn{^{1.}} \code{xxx} is an extended class property.
}
\description{
Extended class properties are not formally defined, but are dynamically evaluated.
\cr\cr \strong{Extended class property values}
\tabular{rll}{
   \code{arr}   \tab Array      \tab   Arrays.
\cr         \tab            \tab   
\cr \code{mat}   \tab Matrix     \tab   Matrices.
\cr         \tab            \tab   
\cr \code{dtf}   \tab Data.frame \tab   Data.frames.
\cr         \tab            \tab   
\cr \code{vls}   \tab Vlist      \tab   Vector-lists (non-data.frame lists).
\cr         \tab            \tab   
\cr \code{gen}   \tab Generic    \tab   Vectors, vlists, arrays.
\cr         \tab            \tab   
\cr \code{scl}   \tab Scalar     \tab   Length-\code{1}  vectors, vlists, arrays.
\cr         \tab            \tab   
\cr \code{mvc}   \tab Multivec   \tab   Length-\verb{2+} vectors, vlists; \link[=ie1D]{effectively 1D} arrays.
\cr         \tab            \tab   
\cr \code{vec}   \tab Vec        \tab   Scalars and multivecs.
}
\strong{Extended class property functions}
\tabular{rl}{
  \code{is_ccc_spec}   \tab Evaluates whether \code{spec} is a valid extended class property specification.
\cr               \tab   
\cr \code{ccc_props}   \tab Gets a character vector of all possible extended class property values.
\cr               \tab   
\cr      \code{ixxx}   \tab Evaluates whether \code{x} matches property \code{xxx}\eqn{^1} (subject to any restrictions in \code{...}).
\cr               \tab   
\cr      \code{iccc}   \tab Evaluates whether \code{x} matches one or more (possibly pipe-delimited) extended class properties in \code{spec} (subject to any restrictions in \code{...}).
\cr               \tab   
\cr       \code{ccc}   \tab Gets a character vector of all extended class properties matching \code{x}.
}
\eqn{^{1.}} An extended class property.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
\verb{max,maxr,maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                   \tab  
\cr \verb{min,minr,minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                   \tab  
\cr   \verb{lt,le,ge,gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                   \tab  
\cr       \verb{n,nr,nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                   \tab  
\cr          \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_ccc_spec("scl|vls")
ccc_props()
iccc(letters, "vec|dtf")
ivec(letters)
ccc(letters)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
