% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\encoding{UTF-8}
\name{ccc}
\alias{ccc}
\alias{ccc_funs}
\alias{ccc_props}
\alias{is_ccc_spec}
\alias{iccc}
\alias{iarr}
\alias{idtf}
\alias{igen}
\alias{imat}
\alias{imvc}
\alias{iscl}
\alias{ivec}
\alias{ivls}
\title{Extended class (xclass) properties}
\usage{
ccc(x)

ccc_funs()

ccc_props()

is_ccc_spec(spec)

iccc(x, spec, ...)

iarr(x, ...)

idtf(x, ...)

igen(x, ...)

imat(x, ...)

imvc(x, ...)

iscl(x, ...)

ivec(x, ...)

ivls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more xclass properties (i.e., from \code{\link{ccc_props}()}). Properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector}
\cr    \code{ccc_props}
\cr    \code{ccc_funs}
\cr    \code{ccc}
\cr\cr \emph{A logical vector}
\cr    \code{is_ccc_spec}
\cr    \code{iCCC}
\cr    \code{iccc}
}
\description{
\code{xclass} properties are defined as follows:
\tabular{rll}{
  \code{arr}   \tab array    \tab   arrays
\cr \code{mat}   \tab matrix   \tab   matrices\verb{\\cr}dtf\verb{   \\tab dtf      \\tab   data.frames}
\cr \code{vls}   \tab vlist    \tab   vector-lists\verb{\eqn{^a} \\cr }gen\verb{   \\tab generic  \\tab   vectors, vlists, and arrays \\cr }scl\verb{   \\tab scalar   \\tab   Length-}1\verb{generics \\cr}mvc\verb{   \\tab multivec \\tab   Length-}2+\verb{ \link[=ilin]\{linear\} generics \\cr }vec\verb{   \\tab vec      \\tab   scalars and multivecs \} \eqn{^{a.}} Non-}data.frame\verb{ lists. \\cr\\cr **Functions** \\tabular\{rl\}\{   }is_ccc_spec\verb{   \\tab Is }spec\verb{an xclass specification? \\cr               \\tab   \\cr}ccc_props\verb{   \\tab What xclass properties are there? \\cr               \\tab   \\cr  }ccc_funs\verb{   \\tab What xclass property functions are there? \\cr               \\tab   \\cr      }iccc\verb{   \\tab Is }x\verb{a match to the xclass specification}spec\verb{? \\cr               \\tab   \\cr      }iCCC\verb{   \\tab Is }x\verb{a match to the single xclass property}'CCC'\verb{? \\cr               \\tab   \\cr       }ccc\verb{   \\tab What are }x`'s xclass properties?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab  
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab  
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
ccc_funs()
ccc_props()
is_ccc_spec("scl|vls")
iccc(letters, "vec|dtf")
ivec(letters)
ccc(letters)
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{unq_mmm_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
