% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\name{ccc}
\alias{ccc}
\alias{ccc_funs}
\alias{ccc_props}
\alias{is_ccc_spec}
\alias{iccc}
\alias{iarr}
\alias{idtf}
\alias{igen}
\alias{imat}
\alias{imvc}
\alias{iscl}
\alias{ivec}
\alias{ivls}
\title{Extended class (\code{xclass}) properties}
\usage{
ccc(x)

ccc_funs()

ccc_props()

is_ccc_spec(spec)

iccc(x, spec, ...)

iarr(x, ...)

idtf(x, ...)

igen(x, ...)

imat(x, ...)

imvc(x, ...)

iscl(x, ...)

ivec(x, ...)

ivls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{xclass} properties (i.e., from \code{ccc_props()}). \code{xclass} properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_ccc_spec}   \tab A logical scalar.
\cr \code{ccc_props}   \tab A character vector.
\cr  \code{ccc_funs}   \tab A character vector.
\cr      \code{iccc}   \tab A logical scalar.
\cr      \code{iCCC}   \tab A logical scalar.
\cr       \code{ccc}   \tab A character vector.
}
}
\description{
\code{xclass} properties are defined as follows:
\tabular{rll}{
   \code{arr}   \tab \code{array}    \tab   \code{arrays}.
\cr \code{mat}   \tab \code{matrix}   \tab   \code{matrices}.
\cr \code{dtf}   \tab \code{dtf}      \tab   \code{data.frames}.
\cr \code{vls}   \tab \code{vlist}    \tab   \code{vector-lists}\eqn{^a}.
\cr \code{gen}   \tab \code{generic}  \tab   \code{vectors}, \code{vlists}, and \code{arrays.}
\cr \code{scl}   \tab \code{scalar}   \tab   Length-\code{1} \code{generics}
\cr \code{mvc}   \tab \code{multivec} \tab   Length-\verb{2+} \link[=ilin]{linear} \code{generics.}
\cr \code{vec}   \tab \code{vec}      \tab   \code{scalars} and \code{multivecs.}
}
\eqn{^{a.}} Non-\code{data.frame} lists.
\cr\cr
\strong{Functions}
\tabular{rl}{
  \code{is_ccc_spec}   \tab Is \code{spec} an \code{xclass} specification?
\cr               \tab   
\cr \code{ccc_props}   \tab What \code{xclass} properties are there?
\cr               \tab   
\cr  \code{ccc_funs}   \tab What \code{xclass} property functions are there?
\cr               \tab   
\cr      \code{iccc}   \tab Is \code{x} a match to the \code{xclass} specification \code{spec}?
\cr               \tab   
\cr      \code{iCCC}   \tab Is \code{x} a match to the single \code{xclass} property \code{'CCC'}?
\cr               \tab   
\cr       \code{ccc}   \tab What are \code{x}'s \code{xclass} properties?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
ccc_funs()
ccc_props()
is_ccc_spec("scl|vls")
iccc(letters, "vec|dtf")
ivec(letters)
ccc(letters)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
