% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\encoding{UTF-8}
\name{ccc}
\alias{ccc}
\alias{ccc_funs}
\alias{ccc_props}
\alias{is_ccc_spec}
\alias{CCC}
\alias{ARR}
\alias{DTF}
\alias{GEN}
\alias{MAT}
\alias{MVC}
\alias{SCL}
\alias{VEC}
\alias{VLS}
\title{Extended class (xclass) properties}
\usage{
ccc(x)

ccc_funs()

ccc_props()

is_ccc_spec(spec)

CCC(x, spec, ...)

ARR(x, ...)

DTF(x, ...)

GEN(x, ...)

MAT(x, ...)

MVC(x, ...)

SCL(x, ...)

VEC(x, ...)

VLS(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more xclass properties (i.e., from \code{\link{ccc_props}()}). Properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \verb{ccc_props, ccc_funs, ccc}
\cr\cr  \strong{A logical vector}   \cr\cr \verb{is_ccc_spec, \{CCC\}, CCC}
}
\description{
\code{xclass} properties are defined as follows:
\tabular{lll}{  \verb{'arr', 'ARR'}   \tab array      \tab arrays                                   \cr
  \verb{'mat', 'MAT'}   \tab matrix     \tab matrices                                 \cr
  \verb{'dtf', 'DTF'}   \tab dtf        \tab data.frames                              \cr
  \verb{'vls', 'VLS'}   \tab vlist      \tab vector-lists  \eqn{^{(1)}}               \cr
  \verb{'gen', 'GEN'}   \tab generic    \tab vectors, vlists, and arrays              \cr
  \verb{'scl', 'SCL'}   \tab scalar     \tab Length-\code{1} generics                      \cr
  \verb{'mvc', 'MVC'}   \tab multivec   \tab Length-\verb{2+} \link[=LIN]{linear} generics \cr
  \verb{'vec', 'VEC'}   \tab vec        \tab scalars and multivecs                      }
\tabular{l}{  \eqn{^{(1)}} Non-\code{data.frame} lists. }
}
\details{
\tabular{ll}{  \code{is_ccc_spec}   \tab Is \code{spec} an xclass specification?                                                                                   \cr   \tab   \cr
  \code{ccc_props}     \tab What xclass properties are there?                                                                                    \cr   \tab   \cr
  \code{ccc_funs}      \tab What xclass property functions are there?                                                                            \cr   \tab   \cr
  \code{{CCC}}         \tab Is \code{x} a match to the single xclass property \code{'{CCC}'} where \code{{CCC}} is a placeholder for any given xclass property? \cr   \tab   \cr
  \code{CCC}           \tab Is \code{x} a match to the xclass specification \code{spec}?                                                                   \cr   \tab   \cr
  \code{ccc}           \tab What are \code{x}'s xclass properties?                                                                                                   }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
ccc_funs()
ccc_props()
is_ccc_spec("scl|vls")
CCC(letters, "vec|dtf")
VEC(letters)
ccc(letters)
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{sss_ccc}()},
\code{\link{unq_mmm_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
