% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\name{ccc}
\alias{ccc}
\alias{ccc_props}
\alias{is_ccc_spec}
\alias{iccc}
\alias{iarr}
\alias{idtf}
\alias{igen}
\alias{imat}
\alias{imvc}
\alias{iscl}
\alias{ivec}
\alias{ivls}
\title{Extended class (\code{exclass}) properties}
\usage{
ccc(x)

ccc_props()

is_ccc_spec(spec)

iccc(x, spec, ...)

iarr(x, ...)

idtf(x, ...)

igen(x, ...)

imat(x, ...)

imvc(x, ...)

iscl(x, ...)

ivec(x, ...)

ivls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{exclass} properties (i.e., from \code{ccc_props()}). \code{exclass} properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_ccc_spec}   \tab A logical
\cr      \code{iCCC}   \tab scalar.
\cr      \code{iccc}   \tab   
\cr               \tab   
\cr \code{ccc_props}   \tab A character
\cr       \code{ccc}   \tab vector.
}
}
\description{
NOTE: \code{exclass}es are dynamically evaluated rather than formally defined.
\cr\cr \strong{exclass properties}
\tabular{rll}{
   \code{arr}   \tab array    \tab   Arrays.
\cr \code{mat}   \tab matrix   \tab   Matrices.
\cr \code{dtf}   \tab dtf      \tab   Data.frames.
\cr \code{vls}   \tab vlist    \tab   Vector-lists\eqn{^1}.
\cr \code{gen}   \tab generic  \tab   Vectors, vlists, & arrays.
\cr \code{scl}   \tab scalar   \tab   Length-\code{1} generics
\cr \code{mvc}   \tab multivec \tab   Length-\verb{2+}, \link[=ilin]{linear} generics.
\cr \code{vec}   \tab vec      \tab   Scalars, multivecs.
}
\eqn{^{1.}} i.e., a non-data.frame list.
\cr\cr \strong{exclass functions}
\tabular{rl}{
  \code{is_ccc_spec}   \tab Is \code{spec} is an \code{exclass} spec?
\cr \code{ccc_props}   \tab Gets all possible \code{exclass} values.
\cr      \code{iCCC}   \tab Does \code{x} match \code{exclass} property \code{CCC}\eqn{^2}?
\cr      \code{iccc}   \tab Does \code{x} match \code{exclass} spec \code{spec}\eqn{^2}?
\cr       \code{ccc}   \tab Gets all \code{exclass} properties matching \code{x}.
}
\eqn{^{2.}} subject to any restrictions in \code{...}.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_ccc_spec("scl|vls")
ccc_props()
iccc(letters, "vec|dtf")
ivec(letters)
ccc(letters)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
