% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{charn.}
\alias{charn.}
\alias{charn}
\title{Extract a single character}
\usage{
charn.()

charn(x, n)
}
\arguments{
\item{x}{\link[ichr]{Atomic character object}.}

\item{n}{\link[cmp_psw_scl]{Complete positive whole-number scalar} or
\link{cmp_psw}{complete positive whole-number atomic object} of the same
dimensions as \code{x}.}
}
\value{
\code{link[cmp_ch1]{Complete onechar}} atomic object of the same
dimension as \code{x}.
}
\description{
Extract a single character
}
\section{Functions}{
\itemize{
\item \code{charn()}: Get the \code{n}-th character of each element of
\code{x}.

}}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{has_pat.}()},
\code{\link{ipat.}()},
\code{\link{max_nch.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()},
\code{\link{xpr.}()}
}
\concept{strings}
