% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid.R
\encoding{UTF-8}
\name{check_valid}
\alias{check_valid}
\title{Check \code{x} Against Valid Values and Against a Test}
\usage{
check_valid(x, test, valid)
}
\arguments{
\item{x}{An object.}

\item{test}{A non-missing logical scalar.}

\item{valid}{Either \code{NULL} or an atomic vector of valid values.}

\item{return}{A logical scalar.}
}
\description{
Check \code{x} Against Valid Values and Against a Test
}
\seealso{
Other utils: 
\code{\link{combo2props}()},
\code{\link{meets_errs}()},
\code{\link{spec2combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()}
}
\concept{utils}
