% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chn.R
\name{chn}
\alias{chn}
\title{Get the \code{n}-th character element-wise.}
\usage{
chn(x, n)
}
\arguments{
\item{x}{A \link[=pop_chr]{populated character object}.}

\item{n}{A \link[=cmp_psw_scl]{positive whole-number scalar} or a \link[=cmp_psw]{complete positive whole-number object} of the same dimensions as \code{x}.}
}
\value{
A \link[=cmp_ch1]{complete onechar} atomic object of the same dimension as \code{x}.
}
\description{
Get the \code{n}-th character element-wise.
}
\examples{
 words <- stringr::fruit[1:10]
 words

 chn(words, 1)
 chn(words, 3)
 chn(words, 10)
}
\seealso{
Other chars: 
\code{\link{max_nch}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()}

Other strings: 
\code{\link{blank}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkd}},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{chars}
\concept{strings}
