% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chn.R
\encoding{UTF-8}
\name{chn}
\alias{chn}
\alias{charn}
\title{Get the \code{n}-th character element-wise.}
\usage{
chn(x, n)

charn(x, n)
}
\arguments{
\item{x}{A \link[=pop_chr]{populated character object}.}

\item{n}{A \link[=cmp_psw_scl]{positive whole-number scalar} or a \link[=cmp_psw]{complete positive whole-number object} of the same dimensions as \code{x}.}
}
\value{
A \link[=cmp_ch1]{complete onechar} atomic object of the same dimension as \code{x}.
}
\description{
Get the \code{n}-th character element-wise.
}
\examples{
words <- stringr::fruit[1:10]
words
chn(words, 1)
chn(words, 3)
charn(words, 10)
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown_help}()},
\code{\link{maxnch}()},
\code{\link{ox_vals}()},
\code{\link{ox}()},
\code{\link{pgrid_help}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss_help}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{maxnch}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss_help}()}
}
\concept{chars}
\concept{strings}
