% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\alias{clean_select}
\alias{clean_rename}
\alias{clean_recode}
\alias{clean_remode}
\title{Functions to Interactively Clean Data Frames}
\usage{
clean_data(X, .clear = TRUE)

clean_select(X, .clear = TRUE)

clean_rename(X, .clear = TRUE)

clean_recode(X, .clear = TRUE)

clean_remode(X, .clear = TRUE)
}
\arguments{
\item{X}{A data frame.}

\item{.clear}{A logical scalar indicating whether to clear the console at the start of each data cleaning step.}
}
\value{
A data frame
}
\description{
This family of functions clean a data frame as follows:
\tabular{ll}{  \code{clean_remode}        \tab Change variable modes.                      \cr
  \code{clean_recode}        \tab Recode variables.                           \cr
  \code{clean_rename}        \tab Rename variables.                           \cr
  \code{clean_select}        \tab Select variables to retain.                 \cr
  \code{clean_data}          \tab Select, rename, recode, and remode variables. }
}
