% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls.R
\encoding{UTF-8}
\name{cls}
\alias{cls}
\alias{rm_cls}
\alias{xcls}
\alias{set_cls}
\alias{add_cls}
\alias{has_cls}
\title{Manage Classes}
\usage{
cls(x)

rm_cls(x, ...)

xcls(x, ...)

set_cls(x, ...)

add_cls(x, ...)

has_cls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unquoted class names and/or names of variables whose contents are \link[=av]{atomized} and converted to character. When \code{...} args are names of variables in the calling environment, they are atomized, otherwise they are treated as unquoted class names.}
}
\value{
\strong{A character vector} \cr\cr \code{cls}
\cr\cr  \strong{A logical scalar}   \cr\cr \code{has_cls}
\cr\cr  \strong{An object}          \cr\cr \verb{add_cls, set_cls, rm_cls, xcls}
}
\description{
Manage Classes
}
\details{
\tabular{ll}{  \verb{rm_cls, xcls}   \tab Removes class(es) in \code{...} from classes of \code{x}.               \cr   \tab   \cr
  \code{has_cls}        \tab Evaluate whether class(es) in \code{...} are (all) classes of \code{x}. \cr   \tab   \cr
  \code{add_cls}        \tab Add class(es) in \code{...} to classes of \code{x}.                     \cr   \tab   \cr
  \code{set_cls}        \tab Set class of \code{x} to class(es) in \code{...}.                       \cr   \tab   \cr
  \code{cls}            \tab Gets class(es) of \code{x}.                                                       }
}
\examples{
egObj1 <- data.frame(letters = letters, numbers = 1:26, stringsAsFactors = F)
egObj2 <- set_cls(egObj1, c("new", cls(egObj1)))
egObj3 <- set_cls(egObj1, "new")
cls(egObj1)
cls(egObj2)
cls(egObj3)
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_ccc_help}()},
\code{\link{bbb_help}()},
\code{\link{bbb_mmm_prop_funs}()},
\code{\link{ccc_help}()},
\code{\link{cmp_ccc_help}()},
\code{\link{cmp_mmm_ccc_help}()},
\code{\link{cmp_mmm_help}()},
\code{\link{compatible}()},
\code{\link{ddd_prop_funs}()},
\code{\link{eee_help}()},
\code{\link{iii_help}()},
\code{\link{meets}()},
\code{\link{mmm_ccc_help}()},
\code{\link{mmm_help}()},
\code{\link{ppp_fast_help}()},
\code{\link{ppp_help}()},
\code{\link{sss_ccc_help}()},
\code{\link{sss_prop_funs}()},
\code{\link{unq_ccc_prop_funs}()},
\code{\link{unq_mmm_ccc_help}()},
\code{\link{unq_mmm_help}()}
}
\concept{properties}
