% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_ccc.R
\name{cmp_ccc.}
\alias{cmp_ccc.}
\alias{cmp_ccc_vals}
\alias{cmp_arr}
\alias{cmp_gen}
\alias{cmp_dtf}
\alias{cmp_vls}
\alias{cmp_mat}
\alias{cmp_mvc}
\alias{cmp_scl}
\alias{cmp_vec}
\title{Completeness + (atomic) Extended Class Properties}
\usage{
cmp_ccc.()

cmp_ccc_vals()

cmp_arr(x)

cmp_gen(x)

cmp_dtf(x)

cmp_vls(x)

cmp_mat(x)

cmp_mvc(x)

cmp_scl(x)

cmp_vec(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\tabular{ll}{
\code{cmp_mmm_vals}   \tab A character vector.                          \cr
\code{cmp_[ccc]}      \tab A logical scalar.                              }
}
\description{
\tabular{ll}{
\code{cmp_ccc_vals}   \tab Gets a character vector of all possible
\link[=icmp]{completeness} + atomic
\link[=ccc]{extended class} properties.      \cr
\code{cmp_[ccc]}      \tab Evaluates whether \code{x} is
\link[=icmp]{complete} and of the atomic
\link[=ccc]{extended class} represented by the
placeholder \code{[ccc]}.                      }
}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
