% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_ccc..R
\name{cmp_ccc}
\alias{cmp_ccc}
\alias{cmp_ccc_props}
\alias{cmp_arr}
\alias{cmp_dtf}
\alias{cmp_gen}
\alias{cmp_mat}
\alias{cmp_mvc}
\alias{cmp_scl}
\alias{cmp_vec}
\alias{cmp_vls}
\title{Completeness + extended class properties}
\usage{
cmp_ccc(x, ccc, ...)

cmp_ccc_props()

cmp_arr(x, ...)

cmp_dtf(x, ...)

cmp_gen(x, ...)

cmp_mat(x, ...)

cmp_mvc(x, ...)

cmp_scl(x, ...)

cmp_vec(x, ...)

cmp_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{ccc}{A character scalar extended class property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{cmp_ccc_props}   \tab A character vector.
\cr     \code{cmp_ccc}   \tab A logical scalar.
\cr     \code{cmp_xxx}   \tab A logical scalar\eqn{^2}.
}
\eqn{^{2.}} \code{xxx} is an extended class property.
}
\description{
\tabular{rl}{
  \code{cmp_ccc_props}   \tab Gets a character vector of all possible completeness + extended class properties.
\cr                 \tab  
\cr     \code{cmp_ccc}   \tab Evaluates whether \code{x} is complete and matches the extended class in argument \code{ccc} (subject to any restrictions in \code{...}).
\cr                 \tab  
\cr     \code{cmp_xxx}   \tab Evaluates whether \code{x} is complete and matches property \code{xxx}\eqn{^1} (subject to any restrictions in \code{...}).
}
\eqn{^{1.}} An extended class property.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
\verb{max,maxr,maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                   \tab  
\cr \verb{min,minr,minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                   \tab  
\cr   \verb{lt,le,ge,gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                   \tab  
\cr       \verb{n,nr,nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                   \tab  
\cr          \code{vals}   \tab A vector of valid values.
}
}

\examples{
cmp_ccc_props()
cmp_ccc(letters, "mvc")
cmp_ccc(1, "scl")
cmp_ccc(NA, "gen")
cmp_scl(1)
cmp_mvc(letters)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
