% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_ccc.R
\name{cmp_ccc}
\alias{cmp_ccc}
\alias{cmp_ccc_props}
\alias{cmp_arr}
\alias{cmp_dtf}
\alias{cmp_gen}
\alias{cmp_mat}
\alias{cmp_mvc}
\alias{cmp_scl}
\alias{cmp_vec}
\alias{cmp_vls}
\title{Completeness + Extended Class Properties}
\usage{
cmp_ccc(x, ccc, ...)

cmp_ccc_props()

cmp_arr(x, ...)

cmp_dtf(x, ...)

cmp_gen(x, ...)

cmp_mat(x, ...)

cmp_mvc(x, ...)

cmp_scl(x, ...)

cmp_vec(x, ...)

cmp_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{ccc}{A character scalar extended class property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{ll}{
FUNCTION               \tab WHAT IT DOES                                \cr
\code{cmp_ccc_props}        \tab A character vector.                         \cr
\code{cmp_ccc}, \code{cmp_CCC}   \tab A logical scalar.                             }
}
\description{
\tabular{ll}{
FUNCTION          \tab WHAT IT DOES                                     \cr
\code{cmp_ccc}         \tab Evaluate whether \code{x} is complete and matches the
extended class specified in the argument \code{ccc}
subject to any restrictions in \code{...}.            \cr
\code{cmp_CCC}         \tab Evaluates whether \code{x} is complete and matches the
extended class property \code{CCC} subject to any
restrictions in \code{...}, where \code{CCC} is a placeholder
for any given extended class property.           \cr
\code{cmp_ccc_props}   \tab Gets a character vector of all possible completeness
+ extended class properties.                       }
}
\section{Specifying Count and Value Restrictions}{
 Specifying restrictions in
\code{...} is optional. The full set of recognized arguments names are defined
in the following table along with the properties each
specifies:\tabular{ll}{
NAME     \tab WHAT IT SPECIFIES                                         \cr
\code{n}      \tab Vector of valid lengths/numbers of elements.              \cr
\code{nr}     \tab Vector of valid numbers of rows.                          \cr
\code{nc}     \tab Vector of valid numbers of columns.                       \cr
\code{min}    \tab Scalar minimum valid length/number of element.            \cr
\code{minr}   \tab Scalar minimum valid number of rows.                      \cr
\code{minc}   \tab Scalar minimum valid number of columns.                   \cr
\code{max}    \tab Scalar maximum valid length/number of element.            \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                      \cr
\code{maxc}   \tab Scalar maximum valid number of columns.                   \cr
\code{vals}   \tab Vector of valid values.                                   \cr
\code{lt}     \tab Scalar less-than (exclusive upper) bound.                 \cr
\code{le}     \tab Scalar less-than-or-equal (inclusive upper) bound         \cr
\code{ge}     \tab Scalar greater-than-or-equal (inclusive lower) bound.     \cr
\code{gt}     \tab Scalar greater-than bound (exclusive lower) bound.          }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
