% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_ccc.R
\name{cmp_ccc}
\alias{cmp_ccc}
\alias{cmp_ccc_props}
\alias{cmp_arr}
\alias{cmp_dtf}
\alias{cmp_gen}
\alias{cmp_mat}
\alias{cmp_mvc}
\alias{cmp_scl}
\alias{cmp_vec}
\alias{cmp_vls}
\title{Complete + \code{xclass} properties}
\usage{
cmp_ccc(x, spec, ...)

cmp_ccc_props()

cmp_arr(x, ...)

cmp_dtf(x, ...)

cmp_gen(x, ...)

cmp_mat(x, ...)

cmp_mvc(x, ...)

cmp_scl(x, ...)

cmp_vec(x, ...)

cmp_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{A character scalar \code{xclass} spec built from values in \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{cmp_ccc_props}   \tab A character vector.
\cr     \code{cmp_ccc}   \tab A logical scalar.
\cr     \code{cmp_CCC}   \tab A logical scalar.
}
}
\description{
\tabular{rl}{
  \code{cmp_ccc_props}   \tab Gets all possible \link[=icmp]{complete} + \code{\link[=ccc]{xclass}} properties.
\cr                 \tab   
\cr     \code{cmp_ccc}   \tab Is \code{x} complete and does it match \code{xclass} spec \code{ccc}?
\cr                 \tab   
\cr     \code{cmp_CCC}   \tab Is \code{x} complete and does it match \code{xclass} property \code{'CCC'}?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
cmp_ccc_props()
cmp_ccc(letters, "mvc")
cmp_ccc(1, "scl")
cmp_ccc(NA, "gen")
cmp_scl(1)
cmp_mvc(letters)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
