% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\name{cmp_mmm.}
\alias{cmp_mmm.}
\alias{cmp_mmm_vals}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{Completeness + Extended Mode Properties}
\usage{
cmp_mmm.()

cmp_mmm_vals()

cmp_atm(x)

cmp_ch1(x)

cmp_chr(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\tabular{ll}{
\code{cmp_mmm_vals}   \tab A character vector.                          \cr
\code{cmp_[mmm]}      \tab A logical scalar.                              }
}
\description{
\tabular{ll}{
\code{cmp_mmm_vals}   \tab Gets a character vector of all possible
\link[=icmp]{completeness} +
\link[=mmm]{extended mode} properties.       \cr
\code{cmp_[mmm]}      \tab Evaluates whether \code{x} is
\link[=icmp]{complete} and of the
\link[=mmm]{extended mode} represented by the
placeholder \code{[mmm]}.                      }
}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
