% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\name{cmp_mmm.}
\alias{cmp_mmm.}
\alias{cmp_mmm_vals}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{Complete plus extended mode properties}
\usage{
cmp_mmm.()

cmp_mmm_vals()

cmp_atm(x)

cmp_ch1(x)

cmp_chr(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE} except for \code{cmp_mmm_vals}, which
returns a character vector.
}
\description{
Combinations of \link[icmp]{completeness} and
\link[mmm]{extended mode}.
}
\section{Functions}{
\itemize{
\item \code{cmp_mmm_vals()}: Get a character vector of all possible complete +
extended mode properties

\item \code{cmp_atm()}: Is \code{x} a \link[icmp]{complete} \link[iatm]{atomic}
object?

\item \code{cmp_ch1()}: Is \code{x} a \link[icmp]{complete} \link[ich1]{onechar}
object?

\item \code{cmp_chr()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} object?

\item \code{cmp_clr()}: Is \code{x} a \link[icmp]{complete} \link[iclr]{color}
object?

\item \code{cmp_evn()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} object?

\item \code{cmp_fac()}: Is \code{x} a \link[icmp]{complete} \link[ifac]{factor}
object?

\item \code{cmp_frc()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} object?

\item \code{cmp_ind()}: Is \code{x} a \link[icmp]{complete} \link[iind]{indexer}
object?

\item \code{cmp_lgl()}: Is \code{x} a \link[icmp]{complete} \link[ilgl]{logical}
object?

\item \code{cmp_neg()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued numeric} object?

\item \code{cmp_ngw()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} object?

\item \code{cmp_nng()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} object?

\item \code{cmp_nnw()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} object?

\item \code{cmp_nps()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric} object?

\item \code{cmp_npw()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} object?

\item \code{cmp_nst()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} object?

\item \code{cmp_num()}: Is \code{x} a \link[icmp]{complete} \link[inum]{numeric}
object?

\item \code{cmp_odd()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} object?

\item \code{cmp_ord()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} object?

\item \code{cmp_pct()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} object?

\item \code{cmp_pos()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric} object?

\item \code{cmp_ppn()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} object?

\item \code{cmp_psw()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued whole-number} object?

\item \code{cmp_srt()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} object?

\item \code{cmp_str()}: Is \code{x} a \link[icmp]{complete} \link[istr]{string}
object?

\item \code{cmp_uno()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} object?

\item \code{cmp_whl()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} object?

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
