% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\name{cmp_mmm}
\alias{cmp_mmm}
\alias{cmp_mmm_props}
\alias{cmp_ch1}
\alias{cmp_ch3}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{Completeness + Extended Mode Properties}
\usage{
cmp_mmm(x, mmm, ...)

cmp_mmm_props()

cmp_ch1(x, ...)

cmp_ch3(x, ...)

cmp_chr(x, ...)

cmp_clr(x, ...)

cmp_evn(x, ...)

cmp_fac(x, ...)

cmp_frc(x, ...)

cmp_ind(x, ...)

cmp_lgl(x, ...)

cmp_neg(x, ...)

cmp_ngw(x, ...)

cmp_nng(x, ...)

cmp_nnw(x, ...)

cmp_nps(x, ...)

cmp_npw(x, ...)

cmp_nst(x, ...)

cmp_num(x, ...)

cmp_odd(x, ...)

cmp_ord(x, ...)

cmp_pct(x, ...)

cmp_pos(x, ...)

cmp_ppn(x, ...)

cmp_psw(x, ...)

cmp_srt(x, ...)

cmp_str(x, ...)

cmp_uno(x, ...)

cmp_whl(x, ...)
}
\arguments{
\item{x}{An R object}

\item{mmm}{A character scalar containing an extended mode property from
\code{mmm_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{ll}{
FUNCTION               \tab RETURN VALUE                                \cr
\code{cmp_mmm_props}        \tab A character vector.                         \cr
\code{cmp_mmm}, \code{cmp_MMM}   \tab A logical scalar.                             }
}
\description{
\tabular{ll}{
FUNCTION          \tab WHAT IT DOES                                     \cr
\code{cmp_mmm}         \tab Evaluates whether \code{x} is complete and matches
the extended mode specified in the argument \code{mmm}
subject to any restrictions in \code{...}.            \cr
\code{cmp_MMM}         \tab Evaluates whether \code{x} is complete and matches the
extended mode property \code{MMM} subject to any
restrictions in \code{...}, where \code{MMM} is a placeholder
for any given extended mode property.            \cr
\code{cmp_mmm_props}   \tab Gets a character vector of all possible completeness
+ extended mode properties.                        }
}
\section{Specifying Count and Value Restrictions}{
 Specifying restrictions in
\code{...} is optional. The full set of recognized arguments names are defined
in the following table along with the properties each
specifies:\tabular{ll}{
NAME     \tab WHAT IT SPECIFIES                                         \cr
\code{n}      \tab Vector of valid lengths/numbers of elements.              \cr
\code{nr}     \tab Vector of valid numbers of rows.                          \cr
\code{nc}     \tab Vector of valid numbers of columns.                       \cr
\code{min}    \tab Scalar minimum valid length/number of element.            \cr
\code{minr}   \tab Scalar minimum valid number of rows.                      \cr
\code{minc}   \tab Scalar minimum valid number of columns.                   \cr
\code{max}    \tab Scalar maximum valid length/number of element.            \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                      \cr
\code{maxc}   \tab Scalar maximum valid number of columns.                   \cr
\code{vals}   \tab Vector of valid values.                                   \cr
\code{lt}     \tab Scalar less-than (exclusive upper) bound.                 \cr
\code{le}     \tab Scalar less-than-or-equal (inclusive upper) bound         \cr
\code{ge}     \tab Scalar greater-than-or-equal (inclusive lower) bound.     \cr
\code{gt}     \tab Scalar greater-than bound (exclusive lower) bound.          }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
