% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\name{cmp_mmm}
\alias{cmp_mmm}
\alias{cmp_mmm_vals}
\alias{cmp_atm}
\alias{cmp_chr}
\alias{cmp_ch1}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{Complete + Extended Mode (cmp + mmm)}
\usage{
cmp_mmm_vals()

cmp_atm(x)

cmp_chr(x)

cmp_ch1(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible complete + extended mode properties.
}
\details{
See \code{\link{xcmp}}, and \code{\link{xmmm}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm}: Is \code{x} a complete atomic generic?

\item \code{cmp_chr}: Is \code{x} a complete character atomic generic?

\item \code{cmp_ch1}: Is \code{x} a complete onechar atomic generic?

\item \code{cmp_clr}: Is \code{x} a complete color atomic generic?

\item \code{cmp_evn}: Is \code{x} a complete even-numeric atomic generic?

\item \code{cmp_fac}: Is \code{x} a complete factor atomic generic?

\item \code{cmp_frc}: Is \code{x} a complete fractional-numeric atomic generic?

\item \code{cmp_ind}: Is \code{x} a complete indexer atomic generic?

\item \code{cmp_lgl}: Is \code{x} a complete logical atomic generic?

\item \code{cmp_neg}: Is \code{x} a complete negative-numeric atomic generic?

\item \code{cmp_ngw}: Is \code{x} a complete negative-whole-numeric atomic
generic?

\item \code{cmp_nng}: Is \code{x} a complete non-negative-numeric atomic
generic?

\item \code{cmp_nnw}: Is \code{x} a complete non-negative-whole-numeric atomic
generic?

\item \code{cmp_nps}: Is \code{x} a complete non-negative-whole atomic generic?

\item \code{cmp_npw}: Is \code{x} a complete non-positive-numeric atomic
generic?

\item \code{cmp_nst}: Is \code{x} a complete non-sortable atomic generic?

\item \code{cmp_num}: Is \code{x} a complete numeric atomic generic?

\item \code{cmp_odd}: Is \code{x} a complete odd-numeric atomic generic?

\item \code{cmp_ord}: Is \code{x} a complete ordered-factor atomic generic?

\item \code{cmp_pct}: Is \code{x} a complete percent-numeric atomic generic?

\item \code{cmp_pos}: Is \code{x} a complete positive-numeric atomic generic?

\item \code{cmp_ppn}: Is \code{x} a complete proportion-numeric atomic generic?

\item \code{cmp_psw}: Is \code{x} a complete positive-whole-numeric atomic
generic?

\item \code{cmp_srt}: Is \code{x} a complete sortable atomic generic?

\item \code{cmp_str}: Is \code{x} a complete string atomic generic?

\item \code{cmp_uno}: Is \code{x} a complete unordered-factor atomic generic?

\item \code{cmp_whl}: Is \code{x} a complete whole-numeric atomic generic?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
