% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\encoding{UTF-8}
\name{cmp_mmm}
\alias{cmp_mmm}
\alias{cmp_mmm_funs}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_ch3}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{complete + xmode combination properties}
\usage{
cmp_mmm(x, mmm, ...)

cmp_mmm_funs()

cmp_atm(x, ...)

cmp_ch1(x, ...)

cmp_ch3(x, ...)

cmp_chr(x, ...)

cmp_clr(x, ...)

cmp_evn(x, ...)

cmp_fac(x, ...)

cmp_frc(x, ...)

cmp_ind(x, ...)

cmp_lgl(x, ...)

cmp_neg(x, ...)

cmp_ngw(x, ...)

cmp_nng(x, ...)

cmp_nnw(x, ...)

cmp_nps(x, ...)

cmp_npw(x, ...)

cmp_nst(x, ...)

cmp_num(x, ...)

cmp_odd(x, ...)

cmp_ord(x, ...)

cmp_pct(x, ...)

cmp_pos(x, ...)

cmp_pos(x, ...)

cmp_psw(x, ...)

cmp_srt(x, ...)

cmp_str(x, ...)

cmp_uno(x, ...)

cmp_whl(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{mmm}{A character scalar xmode property from \code{mmm_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \code{cmp_mmm_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{cmp_\{mmm\}, cmp_mmm}
}
\description{
Check for combination of \link[=CMP]{completeness} and \link[=mmm]{xmode}.
}
\details{
\tabular{ll}{  \code{cmp_mmm_funs}   \tab What \link[=CMP]{complete} + \link[=mmm]{xmode} combination \link[=prop_funs]{property functions} are there?                         \cr   \tab   \cr
  \verb{cmp_\{mmm\}}      \tab Is \code{x} both complete and a match to the single xmode property \code{'{mmm}'} where \code{{mmm}} is a placeholder for any given xmode property? \cr   \tab   \cr
  \code{cmp_mmm}        \tab Is \code{x} both complete and a match to single xmode property in arg \code{mmm}?                                                                             }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
cmp_mmm_funs()
cmp_mmm(letters, "ch1")
cmp_mmm(c("abc", "def"), "ch3")
cmp_psw(0:10)
cmp_ord(factor(letters, ordered = T))
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
