% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm.R
\name{cmp_mmm}
\alias{cmp_mmm}
\alias{cmp_mmm_props}
\alias{cmp_ch1}
\alias{cmp_ch3}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\title{Completeness + extended mode properties}
\usage{
cmp_mmm(x, mmm, ...)

cmp_mmm_props()

cmp_ch1(x, ...)

cmp_ch3(x, ...)

cmp_chr(x, ...)

cmp_clr(x, ...)

cmp_evn(x, ...)

cmp_fac(x, ...)

cmp_frc(x, ...)

cmp_ind(x, ...)

cmp_lgl(x, ...)

cmp_neg(x, ...)

cmp_ngw(x, ...)

cmp_nng(x, ...)

cmp_nnw(x, ...)

cmp_nps(x, ...)

cmp_npw(x, ...)

cmp_nst(x, ...)

cmp_num(x, ...)

cmp_odd(x, ...)

cmp_ord(x, ...)

cmp_pct(x, ...)

cmp_pos(x, ...)

cmp_ppn(x, ...)

cmp_psw(x, ...)

cmp_srt(x, ...)

cmp_str(x, ...)

cmp_uno(x, ...)

cmp_whl(x, ...)
}
\arguments{
\item{x}{An R object}

\item{mmm}{A character scalar containing an extended mode property from \code{mmm_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{cmp_mmm_props}   \tab A character vector.
\cr     \code{cmp_mmm}   \tab A logical scalar.
\cr     \code{cmp_xxx}   \tab A logical  calar\eqn{^1}.
}
\eqn{^{2.}} \code{xxx} is an extended mode property.
}
\description{
\tabular{rl}{
  \code{cmp_mmm_props}   \tab Gets a character vector of all possible \link[=icmp]{completeness} + \link[=mmm]{extended mode} properties.
\cr                 \tab  
\cr     \code{cmp_mmm}   \tab Evaluates whether \code{x} is complete and matches the extended mode spec in argument \code{mmm} (subject to any restrictions in \code{...}).
\cr                 \tab  
\cr     \code{cmp_xxx}   \tab Evaluates whether \code{x} is complete and matches extended mode \code{xxx\eqn{^1}} (subject to any restrictions in \code{...}).
}
\eqn{^{1.}} An extended mode property.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
cmp_mmm_props()
cmp_mmm(letters, "ch1")
cmp_mmm(c("abc", "def"), "ch3")
cmp_psw(0:10)
cmp_ord(factor(letters, ordered = T))
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
