% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_agn.R
\name{cmp_mmm_agn}
\alias{cmp_mmm_agn}
\alias{cmp_mmm_agn_vals}
\alias{cmp_atm_agn}
\alias{cmp_chr_agn}
\alias{cmp_ch1_agn}
\alias{cmp_clr_agn}
\alias{cmp_evn_agn}
\alias{cmp_fac_agn}
\alias{cmp_frc_agn}
\alias{cmp_ind_agn}
\alias{cmp_lgc_agn}
\alias{cmp_cnb_agn}
\alias{cmp_neg_agn}
\alias{cmp_ngw_agn}
\alias{cmp_nng_agn}
\alias{cmp_nnw_agn}
\alias{cmp_nps_agn}
\alias{cmp_npw_agn}
\alias{cmp_nst_agn}
\alias{cmp_num_agn}
\alias{cmp_odd_agn}
\alias{cmp_ord_agn}
\alias{cmp_pct_agn}
\alias{cmp_pos_agn}
\alias{cmp_ppn_agn}
\alias{cmp_psw_agn}
\alias{cmp_srt_agn}
\alias{cmp_str_agn}
\alias{cmp_uno_agn}
\alias{cmp_whl_agn}
\title{Complete + Extended Mode + Atomic Generic (cmp + mmm + agn)}
\usage{
cmp_mmm_agn_vals()

cmp_atm_agn(x)

cmp_chr_agn(x)

cmp_ch1_agn(x)

cmp_clr_agn(x)

cmp_evn_agn(x)

cmp_fac_agn(x)

cmp_frc_agn(x)

cmp_ind_agn(x)

cmp_lgc_agn(x)

cmp_cnb_agn(x)

cmp_neg_agn(x)

cmp_ngw_agn(x)

cmp_nng_agn(x)

cmp_nnw_agn(x)

cmp_nps_agn(x)

cmp_npw_agn(x)

cmp_nst_agn(x)

cmp_num_agn(x)

cmp_odd_agn(x)

cmp_ord_agn(x)

cmp_pct_agn(x)

cmp_pos_agn(x)

cmp_ppn_agn(x)

cmp_psw_agn(x)

cmp_srt_agn(x)

cmp_str_agn(x)

cmp_uno_agn(x)

cmp_whl_agn(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
See \code{\link{mmm}}, and \code{\link{is_cmp_generic}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
