% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_arr.R
\name{cmp_mmm_arr}
\alias{cmp_mmm_arr}
\alias{cmp_mmm_arr_vals}
\alias{cmp_atm_arr}
\alias{cmp_chr_arr}
\alias{cmp_ch1_arr}
\alias{cmp_clr_arr}
\alias{cmp_evn_arr}
\alias{cmp_fac_arr}
\alias{cmp_frc_arr}
\alias{cmp_ind_arr}
\alias{cmp_lgc_arr}
\alias{cmp_cnb_arr}
\alias{cmp_neg_arr}
\alias{cmp_ngw_arr}
\alias{cmp_nng_arr}
\alias{cmp_nnw_arr}
\alias{cmp_nps_arr}
\alias{cmp_npw_arr}
\alias{cmp_nst_arr}
\alias{cmp_num_arr}
\alias{cmp_odd_arr}
\alias{cmp_ord_arr}
\alias{cmp_pct_arr}
\alias{cmp_pos_arr}
\alias{cmp_ppn_arr}
\alias{cmp_psw_arr}
\alias{cmp_srt_arr}
\alias{cmp_str_arr}
\alias{cmp_uno_arr}
\alias{cmp_whl_arr}
\title{Complete + Extended Mode + Atomic Array (cmp + mmm + arr)}
\usage{
cmp_mmm_arr_vals()

cmp_atm_arr(x)

cmp_chr_arr(x)

cmp_ch1_arr(x)

cmp_clr_arr(x)

cmp_evn_arr(x)

cmp_fac_arr(x)

cmp_frc_arr(x)

cmp_ind_arr(x)

cmp_lgc_arr(x)

cmp_cnb_arr(x)

cmp_neg_arr(x)

cmp_ngw_arr(x)

cmp_nng_arr(x)

cmp_nnw_arr(x)

cmp_nps_arr(x)

cmp_npw_arr(x)

cmp_nst_arr(x)

cmp_num_arr(x)

cmp_odd_arr(x)

cmp_ord_arr(x)

cmp_pct_arr(x)

cmp_pos_arr(x)

cmp_ppn_arr(x)

cmp_psw_arr(x)

cmp_srt_arr(x)

cmp_str_arr(x)

cmp_uno_arr(x)

cmp_whl_arr(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_cmp_array}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
