% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_avl.R
\name{cmp_mmm_avl}
\alias{cmp_mmm_avl}
\alias{cmp_mmm_avl_vals}
\alias{cmp_atm_avl}
\alias{cmp_chr_avl}
\alias{cmp_ch1_avl}
\alias{cmp_clr_avl}
\alias{cmp_evn_avl}
\alias{cmp_fac_avl}
\alias{cmp_frc_avl}
\alias{cmp_ind_avl}
\alias{cmp_lgl_avl}
\alias{cmp_neg_avl}
\alias{cmp_ngw_avl}
\alias{cmp_nng_avl}
\alias{cmp_nnw_avl}
\alias{cmp_nps_avl}
\alias{cmp_npw_avl}
\alias{cmp_nst_avl}
\alias{cmp_num_avl}
\alias{cmp_odd_avl}
\alias{cmp_ord_avl}
\alias{cmp_pct_avl}
\alias{cmp_pos_avl}
\alias{cmp_ppn_avl}
\alias{cmp_psw_avl}
\alias{cmp_srt_avl}
\alias{cmp_str_avl}
\alias{cmp_uno_avl}
\alias{cmp_whl_avl}
\title{Complete + Extended Mode + Atomic Vlist (cmp + mmm + avl)}
\usage{
cmp_mmm_avl_vals()

cmp_atm_avl(x)

cmp_chr_avl(x)

cmp_ch1_avl(x)

cmp_clr_avl(x)

cmp_evn_avl(x)

cmp_fac_avl(x)

cmp_frc_avl(x)

cmp_ind_avl(x)

cmp_lgl_avl(x)

cmp_neg_avl(x)

cmp_ngw_avl(x)

cmp_nng_avl(x)

cmp_nnw_avl(x)

cmp_nps_avl(x)

cmp_npw_avl(x)

cmp_nst_avl(x)

cmp_num_avl(x)

cmp_odd_avl(x)

cmp_ord_avl(x)

cmp_pct_avl(x)

cmp_pos_avl(x)

cmp_ppn_avl(x)

cmp_psw_avl(x)

cmp_srt_avl(x)

cmp_str_avl(x)

cmp_uno_avl(x)

cmp_whl_avl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic vlist
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_vls}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_avl}: Is \code{x} a complete atomic vlist?

\item \code{cmp_chr_avl}: Is \code{x} a complete character atomic vlist?

\item \code{cmp_ch1_avl}: Is \code{x} a complete onechar atomic vlist?

\item \code{cmp_clr_avl}: Is \code{x} a complete color atomic vlist?

\item \code{cmp_evn_avl}: Is \code{x} a complete even-numbered atomic vlist?

\item \code{cmp_fac_avl}: Is \code{x} a complete factor atomic vlist?

\item \code{cmp_frc_avl}: Is \code{x} a complete fractional-numbered atomic
vlist?

\item \code{cmp_ind_avl}: Is \code{x} a complete indexer atomic vlist?

\item \code{cmp_lgl_avl}: Is \code{x} a complete logical atomic vlist?

\item \code{cmp_neg_avl}: Is \code{x} a complete negative-number atomic vlist?

\item \code{cmp_ngw_avl}: Is \code{x} a complete negative-whole-number atomic
vlist?

\item \code{cmp_nng_avl}: Is \code{x} a complete non-negative-number atomic
vlist?

\item \code{cmp_nnw_avl}: Is \code{x} a complete non-negative-whole-number
atomic vlist?

\item \code{cmp_nps_avl}: Is \code{x} a complete non-positive-numbered atomic
vlist?

\item \code{cmp_npw_avl}: Is \code{x} a complete non-positive-number atomic
vlist?

\item \code{cmp_nst_avl}: Is \code{x} a complete non-sortable atomic vlist?

\item \code{cmp_num_avl}: Is \code{x} a complete number atomic vlist?

\item \code{cmp_odd_avl}: Is \code{x} a complete odd-number atomic vlist?

\item \code{cmp_ord_avl}: Is \code{x} a complete ordered-factor atomic vlist?

\item \code{cmp_pct_avl}: Is \code{x} a complete percent-number atomic vlist?

\item \code{cmp_pos_avl}: Is \code{x} a complete positive-number atomic vlist?

\item \code{cmp_ppn_avl}: Is \code{x} a complete proportion-number atomic
vlist?

\item \code{cmp_psw_avl}: Is \code{x} a complete positive-whole-number atomic
vlist?

\item \code{cmp_srt_avl}: Is \code{x} a complete sortable atomic vlist?

\item \code{cmp_str_avl}: Is \code{x} a complete string atomic vlist?

\item \code{cmp_uno_avl}: Is \code{x} a complete unordered-factor atomic vlist?

\item \code{cmp_whl_avl}: Is \code{x} a complete whole-number atomic vlist?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
