% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_avt.R
\name{cmp_mmm_avt}
\alias{cmp_mmm_avt}
\alias{cmp_mmm_avt_vals}
\alias{cmp_atm_avt}
\alias{cmp_chr_avt}
\alias{cmp_ch1_avt}
\alias{cmp_clr_avt}
\alias{cmp_evn_avt}
\alias{cmp_fac_avt}
\alias{cmp_frc_avt}
\alias{cmp_ind_avt}
\alias{cmp_lgl_avt}
\alias{cmp_neg_avt}
\alias{cmp_ngw_avt}
\alias{cmp_nng_avt}
\alias{cmp_nnw_avt}
\alias{cmp_nps_avt}
\alias{cmp_npw_avt}
\alias{cmp_nst_avt}
\alias{cmp_num_avt}
\alias{cmp_odd_avt}
\alias{cmp_ord_avt}
\alias{cmp_pct_avt}
\alias{cmp_pos_avt}
\alias{cmp_ppn_avt}
\alias{cmp_psw_avt}
\alias{cmp_srt_avt}
\alias{cmp_str_avt}
\alias{cmp_uno_avt}
\alias{cmp_whl_avt}
\title{Complete + Extended Mode + Atomic Vtype (cmp + mmm + avt)}
\usage{
cmp_mmm_avt_vals()

cmp_atm_avt(x)

cmp_chr_avt(x)

cmp_ch1_avt(x)

cmp_clr_avt(x)

cmp_evn_avt(x)

cmp_fac_avt(x)

cmp_frc_avt(x)

cmp_ind_avt(x)

cmp_lgl_avt(x)

cmp_neg_avt(x)

cmp_ngw_avt(x)

cmp_nng_avt(x)

cmp_nnw_avt(x)

cmp_nps_avt(x)

cmp_npw_avt(x)

cmp_nst_avt(x)

cmp_num_avt(x)

cmp_odd_avt(x)

cmp_ord_avt(x)

cmp_pct_avt(x)

cmp_pos_avt(x)

cmp_ppn_avt(x)

cmp_psw_avt(x)

cmp_srt_avt(x)

cmp_str_avt(x)

cmp_uno_avt(x)

cmp_whl_avt(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic vtype
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_vtp}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_avt}: Is \code{x} a complete atomic vtype?

\item \code{cmp_chr_avt}: Is \code{x} a complete character atomic vtype?

\item \code{cmp_ch1_avt}: Is \code{x} a complete onechar atomic vtype?

\item \code{cmp_clr_avt}: Is \code{x} a complete color atomic vtype?

\item \code{cmp_evn_avt}: Is \code{x} a complete even-number atomic vtype?

\item \code{cmp_fac_avt}: Is \code{x} a complete factor atomic vtype?

\item \code{cmp_frc_avt}: Is \code{x} a complete fractional-number atomic
vtype?

\item \code{cmp_ind_avt}: Is \code{x} a complete indexer atomic vtype?

\item \code{cmp_lgl_avt}: Is \code{x} a complete logical atomic vtype?

\item \code{cmp_neg_avt}: Is \code{x} a complete negative-number atomic vtype?

\item \code{cmp_ngw_avt}: Is \code{x} a complete negative-whole-number atomic
vtype?

\item \code{cmp_nng_avt}: Is \code{x} a complete non-negative-number atomic
vtype?

\item \code{cmp_nnw_avt}: Is \code{x} a complete non-negative-whole-number
atomic vtype?

\item \code{cmp_nps_avt}: Is \code{x} a complete non-positive-number atomic
vtype?

\item \code{cmp_npw_avt}: Is \code{x} a complete non-positive-whole-number
atomic vtype?

\item \code{cmp_nst_avt}: Is \code{x} a complete non-sortable atomic vtype?

\item \code{cmp_num_avt}: Is \code{x} a complete number atomic vtype?

\item \code{cmp_odd_avt}: Is \code{x} a complete odd-number atomic vtype?

\item \code{cmp_ord_avt}: Is \code{x} a complete ordered-factor atomic vtype?

\item \code{cmp_pct_avt}: Is \code{x} a complete percentage-number atomic
vtype?

\item \code{cmp_pos_avt}: Is \code{x} a complete positive-number atomic vtype?

\item \code{cmp_ppn_avt}: Is \code{x} a complete proportional-number atomic
vtype?

\item \code{cmp_psw_avt}: Is \code{x} a complete positive-whole-number atomic
vtype?

\item \code{cmp_srt_avt}: Is \code{x} a complete sortable atomic vtype?

\item \code{cmp_str_avt}: Is \code{x} a complete string atomic vtype?

\item \code{cmp_uno_avt}: Is \code{x} a complete unordered-factor atomic vtype?

\item \code{cmp_whl_avt}: Is \code{x} a complete whole-numbered atomic vtype?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
