% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_ccc.R
\name{cmp_mmm_ccc}
\alias{cmp_mmm_ccc}
\alias{cmp_mmm_ccc_props}
\alias{cmp_ch1_arr}
\alias{cmp_ch1_dtf}
\alias{cmp_ch1_gen}
\alias{cmp_ch1_mat}
\alias{cmp_ch1_mvc}
\alias{cmp_ch1_scl}
\alias{cmp_ch1_vec}
\alias{cmp_ch1_vls}
\alias{cmp_ch3_arr}
\alias{cmp_ch3_dtf}
\alias{cmp_ch3_gen}
\alias{cmp_ch3_mat}
\alias{cmp_ch3_mvc}
\alias{cmp_ch3_scl}
\alias{cmp_ch3_vec}
\alias{cmp_ch3_vls}
\alias{cmp_chr_arr}
\alias{cmp_chr_dtf}
\alias{cmp_chr_gen}
\alias{cmp_chr_mat}
\alias{cmp_chr_mvc}
\alias{cmp_chr_scl}
\alias{cmp_chr_vec}
\alias{cmp_chr_vls}
\alias{cmp_clr_arr}
\alias{cmp_clr_dtf}
\alias{cmp_clr_gen}
\alias{cmp_clr_mat}
\alias{cmp_clr_mvc}
\alias{cmp_clr_scl}
\alias{cmp_clr_vec}
\alias{cmp_clr_vls}
\alias{cmp_evn_arr}
\alias{cmp_evn_dtf}
\alias{cmp_evn_gen}
\alias{cmp_evn_mat}
\alias{cmp_evn_mvc}
\alias{cmp_evn_scl}
\alias{cmp_evn_vec}
\alias{cmp_evn_vls}
\alias{cmp_fac_arr}
\alias{cmp_fac_dtf}
\alias{cmp_fac_gen}
\alias{cmp_fac_mat}
\alias{cmp_fac_mvc}
\alias{cmp_fac_scl}
\alias{cmp_fac_vec}
\alias{cmp_fac_vls}
\alias{cmp_frc_arr}
\alias{cmp_frc_dtf}
\alias{cmp_frc_gen}
\alias{cmp_frc_mat}
\alias{cmp_frc_mvc}
\alias{cmp_frc_scl}
\alias{cmp_frc_vec}
\alias{cmp_frc_vls}
\alias{cmp_ind_arr}
\alias{cmp_ind_dtf}
\alias{cmp_ind_gen}
\alias{cmp_ind_mat}
\alias{cmp_ind_mvc}
\alias{cmp_ind_scl}
\alias{cmp_ind_vec}
\alias{cmp_ind_vls}
\alias{cmp_lgl_arr}
\alias{cmp_lgl_dtf}
\alias{cmp_lgl_gen}
\alias{cmp_lgl_mat}
\alias{cmp_lgl_mvc}
\alias{cmp_lgl_scl}
\alias{cmp_lgl_vec}
\alias{cmp_lgl_vls}
\alias{cmp_neg_arr}
\alias{cmp_neg_dtf}
\alias{cmp_neg_gen}
\alias{cmp_neg_mat}
\alias{cmp_neg_mvc}
\alias{cmp_neg_scl}
\alias{cmp_neg_vec}
\alias{cmp_neg_vls}
\alias{cmp_ngw_arr}
\alias{cmp_ngw_dtf}
\alias{cmp_ngw_gen}
\alias{cmp_ngw_mat}
\alias{cmp_ngw_mvc}
\alias{cmp_ngw_scl}
\alias{cmp_ngw_vec}
\alias{cmp_ngw_vls}
\alias{cmp_nng_arr}
\alias{cmp_nng_dtf}
\alias{cmp_nng_gen}
\alias{cmp_nng_mat}
\alias{cmp_nng_mvc}
\alias{cmp_nng_scl}
\alias{cmp_nng_vec}
\alias{cmp_nng_vls}
\alias{cmp_nnw_arr}
\alias{cmp_nnw_dtf}
\alias{cmp_nnw_gen}
\alias{cmp_nnw_mat}
\alias{cmp_nnw_mvc}
\alias{cmp_nnw_scl}
\alias{cmp_nnw_vec}
\alias{cmp_nnw_vls}
\alias{cmp_nps_arr}
\alias{cmp_nps_dtf}
\alias{cmp_nps_gen}
\alias{cmp_nps_mat}
\alias{cmp_nps_mvc}
\alias{cmp_nps_scl}
\alias{cmp_nps_vec}
\alias{cmp_nps_vls}
\alias{cmp_npw_arr}
\alias{cmp_npw_dtf}
\alias{cmp_npw_gen}
\alias{cmp_npw_mat}
\alias{cmp_npw_mvc}
\alias{cmp_npw_scl}
\alias{cmp_npw_vec}
\alias{cmp_npw_vls}
\alias{cmp_nst_arr}
\alias{cmp_nst_dtf}
\alias{cmp_nst_gen}
\alias{cmp_nst_mat}
\alias{cmp_nst_mvc}
\alias{cmp_nst_scl}
\alias{cmp_nst_vec}
\alias{cmp_nst_vls}
\alias{cmp_num_arr}
\alias{cmp_num_dtf}
\alias{cmp_num_gen}
\alias{cmp_num_mat}
\alias{cmp_num_mvc}
\alias{cmp_num_scl}
\alias{cmp_num_vec}
\alias{cmp_num_vls}
\alias{cmp_odd_arr}
\alias{cmp_odd_dtf}
\alias{cmp_odd_gen}
\alias{cmp_odd_mat}
\alias{cmp_odd_mvc}
\alias{cmp_odd_scl}
\alias{cmp_odd_vec}
\alias{cmp_odd_vls}
\alias{cmp_ord_arr}
\alias{cmp_ord_dtf}
\alias{cmp_ord_gen}
\alias{cmp_ord_mat}
\alias{cmp_ord_mvc}
\alias{cmp_ord_scl}
\alias{cmp_ord_vec}
\alias{cmp_ord_vls}
\alias{cmp_pct_arr}
\alias{cmp_pct_dtf}
\alias{cmp_pct_gen}
\alias{cmp_pct_mat}
\alias{cmp_pct_mvc}
\alias{cmp_pct_scl}
\alias{cmp_pct_vec}
\alias{cmp_pct_vls}
\alias{cmp_pos_arr}
\alias{cmp_pos_dtf}
\alias{cmp_pos_gen}
\alias{cmp_pos_mat}
\alias{cmp_pos_mvc}
\alias{cmp_pos_scl}
\alias{cmp_pos_vec}
\alias{cmp_pos_vls}
\alias{cmp_ppn_arr}
\alias{cmp_ppn_dtf}
\alias{cmp_ppn_gen}
\alias{cmp_ppn_mat}
\alias{cmp_ppn_mvc}
\alias{cmp_ppn_scl}
\alias{cmp_ppn_vec}
\alias{cmp_ppn_vls}
\alias{cmp_psw_arr}
\alias{cmp_psw_dtf}
\alias{cmp_psw_gen}
\alias{cmp_psw_mat}
\alias{cmp_psw_mvc}
\alias{cmp_psw_scl}
\alias{cmp_psw_vec}
\alias{cmp_psw_vls}
\alias{cmp_srt_arr}
\alias{cmp_srt_dtf}
\alias{cmp_srt_gen}
\alias{cmp_srt_mat}
\alias{cmp_srt_mvc}
\alias{cmp_srt_scl}
\alias{cmp_srt_vec}
\alias{cmp_srt_vls}
\alias{cmp_str_arr}
\alias{cmp_str_dtf}
\alias{cmp_str_gen}
\alias{cmp_str_mat}
\alias{cmp_str_mvc}
\alias{cmp_str_scl}
\alias{cmp_str_vec}
\alias{cmp_str_vls}
\alias{cmp_uno_arr}
\alias{cmp_uno_dtf}
\alias{cmp_uno_gen}
\alias{cmp_uno_mat}
\alias{cmp_uno_mvc}
\alias{cmp_uno_scl}
\alias{cmp_uno_vec}
\alias{cmp_uno_vls}
\alias{cmp_whl_arr}
\alias{cmp_whl_dtf}
\alias{cmp_whl_gen}
\alias{cmp_whl_mat}
\alias{cmp_whl_mvc}
\alias{cmp_whl_scl}
\alias{cmp_whl_vec}
\alias{cmp_whl_vls}
\title{Complete + \code{xmode} + \code{xclass} properties}
\usage{
cmp_mmm_ccc(x, mmm, ccc, ...)

cmp_mmm_ccc_props()

cmp_ch1_arr(x, ...)

cmp_ch1_dtf(x, ...)

cmp_ch1_gen(x, ...)

cmp_ch1_mat(x, ...)

cmp_ch1_mvc(x, ...)

cmp_ch1_scl(x, ...)

cmp_ch1_vec(x, ...)

cmp_ch1_vls(x, ...)

cmp_ch3_arr(x, ...)

cmp_ch3_dtf(x, ...)

cmp_ch3_gen(x, ...)

cmp_ch3_mat(x, ...)

cmp_ch3_mvc(x, ...)

cmp_ch3_scl(x, ...)

cmp_ch3_vec(x, ...)

cmp_ch3_vls(x, ...)

cmp_chr_arr(x, ...)

cmp_chr_dtf(x, ...)

cmp_chr_gen(x, ...)

cmp_chr_mat(x, ...)

cmp_chr_mvc(x, ...)

cmp_chr_scl(x, ...)

cmp_chr_vec(x, ...)

cmp_chr_vls(x, ...)

cmp_clr_arr(x, ...)

cmp_clr_dtf(x, ...)

cmp_clr_gen(x, ...)

cmp_clr_mat(x, ...)

cmp_clr_mvc(x, ...)

cmp_clr_scl(x, ...)

cmp_clr_vec(x, ...)

cmp_clr_vls(x, ...)

cmp_evn_arr(x, ...)

cmp_evn_dtf(x, ...)

cmp_evn_gen(x, ...)

cmp_evn_mat(x, ...)

cmp_evn_mvc(x, ...)

cmp_evn_scl(x, ...)

cmp_evn_vec(x, ...)

cmp_evn_vls(x, ...)

cmp_fac_arr(x, ...)

cmp_fac_dtf(x, ...)

cmp_fac_gen(x, ...)

cmp_fac_mat(x, ...)

cmp_fac_mvc(x, ...)

cmp_fac_scl(x, ...)

cmp_fac_vec(x, ...)

cmp_fac_vls(x, ...)

cmp_frc_arr(x, ...)

cmp_frc_dtf(x, ...)

cmp_frc_gen(x, ...)

cmp_frc_mat(x, ...)

cmp_frc_mvc(x, ...)

cmp_frc_scl(x, ...)

cmp_frc_vec(x, ...)

cmp_frc_vls(x, ...)

cmp_ind_arr(x, ...)

cmp_ind_dtf(x, ...)

cmp_ind_gen(x, ...)

cmp_ind_mat(x, ...)

cmp_ind_mvc(x, ...)

cmp_ind_scl(x, ...)

cmp_ind_vec(x, ...)

cmp_ind_vls(x, ...)

cmp_lgl_arr(x, ...)

cmp_lgl_dtf(x, ...)

cmp_lgl_gen(x, ...)

cmp_lgl_mat(x, ...)

cmp_lgl_mvc(x, ...)

cmp_lgl_scl(x, ...)

cmp_lgl_vec(x, ...)

cmp_lgl_vls(x, ...)

cmp_neg_arr(x, ...)

cmp_neg_dtf(x, ...)

cmp_neg_gen(x, ...)

cmp_neg_mat(x, ...)

cmp_neg_mvc(x, ...)

cmp_neg_scl(x, ...)

cmp_neg_vec(x, ...)

cmp_neg_vls(x, ...)

cmp_ngw_arr(x, ...)

cmp_ngw_dtf(x, ...)

cmp_ngw_gen(x, ...)

cmp_ngw_mat(x, ...)

cmp_ngw_mvc(x, ...)

cmp_ngw_scl(x, ...)

cmp_ngw_vec(x, ...)

cmp_ngw_vls(x, ...)

cmp_nng_arr(x, ...)

cmp_nng_dtf(x, ...)

cmp_nng_gen(x, ...)

cmp_nng_mat(x, ...)

cmp_nng_mvc(x, ...)

cmp_nng_scl(x, ...)

cmp_nng_vec(x, ...)

cmp_nng_vls(x, ...)

cmp_nnw_arr(x, ...)

cmp_nnw_dtf(x, ...)

cmp_nnw_gen(x, ...)

cmp_nnw_mat(x, ...)

cmp_nnw_mvc(x, ...)

cmp_nnw_scl(x, ...)

cmp_nnw_vec(x, ...)

cmp_nnw_vls(x, ...)

cmp_nps_arr(x, ...)

cmp_nps_dtf(x, ...)

cmp_nps_gen(x, ...)

cmp_nps_mat(x, ...)

cmp_nps_mvc(x, ...)

cmp_nps_scl(x, ...)

cmp_nps_vec(x, ...)

cmp_nps_vls(x, ...)

cmp_npw_arr(x, ...)

cmp_npw_dtf(x, ...)

cmp_npw_gen(x, ...)

cmp_npw_mat(x, ...)

cmp_npw_mvc(x, ...)

cmp_npw_scl(x, ...)

cmp_npw_vec(x, ...)

cmp_npw_vls(x, ...)

cmp_nst_arr(x, ...)

cmp_nst_dtf(x, ...)

cmp_nst_gen(x, ...)

cmp_nst_mat(x, ...)

cmp_nst_mvc(x, ...)

cmp_nst_scl(x, ...)

cmp_nst_vec(x, ...)

cmp_nst_vls(x, ...)

cmp_num_arr(x, ...)

cmp_num_dtf(x, ...)

cmp_num_gen(x, ...)

cmp_num_mat(x, ...)

cmp_num_mvc(x, ...)

cmp_num_scl(x, ...)

cmp_num_vec(x, ...)

cmp_num_vls(x, ...)

cmp_odd_arr(x, ...)

cmp_odd_dtf(x, ...)

cmp_odd_gen(x, ...)

cmp_odd_mat(x, ...)

cmp_odd_mvc(x, ...)

cmp_odd_scl(x, ...)

cmp_odd_vec(x, ...)

cmp_odd_vls(x, ...)

cmp_ord_arr(x, ...)

cmp_ord_dtf(x, ...)

cmp_ord_gen(x, ...)

cmp_ord_mat(x, ...)

cmp_ord_mvc(x, ...)

cmp_ord_scl(x, ...)

cmp_ord_vec(x, ...)

cmp_ord_vls(x, ...)

cmp_pct_arr(x, ...)

cmp_pct_dtf(x, ...)

cmp_pct_gen(x, ...)

cmp_pct_mat(x, ...)

cmp_pct_mvc(x, ...)

cmp_pct_scl(x, ...)

cmp_pct_vec(x, ...)

cmp_pct_vls(x, ...)

cmp_pos_arr(x, ...)

cmp_pos_dtf(x, ...)

cmp_pos_gen(x, ...)

cmp_pos_mat(x, ...)

cmp_pos_mvc(x, ...)

cmp_pos_scl(x, ...)

cmp_pos_vec(x, ...)

cmp_pos_vls(x, ...)

cmp_ppn_arr(x, ...)

cmp_ppn_dtf(x, ...)

cmp_ppn_gen(x, ...)

cmp_ppn_mat(x, ...)

cmp_ppn_mvc(x, ...)

cmp_ppn_scl(x, ...)

cmp_ppn_vec(x, ...)

cmp_ppn_vls(x, ...)

cmp_psw_arr(x, ...)

cmp_psw_dtf(x, ...)

cmp_psw_gen(x, ...)

cmp_psw_mat(x, ...)

cmp_psw_mvc(x, ...)

cmp_psw_scl(x, ...)

cmp_psw_vec(x, ...)

cmp_psw_vls(x, ...)

cmp_srt_arr(x, ...)

cmp_srt_dtf(x, ...)

cmp_srt_gen(x, ...)

cmp_srt_mat(x, ...)

cmp_srt_mvc(x, ...)

cmp_srt_scl(x, ...)

cmp_srt_vec(x, ...)

cmp_srt_vls(x, ...)

cmp_str_arr(x, ...)

cmp_str_dtf(x, ...)

cmp_str_gen(x, ...)

cmp_str_mat(x, ...)

cmp_str_mvc(x, ...)

cmp_str_scl(x, ...)

cmp_str_vec(x, ...)

cmp_str_vls(x, ...)

cmp_uno_arr(x, ...)

cmp_uno_dtf(x, ...)

cmp_uno_gen(x, ...)

cmp_uno_mat(x, ...)

cmp_uno_mvc(x, ...)

cmp_uno_scl(x, ...)

cmp_uno_vec(x, ...)

cmp_uno_vls(x, ...)

cmp_whl_arr(x, ...)

cmp_whl_dtf(x, ...)

cmp_whl_gen(x, ...)

cmp_whl_mat(x, ...)

cmp_whl_mvc(x, ...)

cmp_whl_scl(x, ...)

cmp_whl_vec(x, ...)

cmp_whl_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{mmm}{A character scalar \code{xmode} property from \code{mmm_props()}.}

\item{ccc}{A character scalar \code{xclass} property from \code{ccc_props()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{cmp_mmm_ccc_props}   \tab A character vector.
\cr                     \tab   
\cr     \code{cmp_mmm_ccc}   \tab A logical scalar.
\cr     \code{cmp_MMM_CCC}   \tab   
}
}
\description{
\tabular{rl}{
  \code{cmp_mmm_ccc_props}   \tab Gets all possible \link[=icmp]{complete} + \link[=mmm]{\code{xmode}} + \link[=ccc]{\code{xclass}} properties.
\cr                     \tab   
\cr     \code{cmp_mmm_ccc}   \tab Is \code{x} complete and does it match \code{xmode} and \code{xclass} specs in arguments \code{mmm} and \code{ccc}, respectively?
\cr                     \tab   
\cr     \code{cmp_MMM_CCC}   \tab Is \code{x} complete and does it match \code{xmode} and \code{xclass} properties \code{'MMM'} and \code{'CCC'}, respectively?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
cmp_mmm_ccc_props()
cmp_mmm_ccc(letters, "ch1", "vec")
cmp_mmm_ccc(letters, "str", "scl")
cmp_ch1_vec(letters)
cmp_str_scl("a")
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
