% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_gen.R
\name{cmp_mmm_gen}
\alias{cmp_mmm_gen}
\alias{cmp_mmm_gen_vals}
\alias{cmp_atm_gen}
\alias{cmp_chr_gen}
\alias{cmp_ch1_gen}
\alias{cmp_clr_gen}
\alias{cmp_evn_gen}
\alias{cmp_fac_gen}
\alias{cmp_frc_gen}
\alias{cmp_ind_gen}
\alias{cmp_lgl_gen}
\alias{cmp_neg_gen}
\alias{cmp_ngw_gen}
\alias{cmp_nng_gen}
\alias{cmp_nnw_gen}
\alias{cmp_nps_gen}
\alias{cmp_npw_gen}
\alias{cmp_nst_gen}
\alias{cmp_num_gen}
\alias{cmp_odd_gen}
\alias{cmp_ord_gen}
\alias{cmp_pct_gen}
\alias{cmp_pos_gen}
\alias{cmp_ppn_gen}
\alias{cmp_psw_gen}
\alias{cmp_srt_gen}
\alias{cmp_str_gen}
\alias{cmp_uno_gen}
\alias{cmp_whl_gen}
\title{Complete + Extended Mode + Atomic Generic (cmp + mmm + avl; includes
atomic tibbles and atomic vlists)}
\usage{
cmp_mmm_gen_vals()

cmp_atm_gen(x)

cmp_chr_gen(x)

cmp_ch1_gen(x)

cmp_clr_gen(x)

cmp_evn_gen(x)

cmp_fac_gen(x)

cmp_frc_gen(x)

cmp_ind_gen(x)

cmp_lgl_gen(x)

cmp_neg_gen(x)

cmp_ngw_gen(x)

cmp_nng_gen(x)

cmp_nnw_gen(x)

cmp_nps_gen(x)

cmp_npw_gen(x)

cmp_nst_gen(x)

cmp_num_gen(x)

cmp_odd_gen(x)

cmp_ord_gen(x)

cmp_pct_gen(x)

cmp_pos_gen(x)

cmp_ppn_gen(x)

cmp_psw_gen(x)

cmp_srt_gen(x)

cmp_str_gen(x)

cmp_uno_gen(x)

cmp_whl_gen(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible complete + extended mode + atomic generic
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_gen}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_gen}: Is \code{x} a complete atomic generic?

\item \code{cmp_chr_gen}: Is \code{x} a complete character atomic generic?

\item \code{cmp_ch1_gen}: Is \code{x} a complete onechar atomic generic?

\item \code{cmp_clr_gen}: Is \code{x} a complete color atomic generic?

\item \code{cmp_evn_gen}: Is \code{x} a complete even-number atomic generic?

\item \code{cmp_fac_gen}: Is \code{x} a complete factor atomic generic?

\item \code{cmp_frc_gen}: Is \code{x} a complete fractional-numeric atomic
generic?

\item \code{cmp_ind_gen}: Is \code{x} a complete indexer atomic generic?

\item \code{cmp_lgl_gen}: Is \code{x} a complete logical atomic generic?

\item \code{cmp_neg_gen}: Is \code{x} a complete negative-number atomic generic?

\item \code{cmp_ngw_gen}: Is \code{x} a complete negative-whole-number atomic
generic?

\item \code{cmp_nng_gen}: Is \code{x} a complete non-negative-number atomic
generic?

\item \code{cmp_nnw_gen}: Is \code{x} a complete non-negative-whole-number
atomic generic?

\item \code{cmp_nps_gen}: Is \code{x} a complete non-positive-number atomic
generic?

\item \code{cmp_npw_gen}: Is \code{x} a complete non-positive-whole-number
atomic generic?

\item \code{cmp_nst_gen}: Is \code{x} a complete non-sortable atomic generic?

\item \code{cmp_num_gen}: Is \code{x} a complete number atomic generic?

\item \code{cmp_odd_gen}: Is \code{x} a complete odd-number atomic generic?

\item \code{cmp_ord_gen}: Is \code{x} a complete ordered-factor atomic generic?

\item \code{cmp_pct_gen}: Is \code{x} a complete percent-number atomic generic?

\item \code{cmp_pos_gen}: Is \code{x} a complete positive-number atomic
generic?

\item \code{cmp_ppn_gen}: Is \code{x} a complete proportion-number atomic
generic?

\item \code{cmp_psw_gen}: Is \code{x} a complete positive-whole-number atomic
generic?

\item \code{cmp_srt_gen}: Is \code{x} a complete sortable atomic generic?

\item \code{cmp_str_gen}: Is \code{x} a complete string atomic generic?

\item \code{cmp_uno_gen}: Is \code{x} a complete unordered-factor atomic
generic?

\item \code{cmp_whl_gen}: Is \code{x} a complete whole-number atomic generic?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
