% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_mmm_scl.R
\name{cmp_mmm_scl}
\alias{cmp_mmm_scl}
\alias{cmp_mmm_scl_vals}
\alias{cmp_atm_scl}
\alias{cmp_chr_scl}
\alias{cmp_ch1_scl}
\alias{cmp_clr_scl}
\alias{cmp_evn_scl}
\alias{cmp_fac_scl}
\alias{cmp_frc_scl}
\alias{cmp_ind_scl}
\alias{cmp_lgl_scl}
\alias{cmp_neg_scl}
\alias{cmp_ngw_scl}
\alias{cmp_nng_scl}
\alias{cmp_nnw_scl}
\alias{cmp_nps_scl}
\alias{cmp_npw_scl}
\alias{cmp_nst_scl}
\alias{cmp_num_scl}
\alias{cmp_odd_scl}
\alias{cmp_ord_scl}
\alias{cmp_pct_scl}
\alias{cmp_pos_scl}
\alias{cmp_ppn_scl}
\alias{cmp_psw_scl}
\alias{cmp_srt_scl}
\alias{cmp_str_scl}
\alias{cmp_uno_scl}
\alias{cmp_whl_scl}
\title{Complete + Extended Mode + Atomic Scalar (cmp + mmm + scl)}
\usage{
cmp_mmm_scl_vals()

cmp_atm_scl(x)

cmp_chr_scl(x)

cmp_ch1_scl(x)

cmp_clr_scl(x)

cmp_evn_scl(x)

cmp_fac_scl(x)

cmp_frc_scl(x)

cmp_ind_scl(x)

cmp_lgl_scl(x)

cmp_neg_scl(x)

cmp_ngw_scl(x)

cmp_nng_scl(x)

cmp_nnw_scl(x)

cmp_nps_scl(x)

cmp_npw_scl(x)

cmp_nst_scl(x)

cmp_num_scl(x)

cmp_odd_scl(x)

cmp_ord_scl(x)

cmp_pct_scl(x)

cmp_pos_scl(x)

cmp_ppn_scl(x)

cmp_psw_scl(x)

cmp_srt_scl(x)

cmp_str_scl(x)

cmp_uno_scl(x)

cmp_whl_scl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Get all possible complete + extended mode + atomic scalar
properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{cmp_scl}}.
}
\section{Functions}{
\itemize{
\item \code{cmp_atm_scl}: Is \code{x} a complete atomic scalar?

\item \code{cmp_chr_scl}: Is \code{x} a complete character atomic scalar?

\item \code{cmp_ch1_scl}: Is \code{x} a complete onechar atomic scalar?

\item \code{cmp_clr_scl}: Is \code{x} a complete color atomic scalar?

\item \code{cmp_evn_scl}: Is \code{x} a complete even-numeric atomic scalar?

\item \code{cmp_fac_scl}: Is \code{x} a complete factor atomic scalar?

\item \code{cmp_frc_scl}: Is \code{x} a complete fractional-numeric atomic
scalar?

\item \code{cmp_ind_scl}: Is \code{x} a complete indexer atomic scalar?

\item \code{cmp_lgl_scl}: Is \code{x} a complete logical atomic scalar?

\item \code{cmp_neg_scl}: Is \code{x} a complete negative-numeric atomic
scalar?

\item \code{cmp_ngw_scl}: Is \code{x} a complete negative-whole-numeric atomic
scalar?

\item \code{cmp_nng_scl}: Is \code{x} a complete non-negative-numeric atomic
scalar?

\item \code{cmp_nnw_scl}: Is \code{x} a complete non-negative-whole-numeric
atomic scalar?

\item \code{cmp_nps_scl}: Is \code{x} a complete non-positive-numeric atomic
scalar?

\item \code{cmp_npw_scl}: Is \code{x} a complete non-positive-whole-numeric
atomic scalar?

\item \code{cmp_nst_scl}: Is \code{x} a complete non-sortable atomic scalar?

\item \code{cmp_num_scl}: Is \code{x} a complete numeric atomic scalar?

\item \code{cmp_odd_scl}: Is \code{x} a complete odd-numeric atomic scalar?

\item \code{cmp_ord_scl}: Is \code{x} a complete ordered-factor atomic scalar?

\item \code{cmp_pct_scl}: Is \code{x} a complete percent-numeric atomic scalar?

\item \code{cmp_pos_scl}: Is \code{x} a complete positive-numeric atomic
scalar?

\item \code{cmp_ppn_scl}: Is \code{x} a complete proportion-numeric atomic
scalar?

\item \code{cmp_psw_scl}: Is \code{x} a complete positive-whole-numeric atomic
scalar?

\item \code{cmp_srt_scl}: Is \code{x} a complete sortable atomic scalar?

\item \code{cmp_str_scl}: Is \code{x} a complete string atomic scalar?

\item \code{cmp_uno_scl}: Is \code{x} a complete unordered-factor atomic
scalar?

\item \code{cmp_whl_scl}: Is \code{x} a complete whole-numeric atomic scalar?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
