% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{color_aliases}
\alias{color_aliases}
\alias{.blk}
\alias{.blu}
\alias{.cyn}
\alias{.grn}
\alias{.mag}
\alias{.red}
\alias{.sil}
\alias{.wht}
\alias{.yel}
\alias{.bgs}
\alias{.fgs}
\title{Color Aliases}
\usage{
color_aliases()

.blk()

.blu()

.cyn()

.grn()

.mag()

.red()

.sil()

.wht()

.yel()

.bgs()

.fgs()
}
\value{
A character vector.
}
\description{
Each color is assigned a set of aliases, all of which will work in this package.
}
\section{Functions}{
\itemize{
\item \code{.blk()}: Aliases for "black".

\item \code{.blu()}: Aliases for "blue".

\item \code{.cyn()}: Aliases for "cyan".

\item \code{.grn()}: Aliases for "green".

\item \code{.mag()}: Aliases for "magenta".

\item \code{.red()}: Aliases for "red".

\item \code{.sil()}: Aliases for "silver".

\item \code{.wht()}: Aliases for "white".

\item \code{.yel()}: Aliases for "yellow".

\item \code{.bgs()}: All aliases for background colors.

\item \code{.fgs()}: All aliases for foreground colors.

}}
\examples{
.blk()
.blu()
.cyn()
.grn()
.mag()
.red()
.sil()
.wht()
.yel()
.bgs()
.fgs()
}
