% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{combo2props}
\alias{combo2props}
\title{Split a Property Combos into a Vector of Single Properties}
\usage{
combo2props(x)
}
\arguments{
\item{x}{A character scalar.}
}
\value{
Either \code{NULL} or a character vector (returns \code{NULL} when \code{x} is not of mode \code{'character'} or is of length zero).
}
\description{
Splits a single property combo (underscore separated) into its constituent single properties.
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{compat}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{not_err}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}
}
\concept{utils}
