% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparable.R
\encoding{UTF-8}
\name{comparable}
\alias{comparable}
\title{Are objects comparable?}
\usage{
comparable(..., rec. = FALSE)
}
\arguments{
\item{...}{An arbitrary number of arguments to be checked for comparability with each other.}

\item{rec.}{A non-\code{NA} logical scalar indicating whether \code{...} arguments must be recyclable to be comparable.}
}
\value{
A logical scalar.
}
\description{
Determines whether modes of all \code{...} arguments are comparable (i.e., sortable and compatible with each other), meaning that the are either all character, all logical, all numeric, or all ordered factor with the same set of levels (in the same order).
}
\examples{
az. <- letters
comparable(1:10, pi, exp(1))
comparable(az., az., "", "!")
comparable(az., factor(az.))
comparable(1:10, 1:3, rec. = TRUE)
comparable(list(az., 1:10), az., 1:10)
comparable(factor(az., az., ordered = T), factor("q", az., ordered = T))
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()}

Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{extensions}
\concept{properties}
