% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{compat}
\alias{compat}
\title{Evaluate Any Object, Returning a Valid Result, Always}
\usage{
compat(x, y)
}
\arguments{
\item{x}{An object to evaluate, even if doing so would result in an error,.}

\item{def}{The value to return if evaluating \code{x} results in an error.}

\item{x, y}{Objects to evaluate.}
}
\value{
Either \code{x} or \code{def}.

A logical scalar.
}
\description{
Evaluate Any Object, Returning a Valid Result, Always

Evaluate Whether \code{x} and \code{y} Are of Compatible Modes
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{not_err}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}

Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{not_err}()},
\code{\link{ox_vals}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}
}
\concept{utils}
