% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayons.R
\name{crayons}
\alias{crayons}
\alias{st_vals}
\alias{bg_vals}
\alias{fg_vals}
\alias{st_bld}
\alias{st_def}
\alias{st_itl}
\alias{st_pln}
\alias{st_und}
\alias{st}
\alias{fg_blk}
\alias{fg_blu}
\alias{fg_cyn}
\alias{fg_def}
\alias{fg_grn}
\alias{fg_gry}
\alias{fg_mag}
\alias{fg_red}
\alias{fg_sil}
\alias{fg_slv}
\alias{fg_wht}
\alias{fg_ylw}
\alias{fg}
\alias{bg_blk}
\alias{bg_blu}
\alias{bg_cyn}
\alias{bg_def}
\alias{bg_grn}
\alias{bg_mag}
\alias{bg_red}
\alias{bg_wht}
\alias{bg_ylw}
\alias{bg}
\alias{txt}
\title{Wrappers for package \code{\link[crayon]{crayon}} functions to styled and color console text}
\usage{
st_vals(.ST = NULL)

bg_vals(.BG = NULL)

fg_vals(.FG = NULL)

st_bld(..., .D = " ")

st_def(..., .D = " ")

st_itl(..., .D = " ")

st_pln(..., .D = " ")

st_und(..., .D = " ")

st(.ST, ..., .D = " ")

fg_blk(..., .D = " ")

fg_blu(..., .D = " ")

fg_cyn(..., .D = " ")

fg_def(..., .D = " ")

fg_grn(..., .D = " ")

fg_gry(..., .D = " ")

fg_mag(..., .D = " ")

fg_red(..., .D = " ")

fg_sil(..., .D = " ")

fg_slv(..., .D = " ")

fg_wht(..., .D = " ")

fg_ylw(..., .D = " ")

fg(.FG, ..., .D = " ")

bg_blk(..., .D = " ")

bg_blu(..., .D = " ")

bg_cyn(..., .D = " ")

bg_def(..., .D = " ")

bg_grn(..., .D = " ")

bg_mag(..., .D = " ")

bg_red(..., .D = " ")

bg_wht(..., .D = " ")

bg_ylw(..., .D = " ")

bg(.BG, ..., .D = " ")

txt(..., .BG = NULL, .FG = NULL, .ST = NULL, .D = " ")
}
\arguments{
\item{.ST}{A quoted or unquoted, case-insensitive, character scalar text style from \code{st_vals()}. May also be \code{NULL} for function \code{st_vals}.}

\item{.BG}{A quoted or unquoted, case-insensitive, character scalar text background color from \code{bg_vals()}. May also be \code{NULL} for function \code{bg_vals}.}

\item{.FG}{A quoted or unquoted, case-insensitive, character scalar text foreground color from \code{fg_vals()}. May also be \code{NULL} for function \code{fg_vals}.}

\item{...}{An arbitrary number of objects to be \link[=glue_dots]{collapsed} to a character scalar to be styled.}

\item{.D}{A non-\code{NA} character scalar delimiter for \link[=glue_dots]{collapsing} \code{...} args into a character scalar.}
}
\value{
\strong{A character vector} \cr\cr \verb{st_vals, fg_vals, bg_vals}
\cr\cr  \strong{A character scalar} \cr\cr All others
}
\description{
Functions in this family apply styles and colors to text for display on the console, where the text to display is identified by \link[=glue_dots]{collapsing} \code{...} args into a character scalar.
}
\section{Background color functions}{

\tabular{ll}{  \code{bg_vals}   \tab Get all valid values of arg \code{.BG}.                                    \cr
  \code{bg_red}    \tab Set to red (\code{\link[crayon:bgRed]{crayon::bgRed}}).              \cr
  \code{bg_blu}    \tab Set to blue (\code{\link[crayon:bgBlue]{crayon::bgBlue}}).           \cr
  \code{bg_cyn}    \tab Set to cyan (\code{\link[crayon:bgCyan]{crayon::bgCyan}}).           \cr
  \code{bg_blk}    \tab Set to black (\code{\link[crayon:bgBlack]{crayon::bgBlack}}).        \cr
  \code{bg_grn}    \tab Set to green (\code{\link[crayon:bgGreen]{crayon::bgGreen}}).        \cr
  \code{bg_wht}    \tab Set to white (\code{\link[crayon:bgWhite]{crayon::bgWhite}}).        \cr
  \code{bg_ylw}    \tab Set to yellow (\code{\link[crayon:bgYellow]{crayon::bgYellow}}).     \cr
  \code{bg_mag}    \tab Set to magenta (\code{\link[crayon:bgMagenta]{crayon::bgMagenta}}).  \cr
  \code{bg_def}    \tab Set to system default.                                               \cr
  \code{bg}        \tab Set to value of arg \code{.BG}.                                              }
\cr\cr Valid values of \code{.BG} are the following:
\tabular{ll}{  magenta   \tab \code{'m'}, \code{'mag'}, \code{'magenta'} \cr
  default   \tab \code{'d'}, \code{'def'}, \code{'default'} \cr
  yellow    \tab \code{'y'}, \code{'ylw'}, \code{'yellow'}  \cr
  black     \tab \code{'k'}, \code{'blk'}, \code{'black'}   \cr
  white     \tab \code{'w'}, \code{'wht'}, \code{'white'}   \cr
  green     \tab \code{'g'}, \code{'grn'}, \code{'green'}   \cr
  blue      \tab \code{'b'}, \code{'blu'}, \code{'blue'}    \cr
  cyan      \tab \code{'c'}, \code{'cyn'}, \code{'cyan'}    \cr
  red       \tab \code{'r'}, \code{'red'}                }
NOTE: Silver (grey) is \emph{not} valid for \emph{background}.
}

\section{Foreground color functions}{

\tabular{ll}{  \code{fg_vals}   \tab Get all valid values of \code{.FG} (foreground text color).              \cr
  \code{fg_red}    \tab Set to red (\code{\link[crayon:red]{crayon::red}}).                 \cr
  \code{fg_blu}    \tab Set to red (\code{\link[crayon:blue]{crayon::blue}}).               \cr
  \code{fg_cyn}    \tab Set to red (\code{\link[crayon:cyan]{crayon::cyan}}).               \cr
  \code{fg_blk}    \tab Set to red (\code{\link[crayon:black]{crayon::black}}).             \cr
  \code{fg_grn}    \tab Set to red (\code{\link[crayon:green]{crayon::green}}).             \cr
  \code{fg_wht}    \tab Set to red (\code{\link[crayon:white]{crayon::white}}).             \cr
  \code{fg_ylw}    \tab Set to red (\code{\link[crayon:yellow]{crayon::yellow}}).           \cr
  \code{fg_mag}    \tab Set to red (\code{\link[crayon:magenta]{crayon::magenta}}).         \cr
  \code{fg_gry}    \tab Set to silver (grey) (\code{\link[crayon:silver]{crayon::silver}}). \cr
  \code{fg_sil}    \tab Set to silver (grey).                                               \cr
  \code{fg_slv}    \tab Set to silver (grey).                                               \cr
  \code{fg_def}    \tab Set to system default.                                              \cr
  \code{fg}        \tab Set to value of arg \code{.FG}.                                            }
\cr\cr Valid values of \code{.FG} are :
\tabular{ll}{  silver (grey)   \tab \code{'s'}, \code{'sil'}, \code{'slv'}, \code{'silver'}, \code{'gry'}, \code{'gray'}, \code{'grey'} \cr
  magenta         \tab \code{'m'}, \code{'mag'}, \code{'magenta'}                                      \cr
  default         \tab \code{'.D'}, \code{'def'}, \code{'default'}                                     \cr
  yellow          \tab \code{'y'}, \code{'ylw'}, \code{'yellow'}                                       \cr
  black           \tab \code{'k'}, \code{'blk'}, \code{'black'}                                        \cr
  white           \tab \code{'w'}, \code{'wht'}, \code{'white'}                                        \cr
  green           \tab \code{'g'}, \code{'grn'}, \code{'green'}                                        \cr
  blue            \tab \code{'b'}, \code{'blu'}, \code{'blue'}                                         \cr
  cyan            \tab \code{'c'}, \code{'cyn'}, \code{'cyan'}                                         \cr
  red             \tab \code{'r'}, \code{'red'}                                                     }
NOTE: Silver (grey) is \emph{only} valid for \emph{foreground}.
}

\section{Style functions}{

\tabular{ll}{  \code{st_vals}   \tab Get all valid values of arg \code{.ST}.                              \cr
  \code{st_bld}    \tab Set to bold (\code{\link[crayon:bgRed]{crayon::bold}}).         \cr
  \code{st_pln}    \tab Set to plain (\code{\link[crayon:bgBlue]{crayon::reset}}).      \cr
  \code{st_itl}    \tab Set to italic (\code{\link[crayon:bgBlue]{crayon::italic}}).    \cr
  \code{st_und}    \tab Set to underline (\code{\link[crayon:bgBlue]{crayon::italic}}). \cr
  \code{st_def}    \tab Set to system default.                                          \cr
  \code{st}        \tab Set to value of arg \code{.ST}.                                        }
\cr\cr Valid values of \code{st} are:
\tabular{ll}{  underline   \tab \code{'u'}, \code{'un'}, \code{'und'}, \code{'under'}, \code{'underline'}, \code{'underlined'}                                                                         \cr   \tab   \cr
  default     \tab \code{'.D'}, \code{'def'}, \code{'default'}                                                                                                             \cr   \tab   \cr
  italic      \tab \code{'i'}, \code{'it'}, \code{'itl'}, \code{'ital'}, \code{'italic'}, \code{'italics'}, \code{'italicized'},\code{'e'}, \code{'em'}, \code{'emp'}, \code{'emph'}, \code{'emphasis'}, \code{'emphasized'} \cr   \tab   \cr
  plain       \tab \code{'p'},\code{ 'pl'}, \code{'pln'}, \code{'plain'}, \code{'r'}, \code{'re'}, \verb{'res}', \code{'reset'}                                                                     \cr   \tab   \cr
  bold        \tab \code{'b'}, \code{'bo'}, \code{'bld'}, \code{'bold'}, \code{'bolded'}, \code{'s'}, \code{'st'}, \code{'str'}, \code{'strong'}                                                                        }
}

\section{Generic function}{

\tabular{ll}{  \code{txt}   \tab Styles text using args \code{.ST}, \code{.BG}, \code{.FG}.}
}

\examples{
egArg1 <- "Some "
egArg2 <- c("text", " to display", "\nP.S. ")

bg_vals()
bg_vals("k")
bg_vals("mag")
bg_vals("yellow")

fg_vals()
fg_vals("s")
fg_vals("blu")
fg_vals("white")

st_vals()
st_vals("b")
st_vals("itl")
st_vals("plain")
st_vals("underline")

cat(st_bld(egArg1, egArg2, "note"))
cat(st_def(egArg1, egArg2, "note"))
cat(st_itl(egArg1, egArg2, "note"))
cat(st_pln(egArg1, egArg2, "note"))
cat(st_und(egArg1, egArg2, "note"))
cat(st_itl(st_und(egArg1, egArg2, "note")))

cat(st("u", egArg1, egArg2, "note"))
cat(st("itl", egArg1, egArg2, "note"))
cat(st("under", egArg1, egArg2, "note"))
cat(st("underline", st("itl", egArg1, egArg2, "note")))

cat(fg_blk(egArg1, egArg2))
cat(fg_blu(egArg1, egArg2))
cat(fg_cyn(egArg1, egArg2))
cat(fg_def(egArg1, egArg2))
cat(fg_grn(egArg1, egArg2))
cat(fg_gry(egArg1, egArg2))
cat(fg_sil(egArg1, egArg2))
cat(fg_slv(egArg1, egArg2))
cat(fg_mag(egArg1, egArg2))
cat(fg_red(egArg1, egArg2))
cat(fg_wht(egArg1, egArg2))
cat(fg_ylw(egArg1, egArg2))
cat(fg_def(egArg1, egArg2))

cat(fg("s", egArg1, egArg2, "note"))
cat(fg("sil", egArg1, egArg2, "note"))
cat(fg("grey", egArg1, egArg2, "note"))

cat(bg_blk(egArg1, egArg2, "note"))
cat(bg_blu(egArg1, egArg2, "note"))
cat(bg_cyn(egArg1, egArg2, "note"))
cat(bg_def(egArg1, egArg2, "note"))
cat(bg_grn(egArg1, egArg2, "note"))
cat(bg_mag(egArg1, egArg2, "note"))
cat(bg_red(egArg1, egArg2, "note"))
cat(bg_wht(egArg1, egArg2, "note"))
cat(bg_ylw(egArg1, egArg2, "note"))
cat(bg_def(egArg1, egArg2, "note"))

cat(bg("y", egArg1, egArg2, "note"))
cat(bg("blu", egArg1, egArg2, "note"))
cat(bg("yellow", egArg1, egArg2, "note"))

cat(bg("y", fg("r", st("u", st("i", egArg1, egArg2, "note")))))
cat(txt(egArg1, egArg2, "note", .BG = "blue", .FG = "w", .ST = c("i", "under")))

}
