% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\name{ddd.}
\alias{ddd.}
\alias{ddd}
\alias{nddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{ddd_vals}
\alias{iddd}
\title{Defined dimensionality properties}
\usage{
ddd.()

ddd(x)

nddd(x)

id0D(x)

id1D(x)

id2D(x)

idHD(x)

ddd_vals()

iddd(x, ddd, ...)
}
\arguments{
\item{x}{An R object.}

\item{ddd}{A \link[=cmp_chr_scl]{complete character scalar} containing one or
more values from \code{ddd_vals()} separated by pipes and/or underscores.
Combinations of defined dimensionality properties can be specified by
separating them with underscores. Separating defined dimensionality
properties or combinations of defined dimensionality properties with pipes
will result in a value of \code{TRUE} if any of them applies to \code{x}.}

\item{...}{
  Arguments passed on to \code{\link[=meets.]{meets.}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{lll}{
\code{nddd}                   \tab  \tab A non-negative whole-number
scalar.                        \cr
\code{ddd_vals} and \code{ddd}\tab  \tab A character scalar or vector.  \cr
\code{iddd} and \code{i[ddd]} \tab  \tab A logical scalar.                }
}
\description{
An object's defined dimensionality is the number of dimensions
on which its components can be indexed. The following table gives levels of
defined dimensionality, property values assigned to levels of defined
dimensionality, property names, and a definition of each value of defined
dimensionality.\tabular{llll}{
          \tab DEFINED          \tab DEFINED          \tab                \cr
NUMBER    \tab DIMENSIONALITY   \tab DIMENSIONALITY   \tab CHARACTERISTICS\cr
OF DEFINED\tab PROPERTY         \tab PROPERTY         \tab OF QUALIFYING  \cr
DIMENSIONS   \tab VALUE         \tab NAME             \tab OBJECTS        \cr
\code{0}     \tab\code{'d0D'}   \tab 0-dimensional structure.
\tab \code{NULL}.                         \cr
\code{1}     \tab\code{'d1D'}   \tab 1-dimensional structure.
\tab Vector, \link[=ivls]{vlist},
1-dimensional array.                 \cr
\code{2}     \tab\code{'d2D'}   \tab 2-dimensional structure.
\tab data.frame or matrix.                \cr
\code{≥ 3}   \tab\code{'dHD'}   \tab Hyper-dimensional structure.
\tab Array with 3+ dimensions.              }
Functions related to defined dimensionality are described in the following
table:\tabular{ll}{
FUNCTION          \tab WHAT THE                                           \cr
FORMAT            \tab FUNCTION DOES                                      \cr
\code{i[ddd]}     \tab Evaluates whether an object is of the defined
dimensionality represented by \code{[ddd]}.        \cr
\code{ddd}        \tab Gets a character scalar containing the defined
dimensionality of an object.                       \cr
\code{iddd}       \tab Evaluates an object for a specific defined
dimensionality and any additional properties specified
in \code{...}.                                     \cr
\code{ddd_vals}   \tab Gets a character vector of all possible defined
dimensionality property values.                      }
}
\section{Specifying Additional Property Requirements}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar value all values of \code{x} must be less
than.                                                \cr
\code{le}     \tab Scalar value all values of \code{x} must be less than or
equal to.                                            \cr
\code{ge}     \tab Scalar value all values of \code{x} must be greater than
or equal to.                                         \cr
\code{gt}     \tab Scalar value all values of \code{x} must be greater
than.                                                  }
}

\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
