% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\name{ddd}
\alias{ddd}
\alias{ddd_props}
\alias{is_ddd_spec}
\alias{iddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{nddd}
\title{Defined dimensionality properties}
\usage{
ddd(x)

ddd_props()

is_ddd_spec(spec)

iddd(x, spec, ...)

id0D(x, ...)

id1D(x, ...)

id2D(x, ...)

idHD(x, ...)

nddd(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec}
containing one or more defined dimensionality properties (i.e., from
\code{ddd_vals()}). \strong{NOTE}: properties may be pipe-separated. If
If there are multiple properties in \code{spec}, \code{x} is inspected for
a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{\code{ddd_vals}}: A character vector.
\cr\cr\strong{\code{ddd}}: A character scalar or character vector.
\cr\cr\strong{\code{iDDD, iddd, is_ddd_spec}}: A logical scalar.
}
\description{
NOTE: \code{DDD} is used as a wildcard representing any given
effective dimensionality property.
\cr\cr
An object's defined dimensionality is the number of dimensions
on which its components can be indexed. The following table gives levels of
defined dimensionality, property values assigned to levels of defined
dimensionality, property names, and a definition of each value of defined
dimensionality.\tabular{llll}{
        \tab DEFINED          \tab DEFINED          \tab                \cr
NUMBER  \tab DIMENSIONALITY   \tab DIMENSIONALITY   \tab CHARACTERISTICS\cr
OF DEFINED\tab PROPERTY       \tab PROPERTY         \tab OF QUALIFYING  \cr
DIMENSIONS   \tab VALUE       \tab NAME             \tab OBJECTS        \cr
\code{0}     \tab\code{'d0D'} \tab 0-dimensional structure.
\tab \code{NULL}.                         \cr
\code{1}     \tab\code{'d1D'} \tab 1-dimensional structure.
\tab Vector, \link[=ivls]{vlist},
1-dimensional array.                 \cr
\code{2}     \tab\code{'d2D'} \tab 2-dimensional structure.
\tab data.frame or matrix.                \cr
\code{≥ 3}   \tab\code{'dHD'} \tab Hyper-dimensional structure.
\tab Array with 3+ dimensions.              }
Functions related to defined dimensionality properties are as follows:
\tabular{ll}{
DDD FUNCTION         \tab WHAT IT DOES                                \cr
\code{iDDD}          \tab Evaluates whether \code{x} matches the
defined dimensionality property \code{DDD}
(subject to any restrictions in \code{...}).\cr
\code{ddd}           \tab Gets a character vector containing all
defined dimensionality properties matching
\code{x}.                                   \cr
\code{iddd}          \tab Evaluates \code{x} against the
defined dimensionality property specification
in \code{spec} (subject to any restrictions in
\code{...}).                                \cr
\code{ddd_props}     \tab Gets a character vector of all possible
defined dimensionality property values.     \cr
\code{is_ddd_spec}   \tab Evaluates whether \code{spec} is a valid
defined dimensionality property
specification.                                }
}
\section{Specifying Count and Value Restrictions}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar less-than bound on values of \code{x}.        \cr
\code{le}     \tab Scalar less-than-or-equal bound on values of
\code{x}.                                            \cr
\code{ge}     \tab Scalar greater-than-or-equal bound on values of
\code{x}.                                            \cr
\code{gt}     \tab Scalar greater-than bound on values of \code{x}.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc_props}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
