% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\name{ddd}
\alias{ddd}
\alias{nddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{ddd_vals}
\alias{iddd}
\title{Defined Dimensionality (ddd) Property Family}
\usage{
ddd(x.)

nddd(x.)

id0D(x.)

id1D(x.)

id2D(x.)

idHD(x.)

ddd_vals()

iddd(x., xxx., ...)
}
\arguments{
\item{x.}{An object.}

\item{xxx.}{A character scalar containing one or more values from
\code{ddd_vals()} separated by pipes and/or underscores. Combinations of
defined dimensionality properties can be specified by separating them with
underscores. Separating defined dimensionality properties or combinations
of defined dimensionality properties with pipes will result in a value of
\code{TRUE} if any of them applies to \code{x}.}

\item{...}{Additional arguments to \code{\link{meets}} containing value and
element/row/column count restrictions.}
}
\value{
\code{nddd} returns a non-negative whole-number scalar.
\code{ddd_vals} and \code{ddd} return a character scalar or vector. All
others return either \code{TRUE} or \code{FALSE}.
}
\description{
Gets a vector of all defined dimensionality properties from
\code{ddd_vals()} that are applicable to \code{x.}.
}
\details{
An object's defined dimensionality (def. dim.) is the number of
dimensions on which its components can be indexed. The following table
gives levels of defined dimensionality, property values assigned to levels
of defined dimensionality, property names, and a definition of each value
of defined dimensionality.\tabular{llll}{
NUMBER OF  \tab DEF. DIM.  \tab DEF. DIM. \tab TYPES OF                \cr
DEFINED    \tab PROPERTY   \tab PROPERTY  \tab QUALIFYING              \cr
DIMENSIONS \tab VALUE      \tab NAME      \tab OBJECTS                 \cr
\code{0}   \tab\code{'d0D'}\tab of 0-dimensional structure 
\tab \code{NULL}                            \cr
\code{1}   \tab\code{'d1D'}\tab of 1-dimensional structure
\tab vector, vlist, 1-dimensional array     \cr
\code{2}   \tab\code{'d2D'}\tab of 2-dimensional structure 
\tab tibble or matrix                       \cr
\code{≥ 3} \tab\code{'dHD'}\tab of hyper-dimensional structure 
\tab array with 3+ dimensions.                }
Submitting additional arguments to \code{is_ddd} via \code{...} allows
for checking not just the defined dimensionality but whether length, number
of rows, number of columns, and element values meet flexible criteria.
}
\section{Functions}{
\itemize{
\item \code{nddd}: Gets the \strong{number} of defined dimensions of \code{x.}
(i.e., not a character scalar defined dimensionality property from
\code{ddd_vals}).

\item \code{id0D}: Evaluate whether \code{x.} has \code{0} defined dimensions.

\item \code{id1D}: Evaluate whether \code{x.} has \code{1} defined dimension.

\item \code{id2D}: Evaluate whether \code{x.} has \code{2} defined dimensions.

\item \code{idHD}: Evaluate whether \code{x.} has \code{3+} defined dimensions.

\item \code{ddd_vals}: Get a character vector of all possible defined dimensionality
property values.

\item \code{iddd}: Evaluates whether at least one of the defined dimensionality
properties in \code{xxx.} is applicable to \code{x.}.
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
