% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\encoding{UTF-8}
\name{ddd}
\alias{ddd}
\alias{ddd_funs}
\alias{ddd_props}
\alias{is_ddd_spec}
\alias{iddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{nddd}
\title{Defined dimensionality (defined.d) properties}
\usage{
ddd(x)

ddd_funs()

ddd_props()

is_ddd_spec(spec)

iddd(x, spec, ...)

id0D(x, ...)

id1D(x, ...)

id2D(x, ...)

idHD(x, ...)

nddd(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more defined.d properties from \code{ddd_props()}. defined.d properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector} \cr   \verb{ddd_props, ddd_funs, ddd}
\cr\cr \emph{A logical scalar} \cr   \verb{is_ddd_spec, iDDD, iddd}
\cr\cr \emph{A numeric scalar} \cr   \code{nddd}
}
\description{
An object's defined dimensionality (defined.d) is the number of dimensions on which its elements can be indexed. The associated property values are:
\tabular{rcl}{
  \code{'d0D'} \tab   \verb{0D}   \tab The \code{NULL} object.
\cr \code{'d1D'} \tab   \verb{1D}   \tab Vectors, \link[=ivls]{vlists}, \verb{1D} arrays.
\cr \code{'d2D'} \tab   \verb{2D}   \tab Data.frames and matrices.
\cr \code{'dHD'} \tab   \code{HD}   \tab Hyper (3+) dimensional arrays
}
\cr \emph{Functions} \tabular{rl}{
  \code{is_ddd_spec}   \tab Is \code{spec} a defined.d property specification?
\cr \code{ddd_props}   \tab What defined.d properties are there?
\cr  \code{ddd_funs}   \tab What defined.d property functions are there?
\cr      \code{nddd}   \tab How many defined dimensions does \code{x} have?
\cr      \code{iddd}   \tab Does \code{x} match defined.d specification \code{spec}?
\cr      \code{iDDD}   \tab Does \code{x} match single defined.d property \code{'DDD'}?
\cr       \code{ddd}   \tab What are \code{x}'s defined.d properties?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
ddd_funs()
ddd_props()
is_ddd_spec("d1D|d2D")
nddd(matrix(1))
nddd(letters)
nddd(1)
iddd(data.frame(letters), "d2D|dHD")
id0D(NULL)
id1D(NULL)
ddd(letters)
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
