% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\name{ddd}
\alias{ddd}
\alias{ddd_props}
\alias{is_ddd_spec}
\alias{iddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{nddd}
\title{Defined dimensionality (\code{defined-D}) properties}
\usage{
ddd(x)

ddd_props()

is_ddd_spec(spec)

iddd(x, spec, ...)

id0D(x, ...)

id1D(x, ...)

id2D(x, ...)

idHD(x, ...)

nddd(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{defined-D} properties from \code{ddd_props()}. \code{defined-D} properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{ddd_vals}   \tab A character vector
\cr   \code{nddd}   \tab A numeric scalar
\cr   \code{iddd}   \tab A logical scalar
\cr   \code{iDDD}   \tab A logical scalar
\cr    \code{ddd}   \tab A character vector
}
}
\description{
An object's defined dimensionality (\code{defined-D}) is the number of dimensions on which its elements can be indexed. The associated property values are:
\tabular{rcl}{
  \code{'d0D'} \tab   \verb{0D} \tab   The \code{NULL} object.
\cr \code{'d1D'} \tab   \verb{1D} \tab   Vectors, \link[=ivls]{vlists}, \verb{1D} arrays.
\cr \code{'d2D'} \tab   \verb{2D} \tab   Data.frames and matrices.
\cr \code{'dHD'} \tab   \code{HD} \tab   Hyper (3+) dimensional arrays
}
\strong{Functions}
\tabular{rl}{
  \code{is_ddd_spec}   \tab Is \code{spec} a \code{defined-D} property specification?
\cr               \tab   
\cr \code{ddd_props}   \tab What \code{defined-D} properties are there?
\cr               \tab   
\cr      \code{nddd}   \tab How many defined dimensions does \code{x} have?
\cr               \tab   
\cr      \code{iddd}   \tab Is \code{x} a match to the \code{defined-D}-specific \code{spec}?
\cr               \tab   
\cr      \code{iDDD}   \tab is \code{x} a match to the single \code{defined-D} property \code{'DDD'}?
\cr               \tab   
\cr       \code{ddd}   \tab What are \code{x}'s \code{defined-D} properties?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_ddd_spec("d1D|d2D")
nddd(matrix(1))
nddd(letters)
nddd(1)
ddd_props()
iddd(data.frame(letters), "d2D|dHD")
id0D(NULL)
id1D(NULL)
ddd(letters)
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_bbb_ccc_prop}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
