% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\name{ddd}
\alias{ddd}
\alias{ddd_props}
\alias{is_ddd_spec}
\alias{iddd}
\alias{id0D}
\alias{id1D}
\alias{id2D}
\alias{idHD}
\alias{nddd}
\title{Defined dimensionality properties}
\usage{
ddd(x)

ddd_props()

is_ddd_spec(spec)

iddd(x, spec, ...)

id0D(x, ...)

id1D(x, ...)

id2D(x, ...)

idHD(x, ...)

nddd(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more defined dimensionality properties from \code{ddd_props()}. Defined dimensionality properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\itemize{
\item \strong{\code{ddd_vals}}: a character vector.
\item \strong{\code{ddd}}: a character scalar/vector.
\item \strong{\code{nddd}}: a numeric scalar.
\item \strong{\verb{iddd, ixxx, is_ddd_spec}}: a logical scalar.
}
}
\description{
An object's defined dimensionality is the number of dimensions on which its elements can be indexed.
\itemize{
\item \strong{\code{'d0D'}}: \code{0}-dimensional structure (the \code{NULL} object).
\item \strong{\code{'d1D'}}: \code{1}-dimensional structure (vectors, \link[=ivls]{vlists}, and \code{1}-dimensional arrays).
\item \strong{\code{'d2D'}}: \code{2}-dimensional structure (data.frames and matrices).
\item \strong{\code{'dHD'}}: Hyper-dimensional structure (arrays with \verb{3+} dimensions).
}
\strong{Functions}
\itemize{
\item \strong{\code{ddd}}: gets a character vector containing all defined dimensionality properties possessed by \code{x}.
\item \strong{\code{ixxx}}: evaluates whether \code{x} possesses the defined dimensionality property \code{xxx} (a placeholder for any given defined dimensionality property value), subject to any restrictions in \code{...}.
\item \strong{\code{iddd}}: evaluates whether \code{x} possesses one or more (possibly pipe-delimited) defined dimensionality properties in \code{spec}, subject to any restrictions in \code{...}.
\item \strong{\code{nddd}}: gets the number of defined dimensions of \code{x}.
\item \strong{\code{ddd_props}}: gets a character vector of all possible defined dimensionality property values.
\item \strong{\code{is_ddd_spec}}: evaluates whether \code{spec} is a valid defined dimensionality property specification.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\itemize{
\item \strong{\code{n}}: a vector of valid lengths/numbers of elements.
\item \strong{\code{nr}}: a vector of valid numbers of rows.
\item \strong{\code{nc}}: a vector of valid numbers of columns.
\item \strong{\code{min}}: a scalar minimum valid length/number of element.
\item \strong{\code{minr}}: a scalar minimum valid number of rows.
\item \strong{\code{minc}}: a scalar minimum valid number of columns.
\item \strong{\code{max}}: a scalar maximum valid length/number of element.
\item \strong{\code{maxr}}: a scalar maximum valid number of rows.
\item \strong{\code{maxc}}: a scalar maximum valid number of columns.
\item \strong{\code{vals}}: a vector of valid values.
\item \strong{\code{lt}}: a scalar less-than (exclusive upper) bound.
\item \strong{\code{le}}: a scalar less-than-or-equal (inclusive upper) bound
\item \strong{\code{ge}}: a scalar greater-than-or-equal (inclusive lower) bound.
\item \strong{\code{gt}}: a scalar greater-than bound (exclusive lower) bound.
}
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
