% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd.R
\encoding{UTF-8}
\name{ddd}
\alias{ddd}
\alias{ddd_funs}
\alias{ddd_props}
\alias{is_ddd_spec}
\alias{DDD}
\alias{D0D}
\alias{D1D}
\alias{D2D}
\alias{DHD}
\alias{nddd}
\title{Defined dimensionality (defined.D) properties}
\usage{
ddd(x)

ddd_funs()

ddd_props()

is_ddd_spec(spec)

DDD(x, spec, ...)

D0D(x, ...)

D1D(x, ...)

D2D(x, ...)

DHD(x, ...)

nddd(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more defined.D properties from \code{ddd_props()}. Defined.D properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \verb{ddd_props, ddd_funs, ddd}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{is_ddd_spec, DDD, \{DDD\}}
\cr\cr  \strong{A numeric scalar}   \cr\cr \code{nddd}
}
\description{
An object's defined dimensionality (defined.D) is the number of dimensions on which its elements can be indexed. The associated property values are:
\tabular{lll}{  \verb{'d0d', 'D0D'}   \tab \verb{0D}   \tab The \code{NULL} object.                          \cr
  \verb{'d1d', 'D1D'}   \tab \verb{1D}   \tab Vectors, \link[=VLS]{vlists}, \verb{1D} arrays.  \cr
  \verb{'d2d', 'D2D'}   \tab \verb{2D}   \tab Data.frames and matrices.                   \cr
  \verb{'dhd', 'DHD'}   \tab \code{HD}   \tab Hyper (3+) dimensional arrays                 }
\cr \strong{Defined dimensionality functions}
\tabular{ll}{  \code{is_ddd_spec}   \tab Is \code{spec} a defined.D property specification?                                                                          \cr   \tab   \cr
  \code{ddd_props}     \tab What defined.D properties are there?                                                                                   \cr   \tab   \cr
  \code{ddd_funs}      \tab What defined.D property functions are there?                                                                           \cr   \tab   \cr
  \code{{DDD}}         \tab Does \code{x} match single defined.D property \code{'{DDD}'}? (where \code{{DDD}} is a placeholder for any given defined.D property). \cr   \tab   \cr
  \code{nddd}          \tab How many defined dimensions does \code{x} have?                                                                             \cr   \tab   \cr
  \code{ddd}           \tab What are \code{x}'s defined.D properties?                                                                                   \cr   \tab   \cr
  \code{DDD}           \tab Does \code{x} match defined.D specification \code{spec}?                                                                                        }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
ddd_funs()
ddd_props()
is_ddd_spec("d1D|d2D")
nddd(matrix(1))
nddd(letters)
nddd(1)
DDD(data.frame(letters), "d2d|dhd")
D0D(NULL)
D1D(NULL)
ddd(letters)
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
