% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim.R
\name{delim}
\alias{delim}
\alias{da}
\alias{dw}
\alias{daw}
\alias{dwa}
\alias{da0}
\alias{da1}
\alias{daB}
\alias{daC}
\alias{daD}
\alias{daG}
\alias{daP}
\alias{daQ}
\alias{daS}
\alias{daT}
\alias{dw0}
\alias{dw1}
\alias{dwB}
\alias{dwC}
\alias{dwD}
\alias{dwG}
\alias{dwP}
\alias{dwQ}
\alias{dwS}
\alias{dwT}
\alias{daw00}
\alias{daw01}
\alias{daw0B}
\alias{daw0C}
\alias{daw0D}
\alias{daw0G}
\alias{daw0P}
\alias{daw0Q}
\alias{daw0S}
\alias{daw0T}
\alias{daw10}
\alias{daw11}
\alias{daw1B}
\alias{daw1C}
\alias{daw1D}
\alias{daw1G}
\alias{daw1P}
\alias{daw1Q}
\alias{daw1S}
\alias{daw1T}
\alias{dawB0}
\alias{dawB1}
\alias{dawBB}
\alias{dawBC}
\alias{dawBD}
\alias{dawBG}
\alias{dawBP}
\alias{dawBQ}
\alias{dawBS}
\alias{dawBT}
\alias{dawC0}
\alias{dawC1}
\alias{dawCB}
\alias{dawCC}
\alias{dawCD}
\alias{dawCG}
\alias{dawCP}
\alias{dawCQ}
\alias{dawCS}
\alias{dawCT}
\alias{dawD0}
\alias{dawD1}
\alias{dawDB}
\alias{dawDC}
\alias{dawDD}
\alias{dawDG}
\alias{dawDP}
\alias{dawDQ}
\alias{dawDS}
\alias{dawDT}
\alias{dawG0}
\alias{dawG1}
\alias{dawGB}
\alias{dawGC}
\alias{dawGD}
\alias{dawGG}
\alias{dawGP}
\alias{dawGQ}
\alias{dawGS}
\alias{dawGT}
\alias{dawP0}
\alias{dawP1}
\alias{dawPB}
\alias{dawPC}
\alias{dawPD}
\alias{dawPG}
\alias{dawPP}
\alias{dawPQ}
\alias{dawPS}
\alias{dawPT}
\alias{dawQ0}
\alias{dawQ1}
\alias{dawQB}
\alias{dawQC}
\alias{dawQD}
\alias{dawQG}
\alias{dawQP}
\alias{dawQQ}
\alias{dawQS}
\alias{dawQT}
\alias{dawS0}
\alias{dawS1}
\alias{dawSB}
\alias{dawSC}
\alias{dawSD}
\alias{dawSG}
\alias{dawSP}
\alias{dawSQ}
\alias{dawSS}
\alias{dawST}
\alias{dawT0}
\alias{dawT1}
\alias{dawTB}
\alias{dawTC}
\alias{dawTD}
\alias{dawTG}
\alias{dawTP}
\alias{dawTQ}
\alias{dawTS}
\alias{dawTT}
\alias{dwa00}
\alias{dwa01}
\alias{dwa0B}
\alias{dwa0C}
\alias{dwa0D}
\alias{dwa0G}
\alias{dwa0P}
\alias{dwa0Q}
\alias{dwa0S}
\alias{dwa0T}
\alias{dwa10}
\alias{dwa11}
\alias{dwa1B}
\alias{dwa1C}
\alias{dwa1D}
\alias{dwa1G}
\alias{dwa1P}
\alias{dwa1Q}
\alias{dwa1S}
\alias{dwa1T}
\alias{dwaB0}
\alias{dwaB1}
\alias{dwaBB}
\alias{dwaBC}
\alias{dwaBD}
\alias{dwaBG}
\alias{dwaBP}
\alias{dwaBQ}
\alias{dwaBS}
\alias{dwaBT}
\alias{dwaC0}
\alias{dwaC1}
\alias{dwaCB}
\alias{dwaCC}
\alias{dwaCD}
\alias{dwaCG}
\alias{dwaCP}
\alias{dwaCQ}
\alias{dwaCS}
\alias{dwaCT}
\alias{dwaD0}
\alias{dwaD1}
\alias{dwaDB}
\alias{dwaDC}
\alias{dwaDD}
\alias{dwaDG}
\alias{dwaDP}
\alias{dwaDQ}
\alias{dwaDS}
\alias{dwaDT}
\alias{dwaG0}
\alias{dwaG1}
\alias{dwaGB}
\alias{dwaGC}
\alias{dwaGD}
\alias{dwaGG}
\alias{dwaGP}
\alias{dwaGQ}
\alias{dwaGS}
\alias{dwaGT}
\alias{dwaP0}
\alias{dwaP1}
\alias{dwaPB}
\alias{dwaPC}
\alias{dwaPD}
\alias{dwaPG}
\alias{dwaPP}
\alias{dwaPQ}
\alias{dwaPS}
\alias{dwaPT}
\alias{dwaQ0}
\alias{dwaQ1}
\alias{dwaQB}
\alias{dwaQC}
\alias{dwaQD}
\alias{dwaQG}
\alias{dwaQP}
\alias{dwaQQ}
\alias{dwaQS}
\alias{dwaQT}
\alias{dwaS0}
\alias{dwaS1}
\alias{dwaSB}
\alias{dwaSC}
\alias{dwaSD}
\alias{dwaSG}
\alias{dwaSP}
\alias{dwaSQ}
\alias{dwaSS}
\alias{dwaST}
\alias{dwaT0}
\alias{dwaT1}
\alias{dwaTB}
\alias{dwaTC}
\alias{dwaTD}
\alias{dwaTG}
\alias{dwaTP}
\alias{dwaTQ}
\alias{dwaTS}
\alias{dwaTT}
\title{Delimit strings}
\usage{
da(a., ...)

dw(w., ...)

daw(a., w., ...)

dwa(w., a., ...)

da0(...)

da1(...)

daB(...)

daC(...)

daD(...)

daG(...)

daP(...)

daQ(...)

daS(...)

daT(...)

dw0(...)

dw1(...)

dwB(...)

dwC(...)

dwD(...)

dwG(...)

dwP(...)

dwQ(...)

dwS(...)

dwT(...)

daw00(...)

daw01(...)

daw0B(...)

daw0C(...)

daw0D(...)

daw0G(...)

daw0P(...)

daw0Q(...)

daw0S(...)

daw0T(...)

daw10(...)

daw11(...)

daw1B(...)

daw1C(...)

daw1D(...)

daw1G(...)

daw1P(...)

daw1Q(...)

daw1S(...)

daw1T(...)

dawB0(...)

dawB1(...)

dawBB(...)

dawBC(...)

dawBD(...)

dawBG(...)

dawBP(...)

dawBQ(...)

dawBS(...)

dawBT(...)

dawC0(...)

dawC1(...)

dawCB(...)

dawCC(...)

dawCD(...)

dawCG(...)

dawCP(...)

dawCQ(...)

dawCS(...)

dawCT(...)

dawD0(...)

dawD1(...)

dawDB(...)

dawDC(...)

dawDD(...)

dawDG(...)

dawDP(...)

dawDQ(...)

dawDS(...)

dawDT(...)

dawG0(...)

dawG1(...)

dawGB(...)

dawGC(...)

dawGD(...)

dawGG(...)

dawGP(...)

dawGQ(...)

dawGS(...)

dawGT(...)

dawP0(...)

dawP1(...)

dawPB(...)

dawPC(...)

dawPD(...)

dawPG(...)

dawPP(...)

dawPQ(...)

dawPS(...)

dawPT(...)

dawQ0(...)

dawQ1(...)

dawQB(...)

dawQC(...)

dawQD(...)

dawQG(...)

dawQP(...)

dawQQ(...)

dawQS(...)

dawQT(...)

dawS0(...)

dawS1(...)

dawSB(...)

dawSC(...)

dawSD(...)

dawSG(...)

dawSP(...)

dawSQ(...)

dawSS(...)

dawST(...)

dawT0(...)

dawT1(...)

dawTB(...)

dawTC(...)

dawTD(...)

dawTG(...)

dawTP(...)

dawTQ(...)

dawTS(...)

dawTT(...)

dwa00(...)

dwa01(...)

dwa0B(...)

dwa0C(...)

dwa0D(...)

dwa0G(...)

dwa0P(...)

dwa0Q(...)

dwa0S(...)

dwa0T(...)

dwa10(...)

dwa11(...)

dwa1B(...)

dwa1C(...)

dwa1D(...)

dwa1G(...)

dwa1P(...)

dwa1Q(...)

dwa1S(...)

dwa1T(...)

dwaB0(...)

dwaB1(...)

dwaBB(...)

dwaBC(...)

dwaBD(...)

dwaBG(...)

dwaBP(...)

dwaBQ(...)

dwaBS(...)

dwaBT(...)

dwaC0(...)

dwaC1(...)

dwaCB(...)

dwaCC(...)

dwaCD(...)

dwaCG(...)

dwaCP(...)

dwaCQ(...)

dwaCS(...)

dwaCT(...)

dwaD0(...)

dwaD1(...)

dwaDB(...)

dwaDC(...)

dwaDD(...)

dwaDG(...)

dwaDP(...)

dwaDQ(...)

dwaDS(...)

dwaDT(...)

dwaG0(...)

dwaG1(...)

dwaGB(...)

dwaGC(...)

dwaGD(...)

dwaGG(...)

dwaGP(...)

dwaGQ(...)

dwaGS(...)

dwaGT(...)

dwaP0(...)

dwaP1(...)

dwaPB(...)

dwaPC(...)

dwaPD(...)

dwaPG(...)

dwaPP(...)

dwaPQ(...)

dwaPS(...)

dwaPT(...)

dwaQ0(...)

dwaQ1(...)

dwaQB(...)

dwaQC(...)

dwaQD(...)

dwaQG(...)

dwaQP(...)

dwaQQ(...)

dwaQS(...)

dwaQT(...)

dwaS0(...)

dwaS1(...)

dwaSB(...)

dwaSC(...)

dwaSD(...)

dwaSG(...)

dwaSP(...)

dwaSQ(...)

dwaSS(...)

dwaST(...)

dwaT0(...)

dwaT1(...)

dwaTB(...)

dwaTC(...)

dwaTD(...)

dwaTG(...)

dwaTP(...)

dwaTQ(...)

dwaTS(...)

dwaTT(...)
}
\arguments{
\item{a., w.}{Character scalar across delimiter vs. within delimiter,
respectively. Defaults to a blank string (i.e., \code{""}).}

\item{...}{An arbitrary number of atomic vector arguments to be delimited.
Argument in \code{...} must be recyclable for \code{da}, \code{daw},
\code{daA}, and \code{dawAW}.}
}
\value{
a character-scalar, character vector, or list of character vectors.
}
\description{
Simplifies and extends \code{paste} and \code{paste0}, allowing
for 'delimiting across' and 'delimiting within' arguments in \code{...},
and a combination of delimiting within and across arguments, in any order.
Also offers convenience functions for specific delimiters.
\cr\cr
\strong{\code{da(a., ...)}}
\cr Delimits across corresponding elements of arguments in \code{...}.
\code{da(delim, ...)} is identical to \code{paste(..., sep = delim)}.
\cr\cr
\strong{\code{dw(w., ...)}}
\cr Delimits within arguments in \code{...}. \code{dw(delim, ...)} is
equivalent to calling \code{sapply(list(...), paste0, collapse = delim)}.
\cr\cr
\strong{\code{daw(a., w., ...)}}
\cr Delimits across then within. \code{daw(delim.a, delimn.w, ...)} is
identical to \code{paste(..., sep = delim.a, collapse = delim.w)}
\cr\cr
\strong{\code{dwa(w., a., ...)}}
\cr Delimits within then across. \code{dwa(delim.w, delim.a, ...)} is
equivalent to \code{paste(sapply(list(...), paste0, collapse = delim.w),
  sep = delim.a)}.
\cr\cr
\strong{Extensions}
\cr Functions are extended for specific delimiters, signified by the
following extension characters:\tabular{lll}{
EXTENSION \tab EXTENSION         \tab DELIMITER  \cr
CHARACTER \tab NAME              \tab INVOKED    \cr
\code{'0'}\tab blank             \tab\code{""}   \cr
\code{'1'}\tab space             \tab\code{" "}  \cr
\code{'B'}\tab broken pipe       \tab\code{"¦"}  \cr
\code{'C'}\tab colon             \tab\code{":"}  \cr
\code{'D'}\tab dot               \tab\code{"."}  \cr
\code{'G'}\tab grammatical comma \tab\code{", "} \cr
\code{'P'}\tab pipe              \tab\code{"|"}  \cr
\code{'Q'}\tab (back)-quote      \tab\code{"`"}  \cr
\code{'S'}\tab simple comma      \tab\code{","}  \cr
\code{'T'}\tab tilde             \tab\code{"~"}    }
\strong{\code{daA}}
\cr Delimits across using the extension character A.
\cr\cr
\strong{\code{dwW}}
\cr Delimits within using the extension character W.
\cr\cr
\strong{\code{dawAW}}
\cr Delimits across using the extension character A then delimits within
using the extension character W.
\cr\cr
\strong{\code{dwaWA}}
\cr Delimits within using the extension character W then delimits across
using the extension character A.
}
\section{Functions}{
\itemize{
\item \code{da}: Delimit across vectors in \code{...} using \code{a.}
resulting in a character vector of \code{max(lengths(list(...)))}
\code{a.}-delimited strings.

\item \code{dw}: Delimit within vectors in \code{...} using \code{w.},
resulting in a character vector of \code{\link[base]{...length()}}
\code{w.}-separated strings.

\item \code{daw}: Delimit across vectors in \code{...} using \code{a.},
then delimit within using \code{w.}, resulting in a character-scalar string
containing \code{max(lengths(list(...)))} \code{a.}-delimited substrings,
each of which contains \code{\link[base]{...length()}} \code{w.}-delimited
subsubstrings.

\item \code{dwa}: Delimit within vectors in \code{...} using \code{w.}, then
delimit across using \code{a.}, resulting in a character-scalar string
containing \code{\link[base]{...length()}} \code{w.}-delimited substrings,
each of which contains \code{max(length(list(...)))} \code{a.}-delimited
subsubstrings.

\item \code{da0}: Delimit across vectors in \code{...} using blanks.

\item \code{da1}: Delimit across vectors in \code{...} using spaces.

\item \code{daB}: Delimit across vectors in \code{...} using broken-pipes.

\item \code{daC}: Delimit across vectors in \code{...} using colons.

\item \code{daD}: Delimit across vectors in \code{...} using dots/periods.

\item \code{daG}: Delimit across vectors in \code{...} using grammatical
commas.

\item \code{daP}: Delimit across vectors in \code{...} using pipes.

\item \code{daQ}: Delimit across vectors in \code{...} using (back)-quotes.

\item \code{daS}: Delimit across vectors in \code{...} using simple commas.

\item \code{daT}: Delimit across vectors in \code{...} using tildes.

\item \code{dw0}: Delimit within vectors in \code{...} using blanks.

\item \code{dw1}: Delimit within vectors in \code{...} using spaces.

\item \code{dwB}: Delimit within vectors in \code{...} using broken-pipes.

\item \code{dwC}: Delimit within vectors in \code{...} using colons.

\item \code{dwD}: Delimit within vectors in \code{...} using dots/periods.

\item \code{dwG}: Delimit within vectors in \code{...} using grammatical
commas.

\item \code{dwP}: Delimit within vectors in \code{...} using pipes.

\item \code{dwQ}: Delimit within vectors in \code{...} using (back)-quotes.

\item \code{dwS}: Delimit within vectors in \code{...} using simple commas.

\item \code{dwT}: Delimit within vectors in \code{...} using tildes.

\item \code{daw00}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks}.

\item \code{daw01}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{spaces}, respectively.

\item \code{daw0B}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{broken-pipes}, respectively.

\item \code{daw0C}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{colons}, respectively.

\item \code{daw0D}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{dots/periods}, respectively.

\item \code{daw0G}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{grammatical commas},
respectively.

\item \code{daw0P}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{pipes}, respectively.

\item \code{daw0Q}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{(back)-quotes}, respectively.

\item \code{daw0S}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{simple commas}, respectively.

\item \code{daw0T}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{blanks} then \emph{tildes}, respectively.

\item \code{daw10}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{blanks}, respectively.

\item \code{daw11}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces}.

\item \code{daw1B}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{broken-pipes}, respectively.

\item \code{daw1C}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{colons}, respectively.

\item \code{daw1D}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{dots/periods}, respectively.

\item \code{daw1G}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{grammatical commas},
respectively.

\item \code{daw1P}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{pipes}, respectively.

\item \code{daw1Q}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{(back)-quotes}, respectively.

\item \code{daw1S}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{simple commas}, respectively.

\item \code{daw1T}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{spaces} then \emph{tildes}, respectively.

\item \code{dawB0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{blanks}, respectively.

\item \code{dawB1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{spaces}, respectively.

\item \code{dawBB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes}.

\item \code{dawBC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{colons}, respectively.

\item \code{dawBD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{dots/periods},
respectively.

\item \code{dawBG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{grammatical commas},
respectively.

\item \code{dawBP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{pipes}, respectively.

\item \code{dawBQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{(back)-quotes},
respectively.

\item \code{dawBS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{simple commas},
respectively.

\item \code{dawBT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{broken-pipes} then \emph{tildes}, respectively.

\item \code{dawC0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{blanks}, respectively.

\item \code{dawC1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{spaces}, respectively.

\item \code{dawCB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{broken-pipes}, respectively.

\item \code{dawCC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons}.

\item \code{dawCD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{dots/periods}, respectively.

\item \code{dawCG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{grammatical commas},
respectively.

\item \code{dawCP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{pipes}, respectively.

\item \code{dawCQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{(back)-quotes}, respectively.

\item \code{dawCS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{simple commas}, respectively.

\item \code{dawCT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{colons} then \emph{tildes}, respectively.

\item \code{dawD0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{blanks}, respectively.

\item \code{dawD1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{spaces}, respectively.

\item \code{dawDB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{broken-pipes},
respectively.

\item \code{dawDC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{colons}, respectively.

\item \code{dawDD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods}.

\item \code{dawDG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{grammatical commas},
respectively.

\item \code{dawDP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{pipes}, respectively.

\item \code{dawDQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{(back)-quotes},
respectively.

\item \code{dawDS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{simple commas},
respectively.

\item \code{dawDT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{dots/periods} then \emph{tildes}, respectively.

\item \code{dawG0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{blanks},
respectively.

\item \code{dawG1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{spaces},
respectively.

\item \code{dawGB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{broken-pipes},
respectively.

\item \code{dawGC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{colons},
respectively.

\item \code{dawGD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{dots/periods},
respectively.

\item \code{dawGG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas}.

\item \code{dawGP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{pipes}, respectively.

\item \code{dawGQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{(back)-quotes},
respectively.

\item \code{dawGS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{simple commas},
respectively.

\item \code{dawGT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{grammatical commas} then \emph{tildes},
respectively.

\item \code{dawP0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{blanks}, respectively.

\item \code{dawP1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{spaces}, respectively.

\item \code{dawPB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{broken-pipes}, respectively.

\item \code{dawPC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{colons}, respectively.

\item \code{dawPD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{dots/periods}, respectively.

\item \code{dawPG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{grammatical commas}, respectively.

\item \code{dawPP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes}.

\item \code{dawPQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{(back)-quotes}, respectively.

\item \code{dawPS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{simple commas}, respectively.

\item \code{dawPT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{pipes} then \emph{tildes}, respectively.

\item \code{dawQ0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{blanks}, respectively.

\item \code{dawQ1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{spaces}, respectively.

\item \code{dawQB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{broken-pipes},
respectively.

\item \code{dawQC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{colons}, respectively.

\item \code{dawQD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{dots/periods},
respectively.

\item \code{dawQG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{grammatical commas},
respectively.

\item \code{dawQP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{pipes}, respectively.

\item \code{dawQQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes}.

\item \code{dawQS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{simple commas},
respectively.

\item \code{dawQT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{(back)-quotes} then \emph{tildes}, respectively.

\item \code{dawS0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{blanks}, respectively.

\item \code{dawS1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{spaces}, respectively.

\item \code{dawSB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{broken-pipes},
respectively.

\item \code{dawSC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{colons}, respectively.

\item \code{dawSD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{dots/periods},
respectively.

\item \code{dawSG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{grammatical commas},
respectively.

\item \code{dawSP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{pipes}, respectively.

\item \code{dawSQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{(back)-quotes},
respectively.

\item \code{dawSS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas}.

\item \code{dawST}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{simple commas} then \emph{tildes}, respectively.

\item \code{dawT0}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{blanks}, respectively.

\item \code{dawT1}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{spaces}, respectively.

\item \code{dawTB}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{broken-pipes}, respectively.

\item \code{dawTC}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{colons}, respectively.

\item \code{dawTD}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{dots/periods}, respectively.

\item \code{dawTG}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{grammatical commas},
respectively.

\item \code{dawTP}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{pipes}, respectively.

\item \code{dawTQ}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{(back)-quotes}, respectively.

\item \code{dawTS}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes} then \emph{simple commas}, respectively.

\item \code{dawTT}: Delimit \emph{across} then \emph{within} vectors in
\code{...} using \emph{tildes}.

\item \code{dwa00}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks}.

\item \code{dwa01}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{spaces}, respectively.

\item \code{dwa0B}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{broken-pipes}, respectively.

\item \code{dwa0C}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{colons}, respectively.

\item \code{dwa0D}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{dots/periods}, respectively.

\item \code{dwa0G}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{grammatical commas},
respectively.

\item \code{dwa0P}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{pipes}, respectively.

\item \code{dwa0Q}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{(back)-quotes}, respectively.

\item \code{dwa0S}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{simple commas}, respectively.

\item \code{dwa0T}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{blanks} then \emph{tildes}, respectively.

\item \code{dwa10}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{blanks}, respectively.

\item \code{dwa11}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces}.

\item \code{dwa1B}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{broken-pipes}, respectively.

\item \code{dwa1C}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{colons}, respectively.

\item \code{dwa1D}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{dots/periods}, respectively.

\item \code{dwa1G}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{grammatical commas},
respectively.

\item \code{dwa1P}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{pipes}, respectively.

\item \code{dwa1Q}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{(back)-quotes}, respectively.

\item \code{dwa1S}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{simple commas}, respectively.

\item \code{dwa1T}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{spaces} then \emph{tildes}, respectively.

\item \code{dwaB0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{blanks}, respectively.

\item \code{dwaB1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{spaces}, respectively.

\item \code{dwaBB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes}.

\item \code{dwaBC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{colons}, respectively.

\item \code{dwaBD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{dots/periods},
respectively.

\item \code{dwaBG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{grammatical commas},
respectively.

\item \code{dwaBP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{pipes}, respectively.

\item \code{dwaBQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{(back)-quotes},
respectively.

\item \code{dwaBS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{simple commas},
respectively.

\item \code{dwaBT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{broken-pipes} then \emph{tildes}, respectively.

\item \code{dwaC0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{blanks}, respectively.

\item \code{dwaC1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{spaces}, respectively.

\item \code{dwaCB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{broken-pipes}, respectively.

\item \code{dwaCC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons}.

\item \code{dwaCD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{dots/periods}, respectively.

\item \code{dwaCG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{grammatical commas},
respectively.

\item \code{dwaCP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{pipes}, respectively.

\item \code{dwaCQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{(back)-quotes}, respectively.

\item \code{dwaCS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{simple commas}, respectively.

\item \code{dwaCT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{colons} then \emph{tildes}, respectively.

\item \code{dwaD0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{blanks}, respectively.

\item \code{dwaD1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{spaces}, respectively.

\item \code{dwaDB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{broken-pipes},
respectively.

\item \code{dwaDC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{colons}, respectively.

\item \code{dwaDD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods}.

\item \code{dwaDG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{grammatical commas},
respectively.

\item \code{dwaDP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{pipes}, respectively.

\item \code{dwaDQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{(back)-quotes},
respectively.

\item \code{dwaDS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{simple commas},
respectively.

\item \code{dwaDT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{dots/periods} then \emph{tildes}, respectively.

\item \code{dwaG0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{blanks},
respectively.

\item \code{dwaG1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{spaces},
respectively.

\item \code{dwaGB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{broken-pipes},
respectively.

\item \code{dwaGC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{colons},
respectively.

\item \code{dwaGD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{dots/periods},
respectively.

\item \code{dwaGG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas}.

\item \code{dwaGP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{pipes}, respectively.

\item \code{dwaGQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{(back)-quotes},
respectively.

\item \code{dwaGS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{simple commas},
respectively.

\item \code{dwaGT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{grammatical commas} then \emph{tildes},
respectively.

\item \code{dwaP0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{blanks}, respectively.

\item \code{dwaP1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{spaces}, respectively.

\item \code{dwaPB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{broken-pipes}, respectively.

\item \code{dwaPC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{colons}, respectively.

\item \code{dwaPD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{dots/periods}, respectively.

\item \code{dwaPG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{grammatical commas}, respectively.

\item \code{dwaPP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes}.

\item \code{dwaPQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{(back)-quotes}, respectively.

\item \code{dwaPS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{simple commas}, respectively.

\item \code{dwaPT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{pipes} then \emph{tildes}, respectively.

\item \code{dwaQ0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{blanks}, respectively.

\item \code{dwaQ1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{spaces}, respectively.

\item \code{dwaQB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{broken-pipes},
respectively.

\item \code{dwaQC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{colons}, respectively.

\item \code{dwaQD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{dots/periods},
respectively.

\item \code{dwaQG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{grammatical commas},
respectively.

\item \code{dwaQP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{pipes}, respectively.

\item \code{dwaQQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes}.

\item \code{dwaQS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{simple commas},
respectively.

\item \code{dwaQT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{(back)-quotes} then \emph{tildes}, respectively.

\item \code{dwaS0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{blanks}, respectively.

\item \code{dwaS1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{spaces}, respectively.

\item \code{dwaSB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{broken-pipes},
respectively.

\item \code{dwaSC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{colons}, respectively.

\item \code{dwaSD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{dots/periods},
respectively.

\item \code{dwaSG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{grammatical commas},
respectively.

\item \code{dwaSP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{pipes}, respectively.

\item \code{dwaSQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{(back)-quotes},
respectively.

\item \code{dwaSS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas}.

\item \code{dwaST}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{simple commas} then \emph{tildes}, respectively.

\item \code{dwaT0}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{blanks}, respectively.

\item \code{dwaT1}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{spaces}, respectively.

\item \code{dwaTB}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{broken-pipes}, respectively.

\item \code{dwaTC}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{colons}, respectively.

\item \code{dwaTD}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{dots/periods}, respectively.

\item \code{dwaTG}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{grammatical commas},
respectively.

\item \code{dwaTP}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{pipes}, respectively.

\item \code{dwaTQ}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{(back)-quotes}, respectively.

\item \code{dwaTS}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes} then \emph{simple commas}, respectively.

\item \code{dwaTT}: Delimit \emph{within} then \emph{across} vectors in
\code{...} using \emph{tildes}.
}}

\seealso{
Other strings: 
\code{\link{ipat}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
