% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\encoding{UTF-8}
\name{dialog}
\alias{dialog}
\alias{alert}
\alias{acknowledge}
\alias{choose1}
\alias{chooseN}
\alias{NO}
\alias{YES}
\alias{OK}
\alias{CANCEL}
\alias{ask}
\alias{asknew}
\alias{choose_dir}
\alias{choose_doc}
\title{Dialog with users using package \code{svDialogs}}
\usage{
alert(
  ...,
  type = "alert",
  title = NULL,
  st = NULL,
  fg = NULL,
  bg = NULL,
  ST = NULL,
  FG = NULL,
  BG = NULL,
  stop = NULL,
  blank = ""
)

acknowledge(
  ...,
  st = NULL,
  bg = NULL,
  fg = NULL,
  ST = NULL,
  BG = NULL,
  FG = NULL
)

choose1(
  options,
  ...,
  st = NULL,
  bg = NULL,
  fg = NULL,
  ST = NULL,
  BG = NULL,
  FG = NULL
)

chooseN(
  options,
  ...,
  all = TRUE,
  none = FALSE,
  n = NULL,
  min.n = NULL,
  max.n = NULL,
  st = NULL,
  bg = NULL,
  fg = NULL,
  ST = NULL,
  BG = NULL,
  FG = NULL
)

NO(..., st = NULL, bg = NULL, fg = NULL, ST = NULL, BG = NULL, FG = NULL)

YES(..., st = NULL, bg = NULL, fg = NULL, ST = NULL, BG = NULL, FG = NULL)

OK(..., st = NULL, bg = NULL, fg = NULL, ST = NULL, BG = NULL, FG = NULL)

CANCEL(..., st = NULL, bg = NULL, fg = NULL, ST = NULL, BG = NULL, FG = NULL)

ask(..., st = NULL, bg = NULL, fg = NULL, ST = NULL, BG = NULL, FG = NULL)

asknew(
  old,
  unq = TRUE,
  st = NULL,
  bg = NULL,
  fg = NULL,
  ST = NULL,
  BG = NULL,
  FG = NULL
)

choose_dir(
  dir.type = "directory",
  st = NULL,
  bg = NULL,
  fg = NULL,
  ST = NULL,
  BG = NULL,
  FG = NULL
)

choose_doc(
  doc.type = "document",
  st = NULL,
  bg = NULL,
  fg = NULL,
  ST = NULL,
  BG = NULL,
  FG = NULL
)
}
\arguments{
\item{...}{An arbitrary, optional number of arguments which are \link[=av]{atomized} into a character scalar message to be posted to the console.}

\item{type}{A character scalar type of alert to post to the console. Predefined alert types are \code{c('alert', 'error', 'update', 'warning')}. Custom types are also accepted, but they should be short to avoid formatting problems.}

\item{title}{\code{NULL} or a character scalar alert title to post to the console. Should be relatively short to avoid formatting problems.}

\item{st}{\code{NULL} or a character scalar alert style specification from \code{link{st_vals}()}.}

\item{fg}{\code{NULL} or a character scalar alert foreground text color specification from \code{\link{fg_vals}()}}

\item{bg}{\code{NULL} or an character scalar alert background text color specification from \code{\link{bg_vals}()}}

\item{ST}{\code{NULL} or a character scalar message style specification from \code{link{st_vals}()}.}

\item{FG}{\code{NULL} or a character scalar message foreground text color specification from \code{\link{fg_vals}()}}

\item{BG}{\code{NULL} or an character scalar message background text color specification from \code{\link{bg_vals}()}}

\item{stop}{An optional scalar \code{TRUE} or \code{FALSE} indicating whether to stop execution. Defaults to \code{TRUE} for \code{type = 'error'}, otherwise defaults to \code{FALSE}.}

\item{blank}{A character scalar containing a default message if \link[=av]{atomizing} and collapsing \code{...} to a character scalar results in a blank string (\code{""}).}

\item{options}{An atomic vector list of options to choose from.}

\item{all}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add an \code{{ ALL }} value to \code{options}.}

\item{none}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add a \code{{ NONE }} value to \code{options} (implying that it is valid to select none).}

\item{n}{An optional \link[=cmp_psw_scl]{complete positive numeric whole-number scalar} (?cmp_psw_scl) indicating the number of options that must be selected. Must be contained in \code{1:length(options)}.}

\item{min.n}{An optional \link[=icmp]{complete} \link[=ipsw]{positive numeric whole-number} \link[=iscl]{scalar} (?icmp, ?ipsw, ?iscl) indicating the minimum number of options that may be selected. Must be \code{NULL} when \code{n} is non-\code{NULL}.}

\item{max.n}{An optional complete positive numeric whole-number scalar indicating the maximum number of options that may be selected. Must be \code{NULL} when \code{n} is non-\code{NULL}.}

\item{old}{A \link[=chr_vec]{character vec} of unique values to be replaced.}

\item{unq}{Scalar \code{TRUE} or \code{FALSE} indicating whether replacement values must be unique.}
}
\value{
\emph{A character scalar} \cr   \code{choose_dir} (a directory path) \cr   \code{choose_doc} (a document path) \cr   \code{ask}
\cr\cr \emph{A character vector} \cr   \code{asknew}
\cr\cr \emph{An atomic vector} \cr   \code{chooseN}
\cr\cr \emph{An atomic scalar} \cr   \code{choose1}
\cr\cr \emph{A logical scalar} \cr   \verb{CANCEL, YES, NO, OK}
\cr\cr \code{NULL} \cr   \verb{acknowledge, alert}
}
\description{
All functions \link[=collapse_dots]{collapses} \code{...} args into a prompt. Each posts an alert to the console, posts the prompt (if any), followed by the actions detailed in the table below.
\tabular{rl}{
  \code{acknowledge}   \tab Waits for user to acknowledge.\eqn{^1}
\cr \code{choose_doc}   \tab Asks user to choose a document.\eqn{^2}
\cr \code{choose_dir}   \tab Asks user to choose a directory.\eqn{^2}
\cr    \code{choose1}   \tab Asks user to choose \code{1} option.\eqn{^2}
\cr    \code{chooseN}   \tab Asks user to choose \verb{1+} options.\eqn{^{2,3}}
\cr      \code{alert}   \tab None.\eqn{^1}
\cr        \code{ask}   \tab Asks for typed input.\eqn{^2}
\cr     \code{asknew}   \tab Asks for a list of new values.\eqn{^2}
\cr    \verb{YES, NO}   \tab Offers \code{YES/NO/CANCEL}.\eqn{^{2,4}}
\cr \verb{OK, CANCEL}   \tab Offers \code{OK/CANCEL}.\eqn{^4}
}
\verb{ }\eqn{^{1.}} Default style, bg color, and fg color depend on \code{type}.
\cr \verb{ }\eqn{^{2.}} Stops execution if user chooses \code{CANCEL}.
\cr \verb{ }\eqn{^{3.}} Stops execution if \code{stop | (type = 'error' & is.null(stop))}.
\cr \verb{ }\eqn{^{4.}} Returns \code{TRUE} if user choice matches the function name.
}
\seealso{
Other user: 
\code{\link{say}()}
}
\concept{dialogs}
\concept{user}
