% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\encoding{UTF-8}
\name{dialog}
\alias{dialog}
\alias{alert}
\alias{acknowledge}
\alias{choose1}
\alias{chooseN}
\alias{NO}
\alias{YES}
\alias{OK}
\alias{CANCEL}
\alias{ask}
\alias{asknew}
\alias{choose_dir}
\alias{choose_doc}
\title{Dialog boxes using package \code{svDialogs}}
\usage{
alert(
  ...,
  type = "alert",
  title = NULL,
  style = NULL,
  fg = NULL,
  bg = NULL,
  stop = NULL
)

acknowledge(..., style = NULL, bg = NULL, fg = NULL)

choose1(options, ...)

chooseN(
  options,
  ...,
  all = TRUE,
  none = FALSE,
  n = NULL,
  min.n = NULL,
  max.n = NULL
)

NO(...)

YES(...)

OK(...)

CANCEL(...)

ask(...)

asknew(old, unq = TRUE)

choose_dir(dir.type = "directory")

choose_doc(doc.type = "document")
}
\arguments{
\item{...}{An arbitrary, optional number of arguments which are \link[=av]{atomized} into a character scalar message to be posted to the console.}

\item{type}{A character scalar type of alert to post to the console. Predefined alert types are \code{c('alert', 'error', 'update', 'warning')}. Custom types are also accepted.}

\item{title}{An optional character scalar alert title to post to the console.}

\item{style}{An optional character scalar style specification. Valid values are \code{c('bold', 'italic', 'plain', 'underline')}. Defaults to \code{'plain'}.}

\item{fg}{An optional character scalar foreground text color. Valid values are \code{c('black', 'red', 'green', 'yellow', 'blue', 'magenta', 'cyan', 'white', 'grey')}. Defaults to system default.}

\item{bg}{An optional character scalar background text color. Valid values are \code{c('black', 'red', 'green', 'yellow', 'blue', 'magenta', 'cyan', 'white')}. Defaults to system default.}

\item{stop}{An optional scalar \code{TRUE} or \code{FALSE} indicating whether to stop execution. Defaults to \code{TRUE} for \code{type = 'error'}, otherwise defaults to \code{FALSE}.}

\item{options}{An atomic vector list of options to choose from.}

\item{all}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add an \code{{ ALL }} value to \code{options}.}

\item{none}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add a \code{{ NONE }} value to \code{options} (implying that it is valid to select none).}

\item{n}{An optional \link[=cmp_psw_scl]{complete positive numeric whole-number scalar} (?cmp_psw_scl) indicating the number of options that must be selected. Must be contained in \code{1:length(options)}.}

\item{min.n}{An optional \link[=icmp]{complete} \link[=ipsw]{positive numeric whole-number} \link[=iscl]{scalar} (?icmp, ?ipsw, ?iscl) indicating the minimum number of options that may be selected. Must be \code{NULL} when \code{n} is non-\code{NULL}.}

\item{max.n}{An optional complete positive numeric whole-number scalar indicating the maximum number of options that may be selected. Must be \code{NULL} when \code{n} is non-\code{NULL}.}

\item{old}{A \link[=chr_vec]{character vec} of unique values to be replaced.}

\item{unq}{Scalar \code{TRUE} or \code{FALSE} indicating whether replacement values must be unique.}
}
\value{
\emph{A character scalar}
\cr   \code{choose_dir} (a directory path)
\cr   \code{choose_doc} (a document path)
\cr   \code{ask}
\cr
\cr \emph{A character vector}
\cr   \code{asknew}
\cr
\cr \emph{An atomic vector}
\cr   \code{chooseN}
\cr
\cr \emph{An atomic scalar}
\cr   \code{choose1}
\cr
\cr \emph{A logical scalar}
\cr   \code{CANCEL}
\cr   \code{YES}
\cr   \code{NO}
\cr   \code{OK}
\cr
\cr \code{NULL}
\cr   \code{acknowledge}
\cr   \code{alert}
}
\description{
All functions in this family take \code{...} arguments, which are \link[=av]{atomized} and collapsed into a prompt to be displayed in a dialog box.
\tabular{rl}{ \code{alert}   \tab Posts an alert to the console. Default style, background color, and foreground color depend on the value of the argument \code{type}, which can be overridden using the arguments \code{style}, \code{fg}, and \code{bg}.
\cr                   \tab  
\cr   \code{acknowledge}   \tab Posts an alert, then posts a message requesting acknowledgement, After user acknowledges, launches file selection dialog box.
\cr                   \tab  
\cr    \code{choose_doc}   \tab Posts an acknowledgement alert. After user acknowledges, launches file selection dialog box.
\cr                   \tab  
\cr    \code{choose_dir}   \tab Posts an acknowledgement alert. After user acknowledges, launches directory selection dialog box.
\cr                   \tab  
\cr       \code{choose1}   \tab Posts an alert followed by a message asking user to select \code{1} option from a list. Stops execution if user select the always-included, always-first option \code{'{ CANCEL }'}.
\cr                   \tab  
\cr       \code{choosen}   \tab Posts an alert followed by a message asking user to select from among a list of options. Stops execution if user selects the always-included, always-first option \code{'{ CANCEL }'}.
\cr                   \tab  
\cr        \code{asknew}   \tab Posts an alert followed by a message asking user to enter a pipe-separated list of values to replace existing values.
\cr                   \tab  
\cr           \code{ask}   \tab Posts an alert followed by a message asking user for typed input.
\cr                   \tab  
\cr           \code{msg}   \tab Posts an alert followed by a message to hit enter to continue.
\cr                   \tab  
\cr           \code{YES}   \tab Posts an alert followed by a message offering choices  of \code{'{ CANCEL }'}, \code{'{ YES }'} and \code{'{ NO }'}. Stops execution if user selects \code{'{ CANCEL }'}, returns \code{TRUE} if user selects \code{'{ YES }'}, otherwise, returns \code{FALSE}.
\cr                   \tab  
\cr            \code{NO}   \tab Posts an alert followed by a message offering choices  of \code{'{ CANCEL }'}, \code{'{ YES }'} and \code{'{ NO }'}. Stops execution if user selects \code{'{ CANCEL }'}, returns \code{TRUE} if user selects \code{'{ NO }'}, otherwise, returns \code{FALSE}.
\cr                   \tab  
\cr            \code{OK}   \tab Posts an alert followed by a message offering choices of \code{'{ CANCEL }'} and \code{'{ OK }'}. Returns \code{TRUE} if user selects \code{{ OK }}, otherwise return \code{FALSE}.
\cr                   \tab  
\cr        \code{CANCEL}   \tab Posts an alert followed by a message offering choices of \code{'{ CANCEL }'} and \code{'{ OK }'}. Returns \code{TRUE} if user selects \code{{ CANCEL }}, otherwise return \code{FALSE}.
}
}
\seealso{
Other user: 
\code{\link{say}()}
}
\concept{dialogs}
\concept{user}
