% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{dialogs}
\alias{dialogs}
\alias{msgbox}
\alias{dirbox}
\alias{docbox}
\alias{ansbox}
\alias{no}
\alias{yes}
\alias{msg}
\alias{okx}
\alias{choose_dir}
\alias{choose_doc}
\alias{ask}
\alias{ask1}
\alias{askn}
\alias{asksn}
\alias{asks1}
\alias{asknew}
\title{Dialog boxes using package \code{svDialogs}}
\usage{
msgbox(..., t = "ok")

dirbox(..., d = getwd())

docbox(...)

ansbox(x = "Enter a value", def = "")

no(...)

yes(...)

msg(x, t = "ok")

okx(x)

choose_dir(type = "file")

choose_doc(type = "document")

ask(..., def = "", cancel = ".stop")

ask1(opts, what, more = FALSE)

askn(opts, what, all = TRUE, none = TRUE, more = FALSE)

asksn(opts, what, per = 9)

asks1(opts, what, per = 9)

asknew(old, what, unq = T)
}
\arguments{
\item{...}{Objects to be atomized to create the question.}

\item{t}{A complete character scalar type of dialog box (valid values are \code{'ok'}, \code{'okcancel'}, \code{'yesno'}, and \code{'yesnocancel'}).}

\item{d}{A complete character scalar default directory.}

\item{x}{A \link[=cmp_chr_scl]{complete character scalar} message.}

\item{def}{A complete character scalar default value.}

\item{type}{A complete character scalar type of directory or document to select.}

\item{cancel}{A complete character scalar value to return if user clicks the \code{cancel} button. The value \code{cancel = '.stop'} is an indicator that an error should be thrown if the user clicks the \code{cancel} button.}

\item{opts}{A \link[=cmp_chr_vec]{complete character vec} of options to choose from.}

\item{what}{A complete character scalar describing type of option being presented.}

\item{more}{A non-\code{NA} logical scalar indicating whether there are more options than could be presented in a single dialog box that still could be presented.}

\item{all, none}{Non-\code{NA} logical scalars indicating whether the user should be able to select all options or none of the options, respectively.}

\item{per}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the maximum number of options to present in a single dialog box interaction.}

\item{unq}{A non-\code{NA} logical scalar indicating whether new values must be unique.}
}
\value{
\tabular{rl}{
\cr \code{choose_dir}   \tab A character scalar.
\cr \code{choose_doc}   \tab   
\cr      \code{asks1}   \tab   
\cr       \code{ask1}   \tab   
\cr        \code{ask}   \tab   
\cr        \code{okx}   \tab   
\cr        \code{msg}   \tab   
\cr                \tab   
\cr     \code{ansbox}   \tab An environment.
\cr     \code{docbox}   \tab   
\cr     \code{dirbox}   \tab   
\cr     \code{msgbox}   \tab   
\cr                \tab   
\cr     \code{asknew}   \tab A character vector.
\cr      \code{asksn}   \tab   
\cr       \code{askn}   \tab   
\cr                \tab   
\cr        \code{yes}   \tab A logical scalar.
\cr         \code{no}   \tab   
}
}
\description{
All functions in this family take \code{...} arguments, which are \link[=av]{atomized} and collapsed into a prompt to be displayed in a dialog box.
\tabular{rl}{
  \code{choose_doc}   \tab asks user to select a document/file.
\cr                \tab   
\cr \code{choose_dir}   \tab asks user to select a directory/folder.
\cr                \tab   
\cr     \code{msgbox}   \tab launches a simple dialog box to give the message.
\cr                \tab   
\cr     \code{dirbox}   \tab asks user to select a directory/folder\eqn{^1}.
\cr                \tab   
\cr     \code{docbox}   \tab asks user to select a document\eqn{^1}.
\cr                \tab   
\cr     \code{ansbox}   \tab asks user to input a text response.
\cr                \tab   
\cr     \code{asknew}   \tab asks user to enter a space- or comma-separated list of replacement values for existing values.
\cr                \tab   
\cr      \code{asks1}   \tab asks user to select \code{1} option from a list, possibly across multiple rounds.
\cr                \tab   
\cr      \code{asksn}   \tab asks user to select a \verb{1+} options from a list of possible options, possibly across multiple rounds.
\cr                \tab   
\cr       \code{ask1}   \tab asks user to select a single option from a list.
\cr                \tab   
\cr       \code{askn}   \tab asks user to select \verb{1+} options from a list.
\cr                \tab   
\cr        \code{ask}   \tab asks user for typed input.
\cr                \tab   
\cr        \code{msg}   \tab returns the name of the button clicked.
\cr                \tab   
\cr        \code{okx}   \tab checks for user selecting the \code{ok} button of an \code{okcancel} dialog.
\cr                \tab   
\cr        \code{yes}   \tab checks for user selecting the \code{yes} button of a \code{yesno} dialog.
\cr                \tab   
\cr         \code{no}   \tab checks for user selecting the \code{no} button of a \code{yesno} dialog.
}
\eqn{^{1.}} Failsafe feature launches a message dialog box prompting the user to take an action in the next dialog box. After user acknowledges, only then launches the selection dialog to avoid problems with prompts not showing up on all operating systems.
}
