% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_uj.R
\name{dots_uj}
\alias{dots_uj}
\alias{dots}
\alias{dot}
\alias{dot_names}
\alias{named_dots}
\alias{unnamed_dots}
\title{Manage \code{...} arguments}
\usage{
dots(names., defs., ...)

dot(name., def., ...)

dot_names(..., subs. = NULL, req. = T, blank. = F, u. = T)

named_dots(...)

unnamed_dots(...)
}
\arguments{
\item{names.}{\code{NULL} or an \link[=atm_vec]{atomic vec} (may include \code{NA} values). Is split along the delimiter \code{'|'} to allow for compactness in submitting multiple names. \code{NULL} will match either an argument in \code{...} or element of \code{defs.} with the name \code{'NULL'}. \code{NA} values will match an argument in \code{...} or an element of \code{defs.} with the name \code{'NA'}.}

\item{defs.}{A named \link[=ivls]{vlist} of default objects/values to return if the specified arguments are not in \code{...}. Elements of \code{defs.} must be uniquely named. If \code{defs.} is a tibble, columns with matching names are returned.}

\item{...}{An arbitrary number of arguments.}

\item{name.}{\code{NULL} or \link[=cmp_scl]{complete atomic scalar}. \code{NULL} is replaced with \code{'NULL'} and \code{NA} is replaced with \code{'NA'}.}

\item{def.}{A default object/value to return if a specified argument is not in \code{...}. Can be, but does not need to be, a list.}

\item{subs.}{\code{NULL} or \link[=cmp_chr_vec]{complete character vec}. If not \code{NULL}, it is split using pipes \code{'|'} as a delimiter. If there are no pipes contained in \code{names.}, it remains unchanged. When this argument is not \code{NULL}, it is substituted for the names of \code{...} arguments; thus, after splitting, its length must equal the number of \code{...} arguments. For example, \code{names. = c('one', 'two', 'three|four|five')} indicates that there should be five \code{...} arguments and the vector \code{c('one', 'two', 'three', 'four', 'five')} is substituted for their names.}

\item{req., blank., u.}{Non-\code{NA} logical scalars indicating whether names are required, whether lank names are allowed, and whether names must be unique.}
}
\description{
\tabular{rl}{
  \code{unnamed_dots}   \tab Extracts unnamed \code{...} arguments as an unnamed list (includes any named with blank strings).
\cr                \tab   
\cr \code{named_dots}   \tab Extracts named \code{...} arguments as a named list (does not include with blank-string names).
\cr                \tab   
\cr       \code{dots}   \tab Extracts one or more\code{...} arguments based on matching values supplied in \code{names.}. If a supplied name matches the name of a \code{...} argument, that argument is returned. Otherwise, the element of \code{defs.} with a matching name is returned. \code{names. = NULL} and \code{names. = NA} are converted to \code{'NULL'} and \code{'NA'}. Reserved words in \code{names.} should be backtick quoted.
\cr                \tab   
\cr        \code{dot}   \tab A convenience version of \code{dots} for extracting a single named \code{...} argument (or if a matching \code{...} argument name is not found, its default value \code{def.}).
\cr                \tab   
\cr  \code{dot_names}   \tab If \code{names. = NULL}, the return value is \code{...names()}, otherwise, \code{names.} is returned. Throws an error in the following cases:\itemize{
\item \verb{0 ...} Arguments are supplied.
\item \code{names. != NULL} and \code{length(names.) != ...length()}.
\item \code{req. = TRUE}, \code{names. = NULL}, and no \code{...} arguments are named.
\item \code{names. = NULL} and either (a) \code{blank. = TRUE} and any \code{...} argument is unnamed or its name is blank or (b) \item \code{u. = TRUE} and any \code{...} argument name is duplicated.
}}
}
