% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\encoding{UTF-8}
\name{dp}
\alias{dp}
\alias{dp_agg}
\alias{dp_n}
\alias{dp_all}
\alias{dp_grp}
\alias{dp_join}
\alias{dp_lab}
\alias{dp_mod}
\title{Thin wrappers of \code{dplyr} functions}
\usage{
dp()

dp_agg(.data, ..., .groups = NULL)

dp_n()

dp_all(x)

dp_grp(X, Keys)

dp_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

dp_lab(.data, ...)

dp_mod(.data, ...)
}
\arguments{
\item{x}{A data.frame}

\item{keys}{A \link[=cmp_chr_vec]{complete character vec} naming grouping variables in \code{x}.}
}
\value{
A \code{data.frame}.
}
\description{
Thin wrappers of \code{dplyr} functions
}
\details{
\tabular{ll}{  \code{dp_join}   \tab Thinly wraps \code{\link[dplyr]{left_join}}.               \cr
  \code{dp_agg}    \tab Thinly wraps \code{\link[dplyr]{summarize}}.               \cr
  \code{dp_all}    \tab Thinly wraps \code{\link[dplyr]{all_of}}.                  \cr
  \code{dp_lab}    \tab Thinly wraps \code{\link[dplyr]{rename}}.                  \cr
  \code{dp_mod}    \tab Thinly wraps \code{\link[dplyr]{mutate}}.                  \cr
  \code{dp_grp}    \tab Limitedly wraps \code{\link[dplyr]{group_by}}\eqn{^{(1)}}  \cr
  \code{dp_n}      \tab Thinly wraps \code{\link[dplyr]{n}}.                         }
\tabular{l}{  \eqn{^{(1)}} Accepts as argument only a data.frame (\code{x}) and a character vector naming grouping variables (\code{keys}).}
}
\seealso{
Other wraps: 
\code{\link{basics}()},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{rd}()},
\code{\link{sca}()},
\code{\link{tb}()}
}
\concept{wraps}
