% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\encoding{UTF-8}
\name{dups}
\alias{dups}
\alias{idups}
\title{Duplicate values.}
\usage{
dups(...)

idups(x, int = F)
}
\arguments{
\item{...}{One or more atomic vectors to be evaluated for duplicated values.}

\item{x}{An \link[=atm_vec]{atomic vec}.}

\item{int}{A non-\code{NA} logical scalar indicating whether to return a logical vector (\code{int = FALSE}) or an integer vector (\code{int = TRUE}).}
}
\value{
\emph{A logical vector} (when \verb{int = FALSE)}
\cr * or an integer vector* (when \verb{int = TRUE})
\cr    \code{idups}
\cr\cr \emph{An atomic vector}
\cr    \code{dups}
}
\description{
Extensions of \code{\link[base]{duplicated}}.
\tabular{rl}{
  \code{idups}   \tab Indexes all elements of \code{x} that are duplicates of its earlier element, either as a logical vector of the same length as \code{x} (when \code{int = FALSE}) or as an integer vector whose length is the same as the number of elements in \code{x} that are duplicates of earlier elements (when \code{int = TRUE}).
\cr          \tab  
\cr \code{dups}   \tab \link[=av]{Atomizes} \code{...}, then reduce the resulting vector to contain one copy of each duplicated element.
}
}
\examples{
dups(0:5, 5:10, 10:15, 15:20)
idups(c(0:5, 5:10, 10:15, 15:20))
idups(c(0:5, 5:10, 10:15, 15:20), int = T)
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unq}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()},
\code{\link{x}()}
}
\concept{extensions}
