% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eee.R
\name{eee}
\alias{eee}
\alias{eee_props}
\alias{is_eee_spec}
\alias{ieee}
\alias{ie0D}
\alias{ie1D}
\alias{ie2D}
\alias{ieHD}
\alias{ieUD}
\alias{neee}
\title{Effective dimensionality (\code{effective-D}) properties}
\usage{
eee(x)

eee_props()

is_eee_spec(spec)

ieee(x, spec, ...)

ie0D(x, ...)

ie1D(x, ...)

ie2D(x, ...)

ieHD(x, ...)

ieUD(x, ...)

neee(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{effective-D} properties from \code{eee_props()}. \code{Effective-D} properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_eee_spec}   \tab A logical scalar.
\cr \code{eee_props}   \tab A character vector.
\cr      \code{ieee}   \tab A logical scalar.
\cr      \code{iEEE}   \tab A logical scalar.
\cr      \code{neee}   \tab A numeric scalar.
\cr       \code{eee}   \tab A character vector.
}
}
\description{
An object's \code{effective-D} is defined as the number of dimensions in which it has multiple populated index positions (i.e., if a dimension has only one index position, that is not an effective dimension).
\tabular{rl}{
  \code{'eUD'}   \tab \emph{Effectively} \code{NaN} \emph{dimensional}.
\cr             \tab Length-\code{0} objects including \code{NULL} (of undefined \code{effective-D}).
\cr             \tab   
\cr   \code{'e0D'}   \tab \emph{Effectively} \code{0} \emph{dimensional}
\cr             \tab \verb{1x1 data.frames} and length-\verb{1 vectors}, \code{\link[=ivls]{vlists}}, and \code{arrays}.
\cr             \tab   
\cr   \code{'e1D'}   \tab \emph{Effectively} \code{1} \emph{dimensional}
\cr             \tab \code{\link[=imvc]{multivecs}}, length-\code{2+ vlists}, \code{\link[=irow]{row}} \code{data.frames}, \code{\link[=icol]{col}} \code{data.frames}, and length-\verb{2+} arrays with \verb{2+} index positions in exactly \code{1} dimension.
\cr             \tab   
\cr   \code{'e2D'}   \tab \emph{Effectively} \code{2} \emph{dimensional}
\cr             \tab \verb{2+ x 2+ data.frames} and length-\code{4+ arrays} with \verb{2+} positions in exactly \code{2} dimensions.
\cr             \tab   
\cr   \code{'eHD'}   \tab \emph{Effectively} \code{hyper} \emph{dimensional}
\cr             \tab Length-\code{8+ arrays} with \verb{2+} positions in \verb{3+} dimensions.
}
\cr\cr
Effective dimensionality property functions are:
\tabular{rl}{
  \code{is_eee_spec}   \tab Is \code{spec} an \code{effective-D} specification?
\cr                \tab   
\cr  \code{eee_props}   \tab What \code{effective-D} properties are there?
\cr                \tab   
\cr       \code{neee}   \tab How many effective dimensions does \code{x} have?
\cr                \tab   
\cr       \code{ieee}   \tab Is \code{x} a match to the \code{effective-D} specification\code{spec}?
\cr                \tab   
\cr       \code{iEEE}   \tab Is \code{x} a match to the single \code{effective-D} property \code{'EEE'}?
\cr                \tab   
\cr        \code{eee}   \tab What are \code{x}'s \code{effective-D} properties?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
a. <- "a"
abc. <- c("a", "b", "c")
udf1 <- NULL
udf2 <- character(0)
udf3 <- matrix(nrow = 0, ncol = 0)
udf4 <- list()
zro1 <- a.
zro2 <- list(abc = abc.)
zro3 <- data.frame(a = a.)
zro4 <- array(a., dim = c(1, 1, 1))
one1 <- matrix(abc., nrow = 1)
one2 <- matrix(abc., ncol = 1)
one3 <- data.frame(a. = abc.)
one4 <- array(abc., dim = c(3, 1, 1))
two1 <- matrix(rep.int(abc., 2), nrow = 2)
two2 <- data.frame(a = abc., b = abc.)
two3 <- array(1:4, dim = c(2, 2, 1))
two4 <- array(1:4, dim = c(1, 2, 2, 1))
hyp1 <- array(1:8, dim = c(2, 2, 2))
hyp2 <- array(1:8, dim = c(2, 2, 2, 1))
hyp3 <- array(1:8, dim = c(1, 2, 2, 2, 1))
hyp4 <- array(1:16, dim = c(2, 2, 2, 2))

is_eee_spec("eUD")
is_eee_spec("d1D")
eee_props()

c(neee(udf1), neee(udf2), neee(udf3), neee(udf4))
c(neee(zro1), neee(zro2), neee(zro3), neee(zro4))
c(neee(one1), neee(one2), neee(one3), neee(one4))
c(neee(two1), neee(two2), neee(two3), neee(two4))
c(neee(hyp1), neee(hyp2), neee(hyp3), neee(hyp4))

c(ieee(one1, "eUD"), ieee(one1, "e0D"), ieee(one1, "e1D"), ieee(one1, "e2D"), ieee(one1, "eHD"))
c(ieUD(one1), ie0D(one1), ie1D(one1), ie2D(one1), ieHD(one1))

c(eee(udf1), eee(zro1), eee(one1), eee(two1), eee(hyp1), eee(hyp4))
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{iii}()},
\code{\link{is_bbb_ccc_prop}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
