% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eee.R
\encoding{UTF-8}
\name{eee}
\alias{eee}
\alias{eee_funs}
\alias{eee_props}
\alias{is_eee_spec}
\alias{EEE}
\alias{E0D}
\alias{E1D}
\alias{E2D}
\alias{EHD}
\alias{EUD}
\alias{neee}
\title{Effective dimensionality (effective.D) properties}
\usage{
eee(x)

eee_funs()

eee_props()

is_eee_spec(spec)

EEE(x, spec, ...)

E0D(x, ...)

E1D(x, ...)

E2D(x, ...)

EHD(x, ...)

EUD(x, ...)

neee(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more effective.D properties from \code{eee_props()}. Effective.D properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \verb{eee_props, eee_funs, eee}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{is_eee_spec, EEE, \{EEE\}}
\cr\cr  \strong{An integer scalar}  \cr\cr \code{neee}
}
\description{
An object's effective.D is defined as the number of dimensions in which it has multiple populated index positions (i.e., if a dimension has only one index position, that is not an effective dimension).
\tabular{ll}{  \verb{'eud', 'EUD'}   \tab \emph{Effectively} \code{NaN} \emph{dimensional}.                                                                                                                                                                                      \cr
  \tab Length-\code{0} objects including \code{NULL} (of undefined effective.D).                                                                                                                                            \cr   \tab   \cr
  \verb{'e0d', 'E0D'}   \tab \emph{Effectively} \code{0} \emph{dimensional}                                                                                                                                                                                         \cr
  \tab \verb{1x1} data.frames and length-\code{1} vectors, \code{\link[=VLS]{vlists}}, and arrays.                                                                                                                          \cr   \tab   \cr
  \verb{'e1d', 'E1D'}   \tab \emph{Effectively} \code{1} \emph{dimensional}                                                                                                                                                                                         \cr
  \tab \code{\link[=MVC]{multivecs}}, length-\verb{2+} vlists, \code{\link[=ROW]{row}} data.frames, \code{\link[=COL]{column}} data.frames, and length-\verb{2+} arrays with \verb{2+} index positions in exactly \code{1} dimension. \cr   \tab   \cr
  \verb{'e2d', 'E2D'}   \tab \emph{Effectively} \code{2} \emph{dimensional}                                                                                                                                                                                         \cr
  \tab \verb{2+ x 2+ }data.frames and length-\verb{4+} arrays with \verb{2+} positions in exactly \code{2} dimensions.                                                                                                                \cr   \tab   \cr
  \verb{'ehd', 'EHD'}   \tab \emph{Effectively hyper dimensional}                                                                                                                                                                            \cr   \tab   \cr
  \tab Length-\verb{8+} arrays with \verb{2+} positions in \verb{3+} dimensions.                                                                                                                                                                }
\cr Effective dimensionality property functions are:
\tabular{ll}{  \code{is_eee_spec}   \tab Is \code{spec} an effective.D specification?                                                                                                    \cr   \tab   \cr
  \code{eee_props}    \tab What effective.D properties are there?                                                                                                      \cr   \tab   \cr
  \code{eee_funs}     \tab What effective.D property functions are there?                                                                                              \cr   \tab   \cr
  \code{{EEE}}        \tab Is \code{x} a match to the single effective.D property \code{'{EEE}'} where \code{{EEE}} is a placeholder for any given effective dimensionality property? \cr   \tab   \cr
  \code{neee}         \tab How many effective dimensions does \code{x} have?                                                                                                \cr   \tab   \cr
  \code{eee}          \tab What are \code{x}'s effective.d properties?                                                                                                      \cr   \tab   \cr
  \code{EEE}          \tab Is \code{x} a match to the effective.D specification in \code{spec}?                                                                                                 }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.MAX, .MAXR, .MAXC}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.MIN, .MINR, .MINC}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.LT, .LE, .GE, .GT}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.N, .NR, .NC}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.VALS}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
a. <- "a"
abc. <- c("a", "b", "c")
udf1 <- NULL
udf2 <- character(0)
udf3 <- matrix(nrow = 0, ncol = 0)
udf4 <- list()
zro1 <- a.
zro2 <- list(abc = abc.)
zro3 <- data.frame(a = a.)
zro4 <- array(a., dim = c(1, 1, 1))
one1 <- matrix(abc., nrow = 1)
one2 <- matrix(abc., ncol = 1)
one3 <- data.frame(a. = abc.)
one4 <- array(abc., dim = c(3, 1, 1))
two1 <- matrix(rep.int(abc., 2), nrow = 2)
two2 <- data.frame(a = abc., b = abc.)
two3 <- array(1:4, dim = c(2, 2, 1))
two4 <- array(1:4, dim = c(1, 2, 2, 1))
hyp1 <- array(1:8, dim = c(2, 2, 2))
hyp2 <- array(1:8, dim = c(2, 2, 2, 1))
hyp3 <- array(1:8, dim = c(1, 2, 2, 2, 1))
hyp4 <- array(1:16, dim = c(2, 2, 2, 2))

eee_funs()
eee_props()
is_eee_spec("eUD")
is_eee_spec("d1D")
c(neee(udf1), neee(udf2), neee(udf3), neee(udf4))
c(neee(zro1), neee(zro2), neee(zro3), neee(zro4))
c(neee(one1), neee(one2), neee(one3), neee(one4))
c(neee(two1), neee(two2), neee(two3), neee(two4))
c(neee(hyp1), neee(hyp2), neee(hyp3), neee(hyp4))
c(EEE(one1, "eud"), EEE(one1, "e0d"), EEE(one1, "e1d"), EEE(one1, "e2d"), EEE(one1, "ehd"))
c(EUD(one1), E0D(one1), E1D(one1), E2D(one1), EHD(one1))
c(eee(udf1), eee(zro1), eee(one1), eee(two1), eee(hyp1), eee(hyp4))
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp_fast}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
