% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eee.R
\name{eee}
\alias{eee}
\alias{eee_props}
\alias{is_eee_spec}
\alias{ieee}
\alias{ie0D}
\alias{ie1D}
\alias{ie2D}
\alias{ieHD}
\alias{ieUD}
\alias{neee}
\title{Effective dimensionality properties}
\usage{
eee(x)

eee_props()

is_eee_spec(spec)

ieee(x, spec, ...)

ie0D(x, ...)

ie1D(x, ...)

ie2D(x, ...)

ieHD(x, ...)

ieUD(x, ...)

neee(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more effective dimensionality properties (i.e., from \code{eee_props()}). Effective dimensionality properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\itemize{
\item \strong{\code{eee_vals}}: a character vector.
\item \strong{\code{eee}}: a character scalar/vector.
\item \strong{\code{neee}}: a numeric scalar.
\item \strong{\verb{ieee, ixxx}}: a logical scalar.
\item \strong{\code{is_eee_spec}}: a logical scalar.
}
}
\description{
Effective dimensionality of a non-empty object is defined as the number of dimensions with \verb{2+} positions. Effective dimensionality is undefined for empty objects.
\itemize{
\item \strong{\code{'eUD'}}: Effectively \code{NaN}-dimensional (\code{NULL} or otherwise of length-\code{0} objects; effective dimensionality is undefined).
\item \strong{\code{'e0D'}}: Effectively \code{0}-dimensional (vectors, vlists, and arrays of length \code{1} and \verb{1 × 1} data.frames).
\item \strong{\code{'e1D'}}: Effectively \code{1}-dimensional (vectors or vlists of length \verb{2+}, length \verb{2+} arrays with \verb{2+} positions in just \code{1} dimension, and \verb{1 x 2+} or \verb{2+ x 1} data.frames).
\item \strong{\code{'e2D'}}: Effectively \code{2}-dimensional (\verb{2+ x 2+} data frames or matrices and length \verb{4+} arrays with \verb{2+} positions in just \code{2} dimensions).
\item \strong{\code{'eHD'}}: Effectively hyper-dimensional (length-\verb{8+} array with \verb{2+} positions in \verb{3+} dimensions).
}
\strong{Functions}
\itemize{
\item \strong{\code{eee}}: Gets a character vector containing all effective dimensionality properties possessed by \code{x}.
\item \strong{\code{ixxx}}: Evaluates whether \code{x} possesses the effective dimensionality property \code{xxx} (a placeholder for any given effective dimensionality property value), subject to any restrictions in \code{...}.
\item \strong{\code{ieee}}: Evaluates whether \code{x} possesses one or more (possibly pipe-delimited) effective dimensionality properties in \code{spec}, subject to any restrictions in \code{...}.
\item \strong{\code{neee}}: Get the number of effective dimensions of \code{x}.
\item \strong{\code{eee_props}}: Gets a character vector of all possible effective dimensionality property values.
\item \strong{\code{is_eee_spec}}: Evaluates whether \code{spec} is a valid effective dimensionality property specification.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
\verb{max,maxr,maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively..
\cr                   \tab  
\cr \verb{min,minr,minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively..
\cr                   \tab  
\cr   \verb{lt,le,ge,gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                   \tab  
\cr       \verb{n,nr,nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively..
\cr                   \tab  
\cr          \code{vals}   \tab A vector of valid values.
}
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
