% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eee.R
\name{eee}
\alias{eee}
\alias{eee_props}
\alias{is_eee_spec}
\alias{ieee}
\alias{ie0D}
\alias{ie1D}
\alias{ie2D}
\alias{ieHD}
\alias{ieUD}
\alias{neee}
\title{Effective dimensionality properties}
\usage{
eee(x)

eee_props()

is_eee_spec(spec)

ieee(x, spec, ...)

ie0D(x, ...)

ie1D(x, ...)

ie2D(x, ...)

ieHD(x, ...)

ieUD(x, ...)

neee(x)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more effective dimensionality properties (i.e., from \code{eee_props()}). Effective dimensionality properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_eee_spec}   \tab A logical scalar.
\cr \code{eee_props}   \tab A character vector.
\cr      \code{neee}   \tab A numeric scalar.
\cr      \code{ieee}   \tab A logical scalar.
\cr      \code{ixxx}   \tab A logical scalar.
\cr       \code{eee}   \tab A character vector.
}
}
\description{
Effective dimensionality of a non-empty object is defined as the number of dimensions with \verb{2+} positions. Effective dimensionality is undefined for empty objects.
\tabular{rl}{
  \code{'eUD'}   \tab Effectively \code{NaN}-dimensional (\code{NULL} or otherwise of length-\code{0} objects; effective dimensionality is undefined).
\cr             \tab  
\cr   \code{'e0D'}   \tab Effectively \code{0}-dimensional (vectors, vlists, and arrays of length \code{1} and \verb{1 × 1} data.frames).
\cr             \tab  
\cr   \code{'e1D'}   \tab Effectively \code{1}-dimensional (vectors or vlists of length \verb{2+}, length \verb{2+} arrays with \verb{2+} positions in \code{1} dimension, and \verb{1 x 2+} or \verb{2+ x 1} data.frames).
\cr             \tab  
\cr   \code{'e2D'}   \tab Effectively \code{2}-dimensional (\verb{2+ x 2+} data frames or matrices and length \verb{4+} arrays with \verb{2+} positions in \code{2} dimensions).
\cr             \tab  
\cr   \code{'eHD'}   \tab Effectively hyper-dimensional (length-\verb{8+} array with \verb{2+} positions in \verb{3+} dimensions).
}
Effective dimensionality property functions are:
\tabular{rl}{
  \code{is_eee_spec}   \tab Evaluates whether \code{spec} is a valid effective dimensionality property specification.
\cr                \tab  
\cr  \code{eee_props}   \tab Gets a character vector of all possible effective dimensionality property values.
\cr                \tab  
\cr       \code{ixxx}   \tab Evaluates whether \code{x} possesses the effective dimensionality property \code{xxx} (a placeholder for any given effective dimensionality property value), subject to any restrictions in \code{...}.
\cr                \tab  
\cr       \code{ieee}   \tab Evaluates whether \code{x} possesses one or more (possibly pipe-delimited) effective dimensionality properties in \code{spec}, subject to any restrictions in \code{...}.
\cr                \tab  
\cr       \code{neee}   \tab Gets the number of effective dimensions of \code{x}.
\cr                \tab  
\cr        \code{eee}   \tab Gets a character vector of all effective dimensionality properties of \code{x}.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
a. <- "a"
abc. <- c("a", "b", "c")
udf1 <- NULL
udf2 <- character(0)
udf3 <- matrix(nrow = 0, ncol = 0)
udf4 <- list()
zro1 <- a.
zro2 <- list(abc = abc.)
zro3 <- data.frame(a = a.)
zro4 <- array(a., dim = c(1, 1, 1))
one1 <- matrix(abc., nrow = 1)
one2 <- matrix(abc., ncol = 1)
one3 <- data.frame(a. = abc.)
one4 <- array(abc., dim = c(3, 1, 1))
two1 <- matrix(rep.int(abc., 2), nrow = 2)
two2 <- data.frame(a = abc., b = abc.)
two3 <- array(1:4, dim = c(2, 2, 1))
two4 <- array(1:4, dim = c(1, 2, 2, 1))
hyp1 <- array(1:8, dim = c(2, 2, 2))
hyp2 <- array(1:8, dim = c(2, 2, 2, 1))
hyp3 <- array(1:8, dim = c(1, 2, 2, 2, 1))
hyp4 <- array(1:16, dim = c(2, 2, 2, 2))

is_eee_spec("eUD")
is_eee_spec("d1D")
eee_props()

c(neee(udf1), neee(udf2), neee(udf3), neee(udf4))
c(neee(zro1), neee(zro2), neee(zro3), neee(zro4))
c(neee(one1), neee(one2), neee(one3), neee(one4))
c(neee(two1), neee(two2), neee(two3), neee(two4))
c(neee(hyp1), neee(hyp2), neee(hyp3), neee(hyp4))

c(ieee(one1, "eUD"), ieee(one1, "e0D"), ieee(one1, "e1D"), ieee(one1, "e2D"), ieee(one1, "eHD"))
c(ieUD(one1), ie0D(one1), ie1D(one1), ie2D(one1), ieHD(one1))

c(eee(udf1), eee(zro1), eee(one1), eee(two1), eee(hyp1), eee(hyp4))

}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
