% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ex}
\alias{ex}
\title{Extract rows, columns, and/or elements}
\usage{
ex(x, r = NULL, c = NULL, e = NULL)
}
\arguments{
\item{x}{Any valid R object.}

\item{r}{A logical or integer vector indexing rows to extract from \code{x}
of \link[=ddim]{defined dimensionality} 2.}

\item{c}{A logical or integer vector indexing columns to extract from
\code{x} of \link[=ddim]{defined dimensionality} 2.}

\item{e}{A logical or integer vector indexing elements to extract from
\code{x} of \link[=ddim]{defined dimensionality} 1.}
}
\description{
Extract specific rows and/or columns of matrices and tibbles
or extract specific elements of vects or vlists.
}
\examples{
Mat <- matrix(1:100, nrow = 10)
Dat   <- as.tibble(MATRIX)
iR <- 1:10 \%in\% 1:5
iC <- 1:10 \%in\% 6:10
iE <- which(1:26 \%in\% 1:13)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, r = 1:5, c = 6:10)
ex(mat, iR = 1:5, iC = 6:10)
ex(mat, r = 20)
ex(Dat, c = 6:10)
ex(Dat, e = 1:10)
ex(letters, e = 1:13)
ex(letters, e = iE)
ex(letters, r = iR, c = iC)
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
