% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failsafe.R
\name{failsafe}
\alias{failsafe}
\alias{isERR}
\alias{notERR}
\alias{msgERR}
\title{Evaluate objects and flag any errors}
\usage{
failsafe(x)

isERR(x)

notERR(x)

msgERR(x)
}
\arguments{
\item{x}{An object or a call to evaluate in the environment of a parent function where the initial call was made.}
}
\value{
\itemize{
\item \strong{\code{msgERR}}: either \code{NULL} or a character scalar.
\item \strong{\code{failsafe}}: either \code{x}, an object of class \code{'error'}, or an object of class \code{'simpleError'}.
\item \strong{\verb{isERR, notERR}}: A logical scalar.
}
}
\description{
\itemize{
\item \strong{\code{isERR}}: Evaluate whether calling \code{identity(x)} produced an error.
\item \strong{\code{notERR}}: Evaluate whether calling \code{identity(x)} \emph{did not} produce an error.
\item \strong{\code{msgERR}}: Get any error message associated with calling \code{identity(x)}. If there is none, returns \code{NULL}.
\item \strong{\code{failsafe}}: Call \code{identity(x)}. If an error is produced in doing so, returns an object of class \code{'error'} or of class \code{'simpleError'}. If an error is not generated, returns the safely evaluated value of \code{x}.
}
}
\seealso{
Other errs: 
\code{\link{err_check}()}

Other failsafe: 
\code{\link{binary_failsafe}},
\code{\link{is_failsafe}}

Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{errs}
\concept{extensions}
\concept{failsafe}
