% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failsafe.R
\name{failsafe}
\alias{failsafe}
\alias{isERR}
\alias{notERR}
\alias{msgERR}
\title{Evaluate objects, flag any errors}
\usage{
failsafe(x)

isERR(x)

notERR(x)

msgERR(x)
}
\arguments{
\item{x}{An object or a call to evaluate in the environment of a parent
function where the initial call was made.}
}
\value{
\tabular{lll}{
\code{failsafe}            \tab   \tab Either \code{x}, an object of class
\code{'error'}, or an object of class
\code{'simpleError'}.             \cr
\code{isERR}, \code{notERR}\tab   \tab A logical scalar.                \cr
\code{msgERR}              \tab   \tab Either \code{NULL} or a character
scalar.                            }
}
\description{
\tabular{ll}{
\code{failsafe}   \tab Attempts to call \code{identity(x)}. If an error is
produced in doing so, returns an object of class
\code{'error'} or of class \code{'simpleError'}. If
an error is not generated, returns the safely
evaluated value of \code{x}.                     \cr
\code{isERR}      \tab Evaluates whether calling \code{identity(x)}
produced an error.                               \cr
\code{notERR}     \tab Evaluates whether calling \code{identity(x)} did not
produce an error.                                \cr
\code{msgERR}     \tab Gets any error message associated with calling
\code{identity(x)}. If there is none, returns
\code{NULL}.                                       }
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put.}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{xb}()}

Other errs: 
\code{\link{errs_uj}}

Other failsafe: 
\code{\link{binfuns}},
\code{\link{is_failsafe}}
}
\concept{errs}
\concept{extensions}
\concept{failsafe}
