% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failsafe.R
\name{failsafe}
\alias{failsafe}
\alias{isERR}
\alias{notERR}
\alias{msgERR}
\title{Evaluate Objects, Flag Any Errors}
\usage{
failsafe(x)

isERR(x)

notERR(x)

msgERR(x)
}
\arguments{
\item{x}{An object or a call to evaluate in the environment of a parent
function where the initial call was made.}
}
\value{
\strong{\code{failsafe}}
\cr Either \code{x}, an object of class \code{'error'}, or an object of
class \code{'simpleError'}.
\cr\cr
\strong{\code{isERR, notERR}}
\cr A logical scalar.
\cr\cr
\strong{\code{msgERR}}
\cr Either \code{NULL} or a character scalar.
}
\description{
Evaluate Objects, Flag Any Errors
}
\section{Functions in This Family}{

\strong{\code{failsafe}}
\cr Attempts to call \code{identity(x)}. If an error is produced in doing
so, returns an object of class \code{'error'} or of class
\code{'simpleError'}. If an error is not generated, returns the safely
evaluated value of \code{x}.
\cr\cr
\strong{\code{isERR}}
\cr Evaluates whether calling \code{identity(x)} produced an error.
\cr\cr
\strong{\code{notERR}}
\cr Evaluates whether calling \code{identity(x)} did not produce an error.
\cr\cr
\strong{\code{msgERR}}
\cr Gets any error message associated with calling \code{identity(x)}. If
there is none, returns \code{NULL}.
}

\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}

Other errs: 
\code{\link{err_check}()}

Other failsafe: 
\code{\link{binfuns}},
\code{\link{is_failsafe}}
}
\concept{errs}
\concept{extensions}
\concept{failsafe}
