% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failsafe.R
\encoding{UTF-8}
\name{failsafe}
\alias{failsafe}
\alias{isERR}
\alias{notERR}
\alias{msgERR}
\title{Evaluate objects and flag any errors}
\usage{
failsafe(x)

isERR(x)

notERR(x)

msgERR(x)
}
\arguments{
\item{x}{An object or a call to evaluate in the environment of a parent function where the initial call was made.}
}
\value{
\emph{Either} \code{x}, \emph{an object of class} \code{'error'} \emph{or of class} \code{'simpleError'}
\cr    \code{failsafe}
\cr\cr \code{NULL} \emph{or a character scalar}
\cr    \code{msgERR}
\cr\cr \emph{A logical scalar}
\cr    \code{notERR}
\cr    \code{isERR}
}
\description{
\tabular{rl}{
  \code{failsafe}   \tab Calls \code{identity(x)}. If an error is produced in doing so, returns an object of class \code{'error'} or of class \code{'simpleError'}. If an error is not generated, returns the safely evaluated value of \code{x}.
\cr            \tab  
\cr \code{notERR}   \tab Evaluates whether calling \code{identity(x)} \emph{does not} produce an error.
\cr            \tab  
\cr \code{msgERR}   \tab Gets any error message associated with calling \code{identity(x)}. If there is none, returns \code{NULL}.
\cr            \tab  
\cr  \code{isERR}   \tab Evaluates whether calling \code{identity(x)} produced an error.
}
}
\examples{
failsafe(pi)
notERR(pi)
msgERR(pi)
isERR(pi)

failsafe(non.existent.variable)
notERR(non.existent.variable)
msgERR(non.existent.variable)
isERR(non.existent.variable)
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{fork}()},
\code{\link{is_unq}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()},
\code{\link{x}()}

Other failsafe: 
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}}

Other errs: 
\code{\link{err_check}()}
}
\concept{errs}
\concept{extensions}
\concept{failsafe}
