% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failsafe.R
\name{failsafe}
\alias{failsafe}
\alias{isERR}
\alias{notERR}
\alias{msgERR}
\title{Evaluate objects and flag any errors}
\usage{
failsafe(x)

isERR(x)

notERR(x)

msgERR(x)
}
\arguments{
\item{x}{An object or a call to evaluate in the environment of a parent function where the initial call was made.}
}
\value{
\tabular{rl}{
\code{failsafe}   \tab Either \code{x}, an object of class \code{'error'}, or an object of class \code{'simpleError'}.
\cr \code{msgERR}   \tab \code{NULL} or a character scalar.
\cr \code{notERR}   \tab A logical scalar.
\cr  \code{isERR}   \tab A logical scalar.
}
}
\description{
\tabular{rl}{
\code{failsafe}   \tab Calls \code{identity(x)}. If an error is produced in doing so, returns an object of class \code{'error'} or of class \code{'simpleError'}. If an error is not generated, returns the safely evaluated value of \code{x}.
\cr \code{notERR}   \tab Evaluates whether calling \code{identity(x)} \emph{does not} produce an error.
\cr \code{msgERR}   \tab Gets any error message associated with calling \code{identity(x)}. If there is none, returns \code{NULL}.
\cr  \code{isERR}   \tab Evaluates whether calling \code{identity(x)} produced an error.
}
}
\examples{
failsafe(pi)
notERR(pi)
msgERR(pi)
isERR(pi)

failsafe(non.existent.variable)
notERR(non.existent.variable)
msgERR(non.existent.variable)
isERR(non.existent.variable)
}
\seealso{
Other errs: 
\code{\link{err_check}()}

Other failsafe: 
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}}
}
\concept{errs}
\concept{failsafe}
