% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\name{fsub}
\alias{fsub}
\title{Fixed-value string substitution}
\usage{
fsub(x, pats, subs)
}
\arguments{
\item{x}{Atomic character object.}

\item{pats}{Character vect containing string patterns to be replaced.}

\item{subs}{Character vect containing replacement strings.}
}
\value{
Atomic character object.
}
\description{
Fixed-value string substitution
}
\details{
Recycles \code{pats} and \code{subs}
}
