% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\name{fsub}
\alias{fsub}
\title{Fixed-value string substitution}
\usage{
fsub(x, pats, subs)
}
\arguments{
\item{x}{An \link[=atm_chr]{atomic character object}.}

\item{pats, subs}{\link[=cmp_chr_vec]{Complete character vecs} containing patterns to replace and their associated replacement strings.}
}
\value{
Atomic character object.
}
\description{
Replace strings in \code{pats} with corresponding (possibly recycled) strings in \code{subs}.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkd}},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
