% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\name{fsub}
\alias{fsub}
\title{Fixed-Value String Substitution}
\usage{
fsub(x, pats, subs)
}
\arguments{
\item{x}{An \link[=atm_chr]{atomic character object}.}

\item{pats, subs}{\link[=cmp_chr_vec]{Complete character vecs} containing
patterns to replace and their associated replacement strings.}
}
\value{
Atomic character object.
}
\description{
Replace strings in \code{pats} with corresponding (possibly recycled)
strings in \code{subs}.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{charn}()},
\code{\link{delim}},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
\concept{strings}
