% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\encoding{UTF-8}
\name{fsub}
\alias{fsub}
\title{Fixed-value string substitution}
\usage{
fsub(X, Pats, Subs)
}
\arguments{
\item{X}{An \link[=atm_chr]{atomic character object}.}

\item{Pats}{A \link[=cmp_chr_vec]{complete character vec} of patterns to replace.}

\item{Subs}{A complete character vec of replacement strings}
}
\value{
Atomic character object.
}
\description{
Finds strings from \code{Pats} in \code{X} and replaces them with corresponding values of (possibly recycled) \code{Subs}.
}
\examples{
egArg1 <- c("a", "few", "words", "in", "a string")
agArg2 <- paste0(egArg1, collapse = " ")
egPats <- c("a", "f", "w", "i", "s", " ")
egSubs <- c("A", "F", "W", "I", "S", "_")
egArg1
egArg2
fsub(egArg1, Pats, Subs)
fsub(egArg2, Pats, Subs)
fsub(egArg1, Pats, "_")
fsub(egArg2, Pats, "_")
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
