% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsub.R
\encoding{UTF-8}
\name{fsub}
\alias{fsub}
\title{Fixed-value string substitution}
\usage{
fsub(x, pats, subs)
}
\arguments{
\item{x}{An \link[=atm_chr]{atomic character object}.}

\item{pats}{A \link[=cmp_chr_vec]{complete character vec} of patterns to replace.}

\item{subs}{A complete character vec of replacement strings}
}
\value{
Atomic character object.
}
\description{
Finds strings from \code{pats} in \code{x} and replaces them with corresponding values of (possibly recycled) \code{subs}.
}
\examples{
arg1. <- c("a", "few", "words", "in", "a string")
arg2. <- paste0(arg1., collapse = " ")
pats. <- c("a", "f", "w", "i", "s", " ")
subs. <- c("A", "F", "W", "I", "S", "_")

arg1.
arg2.

fsub(arg1., pats., subs.)
fsub(arg1., pats., "_")
fsub(arg2., pats., subs.)
fsub(arg2., pats., "_")
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}},
\code{\link{markdown}},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
