% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp.R
\encoding{UTF-8}
\name{ggp}
\alias{ggp}
\alias{gg_aes}
\alias{gg_dot}
\alias{gg_elin}
\alias{gg_emkd}
\alias{gg_enll}
\alias{gg_erct}
\alias{gg_etxt}
\alias{gg_exp}
\alias{gg_grid}
\alias{gg_hlin}
\alias{gg_lab}
\alias{gg_labs}
\alias{gg_lin}
\alias{gg_mar}
\alias{gg_sac}
\alias{gg_scm}
\alias{gg_sdm}
\alias{gg_slm}
\alias{gg_sav}
\alias{gg_sxc}
\alias{gg_syc}
\alias{gg_thm}
\alias{gg_thm0}
\alias{gg_txt}
\alias{gg_unit}
\alias{gg_vars}
\alias{gg_vlin}
\alias{gg_wrap}
\alias{gg_xlab}
\alias{gg_ylab}
\title{Thin wrappers of \code{ggplot2} and \code{ggtext} functions}
\usage{
ggp(...)

gg_aes(...)

gg_dot(...)

gg_elin(...)

gg_emkd(...)

gg_enll(...)

gg_erct(...)

gg_etxt(...)

gg_exp(...)

gg_grid(...)

gg_hlin(...)

gg_lab(...)

gg_labs(...)

gg_lin(...)

gg_mar(...)

gg_sac(...)

gg_scm(...)

gg_sdm(...)

gg_slm(...)

gg_sav(...)

gg_sxc(...)

gg_syc(...)

gg_thm(...)

gg_thm0(...)

gg_txt(...)

gg_unit(...)

gg_vars(...)

gg_vlin(...)

gg_wrap(...)

gg_xlab(...)

gg_ylab(...)
}
\description{
Thin wrappers of \code{ggplot2} and \code{ggtext} functions
}
\details{
These functions thinly wrap \code{ggplot2} functions unless otherwise noted:
\tabular{ll}{  \code{ggp}       \tab \code{\link[ggplot2]{ggplot}}                                   \cr   \tab   \cr
  \code{gg_sav}    \tab \code{\link[ggplot2]{ggsave}}                                   \cr   \tab   \cr
  \code{gg_aes}    \tab \code{\link[ggplot2]{aes}}                                      \cr
  \code{gg_exp}    \tab \code{\link[ggplot2]{expansion}}                                \cr
  \code{gg_mar}    \tab \code{\link[ggplot2]{margin}}                                   \cr   \tab   \cr
  \code{gg_unit}   \tab \code{\link[ggplot2]{unit}}                                     \cr
  \code{gg_vars}   \tab \code{\link[ggplot2]{vars}}                                     \cr   \tab   \cr
  \code{gg_dot}    \tab \code{\link[ggplot2]{geom_point}}                               \cr
  \code{gg_hlin}   \tab \code{\link[ggplot2]{geom_hline}}                               \cr
  \code{gg_lab}    \tab \code{\link[ggplot2]{geom_label}}                               \cr
  \code{gg_lin}    \tab \code{\link[ggplot2]{geom_line}}                                \cr
  \code{gg_txt}    \tab \code{\link[ggplot2]{geom_text}}                                \cr
  \code{gg_vlin}   \tab \code{\link[ggplot2]{geom_vline}}                               \cr   \tab   \cr
  \code{gg_labs}   \tab \code{\link[ggplot2]{labs}}                                     \cr
  \code{gg_xlab}   \tab \code{\link[ggplot2]{xlab}}                                     \cr
  \code{gg_ylab}   \tab \code{\link[ggplot2]{ylab}}                                     \cr   \tab   \cr
  \code{gg_grid}   \tab \code{\link[ggplot2]{facet_grid}}                               \cr
  \code{gg_wrap}   \tab \code{\link[ggplot2]{facet_wrap}}                               \cr   \tab   \cr
  \code{gg_thm}    \tab \code{\link[ggplot2]{theme}}                                    \cr
  \code{gg_tmin}   \tab \code{\link[ggplot2]{theme_minimal}}                            \cr   \tab   \cr
  \code{gg_sac}    \tab \code{\link[ggplot2]{scale_alpha_continuous}}                   \cr
  \code{gg_scm}    \tab \code{\link[ggplot2]{scale_color_manual}}                       \cr
  \code{gg_sdm}    \tab \code{\link[ggplot2]{scale_shape_manual}}                       \cr
  \code{gg_slm}    \tab \code{\link[ggplot2]{scale_linetype_manual}}                    \cr
  \code{gg_sxc}    \tab \code{\link[ggplot2]{scale_x_continuous}}                       \cr
  \code{gg_syc}    \tab \code{\link[ggplot2]{scale_y_continuous}}                       \cr   \tab   \cr
  \code{gg_erct}   \tab \code{\link[ggplot2]{element_rect}}                             \cr
  \code{gg_etxt}   \tab \code{\link[ggplot2]{element_text}}                             \cr
  \code{gg_elin}   \tab \code{\link[ggplot2]{element_line}}                             \cr
  \code{gg_emkd}   \tab \code{\link[ggtext:element_markdown]{ggtext::element_markdown}} \cr
  \code{gg_enll}   \tab \code{\link[ggplot2]{element_blank}}                              }
}
\seealso{
Other plots: 
\code{\link{color}},
\code{\link{gr}},
\code{\link{markdown}()},
\code{\link{pals}},
\code{\link{sca}()},
\code{\link{u2u}()}

Other wraps: 
\code{\link{basics}()},
\code{\link{dp}},
\code{\link{dt}()},
\code{\link{gr}},
\code{\link{rd}()},
\code{\link{sca}()},
\code{\link{tb}()}
}
\concept{plots}
\concept{wraps}
