% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp.R
\encoding{UTF-8}
\name{ggp}
\alias{ggp}
\alias{ggaes}
\alias{ggdot}
\alias{ggelin}
\alias{ggemkd}
\alias{ggenll}
\alias{ggerct}
\alias{ggetxt}
\alias{ggexp}
\alias{gggrid}
\alias{gghlin}
\alias{gglab}
\alias{gglabs}
\alias{gglin}
\alias{ggmar}
\alias{ggsac}
\alias{ggscm}
\alias{ggsdm}
\alias{ggslm}
\alias{ggsav}
\alias{ggsxc}
\alias{ggsyc}
\alias{ggthm}
\alias{ggtmin}
\alias{ggtxt}
\alias{ggunit}
\alias{ggvars}
\alias{ggvlin}
\alias{ggwrap}
\alias{ggxlab}
\alias{ggylab}
\title{Thin wrappers of \code{ggplot2} and \code{ggtext} functions}
\usage{
ggp(...)

ggaes(...)

ggdot(...)

ggelin(...)

ggemkd(...)

ggenll(...)

ggerct(...)

ggetxt(...)

ggexp(...)

gggrid(...)

gghlin(...)

gglab(...)

gglabs(...)

gglin(...)

ggmar(...)

ggsac(...)

ggscm(...)

ggsdm(...)

ggslm(...)

ggsav(...)

ggsxc(...)

ggsyc(...)

ggthm(...)

ggtmin(...)

ggtxt(...)

ggunit(...)

ggvars(...)

ggvlin(...)

ggwrap(...)

ggxlab(...)

ggylab(...)
}
\description{
Thin wrappers of \code{ggplot2} and \code{ggtext} functions
}
\details{
These functions thinly wrap \code{ggplot2} functions unless otherwise noted:
\tabular{ll}{  \code{ggp}      \tab \code{\link[ggplot2]{ggplot}}                                   \cr   \tab   \cr
  \code{ggsav}    \tab \code{\link[ggplot2]{ggsave}}                                   \cr   \tab   \cr
  \code{ggaes}    \tab \code{\link[ggplot2]{aes}}                                      \cr
  \code{ggexp}    \tab \code{\link[ggplot2]{expansion}}                                \cr
  \code{ggmar}    \tab \code{\link[ggplot2]{margin}}                                   \cr   \tab   \cr
  \code{ggunit}   \tab \code{\link[ggplot2]{unit}}                                     \cr
  \code{ggvars}   \tab \code{\link[ggplot2]{vars}}                                     \cr   \tab   \cr
  \code{ggdot}    \tab \code{\link[ggplot2]{geom_point}}                               \cr
  \code{gghlin}   \tab \code{\link[ggplot2]{geom_hline}}                               \cr
  \code{gglab}    \tab \code{\link[ggplot2]{geom_label}}                               \cr
  \code{gglin}    \tab \code{\link[ggplot2]{geom_line}}                                \cr
  \code{ggtxt}    \tab \code{\link[ggplot2]{geom_text}}                                \cr
  \code{ggvlin}   \tab \code{\link[ggplot2]{geom_vline}}                               \cr   \tab   \cr
  \code{gglabs}   \tab \code{\link[ggplot2]{labs}}                                     \cr
  \code{ggxlab}   \tab \code{\link[ggplot2]{xlab}}                                     \cr
  \code{ggylab}   \tab \code{\link[ggplot2]{ylab}}                                     \cr   \tab   \cr
  \code{gggrid}   \tab \code{\link[ggplot2]{facet_grid}}                               \cr
  \code{ggwrap}   \tab \code{\link[ggplot2]{facet_wrap}}                               \cr   \tab   \cr
  \code{ggthm}    \tab \code{\link[ggplot2]{theme}}                                    \cr
  \code{ggtmin}   \tab \code{\link[ggplot2]{theme_minimal}}                            \cr   \tab   \cr
  \code{ggsac}    \tab \code{\link[ggplot2]{scale_alpha_continuous}}                   \cr
  \code{ggscm}    \tab \code{\link[ggplot2]{scale_color_manual}}                       \cr
  \code{ggsdm}    \tab \code{\link[ggplot2]{scale_shape_manual}}                       \cr
  \code{ggslm}    \tab \code{\link[ggplot2]{scale_linetype_manual}}                    \cr
  \code{ggsxc}    \tab \code{\link[ggplot2]{scale_x_continuous}}                       \cr
  \code{ggsyc}    \tab \code{\link[ggplot2]{scale_y_continuous}}                       \cr   \tab   \cr
  \code{ggerct}   \tab \code{\link[ggplot2]{element_rect}}                             \cr
  \code{ggetxt}   \tab \code{\link[ggplot2]{element_text}}                             \cr
  \code{ggelin}   \tab \code{\link[ggplot2]{element_line}}                             \cr
  \code{ggemkd}   \tab \code{\link[ggtext:element_markdown]{ggtext::element_markdown}} \cr
  \code{ggenll}   \tab \code{\link[ggplot2]{element_blank}}                              }
}
\seealso{
Other plots: 
\code{\link{color}},
\code{\link{gr}},
\code{\link{markdown}()},
\code{\link{pals}},
\code{\link{sca}()},
\code{\link{u2u}()}

Other wraps: 
\code{\link{basics}()},
\code{\link{dp}},
\code{\link{dt}()},
\code{\link{gr}},
\code{\link{rd}},
\code{\link{sca}()},
\code{\link{tb}()}
}
\concept{plots}
\concept{wraps}
