% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.R
\encoding{UTF-8}
\name{gr}
\alias{gr}
\alias{st_h}
\alias{str_w}
\title{Thin wrappers of \code{graphics} functions.}
\usage{
st_h(s, units = "user", cex = NULL, font = NULL, vfont = NULL, ...)

str_w(s, units = "user", cex = NULL, font = NULL, vfont = NULL, ...)
}
\description{
Thin wrappers of \code{graphics} functions.
}
\details{
\tabular{ll}{  \code{str_h}   \tab \code{\link[graphics]{strheight}}. \cr
  \code{str_w}   \tab \code{\link[graphics]{strwidth}}.    }
}
\examples{
sent1 <- "First sentence."
sent2 <- "Second sentence."
sent12 <- paste0(sent1, "\n", sent2)
strh(sent1)
strh(sent2)
strh(sent12)
strw(sent1)
strw(sent2)
strw(sent12)
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other plots: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{markdown}()},
\code{\link{pals}()},
\code{\link{sca}()},
\code{\link{u2u}()}

Other wraps: 
\code{\link{basics}()},
\code{\link{dp}()},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{rd}()},
\code{\link{sca}()},
\code{\link{tb}()}
}
\concept{plots}
\concept{strings}
\concept{wraps}
