% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\name{hexdec.}
\alias{hexdec.}
\alias{todec}
\alias{tohex}
\title{Convert Non-Negative Whole Numbers between Decimal and Hexadecimal}
\usage{
hexdec.()

todec(x)

tohex(x)
}
\arguments{
\item{x}{A \link[=cmp_nnw]{complete non-negative whole-number object} or a
\link[=cmp_chr]{complete character vec} containing only non-negative
whole-number hexademical values.}
}
\value{
\tabular{ll}{
\code{todec}   \tab A \link[=cmp_nnw]{complete non-negative whole-number
object}                                             \cr
\code{tohex}   \tab A \link[=cmp_chr]{complete character object} containing
non-negative whole-number hexadecimal values in the
format \code{'#HHH...'} where \code{H} is a placeholder
for an uppercase hexadecimal digit.                   }
}
\description{
\tabular{ll}{
\code{todec}   \tab Converts hexadecimal \link[=cmp_chr]{complete character
object} \code{x} from hexadecimal to decimal. \code{x}
may be formatted as either \code{'hhh...'} or
\code{'#hhh...'} where \code{h} is a placeholder for a
hexidecimal digit in either upper or lower case.    \cr
\code{tohex}   \tab Converts \link[=cmp_psw]{complete positive
whole-number} \code{x} to hexadecimal formatted as
\code{'#HHH...'} where \code{H} is a placeholder for an
uppercase hexadecimal digit.                                   }
}
\seealso{
Other math: 
\code{\link{inc.}()},
\code{\link{stats.}()}
}
\concept{math}
