% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\name{hexdec}
\alias{hexdec}
\alias{todec}
\alias{tohex}
\title{Convert between integer and hexadecimal}
\usage{
todec(x)

tohex(x)
}
\arguments{
\item{x}{Vector of positive whole numbers or strings with hexademical number
representations.}
}
\description{
Convert from hexadecimal to decimal.
}
\details{
Hexademical representation is of mode character and may begin with
"#".
}
\section{Functions}{
\itemize{
\item \code{tohex}: Convert from decimal to hexadecimal.
}}

\seealso{
Other math: 
\code{\link{inc}()}
}
\concept{math}
