% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\name{hexdec}
\alias{hexdec}
\alias{todec}
\alias{tohex}
\title{Convert non-negative whole numbers between decimal and hexadecimal}
\usage{
todec(x)

tohex(x)
}
\arguments{
\item{x}{A \link[=cmp_nnw]{complete non-negative whole-number object} or a \link[=cmp_chr]{complete character vec} containing only non-negative whole-number hexademical values.}
}
\value{
\itemize{
\item \strong{\code{todec}}: a complete non-negative whole-number object.
\item \strong{\code{tohex}}: a complete character object containing non-negative whole-number hexadecimal values in the format \code{'#HHH...'} where \code{H} is a placeholder for an uppercase hexadecimal digit.
}
}
\description{
\itemize{
\item \strong{\code{todec}}: converts hexadecimal \link[=cmp_chr]{complete character object} \code{x} from hexadecimal to decimal. \code{x} may be formatted as either \code{'hhh...'} or \code{'#hhh...'} where \code{h} is a placeholder for a hexidecimal digit in either upper or lower case.
\item \strong{\code{tohex}}: converts \link[=cmp_psw]{complete positive whole-number} \code{x} to hexadecimal formatted as \code{'#HHH...'} where \code{H} is a placeholder for an uppercase hexadecimal digit.
}
}
