% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iii.R
\name{iii}
\alias{iii}
\alias{iii_props}
\alias{is_iii_spec}
\alias{iiii}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\title{Integrity (state of completeness) Properties}
\usage{
iii(x)

iii_props()

is_iii_spec(spec)

iiii(x, spec, ...)

icmp(x, ...)

imss(x, ...)

inas(x, ...)

ioks(x, ...)

iprt(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec}
containing one or more integrity properties (i.e., from \code{iii_props()}).
Integrity properties may be pipe-delimited. If there are multiple
properties in \code{spec}, \code{x} is inspected for a match to any of the specified
properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{ll}{
FUNCTIONS                       \tab RETURN VALUE                       \cr
\code{iii_vals}                      \tab A character vector.                \cr
\code{iii}                           \tab A character scalar or vector.      \cr
\code{iiii}, \code{ixxx}, \code{is_iii_spec}   \tab A logical scalar.                    }
}
\description{
Integrity properties are defined for \link[=ipop]{populated}
\link[=atm_dtf]{atomic data.frame}, populated \link[=atm_vls]{atomic
vlists}, populated atomic vectors, and populated atomic arrays. For all
others, all integrity properties are considered \code{FALSE}. The following
table summarizes valid integrity properties. \tabular{lll}{
NAME        \tab VALUE   \tab QUALIFYING OBJECTS                        \cr
Complete    \tab \code{'cmp'} \tab Populated and atomic vectors, arrays,
\link[=atm_dtf]{data.frames}, and
\link[=atm_vls]{vlists} containing
\strong{no} \code{NA} values.                  \cr
Missing     \tab \code{'mss'} \tab Populated and atomic vectors, arrays,
\link[=atm_dtf]{data.frames}, and
\link[=atm_vls]{vlists} containing
\strong{only} \code{NA} values.                \cr
Partial     \tab \code{'prt'} \tab Populated and atomic vectors, arrays,
\link[=atm_dtf]{data.frames}, and
\link[=atm_vls]{vlists} containing
\strong{both} \code{NA} and non-\code{NA} values.   \cr
NA scalar   \tab \code{'nas'} \tab An atomic scalar of value \code{NA}. May be a
vector of array.                          \cr
OK scalar   \tab \code{'oks'} \tab A non-\code{NA} atomic scalar of value. May be a
vector of array.                            }
Functions in this family are:\tabular{ll}{
FUNCTION   \tab WHAT IT DOES \cr
\code{iii}           \tab Gets a character vector containing all integrity
properties possessed by \code{x}.                       \cr
\code{ixxx}          \tab Evaluates whether \code{x} possesses the integrity property
\code{xxx} (a placeholder for any given integrity property
value) subject to any restrictions in \code{...}.       \cr
\code{iiii}          \tab Evaluates whether \code{x} possesses one or more  (possibly
pipe-delimited) integrity properties in \code{spec} subject
to any restrictions in \code{...}.                      \cr
\code{iii_props}     \tab Gets a character vector of all possible integrity
property values.                                   \cr
\code{is_iii_spec}   \tab Evaluates whether \code{spec} is a valid integrity property
specification.                            }
}
\section{Specifying Count and Value Restrictions}{
 Specifying restrictions in
\code{...} is optional. The full set of recognized arguments names are defined
in the following table along with the properties each
specifies:\tabular{ll}{
NAME     \tab WHAT IT SPECIFIES                                         \cr
\code{n}      \tab Vector of valid lengths/numbers of elements.              \cr
\code{nr}     \tab Vector of valid numbers of rows.                          \cr
\code{nc}     \tab Vector of valid numbers of columns.                       \cr
\code{min}    \tab Scalar minimum valid length/number of element.            \cr
\code{minr}   \tab Scalar minimum valid number of rows.                      \cr
\code{minc}   \tab Scalar minimum valid number of columns.                   \cr
\code{max}    \tab Scalar maximum valid length/number of element.            \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                      \cr
\code{maxc}   \tab Scalar maximum valid number of columns.                   \cr
\code{vals}   \tab Vector of valid values.                                   \cr
\code{lt}     \tab Scalar less-than (exclusive upper) bound.                 \cr
\code{le}     \tab Scalar less-than-or-equal (inclusive upper) bound         \cr
\code{ge}     \tab Scalar greater-than-or-equal (inclusive lower) bound.     \cr
\code{gt}     \tab Scalar greater-than bound (exclusive lower) bound.          }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
