% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iii.R
\name{iii}
\alias{iii}
\alias{iii_props}
\alias{is_iii_spec}
\alias{iiii}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\title{Integrity (state of completeness) Properties}
\usage{
iii(x)

iii_props()

is_iii_spec(spec)

iiii(x, spec, ...)

icmp(x, ...)

imss(x, ...)

inas(x, ...)

ioks(x, ...)

iprt(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec}
containing one or more integrity properties (i.e., from
\code{iii_vals()}). \strong{NOTE}: properties may be pipe-separated. If
If there are multiple properties in \code{spec}, \code{x} is inspected for
a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{\code{iii_vals}}: A character vector.
\cr\cr\strong{\code{iii}}: A character scalar or character vector.
\cr\cr\strong{\code{iIII, iiii, is_iii_spec}}: A logical scalar.
}
\description{
NOTE: \code{III} is used as a wildcard representing any given
integrity property.
\cr\cr
Integrity are defined for \link[=ipop]{populated} \link[=atm_dtf]{atomic
data.frame}, populated \link[=atm_vls]{atomic vlists}, populated atomic
vectors, and populated atomic arrays. For all others, all integrity
properties are considered \code{FALSE}. The following table summarizes
valid state properties.
\tabular{lll}{
PROPERTY    \tab PROPERTY   \tab QUALIFYING                           \cr
NAME        \tab VALUE      \tab OBJECTS                              \cr
Complete    \tab\code{'cmp'}\tab Populated and atomic vectors, arrays,
\link[=atm_dtf]{data.frames}, and
\link[=atm_vls]{vlists} containing
\strong{no} \code{NA} values.        \cr
Missing     \tab\code{'mss'}\tab Populated and atomic vectors, arrays,
\link[=atm_dtf]{data.frames}, and
\link[=atm_vls]{vlists} containing
\strong{only} \code{NA} values.      \cr
Partial     \tab\code{'prt'}\tab Populated and atomic vectors, arrays,
\link[=atm_dtf]{data.frames}, and
\link[=atm_vls]{vlists} containing
\strong{both} \code{NA} and
non-\code{NA} values.                \cr
NA scalar   \tab\code{'nas'}\tab An atomic scalar of value \code{NA}.
May be a vector of array.            \cr
OK scalar   \tab\code{'oks'}\tab An atomic scalar of value not equal to
\code{NA}. May be a vector of array.   }
Functions related to integrity properties are as follows:
\tabular{ll}{
III FUNCTION         \tab WHAT IT DOES                                \cr
\code{iIII}          \tab Evaluates whether \code{x} matches the
integrity property \code{III}
(subject to any restrictions in \code{...}).\cr
\code{iii}           \tab Gets a character vector containing all
integrity properties matching
\code{x}.                                   \cr
\code{iiii}          \tab Evaluates \code{x} against the
integrity property specification
in \code{spec} (subject to any restrictions in
\code{...}).                                \cr
\code{iii_props}     \tab Gets a character vector of all possible
integrity property values.                  \cr
\code{is_iii_spec}   \tab Evaluates whether \code{spec} is a valid
integrity property
specification.                                }
}
\section{Specifying Count and Value Restrictions}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar less-than bound on values of \code{x}.        \cr
\code{le}     \tab Scalar less-than-or-equal bound on values of
\code{x}.                                            \cr
\code{ge}     \tab Scalar greater-than-or-equal bound on values of
\code{x}.                                            \cr
\code{gt}     \tab Scalar greater-than bound on values of \code{x}.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc_props}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{meets}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
