% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iii.R
\name{iii}
\alias{iii}
\alias{iii_props}
\alias{is_iii_spec}
\alias{iiii}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\title{\code{Integrity} properties}
\usage{
iii(x)

iii_props()

is_iii_spec(spec)

iiii(x, spec, ...)

icmp(x, ...)

imss(x, ...)

inas(x, ...)

ioks(x, ...)

iprt(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{integrity} properties (i.e., from \code{iii_props()}). \code{Integrity} properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_iii_spec} \tab   A logical scalar.
\cr \code{iii_props} \tab   A character vector.
\cr      \code{iiii} \tab   A logical scalar.
\cr      \code{iIII} \tab   A logical scalar.
\cr       \code{iii} \tab   A character vector.
}
}
\description{
\code{Integrity} properties are defined for \link[=ipop]{populated} \link[=atm_dtf]{atomic data.frameS}, populated \link[=atm_vls]{atomic vlists}, populated atomic vectors, and populated atomic arrays. For all others, all integrity properties are considered \code{FALSE}. The following table summarizes valid integrity properties.
\tabular{rll}{
  \code{'cmp'}   \tab complete      \tab Populated and atomic vectors, arrays, \link[=atm_dtf]{data.frames}, and \link[=atm_vls]{vlists} containing \strong{no} \code{NA} values.
\cr           \tab               \tab   
\cr \code{'mss'}   \tab missing       \tab Populated and atomic vectors, arrays, data.frames, and vlists containing \strong{only} \code{NA} values.
\cr           \tab               \tab   
\cr \code{'prt'}   \tab partial       \tab Populated and atomic vectors, arrays, data.frames, and vlists containing \code{NA} \strong{and} non-\code{NA} values.
\cr           \tab               \tab   
\cr \code{'nas'}   \tab \code{NA} scalar   \tab Atomic scalar \code{NA}.
\cr \code{'oks'}   \tab \code{OK} scalar   \tab Non-\code{NA} atomic scalar.
}
\strong{\code{Integrity} property functions}
\tabular{rl}{
  \code{is_iii_spec}   \tab Is \code{spec} an \code{integrity} specification?
\cr               \tab   
\cr \code{iii_props}   \tab What \code{integrity} properties have dedicated functions?
\cr               \tab   
\cr      \code{iiii}   \tab Is \code{x} a match to the \code{integrity} specification \code{spec}?
\cr               \tab   
\cr      \code{iIII}   \tab Is \code{x} a match to the single \code{integrity} property \code{'III'}?
\cr               \tab   
\cr       \code{iii}   \tab Gets all of \code{x}'s \code{integrity} properties.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_iii_spec("nas|mss")
iii_props()
icmp(NA)
icmp(NULL)
inas(NA)
ioks(letters)
ioks(letters[1])
imss(rep(NA, 10))
iii(NA)
iii(letters)
iii(letters[1])
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{is_bbb_ccc_prop}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
