% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iii.R
\encoding{UTF-8}
\name{iii}
\alias{iii}
\alias{iii_funs}
\alias{iii_props}
\alias{is_iii_spec}
\alias{iiii}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\alias{idup}
\alias{iunq}
\title{Integrity properties}
\usage{
iii(x)

iii_funs()

iii_props()

is_iii_spec(spec)

iiii(x, spec, ...)

icmp(x, ...)

imss(x, ...)

inas(x, ...)

ioks(x, ...)

iprt(x, ...)

idup(x, ...)

iunq(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more integrity properties (i.e., from \code{iii_props()}). Integrity properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector} \cr   \verb{iii_props, iii_funs, iii}
\cr\cr \emph{A logical scalar} \cr   \verb{is_iii, iIII, iiii}
}
\description{
Integrity properties are defined for \link[=ipop]{populated} \link[=atm_dtf]{atomic data.frameS}, populated \link[=atm_vls]{atomic vlists}, populated atomic vectors, and populated atomic arrays. For all others, all integrity properties are considered \code{FALSE}. The following table summarizes valid integrity properties.
\tabular{rll}{
  \code{'cmp'}   \tab complete      \tab Populated and atomic vectors, arrays, \link[=atm_dtf]{data.frames}, and \link[=atm_vls]{vlists} containing \emph{no} \code{NA} values.
\cr \code{'mss'}   \tab missing       \tab Populated and atomic vectors, arrays, data.frames, and vlists containing \emph{only} \code{NA} values.
\cr \code{'prt'}   \tab partial       \tab Populated and atomic vectors, arrays, data.frames, and vlists containing \code{NA} \emph{and} non-\code{NA} values.
\cr \code{'dup'}   \tab duplicates    \tab Complete and containing duplicate values.
\cr \code{'unq'}   \tab unique        \tab Complete and containing only unique values.
\cr \code{'nas'}   \tab \code{NA} scalar   \tab Atomic scalar \code{NA}.
\cr \code{'oks'}   \tab \code{OK} scalar   \tab Non-\code{NA} atomic scalar.
}
\emph{Integrity property functions}
\tabular{rl}{
  \code{is_iii_spec}   \tab Is \code{spec} an integrity specification?
\cr \code{iii_props}   \tab What integrity properties are there?
\cr  \code{iii_funs}   \tab What integrity property functions are there?
\cr      \code{iiii}   \tab Does \code{x} match integrity specification \code{spec}?
\cr      \code{iIII}   \tab Does \code{x} match single integrity property \code{'III'}?
\cr       \code{iii}   \tab Gets all of \code{x}'s integrity properties.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
iii_funs()
iii_props()
is_iii_spec("nas|mss")
icmp(NA)
icmp(NULL)
inas(NA)
ioks(letters)
ioks(letters[1])
imss(rep(NA, 10))
iii(NA)
iii(letters)
iii(letters[1])
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
