% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iii.R
\name{iii}
\alias{iii}
\alias{iii_props}
\alias{is_iii_spec}
\alias{iiii}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\title{Integrity properties}
\usage{
iii(x)

iii_props()

is_iii_spec(spec)

iiii(x, spec, ...)

icmp(x, ...)

imss(x, ...)

inas(x, ...)

ioks(x, ...)

iprt(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more integrity properties (i.e., from \code{iii_props()}). Integrity properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_iii_spec}   \tab A logical
\cr      \code{iiii}   \tab scalar.
\cr      \code{iIII}   \tab   
\cr               \tab   
\cr \code{iii_props}   \tab A character
\cr       \code{iii}   \tab vector.
}
}
\description{
Integrity properties are defined for \link[=ipop]{populated} \link[=atm_dtf]{atomic data.frame}, populated \link[=atm_vls]{atomic vlists}, populated atomic vectors, and populated atomic arrays. For all others, all integrity properties are considered \code{FALSE}. The following table summarizes valid integrity properties.
\cr\cr \strong{Integrity properties} \tabular{rll}{
  \code{'cmp'}   \tab Complete    \tab Populated and atomic vectors, arrays, \link[=atm_dtf]{data.frames}, and \link[=atm_vls]{vlists} containing \strong{no} \code{NA} values.
\cr \code{'mss'}   \tab Missing     \tab   \code{...}containing \strong{only} \code{NA} values.
\cr \code{'prt'}   \tab Partial     \tab    \code{...}containing \strong{both} \code{NA} and non-\code{NA} values.
\cr           \tab             \tab  
\cr \code{'nas'}   \tab NA scalar   \tab Atomic scalar \code{NA}.
\cr \code{'oks'}   \tab OK scalar   \tab Non-\code{NA} atomic scalar.
}
\strong{Integrity property functions} \tabular{rl}{
  \code{is_iii_spec}   \tab Is \code{spec} an integrity property specification?
\cr \code{iii_props}   \tab Gets all possible integrity property values.
\cr      \code{iIII}   \tab Does \code{x} match integrity property \code{'III'}?
\cr      \code{iiii}   \tab Dpes \code{x} match integrity spec \code{spec}?
\cr       \code{iii}   \tab Gets all integrity properties possessed by \code{x}.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_iii_spec("nas|mss")
iii_props()
icmp(NA)
icmp(NULL)
inas(NA)
ioks(letters)
ioks(letters[1])
imss(rep(NA, 10))
iii(NA)
iii(letters)
iii(letters[1])
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
