% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iii.R
\encoding{UTF-8}
\name{iii}
\alias{iii}
\alias{iii_funs}
\alias{iii_props}
\alias{is_iii_spec}
\alias{III}
\alias{CMP}
\alias{MSS}
\alias{NA0}
\alias{OK0}
\alias{PRT}
\alias{DUP}
\alias{UNQ}
\title{Integrity properties}
\usage{
iii(x)

iii_funs()

iii_props()

is_iii_spec(spec)

III(x, spec, ...)

CMP(x, ...)

MSS(x, ...)

NA0(x, ...)

OK0(x, ...)

PRT(x, ...)

DUP(x, ...)

UNQ(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more integrity properties (i.e., from \code{iii_props()}). Integrity properties may be pipe-delimited. If there are multiple properties in \code{spec}, \code{x} is inspected for a match to any of the specified properties.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr   \verb{iii_props, iii_funs, iii}
\cr\cr  \strong{A logical scalar}   \cr\cr   \verb{is_iii_spec, III, \{III\}}
}
\description{
Integrity properties are defined for \link[=POP]{populated} \link[=atm_dtf]{atomic data.frames}, populated \link[=atm_vls]{atomic vlists}, populated atomic vectors, and populated atomic arrays. For all others, all integrity properties are considered \code{FALSE}. The following table summarizes valid integrity properties.
\tabular{lll}{  \verb{'cmp', 'CMP'}   \tab complete      \tab Populated and atomic vectors, arrays, \link[=atm_dtf]{data.frames}, and \link[=atm_vls]{vlists} containing \emph{no} \code{NA} values. \cr   \tab   \cr
  \verb{'mss', 'MSS'}   \tab missing       \tab Populated and atomic vectors, arrays, data.frames, and vlists containing \emph{only} \code{NA} values.                                 \cr   \tab   \cr
  \verb{'prt', 'PRT'}   \tab partial       \tab Populated and atomic vectors, arrays, data.frames, and vlists containing \code{NA} \emph{and} non-\code{NA} values.                         \cr   \tab   \cr
  \verb{'dup', 'DUP'}   \tab duplicates    \tab Complete and containing duplicate values.                                                                                    \cr   \tab   \cr
  \verb{'unq', 'UNQ'}   \tab unique        \tab Complete and containing only unique values.                                                                                  \cr   \tab   \cr
  \verb{'nas', 'NA0'}   \tab \code{NA} scalar   \tab Atomic scalar \code{NA}.                                                                                                          \cr   \tab   \cr
  \verb{'oks', 'OK0'}   \tab \code{OK} scalar   \tab Non-\code{NA} atomic scalar.                                                                                                                     }
\cr\cr \strong{Integrity property functions}
\tabular{ll}{  \code{is_iii_spec}   \tab Is \code{spec} an integrity specification?                                                                \cr   \tab   \cr
  \code{iii_props}     \tab What integrity properties are there?                                                                 \cr   \tab   \cr
  \code{iii_funs}      \tab What integrity property functions are there?                                                         \cr   \tab   \cr
  \code{iii}           \tab Gets all of \code{x}'s integrity properties.                                                              \cr   \tab   \cr
  \code{III}           \tab Does \code{x} match integrity specification \code{spec}?                                                       \cr   \tab   \cr
  \code{{II}}          \tab Does \code{x} match single integrity property \code{'{III}'} where \code{{III}} is a placeholder for any given integrity property? }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.MAX, .MAXR, .MAXC}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                 \cr   \tab   \cr
  \verb{.MIN, .MINR, .MINC}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                 \cr   \tab   \cr
  \verb{.LT, .LE, .GE, .GT}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.N, .NR, .NC}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.VALS}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
iii_funs()
iii_props()
is_iii_spec("nas|mss")
CMP(NA)
CMP(NULL)
NA0(NA)
OK0(letters)
OK0(letters[1])
MSS(rep(NA, 10))
iii(NA)
iii(letters)
iii(letters[1])
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp_fast}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
