% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R
\name{inc}
\alias{inc}
\alias{dec}
\title{Increment or decrement a variable}
\usage{
inc(x, i = 1, na = F)

dec(x, d = 1, na = F)
}
\arguments{
\item{x}{An \link[=atm_num]{atomic numeric} object.}

\item{i, d}{\link[=cmp_num_scl]{Complete numeric scalars} giving the value by which to increment or decrement, respectively.}

\item{na}{A non\code{NA} logical scalar indicating whether to throw an error if any values of \code{x} are \code{NA}.}
}
\value{
An atomic numeric object of the same dimension as \code{x}.
}
\description{
\itemize{
\item \strong{\code{inc}}: increments \code{x} by \code{i}.
\item \strong{\code{dec}}: decrements \code{x} by \code{d}.
}
}
