% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R
\name{inc}
\alias{inc}
\title{Increment or Decrement a Variable}
\usage{
inc(x, i = 1, na = F)
}
\arguments{
\item{x}{Any atomic numeric object.}

\item{i}{A numeric scalar giving the value by which to increment.}

\item{na}{A logical scalar indicating whether to throw an error if any values
of \code{x} are \code{NA}.}
}
\value{
\code{x} or \code{list(...)}, but with each element incremented or
decremented.
}
\description{
Increment or Decrement a Variable
}
\seealso{
Other math: 
\code{\link{hexdec}}
}
\concept{math}
