% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc..R
\name{inc}
\alias{inc}
\alias{dec}
\title{Increment or decrement a variable}
\usage{
inc(x, i = 1, na = F)

dec(x, d = 1, na = F)
}
\arguments{
\item{x}{An \link[=atm_num]{atomic numeric} object.}

\item{i, d}{\link[=cmp_num_scl]{Complete numeric scalars} giving the value by which to increment or decrement, respectively.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values are allowed.}
}
\value{
An atomic numeric object of the same dimension as \code{x}.
}
\description{
\tabular{rl}{
\code{inc}   \tab Increments \code{x} by \code{i}.
\cr \code{dec}   \tab Decrements \code{x} by \code{d}.
}
}
\examples{
inc(pi, pi)
inc(1:5, 1)
inc(c(1:5, NA), -1, na = T)

dec(pi, pi)
dec(1:5, 1)
dec(c(1:5, NA), -1, na = T)
}
