% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipat.R
\name{ipat}
\alias{ipat}
\alias{has_pat}
\alias{wpat}
\alias{npat}
\alias{ppat}
\alias{ppats}
\title{Simplified string indexing for fixed patterns}
\usage{
ipat(x, pat)

has_pat(x, pat)

wpat(x, pat)

npat(x, pat)

ppat(x, pat)

ppats(x, pat)
}
\arguments{
\item{x}{A \link[=chr_vec]{character vector}.}

\item{pat}{A fixed value \link[=cmp_chr_scl]{complete character scalar} pattern to search for in the atomic vector resulting from collapsing all elements of every argument supplied via \code{...}.}
}
\value{
\itemize{
\item \strong{\code{ppats}}: a \link[=whl_vls]{whole-number vlist}.
\item \strong{\verb{has_pat, ipat}}: a logical vector.
\item \strong{\verb{wpat, npat, ppat}}: A \link[=whl_vec]{whole-number vector}.
}
}
\description{
\itemize{
\item \strong{\code{ipat}}: Logically indexes character vector elements containing a pattern as \code{TRUE} or \code{FALSE}.
\item \strong{\code{wpat}}: Numerically indexes character vector elements containing a pattern as integer element numbers.
\item \strong{\code{npat}}: Counts the number of times a pattern appears in each element of a character vector.
\item \strong{\code{ppat}}: Identifies the position of \emph{the first} character of the first instance of a pattern in each element of a character vector.
\item \strong{\code{ppats}}: Identifies the position of the first character of \emph{every} instance of a pattern in each element of a character vector.
\item \strong{\code{has_pat}}: Evaluate whether each element of \code{x} contains the string in \code{pat}.
}
}
\details{
Always \link[=a]{atomizes} \code{...} to create a single character vector to search for a fixed pattern.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{max_nch}()},
\code{\link{mkd}},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
