% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipat.R
\name{ipat}
\alias{ipat}
\alias{has_pat}
\alias{wpat}
\alias{npat}
\alias{ppat}
\alias{ppats}
\title{Simplified string indexing for fixed patterns}
\usage{
ipat(x, pat)

has_pat(x, pat)

wpat(x, pat)

npat(x, pat)

ppat(x, pat)

ppats(x, pat)
}
\arguments{
\item{x}{\link[=chr_vec]{Character vector}.}

\item{pat}{A fixed (not regular expression) \link[=cmp_chr_scl]{complete
character scalar} pattern to search for in the atomic vector resulting from
collapsing all elements of every argument supplied via \code{...}.}
}
\value{
\strong{\code{has_pat, ipat}}
\cr A logical vector.
\cr\cr
\strong{\code{wpat, npat, ppat}}
\cr A \link[=whl_vec]{whole-number vector}.
\cr\cr
\strong{\code{ppats}}
\cr A \link[=whl_vls]{whole-number vlist}.
}
\description{
Simplified string indexing for fixed patterns
}
\details{
Always \link[=a]{atomizes} \code{...} via \code{av(...)} to create a
single character vector to search for a fixed pattern.
}
\section{Functions in This Family}{

\strong{\code{ipat}}
\cr Logically indexes character vector elements containing a pattern as
\code{TRUE} or \code{FALSE}.
\cr\cr
\strong{\code{has_pat}}
\cr Evaluate whether each element of \code{x} contains the string in
\code{pat}.
\cr\cr
\strong{\code{wpat}}
\cr Numerically indexes character vector elements containing a pattern as
integer element numbers.
\cr\cr
\strong{\code{npat}}
\cr Counts the number of times a pattern appears in each element of a
character vector.
\cr\cr
\strong{\code{ppat}}
\cr Identifies the position of \emph{the first} character of the first
instance of a pattern in each element of a character vector.
\cr\cr
\strong{\code{ppats}}
\cr Identifies the position of the first character of
\emph{every} instance of a pattern in each element of a character vector.
}

\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{charn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{max_nch}()},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
