% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipat.R
\name{ipat}
\alias{ipat}
\alias{has_pat}
\alias{wpat}
\alias{npat}
\alias{ppat}
\alias{ppats}
\title{Simplified string indexing for fixed patterns}
\usage{
ipat(x, pat)

has_pat(x, pat)

wpat(x, pat)

npat(x, pat)

ppat(x, pat)

ppats(x, pat)
}
\arguments{
\item{x}{A \link[=chr_vec]{character vector}.}

\item{pat}{A fixed value \link[=cmp_str_scl]{complete string scalar} pattern to search for in \code{x}.}
}
\value{
\tabular{rl}{
\code{has_pat} \tab   A logical scalar.
\cr \code{ppats} \tab   A positive whole number vector.
\cr  \code{wpat} \tab   A positive whole number vector.
\cr  \code{npat} \tab   A positive whole number vector.
\cr  \code{ppat} \tab   A positive whole number vector.
\cr  \code{ipat} \tab   A logical scalar.
}
}
\description{
\tabular{rl}{
\code{has_pat}   \tab Evaluates whether each element of \code{x} contains the string in \code{pat}.
\cr \code{ppats}   \tab Identifies the position of the first character of \emph{every} instance of a pattern in each element of a character vector.
\cr  \code{ppat}   \tab Identifies the position of \emph{the first} character of the first instance of a pattern in each element of a character vector.
\cr  \code{npat}   \tab Counts the number of times a pattern appears in each element of a character vector.
\cr  \code{ipat}   \tab Logically indexes character vector elements containing a pattern as \code{TRUE} or \code{FALSE}.
\cr  \code{wpat}   \tab Numerically indexes character vector elements containing a pattern as integer element numbers.
}
}
\details{
Always \link[=a]{atomizes} \code{...} to create a single character vector to search for a fixed pattern.
}
\examples{
words <- c("apple", "banana", "carrot")

has_pat(words, "a")
ipat(words, "a")

has_pat(words, "b")
ipat(words, "b")

wpat(words, "a")
wpat(words, "b")

npat(words, "a")
npat(words, "b")

ppat(words, "a")
ppat(words, "b")

ppats(words, "a")
ppats(words, "b")
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{max_nch}()},
\code{\link{mkd}},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
