% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unq}
\alias{is_unq}
\title{Evaluate whether an object has only unique atomic values}
\usage{
is_unq(x, a = T, na = F)
}
\arguments{
\item{x}{An atomic object.}

\item{a}{A non-\code{NA} logical scalar indicating whether to reduce \code{x} to an atomic vector containing all of its atomic values. When \code{FALSE} and \code{x} is not atomic, throws an error.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values are allowed.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Evaluates whether \code{x} contains only only unique atomic values, with options for for atomizing before evaluation and for handling \code{NA} values.
}
