% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make.}
\alias{make.}
\alias{vec}
\alias{vec_na}
\alias{mat}
\alias{dtf0}
\alias{dtf_na}
\alias{dtf}
\alias{vls}
\alias{mkvls}
\alias{mkdiag}
\title{Extended object creation functionality}
\usage{
make.()

vec(..., r. = 1, vn. = NULL)

vec_na(r.)

mat(..., nr. = 1, nc. = NULL, br. = F, rn. = NULL, cn. = NULL)

dtf0(cn.)

dtf_na(cn., nr.)

dtf(..., cn. = NULL)

vls(..., en. = NULL)

mkvls(...)

mkdiag(x. = 1, n. = 1)
}
\arguments{
\item{...}{Objects to placed in an atomic vect, atomic matrix, atomic tibble,
vlist, or square atomic diagonal matrix.}

\item{r.}{Number of replications.}

\item{vn.}{Optional character vector of names to apply to vector or list
elements. May be pipe-delimited in character vector.}

\item{nr.}{Number of rows.}

\item{nc.}{Number of columns.}

\item{br.}{Whether to fill matrices by row.}

\item{rn.}{Optional character vector of names to apply to rows of a matrix or
tibble. May be pipe-delimited in character vector.}

\item{cn.}{Optional character vector of names to apply to columns of a matrix
or tibble. May be pipe-delimited in character vector. For tibbles, this
argument can be used only if the arguments in \code{...} are not named.}

\item{x.}{An numeric, logical, or character object of length 1 or greater.}

\item{n.}{A positive whole number indicating the number of replications of
\code{x}.}
}
\value{
An \link[=is_atm_tibble]{atomic tibble}, an
\link[=is_atm_vect]{atomic vect}, an \link[=is_atm_matrix]{atomic matrix}
(\code{mat.}), or a \link[=is_vlist]{vlist.} All others return the value of
function they are thin wrappers for.
}
\description{
Extended object creation functionality
}
\section{Functions}{
\itemize{
\item \code{vec()}: Creates an atomic vector from the atomic objects in
\code{...} collapsed into a vector, replicated \code{r.} times, with
optional element names.

\item \code{vec_na()}: Creates a vector of \code{r.} \code{NA} values.

\item \code{mat()}: Creates an atomic matrix from the atomic elements in
\code{...} with \code{r.} rows, \code{c.} columns, optionally filling by
row, with optional row and column names.

\item \code{dtf0()}: Creates an  with 0 rows and columns
with names from \code{cn.}, which may be pipe delimited.

\item \code{dtf_na()}: Creates an \link[idtf]{atomic dtf} with \code{nr.} rows and
columns with names from \code{cn.} (which may be pipe delimited). All cells
of the resulting tibble are populated with \code{NA}. values.

\item \code{dtf()}: Creates an \link[idtf]{atomic dtf} from the
\link[ivec]{atomic vec} arguments in \code{...} with optional row names,
plus optional column names to replace any names of arguments in \code{...}.

\item \code{vls()}: Creates a \link[is_vls]{vlist} from the arguments in
\code{...} with optional element names to replace any names of arguments in
\code{...}.

\item \code{mkvls()}: Creates a named vlist by taking the original call and
parsing arguments to get element names. Allows for more concise code such
as \code{mkvls(letters, LETTERS)} giving the same result as
\code{list(letters = letters, LETTERS = LETTERS)}.

\item \code{mkdiag()}: Creates a square atomic diagonal matrix. For numeric
\code{x.}, off-diagonals are \code{0}. For logical \code{x.}, off diagonals
are \code{FALSE}. For character \code{x.}, off diagonals are blank strings
(\code{""}). For all others, off diagonals are \code{NA}.

}}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{put.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
