% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_uj}
\alias{make_uj}
\alias{vec}
\alias{vec_na}
\alias{mat}
\alias{atb0}
\alias{atb_na}
\alias{atb}
\alias{vls}
\alias{mkls}
\alias{mkdiag}
\title{Extended object creation functionality}
\usage{
make_uj()

vec(..., r. = 1, vn. = NULL)

vec_na(r.)

mat(..., nr. = 1, nc. = NULL, br. = F, rn. = NULL, cn. = NULL)

atb0(cn.)

atb_na(cn., nr.)

atb(..., cn. = NULL)

vls(..., en. = NULL)

mkls(...)

mkdiag(x. = 1, n. = 1)
}
\arguments{
\item{...}{Objects to placed in an atomic vect, atomic matrix, atomic tibble,
vlist, or square atomic diagonal matrix.}

\item{r.}{Number of replications.}

\item{vn.}{Optional character vector of names to apply to vector or list
elements. May be pipe-delimited in character vector.}

\item{nr.}{Number of rows.}

\item{nc.}{Number of columns.}

\item{br.}{Whether to fill matrices by row.}

\item{rn.}{Optional character vector of names to apply to rows of a matrix or
tibble. May be pipe-delimited in character vector.}

\item{cn.}{Optional character vector of names to apply to columns of a matrix
or tibble. May be pipe-delimited in character vector. For tibbles, this
argument can be used only if the arguments in \code{...} are not named.}

\item{x.}{An numeric, logical, or character object of length 1 or greater.}

\item{n.}{A positive whole number indicating the number of replications of
\code{x}.}
}
\value{
An \link[=is_atm_tibble]{atomic tibble}, an
\link[=is_atm_vect]{atomic vect}, an \link[=is_atm_matrix]{atomic matrix}
(\code{mat.}), or a \link[=is_vlist]{vlist.} All others return the value of
function they are thin wrappers for.
}
\description{
\strong{\code{vec(..., r, en.)}}
\cr Creates an atomic vector from the atomic objects in \code{...}
collapsed into a vector, replicated \code{r.} times, with optional element
names.
\cr\cr
\strong{\code{vec_na(r.)}}
\cr Creates a vector of \code{r.} \code{NA} values.
\cr\cr
\strong{\code{mat(..., nr., nc., br., rn., cn.)}}
\cr Creates an atomic matrix from the atomic elements in \code{...} with
\code{r.} rows, \code{c.} columns, optionally filling by row, with optional
row and column names.
\cr\cr
\strong{\code{atb0(cn.)}}
\cr Creates an \link[=is_atm_tibble]{atomic tibble} with 0 rows and columns
with names from \code{cn.}, which may be pipe delimited.
\cr\cr
\strong{\code{atb_na(cn., nr.)}}
\cr Creates an \link[=is_atm_tibble]{atomic tibble} with \code{nr.} rows
and columns with names from \code{cn.} (which may be pipe delimited). All
cells of the resulting tibble are populated with \code{NA}.
\cr\cr
\strong{\code{atb(..., rn., cn.)}}
\cr Creates an \link[=is_atm_tibble]{atomic tibble} from the atomic vect
arguments in \code{...} with optional row names, plus optional column names
to replace any names of arguments in \code{...}.
\cr\cr
\strong{\code{vls(..., en. = NULL)}}
\cr Creates a \link[=is_vlist]{vlist} from the arguments in \code{...} with
optional element names to replace any names of arguments in \code{...}.
\cr\cr
\strong{\code{mkvls(...)}}
\cr Creates a named vlist by taking the original call and parsing arguments
to get element names. Allows for more concise code such as
\code{mkvls(letters, LETTERS)} giving the same result as \code{list(letters
  = letters, LETTERS = LETTERS)}.
\cr\cr
\strong{\code{mkdiag(x. = 1, n. = 1)}}
\cr Creates a square, atomic diagonal matrix. For numeric \code{x.},
off-diagonals are \code{0}. For logical \code{x.}, off diagonals are
\code{FALSE}. For character \code{x.}, off diagonals are blank strings
(\code{""}). For all others, off diagonals are \code{NA}.
}
\section{Functions}{
\itemize{
\item \code{vec}: Create an atomic vector.

\item \code{vec_na}: Create a vector of missing/\code{NA} values.

\item \code{mat}: Create an atomic matrix.

\item \code{atb0}: Create an atomic tibble with zero rows.

\item \code{atb_na}: Create an atomic tibble containing only missing values.

\item \code{atb}: Create an atomic tibble from named arguments or from
column names in \code{cn.}.

\item \code{vls}: Create a named vlist from named arguments or from unnamed
arguments and element names provided in \code{en.}.

\item \code{mkls}: Make a list where element names are derived from the
original function call. For example, \code{mkls(letters, LETTERS)} results
in the same output as \code{list(letters = letters, LETTERS = LETTERS)}

\item \code{mkdiag}: Create a diagonal matrix.
}}

\seealso{
Other meta: 
\code{\link{av_uj}()},
\code{\link{bin_uj}()},
\code{\link{callers_uj}()},
\code{\link{case_uj}()},
\code{\link{dots_uj}()},
\code{\link{ex_uj}()},
\code{\link{failsafe_uj}()},
\code{\link{fork_uj.}()},
\code{\link{fx_uj}()},
\code{\link{is_failsafe_uj}()},
\code{\link{is_unique_uj}()},
\code{\link{logicals_uj}()},
\code{\link{n_uj}},
\code{\link{na_vals_uj}()},
\code{\link{names_uj}()},
\code{\link{os_uj}()},
\code{\link{ply_uj}()},
\code{\link{put_uj}()},
\code{\link{reclass_uj}()},
\code{\link{recycling_uj}()},
\code{\link{reps_uj}()},
\code{\link{run_uj}()},
\code{\link{swap_uj}()},
\code{\link{values_uj}()},
\code{\link{x_uj}()},
\code{\link{xb_uj}()}
}
\concept{meta}
