% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\encoding{UTF-8}
\name{markdown}
\alias{markdown}
\alias{subs}
\alias{sups}
\alias{bfblu}
\alias{blblu}
\alias{bmblu}
\alias{bdblu}
\alias{bfcyn}
\alias{blcyn}
\alias{bmcyn}
\alias{bdcyn}
\alias{bfgrn}
\alias{blgrn}
\alias{bmgrn}
\alias{bdgrn}
\alias{bfmag}
\alias{blmag}
\alias{bmmag}
\alias{bdmag}
\alias{bforn}
\alias{blorn}
\alias{bmorn}
\alias{bdorn}
\alias{bfred}
\alias{blred}
\alias{bmred}
\alias{bdred}
\alias{bfvlt}
\alias{blvlt}
\alias{bmvlt}
\alias{bdvlt}
\alias{bfylw}
\alias{blylw}
\alias{bmylw}
\alias{bdylw}
\alias{bwht}
\alias{bg95}
\alias{bg90}
\alias{bg85}
\alias{bg80}
\alias{bg75}
\alias{bg70}
\alias{bg65}
\alias{bg60}
\alias{bg55}
\alias{bg50}
\alias{bg45}
\alias{bg40}
\alias{bg35}
\alias{bg30}
\alias{bg25}
\alias{bg20}
\alias{bg15}
\alias{bg10}
\alias{bg05}
\alias{bblk}
\alias{binv}
\alias{pfblu}
\alias{plblu}
\alias{pmblu}
\alias{pdblu}
\alias{pfcyn}
\alias{plcyn}
\alias{pmcyn}
\alias{pdcyn}
\alias{pfgrn}
\alias{plgrn}
\alias{pmgrn}
\alias{pdgrn}
\alias{pfmag}
\alias{plmag}
\alias{pmmag}
\alias{pdmag}
\alias{pforn}
\alias{plorn}
\alias{pmorn}
\alias{pdorn}
\alias{pfred}
\alias{plred}
\alias{pmred}
\alias{pdred}
\alias{pfvlt}
\alias{plvlt}
\alias{pmvlt}
\alias{pdvlt}
\alias{pfylw}
\alias{plylw}
\alias{pmylw}
\alias{pdylw}
\alias{pwht}
\alias{pg95}
\alias{pg90}
\alias{pg85}
\alias{pg80}
\alias{pg75}
\alias{pg70}
\alias{pg65}
\alias{pg60}
\alias{pg55}
\alias{pg50}
\alias{pg45}
\alias{pg40}
\alias{pg35}
\alias{pg30}
\alias{pg25}
\alias{pg20}
\alias{pg15}
\alias{pg10}
\alias{pg05}
\alias{pblk}
\alias{pinv}
\title{Build markdown expressions for styled + colored \code{ggtext} strings}
\usage{
subs(x)

sups(x)

bfblu(x)

blblu(x)

bmblu(x)

bdblu(x)

bfcyn(x)

blcyn(x)

bmcyn(x)

bdcyn(x)

bfgrn(x)

blgrn(x)

bmgrn(x)

bdgrn(x)

bfmag(x)

blmag(x)

bmmag(x)

bdmag(x)

bforn(x)

blorn(x)

bmorn(x)

bdorn(x)

bfred(x)

blred(x)

bmred(x)

bdred(x)

bfvlt(x)

blvlt(x)

bmvlt(x)

bdvlt(x)

bfylw(x)

blylw(x)

bmylw(x)

bdylw(x)

bwht(x)

bg95(x)

bg90(x)

bg85(x)

bg80(x)

bg75(x)

bg70(x)

bg65(x)

bg60(x)

bg55(x)

bg50(x)

bg45(x)

bg40(x)

bg35(x)

bg30(x)

bg25(x)

bg20(x)

bg15(x)

bg10(x)

bg05(x)

bblk(x)

binv(x)

pfblu(x)

plblu(x)

pmblu(x)

pdblu(x)

pfcyn(x)

plcyn(x)

pmcyn(x)

pdcyn(x)

pfgrn(x)

plgrn(x)

pmgrn(x)

pdgrn(x)

pfmag(x)

plmag(x)

pmmag(x)

pdmag(x)

pforn(x)

plorn(x)

pmorn(x)

pdorn(x)

pfred(x)

plred(x)

pmred(x)

pdred(x)

pfvlt(x)

plvlt(x)

pmvlt(x)

pdvlt(x)

pfylw(x)

plylw(x)

pmylw(x)

pdylw(x)

pwht(x)

pg95(x)

pg90(x)

pg85(x)

pg80(x)

pg75(x)

pg70(x)

pg65(x)

pg60(x)

pg55(x)

pg50(x)

pg45(x)

pg40(x)

pg35(x)

pg30(x)

pg25(x)

pg20(x)

pg15(x)

pg10(x)

pg05(x)

pblk(x)

pinv(x)
}
\arguments{
\item{x}{A \link[=cmp_chr_vec]{character vector} of text to be styled.}
}
\value{
A markdown-styled character vector.
}
\description{
\emph{Subscript/superscript} functions are: \tabular{rl}{
  \code{subs} \tab   subscript
\cr \code{sups} \tab   superscript
}
\cr \emph{All remaining functions} in this family start with a prefix character as follows: \tabular{rl}{
  prefix = \code{b} \tab   bold
\cr prefix = \code{p} \tab   plain
}
\cr \emph{Black, white, and invisible text} functions suffixes are: \tabular{rl}{
  \code{blk} \tab   black
\cr \code{wht} \tab   white
\cr \code{inv} \tab   invisible
}
For example, the function \code{bwht} styles markdown text as bold and white.
\cr\cr \emph{Greyscale} text suffixes take the form: \tabular{rl}{
  \code{g05} \tab   \code{ 5:95} white:black ratio
\cr \code{g10} \tab   \code{10:90} white:black ratio
\cr \code{...} \tab   \code{...}
\cr \code{g90} \tab   \code{90:10} white:black ratio
\cr \code{g95} \tab   \code{95:5 } white:black ratio
}
For example, the function \code{pg45} styles markdown text as plain and grey with a \code{45:55} white:black ratio.
\cr\cr \emph{Other common-color text} function names append an infix and a suffix to the selected prefix, as follows:
\tabular{rl}{
  suffix = \code{blu} \tab   blue
\cr suffix = \code{cyn} \tab   cyan
\cr suffix = \code{grn} \tab   green
\cr suffix = \code{mag} \tab   magenta
\cr suffix = \code{orn} \tab   orange
\cr suffix = \code{red} \tab   red
\cr suffix = \code{vlt} \tab   violet
\cr suffix = \code{ylw} \tab   yellow
\cr                \tab  
\cr    infix = \code{f} \tab   full-intensity
\cr    infix = \code{l} \tab   light
\cr    infix = \code{m} \tab   medium
\cr    infix = \code{d} \tab   dark
}
For example, the function \code{pfred} styles markdown text as plain and full-intensity red.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other colors: 
\code{\link{sca}()}

Other plots: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{pals}},
\code{\link{sca}()},
\code{\link{u2u}()}
}
\concept{colors}
\concept{plots}
\concept{strings}
