% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\encoding{UTF-8}
\name{markdown}
\alias{markdown}
\alias{subs}
\alias{sups}
\alias{bf_blu}
\alias{bl_blu}
\alias{bm_blu}
\alias{bd_blu}
\alias{bf_cyn}
\alias{bl_cyn}
\alias{bm_cyn}
\alias{bd_cyn}
\alias{bf_grn}
\alias{bl_grn}
\alias{bm_grn}
\alias{bd_grn}
\alias{bf_mag}
\alias{bl_mag}
\alias{bm_mag}
\alias{bd_mag}
\alias{bf_orn}
\alias{bl_orn}
\alias{bm_orn}
\alias{bd_orn}
\alias{bf_red}
\alias{bl_red}
\alias{bm_red}
\alias{bd_red}
\alias{bf_vlt}
\alias{bl_vlt}
\alias{bm_vlt}
\alias{bd_vlt}
\alias{bf_ylw}
\alias{bl_ylw}
\alias{bm_ylw}
\alias{bd_ylw}
\alias{b_wht}
\alias{b_g95}
\alias{b_g90}
\alias{b_g85}
\alias{b_g80}
\alias{b_g75}
\alias{b_g70}
\alias{b_g65}
\alias{b_g60}
\alias{b_g55}
\alias{b_g50}
\alias{b_g45}
\alias{b_g40}
\alias{b_g35}
\alias{bg_30}
\alias{bg_25}
\alias{bg_20}
\alias{bg_15}
\alias{bg_10}
\alias{bg_05}
\alias{b_blk}
\alias{b_inv}
\alias{pf_blu}
\alias{pl_blu}
\alias{pm_blu}
\alias{pd_blu}
\alias{pf_cyn}
\alias{pl_cyn}
\alias{pm_cyn}
\alias{pd_cyn}
\alias{pf_grn}
\alias{pl_grn}
\alias{pm_grn}
\alias{pd_grn}
\alias{pf_mag}
\alias{pl_mag}
\alias{pm_mag}
\alias{pd_mag}
\alias{pf_orn}
\alias{pl_orn}
\alias{pm_orn}
\alias{pd_orn}
\alias{pf_red}
\alias{pl_red}
\alias{pm_red}
\alias{pd_red}
\alias{pf_vlt}
\alias{pl_vlt}
\alias{pm_vlt}
\alias{pd_vlt}
\alias{pf_ylw}
\alias{pl_ylw}
\alias{pm_ylw}
\alias{pd_ylw}
\alias{p_wht}
\alias{p_g95}
\alias{p_g90}
\alias{p_g85}
\alias{p_g80}
\alias{p_g75}
\alias{p_g70}
\alias{p_g65}
\alias{p_g60}
\alias{p_g55}
\alias{p_g50}
\alias{p_g45}
\alias{p_g40}
\alias{p_g35}
\alias{p_g30}
\alias{p_g25}
\alias{p_g20}
\alias{p_g15}
\alias{p_g10}
\alias{p_g05}
\alias{p_blk}
\alias{p_inv}
\title{Build markdown expressions for styled + colored \code{ggtext} strings}
\usage{
markdown()

subs(x)

sups(x)

bf_blu(x)

bl_blu(x)

bm_blu(x)

bd_blu(x)

bf_cyn(x)

bl_cyn(x)

bm_cyn(x)

bd_cyn(x)

bf_grn(x)

bl_grn(x)

bm_grn(x)

bd_grn(x)

bf_mag(x)

bl_mag(x)

bm_mag(x)

bd_mag(x)

bf_orn(x)

bl_orn(x)

bm_orn(x)

bd_orn(x)

bf_red(x)

bl_red(x)

bm_red(x)

bd_red(x)

bf_vlt(x)

bl_vlt(x)

bm_vlt(x)

bd_vlt(x)

bf_ylw(x)

bl_ylw(x)

bm_ylw(x)

bd_ylw(x)

b_wht(x)

b_g95(x)

b_g90(x)

b_g85(x)

b_g80(x)

b_g75(x)

b_g70(x)

b_g65(x)

b_g60(x)

b_g55(x)

b_g50(x)

b_g45(x)

b_g40(x)

b_g35(x)

bg_30(x)

bg_25(x)

bg_20(x)

bg_15(x)

bg_10(x)

bg_05(x)

b_blk(x)

b_inv(x)

pf_blu(x)

pl_blu(x)

pm_blu(x)

pd_blu(x)

pf_cyn(x)

pl_cyn(x)

pm_cyn(x)

pd_cyn(x)

pf_grn(x)

pl_grn(x)

pm_grn(x)

pd_grn(x)

pf_mag(x)

pl_mag(x)

pm_mag(x)

pd_mag(x)

pf_orn(x)

pl_orn(x)

pm_orn(x)

pd_orn(x)

pf_red(x)

pl_red(x)

pm_red(x)

pd_red(x)

pf_vlt(x)

pl_vlt(x)

pm_vlt(x)

pd_vlt(x)

pf_ylw(x)

pl_ylw(x)

pm_ylw(x)

pd_ylw(x)

p_wht(x)

p_g95(x)

p_g90(x)

p_g85(x)

p_g80(x)

p_g75(x)

p_g70(x)

p_g65(x)

p_g60(x)

p_g55(x)

p_g50(x)

p_g45(x)

p_g40(x)

p_g35(x)

p_g30(x)

p_g25(x)

p_g20(x)

p_g15(x)

p_g10(x)

p_g05(x)

p_blk(x)

p_inv(x)
}
\arguments{
\item{x}{A \link[=cmp_chr_vec]{character vector} of text to be styled.}
}
\value{
A markdown-styled character vector.
}
\description{
Style markdown text with subscripts, superscripts, and colors.
}
\details{
\strong{Subscript/superscript} functions are:
\tabular{ll}{  \code{subs}   \tab subscript   \cr
  \code{sups}   \tab superscript   }
\cr\cr \strong{All remaining functions} in this family start with a prefix character as follows:
\tabular{ll}{  prefix = \code{b}   \tab bold  \cr
  prefix = \code{p}   \tab plain   }
\cr\cr \strong{Black, white, and invisible text} functions suffixes are:
\tabular{ll}{  \code{blk}   \tab black        \cr
  \code{wht}   \tab white        \cr
  \code{inv}   \tab invisible      }
\cr For example, the function \code{b_wht} styles markdown text as bold and white.
\cr\cr \strong{Greyscale} text suffixes take the form:
\tabular{ll}{  \code{g05}   \tab \code{ 5:95} white:black ratio \cr
  \code{g10}   \tab \code{10:90} white:black ratio \cr
  \code{...}   \tab \code{...}                     \cr
  \code{g90}   \tab \code{90:10} white:black ratio \cr
  \code{g95}   \tab \code{95:5 } white:black ratio   }
\cr For example, the function \code{p_g45} styles markdown text as plain and grey with a \code{45:55} white:black ratio.
\cr\cr \strong{Other common-color text} function names append an infix and a suffix to the selected prefix, as follows:
\tabular{ll}{  infix = \code{f}      \tab full intensity    \cr
  infix = \code{l}      \tab light             \cr
  infix = \code{m}      \tab medium            \cr
  infix = \code{d}      \tab dark \cr   \tab   \cr
  suffix = \code{blu}   \tab blue              \cr
  suffix = \code{cyn}   \tab cyan              \cr
  suffix = \code{grn}   \tab green             \cr
  suffix = \code{mag}   \tab magenta           \cr
  suffix = \code{orn}   \tab orange            \cr
  suffix = \code{red}   \tab red               \cr
  suffix = \code{vlt}   \tab violet            \cr
  suffix = \code{ylw}   \tab yellow              }

\cr For example, the function \code{pf_red} styles markdown text as plain and full-intensity red.
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other colors: 
\code{\link{sca}()}

Other plots: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{pals}},
\code{\link{sca}()},
\code{\link{u2u}()}
}
\concept{colors}
\concept{plots}
\concept{strings}
