% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnch.R
\name{max_nch.}
\alias{max_nch.}
\alias{max_nch}
\title{Length of longest string}
\usage{
max_nch.()

max_nch(x)
}
\arguments{
\item{x}{\link[=atm_chr]{Atomic character object}.}
}
\value{
Integer scalar.
}
\description{
Gets the number of characters in the longest string in
\code{x}.
}
\seealso{
Other strings: 
\code{\link{blank.}()},
\code{\link{charn.}()},
\code{\link{delim.}()},
\code{\link{fsub.}()},
\code{\link{ipat.}()},
\code{\link{mkstr.}()},
\code{\link{ox.}()},
\code{\link{pgrid.}()},
\code{\link{rev_str.}()},
\code{\link{spacing.}()},
\code{\link{ss.}()},
\code{\link{tocase.}()},
\code{\link{weave.}()}
}
\concept{strings}
