% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnch.R
\name{max_nch}
\alias{max_nch}
\title{Length of longest string}
\usage{
max_nch(x)
}
\arguments{
\item{x}{An \link[=ichr]{character object}.}
}
\value{
An integer scalar.
}
\description{
Gets the number of characters in the longest string in \code{x}.
}
\seealso{
Other chars: 
\code{\link{chn}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()}

Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{mkd}},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{chars}
\concept{strings}
