% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meets.R
\name{meets}
\alias{meets}
\title{Does an object meet count and/or value restrictions}
\usage{
meets(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional restrictive arguments. See details.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks whether \code{x} meets a variety of count and value
restrictions supplied through \code{...}.
}
\details{
Additional restrictive arguments are optional. When no additional
arguments are supplied \code{TRUE} is returned. Possible arguments are as
follows:
\tabular{ll}{
\code{n}   \tab The set of valid lengths/numbers of elements.
\cr\code{nr}  \tab The set of valid numbers of rows.
\cr\code{nc}  \tab The set of valid numbers of columns.
\cr\code{min} \tab The minimum valid length/number of element.
\cr\code{minr}\tab The minimum valid number of rows.
\cr\code{minc}\tab The minimum valid number of columns.
\cr\code{max} \tab The maximum valid length/number of element.
\cr\code{maxr}\tab The maximum valid number of rows.
\cr\code{maxc}\tab The maximum valid number of columns.
\cr\code{vals}\tab The set of valid values.
\cr\code{lt}  \tab The value all values of \code{x} must be less than.
\cr\code{le}  \tab The value all values of \code{x} must be less than or
equal to.
\cr\code{ge}  \tab The value all values of \code{x} must be less than or
equal to.
\cr\code{gt}  \tab The value all values of \code{x} must be greater than.
}
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
