% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meets.R
\encoding{UTF-8}
\name{meets}
\alias{meets}
\title{Evaluate whether an object meets count and/or value restrictions}
\usage{
meets(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Optional named arguments count and/or value restrictions for \code{x}. See the \emph{specifying count and value restrictions} section.}
}
\value{
A logical scalar.
}
\description{
Evaluates whether \code{x} meets any count and value restrictions are provided in \code{...}. If none are provided, returns \code{TRUE}.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{.MAX, .MAXR, .MAXC}   \tab Scalar maximum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.MIN, .MINR, .MINC}   \tab Scalar minimum valid numbers of elements, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{.LT, .LE, .GE, .GT}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{.N, .NR, .NC}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{.VALS}                \tab A vector of valid values.                                                                                                                                 }
}

\examples{
chrs <- c("a", "b", "c")
nums <- 1:3
sqrd <- matrix(1:16, nrow = 4)
meets(chrs, .N = 1:5)
meets(chrs, .MAX = 2)
meets(chrs, .MIN = 3)
meets(sqrd, .NR = 3, .NR = 2:10)
meets(sqrd, .MINR = 2, .MAXR = 5, .MINC = 4, .MAXC = 5)
meets(chrs, .VALS = letters)
meets(chrs, .GT = "a", .LT = "c")
meets(sqrd, .GE = 1, .LE = 16)
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp_fast}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}

Other logicals: 
\code{\link{failsafe}()}

Other counts: 
\code{\link{N}()},
\code{\link{maxnch}()}
}
\concept{counts}
\concept{logicals}
\concept{properties}
