% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meets.R
\name{meets}
\alias{meets}
\title{Does an object meet count and/or value restrictions?}
\usage{
meets(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Optional named arguments specifying (additional) property
requirements for \code{x} in terms of element values and/or
element/row/column counts. See the \emph{Specifying Count and Value
Restrictions} section.}
}
\value{
A logical scalar.
}
\description{
Count and value restrictions are provided in \code{...}. If none
are provided
}
\section{Specifying Count and Value Restrictions}{

Specifying additional requirements in \code{...} is optional. The full set
of recognized arguments names are defined in the following table along with
the properties each specifies:\tabular{ll}{
NAME          \tab WHAT IT SPECIFIES                                    \cr
\code{n}      \tab Vec of valid lengths/numbers of elements.            \cr
\code{nr}     \tab Vec of valid numbers of rows.                        \cr
\code{nc}     \tab Vec of valid numbers of columns.                     \cr
\code{min}    \tab Scalar minimum valid length/number of element.       \cr
\code{minr}   \tab Scalar minimum valid number of rows.                 \cr
\code{minc}   \tab Scalar minimum valid number of columns.              \cr
\code{max}    \tab Scalar maximum valid length/number of element.       \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                 \cr
\code{maxc}   \tab Scalar maximum valid number of columns.              \cr
\code{vals}   \tab Vec of valid values.                                 \cr
\code{lt}     \tab Scalar less-than bound on values of \code{x}.        \cr
\code{le}     \tab Scalar less-than-or-equal bound on values of
\code{x}.                                            \cr
\code{ge}     \tab Scalar greater-than-or-equal bound on values of
\code{x}.                                            \cr
\code{gt}     \tab Scalar greater-than bound on values of \code{x}.       }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc_props}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{props}
