% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkd.R
\name{mkd}
\alias{mkd}
\alias{subs}
\alias{sups}
\alias{bfblu}
\alias{blblu}
\alias{bmblu}
\alias{bdblu}
\alias{bfcyn}
\alias{blcyn}
\alias{bmcyn}
\alias{bdcyn}
\alias{bfgrn}
\alias{blgrn}
\alias{bmgrn}
\alias{bdgrn}
\alias{bfmag}
\alias{blmag}
\alias{bmmag}
\alias{bdmag}
\alias{bforn}
\alias{blorn}
\alias{bmorn}
\alias{bdorn}
\alias{bfred}
\alias{blred}
\alias{bmred}
\alias{bdred}
\alias{bfvlt}
\alias{blvlt}
\alias{bmvlt}
\alias{bdvlt}
\alias{bfylw}
\alias{blylw}
\alias{bmylw}
\alias{bdylw}
\alias{bwht}
\alias{bg95}
\alias{bg90}
\alias{bg85}
\alias{bg80}
\alias{bg75}
\alias{bg70}
\alias{bg65}
\alias{bg60}
\alias{bg55}
\alias{bg50}
\alias{bg45}
\alias{bg40}
\alias{bg35}
\alias{bg30}
\alias{bg25}
\alias{bg20}
\alias{bg15}
\alias{bg10}
\alias{bg05}
\alias{bblk}
\alias{binv}
\alias{pfblu}
\alias{plblu}
\alias{pmblu}
\alias{pdblu}
\alias{pfcyn}
\alias{plcyn}
\alias{pmcyn}
\alias{pdcyn}
\alias{pfgrn}
\alias{plgrn}
\alias{pmgrn}
\alias{pdgrn}
\alias{pfmag}
\alias{plmag}
\alias{pmmag}
\alias{pdmag}
\alias{pforn}
\alias{plorn}
\alias{pmorn}
\alias{pdorn}
\alias{pfred}
\alias{plred}
\alias{pmred}
\alias{pdred}
\alias{pfvlt}
\alias{plvlt}
\alias{pmvlt}
\alias{pdvlt}
\alias{pfylw}
\alias{plylw}
\alias{pmylw}
\alias{pdylw}
\alias{pwht}
\alias{pg95}
\alias{pg90}
\alias{pg85}
\alias{pg80}
\alias{pg75}
\alias{pg70}
\alias{pg65}
\alias{pg60}
\alias{pg55}
\alias{pg50}
\alias{pg45}
\alias{pg40}
\alias{pg35}
\alias{pg30}
\alias{pg25}
\alias{pg20}
\alias{pg15}
\alias{pg10}
\alias{pg05}
\alias{pblk}
\alias{pinv}
\title{Build markdown expressions for styled + colored \code{ggtext}}
\usage{
subs(x)

sups(x)

bfblu(x)

blblu(x)

bmblu(x)

bdblu(x)

bfcyn(x)

blcyn(x)

bmcyn(x)

bdcyn(x)

bfgrn(x)

blgrn(x)

bmgrn(x)

bdgrn(x)

bfmag(x)

blmag(x)

bmmag(x)

bdmag(x)

bforn(x)

blorn(x)

bmorn(x)

bdorn(x)

bfred(x)

blred(x)

bmred(x)

bdred(x)

bfvlt(x)

blvlt(x)

bmvlt(x)

bdvlt(x)

bfylw(x)

blylw(x)

bmylw(x)

bdylw(x)

bwht(x)

bg95(x)

bg90(x)

bg85(x)

bg80(x)

bg75(x)

bg70(x)

bg65(x)

bg60(x)

bg55(x)

bg50(x)

bg45(x)

bg40(x)

bg35(x)

bg30(x)

bg25(x)

bg20(x)

bg15(x)

bg10(x)

bg05(x)

bblk(x)

binv(x)

pfblu(x)

plblu(x)

pmblu(x)

pdblu(x)

pfcyn(x)

plcyn(x)

pmcyn(x)

pdcyn(x)

pfgrn(x)

plgrn(x)

pmgrn(x)

pdgrn(x)

pfmag(x)

plmag(x)

pmmag(x)

pdmag(x)

pforn(x)

plorn(x)

pmorn(x)

pdorn(x)

pfred(x)

plred(x)

pmred(x)

pdred(x)

pfvlt(x)

plvlt(x)

pmvlt(x)

pdvlt(x)

pfylw(x)

plylw(x)

pmylw(x)

pdylw(x)

pwht(x)

pg95(x)

pg90(x)

pg85(x)

pg80(x)

pg75(x)

pg70(x)

pg65(x)

pg60(x)

pg55(x)

pg50(x)

pg45(x)

pg40(x)

pg35(x)

pg30(x)

pg25(x)

pg20(x)

pg15(x)

pg10(x)

pg05(x)

pblk(x)

pinv(x)
}
\arguments{
\item{x}{A \link[=cmp_chr_vec]{character vector} of text to be styled.}
}
\value{
Markdown-styled character vector
}
\description{
\strong{Subscript/superscript} functions are:
\itemize{
\item \strong{\code{subs}}: subscript.
\item \strong{\code{sups}}: superscript.
}
\cr
\strong{Black, white, and invisible text} functions are:
\itemize{
\item \strong{\code{bblk} vs. \code{pblk}} for bold vs. plain black text.
\item \strong{\code{bwht} vs. \code{pwht}} for bold vs. plain white text.
\item \strong{\code{binv} vs. \code{pinv}} for bold vs. plain invisible text (use to ensure consistent spacing).
}
\cr
\strong{Greyscale} text function names take the following form:
\cr \strong{\code{bgXX} vs. \code{pgXX}} for bold vs. plain grey text where \code{XX} runs from \code{05} to \code{95} in increments of \code{5} to indicate what percentage white the color is. For example, \code{bg45} produces bold grey text with a white:black ratio of \code{45:55}.
\cr\cr
\strong{Other common-color} text function names take the following forms:
\itemize{
\item \strong{prefix = \code{b} vs. \code{p}} for bold vs. plain text.
\item \strong{infix = \code{f} vs. \code{l} vs. \code{m} vs. \code{d}} for full-intensity, light, medium, and dark versions.
\item \strong{suffix = \code{blu}} for blue.
\item \strong{suffix = \code{cyn}} for cyan.
\item \strong{suffix = \code{grn}} for green.
\item \strong{suffix = \code{mag}} for magenta.
\item \strong{suffix = \code{orn}} for orange.
\item \strong{suffix = \code{red}} for red.
\item \strong{suffix = \code{vlt}} for violet.
\item \strong{suffix = \code{ylw}} for yellow.
}
}
\seealso{
Other plots: 
\code{\link{pals}}

Other colors: 
\code{\link{color}},
\code{\link{pals}}
}
\concept{colors}
\concept{plots}
