% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_agn.R
\name{mmm_agn}
\alias{mmm_agn}
\alias{mmm_agn_vals}
\alias{atm_agn}
\alias{chr_agn}
\alias{ch1_agn}
\alias{clr_agn}
\alias{evn_agn}
\alias{fac_agn}
\alias{frc_agn}
\alias{ind_agn}
\alias{lgc_agn}
\alias{cnb_agn}
\alias{neg_agn}
\alias{ngw_agn}
\alias{nng_agn}
\alias{nnw_agn}
\alias{nps_agn}
\alias{npw_agn}
\alias{nst_agn}
\alias{num_agn}
\alias{odd_agn}
\alias{ord_agn}
\alias{pct_agn}
\alias{pos_agn}
\alias{ppn_agn}
\alias{psw_agn}
\alias{srt_agn}
\alias{str_agn}
\alias{uno_agn}
\alias{whl_agn}
\title{Extended Mode + Atomic Generic (mmm + agn)}
\usage{
mmm_agn_vals()

atm_agn(x)

chr_agn(x)

ch1_agn(x)

clr_agn(x)

evn_agn(x)

fac_agn(x)

frc_agn(x)

ind_agn(x)

lgc_agn(x)

cnb_agn(x)

neg_agn(x)

ngw_agn(x)

nng_agn(x)

nnw_agn(x)

nps_agn(x)

npw_agn(x)

nst_agn(x)

num_agn(x)

odd_agn(x)

ord_agn(x)

pct_agn(x)

pos_agn(x)

ppn_agn(x)

psw_agn(x)

srt_agn(x)

str_agn(x)

uno_agn(x)

whl_agn(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_generic}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
