% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_arr.R
\name{mmm_arr}
\alias{mmm_arr}
\alias{mmm_arr_vals}
\alias{chr_arr}
\alias{ch1_arr}
\alias{clr_arr}
\alias{evn_arr}
\alias{fac_arr}
\alias{frc_arr}
\alias{ind_arr}
\alias{lgl_arr}
\alias{neg_arr}
\alias{ngw_arr}
\alias{nng_arr}
\alias{nnw_arr}
\alias{nps_arr}
\alias{npw_arr}
\alias{nst_arr}
\alias{num_arr}
\alias{odd_arr}
\alias{ord_arr}
\alias{pct_arr}
\alias{pos_arr}
\alias{psw_arr}
\alias{ppn_arr}
\alias{srt_arr}
\alias{str_arr}
\alias{uno_arr}
\alias{whl_arr}
\title{Extended Mode + Atomic array (mmm + arr)}
\usage{
mmm_arr_vals()

chr_arr(x)

ch1_arr(x)

clr_arr(x)

evn_arr(x)

fac_arr(x)

frc_arr(x)

ind_arr(x)

lgl_arr(x)

neg_arr(x)

ngw_arr(x)

nng_arr(x)

nnw_arr(x)

nps_arr(x)

npw_arr(x)

nst_arr(x)

num_arr(x)

odd_arr(x)

ord_arr(x)

pct_arr(x)

pos_arr(x)

psw_arr(x)

ppn_arr(x)

srt_arr(x)

str_arr(x)

uno_arr(x)

whl_arr(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + array properties.
}
\details{
See \code{\link{mmm}} and \code{\link{is_atm_arr}}.
}
\section{Functions}{
\itemize{
\item \code{chr_arr}: Is \code{x} a character atomic array?

\item \code{ch1_arr}: Is \code{x} a onechar atomic array?

\item \code{clr_arr}: Is \code{x} a color atomic array?

\item \code{evn_arr}: Is \code{x} an even-numeric atomic array?

\item \code{fac_arr}: Is \code{x} a factor atomic array?

\item \code{frc_arr}: Is \code{x} a fractional-numeric atomic array?

\item \code{ind_arr}: Is \code{x} an indexer atomic array?

\item \code{lgl_arr}: Is \code{x} a logical atomic array?

\item \code{neg_arr}: Is \code{x} a negative-numeric atomic array?

\item \code{ngw_arr}: Is \code{x} a negative-whole-numeric atomic array?

\item \code{nng_arr}: Is \code{x} a non-negative-numeric atomic array?

\item \code{nnw_arr}: Is \code{x} a non-negative-whole atomic array?

\item \code{nps_arr}: Is \code{x} a non-positive-numeric atomic array?

\item \code{npw_arr}: Is \code{x} a non-positive-whole-numeric atomic array?

\item \code{nst_arr}: Is \code{x} a non-sortable atomic array?

\item \code{num_arr}: Is \code{x} a numeric atomic array?

\item \code{odd_arr}: Is \code{x} an odd-numeric atomic array?

\item \code{ord_arr}: Is \code{x} an ordered-factor atomic array?

\item \code{pct_arr}: Is \code{x} a percent-numeric atomic array?

\item \code{pos_arr}: Is \code{x} a positive-numeric atomic array?

\item \code{psw_arr}: Is \code{x} a positive-whole-numeric atomic array?

\item \code{ppn_arr}: Is \code{x} a proportion-numeric atomic array?

\item \code{srt_arr}: Is \code{x} a sortable atomic array?

\item \code{str_arr}: Is \code{x} a string atomic array?

\item \code{uno_arr}: Is \code{x} a unordered-factor atomic array?

\item \code{whl_arr}: Is \code{x} a whole-numeric atomic array?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
