% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_avl.R
\name{mmm_avl}
\alias{mmm_avl}
\alias{mmm_avl_vals}
\alias{chr_avl}
\alias{ch1_avl}
\alias{clr_avl}
\alias{evn_avl}
\alias{fac_avl}
\alias{frc_avl}
\alias{ind_avl}
\alias{lgl_avl}
\alias{neg_avl}
\alias{ngw_avl}
\alias{nng_avl}
\alias{nnw_avl}
\alias{nps_avl}
\alias{npw_avl}
\alias{nst_avl}
\alias{num_avl}
\alias{odd_avl}
\alias{ord_avl}
\alias{pct_avl}
\alias{pos_avl}
\alias{ppn_avl}
\alias{psw_avl}
\alias{srt_avl}
\alias{str_avl}
\alias{uno_avl}
\alias{whl_avl}
\title{Extended Mode + Atomic vlist (mmm + avl)}
\usage{
mmm_avl_vals()

chr_avl(x)

ch1_avl(x)

clr_avl(x)

evn_avl(x)

fac_avl(x)

frc_avl(x)

ind_avl(x)

lgl_avl(x)

neg_avl(x)

ngw_avl(x)

nng_avl(x)

nnw_avl(x)

nps_avl(x)

npw_avl(x)

nst_avl(x)

num_avl(x)

odd_avl(x)

ord_avl(x)

pct_avl(x)

pos_avl(x)

ppn_avl(x)

psw_avl(x)

srt_avl(x)

str_avl(x)

uno_avl(x)

whl_avl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic vlist properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_vls}}.
}
\section{Functions}{
\itemize{
\item \code{chr_avl}: Is \code{x} a character atomic vlist?

\item \code{ch1_avl}: Is \code{x} a onechar atomic vlist?

\item \code{clr_avl}: Is \code{x} a color atomic vlist?

\item \code{evn_avl}: Is \code{x} an even-numeric atomic vlist?

\item \code{fac_avl}: Is \code{x} a factor atomic vlist?

\item \code{frc_avl}: Is \code{x} a fractional-numeric atomic vlist?

\item \code{ind_avl}: Is \code{x} an indexer atomic vlist?

\item \code{lgl_avl}: Is \code{x} a logical atomic vlist?

\item \code{neg_avl}: Is \code{x} a negative-numeric atomic vlist?

\item \code{ngw_avl}: Is \code{x} a negative-whole-numeric atomic vlist?

\item \code{nng_avl}: Is \code{x} a non-negative-numeric atomic vlist?

\item \code{nnw_avl}: Is \code{x} a non-negative-whole-numeric atomic vlist?

\item \code{nps_avl}: Is \code{x} a non-positive-numeric atomic vlist?

\item \code{npw_avl}: Is \code{x} a non-positive-whole-numeric atomic vlist?

\item \code{nst_avl}: Is \code{x} a non-sortable atomic vlist?

\item \code{num_avl}: Is \code{x} a numeric atomic vlist?

\item \code{odd_avl}: Is \code{x} an odd-numeric atomic vlist?

\item \code{ord_avl}: Is \code{x} an ordered-factor atomic vlist?

\item \code{pct_avl}: Is \code{x} a percent-numeric atomic vlist?

\item \code{pos_avl}: Is \code{x} a positive-numeric atomic vlist?

\item \code{ppn_avl}: Is \code{x} a proportion-numeric atomic vlist?

\item \code{psw_avl}: Is \code{x} a positive-whole-numeric atomic vlist?

\item \code{srt_avl}: Is \code{x} a sortable atomic vlist?

\item \code{str_avl}: Is \code{x} a string atomic vlist?

\item \code{uno_avl}: Is \code{x} an unordered-factor atomic vlist?

\item \code{whl_avl}: Is \code{x} a whole-numeric atomic vlist?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
