% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_avt.R
\name{mmm_avt}
\alias{mmm_avt}
\alias{mmm_avt_vals}
\alias{atm_avt}
\alias{chr_avt}
\alias{ch1_avt}
\alias{clr_avt}
\alias{evn_avt}
\alias{fac_avt}
\alias{frc_avt}
\alias{ind_avt}
\alias{lgc_avt}
\alias{cnb_avt}
\alias{neg_avt}
\alias{ngw_avt}
\alias{nng_avt}
\alias{nnw_avt}
\alias{nps_avt}
\alias{npw_avt}
\alias{nst_avt}
\alias{num_avt}
\alias{odd_avt}
\alias{ord_avt}
\alias{pct_avt}
\alias{pos_avt}
\alias{ppn_avt}
\alias{psw_avt}
\alias{srt_avt}
\alias{str_avt}
\alias{uno_avt}
\alias{whl_avt}
\title{Extended Mode + Atomic vtype (mmm + avt)}
\usage{
mmm_avt_vals()

atm_avt(x)

chr_avt(x)

ch1_avt(x)

clr_avt(x)

evn_avt(x)

fac_avt(x)

frc_avt(x)

ind_avt(x)

lgc_avt(x)

cnb_avt(x)

neg_avt(x)

ngw_avt(x)

nng_avt(x)

nnw_avt(x)

nps_avt(x)

npw_avt(x)

nst_avt(x)

num_avt(x)

odd_avt(x)

ord_avt(x)

pct_avt(x)

pos_avt(x)

ppn_avt(x)

psw_avt(x)

srt_avt(x)

str_avt(x)

uno_avt(x)

whl_avt(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_vtype}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
