% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_ccc.R
\name{mmm_ccc.}
\alias{mmm_ccc.}
\alias{mmm_ccc_vals}
\alias{ch1_arr}
\alias{chr_arr}
\alias{clr_arr}
\alias{evn_arr}
\alias{fac_arr}
\alias{frc_arr}
\alias{ind_arr}
\alias{lgl_arr}
\alias{neg_arr}
\alias{ngw_arr}
\alias{nng_arr}
\alias{nnw_arr}
\alias{nps_arr}
\alias{npw_arr}
\alias{nst_arr}
\alias{num_arr}
\alias{odd_arr}
\alias{ord_arr}
\alias{pct_arr}
\alias{pos_arr}
\alias{ppn_arr}
\alias{psw_arr}
\alias{srt_arr}
\alias{str_arr}
\alias{uno_arr}
\alias{whl_arr}
\alias{ch1_dtf}
\alias{chr_dtf}
\alias{clr_dtf}
\alias{evn_dtf}
\alias{fac_dtf}
\alias{frc_dtf}
\alias{ind_dtf}
\alias{lgl_dtf}
\alias{neg_dtf}
\alias{ngw_dtf}
\alias{nng_dtf}
\alias{nnw_dtf}
\alias{nps_dtf}
\alias{npw_dtf}
\alias{nst_dtf}
\alias{num_dtf}
\alias{odd_dtf}
\alias{ord_dtf}
\alias{pct_dtf}
\alias{pos_dtf}
\alias{ppn_dtf}
\alias{psw_dtf}
\alias{srt_dtf}
\alias{str_dtf}
\alias{uno_dtf}
\alias{whl_dtf}
\alias{ch1_gen}
\alias{chr_gen}
\alias{clr_gen}
\alias{evn_gen}
\alias{fac_gen}
\alias{frc_gen}
\alias{ind_gen}
\alias{lgl_gen}
\alias{neg_gen}
\alias{ngw_gen}
\alias{nng_gen}
\alias{nnw_gen}
\alias{nps_gen}
\alias{npw_gen}
\alias{nst_gen}
\alias{num_gen}
\alias{odd_gen}
\alias{ord_gen}
\alias{pct_gen}
\alias{pos_gen}
\alias{ppn_gen}
\alias{psw_gen}
\alias{srt_gen}
\alias{str_gen}
\alias{uno_gen}
\alias{whl_gen}
\alias{ch1_mat}
\alias{chr_mat}
\alias{clr_mat}
\alias{evn_mat}
\alias{fac_mat}
\alias{frc_mat}
\alias{ind_mat}
\alias{lgl_mat}
\alias{neg_mat}
\alias{ngw_mat}
\alias{nng_mat}
\alias{nnw_mat}
\alias{nps_mat}
\alias{npw_mat}
\alias{nst_mat}
\alias{num_mat}
\alias{odd_mat}
\alias{ord_mat}
\alias{pct_mat}
\alias{pos_mat}
\alias{ppn_mat}
\alias{psw_mat}
\alias{srt_mat}
\alias{str_mat}
\alias{uno_mat}
\alias{whl_mat}
\alias{ch1_mvc}
\alias{chr_mvc}
\alias{clr_mvc}
\alias{evn_mvc}
\alias{fac_mvc}
\alias{frc_mvc}
\alias{ind_mvc}
\alias{lgl_mvc}
\alias{neg_mvc}
\alias{ngw_mvc}
\alias{nng_mvc}
\alias{nnw_mvc}
\alias{nps_mvc}
\alias{npw_mvc}
\alias{nst_mvc}
\alias{num_mvc}
\alias{odd_mvc}
\alias{ord_mvc}
\alias{pct_mvc}
\alias{pos_mvc}
\alias{ppn_mvc}
\alias{psw_mvc}
\alias{srt_mvc}
\alias{str_mvc}
\alias{uno_mvc}
\alias{whl_mvc}
\alias{ch1_scl}
\alias{chr_scl}
\alias{clr_scl}
\alias{evn_scl}
\alias{fac_scl}
\alias{frc_scl}
\alias{ind_scl}
\alias{lgl_scl}
\alias{neg_scl}
\alias{ngw_scl}
\alias{nng_scl}
\alias{nnw_scl}
\alias{nps_scl}
\alias{npw_scl}
\alias{nst_scl}
\alias{num_scl}
\alias{odd_scl}
\alias{ord_scl}
\alias{pct_scl}
\alias{pos_scl}
\alias{ppn_scl}
\alias{psw_scl}
\alias{srt_scl}
\alias{str_scl}
\alias{uno_scl}
\alias{whl_scl}
\alias{ch1_vec}
\alias{chr_vec}
\alias{clr_vec}
\alias{evn_vec}
\alias{fac_vec}
\alias{frc_vec}
\alias{ind_vec}
\alias{lgl_vec}
\alias{neg_vec}
\alias{ngw_vec}
\alias{nng_vec}
\alias{nnw_vec}
\alias{nps_vec}
\alias{npw_vec}
\alias{nst_vec}
\alias{num_vec}
\alias{odd_vec}
\alias{ord_vec}
\alias{pct_vec}
\alias{pos_vec}
\alias{ppn_vec}
\alias{psw_vec}
\alias{srt_vec}
\alias{str_vec}
\alias{uno_vec}
\alias{whl_vec}
\alias{ch1_vls}
\alias{chr_vls}
\alias{clr_vls}
\alias{evn_vls}
\alias{fac_vls}
\alias{frc_vls}
\alias{ind_vls}
\alias{lgl_vls}
\alias{neg_vls}
\alias{ngw_vls}
\alias{nng_vls}
\alias{nnw_vls}
\alias{nps_vls}
\alias{npw_vls}
\alias{nst_vls}
\alias{num_vls}
\alias{odd_vls}
\alias{ord_vls}
\alias{pct_vls}
\alias{pos_vls}
\alias{ppn_vls}
\alias{psw_vls}
\alias{srt_vls}
\alias{str_vls}
\alias{uno_vls}
\alias{whl_vls}
\title{Extended Mode + Extended Class Properties}
\usage{
mmm_ccc.()

mmm_ccc_vals()

ch1_arr(x)

chr_arr(x)

clr_arr(x)

evn_arr(x)

fac_arr(x)

frc_arr(x)

ind_arr(x)

lgl_arr(x)

neg_arr(x)

ngw_arr(x)

nng_arr(x)

nnw_arr(x)

nps_arr(x)

npw_arr(x)

nst_arr(x)

num_arr(x)

odd_arr(x)

ord_arr(x)

pct_arr(x)

pos_arr(x)

ppn_arr(x)

psw_arr(x)

srt_arr(x)

str_arr(x)

uno_arr(x)

whl_arr(x)

ch1_dtf(x)

chr_dtf(x)

clr_dtf(x)

evn_dtf(x)

fac_dtf(x)

frc_dtf(x)

ind_dtf(x)

lgl_dtf(x)

neg_dtf(x)

ngw_dtf(x)

nng_dtf(x)

nnw_dtf(x)

nps_dtf(x)

npw_dtf(x)

nst_dtf(x)

num_dtf(x)

odd_dtf(x)

ord_dtf(x)

pct_dtf(x)

pos_dtf(x)

ppn_dtf(x)

psw_dtf(x)

srt_dtf(x)

str_dtf(x)

uno_dtf(x)

whl_dtf(x)

ch1_gen(x)

chr_gen(x)

clr_gen(x)

evn_gen(x)

fac_gen(x)

frc_gen(x)

ind_gen(x)

lgl_gen(x)

neg_gen(x)

ngw_gen(x)

nng_gen(x)

nnw_gen(x)

nps_gen(x)

npw_gen(x)

nst_gen(x)

num_gen(x)

odd_gen(x)

ord_gen(x)

pct_gen(x)

pos_gen(x)

ppn_gen(x)

psw_gen(x)

srt_gen(x)

str_gen(x)

uno_gen(x)

whl_gen(x)

ch1_mat(x)

chr_mat(x)

clr_mat(x)

evn_mat(x)

fac_mat(x)

frc_mat(x)

ind_mat(x)

lgl_mat(x)

neg_mat(x)

ngw_mat(x)

nng_mat(x)

nnw_mat(x)

nps_mat(x)

npw_mat(x)

nst_mat(x)

num_mat(x)

odd_mat(x)

ord_mat(x)

pct_mat(x)

pos_mat(x)

ppn_mat(x)

psw_mat(x)

srt_mat(x)

str_mat(x)

uno_mat(x)

whl_mat(x)

ch1_mvc(x)

chr_mvc(x)

clr_mvc(x)

evn_mvc(x)

fac_mvc(x)

frc_mvc(x)

ind_mvc(x)

lgl_mvc(x)

neg_mvc(x)

ngw_mvc(x)

nng_mvc(x)

nnw_mvc(x)

nps_mvc(x)

npw_mvc(x)

nst_mvc(x)

num_mvc(x)

odd_mvc(x)

ord_mvc(x)

pct_mvc(x)

pos_mvc(x)

ppn_mvc(x)

psw_mvc(x)

srt_mvc(x)

str_mvc(x)

uno_mvc(x)

whl_mvc(x)

ch1_scl(x)

chr_scl(x)

clr_scl(x)

evn_scl(x)

fac_scl(x)

frc_scl(x)

ind_scl(x)

lgl_scl(x)

neg_scl(x)

ngw_scl(x)

nng_scl(x)

nnw_scl(x)

nps_scl(x)

npw_scl(x)

nst_scl(x)

num_scl(x)

odd_scl(x)

ord_scl(x)

pct_scl(x)

pos_scl(x)

ppn_scl(x)

psw_scl(x)

srt_scl(x)

str_scl(x)

uno_scl(x)

whl_scl(x)

ch1_vec(x)

chr_vec(x)

clr_vec(x)

evn_vec(x)

fac_vec(x)

frc_vec(x)

ind_vec(x)

lgl_vec(x)

neg_vec(x)

ngw_vec(x)

nng_vec(x)

nnw_vec(x)

nps_vec(x)

npw_vec(x)

nst_vec(x)

num_vec(x)

odd_vec(x)

ord_vec(x)

pct_vec(x)

pos_vec(x)

ppn_vec(x)

psw_vec(x)

srt_vec(x)

str_vec(x)

uno_vec(x)

whl_vec(x)

ch1_vls(x)

chr_vls(x)

clr_vls(x)

evn_vls(x)

fac_vls(x)

frc_vls(x)

ind_vls(x)

lgl_vls(x)

neg_vls(x)

ngw_vls(x)

nng_vls(x)

nnw_vls(x)

nps_vls(x)

npw_vls(x)

nst_vls(x)

num_vls(x)

odd_vls(x)

ord_vls(x)

pct_vls(x)

pos_vls(x)

ppn_vls(x)

psw_vls(x)

srt_vls(x)

str_vls(x)

uno_vls(x)

whl_vls(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\tabular{ll}{
\code{mmm_ccc_vals}   \tab A character vector.                          \cr
\code{[mmm]_[ccc]}    \tab A logical scalar.                              }
}
\description{
\tabular{ll}{
\code{mmm_ccc_vals}   \tab Gets a character vector of all possible
\link[=mmm]{extended mode} +
\link[=ccc]{extended class} properties.      \cr
\code{[mmm]_[ccc]}    \tab Evaluates whether \code{x} is of the
\link[=mmm]{extended mode} represented by the
placeholder \code{[mmm]} and of the
\link[=ccc]{extended class} represented by the
placeholder \code{[ccc]}.                      }
}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
