% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_ccc.R
\name{mmm_ccc.}
\alias{mmm_ccc.}
\alias{mmm_ccc_vals}
\alias{ch1_arr}
\alias{ch1_gen}
\alias{ch1_dtf}
\alias{ch1_vls}
\alias{ch1_vtp}
\alias{ch1_mat}
\alias{ch1_mvc}
\alias{ch1_scl}
\alias{ch1_vec}
\alias{chr_arr}
\alias{chr_gen}
\alias{chr_dtf}
\alias{chr_vls}
\alias{chr_vtp}
\alias{chr_mat}
\alias{chr_mvc}
\alias{chr_scl}
\alias{chr_vec}
\alias{clr_arr}
\alias{clr_gen}
\alias{clr_dtf}
\alias{clr_vls}
\alias{clr_vtp}
\alias{clr_mat}
\alias{clr_mvc}
\alias{clr_scl}
\alias{clr_vec}
\alias{evn_arr}
\alias{evn_gen}
\alias{evn_dtf}
\alias{evn_vls}
\alias{evn_vtp}
\alias{evn_mat}
\alias{evn_mvc}
\alias{evn_scl}
\alias{evn_vec}
\alias{fac_arr}
\alias{fac_gen}
\alias{fac_dtf}
\alias{fac_vls}
\alias{fac_vtp}
\alias{fac_mat}
\alias{fac_mvc}
\alias{fac_scl}
\alias{fac_vec}
\alias{frc_arr}
\alias{frc_gen}
\alias{frc_dtf}
\alias{frc_vls}
\alias{frc_vtp}
\alias{frc_mat}
\alias{frc_mvc}
\alias{frc_scl}
\alias{frc_vec}
\alias{ind_arr}
\alias{ind_gen}
\alias{ind_dtf}
\alias{ind_vls}
\alias{ind_vtp}
\alias{ind_mat}
\alias{ind_mvc}
\alias{ind_scl}
\alias{ind_vec}
\alias{lgl_arr}
\alias{lgl_gen}
\alias{lgl_dtf}
\alias{lgl_vls}
\alias{lgl_vtp}
\alias{lgl_mat}
\alias{lgl_mvc}
\alias{lgl_scl}
\alias{lgl_vec}
\alias{neg_arr}
\alias{neg_gen}
\alias{neg_dtf}
\alias{neg_vls}
\alias{neg_vtp}
\alias{neg_mat}
\alias{neg_mvc}
\alias{neg_scl}
\alias{neg_vec}
\alias{ngw_arr}
\alias{ngw_gen}
\alias{ngw_dtf}
\alias{ngw_vls}
\alias{ngw_vtp}
\alias{ngw_mat}
\alias{ngw_mvc}
\alias{ngw_scl}
\alias{ngw_vec}
\alias{nng_arr}
\alias{nng_gen}
\alias{nng_dtf}
\alias{nng_vls}
\alias{nng_vtp}
\alias{nng_mat}
\alias{nng_mvc}
\alias{nng_scl}
\alias{nng_vec}
\alias{nnw_arr}
\alias{nnw_gen}
\alias{nnw_dtf}
\alias{nnw_vls}
\alias{nnw_vtp}
\alias{nnw_mat}
\alias{nnw_mvc}
\alias{nnw_scl}
\alias{nnw_vec}
\alias{nps_arr}
\alias{nps_gen}
\alias{nps_dtf}
\alias{nps_vls}
\alias{nps_vtp}
\alias{nps_mat}
\alias{nps_mvc}
\alias{nps_scl}
\alias{nps_vec}
\alias{npw_arr}
\alias{npw_gen}
\alias{npw_dtf}
\alias{npw_vls}
\alias{npw_vtp}
\alias{npw_mat}
\alias{npw_mvc}
\alias{npw_scl}
\alias{npw_vec}
\alias{nst_arr}
\alias{nst_gen}
\alias{nst_dtf}
\alias{nst_vls}
\alias{nst_vtp}
\alias{nst_mat}
\alias{nst_mvc}
\alias{nst_scl}
\alias{nst_vec}
\alias{num_arr}
\alias{num_gen}
\alias{num_dtf}
\alias{num_vls}
\alias{num_vtp}
\alias{num_mat}
\alias{num_mvc}
\alias{num_scl}
\alias{num_vec}
\alias{odd_arr}
\alias{odd_gen}
\alias{odd_dtf}
\alias{odd_vls}
\alias{odd_vtp}
\alias{odd_mat}
\alias{odd_mvc}
\alias{odd_scl}
\alias{odd_vec}
\alias{ord_arr}
\alias{ord_gen}
\alias{ord_dtf}
\alias{ord_vls}
\alias{ord_vtp}
\alias{ord_mat}
\alias{ord_mvc}
\alias{ord_scl}
\alias{ord_vec}
\alias{pct_arr}
\alias{pct_gen}
\alias{pct_dtf}
\alias{pct_vls}
\alias{pct_vtp}
\alias{pct_mat}
\alias{pct_mvc}
\alias{pct_scl}
\alias{pct_vec}
\alias{pos_arr}
\alias{pos_gen}
\alias{pos_dtf}
\alias{pos_vls}
\alias{pos_vtp}
\alias{pos_mat}
\alias{pos_mvc}
\alias{pos_scl}
\alias{pos_vec}
\alias{ppn_arr}
\alias{ppn_gen}
\alias{ppn_dtf}
\alias{ppn_vls}
\alias{ppn_vtp}
\alias{ppn_mat}
\alias{ppn_mvc}
\alias{ppn_scl}
\alias{ppn_vec}
\alias{psw_arr}
\alias{psw_gen}
\alias{psw_dtf}
\alias{psw_vls}
\alias{psw_vtp}
\alias{psw_mat}
\alias{psw_mvc}
\alias{psw_scl}
\alias{psw_vec}
\alias{srt_arr}
\alias{srt_gen}
\alias{srt_dtf}
\alias{srt_vls}
\alias{srt_vtp}
\alias{srt_mat}
\alias{srt_mvc}
\alias{srt_scl}
\alias{srt_vec}
\alias{str_arr}
\alias{str_gen}
\alias{str_dtf}
\alias{str_vls}
\alias{str_vtp}
\alias{str_mat}
\alias{str_mvc}
\alias{str_scl}
\alias{str_vec}
\alias{uno_arr}
\alias{uno_gen}
\alias{uno_dtf}
\alias{uno_vls}
\alias{uno_vtp}
\alias{uno_mat}
\alias{uno_mvc}
\alias{uno_scl}
\alias{uno_vec}
\alias{whl_arr}
\alias{whl_gen}
\alias{whl_dtf}
\alias{whl_vls}
\alias{whl_vtp}
\alias{whl_mat}
\alias{whl_mvc}
\alias{whl_scl}
\alias{whl_vec}
\title{Extended mode + extended class properties}
\usage{
mmm_ccc.()

mmm_ccc_vals()

ch1_arr(x)

ch1_gen(x)

ch1_dtf(x)

ch1_vls(x)

ch1_vtp(x)

ch1_mat(x)

ch1_mvc(x)

ch1_scl(x)

ch1_vec(x)

chr_arr(x)

chr_gen(x)

chr_dtf(x)

chr_vls(x)

chr_vtp(x)

chr_mat(x)

chr_mvc(x)

chr_scl(x)

chr_vec(x)

clr_arr(x)

clr_gen(x)

clr_dtf(x)

clr_vls(x)

clr_vtp(x)

clr_mat(x)

clr_mvc(x)

clr_scl(x)

clr_vec(x)

evn_arr(x)

evn_gen(x)

evn_dtf(x)

evn_vls(x)

evn_vtp(x)

evn_mat(x)

evn_mvc(x)

evn_scl(x)

evn_vec(x)

fac_arr(x)

fac_gen(x)

fac_dtf(x)

fac_vls(x)

fac_vtp(x)

fac_mat(x)

fac_mvc(x)

fac_scl(x)

fac_vec(x)

frc_arr(x)

frc_gen(x)

frc_dtf(x)

frc_vls(x)

frc_vtp(x)

frc_mat(x)

frc_mvc(x)

frc_scl(x)

frc_vec(x)

ind_arr(x)

ind_gen(x)

ind_dtf(x)

ind_vls(x)

ind_vtp(x)

ind_mat(x)

ind_mvc(x)

ind_scl(x)

ind_vec(x)

lgl_arr(x)

lgl_gen(x)

lgl_dtf(x)

lgl_vls(x)

lgl_vtp(x)

lgl_mat(x)

lgl_mvc(x)

lgl_scl(x)

lgl_vec(x)

neg_arr(x)

neg_gen(x)

neg_dtf(x)

neg_vls(x)

neg_vtp(x)

neg_mat(x)

neg_mvc(x)

neg_scl(x)

neg_vec(x)

ngw_arr(x)

ngw_gen(x)

ngw_dtf(x)

ngw_vls(x)

ngw_vtp(x)

ngw_mat(x)

ngw_mvc(x)

ngw_scl(x)

ngw_vec(x)

nng_arr(x)

nng_gen(x)

nng_dtf(x)

nng_vls(x)

nng_vtp(x)

nng_mat(x)

nng_mvc(x)

nng_scl(x)

nng_vec(x)

nnw_arr(x)

nnw_gen(x)

nnw_dtf(x)

nnw_vls(x)

nnw_vtp(x)

nnw_mat(x)

nnw_mvc(x)

nnw_scl(x)

nnw_vec(x)

nps_arr(x)

nps_gen(x)

nps_dtf(x)

nps_vls(x)

nps_vtp(x)

nps_mat(x)

nps_mvc(x)

nps_scl(x)

nps_vec(x)

npw_arr(x)

npw_gen(x)

npw_dtf(x)

npw_vls(x)

npw_vtp(x)

npw_mat(x)

npw_mvc(x)

npw_scl(x)

npw_vec(x)

nst_arr(x)

nst_gen(x)

nst_dtf(x)

nst_vls(x)

nst_vtp(x)

nst_mat(x)

nst_mvc(x)

nst_scl(x)

nst_vec(x)

num_arr(x)

num_gen(x)

num_dtf(x)

num_vls(x)

num_vtp(x)

num_mat(x)

num_mvc(x)

num_scl(x)

num_vec(x)

odd_arr(x)

odd_gen(x)

odd_dtf(x)

odd_vls(x)

odd_vtp(x)

odd_mat(x)

odd_mvc(x)

odd_scl(x)

odd_vec(x)

ord_arr(x)

ord_gen(x)

ord_dtf(x)

ord_vls(x)

ord_vtp(x)

ord_mat(x)

ord_mvc(x)

ord_scl(x)

ord_vec(x)

pct_arr(x)

pct_gen(x)

pct_dtf(x)

pct_vls(x)

pct_vtp(x)

pct_mat(x)

pct_mvc(x)

pct_scl(x)

pct_vec(x)

pos_arr(x)

pos_gen(x)

pos_dtf(x)

pos_vls(x)

pos_vtp(x)

pos_mat(x)

pos_mvc(x)

pos_scl(x)

pos_vec(x)

ppn_arr(x)

ppn_gen(x)

ppn_dtf(x)

ppn_vls(x)

ppn_vtp(x)

ppn_mat(x)

ppn_mvc(x)

ppn_scl(x)

ppn_vec(x)

psw_arr(x)

psw_gen(x)

psw_dtf(x)

psw_vls(x)

psw_vtp(x)

psw_mat(x)

psw_mvc(x)

psw_scl(x)

psw_vec(x)

srt_arr(x)

srt_gen(x)

srt_dtf(x)

srt_vls(x)

srt_vtp(x)

srt_mat(x)

srt_mvc(x)

srt_scl(x)

srt_vec(x)

str_arr(x)

str_gen(x)

str_dtf(x)

str_vls(x)

str_vtp(x)

str_mat(x)

str_mvc(x)

str_scl(x)

str_vec(x)

uno_arr(x)

uno_gen(x)

uno_dtf(x)

uno_vls(x)

uno_vtp(x)

uno_mat(x)

uno_mvc(x)

uno_scl(x)

uno_vec(x)

whl_arr(x)

whl_gen(x)

whl_dtf(x)

whl_vls(x)

whl_vtp(x)

whl_mat(x)

whl_mvc(x)

whl_scl(x)

whl_vec(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Evaluate whether an object is of a specific combination of
extended mode (\code{\link{mmm}}) and extended class (\code{\link{ccc}}).
}
\section{Functions}{
\itemize{
\item \code{mmm_ccc_vals()}: Get a character vector of all possible extended mode +
extended class properties.

\item \code{ch1_arr()}: Is \code{x} a onechar array? (\code{\link{ich1}},
\code{\link{iarr}})

\item \code{ch1_gen()}: Is \code{x} a onechar generic? (\code{\link{ich1}},
\code{\link{iagn}})

\item \code{ch1_dtf()}: Is \code{x} a onechar dtf? (\code{\link{ich1}},
\code{\link{idtf}})

\item \code{ch1_vls()}: Is \code{x} a onechar vlist? (\code{\link{ich1}},
\code{\link{iavl}})

\item \code{ch1_vtp()}: Is \code{x} a onechar vtype? (\code{\link{ich1}},
\code{\link{iavt}})

\item \code{ch1_mat()}: Is \code{x} a onechar matrix? (\code{\link{ich1}},
\code{\link{imat}})

\item \code{ch1_mvc()}: Is \code{x} a onechar multivec? (\code{\link{ich1}},
\code{\link{imvc}})

\item \code{ch1_scl()}: Is \code{x} a onechar scalar? (\code{\link{ich1}},
\code{\link{iscl}})

\item \code{ch1_vec()}: Is \code{x} a onechar vector? (\code{\link{ich1}},
\code{\link{ivec}})

\item \code{chr_arr()}: Is \code{x} a character array? (\code{\link{ichr}},
\code{\link{iarr}})

\item \code{chr_gen()}: Is \code{x} a character generic? (\code{\link{ichr}},
\code{\link{iagn}})

\item \code{chr_dtf()}: Is \code{x} a character dtf? (\code{\link{ichr}},
\code{\link{idtf}})

\item \code{chr_vls()}: Is \code{x} a character vlist? (\code{\link{ichr}},
\code{\link{iavl}})

\item \code{chr_vtp()}: Is \code{x} a character vtype? (\code{\link{ichr}},
\code{\link{iavt}})

\item \code{chr_mat()}: Is \code{x} a character matrix? (\code{\link{ichr}},
\code{\link{imat}})

\item \code{chr_mvc()}: Is \code{x} a character multivec? (\code{\link{ichr}},
\code{\link{imvc}})

\item \code{chr_scl()}: Is \code{x} a character scalar? (\code{\link{ichr}},
\code{\link{iscl}})

\item \code{chr_vec()}: Is \code{x} a character vector? (\code{\link{ichr}},
\code{\link{ivec}})

\item \code{clr_arr()}: Is \code{x} a color array? (\code{\link{iclr}},
\code{\link{iarr}})

\item \code{clr_gen()}: Is \code{x} a color generic? (\code{\link{iclr}},
\code{\link{igen}})

\item \code{clr_dtf()}: Is \code{x} a color dtf? (\code{\link{iclr}},
\code{\link{idtf}})

\item \code{clr_vls()}: Is \code{x} a color vlist? (\code{\link{iclr}},
\code{\link{iavl}})

\item \code{clr_vtp()}: Is \code{x} a color vtype? (\code{\link{iclr}},
\code{\link{iavt}})

\item \code{clr_mat()}: Is \code{x} a color matrix? (\code{\link{iclr}},
\code{\link{imat}})

\item \code{clr_mvc()}: Is \code{x} a color multivec? (\code{\link{iclr}},
\code{\link{imvc}})

\item \code{clr_scl()}: Is \code{x} a color scalar? (\code{\link{iclr}},
\code{\link{iscl}})

\item \code{clr_vec()}: Is \code{x} a color vector? (\code{\link{iclr}},
\code{\link{ivec}})

\item \code{evn_arr()}: Is \code{x} an even-valued whole-number array?
(\code{\link{ievn}}, \code{\link{iarr}})

\item \code{evn_gen()}: Is \code{x} an even-valued whole-number generic?
(\code{\link{ievn}}, \code{\link{iagn}})

\item \code{evn_dtf()}: Is \code{x} an even-valued whole-number dtf?
(\code{\link{ievn}}, \code{\link{idtf}})

\item \code{evn_vls()}: Is \code{x} an even-valued whole-number vlist?
(\code{\link{ievn}}, \code{\link{iavl}})

\item \code{evn_vtp()}: Is \code{x} an even-valued
whole-number vtype?
(\code{\link{ievn}}, \code{\link{iavt}})

\item \code{evn_mat()}: Is \code{x} an even-valued
whole-number matrix?
(\code{\link{ievn}}, \code{\link{imat}})

\item \code{evn_mvc()}: Is \code{x} an even-valued
whole-number multivec?
(\code{\link{ievn}}, \code{\link{imvc}})

\item \code{evn_scl()}: Is \code{x} an even-valued whole-number scalar?
(\code{\link{ievn}}, \code{\link{iscl}})

\item \code{evn_vec()}: Is \code{x} an even-valued whole-number vector?
(\code{\link{ievn}}, \code{\link{ivec}})

\item \code{fac_arr()}: Is \code{x} a factor array? (\code{\link{ifac}},
\code{\link{iarr}})

\item \code{fac_gen()}: Is \code{x} a factor generic? (\code{\link{ifac}},
\code{\link{iagn}})

\item \code{fac_dtf()}: Is \code{x} a factor dtf? (\code{\link{ifac}},
\code{\link{idtf}})

\item \code{fac_vls()}: Is \code{x} a factor vlist? (\code{\link{ifac}},
\code{\link{iavl}})

\item \code{fac_vtp()}: Is \code{x} a factor vtype? (\code{\link{ifac}},
\code{\link{iavt}})

\item \code{fac_mat()}: Is \code{x} a factor matrix? (\code{\link{ifac}},
\code{\link{imat}})

\item \code{fac_mvc()}: Is \code{x} a factor multivec? (\code{\link{ifac}},
\code{\link{imvc}})

\item \code{fac_scl()}: Is \code{x} a factor scalar? (\code{\link{ifac}},
\code{\link{iscl}})

\item \code{fac_vec()}: Is \code{x} a factor vector? (\code{\link{ifac}},
\code{\link{ivec}})

\item \code{frc_arr()}: Is \code{x} a fractional-valued numeric array?
(\code{\link{ifrc}}, \code{\link{iarr}})

\item \code{frc_gen()}: Is \code{x} a fractional-valued numeric generic?
(\code{\link{ifrc}}, \code{\link{iagn}})

\item \code{frc_dtf()}: Is \code{x} a fractional-valued numeric dtf?
(\code{\link{ifrc}}, \code{\link{idtf}})

\item \code{frc_vls()}: Is \code{x} a fractional-valued numeric vlist?
(\code{\link{ifrc}}, \code{\link{iavl}})

\item \code{frc_vtp()}: Is \code{x} a fractional-valued numeric vtype?
(\code{\link{ifrc}}, \code{\link{iavt}})

\item \code{frc_mat()}: Is \code{x} a fractional-valued
numeric matrix?
(\code{\link{ifrc}}, \code{\link{imat}})

\item \code{frc_mvc()}: Is \code{x} a fractional-valued
numeric multivec?
(\code{\link{ifrc}}, \code{\link{imvc}})

\item \code{frc_scl()}: Is \code{x} a fractional-valued
numeric scalar?
(\code{\link{ifrc}}, \code{\link{iscl}})

\item \code{frc_vec()}: Is \code{x} a fractional-valued
numeric vector?
(\code{\link{ifrc}}, \code{\link{ivec}})

\item \code{ind_arr()}: Is \code{x} an indexer
array?
(\code{\link{iind}}, \code{\link{iarr}})

\item \code{ind_gen()}: Is \code{x} an indexer
generic?
(\code{\link{iind}}, \code{\link{iagn}})

\item \code{ind_dtf()}: Is \code{x} an indexer
dtf?
(\code{\link{iind}}, \code{\link{idtf}})

\item \code{ind_vls()}: Is \code{x} an indexer
vlist?
(\code{\link{iind}}, \code{\link{iavl}})

\item \code{ind_vtp()}: Is \code{x} an indexer
vtype?
(\code{\link{iind}}, \code{\link{iavt}})

\item \code{ind_mat()}: Is \code{x} an indexer
matrix?
(\code{\link{iind}}, \code{\link{imat}})

\item \code{ind_mvc()}: Is \code{x} an indexer
multivec?
(\code{\link{iind}}, \code{\link{imvc}})

\item \code{ind_scl()}: Is \code{x} an indexer
scalar?
(\code{\link{iind}}, \code{\link{iscl}})

\item \code{ind_vec()}: Is \code{x} an indexer
vector?
(\code{\link{iind}}, \code{\link{ivec}})

\item \code{lgl_arr()}: Is \code{x} a logical
array?
(\code{\link{ilgl}}, \code{\link{iarr}})

\item \code{lgl_gen()}: Is \code{x} a logical
generic?
(\code{\link{ilgl}}, \code{\link{iagn}})

\item \code{lgl_dtf()}: Is \code{x} a logical
dtf?
(\code{\link{ilgl}}, \code{\link{idtf}})

\item \code{lgl_vls()}: Is \code{x} a logical
vlist?
(\code{\link{ilgl}}, \code{\link{iavl}})

\item \code{lgl_vtp()}: Is \code{x} a logical
vtype?
(\code{\link{ilgl}}, \code{\link{iavt}})

\item \code{lgl_mat()}: Is \code{x} a logical
matrix?
(\code{\link{ilgl}}, \code{\link{imat}})

\item \code{lgl_mvc()}: Is \code{x} a logical
multivec?
(\code{\link{ilgl}}, \code{\link{imvc}})

\item \code{lgl_scl()}: Is \code{x} a logical
scalar?
(\code{\link{ilgl}}, \code{\link{iscl}})

\item \code{lgl_vec()}: Is \code{x} a logical
vector?
(\code{\link{ilgl}}, \code{\link{ivec}})

\item \code{neg_arr()}: Is \code{x} a negative-valued
numeric array?
(\code{\link{ineg}}, \code{\link{iarr}})

\item \code{neg_gen()}: Is \code{x} a negative-valued
numeric generic?
(\code{\link{ineg}}, \code{\link{iagn}})

\item \code{neg_dtf()}: Is \code{x} a negative-valued
numeric dtf?
(\code{\link{ineg}}, \code{\link{idtf}})

\item \code{neg_vls()}: Is \code{x} a negative-valued
numeric vlist?
(\code{\link{ineg}}, \code{\link{iavl}})

\item \code{neg_vtp()}: Is \code{x} a negative-valued
numeric vtype?
(\code{\link{ineg}}, \code{\link{iavt}})

\item \code{neg_mat()}: Is \code{x} a negative-valued
numeric matrix?
(\code{\link{ineg}}, \code{\link{imat}})

\item \code{neg_mvc()}: Is \code{x} a negative-valued
numeric multivec?
(\code{\link{ineg}}, \code{\link{imvc}})

\item \code{neg_scl()}: Is \code{x} a negative-valued
numeric scalar?
(\code{\link{ineg}}, \code{\link{iscl}})

\item \code{neg_vec()}: Is \code{x} a negative-valued
numeric vector?
(\code{\link{ineg}}, \code{\link{ivec}})

\item \code{ngw_arr()}: Is \code{x} a negative-valued
whole-number array?
(\code{\link{ingw}}, \code{\link{iarr}})

\item \code{ngw_gen()}: Is \code{x} a negative-valued
whole-number generic?
(\code{\link{ingw}}, \code{\link{iagn}})

\item \code{ngw_dtf()}: Is \code{x} a negative-valued
whole-number dtf?
(\code{\link{ingw}}, \code{\link{idtf}})

\item \code{ngw_vls()}: Is \code{x} a negative-valued
whole-number vlist?
(\code{\link{ingw}}, \code{\link{iavl}})

\item \code{ngw_vtp()}: Is \code{x} a negative-valued
whole-number vtype?
(\code{\link{ingw}}, \code{\link{iavt}})

\item \code{ngw_mat()}: Is \code{x} a negative-valued
whole-number matrix?
(\code{\link{ingw}}, \code{\link{imat}})

\item \code{ngw_mvc()}: Is \code{x} a negative-valued
whole-number multivec?
(\code{\link{ingw}}, \code{\link{imvc}})

\item \code{ngw_scl()}: Is \code{x} a negative-valued
whole-number scalar?
(\code{\link{ingw}}, \code{\link{iscl}})

\item \code{ngw_vec()}: Is \code{x} a negative-valued
whole-number vector?
(\code{\link{ingw}}, \code{\link{ivec}})

\item \code{nng_arr()}: Is \code{x} a non-negative-valued
numeric array?
(\code{\link{inng}}, \code{\link{iarr}})

\item \code{nng_gen()}: Is \code{x} a non-negative-valued
numeric generic?
(\code{\link{inng}}, \code{\link{iagn}})

\item \code{nng_dtf()}: Is \code{x} a non-negative-valued
numeric dtf?
(\code{\link{inng}}, \code{\link{idtf}})

\item \code{nng_vls()}: Is \code{x} a non-negative-valued
numeric vlist?
(\code{\link{inng}}, \code{\link{iavl}})

\item \code{nng_vtp()}: Is \code{x} a non-negative-valued
numeric vtype?
(\code{\link{inng}}, \code{\link{iavt}})

\item \code{nng_mat()}: Is \code{x} a non-negative-valued
numeric matrix?
(\code{\link{inng}}, \code{\link{imat}})

\item \code{nng_mvc()}: Is \code{x} a non-negative-valued
numeric multivec?
(\code{\link{inng}}, \code{\link{imvc}})

\item \code{nng_scl()}: Is \code{x} a non-negative-valued
numeric scalar?
(\code{\link{inng}}, \code{\link{iscl}})

\item \code{nng_vec()}: Is \code{x} a non-negative-valued
numeric vector?
(\code{\link{inng}}, \code{\link{ivec}})

\item \code{nnw_arr()}: Is \code{x} a non-negative-valued
whole-number array?
(\code{\link{innw}}, \code{\link{iarr}})

\item \code{nnw_gen()}: Is \code{x} a non-negative-valued
whole-number generic?
(\code{\link{innw}}, \code{\link{iagn}})

\item \code{nnw_dtf()}: Is \code{x} a non-negative-valued
whole-number dtf?
(\code{\link{innw}}, \code{\link{idtf}})

\item \code{nnw_vls()}: Is \code{x} a non-negative-valued
whole-number vlist?
(\code{\link{innw}}, \code{\link{iavl}})

\item \code{nnw_vtp()}: Is \code{x} a non-negative-valued
whole-number vtype?
(\code{\link{innw}}, \code{\link{iavt}})

\item \code{nnw_mat()}: Is \code{x} a non-negative-valued
whole-number matrix?
(\code{\link{innw}}, \code{\link{imat}})

\item \code{nnw_mvc()}: Is \code{x} a non-negative-valued
whole-number multivec?
(\code{\link{innw}}, \code{\link{imvc}})

\item \code{nnw_scl()}: Is \code{x} a non-negative-valued
whole-number scalar?
(\code{\link{innw}}, \code{\link{iscl}})

\item \code{nnw_vec()}: Is \code{x} a non-negative-valued
whole-number vector?
(\code{\link{innw}}, \code{\link{ivec}})

\item \code{nps_arr()}: Is \code{x} a non-positive-valued
numeric array?
(\code{\link{inps}}, \code{\link{iarr}})

\item \code{nps_gen()}: Is \code{x} a non-positive-valued
numeric generic?
(\code{\link{inps}}, \code{\link{iagn}})

\item \code{nps_dtf()}: Is \code{x} a non-positive-valued
numeric dtf?
(\code{\link{inps}}, \code{\link{idtf}})

\item \code{nps_vls()}: Is \code{x} a non-positive-valued
numeric vlist?
(\code{\link{inps}}, \code{\link{iavl}})

\item \code{nps_vtp()}: Is \code{x} a non-positive-valued
numeric vtype?
(\code{\link{inps}}, \code{\link{iavt}})

\item \code{nps_mat()}: Is \code{x} a non-positive-valued
numeric matrix?
(\code{\link{inps}}, \code{\link{imat}})

\item \code{nps_mvc()}: Is \code{x} a non-positive-valued
numeric multivec?
(\code{\link{inps}}, \code{\link{imvc}})

\item \code{nps_scl()}: Is \code{x} a non-positive-valued
numeric scalar?
(\code{\link{inps}}, \code{\link{iscl}})

\item \code{nps_vec()}: Is \code{x} a non-positive-valued
numeric vector?
(\code{\link{inps}}, \code{\link{ivec}})

\item \code{npw_arr()}: Is \code{x} a non-positive-valued
whole-number array?
(\code{\link{inpw}}, \code{\link{iarr}})

\item \code{npw_gen()}: Is \code{x} a non-positive-valued
whole-number generic?
(\code{\link{inpw}}, \code{\link{iagn}})

\item \code{npw_dtf()}: Is \code{x} a non-positive-valued
whole-number dtf?
(\code{\link{inpw}}, \code{\link{idtf}})

\item \code{npw_vls()}: Is \code{x} a non-positive-valued
whole-number vlist?
(\code{\link{inpw}}, \code{\link{iavl}})

\item \code{npw_vtp()}: Is \code{x} a non-positive-valued
whole-number vtype?
(\code{\link{inpw}}, \code{\link{iavt}})

\item \code{npw_mat()}: Is \code{x} a non-positive-valued
whole-number matrix?
(\code{\link{inpw}}, \code{\link{imat}})

\item \code{npw_mvc()}: Is \code{x} a non-positive-valued
whole-number multivec?
(\code{\link{inpw}}, \code{\link{imvc}})

\item \code{npw_scl()}: Is \code{x} a non-positive-valued
whole-number scalar?
(\code{\link{inpw}}, \code{\link{iscl}})

\item \code{npw_vec()}: Is \code{x} a non-positive-valued
whole-number vector?
(\code{\link{inpw}}, \code{\link{ivec}})

\item \code{nst_arr()}: Is \code{x} a non-sortable
array?
(\code{\link{inst}}, \code{\link{iarr}})

\item \code{nst_gen()}: Is \code{x} a non-sortable
generic?
(\code{\link{inst}}, \code{\link{iagn}})

\item \code{nst_dtf()}: Is \code{x} a non-sortable
dtf?
(\code{\link{inst}}, \code{\link{idtf}})

\item \code{nst_vls()}: Is \code{x} a non-sortable
vlist?
(\code{\link{inst}}, \code{\link{iavl}})

\item \code{nst_vtp()}: Is \code{x} a non-sortable
vtype?
(\code{\link{inst}}, \code{\link{iavt}})

\item \code{nst_mat()}: Is \code{x} a non-sortable
matrix?
(\code{\link{inst}}, \code{\link{imat}})

\item \code{nst_mvc()}: Is \code{x} a non-sortable
multivec?
(\code{\link{inst}}, \code{\link{imvc}})

\item \code{nst_scl()}: Is \code{x} a non-sortable
scalar?
(\code{\link{inst}}, \code{\link{iscl}})

\item \code{nst_vec()}: Is \code{x} a non-sortable
vector?
(\code{\link{inst}}, \code{\link{ivec}})

\item \code{num_arr()}: Is \code{x} a numeric
array?
(\code{\link{inum}}, \code{\link{iarr}})

\item \code{num_gen()}: Is \code{x} a numeric
generic?
(\code{\link{inum}}, \code{\link{iagn}})

\item \code{num_dtf()}: Is \code{x} a numeric
dtf?
(\code{\link{inum}}, \code{\link{idtf}})

\item \code{num_vls()}: Is \code{x} a numeric
vlist?
(\code{\link{inum}}, \code{\link{iavl}})

\item \code{num_vtp()}: Is \code{x} a numeric
vtype?
(\code{\link{inum}}, \code{\link{iavt}})

\item \code{num_mat()}: Is \code{x} a numeric
matrix?
(\code{\link{inum}}, \code{\link{imat}})

\item \code{num_mvc()}: Is \code{x} a numeric
multivec?
(\code{\link{inum}}, \code{\link{imvc}})

\item \code{num_scl()}: Is \code{x} a numeric
scalar?
(\code{\link{inum}}, \code{\link{iscl}})

\item \code{num_vec()}: Is \code{x} a numeric
vector?
(\code{\link{inum}}, \code{\link{ivec}})

\item \code{odd_arr()}: Is \code{x} an odd-valued
whole-number array?
(\code{\link{iodd}}, \code{\link{iarr}})

\item \code{odd_gen()}: Is \code{x} an odd-valued
whole-number generic?
(\code{\link{iodd}}, \code{\link{iagn}})

\item \code{odd_dtf()}: Is \code{x} an odd-valued
whole-number dtf?
(\code{\link{iodd}}, \code{\link{idtf}})

\item \code{odd_vls()}: Is \code{x} an odd-valued
whole-number vlist?
(\code{\link{iodd}}, \code{\link{iavl}})

\item \code{odd_vtp()}: Is \code{x} an odd-valued
whole-number vtype?
(\code{\link{iodd}}, \code{\link{iavt}})

\item \code{odd_mat()}: Is \code{x} an odd-valued
whole-number matrix?
(\code{\link{iodd}}, \code{\link{imat}})

\item \code{odd_mvc()}: Is \code{x} an odd-valued
whole-number multivec?
(\code{\link{iodd}}, \code{\link{imvc}})

\item \code{odd_scl()}: Is \code{x} an odd-valued
whole-number scalar?
(\code{\link{iodd}}, \code{\link{iscl}})

\item \code{odd_vec()}: Is \code{x} an odd-valued
whole-number vector?
(\code{\link{iodd}}, \code{\link{ivec}})

\item \code{ord_arr()}: Is \code{x} an ordered-factor
array?
(\code{\link{iord}}, \code{\link{iarr}})

\item \code{ord_gen()}: Is \code{x} an ordered-factor
generic?
(\code{\link{iord}}, \code{\link{iagn}})

\item \code{ord_dtf()}: Is \code{x} an ordered-factor
dtf?
(\code{\link{iord}}, \code{\link{idtf}})

\item \code{ord_vls()}: Is \code{x} an ordered-factor
vlist?
(\code{\link{iord}}, \code{\link{iavl}})

\item \code{ord_vtp()}: Is \code{x} an ordered-factor
vtype?
(\code{\link{iord}}, \code{\link{iavt}})

\item \code{ord_mat()}: Is \code{x} an ordered-factor
matrix?
(\code{\link{iord}}, \code{\link{imat}})

\item \code{ord_mvc()}: Is \code{x} an ordered-factor
multivec?
(\code{\link{iord}}, \code{\link{imvc}})

\item \code{ord_scl()}: Is \code{x} an ordered-factor
scalar?
(\code{\link{iord}}, \code{\link{iscl}})

\item \code{ord_vec()}: Is \code{x} an ordered-factor
vector?
(\code{\link{iord}}, \code{\link{ivec}})

\item \code{pct_arr()}: Is \code{x} a percent-valued
(0-100) numeric array?
(\code{\link{ipct}}, \code{\link{iarr}})

\item \code{pct_gen()}: Is \code{x} a percent-valued
(0-100) numeric generic?
(\code{\link{ipct}}, \code{\link{iagn}})

\item \code{pct_dtf()}: Is \code{x} a percent-valued
(0-100) numeric dtf?
(\code{\link{ipct}}, \code{\link{idtf}})

\item \code{pct_vls()}: Is \code{x} a percent-valued
(0-100) numeric vlist?
(\code{\link{ipct}}, \code{\link{iavl}})

\item \code{pct_vtp()}: Is \code{x} a percent-valued
(0-100) numeric vtype?
(\code{\link{ipct}}, \code{\link{iavt}})

\item \code{pct_mat()}: Is \code{x} a percent-valued
(0-100) numeric matrix?
(\code{\link{ipct}}, \code{\link{imat}})

\item \code{pct_mvc()}: Is \code{x} a percent-valued
(0-100) numeric multivec?
(\code{\link{ipct}}, \code{\link{imvc}})

\item \code{pct_scl()}: Is \code{x} a percent-valued
(0-100) numeric scalar?
(\code{\link{ipct}}, \code{\link{iscl}})

\item \code{pct_vec()}: Is \code{x} a percent-valued
(0-100) numeric vector?
(\code{\link{ipct}}, \code{\link{ivec}})

\item \code{pos_arr()}: Is \code{x} a positive-valued
numeric array?
(\code{\link{ipos}}, \code{\link{iarr}})

\item \code{pos_gen()}: Is \code{x} a positive-valued
numeric generic?
(\code{\link{ipos}}, \code{\link{iagn}})

\item \code{pos_dtf()}: Is \code{x} a positive-valued
numeric dtf?
(\code{\link{ipos}}, \code{\link{idtf}})

\item \code{pos_vls()}: Is \code{x} a positive-valued
numeric vlist?
(\code{\link{ipos}}, \code{\link{iavl}})

\item \code{pos_vtp()}: Is \code{x} a positive-valued
numeric vtype?
(\code{\link{ipos}}, \code{\link{iavt}})

\item \code{pos_mat()}: Is \code{x} a positive-valued
numeric matrix?
(\code{\link{ipos}}, \code{\link{imat}})

\item \code{pos_mvc()}: Is \code{x} a positive-valued
numeric multivec?
(\code{\link{ipos}}, \code{\link{imvc}})

\item \code{pos_scl()}: Is \code{x} a positive-valued
numeric scalar?
(\code{\link{ipos}}, \code{\link{iscl}})

\item \code{pos_vec()}: Is \code{x} a positive-valued
numeric vector?
(\code{\link{ipos}}, \code{\link{ivec}})

\item \code{ppn_arr()}: Is \code{x} a proportion-valued
(0-1) numeric array?
(\code{\link{ippn}}, \code{\link{iarr}})

\item \code{ppn_gen()}: Is \code{x} a proportion-valued
(0-1) numeric generic?
(\code{\link{ippn}}, \code{\link{iagn}})

\item \code{ppn_dtf()}: Is \code{x} a proportion-valued
(0-1) numeric dtf?
(\code{\link{ippn}}, \code{\link{idtf}})

\item \code{ppn_vls()}: Is \code{x} a proportion-valued
(0-1) numeric vlist?
(\code{\link{ippn}}, \code{\link{iavl}})

\item \code{ppn_vtp()}: Is \code{x} a proportion-valued
(0-1) numeric vtype?
(\code{\link{ippn}}, \code{\link{iavt}})

\item \code{ppn_mat()}: Is \code{x} a proportion-valued
(0-1) numeric matrix?
(\code{\link{ippn}}, \code{\link{imat}})

\item \code{ppn_mvc()}: Is \code{x} a proportion-valued
(0-1) numeric multivec?
(\code{\link{ippn}}, \code{\link{imvc}})

\item \code{ppn_scl()}: Is \code{x} a proportion-valued
(0-1) numeric scalar?
(\code{\link{ippn}}, \code{\link{iscl}})

\item \code{ppn_vec()}: Is \code{x} a proportion-valued
(0-1) numeric vector?
(\code{\link{ippn}}, \code{\link{ivec}})

\item \code{psw_arr()}: Is \code{x} a positive-valued
whole-number array?
(\code{\link{ipsw}}, \code{\link{iarr}})

\item \code{psw_gen()}: Is \code{x} a positive-valued
whole-number generic?
(\code{\link{ipsw}}, \code{\link{iagn}})

\item \code{psw_dtf()}: Is \code{x} a positive-valued
whole-number dtf?
(\code{\link{ipsw}}, \code{\link{idtf}})

\item \code{psw_vls()}: Is \code{x} a positive-valued
whole-number vlist?
(\code{\link{ipsw}}, \code{\link{iavl}})

\item \code{psw_vtp()}: Is \code{x} a positive-valued
whole-number vtype?
(\code{\link{ipsw}}, \code{\link{iavt}})

\item \code{psw_mat()}: Is \code{x} a positive-valued
whole-number matrix?
(\code{\link{ipsw}}, \code{\link{imat}})

\item \code{psw_mvc()}: Is \code{x} a positive-valued
whole-number multivec?
(\code{\link{ipsw}}, \code{\link{imvc}})

\item \code{psw_scl()}: Is \code{x} a positive-valued
whole-number scalar?
(\code{\link{ipsw}}, \code{\link{iscl}})

\item \code{psw_vec()}: Is \code{x} a positive-valued
whole-number vector?
(\code{\link{ipsw}}, \code{\link{ivec}})

\item \code{srt_arr()}: Is \code{x} a sortable
array?
(\code{\link{isrt}}, \code{\link{iarr}})

\item \code{srt_gen()}: Is \code{x} a sortable
generic?
(\code{\link{isrt}}, \code{\link{iagn}})

\item \code{srt_dtf()}: Is \code{x} a sortable
dtf?
(\code{\link{isrt}}, \code{\link{idtf}})

\item \code{srt_vls()}: Is \code{x} a sortable
vlist?
(\code{\link{isrt}}, \code{\link{iavl}})

\item \code{srt_vtp()}: Is \code{x} a sortable
vtype?
(\code{\link{isrt}}, \code{\link{iavt}})

\item \code{srt_mat()}: Is \code{x} a sortable
matrix?
(\code{\link{isrt}}, \code{\link{imat}})

\item \code{srt_mvc()}: Is \code{x} a sortable
multivec?
(\code{\link{isrt}}, \code{\link{imvc}})

\item \code{srt_scl()}: Is \code{x} a sortable
scalar?
(\code{\link{isrt}}, \code{\link{iscl}})

\item \code{srt_vec()}: Is \code{x} a sortable
vector?
(\code{\link{isrt}}, \code{\link{ivec}})

\item \code{str_arr()}: Is \code{x} a string
array?
(\code{\link{istr}}, \code{\link{iarr}})

\item \code{str_gen()}: Is \code{x} a string
generic?
(\code{\link{istr}}, \code{\link{iagn}})

\item \code{str_dtf()}: Is \code{x} a string
dtf?
(\code{\link{istr}}, \code{\link{idtf}})

\item \code{str_vls()}: Is \code{x} a string
vlist?
(\code{\link{istr}}, \code{\link{iavl}})

\item \code{str_vtp()}: Is \code{x} a string
vtype?
(\code{\link{istr}}, \code{\link{iavt}})

\item \code{str_mat()}: Is \code{x} a string
matrix?
(\code{\link{istr}}, \code{\link{imat}})

\item \code{str_mvc()}: Is \code{x} a string
multivec?
(\code{\link{istr}}, \code{\link{imvc}})

\item \code{str_scl()}: Is \code{x} a string
scalar?
(\code{\link{istr}}, \code{\link{iscl}})

\item \code{str_vec()}: Is \code{x} a string
vector?
(\code{\link{istr}}, \code{\link{ivec}})

\item \code{uno_arr()}: Is \code{x} an unordered-factor
array?
(\code{\link{iuno}}, \code{\link{iarr}})

\item \code{uno_gen()}: Is \code{x} an unordered-factor
generic?
(\code{\link{iuno}}, \code{\link{iagn}})

\item \code{uno_dtf()}: Is \code{x} an unordered-factor
dtf?
(\code{\link{iuno}}, \code{\link{idtf}})

\item \code{uno_vls()}: Is \code{x} an unordered-factor
vlist?
(\code{\link{iuno}}, \code{\link{iavl}})

\item \code{uno_vtp()}: Is \code{x} an unordered-factor
vtype?
(\code{\link{iuno}}, \code{\link{iavt}})

\item \code{uno_mat()}: Is \code{x} an unordered-factor
matrix?
(\code{\link{iuno}}, \code{\link{imat}})

\item \code{uno_mvc()}: Is \code{x} an unordered-factor
multivec?
(\code{\link{iuno}}, \code{\link{imvc}})

\item \code{uno_scl()}: Is \code{x} an unordered-factor
scalar?
(\code{\link{iuno}}, \code{\link{iscl}})

\item \code{uno_vec()}: Is \code{x} an unordered-factor
vector?
(\code{\link{iuno}}, \code{\link{ivec}})

\item \code{whl_arr()}: Is \code{x} a whole-number
array?
(\code{\link{iwhl}}, \code{\link{iarr}})

\item \code{whl_gen()}: Is \code{x} a whole-number
generic?
(\code{\link{iwhl}}, \code{\link{iagn}})

\item \code{whl_dtf()}: Is \code{x} a whole-number
dtf?
(\code{\link{iwhl}}, \code{\link{idtf}})

\item \code{whl_vls()}: Is \code{x} a whole-number
vlist?
(\code{\link{iwhl}}, \code{\link{iavl}})

\item \code{whl_vtp()}: Is \code{x} a whole-number
vtype?
(\code{\link{iwhl}}, \code{\link{iavt}})

\item \code{whl_mat()}: Is \code{x} a whole-number
matrix?
(\code{\link{iwhl}}, \code{\link{imat}})

\item \code{whl_mvc()}: Is \code{x} a whole-number
multivec?
(\code{\link{iwhl}}, \code{\link{imvc}})

\item \code{whl_scl()}: Is \code{x} a whole-number
scalar?
(\code{\link{iwhl}}, \code{\link{iscl}})

\item \code{whl_vec()}: Is \code{x} a whole-number
vector?
(\code{\link{iwhl}}, \code{\link{ivec}})

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{is_ccc_uj}()},
\code{\link{is_mmm.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()},
\code{\link{xxx.}()}
}
\concept{props}
