% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_mat.R
\name{mmm_mat}
\alias{mmm_mat}
\alias{mmm_mat_vals}
\alias{chr_mat}
\alias{ch1_mat}
\alias{clr_mat}
\alias{evn_mat}
\alias{fac_mat}
\alias{frc_mat}
\alias{ind_mat}
\alias{lgl_mat}
\alias{neg_mat}
\alias{ngw_mat}
\alias{nng_mat}
\alias{nnw_mat}
\alias{nps_mat}
\alias{npw_mat}
\alias{nst_mat}
\alias{num_mat}
\alias{odd_mat}
\alias{ord_mat}
\alias{pct_mat}
\alias{pos_mat}
\alias{ppn_mat}
\alias{psw_mat}
\alias{srt_mat}
\alias{str_mat}
\alias{uno_mat}
\alias{whl_mat}
\title{Extended Mode + Atomic matrix (mmm + mat)}
\usage{
mmm_mat_vals()

chr_mat(x)

ch1_mat(x)

clr_mat(x)

evn_mat(x)

fac_mat(x)

frc_mat(x)

ind_mat(x)

lgl_mat(x)

neg_mat(x)

ngw_mat(x)

nng_mat(x)

nnw_mat(x)

nps_mat(x)

npw_mat(x)

nst_mat(x)

num_mat(x)

odd_mat(x)

ord_mat(x)

pct_mat(x)

pos_mat(x)

ppn_mat(x)

psw_mat(x)

srt_mat(x)

str_mat(x)

uno_mat(x)

whl_mat(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Get all possible extended mode + atomic matrix properties.
}
\details{
See \code{\link{mmm}}, and \code{\link{atm_mat}}.
}
\section{Functions}{
\itemize{
\item \code{chr_mat}: Is \code{x} a character matrix?

\item \code{ch1_mat}: Is \code{x} a onechar matrix?

\item \code{clr_mat}: Is \code{x} a color matrix?

\item \code{evn_mat}: Is \code{x} an even-numeric matrix?

\item \code{fac_mat}: Is \code{x} a factor matrix?

\item \code{frc_mat}: Is \code{x} a fractional-numeric matrix?

\item \code{ind_mat}: Is \code{x} an indexer matrix?

\item \code{lgl_mat}: Is \code{x} a logical matrix?

\item \code{neg_mat}: Is \code{x} a negative-numeric matrix?

\item \code{ngw_mat}: Is \code{x} a negative-whole-numeric matrix?

\item \code{nng_mat}: Is \code{x} a non-negative-numeric matrix?

\item \code{nnw_mat}: Is \code{x} a non-negative-whole-numeric matrix?

\item \code{nps_mat}: Is \code{x} a non-positive-numeric matrix?

\item \code{npw_mat}: Is \code{x} a non-positive-whole-numeric matrix?

\item \code{nst_mat}: Is \code{x} a non-sortable matrix?

\item \code{num_mat}: Is \code{x} a numeric matrix?

\item \code{odd_mat}: Is \code{x} an odd-numeric matrix?

\item \code{ord_mat}: Is \code{x} an ordered-factor matrix?

\item \code{pct_mat}: Is \code{x} a percent-numeric matrix?

\item \code{pos_mat}: Is \code{x} a positive-numeric matrix?

\item \code{ppn_mat}: Is \code{x} a proportion-numeric matrix?

\item \code{psw_mat}: Is \code{x} a positive-whole-numeric matrix?

\item \code{srt_mat}: Is \code{x} a sortable matrix?

\item \code{str_mat}: Is \code{x} a string matrix?

\item \code{uno_mat}: Is \code{x} an unordered-factor matrix?

\item \code{whl_mat}: Is \code{x} whole-numeric matrix?
}}

\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{is_failsafe}},
\code{\link{is_xxx}},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
