% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_mvc.R
\name{mmm_mvc}
\alias{mmm_mvc}
\alias{mmm_mvc_vals}
\alias{atm_mvc}
\alias{chr_mvc}
\alias{ch1_mvc}
\alias{clr_mvc}
\alias{evn_mvc}
\alias{fac_mvc}
\alias{frc_mvc}
\alias{ind_mvc}
\alias{lgc_mvc}
\alias{cnb_mvc}
\alias{neg_mvc}
\alias{ngw_mvc}
\alias{nng_mvc}
\alias{nnw_mvc}
\alias{nps_mvc}
\alias{npw_mvc}
\alias{nst_mvc}
\alias{num_mvc}
\alias{odd_mvc}
\alias{ord_mvc}
\alias{pct_mvc}
\alias{pos_mvc}
\alias{ppn_mvc}
\alias{psw_mvc}
\alias{srt_mvc}
\alias{str_mvc}
\alias{uno_mvc}
\alias{whl_mvc}
\title{Extended Mode + Atomic mvect (mmm + mvc)}
\usage{
mmm_mvc_vals()

atm_mvc(x)

chr_mvc(x)

ch1_mvc(x)

clr_mvc(x)

evn_mvc(x)

fac_mvc(x)

frc_mvc(x)

ind_mvc(x)

lgc_mvc(x)

cnb_mvc(x)

neg_mvc(x)

ngw_mvc(x)

nng_mvc(x)

nnw_mvc(x)

nps_mvc(x)

npw_mvc(x)

nst_mvc(x)

num_mvc(x)

odd_mvc(x)

ord_mvc(x)

pct_mvc(x)

pos_mvc(x)

ppn_mvc(x)

psw_mvc(x)

srt_mvc(x)

str_mvc(x)

uno_mvc(x)

whl_mvc(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_mvect}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
