% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_scl.R
\name{mmm_scl}
\alias{mmm_scl}
\alias{mmm_scl_vals}
\alias{atm_scl}
\alias{chr_scl}
\alias{ch1_scl}
\alias{clr_scl}
\alias{evn_scl}
\alias{fac_scl}
\alias{frc_scl}
\alias{ind_scl}
\alias{lgc_scl}
\alias{cnb_scl}
\alias{neg_scl}
\alias{ngw_scl}
\alias{nng_scl}
\alias{nnw_scl}
\alias{nps_scl}
\alias{npw_scl}
\alias{nst_scl}
\alias{num_scl}
\alias{odd_scl}
\alias{ord_scl}
\alias{pct_scl}
\alias{pos_scl}
\alias{ppn_scl}
\alias{psw_scl}
\alias{srt_scl}
\alias{chr_str}
\alias{uno_scl}
\alias{whl_scl}
\title{Extended Mode + Atomic scalar (mmm + scl)}
\usage{
mmm_scl_vals()

atm_scl(x)

chr_scl(x)

ch1_scl(x)

clr_scl(x)

evn_scl(x)

fac_scl(x)

frc_scl(x)

ind_scl(x)

lgc_scl(x)

cnb_scl(x)

neg_scl(x)

ngw_scl(x)

nng_scl(x)

nnw_scl(x)

nps_scl(x)

npw_scl(x)

nst_scl(x)

num_scl(x)

odd_scl(x)

ord_scl(x)

pct_scl(x)

pos_scl(x)

ppn_scl(x)

psw_scl(x)

srt_scl(x)

chr_str(x)

uno_scl(x)

whl_scl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_scalar}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
