% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_vec.R
\name{mmm_vec}
\alias{mmm_vec}
\alias{mmm_vec_vals}
\alias{atm_vec}
\alias{chr_vec}
\alias{clr_vec}
\alias{evn_vec}
\alias{fac_vec}
\alias{frc_vec}
\alias{ind_vec}
\alias{lgc_vec}
\alias{cnb_vec}
\alias{neg_vec}
\alias{ngw_vec}
\alias{nng_vec}
\alias{nnw_vec}
\alias{nps_vec}
\alias{npw_vec}
\alias{nst_vec}
\alias{num_vec}
\alias{odd_vec}
\alias{ord_vec}
\alias{pct_vec}
\alias{pos_vec}
\alias{ppn_vec}
\alias{psw_vec}
\alias{srt_vec}
\alias{str_vec}
\alias{uno_vec}
\alias{whl_vec}
\title{Extended Mode + Atomic vect (mmm + vec)}
\usage{
mmm_vec_vals()

atm_vec(x)

chr_vec(x)

clr_vec(x)

evn_vec(x)

fac_vec(x)

frc_vec(x)

ind_vec(x)

lgc_vec(x)

cnb_vec(x)

neg_vec(x)

ngw_vec(x)

nng_vec(x)

nnw_vec(x)

nps_vec(x)

npw_vec(x)

nst_vec(x)

num_vec(x)

odd_vec(x)

ord_vec(x)

pct_vec(x)

pos_vec(x)

ppn_vec(x)

psw_vec(x)

srt_vec(x)

str_vec(x)

uno_vec(x)

whl_vec(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
See \code{\link{mmm}}, and \code{\link{is_atm_vect}}.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{ttt}()},
\code{\link{xxx}()}
}
\concept{props}
