% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\name{n_th}
\alias{n_th}
\alias{first_n}
\alias{last_n}
\alias{n_th_last}
\title{Extract elements by index position}
\usage{
n_th(x, n)

first_n(x, n)

last_n(x, n)

n_th_last(x, n)
}
\arguments{
\item{x}{A \link[=pop_vec]{populated vector} to extract elements from.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}.}
}
\value{
\tabular{lll}{
\code{n_th}, \code{n_th_last}\tab   \tab A (possibly scalar) vector.    \cr
\code{first_n}, \code{last_n}\tab   \tab A scalar vector.                 }
}
\description{
\tabular{ll}{
\code{n_th}        \tab Gets the \code{n}-th element(s) of a vector.    \cr
\code{first_n}     \tab Gets the first \code{n} elements of a vector.   \cr
\code{last_n}      \tab Gets the last \code{n} elements of a vector.    \cr
\code{n_th_last}   \tab Gets the \code{n}-th from last element(s) of a
vector.                                           }
}
\examples{
first_n(letters, 5)
last_n(letters, 5)
n_th(letters, 5)
n_th(letters, )
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put.}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{xb}()}
}
\concept{extensions}
