% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_th.R
\name{n_th}
\alias{n_th}
\alias{first_n}
\alias{last_n}
\alias{n_th_last}
\title{Extract elements by numeric position}
\usage{
n_th(x, n)

first_n(x, n)

last_n(x, n)

n_th_last(x, n)
}
\arguments{
\item{x}{A \link[=pop_vec]{populated vector} to extract elements from.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}.}
}
\value{
A scalar, vector, or \link[=ivls]{vlist}.
}
\description{
\itemize{
\item \strong{\code{n_th}}: gets the \code{n}-th element(s) of a vector.
\item \strong{\code{first_n}}: gets the first \verb{n`` elements of a vector. \\item **}last_n\verb{**: gets the last }n\verb{ elements of a vector. \\item **}n_th_last\verb{**: gets the }n`-th-from-last element(s) of a vector.
}
}
\examples{
first_n(letters, 5)
last_n(letters, 5)
n_th(letters, 5)
n_th(letters, )
}
