% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_th.R
\name{n_th}
\alias{n_th}
\alias{n_th_last}
\alias{first_n}
\alias{last_n}
\title{Extract elements by numeric position}
\usage{
n_th(x, n)

n_th_last(x, n)

first_n(x, n)

last_n(x, n)
}
\arguments{
\item{x}{A \link[=pop_vec]{populated vector} to extract elements from.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}.}
}
\value{
A scalar, vector, or \link[=ivls]{vlist}.
}
\description{
\tabular{rl}{
\code{n_th_last} \tab   Gets the \code{n}-th-from-last element(s).
\cr \code{first_n} \tab   Gets the first \code{n} elements.
\cr  \code{last_n} \tab   Gets the last \code{n} elements.
\cr    \code{n_th} \tab   Gets the \code{n}-th element(s) .
}
}
\examples{
n_th_last(letters, 5)
first_n(letters, 5)
last_n(letters, 5)
n_th(letters, 5:7)
}
