% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\encoding{UTF-8}
\name{na}
\alias{na}
\alias{ok}
\alias{sub_na}
\alias{nas}
\alias{oks}
\alias{rm_na}
\title{Manage \code{NA} and non-\code{NA} values}
\usage{
na(x)

ok(x)

sub_na(x, s)

nas(x)

oks(x)

rm_na(x)
}
\arguments{
\item{x}{The argument to be inspected/managed.}

\item{s}{An \link[=atm_scl]{atomic scalar} to replace \code{NA} values. Mode must be \link[=compatible]{compatible} with\code{x}}
}
\value{
\emph{A logical object of the same dimension as \code{x}}
\cr    \verb{na, ok}
\cr\cr \emph{\code{x} with \code{NA} values replaced}
\cr (properties may change with \code{NA} replacement)
\cr    \code{sub_na}
\cr\cr \emph{\code{x} with \code{NA} values removed}
\cr (properties may change with \code{NA} removal)
\cr    \code{rm_na}
\cr\cr \emph{A logical scalaqr}
\cr    \verb{nas, oks}
}
\description{
All functions in this group take objects of atomic mode, \link[=atm_dtf]{atomic dtfs}, or \link[=atm_vls]{atomic vlists}:
\tabular{rl}{
  \code{sub_na}   \tab Substitutes \code{s} for \code{NA} values of\code{x} Modes of \code{x} and \code{s} must be \link[=compatible]{compatible}.
\cr           \tab  
\cr \code{rm_na}   \tab Removes \code{NA} values of \code{x} If there are no \code{NA} values in \code{x}, returns \code{x} unchanged. If there are \code{NA} values in \code{x}, removing them may change the dimensions and/or class of \code{x}.
\cr           \tab  
\cr   \code{nas}   \tab Evaluates whether \code{x} is an \code{NA} scalar.
\cr   \code{oks}   \tab Evaluates whether \code{x} is a non-\code{NA} scalar.
\cr    \code{na}   \tab Indexes \code{NA} values of \code{x}.
\cr    \code{ok}   \tab Indexes non-\code{NA} values of \code{x}.
}
}
\examples{
na_scl <- NA
ok_scl <- "a"
vec <- c(na_scl, ok_scl)
mat <- sample(vec, 10, replace = T)

mat

na(vec)
ok(vec)

na(mat)
ok(mat)

nas(na_scl)
nas(ok_scl)
nas(vec)

oks(ok_scl)
oks(na_scl)
oks(vec)

sub_na(mat, "x")

rm_na(mat)
}
\seealso{
Other missingness: 
\code{\link{stats0}}

Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unq}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()},
\code{\link{x}()}

Other values: 
\code{\link{atomize}()},
\code{\link{n_th}()},
\code{\link{uv}()},
\code{\link{values}()},
\code{\link{wv}()}
}
\concept{extensions}
\concept{missingness}
\concept{values}
