% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na}
\alias{na}
\alias{ok}
\alias{sub_na}
\alias{nas}
\alias{oks}
\alias{rm_na}
\title{Manage \code{NA} and non-\code{NA} values}
\usage{
na(x)

ok(x)

sub_na(x, s)

nas(x)

oks(x)

rm_na(x)
}
\arguments{
\item{x}{The argument to be inspected/managed.}

\item{s}{An \link[=atm_scl]{atomic scalar} to replace \code{NA} values. Mode must be \link[=compatible]{compatible} with\code{x}}
}
\value{
\itemize{
\item \strong{\code{rm_na}}: either an atomic vector or \code{x}.
\item \strong{\code{sub_na}}: \code{x} with \code{NA} values replaced.
\item \strong{\verb{na, ok}}: a logical object of the same dimension as \code{x}.
\item \strong{\verb{nas, oks}}: a logical scalar.
}
}
\description{
All functions in this group take objects of atomic mode, \link[=atm_dtf]{atomic dtfs}, or \link[=atm_vls]{atomic vlists}:
\itemize{
\item \strong{\code{na}}: indexes \code{NA} values of \code{x}.
\item \strong{\code{ok}}: indexes non-\code{NA} values of \code{x}.
\item \strong{\code{nas}}: evaluates whether \code{x} is an \code{NA} scalar.
\item \strong{\code{oks}}: evaluates whether \code{x} is a non-\code{NA} scalar.
\item \strong{\code{sub_na}}: substitutes \code{s} for \code{NA} values of\code{x} Modes of \code{x} and \code{s} must be \link[=compatible]{compatible}.
\item \strong{\code{rm_na}}: removes \code{NA} values of \code{x} If there are no \code{NA} values in \code{x}, returns \code{x} unchanged. If there are \code{NA} values in \code{x}, may changes the dimensions of and/or class of \code{x} in the course of removing \code{NA} values.
}
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}

Other na_values: 
\code{\link{stats0}}
}
\concept{extensions}
\concept{na_values}
