% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_vals.R
\name{na_vals}
\alias{na_vals}
\alias{na}
\alias{ok}
\alias{sub_na}
\alias{rm_na}
\title{Manage NA and Non-NA Values}
\usage{
na(x)

ok(x)

sub_na(x, s)

rm_na(x)
}
\arguments{
\item{x}{The argument to be inspected/managed.}

\item{s}{Atomic scalar to replace \code{NA} values. Mode must be
\link[=compatible]{compatible} with \code{x}.}
}
\value{
An atomic scalar, an atomic vector, an atomic array, an
\link[=is_atm_tibble]{atomic tibble}, or an
\link[=is_atm_vlist]{atomic vlist}.
}
\description{
Identify \code{NA} values, identify OK values (i.e., not
\code{NA}), substitute a value for \code{NA} values, or remove \code{NA}
values.
}
\details{
All functions in this group take atomic objects,
\link[=is_atm_tibble]{atomic tibbles} and
\link[=is_atm_vlist]{atomic vlists}.
\cr\cr
\strong{\code{na}}
\cr Returns an object with the same dimensions as \code{x} where all atomic
values are replaced by either \code{TRUE} or \code{FALSE} indicating
whether the original atomic value was \code{NA}.
\cr\cr
\strong{\code{ok}}
\cr Returns an object with the same dimensions as \code{x} where all atomic
values are replaced by either \code{TRUE} or \code{FALSE} indicating
whether the original atomic value was OK (i.e., not \code{NA}).
\cr\cr
\strong{\code{sub.na}}
\cr Returns an object with the same dimensions as \code{x} where all
\code{NA} values are replaced by \code{s}. Note that the modes of \code{x}
and \code{s} must be \link[=compatible]{compatible}.
\cr\cr
\strong{\code{rm.na}}
If there are no \code{NA} values in \code{x}, this function returns
\code{x} unchanged.
\cr\cr
However, if there are \code{NA} values in \code{x}, by removing \code{NA}
values from \code{x}, this function changes the dimensions of and/or class
of \code{x}. If \code{x} is an array or an atomic tibble, its class is
reduced to an atomic vector sans \code{NA} values. If \code{x} is an atomic
vlist, any elements with \code{NA} values are reduced to atomic vectors
sans their \code{NA} values.
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
