% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named}
\alias{named}
\alias{enamed}
\alias{rnamed}
\alias{cnamed}
\alias{rcnamed}
\alias{named.}
\title{Are elements of an object named?}
\usage{
named(x., dim. = 0, u. = T, blank. = F)

enamed(x., u. = T, blank. = F)

rnamed(x., u. = T, blank. = F)

cnamed(x., u. = T, blank. = F)

rcnamed(x., u. = T, blank. = F)

named.(..., u. = T, blank. = F)
}
\arguments{
\item{x.}{Argument to be inspected for names.}

\item{dim.}{Dimension \code{0} for elements (of atomic or list vectors only),
\code{1} for rows of matrices or tibbles, \code{2} for columns of matrices
or tibbles, or \code{12} for rows and columns of matrices or tibbles.}

\item{u.}{Logical scalar indicating whether names must be unique.}

\item{blank.}{Logical scalar indicating whether blank strings (\code{""}) are
allowed as names.}

\item{...}{Arbitrary number of arguments to be inspected for names.}
}
\value{
Logical scalar.
}
\description{
Determine whether an objects elements, rows, and/or columns are
named, accounting for whether names must be unique, whether the object may
contain empty elements, and whether names may be blank strings ("").
}
\section{Functions}{
\itemize{
\item \code{enamed}: Are elements of \code{x} named?

\item \code{rnamed}: Are rows of \code{x} named?

\item \code{cnamed}: Are columns of \code{x} named?

\item \code{rcnamed}: Are both rows and columns of \code{x} named?

\item \code{named.}: Are all \code{...} arguments named?
}}

\seealso{
Other naming: 
\code{\link{getnames}()}

Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
\concept{naming}
