% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\encoding{UTF-8}
\name{naming}
\alias{naming}
\alias{namee}
\alias{namer}
\alias{namec}
\alias{namel}
\alias{namerc}
\alias{name}
\alias{named}
\alias{enamed}
\alias{rnamed}
\alias{cnamed}
\alias{rcnamed}
\alias{dnamed}
\alias{get_names}
\alias{dnames}
\alias{dn}
\alias{enames}
\alias{en}
\alias{rnames}
\alias{rn}
\alias{cnames}
\alias{cn}
\alias{rcnames}
\alias{rcn}
\title{Naming utilities}
\usage{
namee(x, en)

namer(x, rn)

namec(x, cn)

namel(x, ln)

namerc(x, rn, cn)

name(x, en = NULL, rn = NULL, cn = NULL)

named(x, dim = 0, u = T, bl = F)

enamed(x, u = T, bl = F)

rnamed(x, u = T, bl = F)

cnamed(x, u = T, bl = F)

rcnamed(x, u = T, bl = F)

dnamed(..., U = T, BL = F)

get_names(x, d = 0, u = T, err = F)

dnames(..., U = T, ERR = F)

dn(..., U = T, ERR = F)

enames(x, u = T, err = F)

en(x, u = T, err = F)

rnames(x, u = T, err = F)

rn(x, u = T, err = F)

cnames(x, u = T, err = F)

cn(x, u = T, err = F)

rcnames(x, u = T, err = F)

rcn(x, u = T, err = F)
}
\arguments{
\item{x}{Vector or list for \code{namee} and \code{enames / en}; matrix or data.frame for \code{rnames / rn}, \code{cnames / cn}, \code{rcnames / rcn}, \code{namerc}, \code{namec}, and \code{namer}; or any object for \code{namel}.}

\item{en}{A complete character vec of element names.}

\item{rn}{A complete character vec of row names.}

\item{cn}{A \link[=cmp_chr_vec]{complete character vec} of column names.}

\item{ln}{A non-\code{NA} character scalar to name the single elemnt of the resulting \link[=VLS]{vlist}.}

\item{dim}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the dimension(s) to name: \code{0} for elements (of atomic or list vectors only), \code{1} for rows of matrices or tibbles, \code{2} for columns of matrices or data.frames, or \code{12} for rows and columns of matrices or data.frames.}

\item{u, U}{\code{TRUE} or \code{FALSE} indicating whether names must be unique.}

\item{bl, BL}{\code{TRUE} or \code{FALSE} indicating whether blank strings (\code{""}) are allowed as names.}

\item{...}{An arbitrary number of arguments.}

\item{err}{\code{TRUE} or \code{FALSE} indicating whether names must exist.}

\item{ERR}{\code{TRUE} or \code{FALSE} indicating whether \code{...} args must have names.}
}
\value{
\strong{A \link[=VEC]{vec}, \link[=VLS]{vlist}, matrix, or data.frame} \cr\cr \code{name}
\cr\cr  \strong{A vector or a vlist of length} \code{2}                             \cr\cr \code{get_names}
\cr\cr  \strong{A matrix or data.frame}                                        \cr\cr \verb{namerc, namer,namec}
\cr\cr  \strong{A \link[=VLS]{vlist} of length} \code{2}                            \cr\cr \verb{rcnames, rcn}
\cr\cr  \strong{A \code{1}-element \link[=VLS]{vlist}}                              \cr\cr \code{namel}
\cr\cr  \strong{A character vector}                                            \cr\cr \verb{rnames, cnames, enames}  \cr \verb{rn, cn, en}
\cr\cr  \strong{A logical scalar}                                              \cr\cr \verb{rcnamed, rnamed, cnamed} \cr \verb{dnamed, enamed, named}
\cr\cr  \strong{A vector or \link[=VLS]{vlist}}                                \cr\cr \code{namev}
}
\description{
Get object names (with optional restrictions), check whether names exist (with optional restrictions), and apply names.
}
\details{
\tabular{ll}{  \code{get_names}      \tab Name retrieval                        \cr
  \code{named}          \tab Name existence check                  \cr
  \code{name}           \tab Naming                                \cr   \tab   \cr
  \verb{rcnames, rcn}   \tab Row and column names                  \cr
  \verb{rnames, rn}     \tab Row names                             \cr
  \verb{cnames, cn}     \tab Column names                          \cr
  \verb{enames, en}     \tab Element names                         \cr
  \verb{dnames, dn}     \tab \code{...} arg names                       \cr   \tab   \cr
  \code{rcnamed}        \tab Are both rows and columns named?      \cr
  \code{cnamed}         \tab Are columns of \code{x} named?             \cr
  \code{dnamed}         \tab Are \code{...} arguments named?            \cr
  \code{enamed}         \tab Are elements of \code{x} named?            \cr
  \code{rnamed}         \tab Are rows of \code{x} named?                \cr   \tab   \cr
  \code{namerc}         \tab Name rows and columns of \code{x}          \cr
  \code{namec}          \tab Name columns of \code{x}.                  \cr
  \code{namel}          \tab Create and name a list of length \code{1}. \cr
  \code{namer}          \tab Name rows of \code{x}.                     \cr
  \code{namee}          \tab Name elements of a \code{x}.                 }
}
\examples{
egDotsNamed <- function(...) {dnamed(...)}

egVec <- 1:5
egVls <- list(num = 1:5, ae = letters[1:5])
egMat <- matrix(letters[1:25], nrow = 5)
egDtf <- data.frame(ae = letters[1:5], AE = LETTERS[1:5])

egVec
egVls
egMat
egDtf

named(egVec)
named(egVls)
named(egMat, d = 12)
named(egDtf, d = 2)

enamed(egVls)
rnamed(egMat)
cnamed(egDtf)
rcnamed(egDtf)

dnamed(egVec, egVls)
dnamed(var1 = egVec, var2 = egVls)

get_names(egVls)
get_names(egMat, d = 1)
get_names(egDtf, d = 2)

en(egVec)
en(egVls)

rn(egMat)
cn(egDtf)
rcn(egDtf)

namel(egDtf, "a data.frame")
namee(egVec, letters[egVec])
namer(egMat, letters[egVec])
namec(egDtf, c("letters", "LETTERS"))
namerc(egMat, letters[egVec], egVec)
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{evals}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{na}()},
\code{\link{nis}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}
}
\concept{extensions}
