% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{naming}
\alias{naming}
\alias{namev}
\alias{namer}
\alias{namec}
\alias{namel}
\alias{namerc}
\alias{named}
\alias{enamed}
\alias{rnamed}
\alias{cnamed}
\alias{rcnamed}
\alias{dnamed}
\alias{getnames}
\alias{en}
\alias{rn}
\alias{cn}
\alias{rcn}
\title{Get names, set names, determine whether named}
\usage{
namev(x, en)

namer(x, rn)

namec(x, cn)

namel(x, ln)

namerc(x, rn, cn)

named(x, d = 0, u = T, blank = F)

enamed(x, u = T, blank = F)

rnamed(x, u = T, blank = F)

cnamed(x, u = T, blank = F)

rcnamed(x, u = T, blank = F)

dnamed(..., u = T, blank = F)

getnames(x, d = 0, u = T, err = F)

en(x, u = T, err = F)

rn(x, u = T, err = F)

cn(x, u = T, err = F)

rcn(x, u = T, err = F)
}
\arguments{
\item{x}{Vector or list for \code{namev} and \code{vn}; matrix or data.frame for \code{rn}, \code{namer}, \code{cn}, \code{namec}, and \code{namerc}; or any object for \code{namel}.}

\item{en}{A \link[=cmp_chr_vec]{complete character vec} of element names.}

\item{rn, cn}{Complete character vecs of row or column names, respectively.}

\item{ln}{A \link[=cmp_chr_scl]{complete character scalar} element name.}

\item{d}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the dimension(s) to name: \code{0} for elements (of atomic or list vectors only), \code{1} for rows of matrices or tibbles, \code{2} for columns of matrices or data.frames, or \code{12} for rows and columns of matrices or data.frames.}

\item{u}{A non-\code{NA} logical scalar indicating whether names must be unique.}

\item{blank}{A non-\code{NA} logical scalar indicating whether blank strings (\code{""}) are allowed as names.}

\item{...}{Arbitrary number of arguments to be inspected for names.}

\item{err}{A non-\code{NA} logical scalar indicating whether names must be non-missing.}
}
\value{
\itemize{
\item \strong{\code{rcn}}: a list of \code{2} character vectors.
\item \strong{\verb{namev**: a vector. \\item **}namel}: a \code{1}-element list.
\item \strong{\verb{en, rn, cn**: a character vector. \\item **}namer, namec, namerc}: a matrix or data.frame.
\item \strong{`named, enamed, rnamed, cnamed, rcnamed}: a logical scalar.
}
}
\description{
\itemize{
\item \strong{\code{named}}: evaluates whether an object's elements, rows, and/or columns are named, accounting for whether names must be unique, whether the object may contain empty elements, and whether names may be blank strings.
\item \strong{\code{enamed}}: evaluates whether elements of \code{x} named.
\item \strong{\verb{rnamed cnamed}}: evaluates whether rows vx. columns of \code{x} are named, respectively.
\item \strong{\code{dnamed}}: evaluates whether dot arguments (\code{...}) named.
}\cr\itemize{
\item \strong{\code{namev}}: names elements of a vector or \link[=ivls]{vlist}.
\item \strong{\code{namel}}: creates and names a 1-element list.
\item \strong{\verb{namer namec namerc}}: name rows vs. columns vs. both rows and columns of \code{x}, respectively.
}\cr\itemize{
\item \strong{\code{en}}: gets element names.
\item \strong{\verb{rn cn rcn}}: get row vs. column vs. both row and column names, respectively.
\item \strong{\code{getnames}}: gets element, row, and/or column names with optional restrictions.
}
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
