% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{naming}
\alias{naming}
\alias{namev}
\alias{namer}
\alias{namec}
\alias{namel}
\alias{namerc}
\alias{vn}
\alias{cn}
\alias{rn}
\title{Get or set element, row, and/or column names.}
\usage{
namev(x, vn)

namer(x, rn)

namec(x, cn)

namel(x, n)

namerc(x, r, c)

vn(x)

cn(x)

rn(x)
}
\arguments{
\item{x}{Vector or vlist for \code{namev} and \code{vn}; matrix or data.frame
for \code{namer}, \code{namec}, \code{namerc}, \code{rn}, and \code{cn}; or
any object for \code{namel}.}

\item{vn}{Character vector of element names.}

\item{rn, cn}{Character vectors of row names or column names, respectively.}

\item{n}{Character scalar element-name.}
}
\value{
Named/renamed version of \code{x}.
}
\description{
Name elements of a vector.
}
\section{Functions}{
\itemize{
\item \code{namer}: Name the rows of a matrix or data frame.

\item \code{namec}: Name the columns of a matrix or data frame.

\item \code{namel}: Create a 1-element list from \code{x} and name that
element with the value of \code{e}

\item \code{namerc}: Name the rows and columns of a matrix or data frame.

\item \code{vn}: Get element names.

\item \code{cn}: Get column names.

\item \code{rn}: Get row names.
}}

