% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{naming}
\alias{naming}
\alias{namev}
\alias{namer}
\alias{namec}
\alias{namel}
\alias{namerc}
\alias{named}
\alias{enamed}
\alias{rnamed}
\alias{cnamed}
\alias{rcnamed}
\alias{dnamed}
\alias{getnames}
\alias{en}
\alias{rn}
\alias{cn}
\alias{rcn}
\title{Get names, set names, determine whether named}
\usage{
namev(x, en)

namer(x, rn)

namec(x, cn)

namel(x, ln)

namerc(x, rn, cn)

named(x, d = 0, u = T, blank = F)

enamed(x, u = T, blank = F)

rnamed(x, u = T, blank = F)

cnamed(x, u = T, blank = F)

rcnamed(x, u = T, blank = F)

dnamed(..., u = T, blank = F)

getnames(x, d = 0, u = T, err = F)

en(x, u = T, err = F)

rn(x, u = T, err = F)

cn(x, u = T, err = F)

rcn(x, u = T, err = F)
}
\arguments{
\item{x}{Vector or list for \code{namev} and \code{vn}; matrix or data.frame for \code{rn}, \code{namer}, \code{cn}, \code{namec}, and \code{namerc}; or any object for \code{namel}.}

\item{en}{A \link[=cmp_chr_vec]{complete character vec} of element names.}

\item{rn, cn}{Complete character vecs of row or column names, respectively.}

\item{ln}{A \link[=cmp_chr_scl]{complete character scalar} element name.}

\item{d}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the dimension(s) to name: \code{0} for elements (of atomic or list vectors only), \code{1} for rows of matrices or tibbles, \code{2} for columns of matrices or data.frames, or \code{12} for rows and columns of matrices or data.frames.}

\item{u}{A non-\code{NA} logical scalar indicating whether names must be unique.}

\item{blank}{A non-\code{NA} logical scalar indicating whether blank strings (\code{""}) are allowed as names.}

\item{...}{Arbitrary number of arguments to be inspected for names.}

\item{err}{A non-\code{NA} logical scalar indicating whether names must be non-missing.}
}
\value{
\tabular{rl}{
   \code{rcnamed}   \tab A logical
\cr \code{rnamed}   \tab scalar.
\cr \code{cnamed}   \tab   
\cr \code{enamed}   \tab   
\cr \code{dnamed}   \tab   
\cr \code{ named}   \tab   
\cr            \tab   
\cr \code{namerc}   \tab A matrix or
\cr  \code{namer}   \tab data.frame.
\cr  \code{namec}   \tab   
\cr            \tab   
\cr  \code{namev}   \tab A vector.
\cr            \tab   
\cr  \code{namel}   \tab A \code{1}-element list.
\cr            \tab   
\cr    \code{rcn}   \tab A list of \code{2} vectors.
\cr            \tab   
\cr     \code{en}   \tab A character
\cr     \code{rn}   \tab vector.
\cr     \code{cn}   \tab   
}
}
\description{
\tabular{rl}{
  \code{getnames}   \tab Gets element, row, and/or column names with optional restrictions.
\cr             \tab   
\cr   \code{named}   \tab Evaluates whether an object's elements, rows, and/or columns are named, accounting for whether names must be unique, whether the object may contain empty elements, and whether names may be blank strings.
\cr             \tab   
\cr \code{rcnamed}   \tab Are both rows and columns named?
\cr  \code{enamed}   \tab Are elements of \code{x} named?
\cr  \code{rnamed}   \tab Are rows of \code{x} named?
\cr  \code{cnamed}   \tab Are columns of \code{x} named?
\cr  \code{dnamed}   \tab Are \code{...} arguments named?
\cr             \tab   
\cr  \code{namerc}   \tab Names rows and columns of \code{x}.
\cr   \code{namer}   \tab Names rows of \code{x}.
\cr   \code{namec}   \tab Names columns of \code{x}.
\cr   \code{namev}   \tab Names elements of a vector/\link[=ivls]{vlist}.
\cr   \code{namel}   \tab Creates and names a 1-element list.
\cr             \tab   
\cr     \code{rcn}   \tab Gets row and column names in a list.
\cr      \code{en}   \tab Gets element names.
\cr      \code{rn}   \tab Gets row names.
\cr      \code{cn}   \tab Get column names.
}
}
