% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\encoding{UTF-8}
\name{nth}
\alias{nth}
\alias{.nth}
\alias{.first_n}
\alias{.last_n}
\alias{.nth_from_last}
\alias{.nth_plus}
\title{Extract elements by numeric position}
\usage{
.nth(x, n)

.first_n(x, n)

.last_n(x, n)

.nth_from_last(x, n)

.nth_plus(x, n)
}
\arguments{
\item{x}{A \link[=pop_vec]{populated vector} to extract elements from.}

\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}.}
}
\value{
A scalar, vector, or \link[=VLS]{vlist}.
}
\description{
Extract the \code{n}-th, \code{n}-th from last, first \code{n}, and last \code{n} elements of a vector of list.
\tabular{ll}{  \code{.nth_from_last}   \tab Gets the \code{n}th from last element(s). \cr
  \code{.nth_plus}        \tab Gets element \code{n} and onwards.        \cr
  \code{.first_n}          \tab Gets the first \code{n} elements.         \cr
  \code{.last_n}           \tab Gets the last \code{n} elements.          \cr
  \code{.nth}             \tab Gets the \code{n}-th element(s).            }
}
\examples{
.nth_from_last(letters, 5)
.nth_plus(letters, 5)
.first_n(letters, 5)
.last_n(letters, 5)
.nth(letters, 5:7)
}
\seealso{
Other values: 
\code{\link{atomize_funs}()},
\code{\link{na}()},
\code{\link{values}()}
}
\concept{to_std_fun_form}
\concept{values}
