% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os}
\alias{os}
\title{Identify the current operating system.}
\usage{
os()
}
\value{
Either \code{mac}, \code{unix}, or \code{win}.
}
\description{
Identify the current operating system.
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}

Other environments: 
\code{\link{callers}()},
\code{\link{envir_vals}},
\code{\link{recycling}},
\code{\link{run}()}
}
\concept{environments}
\concept{extensions}
