% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os}
\alias{os}
\title{Identify the current operating system.}
\usage{
os()
}
\value{
A character scalar (either \code{mac}, \code{unix}, or \code{win}).
}
\description{
Identify the current operating system.
}
\examples{
os()
}
\seealso{
Other environments: 
\code{\link{callers}()},
\code{\link{envir_vals}},
\code{\link{recycling}},
\code{\link{run}()}
}
\concept{environments}
