% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{ox_vals}
\alias{ox_vals}
\title{Simple Oxford-Comma Lists}
\usage{
ox_vals(x, join)
}
\arguments{
\item{x}{A character vector to be formatted as an Oxford-comma list.}

\item{join}{A character scalar conjunction for between the next to last and last elements of \code{x}.}
}
\value{
A character scalar.
}
\description{
Simple Oxford-Comma Lists
}
\seealso{
Other utils: 
\code{\link{av}()},
\code{\link{callers}()},
\code{\link{caller}()},
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{fun_pkg_stack}()},
\code{\link{is_err}()},
\code{\link{meets_errs}()},
\code{\link{not_err}()},
\code{\link{spec2_combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()},
\code{\link{stack2pkgs}()}
}
\concept{utils}
