% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{ox_vals}
\alias{ox_vals}
\title{Simple Oxford-Comma Lists}
\usage{
ox_vals(x, join)
}
\arguments{
\item{x}{A character vector to be formatted as an Oxford-comma list.}

\item{join}{A character scalar conjunction for between the next to last and last elements of \code{x}.}
}
\value{
A character scalar.
egOxVals <- function() {
base::list(and.letters = uj::ox_vals(letters, "and"),
or.digits   = uj::ox_vals(0:9    , "or" ))
}

egOxVals()
}
\description{
Simple Oxford-Comma Lists
}
\seealso{
Other utils: 
\code{\link{check_valid}()},
\code{\link{combo2props}()},
\code{\link{compat}()},
\code{\link{meets_errs}()},
\code{\link{spec2combos}()},
\code{\link{spec2props}()},
\code{\link{stack2funs}()}
}
\concept{utils}
