% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\encoding{UTF-8}
\name{pals}
\alias{pals}
\alias{pal_colors}
\alias{pal_lines}
\alias{pal_azdots}
\alias{pal_AZdots}
\alias{pal_fdots}
\alias{pal_edots}
\alias{pal_ldots}
\alias{pal_ndots}
\alias{pal_pdots}
\alias{pal_sdots}
\title{Plotting palettes}
\usage{
pal_colors(N)

pal_lines(N)

pal_azdots(N)

pal_AZdots(N)

pal_fdots(N)

pal_edots(N)

pal_ldots(N)

pal_ndots(N)

pal_pdots(N)

pal_sdots(N)
}
\arguments{
\item{N}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the number of unique values to return.}
}
\value{
\strong{A positive whole number vector} \cr\cr \verb{pal_edots, pal_fdots}  \cr \verb{pal_ldots, pal_sdots}
\cr\cr  \strong{A character vector}             \cr\cr \verb{pal_colors, pal_lines} \cr \verb{pal_azdots, pal_AZdots} \cr \verb{pal_ndots, pal_pdots}
}
\description{
This family of functions uses \code{'dots'} to subsume plotting to \code{'points'}, \code{'shapes'}, and \code{'pch'}.
}
\details{
The functions in this family create palettes of various types from the first \code{N}  \eqn{^{(1)}} values of source vectors as defined below:
\tabular{lll}{  \emph{Function}     \tab \emph{Type}                    \tab \emph{Source}    \cr
  \code{pal_colors}   \tab colors                    \tab \code{v(colors)} \cr
  \code{pal_lines}    \tab line types                \tab \code{v(lines)}  \cr
  \code{pal_azdots}   \tab alpha dots                \tab \code{v(azdots)} \cr
  \code{pal_AZdots}   \tab ALPHA dots                \tab \code{v(AZdots)} \cr
  \code{pal_ndots}    \tab numeral dots              \tab \code{v(ndots)}  \cr
  \code{pal_pdots}    \tab punctuation dots          \tab \code{v(pdots)}  \cr
  \code{pal_ldots}    \tab line dots\eqn{^{(2)}}     \tab \code{v(ldots)}  \cr
  \code{pal_edots}    \tab empty dots\eqn{^{(3)}}    \tab \code{v(edots)}  \cr
  \code{pal_fdots}    \tab filled dots\eqn{^{(4)}}   \tab \code{v(fdots)}  \cr
  \code{pal_sdots}    \tab solid dots                \tab \code{v(sdots)}    }
\tabular{l}{  \eqn{^{(1)}} Recycled if \verb{N > length(<source>)}.  \cr
  \eqn{^{(2)}} Non-enclosing shape.                 \cr
  \eqn{^{(3)}} Enclosing shape with empty interior. \cr
  \eqn{^{(4)}} Enclosing shape with filled interior.  }
}
\seealso{
Other plots: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{markdown}()},
\code{\link{sca}()},
\code{\link{u2u}()}
}
\concept{plots}
