% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals..R
\name{pals}
\alias{pals}
\alias{pal_colors}
\alias{pal_lines}
\alias{pal_fdots}
\alias{pal_ldots}
\alias{pal_edots}
\alias{pal_sdots}
\title{Plotting palettes}
\usage{
pal_colors(n)

pal_lines(n)

pal_fdots(n)

pal_ldots(n)

pal_edots(n)

pal_sdots(n)
}
\arguments{
\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the number of unique values to return.}
}
\value{
\tabular{rl}{
  \code{pal_colors}   \tab A character vector.
\cr \code{pal_lines}   \tab A character vector.
\cr \code{pal_edots}   \tab A positive whole-number vector.
\cr \code{pal_fdots}   \tab A positive whole-number vector.
\cr \code{pal_ldots}   \tab A positive whole-number vector.
\cr \code{pal_sdots}   \tab A positive whole-number vector.
}
}
\description{
This family of functions refers to \code{points} and \code{shapes} as dots.
\cr\cr The functions in this family create palettes of various types from the first \code{n}\eqn{^1} values of source vectors as defined below:\tabular{rcl}{
  \strong{Function}   \tab \strong{ Type}      \tab   \strong{Source}
\cr    \code{pal_colors}   \tab colors              \tab   \code{v(colors)}
\cr     \code{pal_lines}   \tab line types          \tab   \code{v(lines)}
\cr     \code{pal_sdots}   \tab solid dots          \tab   \code{v(sdots)}
\cr     \code{pal_ldots}   \tab line dots\eqn{^2}   \tab   \code{v(ldots)}
\cr     \code{pal_edots}   \tab empty dots\eqn{^3}  \tab   \code{v(edots)}
\cr     \code{pal_fdots}   \tab filled dots\eqn{^4} \tab   \code{v(fdots)}
}
\eqn{^{1.}} Recycled if \verb{n > length(v(.))}.
\cr\eqn{^{2.}} Non-enclosing.
\cr\eqn{^{3.}} Enclosing with empty interior.
\cr\eqn{^{4.}} Enclosing with filled interior.
}
\seealso{
Other colors: 
\code{\link{color}},
\code{\link{mkd}}

Other plotting: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{sca}()},
\code{\link{u2u}()}
}
\concept{colors}
\concept{plotting}
