% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\name{pals}
\alias{pals}
\alias{pal_colors}
\alias{pal_lines}
\alias{pal_fdots}
\alias{pal_ldots}
\alias{pal_edots}
\alias{pal_sdots}
\title{Plotting palettes}
\usage{
pal_colors(n)

pal_lines(n)

pal_fdots(n)

pal_ldots(n)

pal_edots(n)

pal_sdots(n)
}
\arguments{
\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the number of unique values to return.}
}
\value{
\itemize{
\item \strong{\code{pal_colors}, \code{pal_lines}}: a character vector.
\item \strong{\code{pal_edots}, \code{pal_fdots}}: a positive whole-number vector.
\item \strong{\code{pal_ldots}, \code{pal_sdots}}: a positive whole-number vector.
}
}
\description{
\itemize{
\item \strong{\code{pal_colors}}: creates a color palette with the first \code{n} color values from \code{v(colors)}.
\item \strong{\code{pal_lines}}: creates a line-type palette with the first \code{n} line-type values from \code{v(lines)}.
\item \strong{\code{pal_edots}}: creates an empty-dots/shapes palette with the first \code{n} empty-dot values from \code{v(edots)} (empty dots are closed shapes with an unfilled interior).
\item \strong{\code{pal_fdots}}: creates a fillable-dots/shapes palette with the first \code{n} empty-dot values from \code{v(fdots)} (fillable dots are closed shapes with a fillable interior).
\item \strong{\code{pal_ldots}}: creates a line-only dots/shapes palette with the first \code{n} line-only dot values from \code{v(ldots)} (line-only dots are non-enclosing shapes).
\item \strong{\code{pal_sdots}}: creates a solid-dots/shapes palette with the first \code{n} solid-dot values from \code{v(sdots)}.
}
NOTE: Palette values are recycled if \code{n > length(v(.))}.
}
\seealso{
Other colors: 
\code{\link{color}},
\code{\link{mkd}}

Other plotting: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{sca}()},
\code{\link{u2u}()}
}
\concept{colors}
\concept{plotting}
