% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_palettes.R
\name{pals_uj}
\alias{pals_uj}
\alias{pal_colors}
\alias{pal_lines}
\alias{pal_fdots}
\alias{pal_ldots}
\alias{pal_edots}
\alias{pal_sdots}
\title{Plotting palettes}
\usage{
pals_uj()

pal_colors(n)

pal_lines(n)

pal_fdots(n)

pal_ldots(n)

pal_edots(n)

pal_sdots(n)
}
\arguments{
\item{n}{Number of unique values to return.}
}
\value{
Character vector
}
\description{
Plotting palettes
}
\section{Functions}{
\itemize{
\item \code{pal_colors}: Create palette with \code{n} colors from the default
color palette. Colors will be recycled if \code{n} is greater than
\code{length(v(colors))}.

\item \code{pal_lines}: Create palette with \code{n} line types from the default
line type palette. Line types will be recycled if \code{n} is greater than
\code{length(v(lines))}.

\item \code{pal_fdots}: Create palette with \code{n} fillable dots/shapes from
the default line type palette. Dots/shapes will be recycled if \code{n} is
greater than \code{length(v(fdots))}.

\item \code{pal_ldots}: Create palette with \code{n} line-based dots/shapes from
the default line type palette. Dots/shapes will be recycled if \code{n} is
greater than \code{length(v(ldots))}.

\item \code{pal_edots}: Create palette with \code{n} empty dots/shapes from the
default line type palette. Dots/shapes will be recycled if \code{n} is
greater than \code{length(v(edots))}.

\item \code{pal_sdots}: Create palette with \code{n} solid dots/shapes from the
default line type palette. Dots/shapes will be recycled if \code{n} is
greater than \code{length(v(sdots))}.
}}

\seealso{
Other plots: 
\code{\link{markdown_uj}()}
}
\concept{plots}
