% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\docType{data}
\name{pals_uj}
\alias{pals_uj}
\alias{pal_colors}
\alias{pal_lines}
\alias{pal_fdots}
\alias{pal_ldots}
\alias{pal_edots}
\alias{pal_sdots}
\title{Plotting Palettes}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
pals_uj

pal_colors(n)

pal_lines(n)

pal_fdots(n)

pal_ldots(n)

pal_edots(n)

pal_sdots(n)
}
\arguments{
\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar}
indicating the number of unique values to return.}
}
\value{
\tabular{ll}{
FUNCTIONS                   \tab RETURN VALUE                           \cr
\code{pal_colors}, \code{pal_lines}   \tab A character vector.                    \cr
\code{pal_edots}, \code{pal_fdots}    \tab A positive whole-number vector.        \cr
\code{pal_ldots}, \code{pal_sdots}    \tab A positive whole-number vector.          }
}
\description{
\tabular{ll}{
FUNCTION     \tab WHAT IT DOES                                          \cr
pal_colors   \tab Creates a color palette with the first \code{n} color values
from \code{v(colors)}.                                     \cr
pal_lines    \tab Creates a line-type palette with the first \code{n} line-type
values from \code{v(lines)}.                               \cr
pal_edots    \tab Creates an empty-dots/shapes palette with the first \code{n}
empty-dot values from \code{v(edots)}. Empty dots are closed
shapes with an unfilled interior.                     \cr
pal_fdots    \tab Creates a fillable-dots/shapes palette with the first \code{n}
empty-dot values from \code{v(fdots)}. Fillable dots are
closed shapes with a fillable interior.               \cr
pal_ldots    \tab Creates a line-only dots/shapes palette with the first
\code{n} line-only dot values from \code{v(ldots)}. Line-only dots
are non-enclosing shapes.                             \cr
pal_sdots    \tab Creates a solid-dots/shapes palette with the first first
\code{n} solid-dot values from \code{v(sdots)}.                   }
NOTE: Palette values are recycled if \code{n > length(v(.))}.
}
\seealso{
Other plots: 
\code{\link{mkd}}

Other colors: 
\code{\link{color}},
\code{\link{mkd}}
}
\concept{colors}
\concept{plots}
\keyword{datasets}
