% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{is_path}
\alias{as_path}
\alias{object_path}
\alias{parent_path}
\alias{object_name}
\alias{parent_name}
\alias{object_dirs}
\alias{parent_dirs}
\alias{newdirs}
\title{Evaluate and manipulate paths to files or folders}
\usage{
is_path(..., err = F)

as_path(..., err = T)

object_path(path, err = T)

parent_path(path, err = T)

object_name(path, err = T)

parent_name(path, err = T)

object_dirs(path, err = T)

parent_dirs(path, err = T)

newdirs(dirs, path = NULL)
}
\arguments{
\item{...}{Atomic arguments pasted to create}

\item{err}{A non-\code{NA} logical scalar indicating whether an error should be thrown if \code{...} is not a valid object path when collapsed to a character scalar.}

\item{path}{A \link[=cmp_chr_scl]{complete character scalar} path to either a directory or a file.}

\item{dirs}{A \link[=cmp_chr_vec]{complete character vec} of new sub-directory names.}
}
\value{
\tabular{rl}{
  \code{object_path} \tab   A character scalar.
\cr \code{folder_path} \tab   A character scalar.
\cr \code{object_name} \tab   A character scalar.
\cr \code{folder_name} \tab   A character scalar.
\cr \code{object_dirs} \tab   A character vector.
\cr \code{folder_dirs} \tab   A character vector.
\cr     \code{is_path} \tab   A logical scalar.
}
}
\description{
\tabular{rl}{
  \code{parent_dirs}   \tab Calls \code{parent_path} and split the resulting character scalar into a character vector containing the name of the root folder, the names of any intermediate folders, and the name of the parent folder of the object specified by \code{...}.
\cr                 \tab   
\cr \code{object_path}   \tab Collapses \code{...} into a character scalar and expand the path to the object indicated by the result to account for relative paths.
\cr                 \tab   
\cr \code{parent_path}   \tab Calls \code{object_path} and extract from the resulting character scalar just the path to the parent folder of the object specified by \code{...} (i.e., discarding the name of the object itself).
\cr                 \tab   
\cr \code{object_name}   \tab Calls \code{object_path} and extract from the resulting character scalar just the name of the object, discarding the path to its parent.
\cr                 \tab   
\cr \code{parent_name}   \tab Calls \code{parent_path} and extract from the resulting character scalar just the name of last folder in the path.
\cr                 \tab   
\cr \code{object_dirs}   \tab Calls \code{object_path} and split the resulting character scalar into a character vector containing the name of the root folder, the names of any intermediate folders, the name of the parent folder of the object, and the name of the object.
\cr                 \tab   
\cr    \code{new_dirs}   \tab Creates sub-directories within an existing directory, optionally asking the user to choose that existing directory.
\cr                 \tab   
\cr     \code{is_path}   \tab Checks whether \code{...} resolves to a valid path for an object (file or folder) when \code{...} arguments are collapsed into a character scalar.
\cr                 \tab   
\cr     \code{as_path}   \tab Collapses \code{...} into a path using the current platform file path separator \code{.Platform$file.sep}.
}
}
\details{
In most cases, if (optionally) specified, functions throw an error if \code{...} does not resolve to a valid object path.
}
\examples{
path. <- path.expand("~")
parts. <- ss(.Platform$file.sep, path.)
parts. <- parts.[parts. != ""]
parts.[1] <- paste0(.Platform$file.sep, parts.[1])
parts. <- paste0("'", parts., "'")
code. <- paste0(parts., collapse = ", ")
is_path. <- paste0("is_path(", code., ")")
as_path. <- paste0("as_path(", code., ")")

path.
parts.
is_path.
as_path.

parent_dirs(path.)
object_dirs(path.)

parent_path(path.)
object_path(path.)

parent_name(path.)
object_name(path.)

run(is_path.)
run(as_path.)
}
\seealso{
Other files: 
\code{\link{rd}}
}
\concept{files}
