% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\encoding{UTF-8}
\name{paths}
\alias{paths}
\alias{is_path}
\alias{as_path}
\alias{object_dirs}
\alias{parent_dirs}
\alias{object_path}
\alias{parent_path}
\alias{object_name}
\alias{parent_name}
\alias{new_dirs}
\title{Evaluate and manipulate paths to files or folders}
\usage{
is_path(..., Err = F)

as_path(..., Err = T)

object_dirs(Path, Err = T)

parent_dirs(Path, Err = T)

object_path(Path, Err = T)

parent_path(Path, Err = T)

object_name(Path, Err = T)

parent_name(Path, Err = T)

new_dirs(Dirs, Path = NULL)
}
\arguments{
\item{...}{Atomic arguments pasted to create}

\item{Err}{\code{TRUE} or \code{FALSE} indicating whether an error should be thrown if \code{...} is not a valid object path when collapsed to a character scalar.}

\item{Path}{A \link[=cmp_chr_scl]{complete character scalar} path to either a directory or a file.}

\item{Dirs}{A \link[=cmp_chr_vec]{complete character vec} of new sub-directory names.}
}
\value{
\strong{A character scalar} \cr\cr \verb{object_dirs, parent_path} \cr \verb{object_name, parent_name} \cr \code{as_path}
\cr\cr  \strong{A character vector} \cr\cr \verb{object_dirs, parent_dirs}
\cr\cr  \strong{A logical scalar}   \cr\cr \code{is_path}
}
\description{
Take an file reference and get its directories, path, and name; its parent's directories, path, and name. Convert a \code{...} args to a path, check the validity of a path, and create new directories within a given directory.
}
\details{
\tabular{ll}{  \code{parent_dirs}   \tab Calls \code{parent_path} and split the resulting character scalar into a character vector containing the name of the root folder, the names
of any intermediate folders, and the name of the parent folder of the object specified by \code{...}.                                         \cr   \tab   \cr
  \code{parent_path}   \tab Calls \code{object_path} and extract from the resulting character scalar just the path to the parent folder of the object specified by \code{...}
(i.e., discarding the name of the object itself).                                                                                        \cr   \tab   \cr
  \code{parent_name}   \tab Calls \code{parent_path} and extract from the resulting character scalar just the name of last folder in the path.                            \cr   \tab   \cr
  \code{object_path}   \tab Collapses \code{...} into a character scalar and expand the path to the object indicated by the result to account for relative paths.         \cr   \tab   \cr
  \code{object_dirs}   \tab Calls \code{object_path} and split the resulting character scalar into a character vector containing the name of the root folder, the names
of any intermediate folders, the name of the parent folder of the object, and the name of the object.                                    \cr   \tab   \cr
  \code{object_name}   \tab Calls \code{object_path} and extract from the resulting character scalar just the name of the object, discarding the path to its parent.      \cr   \tab   \cr
  \code{new_dirs}      \tab Creates sub-directories within an existing directory, optionally asking the user to choose that existing directory.                      \cr   \tab   \cr
  \code{is_path}       \tab Checks whether \code{...} resolves to a valid path for an object (file or folder) when \code{...} arguments are collapsed into a character scalar. \cr   \tab   \cr
  \code{as_path}       \tab Collapses \code{...} into a path using the current platform file path separator \code{.Platform$file.sep}.                                                        }

In most cases, if (optionally) specified, functions throw an error if \code{...} does not resolve to a valid object path.
}
\examples{
path. <- path.expand("~")
parts. <- ss(.Platform$file.sep, path.)
parts. <- parts.[parts. != ""]
parts.[1] <- paste0(.Platform$file.sep, parts.[1])
parts. <- paste0("'", parts., "'")
code. <- paste0(parts., collapse = ", ")
is_path. <- paste0("is_path(", code., ")")
as_path. <- paste0("as_path(", code., ")")

path.
parts.
is_path.
as_path.

parent_dirs(path.)
object_dirs(path.)
parent_path(path.)
object_path(path.)
parent_name(path.)
object_name(path.)
run(is_path.)
run(as_path.)
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{purge}()},
\code{\link{rd}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}

Other files: 
\code{\link{rd}()}
}
\concept{extensions}
\concept{files}
