% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\title{Pause for a specified number of seconds to give time for screen
updates.}
\usage{
pause(x = 1e-07)
}
\arguments{
\item{x}{Numeric scalar number of seconds to pause.}
}
\value{
\code{NULL} (called for side effect)
}
\description{
Pause for a specified number of seconds to give time for screen
updates.
}
