% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_fast.R
\name{ppp_fast}
\alias{ppp_fast}
\alias{.ATM}
\alias{.DEF}
\alias{.FUN}
\alias{.NIL}
\alias{.NLL}
\alias{.POP}
\alias{.RCR}
\alias{.ARR}
\alias{.DTF}
\alias{.GEN}
\alias{.MAT}
\alias{.MVC}
\alias{.SCL}
\alias{.VLS}
\alias{.VEC}
\alias{.D0D}
\alias{.D1D}
\alias{.D2D}
\alias{.DHD}
\alias{.EUD}
\alias{.E2D}
\alias{.EHD}
\alias{.E0D}
\alias{.E1D}
\alias{.NA0}
\alias{.OK0}
\alias{.DUP}
\alias{.UNQ}
\alias{.CMP}
\alias{.MSS}
\alias{.PRT}
\alias{.CH1}
\alias{.CH3}
\alias{.CHR}
\alias{.STR}
\alias{.CLR}
\alias{.FAC}
\alias{.ORD}
\alias{.UNO}
\alias{.IND}
\alias{.NST}
\alias{.SRT}
\alias{.LGL}
\alias{.NUM}
\alias{.WHL}
\alias{.ODD}
\alias{.EVN}
\alias{.NGW}
\alias{.NPW}
\alias{.PSW}
\alias{.FRC}
\alias{.NEG}
\alias{.POS}
\alias{.NNW}
\alias{.NPS}
\alias{.PCT}
\alias{.PPN}
\alias{.COL}
\alias{.ROW}
\alias{.PNT}
\alias{.SLD}
\alias{.SQR}
\alias{.EMP}
\alias{.RCT}
\alias{.LIN}
\alias{.atm_dtf}
\alias{.atm_mat}
\alias{.atm_scl}
\alias{.atm_vec}
\alias{.atm_vls}
\alias{.chr_arr}
\alias{.chr_dtf}
\alias{.chr_vec}
\alias{.chr_vls}
\alias{.cmp_atm}
\alias{.cmp_chr}
\alias{.cmp_num}
\alias{.cmp_scl}
\alias{.cmp_vec}
\alias{.lgl_scl}
\alias{.nnw_vec}
\alias{.pop_atm}
\alias{.pop_chr}
\alias{.pop_dtf}
\alias{.pop_mat}
\alias{.pop_srt}
\alias{.pop_vec}
\alias{.pop_vls}
\alias{.cmp_atm_vec}
\alias{.cmp_ch1_scl}
\alias{.cmp_ch1_vec}
\alias{.cmp_ch3_scl}
\alias{.cmp_ch3_vec}
\alias{.cmp_chr_arr}
\alias{.cmp_chr_vec}
\alias{.cmp_chr_vls}
\alias{.cmp_chr_gen}
\alias{.cmp_chr_scl}
\alias{.cmp_clr_vec}
\alias{.cmp_lgl_scl}
\alias{.cmp_lgl_vec}
\alias{.cmp_nng_scl}
\alias{.cmp_nnw_scl}
\alias{.cmp_nnw_vec}
\alias{.cmp_num_scl}
\alias{.cmp_num_vec}
\alias{.cmp_pos_vec}
\alias{.cmp_ppn_vec}
\alias{.cmp_psw_scl}
\alias{.cmp_psw_vec}
\alias{.cmp_srt_vec}
\alias{.cmp_str_scl}
\alias{.cmp_str_vec}
\alias{.cmp_whl_scl}
\alias{.cmp_atm_mvc}
\alias{.unq_atm_mvc}
\alias{.unq_atm_vec}
\alias{.unq_ch1_vec}
\alias{.unq_ch3_vec}
\alias{.unq_chr_vec}
\alias{.unq_nnw_vec}
\alias{.unq_num_vec}
\alias{.unq_psw_vec}
\alias{.unq_str_vec}
\alias{.atm_rct_dtf}
\title{Fast property checking for common checks}
\usage{
ppp_fast()

.ATM()

.DEF(x)

.FUN(x)

.NIL(x)

.NLL(x)

.POP(x)

.RCR(x)

.ARR(x)

.DTF(x)

.GEN(x)

.MAT(x)

.MVC(x)

.SCL(x)

.VLS(x)

.VEC(x)

.D0D(x)

.D1D(x)

.D2D(x)

.DHD(x)

.EUD(x)

.E2D(x)

.EHD(x)

.E0D(x)

.E1D(x)

.NA0(x)

.OK0(x)

.DUP(x)

.UNQ(x)

.CMP(x)

.MSS(x)

.PRT(x)

.CH1(x)

.CH3(x)

.CHR(x)

.STR(x)

.CLR(x)

.FAC(x)

.ORD(x)

.UNO(x)

.IND(x)

.NST(x)

.SRT(x)

.LGL(x)

.NUM(x)

.WHL(x)

.ODD(x)

.EVN(x)

.NGW(x)

.NPW(x)

.PSW(x)

.FRC(x)

.NEG(x)

.POS(x)

.NNW(x)

.NPS(x)

.PCT(x)

.PPN(x)

.COL(x)

.ROW(x)

.PNT(x)

.SLD(x)

.SQR(x)

.EMP(x)

.RCT(x)

.LIN(x)

.atm_dtf(x)

.atm_mat(x)

.atm_scl(x)

.atm_vec(x)

.atm_vls(x)

.chr_arr(x)

.chr_dtf(x)

.chr_vec(x)

.chr_vls(x)

.cmp_atm(x)

.cmp_chr(x)

.cmp_num(x)

.cmp_scl(x)

.cmp_vec(x)

.lgl_scl(x)

.nnw_vec(x)

.pop_atm(x)

.pop_chr(x)

.pop_dtf(x)

.pop_mat(x)

.pop_srt(x)

.pop_vec(x)

.pop_vls(x)

.cmp_atm_vec(x)

.cmp_ch1_scl(x)

.cmp_ch1_vec(x)

.cmp_ch3_scl(x)

.cmp_ch3_vec(x)

.cmp_chr_arr(x, .VALID = NULL)

.cmp_chr_vec(x, .VALID = NULL)

.cmp_chr_vls(x, .VALID = NULL)

.cmp_chr_gen(x, .VALID = NULL)

.cmp_chr_scl(x, .VALID = NULL)

.cmp_clr_vec(x)

.cmp_lgl_scl(x)

.cmp_lgl_vec(x)

.cmp_nng_scl(x)

.cmp_nnw_scl(x)

.cmp_nnw_vec(x)

.cmp_num_scl(x, .VALID = NULL)

.cmp_num_vec(x, .VALID = NULL)

.cmp_pos_vec(x)

.cmp_ppn_vec(x)

.cmp_psw_scl(x)

.cmp_psw_vec(x)

.cmp_srt_vec(x)

.cmp_str_scl(x)

.cmp_str_vec(x)

.cmp_whl_scl(x)

.cmp_atm_mvc(x)

.unq_atm_mvc(x)

.unq_atm_vec(x)

.unq_ch1_vec(x)

.unq_ch3_vec(x)

.unq_chr_vec(x, .VALID = NULL)

.unq_nnw_vec(x)

.unq_num_vec(x, .VALID = NULL)

.unq_psw_vec(x)

.unq_str_vec(x)

.atm_rct_dtf(x)
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Fast property checking for common checks
}
\details{
Fast property checking for \code{\link{ppp}} functions without the dot prefix.
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
