% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_fast.R
\name{ppp_fast}
\alias{ppp_fast}
\alias{.ATM}
\alias{.DEF}
\alias{.FUN}
\alias{.NIL}
\alias{.NLL}
\alias{.POP}
\alias{.RCR}
\alias{.ARR}
\alias{.DTF}
\alias{.GEN}
\alias{.MAT}
\alias{.MVC}
\alias{.SCL}
\alias{.VLS}
\alias{.VEC}
\alias{.D0D}
\alias{.D1D}
\alias{.D2D}
\alias{.DHD}
\alias{.EUD}
\alias{.E2D}
\alias{.EHD}
\alias{.E0D}
\alias{.E1D}
\alias{.NA0}
\alias{.OK0}
\alias{.DUP}
\alias{.UNQ}
\alias{.CMP}
\alias{.MSS}
\alias{.PRT}
\alias{.CH1}
\alias{.CH3}
\alias{.CHR}
\alias{.STR}
\alias{.CLR}
\alias{.FAC}
\alias{.ORD}
\alias{.UNO}
\alias{.IND}
\alias{.NST}
\alias{.SRT}
\alias{.LGL}
\alias{.NUM}
\alias{.WHL}
\alias{.ODD}
\alias{.EVN}
\alias{.NGW}
\alias{.NPW}
\alias{.PSW}
\alias{.FRC}
\alias{.NEG}
\alias{.POS}
\alias{.NNW}
\alias{.NPS}
\alias{.PCT}
\alias{.PPN}
\alias{.COL}
\alias{.ROW}
\alias{.PNT}
\alias{.SLD}
\alias{.SQR}
\alias{.EMP}
\alias{.RCT}
\alias{.LIN}
\alias{.atm_dtf}
\alias{.atm_mat}
\alias{.atm_scl}
\alias{.atm_vec}
\alias{.atm_vls}
\alias{.chr_arr}
\alias{.chr_dtf}
\alias{.chr_vec}
\alias{.chr_vls}
\alias{.cmp_atm}
\alias{.cmp_chr}
\alias{.cmp_num}
\alias{.cmp_scl}
\alias{.cmp_vec}
\alias{.lgl_scl}
\alias{.nnw_vec}
\alias{.pop_atm}
\alias{.pop_chr}
\alias{.pop_dtf}
\alias{.pop_mat}
\alias{.pop_srt}
\alias{.pop_vec}
\alias{.pop_vls}
\alias{.cmp_atm_vec}
\alias{.cmp_ch1_scl}
\alias{.cmp_ch1_vec}
\alias{.cmp_ch3_scl}
\alias{.cmp_ch3_vec}
\alias{.cmp_chr_arr}
\alias{.cmp_chr_vec}
\alias{.cmp_chr_vls}
\alias{.cmp_chr_gen}
\alias{.cmp_chr_scl}
\alias{.cmp_clr_vec}
\alias{.cmp_lgl_scl}
\alias{.cmp_lgl_vec}
\alias{.cmp_nng_scl}
\alias{.cmp_nnw_scl}
\alias{.cmp_nnw_vec}
\alias{.cmp_num_scl}
\alias{.cmp_num_vec}
\alias{.cmp_pos_vec}
\alias{.cmp_ppn_vec}
\alias{.cmp_psw_scl}
\alias{.cmp_psw_vec}
\alias{.cmp_srt_vec}
\alias{.cmp_str_scl}
\alias{.cmp_str_vec}
\alias{.cmp_whl_scl}
\alias{.cmp_atm_mvc}
\alias{.unq_atm_mvc}
\alias{.unq_atm_vec}
\alias{.unq_ch1_vec}
\alias{.unq_ch3_vec}
\alias{.unq_chr_vec}
\alias{.unq_nnw_vec}
\alias{.unq_num_vec}
\alias{.unq_psw_vec}
\alias{.unq_str_vec}
\title{Fast property checking for common checks}
\usage{
ppp_fast()

.ATM(X)

.DEF(X)

.FUN(X)

.NIL(X)

.NLL(X)

.POP(X)

.RCR(X)

.ARR(X)

.DTF(X)

.GEN(X)

.MAT(X)

.MVC(X)

.SCL(X)

.VLS(X)

.VEC(X)

.D0D(X)

.D1D(X)

.D2D(X)

.DHD(X)

.EUD(X)

.E2D(X)

.EHD(X)

.E0D(X)

.E1D(X)

.NA0(X)

.OK0(X)

.DUP(X)

.UNQ(X)

.CMP(X)

.MSS(X)

.PRT(X)

.CH1(X)

.CH3(X)

.CHR(X)

.STR(X)

.CLR(X)

.FAC(X)

.ORD(X)

.UNO(X)

.IND(X)

.NST(X)

.SRT(X)

.LGL(X)

.NUM(X)

.WHL(X)

.ODD(X)

.EVN(X)

.NGW(X)

.NPW(X)

.PSW(X)

.FRC(X)

.NEG(X)

.POS(X)

.NNW(X)

.NPS(X)

.PCT(X)

.PPN(X)

.COL(X)

.ROW(X)

.PNT(X)

.SLD(X)

.SQR(X)

.EMP(X)

.RCT(X)

.LIN(X)

.atm_dtf(X)

.atm_mat(X)

.atm_scl(X)

.atm_vec(X)

.atm_vls(X)

.chr_arr(X)

.chr_dtf(X)

.chr_vec(X)

.chr_vls(X)

.cmp_atm(X)

.cmp_chr(X)

.cmp_num(X)

.cmp_scl(X)

.cmp_vec(X)

.lgl_scl(X)

.nnw_vec(X)

.pop_atm(X)

.pop_chr(X)

.pop_dtf(X)

.pop_mat(X)

.pop_srt(X)

.pop_vec(X)

.pop_vls(X)

.cmp_atm_vec(X)

.cmp_ch1_scl(X)

.cmp_ch1_vec(X)

.cmp_ch3_scl(X)

.cmp_ch3_vec(X)

.cmp_chr_arr(X, Valid = NULL)

.cmp_chr_vec(X, Valid = NULL)

.cmp_chr_vls(X, Valid = NULL)

.cmp_chr_gen(X, Valid = NULL)

.cmp_chr_scl(X, Valid = NULL)

.cmp_clr_vec(X)

.cmp_lgl_scl(X)

.cmp_lgl_vec(X)

.cmp_nng_scl(X)

.cmp_nnw_scl(X)

.cmp_nnw_vec(X)

.cmp_num_scl(X, Valid = NULL)

.cmp_num_vec(X, Valid = NULL)

.cmp_pos_vec(X)

.cmp_ppn_vec(X)

.cmp_psw_scl(X)

.cmp_psw_vec(X)

.cmp_srt_vec(X)

.cmp_str_scl(X)

.cmp_str_vec(X)

.cmp_whl_scl(X)

.cmp_atm_mvc(X)

.unq_atm_mvc(X)

.unq_atm_vec(X)

.unq_ch1_vec(X)

.unq_ch3_vec(X)

.unq_chr_vec(X, Valid = NULL)

.unq_nnw_vec(X)

.unq_num_vec(X, Valid = NULL)

.unq_psw_vec(X)

.unq_str_vec(X)
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Fast property checking for common checks
}
\details{
Fast property checking for \code{\link{ppp}} functions without the dot prefix.
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{stopperr}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
