% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\encoding{UTF-8}
\name{purge}
\alias{purge}
\alias{X}
\alias{purge_plot}
\alias{rm_plot}
\alias{xplot}
\alias{purge_plots}
\alias{rm_plots}
\alias{xplots}
\alias{purge_warnings}
\alias{purge_warn}
\alias{xwarnings}
\alias{xwarn}
\alias{purge_console}
\alias{purge_con}
\alias{xconsole}
\alias{xcon}
\alias{purge_pref}
\alias{rm_pref}
\alias{xpref}
\alias{purge_global}
\alias{purge_all}
\alias{rm_global}
\alias{rm_all}
\alias{xglobal}
\alias{xall}
\title{Purge objects and parts of objects}
\usage{
purge(...)

X(...)

purge_plot(X = NULL)

rm_plot(X = NULL)

xplot(X = NULL)

purge_plots(X = NULL)

rm_plots(X = NULL)

xplots(X = NULL)

purge_warnings()

purge_warn()

xwarnings()

xwarn()

purge_console()

purge_con()

xconsole()

xcon()

purge_pref(Funs)

rm_pref(Funs)

xpref(Funs)

purge_global()

purge_all()

rm_global()

rm_all()

xglobal()

xall()
}
\arguments{
\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs} containing names of objects to purge from the global environment.}

\item{X}{Anticipated to be a plot object, but any object can be handled.}

\item{Funs}{\link[=cmp_chr_vec]{A complete character vec} of function names, possibly with package prefixes.}
}
\value{
\strong{A character vector}    \cr\cr \verb{purge_pref, rm_pref, xpref}
\cr\cr  \strong{The} \code{NULL} \strong{object} \cr\cr all others
}
\description{
Purge objects and parts of objects
}
\details{
All purge (\code{X}) functions include garbage collection.
\tabular{ll}{  \verb{purge_pref, rm__pref, xpref}   \tab Purge package prefixes\eqn{^{(1)}}    \cr
  \verb{purge_plot, rm_plot, xplot}    \tab Purge plots\eqn{^{(2)}}               \cr
  \verb{purge_warn, rm_warn, xwarn}    \tab Purge warnings                        \cr
  \verb{purge_con, rm_con, xcon}       \tab Purge console                         \cr
  \verb{purge, X}                      \tab Purge specific objects                  }
\tabular{l}{  \eqn{^{(1)}} Converts \code{'base::paste'} to \code{'paste'}, for example.           \cr
  \eqn{^{(2)}} Followed by printing the object \code{X} if not \code{NULL}.            \cr   }
\tabular{ll}{  \verb{purge_global, purge_all}      \tab \strong{WARNING: Use with care!}            \cr
  \verb{rm_global, rm_all}            \tab    *These functions purge *            \cr
  \verb{xglobal, xall}                \tab    \emph{the global environment}              }
}
\examples{
egFunNames <- c("base::c", "uj::cmp_chr_scl", "run")
xpref(egFunNames)
\dontrun{
  ## build a ggplot scatterplot object (y.)
  library(ggplot2)
  egObj <- data.frame(X = 0:100, y = sqrt(0:100)
  egGgp <- ggplot(egObj, aes(X, y)) + geom_point()

  ## print to console, wait for user, purge console
  say("\n A message")
  continue()
  xcon()

  ## display a scatterplot, wait for user
  plot(egObj$X, egObj$y)
  continue()

  ## purge plots and print ggplot
  xplot(egGgp)

  ## wait for user, purge plots again
  continue()
  xplot()

  ## generate warning, print warnings
  sqrt(-1)
  warnings()

  ## purge warnings, validate that warnings are purged
  xwarn()
  warnings()
}
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}

Other meta: 
\code{\link{callers}()},
\code{\link{dot_args}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{recyclable_ns}()},
\code{\link{run}()},
\code{\link{say}()}

Other purge: 
\code{\link{rm_vals}()}
}
\concept{extensions}
\concept{meta}
\concept{purge}
