% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put.}
\alias{put.}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put values into objects}
\usage{
put.()

put(x, v, e)

putr(x, v, r)

putc(x, v, c)
}
\arguments{
\item{x}{Object into which values will be placed. May be a \code{\link{vec}},
\code{\link[ivls]{vlist}}, matrix, or data.frame.}

\item{v}{A vector of replacement values.}

\item{e}{A \link[=cmp_ind_vec]{complete indexer vec} or
\link[=cmp_psw_scl]{complete positive whole-number scalar} indexing
element(s) to replace.}

\item{r, c}{A \link[=cmp_ind_scl]{complete indexer scalar} indicating row or
column to replace, respectively.}
}
\value{
An object of the same dimension as \code{x}.
}
\description{
\tabular{ll}{
\code{put}    \tab Put values in \code{v} into \code{e}-indexed elements of
the vector or \code{\link[=ivls]{vlist}} \code{x}.   \cr
\code{putr}   \tab Put the vector \code{v} into the \code{r}-indexed row of
the matrix or data frame \code{x}.                   \cr
\code{putc}   \tab Put the vector \code{v} into the \code{c}-indexed of the
matrix or data frame \code{x}.                         }
}
\seealso{
Other extensions: 
\code{\link{a.}()},
\code{\link{binfuns.}()},
\code{\link{callers.}()},
\code{\link{case.}()},
\code{\link{dots.}()},
\code{\link{ex.}()},
\code{\link{failsafe.}()},
\code{\link{files.}()},
\code{\link{fork.}()},
\code{\link{fsub.}()},
\code{\link{is_failsafe.}()},
\code{\link{is_unq.}()},
\code{\link{logicals.}()},
\code{\link{make.}()},
\code{\link{n_th.}()},
\code{\link{na.}()},
\code{\link{naming.}()},
\code{\link{nx.}()},
\code{\link{os.}()},
\code{\link{pause.}()},
\code{\link{ply.}()},
\code{\link{reclass.}()},
\code{\link{recycling.}()},
\code{\link{run.}()},
\code{\link{stats.}()},
\code{\link{swap.}()},
\code{\link{values.}()},
\code{\link{x.}()},
\code{\link{xb.}()}
}
\concept{extensions}
