% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\encoding{UTF-8}
\name{put}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put values into objects}
\usage{
put(X, V, E)

putr(X, V, R)

putc(X, V, C)
}
\arguments{
\item{X}{For \code{put}, a vector or \link[=ccc]{vlist}. For \code{putc} and \code{putr}, a data.frame or matrix.}

\item{V}{A vector of replacement values. Must be of the same length as \code{E} for \code{put}. Length must equal the number of \code{nrow(X)} for \code{putc}. Length must equal \code{ncol(X)} for \code{putr}}

\item{E}{A \link[=cmp_ind_vec]{complete indexer vec} of length equal to that of \code{V}.}

\item{R, C}{\link[=cmp_ind_scl]{Complete indexer scalars} indicating row or column to replace, respectively.}
}
\value{
An object of the same class and dimension as \code{X}.
}
\description{
Put values into a vector/list, vectors into rows, and vectors into columns.
}
\details{
\tabular{ll}{  \code{put}    \tab Puts vector \code{V} into \code{E}-indexed elements of \code{X} (assumed to be a vector or \link[=VLS]{vlist}). \cr   \tab   \cr
  \code{putc}   \tab Puts vector \code{V} into \code{C}-indexed column of \code{X} (assumed to be a data.frame or matrix).           \cr   \tab   \cr
  \code{putr}   \tab Puts vector \code{V} into \code{R}-indexed row of \code{X} (assumed to be a data.frame or matrix).                             }
}
\examples{
egVec <- letters[1:5]
egVls <- list(a = letters[1:5], b = 1:5)
egMat <- matrix(1:25, nrow = 5)
egDtf <- data.frame(a = letters[1:5], b = 1:5, stringsAsFactors = FALSE)

egVec
egVls
egMat
egDtf

put(egVec, "z", 1:2)
put(egVls, 7, 2)
putr(egMat, 1:5, 1)
putc(egDtf, 1:5, 1)
}
\seealso{
Other function_form: 
\code{\link{insert_before_pos}()},
\code{\link{move_pos}()}
}
\concept{function_form}
