% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put Values into Objects}
\usage{
put(x, rep, elt)

putr(x, rep, row)

putc(x, rep, col)
}
\arguments{
\item{x}{Object into which values will be placed. May be a vector, list,
matrix, or data.frame.}

\item{rep}{Vector of replacement values.}

\item{elt}{Logical vector, integer vector, or integer scalar indexing
element(s) to replace.}

\item{row}{Logical vector or integer scalar indexing row to replace.}

\item{col}{Logical vector or integer scalar indexing column to replace.}
}
\value{
An object of the same dimension as \code{x}.
}
\description{
Place values into indexed elements of vector or list \code{x}.
}
\section{Functions}{
\itemize{
\item \code{putr}: Place vector into indexed row of matrix or data frame
\code{x}.

\item \code{putc}: Place vector into indexed col of matrix or data frame
\code{x}.
}}

