% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put}
\alias{put}
\alias{putr}
\alias{putc}
\title{Put values into objects}
\usage{
put(x, v, e)

putr(x, v, r)

putc(x, v, c)
}
\arguments{
\item{x}{For \code{put}, a vector or \link[=ccc]{vlist}. For \code{putc} and \code{putr}, a data.frame or matrix.}

\item{v}{A vector of replacement values. Must be of the same length as \code{e} for \code{put}. Length must equal the number of \code{nrow(x)} for \code{putc}. Length must equal the \code{ncol(x)} for \code{putr}}

\item{e}{A \link[=cmp_ind_vec]{complete indexer vec} of length equal to that of \code{v}.}

\item{r, c}{\link[=cmp_ind_scl]{Complete indexer scalars} indicating row or column to replace, respectively.}
}
\value{
An object of the same class and dimension as \code{x}.
}
\description{
(These functions are not error checked.)
\itemize{
\item \strong{\code{put}}: puts values of \code{v} into \code{e}-indexed elements of \code{x} (assumed to be a vector or \link[=ccc]{vlist}).
\item \strong{\code{putc}}: puts the vector \code{v} into the \code{c}-indexed column of \code{x} (assumed to be a data.frame or matrix).
\item \strong{\code{putr}}: puts the vector \code{v} into the \code{r}-indexed row of \code{x} (assumed to be a data.frame or matrix).
}
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
