% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r.}
\alias{r.}
\alias{r}
\alias{e}
\alias{re}
\title{Error-Checked Wraps of \code{\link[=base:rep]{base::rep}}.}
\usage{
r.()

r(r, ...)

e(e, ...)

re(e, r, ...)

re(e, r, ...)
}
\arguments{
\item{r}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving
the number of replications of entire vectors.}

\item{...}{One or more atomic vectors to be replicated, reduced to unique
values, or reduced to unique duplicated values.}

\item{e}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving
the number of times each value is to be replicated in place before
replication of the entire vector.}
}
\value{
An atomic vector.
}
\description{
\tabular{ll}{
\code{r}            \tab   \tab Atomizes \code{...} and, with the resulting
vector \code{dots}, calls
\code{base::rep.int(dots, r)}.            \cr
\code{e}            \tab   \tab Atomizes \code{...} and, with the resulting
vector \code{dots}, calls
\code{base::rep(dots, each = e)}.         \cr
\code{er}, \code{re}\tab   \tab Both functions atomize \code{...}, and, with
the resulting vector, calls
\code{base::rep(dots, times = r, each = e)}.}
\tabular{ll}{
NOTE \tab Although \code{er} and \code{re} produce identical results when
their arguments are the same, both are included for convenience.  }
}
\section{Comparing Function Output}{
 The following console excerpt
demonstrates how each function performs.

\if{html}{\out{<div class="sourceCode">}}\preformatted{> list(r  =  r(2,    0:1, 9),
+      e  =  e(3,    0:1, 9),
+      re = re(2, 3, 0:1, 9),
+      er = er(3, 2, 0:1, 9))
$r
[1] 0 1 9 0 1 9

$e
[1] 0 0 0 1 1 1 9 9 9

$re
 [1] 0 0 0 1 1 1 9 9 9 0 0 0 1 1 1 9 9 9

$er
 [1] 0 0 1 1 9 9 0 0 1 1 9 9 0 0 1 1 9 9
}\if{html}{\out{</div>}}
}

\examples{
r(2, 0:4, 5:9)
e(3, 0:4, 5:9)
re(2, 3, 0:4, 5:9)
er(3, 2, 0:4, 5:9)
}
\concept{extended}
