% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r}
\alias{r}
\alias{e}
\alias{re}
\alias{er}
\title{Error-checked wrappers for \code{base::rep}}
\usage{
r(r, ...)

e(e, ...)

re(r, e, ...)

er(e, r, ...)
}
\arguments{
\item{r}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the number of replications of entire vectors.}

\item{...}{One or more atomic vectors to be replicated (\link[=a]{atomized} before replication).}

\item{e}{A complete positive whole-number scalar giving the number of times each value is to be replicated in place.}
}
\value{
An atomic vector.
}
\description{
These functions \link[=av]{atomize} \code{...} arguments, converting them to an atomic vector (represented here as \code{dots}) before replication.
\tabular{rl}{
\verb{re,er} \tab   Calls \verb{rep(dots, times = r, each = e)}.
\cr  \code{e} \tab   Calls \code{rep.int(dots, each = e)}.
\cr  \code{r} \tab   Calls \code{rep.int(dots, r)}.
}
}
\examples{
r(2, 0:2, 3)
e(3, 0:2, 3)
re(2, 3, 0:2, 3)
}
