% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r}
\alias{r}
\alias{e}
\alias{re}
\title{Error-Checked Wraps of \code{base::rep}}
\usage{
r(r, ...)

e(e, ...)

re(e, r, ...)

re(e, r, ...)
}
\arguments{
\item{r}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving
the number of replications of entire vectors.}

\item{...}{One or more atomic vectors to be replicated, reduced to unique
values, or reduced to unique duplicated values.}

\item{e}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving
the number of times each value is to be replicated in place before
replication of the entire vector.}
}
\value{
An atomic vector.
}
\description{
\tabular{ll}{
FUNCTIONS    \tab WHAT IT DOES                                          \cr
\code{r}          \tab \link[=av]{Atomize} \code{...} and with the result \code{dots},
call \code{rep.int(dots, r)}.                              \cr
\code{e}          \tab Atomize \code{...} and with the result \code{dots}, call
\code{rep.int(dots, each = e)}.                            \cr
\code{re}, \code{er}   \tab Atomize \code{...}, and with the result \code{dots}, call
\code{rep(dots, times = r, each = e)}.                       }
NOTE
\cr Although \code{er} and \code{re} produce identical results when their arguments
are the same, both are included for convenience.
}
\section{Comparing Function Output}{
 The following console excerpt
demonstrates how each function performs.

\if{html}{\out{<div class="sourceCode">}}\preformatted{> list(r  =  r(2,    0:1, 9),
+      e  =  e(3,    0:1, 9),
+      re = re(2, 3, 0:1, 9),
+      er = er(3, 2, 0:1, 9))
$r
[1] 0 1 9 0 1 9

$e
[1] 0 0 0 1 1 1 9 9 9

$re
 [1] 0 0 0 1 1 1 9 9 9 0 0 0 1 1 1 9 9 9

$er
 [1] 0 0 1 1 9 9 0 0 1 1 9 9 0 0 1 1 9 9
}\if{html}{\out{</div>}}
}

\examples{
r(2, 0:4, 5:9)
e(3, 0:4, 5:9)
re(2, 3, 0:4, 5:9)
er(3, 2, 0:4, 5:9)
}
\concept{extended}
