% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\encoding{UTF-8}
\name{r}
\alias{r}
\alias{e}
\alias{re}
\alias{er}
\title{Error-checked wrappers for \code{base::rep}}
\usage{
r(R, ...)

e(E, ...)

re(R, E, ...)

er(E, R, ...)
}
\arguments{
\item{R}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} giving the number of replications of entire vectors.}

\item{...}{One or more objects containing atomic values to be replicated (\link[=a]{atomized} before replication).}

\item{E}{A complete positive whole-number scalar giving the number of times each value is to be replicated in place.}
}
\value{
An atomic vector.
}
\description{
These functions \link[=av]{atomize} \code{...} arguments, converting them to an atomic vector (represented here as \code{dots}) before replication.
}
\details{
\tabular{ll}{  \verb{re, er}   \tab Calls \code{rep(dots, times = R, each = E)}. \cr   \tab   \cr
  \code{e}        \tab Calls \code{rep.int(dots, each = E)}.        \cr   \tab   \cr
  \code{r}        \tab Calls \code{rep.int(dots, R)}.                              }
}
\examples{
r(2, 0:2, 3)
e(3, 0:2, 3)
re(2, 3, 0:2, 3)
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}
}
\concept{extensions}
