% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\encoding{UTF-8}
\name{rd}
\alias{rd}
\alias{rd_csv}
\alias{rd_clip}
\alias{rd_xsv}
\alias{rd_tsv}
\title{Thin and extended functionality wrappers of \code{readr} functions.}
\usage{
rd()

rd_csv(File = NULL, ...)

rd_clip()

rd_xsv(File = NULL, Delim = NULL, ...)

rd_tsv(File = NULL, ...)
}
\arguments{
\item{File}{Either \code{NULL} or a \link[=cmp_chr_vec]{complete character vector} that resolves to a file path. When \code{File = NULL} the user is asked to select a file using a system dialog box.}
}
\value{
\strong{Varies}       \cr\cr \code{rd_clip}
\cr\cr  \strong{A data.frame} \cr\cr \verb{rd_csv, rd_tsv, rd_xsv}
}
\description{
Read the clipboard and text files, but prompting user to choose a file if no file name/path is provided.
}
\details{
\tabular{ll}{  \code{rd_clip}   \tab Thinly wraps \code{\link[readr]{clipboard}}         \cr   \tab   \cr
  \code{rd_csv}    \tab Extends \code{\link[readr]{read_csv}}\eqn{^{(1)}}   \cr
  \code{rd_tsv}    \tab Extends \code{\link[readr]{read_tsv}}\eqn{^{(1)}}   \cr
  \code{rd_xsv}    \tab Extends \code{\link[readr]{read_delim}}\eqn{^{(1)}} \cr   \tab     }
\tabular{l}{  \eqn{^{(1)}} \code{File = NULL} prompts the user to select a file.                       }
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{swap}()},
\code{\link{value_exists}()}

Other wraps: 
\code{\link{basics}()},
\code{\link{dp}()},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{sca}()},
\code{\link{tb}()}

Other files: 
\code{\link{paths}}
}
\concept{extensions}
\concept{files}
\concept{wraps}
