% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rct_delim.R
\name{read_rct_delim}
\alias{read_rct_delim}
\title{Interactively Select and Read a Rectangular Delimited Text Data File}
\usage{
read_rct_delim(type, d = ",", clear = TRUE)
}
\arguments{
\item{type}{A \link[=cmp_str_scl]{complete string scalar} description of the type of data file to be read.}

\item{d}{A \link[=cmp_ch1_scl]{complete onechar scalar} text delimiter.}

\item{clear}{A logical scalar indicating whether to clear the console to denote a new user interaction.}
}
\value{
A data.frame.
}
\description{
Interactively Select and Read a Rectangular Delimited Text Data File
}
\seealso{
Other files: 
\code{\link{paths_help}()},
\code{\link{rd_help}()}
}
\concept{files}
