% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\encoding{UTF-8}
\name{reclass}
\alias{reclass}
\title{Set object class}
\usage{
reclass(x, cls)
}
\arguments{
\item{x}{An R object.}

\item{cls}{A \link[=cmp_chr_vec]{complete character vec} of classes to apply.}
}
\value{
An object of class \code{cls}.
}
\description{
Set object class
}
\examples{
obj1. <- tb(letters = letters, numbers = 1:26)
obj2. <- reclass(obj1., c("new", class(obj1.)))
obj3. <- reclass(obj1., "new")
class(obj1.)
class(obj2.)
class(obj3.)
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unq}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
