% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Set object class}
\usage{
reclass(x, cls)
}
\arguments{
\item{x}{An R object.}

\item{cls}{A \link[=cmp_chr_vec]{complete character vec} of classes to apply.}
}
\value{
An object of class \code{cls}.
}
\description{
Set object class
}
