% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling.R
\name{recycling}
\alias{recycling}
\alias{recyclable_n}
\alias{recyclable}
\alias{recycle}
\title{Recycling}
\usage{
recyclable_n(n., targ. = max(n.))

recyclable(..., n. = NULL, min. = 1, max. = NULL, targ. = NULL, err. = F)

recycle(..., n. = NULL, min. = 1, max. = NULL, targ. = NULL)
}
\arguments{
\item{n.}{for \code{recyclable_n}, the lengths of arguments to be recycled;
for \code{recyclable} and \code{recycle}, \code{NULL} or the set of valid
recycled argument lengths.}

\item{targ.}{target length of recycled arguments.}

\item{...}{named arguments to be recycled in the calling function.}

\item{min.}{\code{NULL} or the minimum valid recycled argument length.}

\item{max.}{\code{NULL} or the maximum valid recycled argument length.}

\item{err.}{\code{TRUE} or \code{FALSE} indicating whether to throw an error
if the arguments in \code{...} are not recyclable.}

\item{lengths.}{vector of lengths to check for recyclability.}
}
\value{
\code{recyclable_n} and \code{recyclable} return \code{TRUE} or
\code{FALSE}. \code{recycle} returns \code{NULL} as it is called for the
side effect of recycling arguments in the environment of a parent function.
}
\description{
Check for recyclability and recycle arguments.
}
\details{
\strong{\code{recyclable_n}}
\cr Checks whether a vector of lengths represents recyclable arguments.
\cr\cr
\strong{\code{recyclable}}
\cr Checks arguments in \code{...} for recyclability.
\cr\cr
\strong{\code{recycle}}
\cr Recycles arguments in \code{...} in the environment of the function
\code{gens.} generations back in the function call stack.
}
\seealso{
Other meta: 
\code{\link{as_xxx}},
\code{\link{av}()},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fx}()},
\code{\link{getnames}()},
\code{\link{make}},
\code{\link{na_vals}},
\code{\link{named}()},
\code{\link{nx}()},
\code{\link{os}()},
\code{\link{ply}()},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{values}},
\code{\link{xb}()}
}
\concept{meta}
