% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal.R
\name{removal}
\alias{removal}
\alias{xat}
\alias{xat0}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{Remove various objects or characteristics.}
\usage{
xat(x, ...)

xat0(x)

xcls(x, ...)

xwarn()

xplot(x = NULL)

xcon()
}
\arguments{
\item{x}{An R object.}

\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs}
containing names of attributes or classes to remove.}
}
\value{
\tabular{lll}{
\code{xat}, \code{xat0}, \code{xcls}
\tab    \tab An R object.                                             \cr
  \tab    \tab                                                          \cr
\code{xwarn}, \code{splot}, \code{xcon}
\tab    \tab {NULL} (call for side effects).                            }
An object (\code{xat} and \code{xcls}) or \code{NULL} (\code{xwarn}
and \code{xcls}).
}
\description{
\tabular{ll}{
\code{xat}     \tab Remove one or more attributes by name.              \cr
\code{xat0}    \tab Remove all attributes.                              \cr
\code{xcls}    \tab Remove one or more class attributes by name.        \cr
\code{xwarn}   \tab Purge all warnings.                                 \cr
\code{xplot}   \tab Purge all plots (and if \code{x} is not \code{NULL},
print {x}, which is meant to be a plot object).     \cr
\code{xcon}    \tab Clear the console.                                    }
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put.}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{xb}()}
}
\concept{extensions}
