% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal..R
\encoding{UTF-8}
\name{removal}
\alias{removal}
\alias{xat}
\alias{xat0}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{Remove objects or object characteristics}
\usage{
xat(x, ...)

xat0(x)

xcls(x, ...)

xwarn()

xplot(x = NULL)

xcon()
}
\arguments{
\item{x}{An R object.}

\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs} containing names of attributes or classes to remove.}
}
\value{
\tabular{rl}{
\verb{xcon,xplot,xwarn}   \tab \code{NULL}.
\cr \verb{xat,xat0,xcls}   \tab An R object.
}
}
\description{
\tabular{rl}{
\code{xplot}   \tab Purges plots + prints \code{x} if not \code{NULL}.
\cr \code{xwarn}   \tab Purges warnings.
\cr  \code{xcon}   \tab Purges the console.
\cr  \code{xcls}   \tab Removes class(es) by name.
\cr  \code{xat0}   \tab Removes all attributes.
\cr   \code{xat}   \tab Removes attribute(s) by name.
}
}
\examples{
x. <- data.frame(x = -5:5, y = sqrt(abs(-5:5)))
class(x.) <- c(class(x.), "ccc")
attr(x., "aaa") <- "aaa"

## Structure with custom class and attribute.
str(x.)

## After removing just the custom class.
str(xcls(x., "ccc"))

## After removing just the custom attribute.
str(xat(x., "aaa"))

## After removing all attributes.
str(xat0(x.))

\dontrun{
## Build a ggplot scatterplot object (y.)
library(ggplot2)
y. <- ggplot(x., aes(x, y)) + geom_point()

## Print to console, wait for user, purge console
cat("\n A message")
continue()
xcon()

## Display a scatterplot, wait for user
plot(x.$x, x.$y)
continue()

## Purge plots + print ggplot, wait for user, purge plots
xplot(y.)
continue()
xplot()

## Generate warning, print, purge, validate purge
sqrt(-1)
warnings()
xwarn()
warnings()
}
}
