% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal.R
\name{removal}
\alias{removal}
\alias{xat}
\alias{xat0}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{Remove objects or object characteristics}
\usage{
xat(x, ...)

xat0(x)

xcls(x, ...)

xwarn()

xplot(x = NULL)

xcon()
}
\arguments{
\item{x}{An R object.}

\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs} containing names of attributes or classes to remove.}
}
\value{
\itemize{
\item \strong{\verb{xat, xat0, xcls}}: An R object.
\item \strong{\verb{xwarn, splot, xcon}}: \code{NULL} (called for side effects).
}
}
\description{
\itemize{
\item \strong{\code{xat}}: remove one or more attributes by name.
\item \strong{\code{xat0}}: removes all attributes.
\item \strong{\code{xcls}}: removes one or more class attributes by name.
\item \strong{\code{xwarn}}: purges all warnings.
\item \strong{\code{xplot}}: purges all plots (and if \code{x} is not \code{NULL}, print it).
\item \strong{\code{xcon}}: clears the console.
}
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}
}
\concept{extensions}
