% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.R
\name{reps}
\alias{reps}
\alias{each}
\alias{reps_each}
\alias{each_reps}
\title{Replication}
\usage{
reps(..., r = 1)

each(..., e = 1)

reps_each(..., r = 1, e = 1)

each_reps(..., e = 1, r = 1)
}
\arguments{
\item{...}{One or more atomic vectors to be replicated, reduced to unique
values, or reduced to unique duplicated values.}

\item{r}{An integer scalar giving the number of replications of entire
vectors.}

\item{e}{An integer scalar giving the number of times each value is to be
replicated in place before replication of the entire vector.}
}
\value{
An atomic vector.
}
\description{
Extended functionality for \code{\link[base:rep]{base::rep}}.
}
\details{
\strong{\code{reps}}
\cr \code{\link[=a]{atomizes}} \code{...} and concatenates \code{r} copies
of the resulting vector. For example, \code{reps(0:1, 9, r = 2)} gives the
result \code{c(0, 1, 9, 0, 1, 9)}.
\cr\cr
\strong{\code{each}}
\cr \code{\link[=a.]{atomizes}} \code{...} and creates \code{e} copies of
each element of the resulting vector in place. For example, \code{each(0:1,
  9, e = 2)} gives the result \code{c(0, 0, 1, 1, 9, 9)}.
\cr\cr
\strong{\code{reps_each} and \code{each_reps}}
\cr Call \code{reps} then \code{each} and vice versa, respectively.
}
\examples{
reps(0:4, 5:9, r = 2)
each(0:4, 5:9, r = 2)
reps_each(0:4, 5:9, r = 2, e = 2)
each_reps(0:4, 5:9, e = 2, r = 2)
}
\seealso{
Other values: 
\code{\link{dups}()},
\code{\link{uv}()}
}
\concept{values}
