% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revstr.R
\encoding{UTF-8}
\name{revstr}
\alias{revstr}
\title{Reverse strings}
\usage{
revstr(x)
}
\arguments{
\item{x}{A \link[=chr_vec]{character vec}.}

\item{na}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values are acceptable.}
}
\value{
An object of the same structure and dimension as \code{x}.
}
\description{
Reverses the order of characters in each element of a non-empty \link[=atm_chr]{atomic character object}, \link[=chr_vls]{character vlist} or \link[=chr_dtf]{character data.frame}.
}
\examples{
egRevStr <- function() {
  egBanana <- "banana"
  egApple  <- "apple"
  egCarrot <- "torrac"
  egZero   <- "orez"
  egFront  <- base::c(egApple, egBanana)
  egBack   <- base::c(egZero , egCarrot)
  egMat22  <- base::matrix(base::c(egFront, egBack), nrow = 2)
  egVls13  <- base::list(front = egFront, back = egBack)
  egDtf22  <- base::data.frame(front = egFront, back = egBack)
  base::list(
    egBanana = egBanana, revBanana = uj::revstr(egBanana),
    egCarrot = egCarrot, revCarrot = uj::revstr(egCarrot),
    egApple  = egApple , revApple  = uj::revstr(egApple ),
    egZero   = egZero  , revZero   = uj::revstr(egZero  ),
    egFront  = egFront , revFront  = uj::revstr(egFront ),
    egBack   = egBack  , revBack   = uj::revstr(egBack  ),
    egMat22  = egMat22 , revMat22  = uj::revstr(egMat22 ),
    egVls13  = egVls13 , revVls13  = uj::revstr(egVls13 ),
    egDtf22  = egDtf22 , revDtf22  = uj::revstr(egDtf22 ))
}
egRevStr <- egRevStr()
egRevStr
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{chn}()},
\code{\link{maxnch}()},
\code{\link{spaces}()},
\code{\link{ss}()}
}
\concept{chars}
\concept{strings}
