% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revstr.R
\encoding{UTF-8}
\name{revstr}
\alias{revstr}
\title{Reverse strings}
\usage{
revstr(x, na = FALSE)
}
\arguments{
\item{x}{A \link[=cmp_chr_vec]{character vec}.}

\item{na}{A non-\code{NA} logical scalar indicating whether \code{NA} values are acceptable.}
}
\value{
An object of the same structure and dimension as \code{x}.
}
\description{
Reverses the order of characters in each element of a non-empty \link[=pop_chr]{atomic character object}, \link[=chr_vls]{character vlist} or \link[=chr_dtf]{character data.frame}.
}
\examples{
zero. <- c(zero = "orez")
apple. <- c(apple = "apple")
banana. <- c(banana = "banana")
carrot. <- c(carrot = "torrac")
back. <- c(zero., carrot.)
front. <- c(apple., banana.)
mat22. <- matrix(c(front., back.), nrow = 2)
vls13. <- list(scl = av(apple.), vec = av(banana., back.))
dtf22. <- data.frame(front = av(front.), back = av(back.))
dtfNA. <- data.frame(front = av(NA, front.), back = av(back., NA))

back.
revstr(back.)

front.
revstr(front.)

mat22.
revstr(mat22.)

vls13.
revstr(vls13.)

dtf22.
revstr(dtf22.)

dtfNA.
revstr(dtfNA., na = TRUE)
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}},
\code{\link{markdown}},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{chn}()},
\code{\link{max_nch}()},
\code{\link{spaces}()},
\code{\link{ss}()}
}
\concept{chars}
\concept{strings}
