% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revstr.R
\encoding{UTF-8}
\name{revstr}
\alias{revstr}
\title{Reverse strings}
\usage{
revstr(x, nas = FALSE)
}
\arguments{
\item{x}{A \link[=chr_vec]{character vec}.}

\item{nas}{\code{TRUE} or \code{FALSE} indicating whether \code{NA} values are acceptable.}
}
\value{
An object of the same structure and dimension as \code{x}.
}
\description{
Reverses the order of characters in each element of a non-empty \link[=atm_chr]{atomic character object}, \link[=chr_vls]{character vlist} or \link[=chr_dtf]{character data.frame}.
}
\examples{
egBanana <- c(banana = "banana")
egCarrot <- c(carrot = "torrac")
egApple <- c(apple = "apple")
eg0 <- c(zero = "orez")

egFront <- c(egApple, egBanana)
egBack <- c(eg0, egCarrot)

egMat22 <- matrix(c(egFront, egBack.), nrow = 2)
egVls13 <- list(scl = av(egApple), vec = av(egBanana, egBack))
egDtf22 <- data.frame(front = av(egFront), back = av(egBack))
egDtfNA <- data.frame(front = av(NA, egFront), back = av(egBack, NA))

egBack
revstr(egBack)

egFront
revstr(egFront)

egMat22
revstr(egMat22)

egVLS13
revstr(egVLS13)

egDtf22
revstr(egDtf22)

egDtfNA
revstr(egDtfNA, na = TRUE)
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{chn}()},
\code{\link{max_nch}()},
\code{\link{spaces}()},
\code{\link{ss}()}
}
\concept{chars}
\concept{strings}
