% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\encoding{UTF-8}
\name{rm_vals}
\alias{rm_vals}
\alias{rm_elts}
\alias{rm_rows}
\alias{rm_cols}
\alias{xvals}
\alias{xelts}
\alias{xrows}
\alias{xcols}
\title{Manipulate positions, rows, and columns}
\usage{
rm_vals(x, vals)

rm_elts(x, elts)

rm_rows(x, rows)

rm_cols(x, cols)

xvals(x, vals)

xelts(x, elts)

xrows(x, rows)

xcols(x, cols)
}
\arguments{
\item{x}{For \verb{*vals} and \verb{*elts} an atomic vector. For \verb{*cols} and \verb{*rows}, a atomic data.frame (?atm_dtf) or an atomic matrix.}

\item{vals}{A unique-valued atomic vector (?uj::unq_atm_vec) mode-compatible with \code{x}.}

\item{elts, rows, cols}{Atomic vector indexing elements, rows, or columns of \code{x} to be deleted, respectively. These are either
\itemize{\item A uniquely-valued positive whole-number vector.
\item A complete logical vector (?uj::cmp_lgl_vec) of length \code{length(x)}.
\item A uniquely-valued character vector of element/column/row names.   }}
}
\value{
An object of the same class as \code{x}.
}
\description{
Manipulate positions, rows, and columns
}
\details{
\tabular{ll}{  \verb{purge_vals, rm_vals, xvals}   \tab Remove values in atomic vector \code{vals} from atomic vector \code{x}. \cr   \tab   \cr
  \verb{purge_elts, rm_elts, xelts}   \tab Delete elements of \code{x} indexed in \code{elts}                      \cr   \tab   \cr
  \verb{purge_rows, rm_rows, xrows}   \tab Delete rows of \code{x} indexed in \code{rows}.                         \cr   \tab   \cr
  \verb{purge_cols, rm_cols, xcols}   \tab Delete columns of \code{x} indexed in \code{cols}                                      }
}
\seealso{
Other function_form: 
\code{\link{insert_elts_before}()},
\code{\link{move_elt}()},
\code{\link{put}()}

Other purge: 
\code{\link{purge}()}
}
\concept{function_form}
\concept{purge}
