% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng2lev.R
\name{rng2lev}
\alias{rng2lev}
\title{Convert sortable values in levels by range}
\usage{
rng2lev(x, cuts, levs)
}
\arguments{
\item{x}{An atomic object of \link[=isrt]{sortable} values.}

\item{cuts}{An \link[=num_vec]{atomic vec} of \code{n-1} cuts for \code{n} levels.}

\item{levs}{An atomic vec of \code{n} level labels.}
}
\value{
An atomic object of the same dimension as \code{x} and the same mode as \code{levs}.
}
\description{
Convert values of \code{x} to the \code{n} levels in \code{levs} distinguished by the \code{n-1} thresholds in \code{cuts}.
\cr\cr When values of \code{x} are equal to a value of \code{cuts}, assignment to level is dependent on whether the value of each element of \code{cut} is less than \code{0} vs. greater than or equal to \code{0}.
\cr\cr For a value in \code{cuts} less than \code{0}, the level assigned is the lower of the levels distinguished by that value of \code{cuts}.
\cr\cr Otherwise, the level assigned is the upper of the levels distinguished by that value of \code{cuts}.
}
\examples{
num_vals <- -5:5
chr_vals <- letters[1:11]
ord_vals <- factor(chr_vals, levels = chr_vals, ordered = TRUE)
num_cuts <- c(-3, 0, 3)
chr_cuts <- c("c", "f", "i")
ord_cuts <- factor(chr_cuts, levels = chr_vals, ordered = TRUE)
rng_levs <- c("A", "B", "C", "D")
rng2lev(num_vals, num_cuts, rng_levs)
rng2lev(chr_vals, chr_cuts, rng_levs)
rng2lev(ord_vals, ord_cuts, rng_levs)
}
