% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng2lev.R
\encoding{UTF-8}
\name{rng2lev}
\alias{rng2lev}
\title{Convert sortable values in levels by range}
\usage{
rng2lev(x, cuts, levs)
}
\arguments{
\item{x}{An \link[=atm_srt]{atomic object of sortable} values.}

\item{cuts}{An \link[=srt_vec]{atomic vec} of \code{N-1} cuts for \code{N} levels.}

\item{levs}{An atomic vec of \code{N} level labels.}
}
\value{
An atomic object of the same dimension as \code{x} and the same mode as \code{levs}.
}
\description{
Convert values of \code{x} to the \code{N} levels in \code{levs} distinguished by the \code{N-1} thresholds in \code{cuts}.
\cr\cr When values of \code{x} are equal to a value of \code{cuts}, assignment to level is dependent on whether the value of each element of \code{cuts} is less than \code{0} vs. greater than or equal to \code{0}.
\cr\cr For a value in \code{cuts} less than \code{0}, the level assigned is the lower of the levels distinguished by that value of \code{cuts}.
\cr\cr Otherwise, the level assigned is the upper of the levels distinguished by that value of \code{cuts}.
}
\examples{
numVALS <- -5:5
chrVALS <- letters[1:11]
ordVALS <- factor(chrVALS, levels = chrVALS, ordered = TRUE)
numCUTS <- c(-3, 0, 3)
chrCUTS <- c("c", "f", "i")
ordCUTS <- factor(chrVALS, levels = chrVALS, ordered = TRUE)
rngLEVS <- c("A", "B", "C", "D")
rng2lev(numVALS, numCUTS, rngLEVS)
rng2lev(chrVALS, chrCUTS, rngLEVS)
rng2lev(ordVALS, ordCUTS, rngLEVS)
}
\seealso{
Other conversions: 
\code{\link{hex_dec_funs}()},
\code{\link{tocase}()},
\code{\link{u2u}()}
}
\concept{conversions}
