% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Build and evaluate code}
\usage{
run(...)
}
\arguments{
\item{...}{Arguments to be pasted into a character scalar command to be run (i.e., parsed and evaluated) in the environment where \code{run(...)} was called.}
}
\value{
\code{NULL}. Called for its side effect of executing code in the environment where \code{run(...)} was called.
}
\description{
Collapses \code{...} into a character scalar and evaluates it as code in the environment of the calling function.
}
\examples{
Join <- ""
CallVector <- c("paste0(", 0:3, ", collapse = Join)")
CallScalar <- paste0(CallVector, collapse = Join)
CallVector <- c("Vector <- ", CallVector)
CallScalar <- paste0("Scalar <- ", CallScalar)
run(CallVector)
run(CallScalar)
print(CallVector)
print(CallScalar)
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binary_failsafe}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{paths}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{r}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}},
\code{\link{xb}()}

Other environments: 
\code{\link{callers}()},
\code{\link{envir_vals}},
\code{\link{os}()},
\code{\link{recycling}}
}
\concept{code}
\concept{environments}
\concept{extensions}
