% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Build and evaluate code}
\usage{
run(...)
}
\arguments{
\item{...}{Arguments to be pasted into a character scalar command to be run (i.e., parsed and evaluated) in the environment where \code{run(...)} was called.}
}
\value{
\code{NULL}. Called for its side effect of executing code in the environment where \code{run(...)} was called.
}
\description{
Collapses \code{...} into a character scalar and evaluates it as code in the environment of the calling function.
}
\examples{
CallVec <- paste0("'", 0:3, "'")
CallVec <- c("paste0(", CallVec, ", collapse = '')")
CallVec <- c("ResultVec <- ", CallVec)
run(CallVec)

CallScl <- paste0(CallVec, collapse = "")
CallScl <- paste0("ResultScl <- ", CallScl)
run(CallScl)

print(CallVec)
print(CallScl)
print(ResultVec)
print(ResultScl)
}
\seealso{
Other environments: 
\code{\link{callers}()},
\code{\link{envir_vals}},
\code{\link{os}()},
\code{\link{recycling}}
}
\concept{code}
\concept{environments}
