% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\encoding{UTF-8}
\name{run}
\alias{run}
\title{Build and evaluate code}
\usage{
run(...)
}
\arguments{
\item{...}{Arguments to be pasted into a character scalar command to be run (i.e., parsed and evaluated) in the environment where \code{run(...)} was called.}
}
\value{
The value returned upon evaluating the character scalar created from collapsing \code{...} arguments.
}
\description{
\code{run(...)} Collapses \code{...} into a character scalar and evaluates it as code in the environment of the calling function.
}
\examples{
egCallVec <- paste0("'", 0:3, "'")
egCallVec <- c("paste0(", egCallVec, ", collapse = '')")
egCallVec <- c("egResultVec <- ", egCallVec)
run(egCallVec)

egCallScl <- paste0(egCallVec, collapse = "")
egCallScl <- paste0("egResultScl <- ", egCallScl)
run(egCallScl)

print(egCallVec)
print(egCallScl)
print(egResultVec)
print(egResultScl)

egRUNalias1 <- function(x, y) {egDTFalias(var.x = x, var.y = y)}
egRUNalias2 <- function(x, y) {egRUNalias1(x, y)}

}
\seealso{
Other meta: 
\code{\link{callers}()},
\code{\link{dot_args}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{purge}()},
\code{\link{recyclable_ns}()}
}
\concept{meta}
