% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\encoding{UTF-8}
\name{say}
\alias{say}
\alias{cat0}
\alias{continue}
\title{User updates and interaction via the console in default color and style}
\usage{
say(..., Lev = 0, Sub = 0, .clear = FALSE)

cat0(..., .clear = FALSE)

continue(.clear = FALSE)
}
\arguments{
\item{...}{An arbitrary number of atomic arguments to be atomized collapsed into a character scalar message to the user. If no values are supplied and \code{Lev = NA} or \code{Lev = 6}, prints a single dot (\code{.}) immediately following the current contents of the console.}

\item{Lev}{Either \code{NA} to print the message in \code{...} immediately following the current contents of the console or a value from \code{1:6} to use the console for up to \code{6} levels of user updates. See details.}
}
\value{
\code{NULL}.
}
\description{
User updates and interaction via the console in default color and style
}
\details{
\tabular{ll}{  \code{continue}   \tab Prints \code{"Hit [enter] or [return] to continue"} to the console and pausse execution until the user hits the enter or return key.                                             \cr   \tab   \cr
  \code{cat0}       \tab Collapses \code{...} to a character scalar message prefixed and suffixed with newlines and prints it to the console (ensures the message occurs on its own line of the console). \cr   \tab   \cr
  \code{say}        \tab Collapses \code{...} to a character scalar message and either (a) prints it to the console or (b) structures it as an update and prints it to the console. See the next table.                  }
When \code{...} is empty, the message defaults to a single dot (\code{"."}).
\cr\cr The function \code{say} allows for structuring user update messages at hierarchical levels as explained in the following table:
\tabular{ll}{  VALUE      \tab PREFIX OF MESSAGE          \cr
  OF \code{Lev}   \tab POSTED TO CONSOLE          \cr
  \code{0}        \tab < none >                   \cr
  \code{1}        \tab < newline >                \cr
  \code{2}        \tab < newline > + \code{'| '}       \cr
  \code{3}        \tab < newline > + \code{'| > '}     \cr
  \code{4}        \tab < newline > + \code{'| | > '}   \cr
  \code{5}        \tab < newline > + \code{'| | | > '} \cr
  \code{...}      \tab \code{...}                        }
In addition, Sub-levels are defined via the argument \code{Sub} where \code{Sub = 0} leaves the message as is, \code{Sub = 1} prepends \code{' > '} to the message, and \code{Sub = 2}, prepends \code{"..("} and appends \code{")"} to the message.
}
\examples{
cat0("msg on its own line")
chat. <- function() {
  say("msg 1A", Lev = 1)
  say("msg 2A", Lev = 2)
  say("msg 2B", Lev = 2)
  say("msg 3A", Lev = 3)
  say("msg 3B", Lev = 3)
  say()
  say()
  say()
  say("msg 3C", Lev = 3)
  say("msg 4A", Lev = 4)
  say("msg 4B", Lev = 4)
  say("msg 5A", Lev = 5)
  say(Lev = 6)
  say(Lev = 6)
  say("msg 5B", Lev = 5)
  say("msg 3D", Lev = 3)
  say()
  say()
  say("msg 1B", Lev = 1)
  say(" + a_random_message")
}
chat.()
\dontrun{continue()}
}
\seealso{
Other meta: 
\code{\link{callers}()},
\code{\link{dot_args}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{purge}()},
\code{\link{recyclable_ns}()},
\code{\link{run}()}

Other user: 
\code{\link{dialog}}
}
\concept{meta}
\concept{user}
