% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca.R
\name{sca}
\alias{sca}
\title{Thin wrappers of \code{scales} functions}
\usage{
sca(color, alpha = NA)
}
\arguments{
\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA},
existing alpha values are preserved.}
}
\description{
\tabular{rl}{
  \code{sca}   \tab Thinly wraps \code{\link[scales]{alpha}}.
}
}
\examples{
alpha("red", 0.1)
alpha(colours(), 0.5)
alpha("red", seq(0, 1, length.out = 10))
alpha(c("first" = "gold", "second" = "lightgray", "third" = "#cd7f32"), .5)
}
\seealso{
Other wraps: 
\code{\link{basics}},
\code{\link{dp}},
\code{\link{dt}()},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{rd}},
\code{\link{tb}()}

Other plotting: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{pals}},
\code{\link{u2u}()}
}
\concept{plotting}
\concept{wraps}
