% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{sets}
\alias{sets}
\title{Get All Possible Non-Empty Sets}
\usage{
sets(n)
}
\description{
Create a list containing elements of indices to create all
possible sets of 1 or more elements of an object of size \code{n}.
}
