% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{sets}
\alias{sets}
\title{Create Indices to Build All Possible Non-Empty Sets}
\usage{
sets(n)
}
\description{
Create a list of elements containing indices to create all
possible sets of size \code{1} or greater from a length-\code{n} vector of unique
values.
}
