% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\encoding{UTF-8}
\name{sets}
\alias{sets}
\title{Create indices to build all possible non-empty sets}
\usage{
sets(n)
}
\arguments{
\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the length of set of unique values.}
}
\value{
A \link[=cmp_psw_vls]{complete positive whole-number vlist}.
}
\description{
Creates a list of elements containing indices to create all possible unique-value sets of size \code{1} to \code{n} from a length-\code{n} vector of unique values.
}
\examples{
sets(1)
sets(2)
sets(3)
sets(4)
}
