% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\encoding{UTF-8}
\name{spaces}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{String spacing functions}
\usage{
spaces(n)

pad(..., n = 0, s = "r", p = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n}{A \link[=cmp_psw_scl]{complete positive whole-number scalar} indicating the number of spaces or the number of pad characters.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{s}{A \link[=cmp_ch1_vec]{complete onechar scalar} indicating side(s) to pad: \code{'l'}, \code{'r'}, or \code{'b'} for left, right, or both, respectively.}

\item{p}{A complete onechar scalar containing a single character used to pad strings.}
}
\value{
\strong{A character scalar} \cr\cr   \code{spaces}
\cr\cr  \strong{A character object} \cr\cr   \verb{pad, trm, sqz}
}
\description{
These functions create strings of spaces, pad strings with space, trim edge spaces, and squeeze out unnecessary spaces.
}
\details{
\tabular{ll}{  \code{spaces}   \tab Creates a character scalar of \code{n} spaces.                                                                            \cr   \tab   \cr
  \code{pad}      \tab Pads strings with leading, trailing white space using \code{\link[stringr:str_pad]{stringr::str_pad}}.               \cr   \tab   \cr
  \code{sqz}      \tab Trims leading, trailing, and extra internal white space using \code{\link[stringr:str_squish]{stringr::str_squish}}. \cr   \tab   \cr
  \code{trm}      \tab Trims leading and trailing white space using \code{\link[stringr:str_trim]{stringr::str_trim}}.                                     }
}
\examples{
spaces(0)
spaces(4)
pad(0:9, n = 3)
pad(0:9, n = 3, s = "r")
pad(0:9, n = 3, s = "l")
pad(0:9, n = 3, s = "b")
pad(0:9, n = 3, s = "b", p. = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}

Other chars: 
\code{\link{chn}()},
\code{\link{max_nch}()},
\code{\link{revstr}()},
\code{\link{ss}()}
}
\concept{chars}
\concept{strings}
