% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.R
\name{spacing}
\alias{spacing}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{Simple spacing functions}
\usage{
spaces(n.)

pad(..., n. = 0, s. = "r", p. = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n.}{Whole-number scalar indicating the number of spaces or pad
characters.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{s.}{Character scalar "side" padded ("l", "r", or "b" for left, right,
or both).}

\item{p.}{Character scalar containing a single character used to pad strings.}
}
\description{
Create a string of spaces, pad a string with spaces, trim spaces
from strings, and squeeze extra spaces from the middle of strings.
}
\details{
\strong{\code{spaces}}
\cr Creates a string of \code{n.} spaces.
\cr\cr
\strong{\code{pad}}
\cr Pads strings to length \code{N} using the padding character \code{p} on
side(s) \code{s.} (via \code{\link[stringr]{str_pad}}).
\cr\cr
\strong{\code{trm}}
\cr Trims white space the the left and right ends of strings (via
\code{\link[stringr]{str_trim}}).
\cr\cr
\strong{\code{sqz}}
\cr Trims white space from the left and right ends and extra white space
rom the middle of strings (via \code{\link[stringr]{str_squish}}).
}
\examples{
spaces(0)
spaces(4)
pad(0:9, N = 3)
pad(0:9, N = 3, S = "r")
pad(0:9, N = 3, S = "l")
pad(0:9, N = 3, S = "b")
pad(0:9, N = 3, S = "b", P = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{Oxford}},
\code{\link{delim}},
\code{\link{ipat}()},
\code{\link{pgrid}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
