% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.R
\name{spacing}
\alias{spacing}
\alias{spaces}
\alias{pad}
\alias{sqz}
\alias{trm}
\title{Error-Checked \code{stringr}-Based Spacing Functions}
\usage{
spaces(n)

pad(..., n. = 0, s. = "r", p. = " ")

sqz(...)

trm(...)
}
\arguments{
\item{n}{\link[=cmp_psw_scl]{Complete positive whole-number scalar}
indicating the number of spaces.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{n.}{\link[=cmp_psw_scl]{Complete positive whole-number scalar}
indicating the number of pad characters.}

\item{s.}{\link[=cmp_ch1_scl]{Complete onechar scalar} indicate side(s) to
pad: "l", "r", or "b" for left, right, or both, respectively.}

\item{p.}{\link[=cmp_ch1_scl]{Complete onechar scalar} containing a single
character used to pad strings.}
}
\value{
\tabular{lll}{
\code{spaces}                     \tab   \tab A character scalar.       \cr
\code{pad}, \code{trm}, \code{sqz}\tab   \tab A character object.         }
}
\description{
\tabular{ll}{
\code{spaces}   \tab Create a character vector of \code{n} spaces.      \cr
\code{pad}      \tab Pad a string with leading and/or trailing spaces using
\code{\link[stringr]{pad}}.                        \cr
\code{sqz}      \tab Trim leading, trailing, and extra internal spaces
using \code{\link[stringr]{str_squish}}.           \cr
\code{trm}      \tab Trim leading and trailing spaces using
\code{\link[stringr]{str_trim}}.                     }
}
\examples{
spaces(0)
spaces(4)
pad(0:9, n. = 3)
pad(0:9, n. = 3, s. = "r")
pad(0:9, n. = 3, s. = "l")
pad(0:9, n. = 3, s. = "b")
pad(0:9, n. = 3, s. = "b", p. = "*")
trm(c("a ", "b", "c", " ", "d  ", ""))
sqz(c("a ", "b", "c", " ", "d  ", ""))
sqz(trm(c("a ", "b", "c", " ", "d  ", "")))
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{charn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkstr}},
\code{\link{ox.}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{ss}()},
\code{\link{tocase}()},
\code{\link{weave}()}
}
\concept{strings}
