% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss.}
\alias{sss.}
\alias{sss}
\alias{icmp}
\alias{imss}
\alias{inas}
\alias{ioks}
\alias{iprt}
\alias{sss_vals}
\alias{isss}
\title{State of completeness properties}
\usage{
sss.()

sss(x)

icmp(x)

imss(x)

inas(x)

ioks(x)

iprt(x)

sss_vals()

isss(x, sss, ...)
}
\arguments{
\item{x}{An object.}

\item{sss}{\link[cmp_chr_scl]{Complete character scalar} containing one or
more values from \code{sss_vals()} separated by pipes.}

\item{...}{Additional arguments to \code{\link{meets}} containing value and
element/row/column count restrictions.}
}
\value{
\code{sss_vals()} returns a character vector. \code{sss(x)}
returns a character scalar or \code{NULL}. All others return either
\code{TRUE} or \code{FALSE}.
}
\description{
An object's state of completeness is defined for non-empty
atomic tibbles, non-empty atomic lists, non-empty atomic vectors, and
non-empty atomic arrays. For all others, state is undefined and returned as
\code{NULL}. The following table summarizes valid state properties.
\tabular{lll}{
STATE       \tab STATE      \tab CHARACTERISTICS                          \cr
PROPERTY    \tab PROPERTY   \tab OF QUALIFYING                            \cr
VALUE       \tab NAME       \tab OBJECTS                                  \cr
\code{'cmp'}\tab Complete   \tab Non-empty, atomic, no \code{NA} values.  \cr
\code{'mss'}\tab Missing    \tab Non-empty, atomic, only \code{NA} values.\cr
\code{'prt'}\tab Partial    \tab Non-empty, atomic, both \code{NA} and
non-\code{NA} values.                    \cr
\code{'nas'}\tab NA scalar  \tab Atomic \code{NA} scalar.                 \cr
\code{'oks'}\tab OK scalar  \tab Atomic non-\code{NA} scalar.               }
Functions related to state of completeness are described in the following
table:\tabular{ll}{
FUNCTION          \tab WHAT THE                                           \cr
FORMAT            \tab FUNCTION DOES                                      \cr
\code{i***}       \tab Evaluates whether an object is of the state of
completeness represented by \code{***}.            \cr
\code{sss}        \tab Gets a character vector containing all state of
completeness properties of an object.              \cr
\code{isss}       \tab Evaluates an object for a specific state of
completeness and any additional properties specified
in \code{...}.                                     \cr
\code{sss_vals}   \tab Gets a character vector of all possible state of
completeness property values.                        }
}
\section{Functions}{
\itemize{
\item \code{sss()}: Get a character scalar containing all state properties
from \code{sss_vals()} that is applicable to \code{x}, if defined. If not
defined, return \code{NULL}.

\item \code{icmp()}: Is \code{x} complete (atomic, of length 1 or greater, and
containing no \code{NA} values)?

\item \code{imss()}: Is \code{x} a missing object (atomic, of length 1 or
greater, and containing only \code{NA} values)?

\item \code{inas()}: Is \code{x} an atomic scalar \code{NA} value?

\item \code{ioks()}: Is \code{x} an atomic scalar non-\code{NA} value?

\item \code{iprt()}: Is \code{x} an atomic object with some \code{NA} and some
non-\code{NA} values?

\item \code{sss_vals()}: Get a character vector of all possible state property
values.

\item \code{isss()}: Evaluate whether \code{x} has a state property represented
in \code{sss}.

}}
\section{Additional Arguments in \code{...}}{

Submitting additional arguments to \code{isss} via \code{...}
allows for checking not just the state but whether length, number of rows,
number of columns, and element values meet flexible criteria.
}

\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()},
\code{\link{ttt_mmm.}()}
}
\concept{props}
