% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss}
\alias{sss}
\alias{sss_props}
\alias{is_sss_spec}
\alias{isss}
\alias{icol}
\alias{iemp}
\alias{ilin}
\alias{ipnt}
\alias{irct}
\alias{irow}
\alias{isld}
\alias{isqr}
\title{Shape properties}
\usage{
sss(x)

sss_props()

is_sss_spec(spec)

isss(x, spec, ...)

icol(x, ...)

iemp(x, ...)

ilin(x, ...)

ipnt(x, ...)

irct(x, ...)

irow(x, ...)

isld(x, ...)

isqr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{\code{NULL} or a \link[=cmp_chr_scl]{complete character vec}
containing one or more shape properties (i.e., from \code{sss_vals()}). Shape
properties in \code{spec} may be pipe-delimited. If there are multiple shape
properties in \code{spec}, \code{x} is inspected for any match to any shape property
in \code{spec}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{ll}{
FUNCTIONS                       \tab RETURN VALUE                       \cr
\code{sss_vals}                      \tab A character vector.                \cr
\code{sss}                           \tab A character scalar or vector.      \cr
\code{ixxx}, \code{isss}, \code{is_sss_spec}   \tab A logical scalar.                    }
}
\description{
An object's shape properties have to do with the number of
\link[=ddd]{defined dimensions} and/or \link[=eee]{effective dimensions}
with some additional restrictions as follows:\tabular{lll}{
NAME          \tab VALUE     \tab QUALIFYING OBJECTS                    \cr
Empty         \tab \code{'emp'}   \tab Non-\code{NULL}, length-\code{1} objects.       \cr
Point         \tab \code{'pnt'}   \tab Length-\code{1} vectors, length-\code{1}
\link[=ivls]{vlists}, length-\code{1} arrays,
and \verb{1 x 1} data.frames.              \cr
Linear        \tab \code{'lin'}   \tab Length \verb{2+} vectors, length \verb{2+}
\link[=ivls]{vlists}, length \verb{2+} arrays
with multiple index positions in just
\code{1} dimension, and data.frames with \code{1}
row and multiple columns or \code{1} column
and multiple rows.                    \cr
Row           \tab \code{'row'}   \tab Matrices and data.frames with \code{1} row and
multiple columns.                     \cr
Column        \tab \code{'col'}   \tab Matrices and data.frames with \code{1} column
and multiple rows.                    \cr
Rectangular   \tab \code{'rct'}   \tab Matrices and data.frames with multiple
rows and multiple columns.            \cr
Square        \tab \code{'sqr'}   \tab Matrices and data.frames with multiple
rows, multiple columns, and the same
number of rows and columns.           \cr
Solid         \tab \code{'sld'}   \tab Arrays with multiple index positions in
\verb{3+} dimensions.                        }
Functions in this family are:\tabular{ll}{
FUNCTION        \tab WHAT IT DOES                                       \cr
\code{sss}           \tab Gets a character vector containing all shape
properties possessed by \code{x}                        \cr
\code{ixxx}          \tab Evaluates whether \code{x} possesses the shape property
\code{xxx} (a placeholder for any given shape property
value), subject to any restrictions in \code{...}.      \cr
\code{isss}          \tab Evaluates whether \code{x} possesses one or more
(possibly pipe-delimited) shape properties in
\code{spec}, subject to any restrictions in \code{...}.      \cr
\code{sss_props}     \tab Gets a character vector of all possible shape
property values.                                   \cr
\code{is_sss_spec}   \tab Evaluates whether \code{spec} is a valid shape property
specification.                                       }
}
\section{Specifying Count and Value Restrictions}{
 Specifying restrictions in
\code{...} is optional. The full set of recognized arguments names are defined
in the following table along with the properties each
specifies:\tabular{ll}{
NAME     \tab WHAT IT SPECIFIES                                         \cr
\code{n}      \tab Vector of valid lengths/numbers of elements.              \cr
\code{nr}     \tab Vector of valid numbers of rows.                          \cr
\code{nc}     \tab Vector of valid numbers of columns.                       \cr
\code{min}    \tab Scalar minimum valid length/number of element.            \cr
\code{minr}   \tab Scalar minimum valid number of rows.                      \cr
\code{minc}   \tab Scalar minimum valid number of columns.                   \cr
\code{max}    \tab Scalar maximum valid length/number of element.            \cr
\code{maxr}   \tab Scalar maximum valid number of rows.                      \cr
\code{maxc}   \tab Scalar maximum valid number of columns.                   \cr
\code{vals}   \tab Vector of valid values.                                   \cr
\code{lt}     \tab Scalar less-than (exclusive upper) bound.                 \cr
\code{le}     \tab Scalar less-than-or-equal (inclusive upper) bound         \cr
\code{ge}     \tab Scalar greater-than-or-equal (inclusive lower) bound.     \cr
\code{gt}     \tab Scalar greater-than bound (exclusive lower) bound.          }
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()}
}
\concept{props}
