% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\encoding{UTF-8}
\name{sss}
\alias{sss}
\alias{sss_props}
\alias{sss_funs}
\alias{is_sss_spec}
\alias{isss}
\alias{icol}
\alias{iemp}
\alias{ilin}
\alias{ipnt}
\alias{irct}
\alias{irow}
\alias{isld}
\alias{isqr}
\title{Shape properties}
\usage{
sss(x)

sss_props()

sss_funs()

is_sss_spec(spec)

isss(x, spec, ...)

icol(x, ...)

iemp(x, ...)

ilin(x, ...)

ipnt(x, ...)

irct(x, ...)

irow(x, ...)

isld(x, ...)

isqr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{Either \code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{shape} properties (i.e., from \code{sss_vals()}). \code{Shape} properties in \code{spec} may be pipe-delimited. If there are multiple \code{shape} properties in \code{spec}, \code{x} is inspected for any match to any shape property in \code{spec}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A logical scalar}   \cr   \verb{is_sss_spec, iSSS, isss}
\cr\cr \emph{A character vector} \cr   \verb{sss_props, sss_funs}
}
\description{
Shape properties are defined as follows: \tabular{rll}{
  \code{'emp'}   \tab empty    \tab Of length \code{0} (but not \code{NULL}).
\cr \code{'pnt'}   \tab point    \tab Of length \code{1}\eqn{^a}.
\cr \code{'col'}   \tab column   \tab \verb{1 × 2+} matrix/data.frame.
\cr \code{'row'}   \tab row      \tab \verb{2+ × 1} matrix/data.frame
\cr \code{'lin'}   \tab linear   \tab \verb{2+} positions in \code{1} dimension\eqn{^b}.
\cr \code{'rct'}   \tab rect     \tab \verb{2+} positions in \code{2} dimensions.
\cr \code{'sqr'}   \tab square   \tab \verb{N × N} matrix where \verb{N ≥ 2}.
\cr \code{'sld'}   \tab solid    \tab \verb{2+} positions in \verb{3+} dimensions.
}
\verb{ }\eqn{^{a.}} Includes \verb{1x1} data.frames.
\cr \verb{ }\eqn{^{b.}} Length-\verb{2+} vectors, arrays of \verb{2+} positions in \code{1} dimension, row/column data.frames.
\cr\cr
\strong{Shape property functions} \tabular{rl}{
  \code{is_sss_spec}   \tab Is \code{spec} a shape specification?
\cr \code{sss_props}   \tab What shape properties are there?
\cr  \code{sss_funs}   \tab What shape property functions are there?
\cr      \code{isss}   \tab Does \code{x} match the shape spec in argument \code{spec}?
\cr      \code{iSSS}   \tab Does \code{x} match single shape property \code{'SSS'}?
\cr       \code{sss}   \tab What are \code{x}'s shape properties?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
sss_funs()
sss_props()
is_sss_spec("emp|pnt")
sss(matrix(letters, nrow = 1))
isss(matrix(letters, nrow = 2), "rct")
ipnt(data.frame(a = 1))
}
\seealso{
Other properties: 
\code{\link{as_mmm}},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{err_check}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
