% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\encoding{UTF-8}
\name{sss}
\alias{sss}
\alias{sss_props}
\alias{sss_funs}
\alias{is_sss_spec}
\alias{SSS}
\alias{COL}
\alias{EMP}
\alias{LIN}
\alias{PNT}
\alias{RCT}
\alias{ROW}
\alias{SLD}
\alias{SQR}
\title{Shape properties}
\usage{
sss(x)

sss_props()

sss_funs()

is_sss_spec(spec)

SSS(x, spec, ...)

COL(x, ...)

EMP(x, ...)

LIN(x, ...)

PNT(x, ...)

RCT(x, ...)

ROW(x, ...)

SLD(x, ...)

SQR(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{Either \code{NULL} or a \link[=cmp_chr_vec]{complete character vec} containing one or more shape properties (i.e., from \code{sss_props()}). Shape properties in \code{spec} may be pipe-delimited. If there are multiple shape properties in \code{spec}, \code{x} is inspected for any match to any shape property in \code{spec}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \verb{sss_props, sss_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{is_sss_spec, \{SSS\}, SSS, sss}
}
\description{
Shape properties are defined as follows:
\tabular{lll}{  \verb{'emp', 'EMP'}   \tab empty    \tab of length \code{0} (but not \code{NULL})              \cr
  \verb{'pnt', 'PNT'}   \tab point    \tab of length \code{1}\eqn{^{(1)}}                   \cr
  \verb{'col', 'COL'}   \tab column   \tab \verb{1 × 2+} matrix/data.frame                  \cr
  \verb{'row', 'ROW'}   \tab row      \tab \verb{2+ × 1} matrix/data.frame                  \cr
  \verb{'lin', 'LIN'}   \tab linear   \tab \verb{2+} positions in \code{1} dimension\eqn{^{(2)}} \cr
  \verb{'rct', 'RCT'}   \tab rect     \tab \verb{2+} positions in \code{2} dimensions            \cr
  \verb{'sqr', 'SQR'}   \tab square   \tab \verb{N × N} matrix where \verb{N ≥ 2}                \cr
  \verb{'sld', 'SLD'}   \tab solid    \tab \verb{2+} positions in \verb{3+} dimensions             }
\tabular{l}{  \eqn{^{(1)}} Includes \verb{1x1} data.frames.                                                            \cr
  \eqn{^{(2)}} Length-\verb{2+} vectors, arrays of \verb{2+} positions in \code{1} dimension, row/column data.frames.  }
}
\details{
\tabular{ll}{  \code{is_sss_spec}   \tab Is \code{spec} a shape specification?                                                                              \cr
  \code{sss_props}     \tab What shape properties are there?                                                                              \cr
  \code{sss_funs}      \tab What shape property functions are there?                                                                      \cr   \tab  }
\tabular{ll}{  \code{{SSS}}         \tab Does \code{x} match single shape property \code{'{SSS}'}? (where \code{{SSS}} is a placeholder for any given shape property. \cr   \tab   \cr
  \code{SSS}           \tab Does \code{x} match the shape spec in argument \code{spec}?                                                             \cr
  \code{sss}           \tab What are \code{x}'s shape properties?                                                                                }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
Cl1 <- data.frame(letters = letters)
Cl2 <- matrix(1:10)
Emp <- vector(0)
Ln1 <- letters
Ln2 <- Cl1
Pt1 <- 1
Pt2 <- data.frame(letters = "a")
Rc1 <- data.frame(letters = letters, numbers = 1:26)
Rc2 <- matrix(1:10, nrow = 2)
Rw1 <- data.frame(a = "a", b = 1)
Rw2 <- matrix(1:10, nrow = 1)
Sld <- array(1:27, dim = c(3, 3, 3))
Sqr <- matrix(1:9, nrow = 3)
sss_props()
sss_funs()
is_sss_sped("emp|pnt")
sss(Cl1)
sss(Sqr)
c(SSS(Cl1, "col"), SSS(Rw1, "col"))
c(COL(Cl1), COL(Rw1))
c(COL(Cl2), COL(Rw2))
c(EMP(Emp), EMP(Ln1))
c(LIN(Ln1), LIN(Pt1))
c(LIN(Ln2), LIN(Rc1))
c(PNT(Pt1), PNT(Rc1))
c(PNT(Pt2), PNT(Rw1))
c(RCT(Rc1), RCT(Pt1))
c(RCT(Rc2), RCT(Ln1))
c(ROW(Rw1), ROW(Cl1))
c(ROW(Rw2), ROW(Rc1))
c(SLD(Sld), SLD(Rc1))
c(SQR(Sqr), SQR(Pt1))
}
\seealso{
Other properties: 
\code{\link{asMODE}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{unq_ccc}()}
}
\concept{properties}
