% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss}
\alias{sss}
\alias{sss_props}
\alias{is_sss_spec}
\alias{isss}
\alias{icol}
\alias{iemp}
\alias{ilin}
\alias{ipnt}
\alias{irct}
\alias{irow}
\alias{isld}
\alias{isqr}
\title{Shape properties}
\usage{
sss(x)

sss_props()

is_sss_spec(spec)

isss(x, spec, ...)

icol(x, ...)

iemp(x, ...)

ilin(x, ...)

ipnt(x, ...)

irct(x, ...)

irow(x, ...)

isld(x, ...)

isqr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{Either\code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more shape properties (i.e., from \code{sss_vals()}). Shape properties in \code{spec} may be pipe-delimited. If there are multiple shape properties in \code{spec}, \code{x} is inspected for any match to any shape property in \code{spec}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\itemize{
\item \strong{\code{sss}}: a character scalar/vector.
\item \strong{\code{sss_vals}}: a character vector.
\item \strong{\verb{ixxx, isss, is_sss_spec}}: a logical scalar.
}
}
\description{
Shape property values, names, and qualifying objects are as follows:
\itemize{
\item \strong{\code{'emp'}} (empty): non-\code{NULL}, length-\code{1} objects.
\item \strong{\code{'pnt'}} (point): \verb{1 x 1} data/frames and arrays, vectors, \link[=ivls]{vlists} with \code{1} element.
\item \strong{\code{'col'}} (column): \verb{1 x 2+} data.frames and matrices.
\item \strong{\code{'row'}} (row): \verb{2+ x 1} data.frames and matrices.
\item \strong{\code{'lin'}} (linear): length \verb{2+} vectors and vlists, length \verb{2+} arrays elements in just \code{1} dimension, and column and row data.frames and matrices.
\item \strong{\code{'rct'}} (rectangular): \verb{2+ x 2+} data.frames and matrices.
\item \strong{\code{'sqr'}} (square): \verb{n × n} matrices where \code{n > 1}.
\item \strong{\code{'sld'}} (solid): arrays with \verb{2+} elements in \verb{3+} dimensions.
}
\strong{Functions}
\itemize{
\item \strong{\code{sss}}: gets a vector of shape properties applicable to \code{x}.
\item \strong{\code{isss}}: evaluates whether \code{x} has one or more (possibly pipe-delimited) shape properties in \code{spec}, subject to any restrictions in \code{...}.
\item \strong{\code{ixxx}}: evaluates whether \code{x} has the shape property \code{xxx} subject to any restrictions in \code{...} (where \code{xxx} is a placeholder for any shape property).
\item \strong{\code{sss_props}}: gets a vector of all possible shape properties.
\item \strong{\code{is_sss_spec}}: evaluates whether \code{spec} is a valid shape property specification.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\itemize{
\item \strong{\code{n}}: a vector of valid lengths/numbers of elements.
\item \strong{\code{nr}}: a vector of valid numbers of rows.
\item \strong{\code{nc}}: a vector of valid numbers of columns.
\item \strong{\code{min}}: a scalar minimum valid length/number of element.
\item \strong{\code{minr}}: a scalar minimum valid number of rows.
\item \strong{\code{minc}}: a scalar minimum valid number of columns.
\item \strong{\code{max}}: a scalar maximum valid length/number of element.
\item \strong{\code{maxr}}: a scalar maximum valid number of rows.
\item \strong{\code{maxc}}: a scalar maximum valid number of columns.
\item \strong{\code{vals}}: a vector of valid values.
\item \strong{\code{lt}}: a scalar less-than (exclusive upper) bound.
\item \strong{\code{le}}: a scalar less-than-or-equal (inclusive upper) bound
\item \strong{\code{ge}}: a scalar greater-than-or-equal (inclusive lower) bound.
\item \strong{\code{gt}}: a scalar greater-than bound (exclusive lower) bound.
}
}

\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()}
}
\concept{props}
