% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sss}
\alias{sss}
\alias{sss_props}
\alias{is_sss_spec}
\alias{isss}
\alias{icol}
\alias{iemp}
\alias{ilin}
\alias{ipnt}
\alias{irct}
\alias{irow}
\alias{isld}
\alias{isqr}
\title{\code{Shape} properties}
\usage{
sss(x)

sss_props()

is_sss_spec(spec)

isss(x, spec, ...)

icol(x, ...)

iemp(x, ...)

ilin(x, ...)

ipnt(x, ...)

irct(x, ...)

irow(x, ...)

isld(x, ...)

isqr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{spec}{Either \code{NULL} or a \link[=cmp_chr_scl]{complete character vec} containing one or more \code{shape} properties (i.e., from \code{sss_vals()}). \code{Shape} properties in \code{spec} may be pipe-delimited. If there are multiple \code{shape} properties in \code{spec}, \code{x} is inspected for any match to any shape property in \code{spec}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\tabular{rl}{
  \code{is_sss_spec}   \tab A logical scalar.
\cr      \code{iSSS}   \tab   
\cr      \code{isss}   \tab   
\cr               \tab   
\cr  \code{sss_vals}   \tab A character vector.
\cr       \code{sss}   \tab   
}
}
\description{
\code{Shape} properties are as defined in the following table: \tabular{rcl}{
  \code{'emp'}   \tab empty  \tab   \code{0} length (but not \code{NULL}).
\cr \code{'pnt'}   \tab point  \tab   \code{1} element or cell\eqn{^1}.
\cr \code{'col'}   \tab column \tab   \code{1} row by \verb{2+} columns.
\cr \code{'row'}   \tab row    \tab   \verb{2+} rows by \code{1} column.
\cr \code{'lin'}   \tab linear \tab   \verb{2+} positions in 1 dimension\eqn{^2}.
\cr \code{'rct'}   \tab rect   \tab   \verb{2+} positions in \code{2} dimensions.
\cr \code{'sqr'}   \tab square \tab   \verb{N × N} matrix where \verb{N ≥ 2}.
\cr \code{'sld'}   \tab solid  \tab   \verb{2+} positions in \verb{3+} dimensions.
}
  \eqn{^{1.}} \verb{1x1} data.frames; length-\code{1} arrays, vectors, \link[=ivls]{vlists}.
\cr
  \eqn{^{2.}} Length-\verb{2+} vectors, arrays of \verb{2+} positions in \code{1} dimension, row/column data.frames.
\cr\cr
\strong{Shape property functions}
\tabular{rl}{
  \code{is_sss_spec}   \tab Is \code{spec} a \code{shape} property spec?
\cr \code{sss_props}   \tab Gets all possible \code{shape} properties.
\cr      \code{isss}   \tab Does \code{x} match \code{shape} spec \code{spec}?
\cr      \code{iSSS}   \tab Does \code{x} match the \code{shape} property \code{SSS}?
\cr       \code{sss}   \tab Gets all \code{shape} properties matched to \code{x}.
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab   
\cr  \verb{lt, le, ge, gt}   \tab Scalar less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab   
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab   
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
is_sss_spec("emp|pnt")
sss_props()
sss(matrix(letters, nrow = 1))
isss(matrix(letters, nrow = 2), "rct")
ipnt(data.frame(a = 1))
}
\seealso{
Other props: 
\code{\link{as_mmm}},
\code{\link{bbb_ccc}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{ccc}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()}
}
\concept{props}
