% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss_ccc.R
\encoding{UTF-8}
\name{sss_ccc}
\alias{sss_ccc}
\alias{sss_ccc_funs}
\alias{emp_arr}
\alias{emp_dtf}
\alias{emp_gen}
\alias{emp_mat}
\alias{emp_vec}
\alias{emp_vls}
\alias{pnt_arr}
\alias{pnt_dtf}
\alias{pnt_gen}
\alias{pnt_mat}
\alias{pnt_scl}
\alias{pnt_vec}
\alias{pnt_vls}
\alias{lin_arr}
\alias{lin_dtf}
\alias{lin_gen}
\alias{lin_mat}
\alias{linMVC}
\alias{lin_vec}
\alias{lin_vls}
\alias{col_dtf}
\alias{col_mat}
\alias{row_dtf}
\alias{row_mat}
\alias{rct_dtf}
\alias{rct_mat}
\alias{sqr_mat}
\alias{sld_arr}
\title{Shape + xclass combination properties}
\usage{
sss_ccc(x, sss, ccc, ...)

sss_ccc_funs()

emp_arr(x, ...)

emp_dtf(x, ...)

emp_gen(x, ...)

emp_mat(x, ...)

emp_vec(x, ...)

emp_vls(x, ...)

pnt_arr(x, ...)

pnt_dtf(x, ...)

pnt_gen(x, ...)

pnt_mat(x, ...)

pnt_scl(x, ...)

pnt_vec(x, ...)

pnt_vls(x, ...)

lin_arr(x, ...)

lin_dtf(x, ...)

lin_gen(x, ...)

lin_mat(x, ...)

linMVC(x, ...)

lin_vec(x, ...)

lin_vls(x, ...)

col_dtf(x, ...)

col_mat(x, ...)

row_dtf(x, ...)

row_mat(x, ...)

rct_dtf(x, ...)

rct_mat(x, ...)

sqr_mat(x, ...)

sld_arr(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{sss}{A character scalar single basic property from \code{link{sss_props}()}.}

\item{ccc}{A character scalar single xclass property from \code{link{ccc_props}()}}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \code{sss_ccc_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{\{sss\}_\{ccc\}, sss_ccc}
}
\description{
These functions check for the combination of \link[=sss]{shape} and \link[=ccc]{xlass} properties.
}
\details{
\tabular{ll}{  \code{sss_ccc_funs}   \tab What \link[=sss]{shape} + \link[=ccc]{xclass} combination \link[=prop_funs]{property functions} are there?                                                                                                   \cr   \tab   \cr
  \verb{\{sss\}_\{ccc\}}    \tab Is \code{x} a match to single shape property \code{'{sss}'} and single xclass property \code{'{ccc}'}, where \code{{sss}} and \code{{ccc}} are placeholders for any given shape property and any given xclass property, respectively? \cr   \tab   \cr
  \code{sss_ccc}        \tab Is \code{x} a match to the single shape and xclass properties in \code{bbb} and \code{ccc}, respectively?                                                                                                                                  }
\cr Some combinations of shape and xclass are nonsensical. That includes:
\itemize{\item Shapes \code{emp} (empty) and \code{pnt} (point) combined with xclasses \code{mvc} (multivec) or \code{scl} (scalar).
\item Shapes \code{row}, \code{col} (column), or \code{rct} (rectangular) combined with xclasses other than \code{dtf} (data.frame) or \code{mat} (matrix),
\item Shape \code{sqr} (square) combined with any xclass other than \code{mat} (matrix).}
\cr Nonsensical combinations do not have corresponding \code{sss_ccc} property functions.
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
sss_cccfuns()
sss_cccprops()
sss_ccc(letters, "lin", "vec")
sss_ccc(1, "emp", "vec")
lin_gen(letters)
pnt_vec(1)
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{unq_mmm_ccc}()},
\code{\link{unq_mmm}()}
}
\concept{props}
