% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats0.R
\name{stats0}
\alias{stats0}
\alias{sum0}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\title{Compute stats ignoring \code{NA} values}
\usage{
sum0(...)

sd0(...)

cor0(x, y = NULL)

cov0(x, y = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x)

rsds0(x)

csums0(x)

rsums0(x)

cmeans0(x)

rmeans0(x)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x}{A \link[=cmp_num_vec]{complete numeric vec} or a \link[=cmp_num_mat]{complete numeric matrix}.}

\item{y}{An optional complete numeric vec or a complete numeric matrix.}

\item{n}{Positive whole-number scalar size of the complete set each subset is drawn from.}
}
\value{
\itemize{
\item \strong{\verb{sum0, sd0, var0, min0, max0, mean0}}: a numeric scalar.
\item \strong{\verb{cor0, cov0, csds0, rsds0, pmin0, pmax0, csums0, rsums0, cmeans0, rmeans0}}: a numeric scalar or vector.
}
}
\description{
\itemize{
\item \strong{\code{sum0}}: calculates sum of non-\code{NA} atomic values in \code{...}.
\item \strong{\code{sd0}}: calculates standard deviation of non-\code{NA} atomic values in \code{...}.
\item \strong{\code{var0}}: calculates variance of non-\code{NA} atomic values in \code{...}.
\item \strong{\code{min0}}: calculates minimum of non-\code{NA} atomic values in \code{...}.
\item \strong{\code{max0}}: calculates maximum of non-\code{NA} atomic values in \code{...}.
\item \strong{\code{mean0}}: calculates mean of all non-\code{NA} atomic values in \code{...}.
\item \strong{\code{cor0}}: calls \code{cor(x, y, use = "complete.obs")}.
\item \strong{\code{cov0}}: calls \code{cov(x, y, use = "complete.obs")}.
\item \strong{\code{csds0}}: calls \code{apply(x, 2, sd, na.rm = T)}.
\item \strong{\code{rsds0}}: calls \code{apply(x, 1, sd, na.rm = T)}.
\item \strong{\code{pmin0}}: calls \code{pmin(..., na.rm = T)}
\item \strong{\code{pmax0}}: calls \code{pmax(..., na.rm = T)}
\item \strong{\code{csums0}}: calls \code{colSums(x, na.rm = T)}.
\item \strong{\code{rsums0}}: calls \code{rowSums(x, na.rm = T)}.
\item \strong{\code{cmeans0}}: calls \code{colMeans(x, na.rm = T)}.
\item \strong{\code{rmeans0}}: calls \code{rowMeans(x, na.rm = T)}.
}
}
\seealso{
Other missingness: 
\code{\link{na}()}
}
\concept{missingness}
