% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats0.R
\name{stats0}
\alias{stats0}
\alias{sum0}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\title{Compute stats ignoring \code{NA} values}
\usage{
sum0(...)

sd0(...)

cor0(x, y = NULL)

cov0(x, y = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x)

rsds0(x)

csums0(x)

rsums0(x)

cmeans0(x)

rmeans0(x)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x}{A \link[=cmp_num_vec]{complete numeric vec} or a \link[=cmp_num_mat]{complete numeric matrix}.}

\item{y}{An optional complete numeric vec or a complete numeric matrix.}

\item{n}{Positive whole-number scalar size of the complete set each subset is drawn from.}
}
\value{
\tabular{rl}{
  \verb{cmeans0,rmeans0}   \tab A numeric vector.
\cr \verb{csums0,rsums0}   \tab  
\cr   \verb{csds0,rsds0}   \tab  
\cr   \verb{pmin0,pmax0}   \tab  
\cr                   \tab  
\cr    \verb{mean0,sum0}   \tab A numeric scalar.
\cr     \verb{min0,max0}   \tab  
\cr      \verb{sd0,var0}   \tab  
\cr                   \tab  
\cr     \verb{cor0,cov0}   \tab A numeric vector/matrix.
}
}
\description{
\tabular{rl}{
  \code{cmeans0}   \tab Calls \verb{colMeans(x, na.rm = T)}.
\cr \code{rmeans0}   \tab Calls \verb{rowMeans(x, na.rm = T)}.
\cr  \code{csums0}   \tab Calls \verb{colSums(x, na.rm = T)}.
\cr  \code{rsums0}   \tab Calls \verb{rowSums(x, na.rm = T)}.
\cr   \code{csds0}   \tab Calls \verb{apply(x, 2, sd, na.rm = T)}.
\cr   \code{rsds0}   \tab Calls \verb{apply(x, 1, sd, na.rm = T)}.
\cr   \code{pmin0}   \tab Calls \verb{pmin(..., na.rm = T)}.
\cr   \code{pmax0}   \tab Calls \verb{pmax(..., na.rm = T)}.
\cr   \code{mean0}   \tab Calls \verb{mean(av(...), na.rm = T)}.
\cr    \code{min0}   \tab Calls \verb{min(av(...), na.rm = T)}.
\cr    \code{max0}   \tab Calls \verb{max(av(...), na.rm = T)}.
\cr    \code{sum0}   \tab Calls \verb{sum(av(...), na.rm = T)}.
\cr    \code{var0}   \tab Calls \verb{var(av(...), na.rm = T)}.
\cr     \code{sd0}   \tab Calls \verb{sd(av(...), na.rm = T)}.
\cr    \code{cor0}   \tab Calls \verb{cor(x, y, use = "complete.obs")}.
\cr    \code{cov0}   \tab Calls \verb{cov(x, y, use = "complete.obs")}.
}
}
\examples{
vec1. <- sample(0:99, 10)
vec2. <- sample(0:99, 20)
vec3. <- sample(0:99, 20)
mat1. <- sample(0:99, 100)

vec1.[sample(1:10, 1)] <- NA
vec2.[sample(1:20, 1)] <- NA
vec3.[sample(1:20, 2)] <- NA
mat1.[sample(1:100, 5)] <- NA
mat1. <- matrix(mat1., nrow = 10)
rownames(mat1.) <- paste0("R", 1:10)
colnames(mat1.) <- paste0("C", 1:10)
dtf1. <- as.data.frame(mat1.)
x. <- list(v1 = vec1., v2 = vec2., v3 = vec3., m = mat1., d = dtf1.)

vec1.
vec2.
vec3.
mat1.
dtf1.
x.

matrix(c(cmeans0(mat1.), cmeans0(dtf1.)), ncol = 2)
matrix(c(rmeans0(mat1.), rmeans0(dtf1.)), ncol = 2)
matrix(c(csums0(mat1.), csums0(dtf1.)), ncol = 2)
matrix(c(rsums0(mat1.), rsums0(dtf1.)), ncol = 2)
matrix(c(csds0(mat1.), csds0(dtf1.)), ncol = 2)
matrix(c(rsds0(mat1.), rsds0(dtf1.)), ncol = 2)

pmin0(vec2., vec3.)
pmax0(vec2., vec3.)

cor0(vec2., vec3.)
cov0(vec2., vec3.)

cor0(mat1.)
cov0(mat1.)

list(mean = mean0(x.), min = min0(x.), max = max0(x.), sum = sum0(x.), var = var0(x.), sd = sd0(x.))
}
\seealso{
Other missingness: 
\code{\link{na}()}
}
\concept{missingness}
\concept{unchecked}
