% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\docType{data}
\name{stats0}
\alias{stats0}
\alias{sum0}
\alias{sd0}
\alias{cor0}
\alias{cov0}
\alias{min0}
\alias{max0}
\alias{var0}
\alias{pmin0}
\alias{pmax0}
\alias{mean0}
\alias{csds0}
\alias{rsds0}
\alias{csums0}
\alias{rsums0}
\alias{cmeans0}
\alias{rmeans0}
\alias{sets}
\title{Compute stats ignoring NA}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
stats0

sum0(...)

sd0(...)

cor0(x, y = NULL)

cov0(x, y = NULL)

min0(...)

max0(...)

var0(...)

pmin0(...)

pmax0(...)

mean0(...)

csds0(x)

rsds0(x)

csums0(x)

rsums0(x)

cmeans0(x)

rmeans0(x)

sets(n)
}
\arguments{
\item{...}{Scalars, vectors, or matrices. Reduced a single vector of atomic
values for \code{sd0}, \code{min0}, \code{max0}, and \code{mean0}.}

\item{x}{\link[=cmp_num_vec]{Complete numeric vec} or
\link[=cmp_num_mat]{complete numeric matrix}.}

\item{y}{Optional \link[=cmp_num_vec]{complete numeric vec} or
\link[=cmp_num_mat]{complete numeric matrix}.}

\item{n}{Positive whole-number scalar size of the complete set each subset is
drawn from.}
}
\description{
Compute stats ignoring NA
}
\section{Functions}{
\itemize{
\item \code{sum0()}: Standard deviation of non-\code{NA} atomic elements
contained in \code{...} as a whole.

\item \code{sd0()}: Standard deviation of non-\code{NA} atomic elements
contained in \code{...} as a whole.

\item \code{cor0()}: Correlation of complete cases in \code{x.} and/or
\code{y.}.

\item \code{cov0()}: Covariance of complete cases in \code{x.} and/or
\code{y.}.

\item \code{min0()}: Minimum value of non-\code{NA} atomic elements contained
in \code{...} as a whole.

\item \code{max0()}: Maximum value of non-\code{NA} atomic elements contained
in \code{...} as a whole.

\item \code{var0()}: Variance of non-\code{NA} atomic elements contained in
\code{...} as a whole.

\item \code{pmin0()}: Pairwise minimum values ignoring \code{NA} values.

\item \code{pmax0()}: Pairwise maximum values ignoring \code{NA} values.

\item \code{mean0()}: Mean value of non-\code{NA} atomic elements contained in
\code{...} as a whole.

\item \code{csds0()}: Column standard deviations of \code{x} removing
non-\code{NA} values separately for each column.

\item \code{rsds0()}: Row standard deviations of \code{x} removing non-\code{NA}
values separately for each row.

\item \code{csums0()}: Column sums of \code{x} removing non-\code{NA} values
separately for each column.

\item \code{rsums0()}: Row sums of \code{x} removing non-\code{NA} values
separately for each row.

\item \code{cmeans0()}: Column means of \code{x} removing non-\code{NA} values
separately for each column.

\item \code{rmeans0()}: Row means of \code{x} removing non-\code{NA} values
separately for each row.

\item \code{sets()}: Create a list containing elements of indices to create all
possible combinations of 1 or more elements of an object of size \code{n}.

}}
\seealso{
Other math: 
\code{\link{hexdec}},
\code{\link{inc}()}

Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put.}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{swap}()},
\code{\link{xb}()}
}
\concept{extensions}
\concept{math}
\keyword{datasets}
