% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{style_aliases}
\alias{style_aliases}
\alias{.bld}
\alias{.itl}
\alias{.pln}
\alias{.def}
\alias{.und}
\alias{.res}
\alias{.sts}
\title{Style Aliases}
\usage{
style_aliases()

.bld()

.itl()

.pln()

.def()

.und()

.res()

.sts()
}
\value{
A character vector.
}
\description{
Each style is assigned a set of aliases, all of which will work in this package.
}
\section{Functions}{
\itemize{
\item \code{.bld()}: All aliases for bold style.

\item \code{.itl()}: All aliases for italic style.

\item \code{.pln()}: All aliases for plain style

\item \code{.def()}: All aliases for default style

\item \code{.und()}: All aliases for underline style.

\item \code{.res()}: All aliases for resetting styles and colors back to default values.

\item \code{.sts()}: All aliases for styles.

}}
\examples{
.bld()
.itl()
.pln()
.def()
.und()
.res()
.sts()
}
