% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\encoding{UTF-8}
\name{swap}
\alias{swap}
\alias{swap_vals}
\alias{swap_elt}
\alias{swap_rows}
\alias{swap_cols}
\title{Swap by Row Index, Column Index, Element Index, and Atomic Value.}
\usage{
swap(X, DIM = 0, From = NULL, To = NULL, Old = NULL, New = NULL, All = FALSE)

swap_vals(X, Old, New, All = FALSE)

swap_elt(X, From, To)

swap_rows(X, From, To)

swap_cols(X, From, To)
}
\arguments{
\item{X}{A non-empty data.frame, matrix, vlist (?uj::VLS), or an atomic object of any dimension.}

\item{DIM}{Dimension on which swapping occurs. \code{0} indicates swapping by element index of a non-empty vector, vlist, or other atomic object. \code{1} and \code{2} indicate, respectively, swapping rows and columns of a data.frame or matrix.}

\item{From, To}{Uniquely valued non-zero whole number vectors indexing source (\code{From}) and destination (\code{To}) positions, rows, or columns. \code{From < 0} and \code{To < 0} index from the last value rather than the first.}

\item{Old}{\link[=unq_atm_vec]{A uniquely valued atomic vec} of values to be replaced.}

\item{New}{\link[=atm_vec]{An atomic vec} of replacement values. \code{length(New)} must be in \code{c(1, length(Old))}.}

\item{All}{\code{TRUE} or \code{FALSE} indicating whether all values of \code{X} must be contained in \code{Old} (ignored when \code{Old} and \code{New} are \code{NULL}).}
}
\value{
An object of the same class and dimensions as \code{X}.
}
\description{
Swap by Row Index, Column Index, Element Index, and Atomic Value.
}
\examples{
letters[1:6]
swap(letters[1:6], c("a", "b", "c"), ".")
swap(letters[1:6], c("a", "b", "c"), c("-A-", "-B-", "-C-"))
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{value_exists}()}

Other forks: 
\code{\link{case}()},
\code{\link{fork}()}
}
\concept{extensions}
\concept{forks}
