% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\encoding{UTF-8}
\name{swap}
\alias{swap}
\title{Swap atomic values.}
\usage{
swap(x, old, new, all = FALSE)
}
\arguments{
\item{x}{An atomic object.}

\item{old, new}{\link[=atm_vec]{Atomic vecs} of unique values to be replaced and replacement values, respectively. \code{length(new)} must be in \code{c(1, length(old))}.}

\item{all}{A non-\code{NA} logical scalar indicating whether all values of \code{x} must be contained in \code{old}.}
}
\value{
An atomic object of the same dimensions as \code{x}.
}
\description{
Swap atomic values.
}
\examples{
letters[1:6]
swap(letters[1:6], c("a", "b", "c"), ".")
swap(letters[1:6], c("a", "b", "c"), c("-A-", "-B-", "-C-"))
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}}

Other forks: 
\code{\link{case}()},
\code{\link{fork}()}
}
\concept{extensions}
\concept{forks}
