% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\encoding{UTF-8}
\name{swap}
\alias{swap}
\alias{swap_vals}
\alias{swap_elt}
\alias{swap_rows}
\alias{swap_cols}
\title{Swap by Row Index, Column Index, Element Index, and Atomic Value.}
\usage{
swap(x, from = NULL, to = NULL, old = NULL, new = NULL, .dim = 0, .all = FALSE)

swap_vals(x, old, new, .all = FALSE)

swap_elt(x, from, to)

swap_rows(x, from, to)

swap_cols(x, from, to)
}
\arguments{
\item{x}{A non-empty data.frame, matrix, vlist (?VLS), or an atomic object of any dimension.}

\item{from, to}{Uniquely valued non-zero whole number vectors indexing source (\code{from}) and destination (\code{to}) positions, rows, or columns. \code{from < 0} and \code{to < 0} index from the last value rather than the first.}

\item{old}{\link[=unq_atm_vec]{A uniquely valued atomic vec} of values to be replaced.}

\item{new}{\link[=atm_vec]{An atomic vec} of replacement values. \code{length(new)} must be in \code{c(1, length(old))}.}

\item{.dim}{Dimension on which swapping occurs. \code{0} indicates swapping by element index of a non-empty vector, vlist, or other atomic object. \code{1} and \code{2} indicate, respectively, swapping rows and columns of a data.frame or matrix.}

\item{.all}{\code{TRUE} or \code{FALSE} indicating whether all values of \code{x} must be contained in \code{old} (ignored when \code{old} and \code{new} are \code{NULL}).}
}
\value{
An object of the same class and dimensions as \code{x}.
}
\description{
Swap by Row Index, Column Index, Element Index, and Atomic Value.
}
\examples{
letters[1:6]
swap(letters[1:6], c("a", "b", "c"), ".")
swap(letters[1:6], c("a", "b", "c"), c("-A-", "-B-", "-C-"))
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{flex_rbind}()},
\code{\link{fork}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{value_exists}()}

Other forks: 
\code{\link{case}()},
\code{\link{fork}()}
}
\concept{extensions}
\concept{forks}
