% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\encoding{UTF-8}
\name{swap}
\alias{swap}
\title{Swap atomic values.}
\usage{
swap(X, Old, New, All = FALSE)
}
\arguments{
\item{X}{An atomic object.}

\item{Old, New}{\link[=atm_vec]{Atomic vecs} of unique values to be replaced and replacement values, respectively. \code{length(New)} must be in \code{c(1, length(Old))}.}

\item{All}{\code{TRUE} or \code{FALSE} indicating whether All values of \code{X} must be contained in \code{Old}.}
}
\value{
An atomic object of the same dimensions as \code{X}.
}
\description{
Swap atomic values.
}
\examples{
letters[1:6]
swap(letters[1:6], c("a", "b", "c"), ".")
swap(letters[1:6], c("a", "b", "c"), c("-A-", "-B-", "-C-"))
}
\seealso{
Other extensions: 
\code{\link{N}()},
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{purge}()},
\code{\link{rd}()},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{sum0}()},
\code{\link{value_exists}()}

Other forks: 
\code{\link{case}()},
\code{\link{fork}()}
}
\concept{extensions}
\concept{forks}
