% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdec.R
\encoding{UTF-8}
\name{to_dec}
\alias{to_dec}
\alias{to_hex}
\alias{todec}
\alias{tohex}
\title{Convert non-negative whole numbers between decimal and hexadecimal}
\usage{
to_dec(X)

to_hex(X)

todec(X)

tohex(X)
}
\arguments{
\item{X}{A \link[=cmp_nnw]{complete non-negative whole-number object} or a \link[=cmp_chr]{complete character vec} containing only non-negative whole-number hexademical values.}
}
\value{
\strong{A non-negative whole number object}                                 \cr \verb{to_dec, todec}
\cr\cr  \strong{A character object of non-negative whole number hexadecimal values} \cr \verb{to_hex, tohex}
}
\description{
\tabular{ll}{  \verb{to_dec, todec}   \tab Converts hexadecimal \link[=cmp_chr]{complete character object} \code{X} from hexadecimal to decimal. \code{X} may be formatted as either \code{'hhh...'} or \code{'#hhh...'} where \code{h} is a placeholder for a hexidecimal digit in either upper or lower case. \cr   \tab   \cr
  \verb{to_hex, tohex}   \tab Converts \link[=cmp_psw]{complete positive whole-number} \code{X} to hexadecimal formatted as \code{'#HHH...'} where \code{H} is a placeholder for an uppercase hexadecimal digit.                                                                                        }
}
\examples{
egDEC <- sample(1:100, 5)
egHEX <- c("10", "#B25", "2bc4a", "#FF")
egDEC
egHEX
to_dec(egHEX)
to_hex(egDEC)
}
\seealso{
Other conversions: 
\code{\link{rng2lev}()},
\code{\link{tocase}()},
\code{\link{u2u}()}
}
\concept{conversions}
