% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocase.R
\name{tocase}
\alias{tocase}
\alias{tosentence}
\alias{totitle}
\alias{lc}
\alias{sc}
\alias{tc}
\alias{uc}
\title{Convert String Case}
\usage{
tocase(case, ..., a. = T)

tosentence(..., a. = T)

totitle(..., a. = T)

lc(..., a. = T)

sc(..., a. = T)

tc(..., a. = T)

uc(..., a. = T)
}
\arguments{
\item{case}{\code{'l'}, \code{'s'}, \code{'t'}, or \code{'u'} to indicate
lower, sentence, title, or upper case. Case insensitive.}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{a}{.\code{TRUE} or \code{FALSE} indicating whether to
\link[=a.]{atomize} \code{...} before processing.}
}
\description{
Convert strings to lower, upper, sentence, and title cases
(supplements \code{tolower} and \code{toupper}).

\strong{\code{tocase}}
\cr Converts to the case specified in \code{case}.
\cr\cr
\strong{\code{tosentence, sc}}
\cr Converts to sentence case via \code{\link[stringr]{str_to_sentence}}.
\cr\cr
\strong{\code{totitle, tc}}
\cr Converts to sentence case via \code{\link[stringr]{str_to_title}}.
\cr\cr
\strong{\code{lc}}
\cr Converts to lower case via \code{\link[base]{tolower}}.
\cr\cr
\strong{\code{uc}}
\cr Converts to upper case via \code{\link[base]{toupper}}.
}
\examples{
TestTitle    <- "This is a Title"
TestSentence <- "This is a sentence."
TestWords    <- "this is an ordered group of words"
TestSet      <- c("a", "set", "of", "words")
TestVector   <- c(TestTitle, TestSentence, TestWords, TestSet)
lc(TestVector)
lc(TestTitle, TestSentence, TestWords, TestSet)
sc(TestTitle, TestSentence, TestWords, TestSet)
tc(TestTitle, TestSentence, TestWords, TestSet)
uc(TestTitle, TestSentence, TestWords, TestSet)
scase(TestTitle, TestSentence, TestWords, TestSet, case = "LOWER")
scase(TestVector, case = "l")
scase(TestVector, case = "s")
scase(TestTitle, TestSentence, TestWords, TestSet, case = "T")
scase(TestVector, case = "UPPER")
}
\seealso{
Other strings: 
\code{\link{Oxford}},
\code{\link{delim}},
\code{\link{ipat}()},
\code{\link{pgrid}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{weave}()}
}
\concept{strings}
