% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_case.R
\encoding{UTF-8}
\name{tocase}
\alias{tocase}
\alias{tosent}
\alias{totitle}
\alias{lc}
\alias{sc}
\alias{tc}
\alias{uc}
\title{Convert string case}
\usage{
tocase(case, ...)

tosent(...)

totitle(...)

lc(...)

sc(...)

tc(...)

uc(...)
}
\arguments{
\item{case}{A case insensitive \link[=cmp_ch1_scl]{complete onechar scalar}. Either \code{'l'}, \code{'s'}, \code{'t'}, or \code{'u'} to indicate lower, sentence, title, or upper case, respectively.}

\item{...}{An arbitrary number of non-empty character arguments.}
}
\value{
A character vector.
}
\description{
Convert among lower, sentence, title, and upper cases.
}
\details{
Functions in this family convert string case as follows:
\tabular{ll}{  \verb{totitle, tc}   \tab Convert to \link[stringr:str_to_sentence]{title case}.    \cr
  \verb{tosent, sc}    \tab Convert to \link[stringr:str_to_sentence]{sentence case}. \cr
  \code{tocase}        \tab Convert to case \code{case}.                                   \cr
  \code{lc}            \tab Convert to \link[base:tolower]{lower case}.               \cr
  \code{uc}            \tab Convert to \link[base:toupper]{upper case}.                 }
}
\examples{
egSentence <- "This is a sentence."
egTitle <- "This is a Title"
egWords <- c("word1", "word2")
egList <- list(egWords, "word3")

egWords
egList

tc(egSentence, egTitle)
sc(egSentence, egTitle)
lc(egSentence, egTitle)
uc(egSentence, egTitle)

tosent(egSentence, egTitle, egWords, egList)
totitle(egSentence, egTitle, egWords, egList)

tocase("s", egTitle, egList)
tocase("u", egList)
}
\seealso{
Other conversions: 
\code{\link{hex_dec_funs}()},
\code{\link{rng2lev}()},
\code{\link{u2u}()}

Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}()},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}()},
\code{\link{markdown}()},
\code{\link{maxnch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{weave}()}
}
\concept{conversions}
\concept{strings}
