% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocase.R
\name{tocase}
\alias{tocase}
\alias{tosentence}
\alias{totitle}
\alias{lc}
\alias{sc}
\alias{tc}
\alias{uc}
\title{Convert string case}
\usage{
tocase(case., ..., a. = T)

tosentence(..., a. = T)

totitle(..., a. = T)

lc(..., a. = T)

sc(..., a. = T)

tc(..., a. = T)

uc(..., a. = T)
}
\arguments{
\item{case.}{A \link[=cmp_ch1_scl]{complete onechar scalar}. Either \code{'l'}, \code{'s'}, \code{'t'}, or \code{'u'} to indicate lower, sentence, title, or upper case, respectively (case insensitive).}

\item{...}{An arbitrary number of atomic arguments to be processed.}

\item{a.}{A non-\code{NA} logical scalar indicating whether to \link[=a]{atomize} \code{...} before processing.}
}
\value{
An atomic character object or a \link[=chr_vls]{character vlist}.
}
\description{
\itemize{
\item \strong{\code{lc}}: converts to lower case via \code{\link[base:tolower]{base::tolower}}.
\item \strong{\code{uc}}: converts to upper case via \code{\link[base:toupper]{base::toupper}}.
\item \strong{\code{tocase}}: converts among lower, upper, sentence, and title case based on the value of the argument \code{case}.
\item \strong{\verb{totitle, tc}}: convert to title case via \code{\link[stringr:str_to_title]{stringr::str_to_title}}.
\item \strong{\verb{tosentence, sc}}: convert to sentence case via \code{\link[stringr:str_to_sentence]{stringr::str_to_sentence}}.
}
}
\examples{
TestSentence <- "This is a sentence."
TestTitle    <- "This is a Title"
TestWords    <- "this is an ordered group of words"
TestSet      <- c("a", "set", "of", "words")
TestVector   <- c(TestTitle, TestSentence, TestWords, TestSet)
lc(TestVector)
lc(TestTitle, TestSentence, TestWords, TestSet)
sc(TestTitle, TestSentence, TestWords, TestSet)
tc(TestTitle, TestSentence, TestWords, TestSet)
uc(TestTitle, TestSentence, TestWords, TestSet)
scase(TestTitle, TestSentence, TestWords, TestSet, case = "LOWER")
scase(TestVector, case = "l")
scase(TestVector, case = "s")
scase(TestTitle, TestSentence, TestWords, TestSet, case = "T")
scase(TestVector, case = "UPPER")
}
\seealso{
Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{max_nch}()},
\code{\link{mkd}},
\code{\link{mkstr}},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spacing}},
\code{\link{ss}()},
\code{\link{weave}()}
}
\concept{strings}
