% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocase.R
\encoding{UTF-8}
\name{tocase}
\alias{tocase}
\alias{tosent}
\alias{totitle}
\alias{lc}
\alias{sc}
\alias{tc}
\alias{uc}
\title{Convert string case}
\usage{
tocase(case., ...)

tosent(...)

totitle(...)

lc(...)

sc(...)

tc(...)

uc(...)
}
\arguments{
\item{case.}{A \link[=cmp_ch1_scl]{complete onechar scalar}. Either \code{'l'}, \code{'s'}, \code{'t'}, or \code{'u'} to indicate lower, sentence, title, or upper case, respectively (case insensitive).}

\item{...}{An arbitrary number of non-empty character arguments.}
}
\value{
A character vector.
}
\description{
Functions in this family convert string case as follows:
\tabular{rl}{
  \verb{totitle, tc} \tab   Convert to \link[stringr:str_to_sentence]{title case}.
\cr \verb{tosent, ts} \tab   Convert to \link[stringr:str_to_sentence]{sentence case}.
\cr         \code{lc} \tab   Convert to \link[base:tolower]{lower case}.
\cr         \code{uc} \tab   Convert to \link[base:toupper]{upper case}.
\cr     \code{tocase} \tab   Convert to case \code{case.}.
}
}
\examples{
sentence. <- "This is a sentence."
title. <- "This is a Title"
words. <- c("word1", "word2")
list. <- list(words., "word3")

words.
list.

tc(sentence., title.)
sc(sentence., title.)
lc(sentence., title.)
uc(sentence., title.)

tosent(sentence., title., words., list.)
totitle(sentence., title., words., list.)

tocase("s", title., list.)
tocase("u", list.)
}
\seealso{
Other conversions: 
\code{\link{hexdec}},
\code{\link{rng2lev}()},
\code{\link{u2u}()}

Other strings: 
\code{\link{blank}()},
\code{\link{chn}()},
\code{\link{delim}},
\code{\link{fsub}()},
\code{\link{gr}},
\code{\link{ipat}()},
\code{\link{makestr}},
\code{\link{markdown}},
\code{\link{max_nch}()},
\code{\link{ox}()},
\code{\link{pgrid}()},
\code{\link{revstr}()},
\code{\link{spaces}()},
\code{\link{ss}()},
\code{\link{weave}()}
}
\concept{conversions}
\concept{strings}
