% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt.R
\name{ttt}
\alias{ttt}
\alias{ttt_vals}
\alias{xnll}
\alias{xdef}
\alias{xnil}
\alias{xpop}
\alias{xatm}
\alias{xfun}
\alias{xrcr}
\alias{is_ttt}
\title{Fundamental Type (ttt) Property Family}
\usage{
ttt_vals()

xnll(x)

xdef(x)

xnil(x)

xpop(x)

xatm(x)

xfun(x)

xrcr(x)

ttt(x)

is_ttt(x, xxx, ...)
}
\arguments{
\item{x}{An object.}

\item{xxx}{A character scalar containing one or more values from
\code{ttt_vals()} separated by pipes and/or underscores ("."). Combinations
of fundamental types can be specified by separating them with underscores.
Separating fundamental types or combinations of fundamental types with
pipes will result in a value of \code{TRUE} if any of them applies to
\code{x}.}

\item{...}{Additional arguments to \code{\link{meets}} containing value and
element/row/column count restrictions.}
}
\value{
\code{ttt_vals} and \code{ttt} return a character scalar or
vector. All others return either \code{TRUE} or \code{FALSE}.
}
\description{
An object's fundamental type (ttt) is defined by its most
basic structural properties as described in the following table:
\tabular{lll}{
\emph{Value}   \tab\emph{Name}    \tab\emph{Requirements}
\cr\code{'nll'}\tab null          \tab\code{NULL}.
\cr\code{'def'}\tab defined       \tab not \code{NULL}.
\cr\code{'nil'}\tab nil-length    \tab length is 0 (includes \code{NULL}).
\cr\code{'pop'}\tab populated     \tab length is > 0.
\cr\code{'fun'}\tab function ref  \tab function object or character scalar
containing a function name.
\cr\code{'atm'}\tab atomic        \tab atomic.
\cr\code{'rcr'}\tab recursive     \tab list-like (tibble, lists, or objects
containing an element that
is not an atomic scalar).
}
}
\details{
\strong{\code{ttt_vals}}
\cr Gets all valid fundamental type property values.
\cr\cr
\strong{\code{ttt}}
\cr Gets a vector of fundamental properties from \code{ttt_vals()} that are
applicable to \code{x}.
\cr\cr
\strong{\code{is_ttt}}
\cr Evaluates whether \code{xxx} contains any fundamental type property (or
combination of fundamental type properties) applicable to \code{x}.
\cr\cr
\strong{\code{xttt}}
\cr All other functions evaluate whether \code{x} has the fundamental type
property represented by \code{ttt}.
\cr\cr
\strong{Additional Arguments in \code{...}}
\cr Submitting additional arguments to \code{is_ttt} via \code{...} allows
for checking not just the fundamental but whether length, number of rows,
number of columns, and element values meet flexible criteria.
}
\seealso{
Other props: 
\code{\link{ccc}()},
\code{\link{cmp_ccc}},
\code{\link{cmp_mmm_agn}},
\code{\link{cmp_mmm_arr}},
\code{\link{cmp_mmm_atb}},
\code{\link{cmp_mmm_avl}},
\code{\link{cmp_mmm_avt}},
\code{\link{cmp_mmm_gen}},
\code{\link{cmp_mmm_mat}},
\code{\link{cmp_mmm_mvc}},
\code{\link{cmp_mmm_scl}},
\code{\link{cmp_mmm_vec}},
\code{\link{cmp_mmm}},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{fff}()},
\code{\link{meets}()},
\code{\link{mmm_agn}},
\code{\link{mmm_arr}},
\code{\link{mmm_atb}},
\code{\link{mmm_avl}},
\code{\link{mmm_avt}},
\code{\link{mmm_mat}},
\code{\link{mmm_mvc}},
\code{\link{mmm_scl}},
\code{\link{mmm_vec}},
\code{\link{mmm}()},
\code{\link{pop_xxx}},
\code{\link{sss}()},
\code{\link{xxx}()}
}
\concept{props}
