% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_mmm.R
\name{ttt_mmm.}
\alias{ttt_mmm.}
\alias{ttt_mmm_vals}
\alias{pop_atm}
\alias{pop_ch1}
\alias{pop_chr}
\alias{pop_clr}
\alias{pop_evn}
\alias{pop_fac}
\alias{pop_frc}
\alias{pop_ind}
\alias{pop_lgl}
\alias{pop_neg}
\alias{pop_ngw}
\alias{pop_nng}
\alias{pop_nnw}
\alias{pop_nps}
\alias{pop_npw}
\alias{pop_nst}
\alias{pop_num}
\alias{pop_odd}
\alias{pop_ord}
\alias{pop_pct}
\alias{pop_pos}
\alias{pop_ppn}
\alias{pop_psw}
\alias{pop_srt}
\alias{pop_str}
\alias{pop_uno}
\alias{pop_whl}
\alias{cmp_atm}
\alias{cmp_ch1}
\alias{cmp_chr}
\alias{cmp_clr}
\alias{cmp_evn}
\alias{cmp_fac}
\alias{cmp_frc}
\alias{cmp_ind}
\alias{cmp_lgl}
\alias{cmp_neg}
\alias{cmp_ngw}
\alias{cmp_nng}
\alias{cmp_nnw}
\alias{cmp_nps}
\alias{cmp_npw}
\alias{cmp_nst}
\alias{cmp_num}
\alias{cmp_odd}
\alias{cmp_ord}
\alias{cmp_pct}
\alias{cmp_pos}
\alias{cmp_ppn}
\alias{cmp_psw}
\alias{cmp_srt}
\alias{cmp_str}
\alias{cmp_uno}
\alias{cmp_whl}
\alias{emp_atm}
\alias{emp_ch1}
\alias{emp_chr}
\alias{emp_clr}
\alias{emp_evn}
\alias{emp_fac}
\alias{emp_frc}
\alias{emp_ind}
\alias{emp_lgl}
\alias{emp_neg}
\alias{emp_ngw}
\alias{emp_nng}
\alias{emp_nnw}
\alias{emp_nps}
\alias{emp_npw}
\alias{emp_nst}
\alias{emp_num}
\alias{emp_odd}
\alias{emp_ord}
\alias{emp_pct}
\alias{emp_pos}
\alias{emp_ppn}
\alias{emp_psw}
\alias{emp_srt}
\alias{emp_str}
\alias{emp_uno}
\alias{emp_whl}
\alias{rcr_atm}
\alias{rcr_ch1}
\alias{rcr_chr}
\alias{rcr_clr}
\alias{rcr_evn}
\alias{rcr_fac}
\alias{rcr_frc}
\alias{rcr_ind}
\alias{rcr_lgl}
\alias{rcr_neg}
\alias{rcr_ngw}
\alias{rcr_nng}
\alias{rcr_nnw}
\alias{rcr_nps}
\alias{rcr_npw}
\alias{rcr_nst}
\alias{rcr_num}
\alias{rcr_odd}
\alias{rcr_ord}
\alias{rcr_pct}
\alias{rcr_pos}
\alias{rcr_ppn}
\alias{rcr_psw}
\alias{rcr_srt}
\alias{rcr_str}
\alias{rcr_uno}
\alias{rcr_whl}
\title{Fundamental type plus extended mode properties}
\usage{
ttt_mmm.()

ttt_mmm_vals()

pop_atm(x)

pop_ch1(x)

pop_chr(x)

pop_clr(x)

pop_evn(x)

pop_fac(x)

pop_frc(x)

pop_ind(x)

pop_lgl(x)

pop_neg(x)

pop_ngw(x)

pop_nng(x)

pop_nnw(x)

pop_nps(x)

pop_npw(x)

pop_nst(x)

pop_num(x)

pop_odd(x)

pop_ord(x)

pop_pct(x)

pop_pos(x)

pop_ppn(x)

pop_psw(x)

pop_srt(x)

pop_str(x)

pop_uno(x)

pop_whl(x)

cmp_atm(x)

cmp_ch1(x)

cmp_chr(x)

cmp_clr(x)

cmp_evn(x)

cmp_fac(x)

cmp_frc(x)

cmp_ind(x)

cmp_lgl(x)

cmp_neg(x)

cmp_ngw(x)

cmp_nng(x)

cmp_nnw(x)

cmp_nps(x)

cmp_npw(x)

cmp_nst(x)

cmp_num(x)

cmp_odd(x)

cmp_ord(x)

cmp_pct(x)

cmp_pos(x)

cmp_ppn(x)

cmp_psw(x)

cmp_srt(x)

cmp_str(x)

cmp_uno(x)

cmp_whl(x)

emp_atm(x)

emp_ch1(x)

emp_chr(x)

emp_clr(x)

emp_evn(x)

emp_fac(x)

emp_frc(x)

emp_ind(x)

emp_lgl(x)

emp_neg(x)

emp_ngw(x)

emp_nng(x)

emp_nnw(x)

emp_nps(x)

emp_npw(x)

emp_nst(x)

emp_num(x)

emp_odd(x)

emp_ord(x)

emp_pct(x)

emp_pos(x)

emp_ppn(x)

emp_psw(x)

emp_srt(x)

emp_str(x)

emp_uno(x)

emp_whl(x)

rcr_atm(x)

rcr_ch1(x)

rcr_chr(x)

rcr_clr(x)

rcr_evn(x)

rcr_fac(x)

rcr_frc(x)

rcr_ind(x)

rcr_lgl(x)

rcr_neg(x)

rcr_ngw(x)

rcr_nng(x)

rcr_nnw(x)

rcr_nps(x)

rcr_npw(x)

rcr_nst(x)

rcr_num(x)

rcr_odd(x)

rcr_ord(x)

rcr_pct(x)

rcr_pos(x)

rcr_ppn(x)

rcr_psw(x)

rcr_srt(x)

rcr_str(x)

rcr_uno(x)

rcr_whl(x)
}
\arguments{
\item{x}{An ℝ object.}
}
\value{
Logical scalar.
}
\description{
Combinations of \link[ttt]{fundamental type} and
\link[mmm]{extended mode} properties.
}
\section{Functions}{
\itemize{
\item \code{ttt_mmm_vals()}: Get a character vector of all possible fundamental type
plus extended class properties.

\item \code{pop_atm()}: Is \code{x} a \link[ipop]{populated} \link[iatm]{atomic}
object?

\item \code{pop_ch1()}: Is \code{x} a \link[ipop]{populated}
\link[ich1]{onechar} object?

\item \code{pop_chr()}: Is \code{x} a \link[ipop]{populated}
\link[ichr]{character} object?

\item \code{pop_clr()}: Is \code{x} a \link[ipop]{populated} \link[iclr]{color}
object?

\item \code{pop_evn()}: Is \code{x} a \link[ipop]{populated}
\link[ievn]{even-valued whole-number} object?

\item \code{pop_fac()}: Is \code{x} a \link[ipop]{populated} \link[ifac]{factor}
object?

\item \code{pop_frc()}: Is \code{x} a \link[ipop]{populated}
\link[ifrc]{fractional-valued numeric} object?

\item \code{pop_ind()}: Is \code{x} a \link[ipop]{populated}
\link[iind]{indexer} object?

\item \code{pop_lgl()}: Is \code{x} a \link[ipop]{populated}
\link[ilgl]{logical} object?

\item \code{pop_neg()}: Is \code{x} a \link[ipop]{populated}
\link[ineg]{negative-valued mumeric} object?

\item \code{pop_ngw()}: Is \code{x} a \link[ipop]{populated}
\link[ingw]{negative-valued whole-number} object?

\item \code{pop_nng()}: Is \code{x} a \link[ipop]{populated}
\link[inng]{non-negative-valued numeric} object?

\item \code{pop_nnw()}: Is \code{x} a \link[ipop]{populated}
\link[innw]{non-negative-valued whole-number} object?

\item \code{pop_nps()}: Is \code{x} a \link[ipop]{populated}
\link[inps]{non-positive-valued numeric object} object?

\item \code{pop_npw()}: Is \code{x} a \link[ipop]{populated}
\link[inpw]{non-positive-valued whole-number} object?

\item \code{pop_nst()}: Is \code{x} a \link[ipop]{populated}
\link[inst]{non-sortable} object?

\item \code{pop_num()}: Is \code{x} a \link[ipop]{populated}
\link[inum]{numeric} object?

\item \code{pop_odd()}: Is \code{x} a \link[ipop]{populated}
\link[iodd]{odd-valued whole-number} object?

\item \code{pop_ord()}: Is \code{x} a \link[ipop]{populated}
\link[iord]{ordered-factor} object?

\item \code{pop_pct()}: Is \code{x} a \link[ipop]{populated}
\link[ipct]{percent-valued (0-100) numeric} object?

\item \code{pop_pos()}: Is \code{x} a \link[ipop]{populated}
\link[ipos]{positive-valued numeric object} object?

\item \code{pop_ppn()}: Is \code{x} a \link[ipop]{populated}
\link[ippn]{proportion-valued (0-1) numeric} object?

\item \code{pop_psw()}: Is \code{x} a \link[ipop]{populated}
\link[ipsw]{positive-valued whole-number} object?

\item \code{pop_srt()}: Is \code{x} a \link[ipop]{populated}
\link[isrt]{sortable} object?

\item \code{pop_str()}: Is \code{x} a \link[ipop]{populated} \link[istr]{string}
object?

\item \code{pop_uno()}: Is \code{x} a \link[ipop]{populated}
\link[iuno]{unordered-factor} object?

\item \code{pop_whl()}: Is \code{x} a \link[ipop]{populated}
\link[iwhl]{whole-number} object?

\item \code{cmp_atm()}: Is \code{x} a \link[icmp]{complete} \link[iatm]{atomic}
object?

\item \code{cmp_ch1()}: Is \code{x} a \link[icmp]{complete} \link[ich1]{onechar}
object?

\item \code{cmp_chr()}: Is \code{x} a \link[icmp]{complete}
\link[ichr]{character} object?

\item \code{cmp_clr()}: Is \code{x} a \link[icmp]{complete} \link[iclr]{color}
object?

\item \code{cmp_evn()}: Is \code{x} a \link[icmp]{complete}
\link[ievn]{even-valued whole-number} object?

\item \code{cmp_fac()}: Is \code{x} a \link[icmp]{complete} \link[ifac]{factor}
object?

\item \code{cmp_frc()}: Is \code{x} a \link[icmp]{complete}
\link[ifrc]{fractional-valued numeric} object?

\item \code{cmp_ind()}: Is \code{x} a \link[icmp]{complete} \link[iind]{indexer}
object?

\item \code{cmp_lgl()}: Is \code{x} a \link[icmp]{complete} \link[ilgl]{logical}
object?

\item \code{cmp_neg()}: Is \code{x} a \link[icmp]{complete}
\link[ineg]{negative-valued mumeric} object?

\item \code{cmp_ngw()}: Is \code{x} a \link[icmp]{complete}
\link[ingw]{negative-valued whole-number} object?

\item \code{cmp_nng()}: Is \code{x} a \link[icmp]{complete}
\link[inng]{non-negative-valued numeric} object?

\item \code{cmp_nnw()}: Is \code{x} a \link[icmp]{complete}
\link[innw]{non-negative-valued whole-number} object?

\item \code{cmp_nps()}: Is \code{x} a \link[icmp]{complete}
\link[inps]{non-positive-valued numeric object} object?

\item \code{cmp_npw()}: Is \code{x} a \link[icmp]{complete}
\link[inpw]{non-positive-valued whole-number} object?

\item \code{cmp_nst()}: Is \code{x} a \link[icmp]{complete}
\link[inst]{non-sortable} object?

\item \code{cmp_num()}: Is \code{x} a \link[icmp]{complete} \link[inum]{numeric}
object?

\item \code{cmp_odd()}: Is \code{x} a \link[icmp]{complete}
\link[iodd]{odd-valued whole-number} object?

\item \code{cmp_ord()}: Is \code{x} a \link[icmp]{complete}
\link[iord]{ordered-factor} object?

\item \code{cmp_pct()}: Is \code{x} a \link[icmp]{complete}
\link[ipct]{percent-valued (0-100) numeric} object?

\item \code{cmp_pos()}: Is \code{x} a \link[icmp]{complete}
\link[ipos]{positive-valued numeric object} object?

\item \code{cmp_ppn()}: Is \code{x} a \link[icmp]{complete}
\link[ippn]{proportion-valued (0-1) numeric} object?

\item \code{cmp_psw()}: Is \code{x} a \link[icmp]{complete}
\link[ipsw]{positive-valued whole-number} object?

\item \code{cmp_srt()}: Is \code{x} a \link[icmp]{complete}
\link[isrt]{sortable} object?

\item \code{cmp_str()}: Is \code{x} a \link[icmp]{complete} \link[istr]{string}
object?

\item \code{cmp_uno()}: Is \code{x} a \link[icmp]{complete}
\link[iuno]{unordered-factor} object?

\item \code{cmp_whl()}: Is \code{x} a \link[icmp]{complete}
\link[iwhl]{whole-number} object?

\item \code{emp_atm()}: Is \code{x} an \link[iemp]{empty}
\link[iatm]{atomic} object?

\item \code{emp_ch1()}: Is \code{x} an \link[iemp]{empty}
\link[ich1]{onechar} object?

\item \code{emp_chr()}: Is \code{x} an \link[iemp]{empty}
\link[ichr]{character} object?

\item \code{emp_clr()}: Is \code{x} an \link[iemp]{empty}
\link[iclr]{color} object?

\item \code{emp_evn()}: Is \code{x} an \link[iemp]{empty}
\link[ievn]{even-valued whole-number} object?

\item \code{emp_fac()}: Is \code{x} an \link[iemp]{empty}
\link[ifac]{factor} object?

\item \code{emp_frc()}: Is \code{x} an \link[iemp]{empty}
\link[ifrc]{fractional-valued numeric} object?

\item \code{emp_ind()}: Is \code{x} an \link[iemp]{empty}
\link[iind]{indexer} object?

\item \code{emp_lgl()}: Is \code{x} an \link[iemp]{empty}
\link[ilgl]{logical} object?

\item \code{emp_neg()}: Is \code{x} an \link[iemp]{empty}
\link[ineg]{negative-valued mumeric} object?

\item \code{emp_ngw()}: Is \code{x} an \link[iemp]{empty}
\link[ingw]{negative-valued whole-number} object?

\item \code{emp_nng()}: Is \code{x} an \link[iemp]{empty}
\link[inng]{non-negative-valued numeric} object?

\item \code{emp_nnw()}: Is \code{x} an \link[iemp]{empty}
\link[innw]{non-negative-valued whole-number} object?

\item \code{emp_nps()}: Is \code{x} an \link[iemp]{empty}
\link[inps]{non-positive-valued numeric object} object?

\item \code{emp_npw()}: Is \code{x} an \link[iemp]{empty}
\link[inpw]{non-positive-valued whole-number} object?

\item \code{emp_nst()}: Is \code{x} an \link[iemp]{empty}
\link[inst]{non-sortable} object?

\item \code{emp_num()}: Is \code{x} an \link[iemp]{empty}
\link[inum]{numeric} object?

\item \code{emp_odd()}: Is \code{x} an \link[iemp]{empty}
\link[iodd]{odd-valued whole-number} object?

\item \code{emp_ord()}: Is \code{x} an \link[iemp]{empty}
\link[iord]{ordered-factor} object?

\item \code{emp_pct()}: Is \code{x} an \link[iemp]{empty}
\link[ipct]{percent-valued (0-100) numeric} object?

\item \code{emp_pos()}: Is \code{x} an \link[iemp]{empty}
\link[ipos]{positive-valued numeric object} object?

\item \code{emp_ppn()}: Is \code{x} an \link[iemp]{empty}
\link[ippn]{proportion-valued (0-1) numeric} object?

\item \code{emp_psw()}: Is \code{x} an \link[iemp]{empty}
\link[ipsw]{positive-valued whole-number} object?

\item \code{emp_srt()}: Is \code{x} an \link[iemp]{empty}
\link[isrt]{sortable} object?

\item \code{emp_str()}: Is \code{x} an \link[iemp]{empty}
\link[istr]{string} object?

\item \code{emp_uno()}: Is \code{x} an \link[iemp]{empty}
\link[iuno]{unordered-factor} object?

\item \code{emp_whl()}: Is \code{x} an \link[iemp]{empty}
\link[iwhl]{whole-number} object?

\item \code{rcr_atm()}: Is \code{x} a \link[ircr]{recursive}
\link[iatm]{atomic} object?

\item \code{rcr_ch1()}: Is \code{x} a \link[ircr]{recursive}
\link[ich1]{onechar} object?

\item \code{rcr_chr()}: Is \code{x} a \link[ircr]{recursive}
\link[ichr]{character} object?

\item \code{rcr_clr()}: Is \code{x} a \link[ircr]{recursive}
\link[iclr]{color} object?

\item \code{rcr_evn()}: Is \code{x} a \link[ircr]{recursive}
\link[ievn]{even-valued whole-number} object?

\item \code{rcr_fac()}: Is \code{x} a \link[ircr]{recursive}
\link[ifac]{factor} object?

\item \code{rcr_frc()}: Is \code{x} a \link[ircr]{recursive}
\link[ifrc]{fractional-valued numeric} object?

\item \code{rcr_ind()}: Is \code{x} a \link[ircr]{recursive}
\link[iind]{indexer} object?

\item \code{rcr_lgl()}: Is \code{x} a \link[ircr]{recursive}
\link[ilgl]{logical} object?

\item \code{rcr_neg()}: Is \code{x} a \link[ircr]{recursive}
\link[ineg]{negative-valued mumeric} object?

\item \code{rcr_ngw()}: Is \code{x} a \link[ircr]{recursive}
\link[ingw]{negative-valued whole-number} object?

\item \code{rcr_nng()}: Is \code{x} a \link[ircr]{recursive}
\link[inng]{non-negative-valued numeric} object?

\item \code{rcr_nnw()}: Is \code{x} a \link[ircr]{recursive}
\link[innw]{non-negative-valued whole-number} object?

\item \code{rcr_nps()}: Is \code{x} a \link[ircr]{recursive}
\link[inps]{non-positive-valued numeric object} object?

\item \code{rcr_npw()}: Is \code{x} a \link[ircr]{recursive}
\link[inpw]{non-positive-valued whole-number} object?

\item \code{rcr_nst()}: Is \code{x} a \link[ircr]{recursive}
\link[inst]{non-sortable} object?

\item \code{rcr_num()}: Is \code{x} a \link[ircr]{recursive}
\link[inum]{numeric} object?

\item \code{rcr_odd()}: Is \code{x} a \link[ircr]{recursive}
\link[iodd]{odd-valued whole-number} object?

\item \code{rcr_ord()}: Is \code{x} a \link[ircr]{recursive}
\link[iord]{ordered-factor} object?

\item \code{rcr_pct()}: Is \code{x} a \link[ircr]{recursive}
\link[ipct]{percent-valued (0-100) numeric} object?

\item \code{rcr_pos()}: Is \code{x} a \link[ircr]{recursive}
\link[ipos]{positive-valued numeric object} object?

\item \code{rcr_ppn()}: Is \code{x} a \link[ircr]{recursive}
\link[ippn]{proportion-valued (0-1) numeric} object?

\item \code{rcr_psw()}: Is \code{x} a \link[ircr]{recursive}
\link[ipsw]{positive-valued whole-number} object?

\item \code{rcr_srt()}: Is \code{x} a \link[ircr]{recursive}
\link[isrt]{sortable} object?

\item \code{rcr_str()}: Is \code{x} a \link[ircr]{recursive} \link[istr]{string}
object?

\item \code{rcr_uno()}: Is \code{x} a \link[ircr]{recursive}
\link[iuno]{unordered-factor} object?

\item \code{rcr_whl()}: Is \code{x} a \link[ircr]{recursive}
\link[iwhl]{whole-number} object?

}}
\seealso{
Other props: 
\code{\link{as_mmm.}()},
\code{\link{ccc.}()},
\code{\link{cmp_ccc.}()},
\code{\link{cmp_mmm.}()},
\code{\link{cmp_mmm_ccc.}()},
\code{\link{comparable.}()},
\code{\link{compatible.}()},
\code{\link{ddd.}()},
\code{\link{eee.}()},
\code{\link{fff.}()},
\code{\link{meets.}()},
\code{\link{mmm.}()},
\code{\link{mmm_ccc.}()},
\code{\link{ppp.}()},
\code{\link{sss.}()},
\code{\link{ttt.}()},
\code{\link{ttt_ccc.}()}
}
\concept{props}
