% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u2u.R
\name{u2u}
\alias{u2u}
\alias{u2cm}
\alias{u2in}
\alias{u2mm}
\alias{u2pt}
\alias{cm2u}
\alias{in2u}
\alias{mm2u}
\alias{pt2u}
\alias{cm2in}
\alias{cm2mm}
\alias{cm2pt}
\alias{in2cm}
\alias{in2mm}
\alias{in2pt}
\alias{mm2cm}
\alias{mm2in}
\alias{mm2pt}
\alias{pt2cm}
\alias{pt2in}
\alias{pt2mm}
\title{Convert plotting units}
\usage{
u2u(x, old, new)

u2cm(x, old)

u2in(x, old)

u2mm(x, old)

u2pt(x, old)

cm2u(x, new)

in2u(x, new)

mm2u(x, new)

pt2u(x, new)

cm2in(x)

cm2mm(x)

cm2pt(x)

in2cm(x)

in2mm(x)

in2pt(x)

mm2cm(x)

mm2in(x)

mm2pt(x)

pt2cm(x)

pt2in(x)

pt2mm(x)
}
\arguments{
\item{x}{\link[=atm_num]{Atomic, numeric object}.}

\item{old, new}{\link[=cmp_chr_scl]{Complete character scalars} giving old and new units of distance, respectively: \code{'cm'}, \code{'in'}, \code{'mm'}, \code{'pt'} for centimeters, inches, millimeters, and points, respectively.}
}
\value{
An atomic, numeric object.
}
\description{
Unit functions take the following forms, where \code{aa} and \code{bb} are wildcard unit codes and \code{old} and \code{new} are user-supplied argument values.
\tabular{rl}{
\code{aa2bb}   \tab Converts units \emph{from} \code{aa} \emph{to} \code{bb}.
\cr \code{aa2u}   \tab    \emph{from} \code{aa} \emph{to} \code{new}.
\cr \code{u2bb}   \tab    \emph{from} \code{old} \emph{to} \code{bb}.
\cr  \code{u2u}   \tab    \emph{from} \code{old} \emph{to} \code{new}.
}
Available units are:
\tabular{rl}{
\code{'cm'}   \tab centimeters.
\cr \code{'in'}   \tab inches.
\cr \code{'mm'}   \tab millimeters.
\cr \code{'pt'}   \tab points (72/inch).
}
}
\examples{
u2u(1:3, "cm", "pt")
u2u(1:3, "in", "mm")
u2cm(1:3, "in")
in2u(1:3, "mm")
in2cm(1:3)
pt2mm(1:3)
}
\seealso{
Other plotting: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{pals}},
\code{\link{sca}()}
}
\concept{plotting}
