% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u2u.R
\encoding{UTF-8}
\name{u2u}
\alias{u2u}
\alias{u2cm}
\alias{u2in}
\alias{u2mm}
\alias{u2pt}
\alias{cm2u}
\alias{in2u}
\alias{mm2u}
\alias{pt2u}
\alias{cm2in}
\alias{cm2mm}
\alias{cm2pt}
\alias{in2cm}
\alias{in2mm}
\alias{in2pt}
\alias{mm2cm}
\alias{mm2in}
\alias{mm2pt}
\alias{pt2cm}
\alias{pt2in}
\alias{pt2mm}
\title{Convert plotting units}
\usage{
u2u(X, Old, New)

u2cm(X, Old)

u2in(X, Old)

u2mm(X, Old)

u2pt(X, Old)

cm2u(X, New)

in2u(X, New)

mm2u(X, New)

pt2u(X, New)

cm2in(X)

cm2mm(X)

cm2pt(X)

in2cm(X)

in2mm(X)

in2pt(X)

mm2cm(X)

mm2in(X)

mm2pt(X)

pt2cm(X)

pt2in(X)

pt2mm(X)
}
\arguments{
\item{X}{An \link[=atm_num]{atomic, numeric object}.}

\item{Old, New}{\link[=cmp_chr_scl]{Complete character scalars} giving Old and New units of distance, respectively: \code{'cm'} for centimeters, \code{'in'} for inches, \code{'mm'} for millimeters, and \code{'pt'} for points.}
}
\value{
An atomic numeric object.
}
\description{
Functions in this family convert among common plotting units.
}
\details{
Unit conversion functions take the following forms.
\tabular{ll}{  \strong{Function}   \tab \strong{Type of}                                              \cr
  \strong{Form}       \tab \strong{Conversion}                                           \cr
  \verb{\{aa\}2\{bb\}}    \tab from \code{'{aa}'} units to \code{'{bb}'} units\eqn{^{(1)}}.       \cr
  \verb{\{aa\}2u}       \tab from \code{'{aa}'} units to \code{New} units\eqn{^{(1,2)}}.        \cr
  \verb{u2\{bb\}}       \tab from \code{Old} units to \code{'{bb}'} units\eqn{^{(1,2)}}.        \cr
  \code{u2u}          \tab from \code{Old} units to \code{New} units.\eqn{^{(2)}}               }
\tabular{l}{  \eqn{^{(1)}} \code{{aa}} and \code{{bb}} are placeholders for any given plotting unit. \cr
  \eqn{^{(2)}} \code{Old} and \code{New} are user supplied arguments.                      }
\cr\cr Available unit codes are:
\tabular{ll}{  \code{'cm'}   \tab centimeters    \cr
  \code{'in'}   \tab inches         \cr
  \code{'mm'}   \tab millimeters    \cr
  \code{'pt'}   \tab points (72/inch) }
}
\examples{
u2u(1:3, "cm", "pt")
u2u(1:3, "in", "mm")
u2cm(1:3, "in")
u2in(1:3, "cm")
u2pt(1:3, "cm")
cm2u(1:3, "in")
in2u(1:3, "cm")
pt2u(1:3, "cm")
cm2in(1:3)
cm2mm(1:3)
cm2pt(1:3)
in2cm(1:3)
in2mm(1:3)
in2pt(1:3)
mm2cm(1:3)
mm2in(1:3)
mm2pt(1:3)
pt2cm(1:3)
pt2in(1:3)
pt2mm(1:3)
}
\seealso{
Other conversions: 
\code{\link{rng2lev}()},
\code{\link{to_dec}()},
\code{\link{tocase}()}

Other plots: 
\code{\link{color}},
\code{\link{ggp}()},
\code{\link{gr}},
\code{\link{markdown}()},
\code{\link{pals}},
\code{\link{sca}()}
}
\concept{conversions}
\concept{plots}
