% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{u2u}
\alias{u2u}
\alias{u2cm}
\alias{u2in}
\alias{u2mm}
\alias{u2pt}
\alias{cm2u}
\alias{in2u}
\alias{mm2u}
\alias{pt2u}
\alias{cm2in}
\alias{cm2mm}
\alias{cm2pt}
\alias{in2cm}
\alias{in2mm}
\alias{in2pt}
\alias{mm2cm}
\alias{mm2in}
\alias{mm2pt}
\alias{pt2cm}
\alias{pt2in}
\alias{pt2mm}
\title{Convert Plotting Units}
\usage{
u2u(x, old, new)

u2cm(x, old)

u2in(x, old)

u2mm(x, old)

u2pt(x, old)

cm2u(x, new)

in2u(x, new)

mm2u(x, new)

pt2u(x, new)

cm2in(x)

cm2mm(x)

cm2pt(x)

in2cm(x)

in2mm(x)

in2pt(x)

mm2cm(x)

mm2in(x)

mm2pt(x)

pt2cm(x)

pt2in(x)

pt2mm(x)
}
\arguments{
\item{x}{\link[=atm_num]{Atomic, numeric object}.}

\item{old, new}{\link[=cmp_chr_scl]{Complete character scalars} giving old and
new units of distance, respectively: \code{c('cm', 'in', 'mm', 'pt')} for
centimeters, inches, millimeters, and points, respectively.}
}
\value{
An atomic, numeric object.
}
\description{
Available units are:\tabular{ll}{
CODE   \tab NAME                                                      \cr
\code{cm}   \tab Centimeters                                               \cr
\verb{in}   \tab Inches                                                    \cr
\code{mm}   \tab Millimeters                                               \cr
\code{pt}   \tab Points (A)                                                  }
(A) Points are \code{1/72} of an inch.
\cr\cr
Unit functions take the following forms, where \code{aa} and \code{bb} are wildcard
units and \code{old} and \code{new} are user-supplied argument values:\tabular{ll}{
FUNCTION   \tab CONVERTS UNITS                                        \cr
\code{aa2bb}    \tab From \code{aa} to \code{bb}.                                    \cr
\code{aa2u}     \tab From \code{aa} to \code{new}.                                   \cr
\code{u2bb}     \tab From \code{old} to \code{bb}.                                   \cr
\code{u2u}      \tab From \code{old} to \code{new}.                                    }
}
