% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{u2u}
\alias{u2u}
\alias{u2cm}
\alias{u2in}
\alias{u2mm}
\alias{u2pt}
\alias{cm2u}
\alias{in2u}
\alias{mm2u}
\alias{pt2u}
\alias{cm2in}
\alias{cm2mm}
\alias{cm2pt}
\alias{in2cm}
\alias{in2mm}
\alias{in2pt}
\alias{mm2cm}
\alias{mm2in}
\alias{mm2pt}
\alias{pt2cm}
\alias{pt2in}
\alias{pt2mm}
\title{Convert among units of distance commonly used in plotting}
\usage{
u2u(x, old, new)

u2cm(x, old)

u2in(x, old)

u2mm(x, old)

u2pt(x, old)

cm2u(x, new)

in2u(x, new)

mm2u(x, new)

pt2u(x, new)

cm2in(x)

cm2mm(x)

cm2pt(x)

in2cm(x)

in2mm(x)

in2pt(x)

mm2cm(x)

mm2in(x)

mm2pt(x)

pt2cm(x)

pt2in(x)

pt2mm(x)
}
\arguments{
\item{x}{An atomic, numeric object.}

\item{old, new}{Character scalars giving old and new units of length. Valid
values are \code{c('cm', 'in', 'mm', 'pt')} for centimeters, inches,
millimeters, and points, respectively.}
}
\value{
An atomic, numeric object.
}
\description{
Convert units specifying both original and new units.
}
\section{Functions}{
\itemize{
\item \code{u2cm}: Convert to centimeter units.

\item \code{u2in}: Convert to inch units.

\item \code{u2mm}: Convert to millimeter units.

\item \code{u2pt}: Convert to point units.

\item \code{cm2u}: Convert centimeters to some other units.

\item \code{in2u}: Convert inches to some other units.

\item \code{mm2u}: Convert millimeters to some other units.

\item \code{pt2u}: Convert points to some other units.

\item \code{cm2in}: Convert centimeters to inches.

\item \code{cm2mm}: Convert centimeters to millimeters.

\item \code{cm2pt}: Convert centimeters to points.

\item \code{in2cm}: Convert inches to centimeters

\item \code{in2mm}: Convert inches to millimeters.

\item \code{in2pt}: Convert inches to points.

\item \code{mm2cm}: Convert millimeters to centimeters.

\item \code{mm2in}: Convert millimeters to inches.

\item \code{mm2pt}: Convert millimeters to points.

\item \code{pt2cm}: Convert points to centimeters.

\item \code{pt2in}: Convert points to inches.

\item \code{pt2mm}: Convert points to millimeters.
}}

