% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_ccc.R
\encoding{UTF-8}
\name{unq_ccc}
\alias{unq_ccc}
\alias{unq_ccc_funs}
\alias{unq_arr}
\alias{unq_dtf}
\alias{unq_gen}
\alias{unq_mat}
\alias{unq_mvc}
\alias{unq_scl}
\alias{unq_vec}
\alias{unq_vls}
\title{Unique + xclass combination properties}
\usage{
unq_ccc(x, ccc, ...)

unq_ccc_funs()

unq_arr(x, ...)

unq_dtf(x, ...)

unq_gen(x, ...)

unq_mat(x, ...)

unq_mvc(x, ...)

unq_scl(x, ...)

unq_vec(x, ...)

unq_vls(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{ccc}{A character scalar single xclass property from \code{\link{ccc_props}()}.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\strong{A character vector} \cr\cr \code{unq_ccc_funs}
\cr\cr  \strong{A logical scalar}   \cr\cr \verb{unq_ccc, unq_\{ccc\}}
}
\description{
Functions checking for combinations of \link[=UNQ]{uniqueness} and \link[=ccc]{xclass}.
}
\details{
\tabular{ll}{  \code{unq_ccc_funs}   \tab What \link[=UNQ]{unique} + \link[=ccc]{xclass} combination \link[=prop_funs]{property functions} are there?                          \cr   \tab   \cr
  \verb{unq_\{ccc\}}      \tab Is \code{x} both unique and a match to the single xclass property \code{'{ccc}'} where \code{{ccc}} is a placeholder for any given xclass property? \cr   \tab   \cr
  \code{unq_ccc}        \tab Is \code{x} both unique and a match to the single xclass property in argument \code{ccc}?                                                                     }
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{ll}{  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.                                                                  \cr   \tab   \cr
  \verb{lt, le, ge, gt}    \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively. \cr   \tab   \cr
  \verb{n, nr, nc}         \tab A vector of valid numbers of elements, rows, and columns, respectively.                                                                    \cr   \tab   \cr
  \code{vals}              \tab A vector of valid values.                                                                                                                                 }
}

\examples{
UnqArr <- array(1:27, dim = c(3, 3, 3))
UnqDtf <- data.frame(a = c("a", "b", "c"), b = 1:3)
UnqMat <- matrix(1:9, nrow = 3)
UnqMvc <- 1:2
UnqScl <- 1
UnqVec <- "a"
UnqVls <- list(a = letters, b = 1:3)

unq_ccc_funs()
unq_ccc(UnqArr, "arr")
unq_ccc(UnqVls, "dtf")
c(unq_dtf(UnqDtf), unq_arr(UnqArr), unq_arr(rbind(UnqDtf, UnqDtf)))
c(unq_mat(UnqMat), unq_arr(UnqVls), unq_arr(rbind(UnqMat, UnqMat)))
c(unq_arr(UnqArr), unq_arr(UnqDtf), unq_arr(c(UnqArr, UnqArr)))
c(unq_gen(UnqArr), unq_arr(UnqDtf), unq_arr(c(UnqArr, UnqArr)))
c(unq_gen(UnqVec), unq_arr(UnqArr), unq_arr(c(UnqVec, UnqCec)))
c(unq_gen(UnqMvc), unq_arr(UnqScl), unq_arr(c(UnqMvc, UnqMvc)))
c(unq_gen(UnqScl), unq_arr(UnqMvc), unq_arr(c(UnqScl, UnqScl)))
c(unq_mvc(UnqMvc), unq_arr(UnqVec), unq_arr(c(UnqMvc, UnqMvc)))
c(unq_scl(UnqScl), unq_arr(UnqMvc), unq_arr(c(UnqScl, UnqScl)))
c(unq_vec(UnqVec), unq_arr(UnqDtf), unq_arr(c(UnqVec, UnqVec)))
c(unq_vls(UnqVls), unq_arr(UnqVec), unq_arr(c(UnqVls, UnqVls)))
}
\seealso{
Other properties: 
\code{\link{as_mmm}()},
\code{\link{bbb_mmm}()},
\code{\link{bbb}()},
\code{\link{checker}},
\code{\link{cls}()},
\code{\link{cmp_ccc}()},
\code{\link{cmp_mmm}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{ddd}()},
\code{\link{eee}()},
\code{\link{iii}()},
\code{\link{is_unique}()},
\code{\link{meets}()},
\code{\link{mmm_ccc}()},
\code{\link{mmm}()},
\code{\link{ppp}()},
\code{\link{sss}()}
}
\concept{properties}
