% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_mmm.R
\encoding{UTF-8}
\name{unq_mmm}
\alias{unq_mmm}
\alias{unq_mmm_funs}
\alias{unq_atm}
\alias{unq_ch1}
\alias{unq_ch3}
\alias{unq_chr}
\alias{unq_clr}
\alias{unq_evn}
\alias{unq_fac}
\alias{unq_frc}
\alias{unq_ind}
\alias{unq_lgl}
\alias{unq_neg}
\alias{unq_ngw}
\alias{unq_nng}
\alias{unq_nnw}
\alias{unq_nps}
\alias{unq_npw}
\alias{unq_nst}
\alias{unq_num}
\alias{unq_odd}
\alias{unq_ord}
\alias{unq_pct}
\alias{unq_pos}
\alias{unq_ppn}
\alias{unq_psw}
\alias{unq_srt}
\alias{unq_str}
\alias{unq_uno}
\alias{unq_whl}
\title{unique + xmode combination properties}
\usage{
unq_mmm(x, mmm, ...)

unq_mmm_funs()

unq_atm(x, ...)

unq_ch1(x, ...)

unq_ch3(x, ...)

unq_chr(x, ...)

unq_clr(x, ...)

unq_evn(x, ...)

unq_fac(x, ...)

unq_frc(x, ...)

unq_ind(x, ...)

unq_lgl(x, ...)

unq_neg(x, ...)

unq_ngw(x, ...)

unq_nng(x, ...)

unq_nnw(x, ...)

unq_nps(x, ...)

unq_npw(x, ...)

unq_nst(x, ...)

unq_num(x, ...)

unq_odd(x, ...)

unq_ord(x, ...)

unq_pct(x, ...)

unq_pos(x, ...)

unq_ppn(x, ...)

unq_psw(x, ...)

unq_srt(x, ...)

unq_str(x, ...)

unq_uno(x, ...)

unq_whl(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{mmm}{A character scalar \code{xmode} property from \code{\link{mmm_props}()}`.}

\item{...}{
  Arguments passed on to \code{\link[=meets]{meets}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\emph{A character vector}
\cr   \code{unq_mmm_funs}
\cr\cr \emph{A logical scalar}
\cr   \code{unq_mmm}
\cr   \code{unq_MMM}
}
\description{
\tabular{rl}{
  \code{unq_mmm_funs}   \tab What \link[=iunq]{unique} + link[=mmm]{xmode} combination \link[=prop_funs]{property functions} are there?
\cr                \tab  
\cr    \code{unq_mmm}   \tab Is \code{x} both unique and a match to the single xmode property in argument \code{mmm}?
\cr                \tab  
\cr    \code{unq_MMM}   \tab Is \code{x} both unique and a match to single xmode property \code{'MMM'}?
}
}
\section{Specifying count and value restrictions}{
 Specifying restrictions in \code{...} is optional. The full set of recognized arguments names are defined in the following table along with the properties each specifies:
\tabular{rl}{
  \verb{max, maxr, maxc}   \tab Scalar maximum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr \verb{min, minr, minc}   \tab Scalar minimum valid numbers of element, rows, and columns, respectively.
\cr                     \tab  
\cr  \verb{lt, le, ge, gt}   \tab \link[=cmp_srt_scl]{Complete sortable scalar} less-than, less-than-or-equal, greater-than-or-equal, and greater-than bounds, respectively.
\cr                     \tab  
\cr       \verb{n, nr, nc}   \tab A vector of valid numbers of elements, rows, and columns, respectively.
\cr                     \tab  
\cr            \code{vals}   \tab A vector of valid values.
}
}

\examples{
unq_mmm_funs()
unq_mmm(letters, "ch1")
unq_mmm(c("abc", "def", "fgh"), "ch3")
unq_psw(0:10)
unq_ord(factor(letters, ordered = T))
}
\seealso{
Other props: 
\code{\link{bbb_ccc}()},
\code{\link{ccc}()},
\code{\link{cmp_mmm_ccc}()},
\code{\link{sss_ccc}()},
\code{\link{unq_mmm_ccc}()}
}
\concept{props}
