% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.r
\name{uv}
\alias{uv}
\alias{uav}
\alias{suv}
\title{Unique atomic values}
\usage{
uv(...)

uav(...)

suv(...)
}
\arguments{
\item{...}{One or more objects to be reduced to a single atomic vector and then further reduced to unique atomic values.}
}
\value{
An atomic vector.
}
\description{
\tabular{rl}{
  \verb{uv,uav}   \tab Gets a vector of unique atomic values contained across all \code{...} arguments..
\cr          \tab   
\cr  \code{suv}   \tab Gets a sorted vector of unique atomic values contained across all \code{...} arguments.
}
}
\examples{
abc <- letters[1:3]
bcd <- data.frame(letters[2:4])
num <- list(1, 2, 3)

abc
bcd
num

uv(0:3, 3:6, 6:9)
uav(abc, bcd, num)
suv(abc, bcd, num)
}
