% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.r
\encoding{UTF-8}
\name{uv}
\alias{uv}
\alias{uav}
\alias{suv}
\title{Unique atomic values}
\usage{
uv(...)

uav(...)

suv(...)
}
\arguments{
\item{...}{One or more objects to be reduced to a single atomic vector and then further reduced to unique atomic values.}
}
\value{
An atomic vector.
}
\description{
\tabular{rl}{
  \verb{uv, uav}   \tab Gets unique atomic values in all \code{...} arguments.
\cr           \tab  
\cr   \code{suv}   \tab Gets unique sorted atomic values in all \code{...} arguments.
}
}
\examples{
abc <- letters[1:3]
bcd <- data.frame(letters[2:4])
num <- list(1, 2, 3)

abc
bcd
num

uv(0:3, 3:6, 6:9)
uav(abc, bcd, num)
suv(abc, bcd, num)
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unique}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{wv}()},
\code{\link{xat}()}

Other values: 
\code{\link{atomize}()},
\code{\link{n_th}()},
\code{\link{na}()},
\code{\link{values}()},
\code{\link{wv}()}
}
\concept{extensions}
\concept{values}
