% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\alias{vnames}
\alias{vtable}
\alias{v}
\title{Package \code{uj} defined values}
\usage{
values()

vnames()

vtable()

v(...)
}
\arguments{
\item{...}{Unquoted, comma-separated list of names of constants to return. If multiple constants are specified, they are coerced into a single atomic vector result.}
}
\value{
\tabular{rl}{
\code{values}   \tab A \code{\link[=ivls]{vlist}}.
\cr \code{vtable}   \tab A data.frame.
\cr \code{vnames}   \tab A character vector.
\cr      \code{v}   \tab A character vector.
}
}
\description{
\tabular{rl}{
\code{vtable}   \tab Gets a data.frame of all \code{uj} named values with three columns: \code{name}, \code{value}, and \code{description}, where named values with multiple elements are pipe-delimited and all values are converted to mode character.
\cr \code{values}   \tab Gets a named list where names are the arguments submitted to \code{v(...)}, element values are the values of the contents of the named element, and the attribute \code{'descriptions'} gives verbose descriptions if needed.
\cr \code{vnames}   \tab Gets names of all package values.
\cr      \code{v}   \tab Gets package values by name.
}
}
\examples{
v(.)
v(pipe, pipe0, eq)
vnames()
values()
vtable()
}
