% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\alias{vnames}
\alias{v}
\title{Package Constants for \code{uj}}
\usage{
values()

vnames()

v(...)
}
\arguments{
\item{...}{unquoted, comma-separated list of the names of constants to
return. If multiple constants are specified, they are coerced into a single
atomic vector result.}
}
\value{
An atomic vector of length 1 or greater.
}
\description{
Package Constants for \code{uj}
}
\section{\code{values}}{
 Gets a named list where names are the arguments
submitted to \code{v(...)}, element values are the values of the constants,
and the attribute \code{'descriptions'} gives verbose descriptions if
needed.
}

\section{\code{vnames}}{
 Gets a character vector of all available \code{uj}
package constant names.
}

\section{\code{v}}{
 Gets one or more \code{uj} package constants by name.
}

