% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\encoding{UTF-8}
\name{values}
\alias{values}
\alias{vnames}
\alias{vtable}
\alias{v}
\title{Named values (unique to package \code{uj})}
\usage{
values()

vnames()

vtable()

v(...)
}
\arguments{
\item{...}{Unquoted, comma-separated list of names of constants to return. If multiple constants are specified, they are coerced into a single atomic vector result.}
}
\value{
\emph{A character vector}
\cr    \code{vnames}
\cr    \code{v}
\cr
\cr \emph{A \link[ivls]{vlist}}
\cr    \code{values}
\cr
\cr \emph{A data.frame}
\cr    \code{vtable}
}
\description{
\tabular{rl}{
  \code{vtable}   \tab Gets a data.frame of all \code{uj}-specific named values with three columns: \code{name}, \code{value}, and \code{description}, where named values with multiple elements are pipe-delimited and all values are converted to mode character.
\cr            \tab  
\cr \code{values}   \tab Gets a named list where names are the arguments submitted to \code{v(...)}, element values are the values of the contents of the named element, and the attribute \code{'descriptions'} gives associated verbose descriptions.
\cr            \tab  
\cr \code{vnames}   \tab Gets names of all package values.
\cr      \code{v}   \tab Gets package values by name.
}
}
\examples{
v(.)
v(pipe, pipe0, eq)
vnames()
values()
vtable()
}
\seealso{
Other values: 
\code{\link{atomize}()},
\code{\link{n_th}()},
\code{\link{na}()},
\code{\link{uv}()},
\code{\link{wv}()}
}
\concept{values}
