% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\encoding{UTF-8}
\name{values}
\alias{values}
\alias{value_names}
\alias{value_table}
\alias{value}
\alias{val_names}
\alias{val_table}
\alias{val}
\alias{vs}
\alias{vnames}
\alias{vtable}
\alias{v}
\title{Named values (unique to package \code{uj})}
\usage{
values()

value_names()

value_table()

value(...)

values()

val_names()

val_table()

val(...)

vs()

vnames()

vtable()

v(...)
}
\arguments{
\item{...}{Unquoted, comma-separated list of names of values to return. If multiple values are specified, they are coerced into a single atomic vector result.}
}
\value{
\strong{A character vector}         \cr\cr \verb{value_names, val_names, vnames}
\cr\cr  \strong{A data.frame}               \cr\cr \verb{value_table, val_table, vtable}
\cr\cr  \strong{A \link[=VLS]{vlist}}       \cr\cr \verb{values, values, vs}
\cr\cr  \strong{An atomic scalar or vector} \cr\cr \verb{value, val, v}
}
\description{
Get names, definitions, and values of special \code{uj} package named values.
}
\details{
\tabular{ll}{  \verb{value_names, val_names, vnames}   \tab Get names of all package values.                                                                             \cr   \tab   \cr
  \verb{value_table, val_table, vtable}   \tab Get a data.frame of all \code{uj}-specific named values with three columns: \code{name}, \code{value}, and \code{description},
where named values with multiple elements are pipe-delimited and all values are converted to mode character. \cr   \tab   \cr
  \verb{values, values, vs}               \tab Get a named list where names are the arguments submitted to \code{v(...)}, element values are the values of
the contents of the named element, and the attribute \code{'descriptions'} gives associated verbose descriptions. \cr   \tab   \cr
  \verb{value, val, v}                    \tab Gets package values by name.                                                                                                }
}
\examples{
v(.)
v(pipe, pipe0, eq)
vnames()
values()
vtable()
}
\seealso{
Other values: 
\code{\link{Nth}()},
\code{\link{atomize}()},
\code{\link{na}()}
}
\concept{values}
