% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\encoding{UTF-8}
\name{x}
\alias{x}
\alias{removal}
\alias{xall}
\alias{xat}
\alias{xat0}
\alias{xcls}
\alias{xwarn}
\alias{xplot}
\alias{xcon}
\title{'\code{x}' out objects or their characteristics}
\usage{
x(...)

xall()

xat(x, ...)

xat0(x)

xcls(x, ...)

xwarn()

xplot(x = NULL)

xcon()
}
\arguments{
\item{...}{One or more \link[=cmp_chr_vec]{complete character vecs} containing names of attributes or classes to remove.}

\item{x}{An R object.}
}
\value{
\code{NULL}
\cr    \verb{xplot \\cr    }xwarn\verb{ \\cr    }xcon\verb{ \\cr    }xall\verb{ \\cr    }x\verb{ \\cr \\cr *An* R *object* \\cr    }xcls\verb{ \\cr    }xat0\verb{ \\cr    }xat`
}
\description{
All \code{x} (removal) functions are followed by garbage collection.
\tabular{rl}{
\code{xall}   \tab \strong{WARNING: Use with care!}
\cr             \tab \strong{Purges the global environment.}
\cr             \tab  
\cr   \code{xplot}   \tab Purges plots and prints \code{x}.
\cr   \code{xwarn}   \tab Purges warnings.
\cr    \code{xcon}   \tab Purges console.
\cr    \code{xcls}   \tab Removes classes by name.
\cr    \code{xat0}   \tab Removes all attributes.
\cr     \code{xat}   \tab Removes attributes by name.
\cr       \code{x}   \tab Deletes \code{...} args in the calling environment.
}
}
\examples{
obj. <- data.frame(x = -5:5, y = sqrt(abs(-5:5)))
class(obj.) <- c(class(obj.), "ccc")
attr(obj., "aaa") <- "aaa"

## structure with custom class and attribute.
str(obj.)

## structure after removing just the custom class.
str(xcls(obj., "ccc"))

## structure after removing just the custom attribute.
str(xat(obj., "aaa"))

## structure after removing all attributes.
str(xat0(obj.))

\dontrun{
## build a ggplot scatterplot object (y.)
library(ggplot2)
ggp. <- ggplot(obj., aes(x, y)) + geom_point()

## print to console, wait for user, purge console
say("\n A message")
continue()
xcon()

## display a scatterplot, wait for user
plot(obj.$x, obj.$y)
continue()

## purge plots and print ggplot
xplot(ggp.)

## wait for user, purge plots again
continue()
xplot()

## generate warning, print warnings
sqrt(-1)
warnings()

## purge warnings, validate that warnings are purged
xwarn()
warnings()
}
}
\seealso{
Other extensions: 
\code{\link{atomize}()},
\code{\link{case}()},
\code{\link{cb}()},
\code{\link{comparable}()},
\code{\link{compatible}()},
\code{\link{declare}},
\code{\link{dups}()},
\code{\link{envir_vals}},
\code{\link{evals}},
\code{\link{failsafe_binaries}},
\code{\link{failsafe_is}},
\code{\link{failsafe}()},
\code{\link{fork}()},
\code{\link{is_unq}()},
\code{\link{n_is}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{paths}},
\code{\link{rd}},
\code{\link{r}()},
\code{\link{sets}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uv}()},
\code{\link{wv}()}

Other meta: 
\code{\link{callers}()},
\code{\link{dots_uj}},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{recycling}},
\code{\link{run}()},
\code{\link{say}()}
}
\concept{extensions}
\concept{meta}
