% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xb.R
\name{xb}
\alias{xb}
\alias{cb}
\alias{rb}
\title{Error-Checked Row and Column Binding}
\usage{
xb(...)

cb(...)

rb(...)
}
\arguments{
\item{...}{Multiple \code{\link{compatible}} atomic matrices or multiple
compatible \link[=atm_dtf]{atomic data.frames}.}
}
\value{
An \link[=atm_dtf]{atomic data.frame} or an atomic matrix.
}
\description{
\tabular{ll}{
FUNCTION   \tab WHAT IT DOES \cr
\code{xb}       \tab Row binds if \code{...} arguments are compatible only for row
binding. Column binds if arguments are compatible only for
column binding. Throws an error if arguments are either
compatible for both row and column binding or not
compatible for binding at all.                          \cr
\code{cb}       \tab Column bind an arbitrary number of atomic matrices or
\link[=atm_dtf]{atomic data.frames}.                    \cr
\code{rb}       \tab Row bind an arbitrary number of atomic matrices or
\link[=atm_dtf]{atomic data.frames}.                      }
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()},
\code{\link{uj_logicals}}
}
\concept{extensions}
