% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xb.R
\name{xb}
\alias{xb}
\alias{cb}
\alias{rb}
\title{Error-checked row and column binding}
\usage{
xb(..., bind = NA)

cb(...)

rb(...)
}
\arguments{
\item{...}{Multiple \code{\link{compatible}} atomic matrices or multiple
compatible \link[=atm_dtf]{atomic data.frames}.}

\item{bind}{\link[=ch1_scl]{Onechar scalar} indicating how to bind:
\code{'r'} for row binding, \code{'c'} for column-binding, \code{NA} to
identify whether either row or column binding is successful and return
whichever was successful.}
}
\value{
An \link[=atm_dtf]{atomic data.frame} or an atomic matrix.
}
\description{
\tabular{ll}{
\code{xb.}   \tab Column or row bind or \link[=atm_dtf]{atomic data.frames}
in \code{...}. If \code{bind = NA}, \code{xb} attempts to
both row and column bind arguments in \code{...}. If both
are successful, processes an error indicating ambiguous
dimension for binding. If neither is successful,
processes an associated error. Otherwise, returns the
result that was successful.                           \cr
\code{cb}    \tab Column bind an arbitrary number of atomic matrices or
\link[=atm_dtf]{atomic data.frames}.                  \cr
\code{rb}    \tab Row bind an arbitrary number of atomic matrices or
\link[=atm_dtf]{atomic data.frames}.                    }
Requires all arguments in \code{...} to be \code{\link{compatible}} atomic
matrices or \link[=atm_dtf]{atomic data.frames}.
}
\seealso{
Other extensions: 
\code{\link{atomize}},
\code{\link{binfuns}},
\code{\link{callers}()},
\code{\link{case}()},
\code{\link{dots_uj}},
\code{\link{envir_vals}},
\code{\link{ex}()},
\code{\link{failsafe}()},
\code{\link{files_uj}},
\code{\link{fork}()},
\code{\link{fsub}()},
\code{\link{is_failsafe}},
\code{\link{is_unq}()},
\code{\link{logicals_uj}()},
\code{\link{make_uj}},
\code{\link{n_is}()},
\code{\link{n_th}()},
\code{\link{naming}},
\code{\link{na}()},
\code{\link{os}()},
\code{\link{pause}()},
\code{\link{ply}()},
\code{\link{put.}()},
\code{\link{reclass}()},
\code{\link{recycling}},
\code{\link{removal}},
\code{\link{run}()},
\code{\link{stats0}},
\code{\link{swap}()}
}
\concept{extensions}
